% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_completion.R
\name{activate_polars_rstudio_completion}
\alias{activate_polars_rstudio_completion}
\title{activate_polars_rstudio_completion}
\description{
activate_polars_rstudio_completion
}
\examples{
.dev$activate_polars_rstudio_completion()
}
