% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{and_then}
\alias{and_then}
\title{map an ok-value or pass on err-value}
\usage{
and_then(x, f)
}
\arguments{
\item{x}{any R object}

\item{f}{a closure that takes the ok part as input}
}
\value{
same R object wrapped in a Err-result
}
\description{
map an ok-value or pass on err-value
}
