% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_arg_min}
\alias{arr_arg_min}
\alias{arr.arg_min}
\title{Arg min sublists}
\format{
function
}
\value{
Expr
}
\description{
Retrieve the index of the minimal value in every sublist.
}
\examples{
df = pl$DataFrame(list(s = list(1:2, 2:1)))
df$select(pl$col("s")$list$arg_min())
}
\keyword{ExprArr}
