% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_concat}
\alias{arr_concat}
\alias{arr.concat}
\title{concat another list}
\format{
function
}
\arguments{
\item{other}{Rlist, Expr or column of same type as self.}
}
\value{
Expr
}
\description{
Concat the arrays in a Series dtype List in linear time.
}
\examples{
df = pl$DataFrame(
  a = list("a", "x"),
  b = list(c("b", "c"), c("y", "z"))
)
df$select(pl$col("a")$list$concat(pl$col("b")))

df$select(pl$col("a")$list$concat("hello from R"))

df$select(pl$col("a")$list$concat(list("hello", c("hello", "world"))))
}
\keyword{ExprArr}
