% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_contains}
\alias{arr_contains}
\alias{arr.contains}
\title{Sublists contains}
\format{
function
}
\arguments{
\item{item}{any into Expr/literal}
}
\value{
Expr of a boolean mask
}
\description{
Check if sublists contain the given item.
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) # NULL or integer() or list()
df$select(pl$col("a")$arr$contains(1L))
}
\keyword{ExprArr}
