% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_diff}
\alias{arr_diff}
\alias{arr_max}
\alias{arr.arg_max}
\title{Diff sublists}
\format{
function
}
\arguments{
\item{n}{Number of slots to shift}

\item{null_behavior}{choice "ignore"(default) "drop"}
}
\value{
Expr
}
\description{
Calculate the n-th discrete difference of every sublist.
}
\examples{
df = pl$DataFrame(list(s = list(1:4,c(10L,2L,1L))))
df$select(pl$col("s")$arr$diff())
}
\keyword{ExprArr}
