% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_eval}
\alias{arr_eval}
\alias{arr.eval}
\title{eval sublists (kinda like lapply)}
\format{
function
}
\arguments{
\item{Expr}{Expression to run. Note that you can select an element with \code{pl$first()}, or
\code{pl$col()}}

\item{parallel}{bool
Run all expression parallel. Don't activate this blindly.
Parallelism is worth it if there is enough work to do per thread.
This likely should not be use in the groupby context, because we already
parallel execution per group}
}
\value{
Expr
}
\description{
Run any polars expression against the lists' elements.
}
\examples{
df = pl$DataFrame(a = list(c(1, 8, 3), b = c(4, 5, 2)))
df$select(pl$all()$cast(pl$dtypes$Int64))$with_columns(
  pl$concat_list(c("a", "b"))$list$eval(pl$element()$rank())$alias("rank")
)
}
\keyword{ExprArr}
