% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_get}
\alias{arr_get}
\alias{arr.get}
\title{Get list}
\format{
function
}
\value{
Expr
}
\description{
Get the value by index in the sublists.
So index \code{0} would return the first item of every sublist
and index \code{-1} would return the last item of every sublist
if an index is out of bounds, it will return a \code{None}.
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) #NULL or integer() or list()
df$select(pl$col("a")$arr$get(0))
}
\keyword{ExprArr}
