% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_lengths}
\alias{arr_lengths}
\alias{lengths}
\alias{arr.lengths}
\title{Lengths arrays in list}
\format{
function
}
\value{
Expr
}
\description{
Get the length of the arrays as UInt32
}
\examples{
df = pl$DataFrame(list_of_strs = pl$Series(list(c("a","b"),"c")))
df$with_column(pl$col("list_of_strs")$arr$lengths()$alias("list_of_strs_lengths"))
}
\keyword{ExprArr}
