% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_max}
\alias{arr_max}
\alias{Expr_arr_max}
\alias{Expr_arr.max}
\title{Max lists}
\format{
function
}
\value{
Expr
}
\description{
Compute the max value of the lists in the array.
}
\examples{
df = pl$DataFrame(values = pl$Series(list(1L,2:3)))
df$select(pl$col("values")$arr$max())
}
\keyword{ExprArr}
