% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_min}
\alias{arr_min}
\alias{Expr_arr_min}
\alias{Expr_arr.min}
\title{#' Min lists}
\format{
function
}
\value{
Expr
}
\description{
Compute the min value of the lists in the array.
}
\examples{
df = pl$DataFrame(values = pl$Series(list(1L,2:3)))
df$select(pl$col("values")$arr$min())
}
\keyword{ExprArr}
