% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_shift}
\alias{arr_shift}
\alias{arr.shift}
\title{Shift sublists}
\format{
function
}
\arguments{
\item{periods}{Value. Number of places to shift (may be negative).}
}
\value{
Expr
}
\description{
Shift values by the given period.
}
\examples{
df = pl$DataFrame(list(s = list(1:4, c(10L, 2L, 1L))))
df$select(pl$col("s")$arr$shift())
}
\keyword{ExprArr}
