% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_sort}
\alias{arr_sort}
\title{Expr_sort}
\arguments{
\item{descending}{Sort values in descending order}
}
\value{
Expr
}
\description{
Sort this column. In projection/ selection context the whole column is sorted.
If used in a groupby context, the groups are sorted.
}
