% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_take}
\alias{arr_take}
\alias{arr.take}
\title{take in sublists}
\format{
function
}
\value{
Expr
}
\description{
Get the take value of the sublists.
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) #NULL or integer() or list()
idx = pl$Series(list(0:1,1L,1L))
df$select(pl$col("a")$arr$take(99))
}
\keyword{ExprArr}
