% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_unique}
\alias{arr_unique}
\alias{arr.unique}
\title{Unique list}
\format{
function
}
\value{
Expr
}
\description{
Get the unique/distinct values in the list.
}
\examples{
df = pl$DataFrame(list(a = list(1, 1, 2)))
df$select(pl$col("a")$arr$unique())
}
\keyword{ExprArr}
