% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{as.character.RThreadHandle}
\alias{as.character.RThreadHandle}
\title{RThreadHandle to string}
\usage{
\method{as.character}{RThreadHandle}(x, ...)
}
\arguments{
\item{x}{RThreadHandle}

\item{...}{not used}
}
\description{
get description of RThreadHandle as string
}
\keyword{internal}
