% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{as.data.frame.DataFrame}
\alias{as.data.frame.DataFrame}
\title{as.data.frame.DataFrame S3 method}
\usage{
\method{as.data.frame}{DataFrame}(x, ...)
}
\arguments{
\item{x}{DataFrame}

\item{...}{any params passed to as.data.frame}
}
\value{
data.frame
}
\description{
as.data.frame.DataFrame S3 method
}
\examples{
as.data.frame(pl$DataFrame(iris[1:3,]))
}
