% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{as.list.Expr}
\alias{as.list.Expr}
\title{as.list Expr}
\usage{
\method{as.list}{Expr}(x, ...)
}
\arguments{
\item{x}{Expr}

\item{...}{not used}
}
\value{
One Expr wrapped in a list
}
\description{
wraps an Expr in a list
}
\keyword{Expr}
