% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{c.Series}
\alias{c.Series}
\title{polars_pl$Series = \(x, name=NULL) {}
\usage{
\method{c}{Series}(x, ...)
}
\description{
if(!is.null(name) && !is_string(name)) abort("name must be a string")
}
\details{

}
