% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{$.pl_polars_env}
\alias{$.pl_polars_env}
\title{get public function from pl namespace/env}
\usage{
\method{$}{pl_polars_env}(self, name)
}
\description{
get public function from pl namespace/env
}
\details{
This method if polars_optenv$debug_polars == TRUE will print what methods are called
}
\keyword{internal}
