% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{$.private_polars_env}
\alias{$.private_polars_env}
\title{get private method from Class}
\usage{
\method{$}{private_polars_env}(self, name)
}
\description{
get private method from Class
}
\details{
This method if polars_optenv$debug_polars == TRUE will print what methods are called
}
\keyword{internal}
