% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{$<-.DataFrame}
\alias{$<-.DataFrame}
\title{generic setter method}
\usage{
\method{$}{DataFrame}(self, name) <- value
}
\value{
value
}
\description{
set value of properties of DataFrames
}
\examples{
#For internal use
#is only activated for following methods of DataFrame
ls(minipolars:::DataFrame.property_setters)

#specific use case for one object property 'columns' (names)
df = pl$DataFrame(iris)

#get values
df$columns

#set + get values
df$columns = letters[1:5] #<- is fine too
df$columns
}
\keyword{DataFrame}
