% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_tz_to_result}
\alias{check_tz_to_result}
\title{Verify correct time zone}
\usage{
check_tz_to_result(tz, allow_null = TRUE)
}
\arguments{
\item{tz}{time zone string or NULL}

\item{allow_null}{bool, if TRUE accept NULL}
}
\value{
a result object, with either a valid string or an Err
}
\description{
Verify correct time zone
}
\examples{
check_tz_to_result = polars:::check_tz_to_result # expose internal
# return Ok
check_tz_to_result("GMT")
check_tz_to_result(NULL)

# return Err
check_tz_to_result("Alice")
check_tz_to_result(42)
check_tz_to_result(NULL, allow_null = FALSE)
}
\keyword{internal}
