% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose}
\alias{choose}
\title{Simple match/switch handler}
\usage{
choose(..., or_else = NULL)
}
\arguments{
\item{...}{odd arugments are bool statements, a next even is returned if prior bool statement is the first true}

\item{or_else}{return this if no bool statements were true}
}
\value{
any return given first true bool statement otherwise value of or_else
}
\description{
Simple match/switch handler
}
\examples{
n=5
choose(
 n<5,"nope",
 n>6,"yeah",
 or_else = abort(paste("failed to have a case for n=",n))
)
}
