% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_functions.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce}
\arguments{
\item{...}{is a:
If one arg:
\itemize{
\item Series or Expr, same as \code{column$sum()}
\item string, same as \code{pl$col(column)$sum()}
\item numeric, same as \code{pl$lit(column)$sum()}
\item list of strings(column names) or exprressions to add up as expr1 + expr2 + expr3 + ...
}

If several args, then wrapped in a list and handled as above.}

\item{exprs}{list of Expr or Series or strings or a mix, or a char vector}
}
\value{
Expr

Expr
}
\description{
Folds the expressions from left to right, keeping the first non-null value.

Folds the expressions from left to right, keeping the first non-null value.
}
\examples{
df = pl$DataFrame(
  a = NA_real_,
  b = c(1:2,NA_real_,NA_real_),
  c = c(1:3,NA_real_)
)
#use coalesce to get first non Null value for each row, otherwise insert 99.9
df$with_column(
  pl$coalesce("a", "b", "c", 99.9)$alias("d")
)

#Create lagged columns and collect them into a list. This mimics a rolling window.
df = pl$DataFrame(A = c(1,2,9,2,13))
df$with_columns(lapply(
  0:2,
  \(i) pl$col("A")$shift(i)$alias(paste0("A_lag_",i))
))$select(
  pl$concat_list(lapply(2:0,\(i) pl$col(paste0("A_lag_",i))))$alias(
  "A_rolling"
 )
)

#concat Expr a Series and an R obejct
pl$concat_list(list(
  pl$lit(1:5),
  pl$Series(5:1),
  rep(0L,5)
))$alias("alice")$lit_to_s()

}
\keyword{Expr_new}
