% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R, R/functions.R
\name{concat}
\alias{concat}
\alias{arr_concat}
\alias{arr.concat}
\title{concat another list}
\format{
function
}
\usage{
concat(
  l,
  rechunk = TRUE,
  how = c("vertical", "horizontal", "diagnoal"),
  parallel = TRUE
)
}
\arguments{
\item{l}{list of DataFrame, or Series, LazyFrame or Expr}

\item{rechunk}{perform a rechunk at last}

\item{how}{choice of bind direction "vertical"(rbind) "horizontal"(cbind) "diagnoal" diagonally}

\item{parallel}{BOOL default TRUE, only used for LazyFrames}

\item{other}{Rlist, Expr or column of same tyoe as self.}
}
\value{
Expr

DataFrame, or Series, LazyFrame or Expr
}
\description{
Concat the arrays in a Series dtype List in linear time.
}
\examples{
df = pl$DataFrame(
  a = list("a","x"),
  b = list(c("b","c"),c("y","z"))
)
df$select(pl$col("a")$arr$concat(pl$col("b")))

df$select(pl$col("a")$arr$concat("hello from R"))

df$select(pl$col("a")$arr$concat(list("hello",c("hello","world"))))
#vertical
 l_ver = lapply(1:10, function(i) {
l_internal = list(
 a = 1:5,
 b = letters[1:5]
)
pl$DataFrame(l_internal)
})
pl$concat(l_ver, how="vertical")


#horizontal
l_hor = lapply(1:10, function(i) {
 l_internal = list(
   1:5,
   letters[1:5]
 )
 names(l_internal) = paste0(c("a","b"),i)
 pl$DataFrame(l_internal)
})
pl$concat(l_hor, how = "horizontal")


#diagonal
pl$concat(l_hor, how = "diagonal")
}
\keyword{ExprArr}
