% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_DataTypeVector}
\alias{construct_DataTypeVector}
\title{construct protoArrayExpr}
\usage{
construct_DataTypeVector(l)
}
\arguments{
\item{l}{list of Expr or string}
}
\value{
extptr to ProtoExprArray with all exprs or strings
}
\description{
construct protoArrayExpr
}
\examples{
minipolars:::construct_protoArrayExpr(list("column_a",pl$col("column_b")))
}
