% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_protoArrayExpr}
\alias{construct_protoArrayExpr}
\title{construct protoArrayExpr}
\usage{
construct_protoArrayExpr(l)
}
\arguments{
\item{l}{list of Expr or string}
}
\value{
extptr to ProtoExprArray with all exprs or strings
}
\description{
construct protoArrayExpr
}
\examples{
rpolars:::construct_protoArrayExpr(list("column_a",pl$col("column_b")))
}
