% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{construct_ProtoExprArray}
\alias{construct_ProtoExprArray}
\title{construct proto Expr array from args}
\usage{
construct_ProtoExprArray(...)
}
\arguments{
\item{...}{any Expr or string}
}
\value{
ProtoExprArray object
}
\description{
construct proto Expr array from args
}
\examples{
construct_ProtoExprArray(pl$col("Species"),"Sepal.Width")
}
