% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__array}
\alias{cs__array}
\title{Select all array columns}
\usage{
cs__array(inner = NULL, ..., width = NULL)
}
\arguments{
\item{inner}{An optional inner \link[=cs]{selector} to select columns having
specific inner \link[=DataType]{data types}. If \code{NULL}, all inner types are selected.}

\item{...}{These dots are for future extensions and must be empty.}

\item{width}{An optional integer specifying the width of the array columns to select.
If \code{NULL}, all widths are selected.}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(
  foo = list(c("xx", "yy"), c("x", "y")),
  bar = list(123, 456),
  baz = c(2.0, 5.5),
  .schema_overrides = list(
    foo = pl$Array(pl$String, 2),
    bar = pl$Array(pl$Int64, 1)
  )
)

# Select all array columns:
df$select(cs$array())

# Select all columns except for those that are array:
df$select(!cs$array())

# If you want to select specific array columns,
# you can specify the inner data type and/or width:
df$select(cs$array(cs$string()))
df$select(cs$array(width = 1))
df$select(cs$array(cs$string() | cs$numeric(), width = 2))
}
\seealso{
\itemize{
\item \link{cs} for the documentation on operators supported by selectors.
\item \code{\link[=cs__by_dtype]{cs$by_dtype()}}: Select all columns matching the given dtype(s).
\item \code{\link[=cs__list]{cs$list()}}: Select all list columns.
\item \code{\link[=cs__nested]{cs$nested()}}: Select all nested columns.
}
}
