% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__binary}
\alias{cs__binary}
\title{Select all binary columns}
\usage{
cs__binary()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all binary columns
}
\examples{
df <- pl$select(
  a = charToRaw("hello"),
  b = pl$lit("world"),
  c = charToRaw("!"),
  d = pl$lit(":"),
)

# Select binary columns:
df$select(cs$binary())

# Select all columns except for those that are binary:
df$select(!cs$binary())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
