% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__boolean}
\alias{cs__boolean}
\title{Select all boolean columns}
\usage{
cs__boolean()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all boolean columns
}
\examples{
df <- pl$DataFrame(
  a = 1:4,
  b = c(FALSE, TRUE, FALSE, TRUE)
)

# Select and invert boolean columns:
df$with_columns(inverted = cs$boolean()$not())

# Select all columns except for those that are boolean:
df$select(!cs$boolean())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
