% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__by_index}
\alias{cs__by_index}
\title{Select all columns matching the given indices (or range objects)}
\usage{
cs__by_index(indices, ..., require_all = TRUE)
}
\arguments{
\item{indices}{0-based column indices to select.
Negative indexing is supported.}

\item{...}{These dots are for future extensions and must be empty.}

\item{require_all}{Whether to match all indices (the default) or any of the indices.}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all columns matching the given indices (or range objects)
}
\details{
Matching columns are returned in the order in which their indexes appear in
the selector, not the underlying schema order.
}
\examples{
vals <- as.list(0.5 * 0:100)
names(vals) <- paste0("c", 0:100)
df <- pl$DataFrame(!!!vals)
df

# Select columns by index (the two first/last columns):
df$select(cs$by_index(c(0, 1, -2, -1)))

# Use seq()
df$select(cs$by_index(c(0, seq(1, 101, 20)), require_all = FALSE))
df$select(cs$by_index(c(0, seq(101, 0, -25)), require_all = FALSE))

# Select only odd-indexed columns:
df$select(!cs$by_index(seq(0, 100, 2)))
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
