% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__by_name}
\alias{cs__by_name}
\title{Select all columns matching the given names}
\usage{
cs__by_name(..., names, require_all = TRUE, expand_patterns = FALSE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Deprecated compatibility
interface for passing multiple column names.}

\item{names}{A character vector of column names to select.}

\item{require_all}{Whether to match all names (the default) or any of the names.}

\item{expand_patterns}{Whether to expand regex patterns (\verb{^...$}) and
wildcards (\code{*}) in names. Default is \code{FALSE} (treat names as literals).}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all columns matching the given names
}
\details{
Matching columns are returned in the order in which their indexes appear in
the selector, not the underlying schema order.
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123, 456),
  baz = c(2.0, 5.5),
  zap = c(FALSE, TRUE)
)

# Select columns by name:
df$select(cs$by_name(c("foo", "bar")))

# Match any of the given columns by name:
df$select(cs$by_name(c("baz", "moose", "foo", "bear"), require_all = FALSE))

# Match all columns except for those given:
df$select(!cs$by_name(c("foo", "bar")))
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
