% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__date}
\alias{cs__date}
\title{Select all date columns}
\usage{
cs__date()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all date columns
}
\examples{
df <- pl$DataFrame(
  dtm = as.POSIXct(c("2001-5-7 10:25", "2031-12-31 00:30")),
  dt = as.Date(c("1999-12-31", "2024-8-9"))
)

# Select date columns:
df$select(cs$date())

# Select all columns except for those that are dates:
df$select(!cs$date())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
