% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__duration}
\alias{cs__duration}
\title{Select all duration columns, optionally filtering by time unit}
\usage{
cs__duration(time_unit = c("ms", "us", "ns"))
}
\arguments{
\item{time_unit}{One (or more) of the allowed time unit precision strings,
\code{"ms"}, \code{"us"}, and \code{"ns"}. Default is to select columns with any valid
timeunit.}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all duration columns, optionally filtering by time unit
}
\examples{
\dontshow{if (\{
requireNamespace("clock", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
df <- pl$DataFrame(
  dtm = as.POSIXct(c("2001-5-7 10:25", "2031-12-31 00:30")),
  dur_ms = clock::duration_milliseconds(1:2),
  dur_us = clock::duration_microseconds(1:2),
  dur_ns = clock::duration_nanoseconds(1:2),
)

# Select duration columns:
df$select(cs$duration())

# Select all duration columns that have "ms" precision:
df$select(cs$duration("ms"))

# Select all duration columns that have "ms" OR "ns" precision:
df$select(cs$duration(c("ms", "ns")))

# Select all columns except for those that are duration:
df$select(!cs$duration())
\dontshow{\}) # examplesIf}
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
