% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__empty}
\alias{cs__empty}
\title{Select no columns}
\usage{
cs__empty()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
This is useful for composition with other selectors.
}
\examples{
pl$DataFrame(a = 1, b = 2)$select(cs$empty())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
