% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__ends_with}
\alias{cs__ends_with}
\title{Select columns that end with the given substring(s)}
\usage{
cs__ends_with(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Substring(s) that matching
column names should end with.}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select columns that end with the given substring(s)
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123, 456),
  baz = c(2.0, 5.5),
  zap = c(FALSE, TRUE)
)

# Select columns that end with the substring "z":
df$select(cs$ends_with("z"))

# Select columns that end with either the letter "z" or "r":
df$select(cs$ends_with("z", "r"))

# Select all columns except for those that end with the substring "z":
df$select(!cs$ends_with("z"))
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
