% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__integer}
\alias{cs__integer}
\title{Select all integer columns.}
\usage{
cs__integer()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all integer columns.
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123L, 456L),
  baz = c(2.0, 5.5),
  zap = 0:1
)

# Select all integer columns:
df$select(cs$integer())

# Select all columns except for those that are integer:
df$select(!cs$integer())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
