% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__list}
\alias{cs__list}
\title{Select all list columns}
\usage{
cs__list(inner = NULL)
}
\arguments{
\item{inner}{An optional inner \link[=cs]{selector} to select columns having
specific inner \link[=DataType]{data types}. If \code{NULL}, all inner types are selected.}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(
  foo = list(c("xx", "yy"), "x"),
  bar = list(c(123, 456), 789),
  baz = c(2.0, 5.5),
)

# Select all list columns:
df$select(cs$list())

# Select all columns except for those that are list:
df$select(!cs$list())

# If you want to select specific list columns,
# you can specify the inner data type with a selector:
df$select(cs$list(cs$string()))
}
\seealso{
\itemize{
\item \link{cs} for the documentation on operators supported by selectors.
\item \code{\link[=cs__by_dtype]{cs$by_dtype()}}: Select all columns matching the given dtype(s).
\item \code{\link[=cs__array]{cs$array()}}: Select all array columns.
\item \code{\link[=cs__nested]{cs$nested()}}: Select all nested columns.
}
}
