% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__numeric}
\alias{cs__numeric}
\title{Select all numeric columns.}
\usage{
cs__numeric()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all numeric columns.
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123L, 456L),
  baz = c(2.0, 5.5),
  zap = 0:1,
  .schema_overrides = list(bar = pl$Int16, baz = pl$Float32, zap = pl$UInt8),
)

# Select all numeric columns:
df$select(cs$numeric())

# Select all columns except for those that are numeric:
df$select(!cs$numeric())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
