% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__signed_integer}
\alias{cs__signed_integer}
\title{Select all signed integer columns}
\usage{
cs__signed_integer()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all signed integer columns
}
\examples{
df <- pl$DataFrame(
  foo = c(-123L, -456L),
  bar = c(3456L, 6789L),
  baz = c(7654L, 4321L),
  zap = c("ab", "cd"),
  .schema_overrides = list(bar = pl$UInt32, baz = pl$UInt64),
)

# Select signed integer columns:
df$select(cs$signed_integer())

# Select all columns except for those that are signed integer:
df$select(!cs$signed_integer())

# Select all integer columns (both signed and unsigned):
df$select(cs$integer())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
