% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__starts_with}
\alias{cs__starts_with}
\title{Select columns that start with the given substring(s)}
\usage{
cs__starts_with(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Substring(s) that matching
column names should end with.}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select columns that start with the given substring(s)
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123, 456),
  baz = c(2.0, 5.5),
  zap = c(FALSE, TRUE)
)

# Select columns that start with the substring "b":
df$select(cs$starts_with("b"))

# Select columns that start with either the letter "b" or "z":
df$select(cs$starts_with("b", "z"))

# Select all columns except for those that start with the substring "b":
df$select(!cs$starts_with("b"))
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
