% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__temporal}
\alias{cs__temporal}
\title{Select all temporal columns}
\usage{
cs__temporal()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all temporal columns
}
\examples{
df <- pl$DataFrame(
  dtm = as.POSIXct(c("2001-5-7 10:25", "2031-12-31 00:30")),
  dt = as.Date(c("1999-12-31", "2024-8-9")),
  value = 1:2
)

# Match all temporal columns:
df$select(cs$temporal())

# Match all temporal columns except for time columns:
df$select(cs$temporal() - cs$datetime())

# Match all columns except for temporal columns:
df$select(!cs$temporal())
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
