% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{DataFrame-class}
\alias{DataFrame-class}
\alias{DataFrame}
\title{General class of Polars DataFrame}
\format{
An object of class \code{environment} of length 37.
}
\usage{
DataFrame
}
\description{
General class of Polars DataFrame
}
\keyword{datasets}
