% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{DataFrame-shape}
\alias{DataFrame-shape}
\title{Shape of  DataFrame}
\usage{
<DataFrame>$shape()
}
\value{
Numeric vector of length two: number of rows, number of columns
}
\description{
Get shape/dimensions of DataFrame
}
\examples{
df = pl$DataFrame(iris)$shape
}
