% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__clear}
\alias{dataframe__clear}
\title{Create an empty or \code{n}-row null-filled copy of the frame}
\usage{
dataframe__clear(n = 0)
}
\arguments{
\item{n}{Number of (null-filled) rows to return in the cleared frame.}
}
\value{
A polars \link{DataFrame}
}
\description{
Returns a \code{n}-row null-filled frame with an identical schema. \code{n} can be
greater than the current number of rows in the frame.
}
\examples{
df <- pl$DataFrame(
  a = c(NA, 2, 3, 4),
  b = c(0.5, NA, 2.5, 13),
  c = c(TRUE, TRUE, FALSE, NA)
)
df$clear()

df$clear(n = 2)
}
