% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__count}
\alias{dataframe__count}
\title{Return the number of non-null elements for each column}
\usage{
dataframe__count()
}
\value{
A polars \link{DataFrame}
}
\description{
Return the number of non-null elements for each column
}
\examples{
df <- pl$DataFrame(a = 1:4, b = c(1, 2, 1, NA), c = rep(NA, 4))
df$count()
}
