% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__describe}
\alias{dataframe__describe}
\title{Summary statistics for a DataFrame.}
\usage{
dataframe__describe(
  percentiles = c(0.25, 0.5, 0.75),
  ...,
  interpolation = c(
    "nearest",
    "higher",
    "lower",
    "midpoint",
    "linear",
    "equiprobable"
  )
)
}
\arguments{
\item{percentiles}{One or more percentiles to include in the summary
statistics. All values must be in the range \verb{[0; 1]}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{interpolation}{Interpolation method for computing quantiles. Must be
one of \code{"nearest"}, \code{"higher"}, \code{"lower"}, \code{"midpoint"}, or \code{"linear"}.}
}
\value{
A polars \link{DataFrame}
}
\description{
Summary statistics for a DataFrame.
}
\examples{
df <- pl$DataFrame(
  int = 1:3,
  float = c(0.5, NA, 2.5),
  string = c(letters[1:2], NA),
  date = c(as.Date("2024-01-20"), as.Date("2024-01-21"), NA),
  cat = factor(c(letters[1:2], NA)),
  bool = c(TRUE, FALSE, NA)
)
df

# Show default frame statistics:
df$describe()

# Customize which percentiles are displayed, applying linear interpolation:
df$describe(
  percentiles = c(0.1, 0.3, 0.5, 0.7, 0.9),
  interpolation = "linear"
)
}
