% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__drop}
\alias{dataframe__drop}
\title{Remove columns}
\usage{
dataframe__drop(..., strict = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names or \link[=polars_selector]{selectors}
that should be removed.}

\item{strict}{Validate that all column names exist in the current schema,
and throw an exception if any do not.}
}
\value{
A polars \link{DataFrame}
}
\description{
Remove columns
}
\examples{
# Drop columns by passing the name of those columns
df <- pl$DataFrame(
  foo = 1:3,
  bar = c(6, 7, 8),
  ham = c("a", "b", "c")
)
df$drop("ham")
df$drop("ham", "bar")

# Drop multiple columns by passing a selector
df$drop(cs$all())
}
