% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__equals}
\alias{dataframe__equals}
\title{Check whether the DataFrame is equal to another DataFrame}
\usage{
dataframe__equals(other, ..., null_equal = TRUE)
}
\arguments{
\item{other}{DataFrame to compare with.}

\item{...}{These dots are for future extensions and must be empty.}

\item{null_equal}{Consider null values as equal.}
}
\value{
A logical value
}
\description{
Check whether the DataFrame is equal to another DataFrame
}
\examples{
dat1 <- as_polars_df(iris)
dat2 <- as_polars_df(iris)
dat3 <- as_polars_df(mtcars)
dat1$equals(dat2)
dat1$equals(dat3)
}
