% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__explode}
\alias{dataframe__explode}
\title{Explode the frame to long format by exploding the given columns}
\usage{
dataframe__explode(..., empty_as_null = FALSE, keep_nulls = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names or \link[=polars_selector]{selectors}
defining them. The underlying columns being exploded must be of
the \code{List} or \code{Array} data type.}

\item{empty_as_null}{Indicates to explode an empty list/array into a \code{null}.
Defaults to \code{FALSE}.}

\item{keep_nulls}{Indicates to explode a \code{null} list/array into a \code{null}.}
}
\value{
A polars \link{DataFrame}
}
\description{
Explode the frame to long format by exploding the given columns
}
\examples{
df <- pl$DataFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)

df$explode("numbers")
}
