% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__head}
\alias{dataframe__head}
\title{Get the first \code{n} rows}
\usage{
dataframe__head(n = 5)
}
\arguments{
\item{n}{Number of rows to return. If a negative value is passed,
return all rows except the last \code{\link[=abs]{abs(n)}}.}
}
\value{
A polars \link{DataFrame}
}
\description{
Get the first \code{n} rows
}
\examples{
df <- pl$DataFrame(foo = 1:5, bar = 6:10, ham = letters[1:5])

df$head(3)

# Pass a negative value to get all rows except the last `abs(n)`.
df$head(-3)
}
