% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__is_empty}
\alias{dataframe__is_empty}
\title{Returns \code{TRUE} if the DataFrame contains no rows.}
\usage{
dataframe__is_empty()
}
\value{
A logical value
}
\description{
Returns \code{TRUE} if the DataFrame contains no rows.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, 3, 1),
  b = c("x", "y", "z", "x")
)
df$is_empty()
df$filter(pl$col("a") > 99)$is_empty()
}
