% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__max_horizontal}
\alias{dataframe__max_horizontal}
\title{Get the maximum value horizontally across columns.}
\usage{
dataframe__max_horizontal()
}
\value{
A \link[=Series]{polars Series}
}
\description{
Get the maximum value horizontally across columns.
}
\examples{
df <- pl$DataFrame(
  foo = c(1, 2, 3),
  bar = c(4.0, 5.0, 6.0),
)
df$max_horizontal()
}
