% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__mean}
\alias{dataframe__mean}
\title{Aggregate the columns in the DataFrame to their mean value}
\usage{
dataframe__mean()
}
\value{
A polars \link{DataFrame}
}
\description{
Aggregate the columns in the DataFrame to their mean value
}
\examples{
df <- pl$DataFrame(a = 1:4, b = c(1, 2, 1, 1))
df$mean()
}
