% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__median}
\alias{dataframe__median}
\title{Aggregate the columns in the DataFrame to their median value}
\usage{
dataframe__median()
}
\value{
A polars \link{DataFrame}
}
\description{
Aggregate the columns in the DataFrame to their median value
}
\examples{
df <- pl$DataFrame(a = 1:4, b = c(1, 2, 1, 1))
df$median()
}
