% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__merge_sorted}
\alias{dataframe__merge_sorted}
\title{Take two sorted DataFrames and merge them by the sorted key}
\usage{
dataframe__merge_sorted(other, key, ..., maintain_order = FALSE)
}
\arguments{
\item{other}{Other DataFrame that must be merged.}

\item{key}{Key that is sorted.}

\item{...}{These dots are for future extensions and must be empty.}

\item{maintain_order}{If \code{TRUE}, the output is guaranteed to have left-biased
ordering for equal keys: rows from the left frame appear before rows from
the right frame when their keys are equal.}
}
\value{
A polars \link{DataFrame}
}
\description{
The output of this operation will also be sorted. It is the callers
responsibility that the frames are sorted by that key, otherwise the output
will not make sense. The schemas of both DataFrames must be equal.
}
\examples{
df1 <- pl$DataFrame(
  name = c("steve", "elise", "bob"),
  age = c(42, 44, 18)
)$sort("age")

df2 <- pl$DataFrame(
  name = c("anna", "megan", "steve", "thomas"),
  age = c(21, 33, 42, 20)
)$sort("age")

df1$merge_sorted(df2, key = "age")
}
