% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__n_chunks}
\alias{dataframe__n_chunks}
\title{Get number of chunks used by the ChunkedArrays of this DataFrame}
\usage{
dataframe__n_chunks(strategy = c("first", "all"))
}
\arguments{
\item{strategy}{Return the number of chunks of the \code{"first"} column, or
\code{"all"} columns in this DataFrame.}
}
\value{
An integer vector.
}
\description{
Get number of chunks used by the ChunkedArrays of this DataFrame
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, 3, 4),
  b = c(0.5, 4, 10, 13),
  c = c(TRUE, TRUE, FALSE, TRUE)
)

df$n_chunks()
df$n_chunks(strategy = "all")
}
