% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__partition_by}
\alias{dataframe__partition_by}
\title{Group by the given columns and return the groups as separate dataframes}
\usage{
dataframe__partition_by(..., maintain_order = TRUE, include_key = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names or \link[=polars_selector]{selectors}
to group by. Must contain at least one column.}

\item{maintain_order}{Ensure that the order of the groups is consistent with
the input data. This is slower than a default partition by operation.}

\item{include_key}{Include the columns used to partition the DataFrame in
the output.}
}
\value{
A list of polars \link{DataFrame}s
}
\description{
Group by the given columns and return the groups as separate dataframes
}
\examples{
# Pass a single column name to partition by that column.
df <- pl$DataFrame(
  a = c("a", "b", "a", "b", "c"),
  b = c(1, 2, 1, 3, 3),
  c = c(5, 4, 3, 2, 1)
)
df$partition_by("a")

# Partition by multiple columns:
df$partition_by("a", "b")
}
