% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__reverse}
\alias{dataframe__reverse}
\title{Reverse the DataFrame}
\usage{
dataframe__reverse()
}
\value{
A polars \link{DataFrame}
}
\description{
Reverse the DataFrame
}
\examples{
df <- pl$DataFrame(key = c("a", "b", "c"), val = 1:3)

df$reverse()
}
