% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__shift}
\alias{dataframe__shift}
\title{Shift values by the given number of indices}
\usage{
dataframe__shift(n = 1, ..., fill_value = NULL)
}
\arguments{
\item{n}{Number of indices to shift forward. If a negative value is passed,
values are shifted in the opposite direction instead.}

\item{...}{These dots are for future extensions and must be empty.}

\item{fill_value}{Fill the resulting null values with this value. Accepts
expression input. Non-expression inputs are parsed as literals.}
}
\value{
A polars \link{DataFrame}
}
\description{
Shift values by the given number of indices
}
\examples{
df <- pl$DataFrame(a = 1:4, b = 5:8)

# By default, values are shifted forward by one index.
df$shift()

# Pass a negative value to shift in the opposite direction instead.
df$shift(-2)

# Specify fill_value to fill the resulting null values.
df$shift(-2, fill_value = 100)
}
