% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__slice}
\alias{dataframe__slice}
\title{Get a slice of the DataFrame.}
\usage{
dataframe__slice(offset, length = NULL)
}
\arguments{
\item{offset}{Start index, can be a negative value. This is 0-indexed, so
\code{offset = 1} skips the first row.}

\item{length}{Length of the slice. If \code{NULL} (default), all rows starting at
the offset will be selected.}
}
\value{
A polars \link{DataFrame}
}
\description{
Get a slice of the DataFrame.
}
\examples{
# skip the first 2 rows and take the 4 following rows
as_polars_df(mtcars)$slice(2, 4)

# this is equivalent to:
mtcars[3:6, ]
}
