% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__sort}
\alias{dataframe__sort}
\title{Sort a DataFrame by the given columns}
\usage{
dataframe__sort(
  ...,
  descending = FALSE,
  nulls_last = FALSE,
  multithreaded = TRUE,
  maintain_order = FALSE
)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column(s) to sort by. Can be
character values indicating column names or Expr(s).}

\item{descending}{Sort in descending order. When sorting by multiple
columns, this can be specified per column by passing a logical vector.}

\item{nulls_last}{Place null values last. When sorting by multiple
columns, this can be specified per column by passing a logical vector.}

\item{multithreaded}{Sort using multiple threads.}

\item{maintain_order}{Whether the order should be maintained if elements are
equal. If \code{TRUE}, streaming is not possible and performance might be worse
since this requires a stable search.}
}
\value{
A polars \link{DataFrame}
}
\description{
Sort a DataFrame by the given columns
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, NA, 4),
  b = c(6, 5, 4, 3),
  c = c("a", "c", "b", "a")
)

# Pass a single column name to sort by that column.
df$sort("a")

# Sorting by expressions is also supported
df$sort(pl$col("a") + pl$col("b") * 2, nulls_last = TRUE)

# Sort by multiple columns by passing a vector of columns
df$sort(c("c", "a"), descending = TRUE)

# Or use positional arguments to sort by multiple columns in the same way
df$sort("c", "a", descending = c(FALSE, TRUE))
}
