% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__std}
\alias{dataframe__std}
\title{Aggregate the columns of this DataFrame to their standard deviation values}
\usage{
dataframe__std(ddof = 1)
}
\arguments{
\item{ddof}{"Delta Degrees of Freedom": the divisor used in the calculation
is \code{N - ddof}, where \code{N} represents the number of elements. By default
\code{ddof} is 1.}
}
\value{
A polars \link{DataFrame}
}
\description{
Aggregate the columns of this DataFrame to their standard deviation values
}
\examples{
df <- pl$DataFrame(a = 1:4, b = c(1, 2, 1, 1))
df$std()
df$std(ddof = 0)
}
