% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__sum_horizontal}
\alias{dataframe__sum_horizontal}
\title{Sum all values horizontally across columns.}
\usage{
dataframe__sum_horizontal(..., ignore_nulls = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{ignore_nulls}{Ignore null values (default). If \code{FALSE}, any null value
in the input will lead to a null output.}
}
\value{
A \link[=Series]{polars Series}
}
\description{
Sum all values horizontally across columns.
}
\examples{
df <- pl$DataFrame(
  foo = c(1, 2, 3),
  bar = c(4.0, 5.0, 6.0),
)
df$sum_horizontal()
}
