% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__unique}
\alias{dataframe__unique}
\title{Drop duplicate rows}
\usage{
dataframe__unique(
  ...,
  keep = c("any", "none", "first", "last"),
  maintain_order = FALSE
)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names or \link[=polars_selector]{selectors}
for which are considered. If empty (default), use all columns
(same as specifying with the selector \code{\link[=cs__all]{cs$all()}}).}

\item{keep}{Which of the duplicate rows to keep. Must be one of:
\itemize{
\item \code{"any"}: does not give any guarantee of which row is kept. This allows
more optimizations.
\item \code{"none"}: don’t keep duplicate rows.
\item \code{"first"}: keep first unique row.
\item \code{"last"}: keep last unique row.
}}

\item{maintain_order}{Keep the same order as the original data. This is
more expensive to compute. Setting this to \code{TRUE} blocks the possibility to
run on the streaming engine.}
}
\value{
A polars \link{DataFrame}
}
\description{
Drop duplicate rows
}
\examples{
df <- pl$DataFrame(
  foo = c(1, 2, 3, 1),
  bar = c("a", "a", "a", "a"),
  ham = c("b", "b", "b", "b"),
)
df$unique(maintain_order = TRUE)

df$unique(c("bar", "ham"), maintain_order = TRUE)

df$unique(keep = "last", maintain_order = TRUE)
}
