% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__unpivot}
\alias{dataframe__unpivot}
\title{Unpivot a frame from wide to long format}
\usage{
dataframe__unpivot(
  on = NULL,
  ...,
  index = NULL,
  variable_name = NULL,
  value_name = NULL
)
}
\arguments{
\item{on}{Column(s) or selector(s) to use as values variables.
If \code{on} is empty (\code{\link[=cs__empty]{cs$empty()}}), no columns will be used.
If set to \code{NULL} (default), all columns that are not in \code{index} will be used.}

\item{...}{These dots are for future extensions and must be empty.}

\item{index}{Column(s) or selector(s) to use as identifier variables.}

\item{variable_name}{Name to give to the new column containing the names of
the melted columns. Defaults to \code{"variable"}.}

\item{value_name}{Name to give to the new column containing the values of
the melted columns. Defaults to \code{"value"}.

The row order of the result is not guaranteed; use \verb{$sort()} when a stable
order is required.}
}
\value{
A polars \link{LazyFrame}
}
\description{
This function is useful to massage a frame into a format where one or
more columns are identifier variables (\code{index}) while all other columns,
considered measured variables (\code{on}), are "unpivoted" to the row axis
leaving just two non-identifier columns, "variable" and "value".
}
\examples{
df <- pl$DataFrame(
  a = c("x", "y", "z"),
  b = c(1, 3, 5),
  c = c(2, 4, 6)
)
df$unpivot(index = "a", on = c("b", "c"))
}
