% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/output-json.R
\name{dataframe__write_json}
\alias{dataframe__write_json}
\title{Serialize to JSON representation}
\usage{
dataframe__write_json(file)
}
\arguments{
\item{file}{File path to which the result will be written.}
}
\value{
\code{NULL} invisibly.
}
\description{
Serialize to JSON representation
}
\examples{
\dontshow{if (\{
requireNamespace("jsonlite", quiet = TRUE)
\}) withAutoprint(\{ # examplesIf}
dat <- as_polars_df(head(mtcars))
destination <- tempfile()

dat$select(pl$col(c("drat", "mpg")))$write_json(destination)
jsonlite::fromJSON(destination)
\dontshow{\}) # examplesIf}
}
