% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_clone}
\alias{DataFrame_clone}
\alias{clone}
\title{Clone a DataFrame}
\usage{
DataFrame_clone()
}
\value{
DataFrame
}
\description{
Rarely useful as DataFrame is nearly 100\% immutable
Any modification of a DataFrame would lead to a clone anyways.
}
\examples{
df1 = pl$DataFrame(iris);
df2 =  df1$clone();
df3 = df1
xptr::xptr_address(df1) != xptr::xptr_address(df2)
xptr::xptr_address(df1) == xptr::xptr_address(df3)

}
\keyword{DataFrame}
