% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{DataFrame_constructor}
\alias{DataFrame_constructor}
\title{create new DataFrame}
\usage{
DataFrame_constructor(data)
}
\arguments{
\item{data}{a data.frame or list of mixed vectors and Series of equal length.}
}
\value{
DataFrame
}
\description{
create new DataFrame
}
\examples{
minipolars:::new_pf(iris)
#with namespace
pl$DataFrame(iris)
pl$DataFrame(list(some_column_name = c(1,2,3,4,5)))
}
