% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_drop}
\alias{DataFrame_drop}
\title{Drop columns of a DataFrame}
\usage{
DataFrame_drop(..., strict = TRUE)
}
\arguments{
\item{...}{Characters of column names to drop. Passed to \code{\link[=pl_col]{pl$col()}}.}

\item{strict}{Validate that all column names exist in the schema and throw an
exception if a column name does not exist in the schema.}
}
\value{
DataFrame
}
\description{
Drop columns of a DataFrame
}
\examples{
as_polars_df(mtcars)$drop(c("mpg", "hp"))

# equivalent
as_polars_df(mtcars)$drop("mpg", "hp")
}
