% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_drop_nulls}
\alias{DataFrame_drop_nulls}
\title{Drop nulls (missing values)}
\usage{
<DataFrame>$drop_nulls(subset = NULL)
}
\arguments{
\item{subset}{A character vector with the names of the column(s) for which
nulls are considered. If \code{NULL} (default), use all columns.}
}
\value{
DataFrame
}
\description{
Drop all rows that contain nulls (which correspond to \code{NA} in R).
}
\examples{
tmp = mtcars
tmp[1:3, "mpg"] = NA
tmp[4, "hp"] = NA
tmp = pl$DataFrame(tmp)

# number of rows in `tmp` before dropping nulls
tmp$height

tmp$drop_nulls()$height
tmp$drop_nulls("mpg")$height
tmp$drop_nulls(c("mpg", "hp"))$height
}
\keyword{DataFrame}
