% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_estimated_size}
\alias{DataFrame_estimated_size}
\title{Estimated size}
\format{
function
}
\usage{
DataFrame_estimated_size
}
\value{
Estimated size in bytes
}
\description{
Return an estimation of the total (heap) allocated size of the
DataFrame.
}
\examples{
pl$DataFrame(mtcars)$estimated_size()
}
\keyword{DataFrame}
