% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_explode}
\alias{DataFrame_explode}
\title{Explode the DataFrame to long format by exploding the given columns}
\usage{
DataFrame_explode(columns, ...)
}
\arguments{
\item{columns}{Column(s) to be exploded. \verb{Into<Expr>}, list of \verb{Into<Expr>} or a char vec.
Only columns of DataType \code{List} or \code{Utf8} can be exploded.}

\item{...}{More columns to explode as above but provided as separate arguments}
}
\value{
DataFrame
}
\description{
Explode the DataFrame to long format by exploding the given columns
}
\examples{
df = pl$DataFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)
df

df$explode("numbers")
}
\keyword{DataFrame}
