% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_export_stream}
\alias{DataFrame_export_stream}
\title{Stream a DataFrame using nanoarrow}
\usage{
DataFrame_export_stream()
}
\value{
A \link[nanoarrow:as_nanoarrow_array_stream]{nanoarrow_array_stream}
}
\description{
Stream a DataFrame using nanoarrow
}
