% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_fetch}
\alias{DataFrame_fetch}
\title{Fetch limited number of rows of DataFrame}
\usage{
DataFrame_fetch(n_rows)
}
\arguments{
\item{n_rows}{number of rows to fetch at maximum.}
}
\value{
A DataFrame of maximum n_rows
}
\description{
Get rows
}
\examples{
pl$DataFrame(iris)$fetch(2)
}
\keyword{DataFrame}
