% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_fill_nan}
\alias{DataFrame_fill_nan}
\title{Fill floating point NaN value with a fill value}
\usage{
DataFrame_fill_nan(value)
}
\arguments{
\item{value}{Value used to fill \code{NaN} values.}
}
\value{
DataFrame
}
\description{
Fill floating point NaN value with a fill value
}
\examples{
df = pl$DataFrame(
  a = c(1.5, 2, NaN, 4),
  b = c(1.5, NaN, NaN, 4)
)
df$fill_nan(99)
}
