% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_fill_null}
\alias{DataFrame_fill_null}
\title{Fill null}
\usage{
DataFrame_fill_null(fill_value)
}
\arguments{
\item{fill_value}{Value to fill \code{NA} with.}
}
\value{
DataFrame
}
\description{
Fill null values using the specified value or strategy.
}
\examples{
pl$DataFrame(
  a = c(1.5, 2, NA, 4),
  b = c(1.5, NA, NA, 4)
)$fill_null(99)
}
\keyword{DataFrame}
