% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_filter}
\alias{DataFrame_filter}
\title{Filter rows based on conditions}
\usage{
DataFrame_filter(expr)
}
\arguments{
\item{expr}{one Expr or string naming a column}
}
\value{
\code{DataFrame} with rows matching the conditions
}
\description{
Filter rows based on conditions
}
\examples{
pl$DataFrame(iris)$filter(pl$col("Sepal.Length") > 5)
}
\keyword{DataFrame}
