% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_first}
\alias{DataFrame_first}
\title{Get the first row of the DataFrame.}
\usage{
<DataFrame>$first()
}
\value{
A DataFrame with one row.
}
\description{
Get the first row of the DataFrame.
}
\examples{
pl$DataFrame(mtcars)$first()
}
\keyword{DataFrame}
