% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_flags}
\alias{DataFrame_flags}
\title{Get flags for each column}
\usage{
DataFrame_flags()
}
\value{
A nested list with column names at the top level and column flags
in each sublist.
}
\description{
Flags are used internally to avoid doing unnecessary computations, such as
sorting a variable that we know is already sorted. The number of flags varies
depending on the column type: columns of type \code{array} and \code{list} have the
flags \code{SORTED_ASC}, \code{SORTED_DESC}, and \code{FAST_EXPLODE}, while other column
types only have the former two.

\code{SORTED_ASC} is set to \code{TRUE} when we sort a column in increasing order, so
that we can use this information later on to avoid re-sorting it. \code{SORTED_DESC}
is similar but applies to sort in decreasing order.
}
\examples{
pl$DataFrame(a = c(2, 1), b = c(3, 4), c = list(c(1, 2), 4))$sort("a")$flags
}
