% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_frame_equal}
\alias{DataFrame_frame_equal}
\title{Drop in place}
\usage{
DataFrame_frame_equal(other)
}
\arguments{
\item{other}{DataFrame to compare with.}
}
\value{
bool
}
\description{
Check if DataFrame is equal to other.
}
\examples{
dat1 = pl$DataFrame(iris)
dat2 = pl$DataFrame(iris)
dat3 = pl$DataFrame(mtcars)
dat1$frame_equal(dat2)
dat1$frame_equal(dat3)
}
\keyword{DataFrame}
