% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_gather_every}
\alias{DataFrame_gather_every}
\title{Take every nth row in the DataFrame}
\usage{
DataFrame_gather_every(n, offset = 0)
}
\arguments{
\item{n}{Gather every \code{n}-th row.}

\item{offset}{Starting index.}
}
\value{
A DataFrame
}
\description{
Take every nth row in the DataFrame
}
\examples{
df = pl$DataFrame(a = 1:4, b = 5:8)
df$gather_every(2)

df$gather_every(2, offset = 1)
}
