% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_head}
\alias{DataFrame_head}
\alias{DataFrame_limit}
\title{Get the first \code{n} rows.}
\usage{
DataFrame_head(n = 5L)

DataFrame_limit(n = 5L)
}
\arguments{
\item{n}{Number of rows to return. If a negative value is passed,
return all rows except the last \code{\link[=abs]{abs(n)}}.}
}
\value{
A \link[=DataFrame_class]{DataFrame}
}
\description{
Get the first \code{n} rows.
}
\details{
\verb{$limit()} is an alias for \verb{$head()}.
}
\examples{
df = pl$DataFrame(foo = 1:5, bar = 6:10, ham = letters[1:5])

df$head(3)

# Pass a negative value to get all rows except the last `abs(n)`.
df$head(-3)
}
