% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_height}
\alias{DataFrame_height}
\alias{height}
\alias{nrow}
\title{Number of rows of a DataFrame}
\usage{
DataFrame_height()
}
\value{
The number of rows of the DataFrame
}
\description{
Get the number of rows (height) of a DataFrame
}
\examples{
pl$DataFrame(iris)$height
}
\keyword{DataFrame}
