% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_item}
\alias{DataFrame_item}
\title{Return the element at the given row/column.}
\usage{
DataFrame_item(row = NULL, column = NULL)
}
\arguments{
\item{row}{Optional row index (0-indexed).}

\item{column}{Optional column index (0-indexed) or name.}
}
\value{
A value of length 1
}
\description{
If row and column location are not specified, the \link[=DataFrame_class]{DataFrame}
must have dimensions (1, 1).
}
\examples{
df = pl$DataFrame(a = c(1, 2, 3), b = c(4, 5, 6))

df$select((pl$col("a") * pl$col("b"))$sum())$item()

df$item(1, 1)

df$item(2, "b")
}
