% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_last}
\alias{DataFrame_last}
\title{Last}
\usage{
DataFrame_last()
}
\value{
A new \code{DataFrame} object with applied filter.
}
\description{
Get the last row of the DataFrame.
}
\examples{
pl$DataFrame(mtcars)$last()
}
\keyword{DataFrame}
