% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_lazy}
\alias{DataFrame_lazy}
\alias{lazy}
\title{Convert an existing DataFrame to a PolarsLazyFrame}
\usage{
DataFrame_lazy
}
\value{
A PolarsLazyFrame
}
\description{
Start a new lazy query from a DataFrame.
}
\examples{
pl$DataFrame(iris)$lazy()

}
\keyword{DataFrame}
\keyword{PolarsLazyFrame_new}
