% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_limit}
\alias{DataFrame_limit}
\alias{limit}
\title{Limit a DataFrame}
\usage{
DataFrame_limit(n)
}
\arguments{
\item{n}{positive numeric or integer number not larger than 2^32}
}
\value{
DataFrame
}
\description{
take limit of n rows of query
}
\details{
any number will converted to u32. Negative raises error
}
\keyword{DataFrame}
