% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_max}
\alias{DataFrame_max}
\title{Max}
\usage{
DataFrame_max()
}
\value{
A DataFrame with one row.
}
\description{
Aggregate the columns in the DataFrame to their maximum value.
}
\examples{
pl$DataFrame(mtcars)$max()
}
\keyword{DataFrame}
