% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_median}
\alias{DataFrame_median}
\title{Median}
\usage{
<DataFrame>$median()
}
\value{
A DataFrame with one row.
}
\description{
Aggregate the columns in the DataFrame to their median value.
}
\examples{
pl$DataFrame(mtcars)$median()
}
\keyword{DataFrame}
