% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_min}
\alias{DataFrame_min}
\title{Min}
\usage{
DataFrame_min()
}
\value{
A new \code{DataFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their minimum value.
}
\examples{
pl$DataFrame(mtcars)$min()
}
\keyword{DataFrame}
