% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_null_count}
\alias{DataFrame_null_count}
\title{Null count}
\format{
function
}
\usage{
DataFrame_null_count
}
\value{
DataFrame
}
\description{
Create a new DataFrame that shows the null counts per column.
}
\examples{
x = mtcars
x[1, 2:3] = NA
pl$DataFrame(x)$null_count()
}
\keyword{DataFrame}
