% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_pivot}
\alias{DataFrame_pivot}
\title{Pivot data from long to wide}
\usage{
<DataFrame>$pivot(
  values,
  index,
  columns,
  aggregate_function = NULL,
  maintain_order = TRUE,
  sort_columns = FALSE,
  separator = "_"
)
}
\arguments{
\item{values}{Column values to aggregate. Can be multiple columns if the
\code{columns} arguments contains multiple columns as well.}

\item{index}{One or multiple keys to group by.}

\item{columns}{Name of the column(s) whose values will be used as the header
of the output DataFrame.}

\item{aggregate_function}{One of:
\itemize{
\item string indicating the expressions to aggregate with, such as 'first',
'sum', 'max', 'min', 'mean', 'median', 'last', 'count'),
\item an Expr e.g. \code{pl$element()$sum()}
}}

\item{maintain_order}{Keep the same order as the original \code{DataFrame}. Setting
this to \code{TRUE} makes it more expensive to compute and blocks the possibility
to run on the streaming engine. The default value can be changed with
\code{pl$set_options(maintain_order = TRUE)}.}

\item{sort_columns}{Sort the transposed columns by name. Default is by order
of discovery.}

\item{separator}{Used as separator/delimiter in generated column names.}
}
\value{
DataFrame
}
\description{
Pivot data from long to wide
}
\examples{
df = pl$DataFrame(
  foo = c("one", "one", "one", "two", "two", "two"),
  bar = c("A", "B", "C", "A", "B", "C"),
  baz = c(1, 2, 3, 4, 5, 6)
)
df

df$pivot(
  values = "baz", index = "foo", columns = "bar"
)

# Run an expression as aggregation function
df = pl$DataFrame(
  col1 = c("a", "a", "a", "b", "b", "b"),
  col2 = c("x", "x", "x", "x", "y", "y"),
  col3 = c(6, 7, 3, 2, 5, 7)
)
df

df$pivot(
  index = "col1",
  columns = "col2",
  values = "col3",
  aggregate_function = pl$element()$tanh()$mean()
)
}
\keyword{DataFrame}
