% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_print}
\alias{DataFrame_print}
\title{internal method print DataFrame}
\usage{
DataFrame_print()
}
\value{
self
}
\description{
internal method print DataFrame
}
\examples{
pl$DataFrame(iris)
}
