% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_quantile}
\alias{DataFrame_quantile}
\title{Quantile}
\usage{
DataFrame_quantile(quantile, interpolation = "nearest")
}
\arguments{
\item{quantile}{numeric Quantile between 0.0 and 1.0.}

\item{interpolation}{string Interpolation method: "nearest", "higher", "lower", "midpoint", or "linear".}
}
\value{
DataFrame
}
\description{
Aggregate the columns in the DataFrame to their quantile value.
}
\examples{
pl$DataFrame(mtcars)$quantile(.4)
}
\keyword{DataFrame}
