% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_rename}
\alias{DataFrame_rename}
\title{Rename column names of a DataFrame}
\usage{
DataFrame_rename(...)
}
\arguments{
\item{...}{One of the following:
\itemize{
\item params like \code{new_name = "old_name"} to rename selected variables.
\item as above but with params wrapped in a list
}}
}
\value{
\link[=DataFrame_class]{DataFrame}
}
\description{
Rename column names of a DataFrame
}
\details{
If existing names are swapped (e.g. \code{A} points to \code{B} and \code{B} points to \code{A}),
polars will block projection and predicate pushdowns at this node.
}
\examples{
df = pl$DataFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)

df$rename(apple = "foo")
}
\seealso{
\itemize{
\item \code{\link[=DataFrame_rename_with]{<DataFrame>$rename_with()}}
}
}
