% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_rename_with}
\alias{DataFrame_rename_with}
\title{Rename column names of a DataFrame with a function}
\usage{
DataFrame_rename_with(fun)
}
\arguments{
\item{fun}{An R function that takes the old names character vector as input and
returns the new names character vector.}
}
\value{
\link[=DataFrame_class]{DataFrame}
}
\description{
This method is currently experimental and may
change without it being considered a breaking change.
}
\details{
If existing names are swapped (e.g. \code{A} points to \code{B} and \code{B} points to \code{A}),
polars will block projection and predicate pushdowns at this node.
}
\examples{
df = pl$DataFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)

df$rename_with(
  \(column_name) paste0("c", substr(column_name, 2, 100))
)
}
\seealso{
\itemize{
\item \code{\link[=DataFrame_rename]{<DataFrame>$rename()}}
}
}
