% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_reverse}
\alias{DataFrame_reverse}
\title{Reverse}
\usage{
<DataFrame>$reverse()
}
\value{
DataFrame
}
\description{
Reverse the DataFrame (the last row becomes the first one, etc.).
}
\examples{
pl$DataFrame(mtcars)$reverse()
}
