% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_sample}
\alias{DataFrame_sample}
\title{Take a sample of rows from a DataFrame}
\usage{
DataFrame_sample(
  n = NULL,
  fraction = NULL,
  with_replacement = FALSE,
  shuffle = FALSE,
  seed = NULL
)
}
\arguments{
\item{n}{Number of rows to return. Cannot be used with \code{fraction}.}

\item{fraction}{Fraction of rows to return (between 0 and 1). Cannot be used
with \code{n}.}

\item{with_replacement}{Allow values to be sampled more than once.}

\item{shuffle}{If \code{TRUE}, the order of the sampled rows will be shuffled. If
\code{FALSE} (default), the order of the returned rows will be neither stable nor
fully random.}

\item{seed}{Seed for the random number generator. If set to \code{NULL} (default),
a random seed is generated for each sample operation.}
}
\value{
DataFrame
}
\description{
Take a sample of rows from a DataFrame
}
\examples{
df = pl$DataFrame(iris)
df$sample(n = 20)
df$sample(frac = 0.1)
}
\keyword{DataFrame}
