% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_select}
\alias{DataFrame_select}
\alias{select}
\title{perform select on DataFrame}
\usage{
DataFrame_select(...)
}
\arguments{
\item{...}{expresssions or strings defining columns to select(keep) in context the DataFrame}
}
\description{
related to dplyr \code{mutate()} However discards unmentioned columns as data.table \code{.()}.
}
\keyword{#'}
\keyword{(pl$col("Sepal.Length")+2)$alias("add_2_SL")}
\keyword{)}
\keyword{DataFrame}
\keyword{pl$DataFrame(iris)$select(}
\keyword{pl$col("Sepal.Length")$abs()$alias("abs_SL"),}
