% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_shape}
\alias{DataFrame_shape}
\title{Dimensions of a DataFrame}
\usage{
<DataFrame>$shape()
}
\value{
Numeric vector of length two with the number of rows and the number
of columns.
}
\description{
Get shape/dimensions of DataFrame
}
\examples{
pl$DataFrame(iris)$shape
}
\keyword{DataFrame}
