% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_shift}
\alias{DataFrame_shift}
\title{Shift a DataFrame}
\usage{
DataFrame_shift(n = 1, fill_value = NULL)
}
\arguments{
\item{n}{Number of indices to shift forward. If a negative value is
passed, values are shifted in the opposite direction instead.}

\item{fill_value}{Fill the resulting null values with this value. Accepts
expression input. Non-expression inputs are parsed as literals.}
}
\value{
DataFrame
}
\description{
Shift the values by a given period. If the period (\code{n}) is positive,
then \code{n} rows will be inserted at the top of the DataFrame and the last \code{n}
rows will be discarded. Vice-versa if the period is negative. In the end,
the total number of rows of the DataFrame doesn't change.
}
\examples{
df = pl$DataFrame(a = 1:4, b = 5:8)

df$shift(2)

df$shift(-2)

df$shift(-2, fill_value = 100)
}
