% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_shift_and_fill}
\alias{DataFrame_shift_and_fill}
\title{Shift and fill}
\usage{
<DataFrame>$shift_and_fill(fill_value, periods = 1)
}
\arguments{
\item{fill_value}{Fill new \code{NULL} values with this value. Must of length 1.
A logical value will be converted to numeric.}

\item{periods}{Number of periods to shift (can be negative).}
}
\value{
DataFrame
}
\description{
Shift the values by a given period and fill the resulting null
values. See the docs of \verb{$shift()} for more details on shifting.
}
\examples{
df = pl$DataFrame(mtcars)

# insert two rows filled with 0 at the top of the DataFrame
df$shift_and_fill(0, 2)

# automatic conversion of logical value to numeric
df$shift_and_fill(TRUE, 2)
}
\keyword{DataFrame}
