% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_slice}
\alias{DataFrame_slice}
\title{Slice}
\usage{
<DataFrame>$slice(offset, length = NULL)
}
\arguments{
\item{offset}{Start index, can be a negative value. This is 0-indexed, so
\code{offset = 1} doesn't include the first row.}

\item{length}{Length of the slice. If \code{NULL} (default), all rows starting at
the offset will be selected.}
}
\value{
DataFrame
}
\description{
Get a slice of the DataFrame.
}
\examples{
# skip the first 2 rows and take the 4 following rows
pl$DataFrame(mtcars)$slice(2, 4)

# this is equivalent to:
mtcars[3:6, ]
}
