% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_sql}
\alias{DataFrame_sql}
\title{Execute a SQL query against the DataFrame}
\usage{
DataFrame_sql(query, ..., table_name = NULL, envir = parent.frame())
}
\arguments{
\item{query}{A character of the SQL query to execute.}

\item{...}{Ignored.}

\item{table_name}{\code{NULL} (default) or a character of an explicit name for the table
that represents the calling frame (the alias \code{"self"} will always be registered/available).}

\item{envir}{The environment to search for polars
\link[=DataFrame_class]{DataFrames}/\link[=LazyFrame_class]{LazyFrames}.}
}
\value{
\link[=DataFrame_class]{DataFrame}
}
\description{
The calling frame is automatically registered as a table in the SQL context
under the name \code{"self"}. All \link[=DataFrame_class]{DataFrames} and
\link[=LazyFrame_class]{LazyFrames} found in the \code{envir} are also registered,
using their variable name.
More control over registration and execution behaviour is available by
the \link[=SQLContext_class]{SQLContext} object.
}
\details{
This functionality is considered \strong{unstable}, although it is close to
being considered stable. It may be changed at any point without it being
considered a breaking change.
}
\examples{
\dontshow{if (polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df1 = pl$DataFrame(
  a = 1:3,
  b = c("zz", "yy", "xx"),
  c = as.Date(c("1999-12-31", "2010-10-10", "2077-08-08"))
)

# Query the DataFrame using SQL:
df1$sql("SELECT c, b FROM self WHERE a > 1")

# Join two DataFrames using SQL.
df2 = pl$DataFrame(a = 3:1, d = c(125, -654, 888))
df1$sql(
  "
SELECT self.*, d
FROM self
INNER JOIN df2 USING (a)
WHERE a > 1 AND EXTRACT(year FROM c) < 2050
"
)

# Apply transformations to a DataFrame using SQL, aliasing "self" to "frame".
df1$sql(
  query = r"(
SELECT
a,
(a \% 2 == 0) AS a_is_even,
CONCAT_WS(':', b, b) AS b_b,
EXTRACT(year FROM c) AS year,
0::float AS 'zero'
FROM frame
)",
  table_name = "frame"
)
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \link[=SQLContext_class]{SQLContext}
}
}
