% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_sum}
\alias{DataFrame_sum}
\title{Sum}
\usage{
DataFrame_sum()
}
\value{
A DataFrame with one row.
}
\description{
Aggregate the columns of this DataFrame to their sum values.
}
\examples{
as_polars_df(mtcars)$sum()
}
\keyword{DataFrame}
