% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_dummies}
\alias{DataFrame_to_dummies}
\title{Convert variables into dummy/indicator variables}
\usage{
DataFrame_to_dummies(columns = NULL, ..., separator = "_", drop_first = FALSE)
}
\arguments{
\item{columns}{Column name(s) or selector(s) that should be converted to
dummy variables. If \code{NULL} (default), convert all columns.}

\item{...}{Ignored.}

\item{separator}{Separator/delimiter used when generating column names.}

\item{drop_first}{Remove the first category from the variables being encoded.}
}
\value{
A DataFrame
}
\description{
Convert variables into dummy/indicator variables
}
\examples{
df = pl$DataFrame(foo = 1:2, bar = 3:4, ham = c("a", "b"))

df$to_dummies()

df$to_dummies(drop_first = TRUE)

df$to_dummies(c("foo", "bar"), separator = "::")
}
