% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R, R/lazyframe__lazy.R
\name{DataFrame_to_series}
\alias{DataFrame_to_series}
\alias{DataFrame_sort}
\alias{LazyFrame_sort}
\title{Get Series by idx, if there}
\usage{
DataFrame_to_series(idx = 0)

DataFrame_sort(by, ..., descending = FALSE, nulls_last = FALSE)

LazyFrame_sort(by, ..., descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{idx}{numeric default 0, zero-index of what column to return as Series}

\item{by}{Column(s) to sort by. Iterable Into\if{html}{\out{<Expr>}}, e.g. one Expr or list of two, or plain
charvec naming columns.}

\item{...}{more columns to sort by. Into\if{html}{\out{<Expr>}} as above but not e.g. in a list, can be combined
with above}

\item{descending}{Sort desceding? logical vector of len 1 or same as number of Expr's above}

\item{nulls_last}{place all nulls_last? Bool}
}
\value{
Series or NULL

Series or NULL

Series or NULL
}
\description{
get one column by idx as series from DataFrame.
Unlike get_column this method will not fail if no series found at idx but
return a NULL, idx is zero idx.

get one column by idx as series from DataFrame.
Unlike get_column this method will not fail if no series found at idx but
return a NULL, idx is zero idx.

sort a LazyFrame by on or more Expr

get one column by idx as series from DataFrame.
Unlike get_column this method will not fail if no series found at idx but
return a NULL, idx is zero idx.
}
\examples{
pl$DataFrame(a=1:4)$to_series()
}
\keyword{DataFrame}
