% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_struct}
\alias{DataFrame_to_struct}
\alias{to_struct}
\title{Convert DataFrame to a Series of type "struct"}
\usage{
<DataFrame>$to_struct(name = "")
}
\arguments{
\item{name}{Name given to the new Series}
}
\value{
A Series of type "struct"
}
\description{
Convert DataFrame to a Series of type "struct"
}
\examples{
# round-trip conversion from DataFrame with two columns
df = pl$DataFrame(a = 1:5, b = c("one", "two", "three", "four", "five"))
s = df$to_struct()
s

# convert to an R list
s$to_r()

# Convert back to a DataFrame
df_s = s$to_frame()
df_s
}
\keyword{DataFrame}
