% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_Struct_unnest}
\alias{DataFrame_to_Struct_unnest}
\alias{DataFrame_to_struct}
\alias{to_struct}
\alias{DataFrame_unnest}
\title{to_struct and unnest again}
\usage{
DataFrame_to_struct(name = "")

DataFrame_unnest(names = NULL)
}
\arguments{
\item{name}{name of new Series}

\item{names}{names of struct columns to unnest, default NULL unnest any struct column}
}
\value{
@to_struct() returns a Series

$unnest() returns a DataFrame with all column including any that has been unnested
}
\description{
to_struct and unnest again

Unnest a DataFrame struct columns.
}
\examples{
#round-trip conversion from DataFrame with two columns
df = pl$DataFrame(a=1:5,b=c("one","two","three","four","five"))
s = df$to_struct()
s
s$to_r() # to r list
df_s = s$to_frame() #place series in a new DataFrame
df_s$unnest() # back to starting df
}
\keyword{DataFrame}
