% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_unnest}
\alias{DataFrame_unnest}
\title{Unnest a DataFrame struct columns.}
\usage{
DataFrame_unnest(names = NULL)
}
\arguments{
\item{names}{names of struct columns to unnest, default NULL unnest any struct column}
}
\value{
$unnest() returns a DataFrame with all column including any that has been unnested
}
\description{
Unnest a DataFrame struct columns.
}
\keyword{DataFrame}
