% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_unpivot}
\alias{DataFrame_unpivot}
\title{Unpivot a Frame from wide to long format}
\usage{
DataFrame_unpivot(
  on = NULL,
  ...,
  index = NULL,
  variable_name = NULL,
  value_name = NULL
)
}
\arguments{
\item{on}{Values to use as identifier variables. If \code{value_vars} is
empty all columns that are not in \code{id_vars} will be used.}

\item{...}{Not used.}

\item{index}{Columns to use as identifier variables.}

\item{variable_name}{Name to give to the new column containing the names of
the melted columns. Defaults to "variable".}

\item{value_name}{Name to give to the new column containing the values of
the melted columns. Defaults to "value"}
}
\value{
A new \code{DataFrame}
}
\description{
Unpivot a Frame from wide to long format
}
\details{
Optionally leaves identifiers set.

This function is useful to massage a Frame into a format where one or more
columns are identifier variables (id_vars), while all other columns, considered
measured variables (value_vars), are "unpivoted" to the row axis, leaving just
two non-identifier columns, 'variable' and 'value'.
}
\examples{
df = pl$DataFrame(
  a = c("x", "y", "z"),
  b = c(1, 3, 5),
  c = c(2, 4, 6),
  d = c(7, 8, 9)
)
df$unpivot(index = "a", on = c("b", "c", "d"))
}
\keyword{DataFrame}
