% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_width}
\alias{DataFrame_width}
\title{Number of columns of a DataFrame}
\usage{
<DataFrame>$width()
}
\value{
The number of columns of a DataFrame
}
\description{
Get the number of columns (width) of a DataFrame
}
\examples{
pl$DataFrame(iris)$width
}
\keyword{DataFrame}
