% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_Array}
\alias{DataType_Array}
\title{Create Array DataType}
\format{
function
}
\usage{
DataType_Array(datatype = "unknown", width)
}
\arguments{
\item{datatype}{an inner DataType, default is "Unknown" (placeholder for when inner DataType
does not matter, e.g. as used in example)}

\item{width}{The length of the arrays}
}
\value{
an array DataType with an inner DataType
}
\description{
Create Array DataType
}
\examples{
# basic Array
pl$Array(pl$Int32, 4)
# some nested Array
pl$Array(pl$Array(pl$Boolean, 3), 2)
}
\keyword{pl}
