% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_Categorical}
\alias{DataType_Categorical}
\title{Create Categorical DataType}
\usage{
DataType_Categorical(ordering = "physical")
}
\arguments{
\item{ordering}{Either \code{"physical"} (default) or \code{"lexical"}.}
}
\value{
A Categorical DataType
}
\description{
Create Categorical DataType
}
\details{
When a categorical variable is created, its string values (or "lexical"
values) are stored and encoded as integers ("physical" values) by
order of appearance. Therefore, sorting a categorical value can be done
either on the lexical or on the physical values. See Examples.
}
\examples{
# default is to order by physical values
df = pl$DataFrame(x = c("z", "z", "k", "a", "z"), schema = list(x = pl$Categorical()))
df$sort("x")

# when setting ordering = "lexical", sorting will be based on the strings
df_lex = pl$DataFrame(
  x = c("z", "z", "k", "a", "z"),
  schema = list(x = pl$Categorical("lexical"))
)
df_lex$sort("x")
}
