% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_contains_categoricals}
\alias{DataType_contains_categoricals}
\title{Check whether the data type contains categoricals}
\usage{
DataType_contains_categoricals()
}
\value{
A logical value
}
\description{
Check whether the data type contains categoricals
}
\examples{
pl$List(pl$Categorical())$contains_categoricals()
pl$List(pl$Enum(c("a", "b")))$contains_categoricals()
pl$List(pl$Float32)$contains_categoricals()
pl$List(pl$List(pl$Categorical()))$contains_categoricals()
}
