% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_contains_views}
\alias{DataType_contains_views}
\title{Check whether the data type contains views}
\usage{
DataType_contains_views()
}
\value{
A logical value
}
\description{
Check whether the data type contains views
}
\examples{
pl$List(pl$String)$contains_views()
pl$List(pl$Binary)$contains_views()
pl$List(pl$Float32)$contains_views()
pl$List(pl$List(pl$Binary))$contains_views()
}
