% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_Datetime}
\alias{DataType_Datetime}
\title{Create Datetime DataType}
\usage{
DataType_Datetime(tu = "us", tz = NULL)
}
\arguments{
\item{tu}{string option either "ms", "us" or "ns"}

\item{tz}{string the Time Zone, see details}
}
\value{
Datetime DataType
}
\description{
Datetime DataType constructor
}
\details{
all allowed TimeZone designations can be found in \code{base::OlsonNames()}
}
\examples{
pl$Datetime("ns", "Pacific/Samoa")
}
\keyword{pl}
