% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_Duration}
\alias{DataType_Duration}
\title{Data type representing a time duration}
\usage{
DataType_Duration(time_unit = "us")
}
\arguments{
\item{time_unit}{Unit of time. One of \code{"ms"}, \code{"us"} (default) or \code{"ns"}.}
}
\value{
Duration DataType
}
\description{
Data type representing a time duration
}
\examples{
test = pl$DataFrame(
  a = 1:2,
  b = c("a", "b"),
  c = pl$duration(weeks = c(1, 2), days = c(0, 2))
)

# select all columns of type "duration"
test$select(pl$col(pl$Duration()))
}
