% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/datatype-expr-expr.R
\name{datatype_expr__display}
\alias{datatype_expr__display}
\title{Get a formatted version of the output DataType}
\usage{
datatype_expr__display()
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get a formatted version of the output DataType.
}
\examples{
df <- pl$DataFrame(x = 1:2, y = c("a", "b"), z = c(1, 2))
df$select(
  x = pl$dtype_of("x")$display(),
  y = pl$dtype_of("y")$display(),
  z = pl$dtype_of("z")$display(),
)$transpose(include_header = TRUE)
}
