% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/datatype-expr-expr.R
\name{datatype_expr__matches}
\alias{datatype_expr__matches}
\title{Get whether the output DataType matches a certain selector}
\usage{
datatype_expr__matches(selector)
}
\arguments{
\item{selector}{A \link[=cs]{selector} presenting the data types to match.}
}
\value{
A polars datatype expression.
This is not the same as a \link[=polars_expr]{polars expression}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get whether the output DataType matches a certain selector.
}
\examples{
df <- pl$DataFrame(a = 1:3)
df$select(
  a_is_string = pl$dtype_of("a")$matches(cs$string()),
  a_is_integer = pl$dtype_of("a")$matches(cs$integer()),
)
}
