% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/datatype-expr-expr.R
\name{datatype_expr__wrap_in_array}
\alias{datatype_expr__wrap_in_array}
\title{Get the DataType wrapped in an array}
\usage{
datatype_expr__wrap_in_array(..., width)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{width}{Width of the array.}
}
\value{
A polars datatype expression.
This is not the same as a \link[=polars_expr]{polars expression}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get the DataType wrapped in an array.
}
\examples{
df <- pl$DataFrame(
  a = 1L,
  b = list("a"),
  c = data.frame(x = 1, y = 2)
)

df$select(
  a_wrapped = pl$dtype_of("a")$wrap_in_array(width = 1)$display(),
  b_wrapped = pl$dtype_of("b")$wrap_in_array(width = 1)$display(),
  c_wrapped = pl$dtype_of("c")$wrap_in_array(width = 1)$display()
)
}
