% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/datatype-expr-expr.R
\name{datatype_expr__wrap_in_list}
\alias{datatype_expr__wrap_in_list}
\title{Get the DataType wrapped in a list}
\usage{
datatype_expr__wrap_in_list()
}
\value{
A polars datatype expression.
This is not the same as a \link[=polars_expr]{polars expression}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get the DataType wrapped in a list.
}
\examples{
df <- pl$DataFrame(
  a = 1L,
  b = list("a"),
  c = data.frame(x = 1, y = 2)
)

df$select(
  a_wrapped = pl$dtype_of("a")$wrap_in_list()$display(),
  b_wrapped = pl$dtype_of("b")$wrap_in_list()$display(),
  c_wrapped = pl$dtype_of("c")$wrap_in_list()$display()
)
}
