% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype-expr-expr.R
\name{datatype_expr_matches}
\alias{datatype_expr_matches}
\title{Get whether the output DataType is matches a certain selector}
\usage{
datatype_expr_matches(selector)
}
\value{
A polars \link{expression}
}
\description{
Null values are counted in the total.
}
\examples{
df <- pl$DataFrame(a = 1:3)
df$select(
  a_is_string = pl$dtype_of("a")$matches(cs$string()),
  a_is_integer = pl$dtype_of("a")$matches(cs$integer()),
)
}
