% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_array}
\alias{DataType_is_array}
\title{Check whether the data type is an array type}
\usage{
DataType_is_array()
}
\value{
A logical value
}
\description{
Check whether the data type is an array type
}
\examples{
pl$Array(width = 2)$is_array()
pl$Float32$is_array()
}
