% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_binary}
\alias{DataType_is_binary}
\title{Check whether the data type is a binary type}
\usage{
DataType_is_binary()
}
\value{
A logical value
}
\description{
Check whether the data type is a binary type
}
\examples{
pl$Binary$is_binary()
pl$Float32$is_binary()
}
