% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_bool}
\alias{DataType_is_bool}
\title{Check whether the data type is a boolean type}
\usage{
DataType_is_bool()
}
\value{
A logical value
}
\description{
Check whether the data type is a boolean type
}
\examples{
pl$Boolean$is_bool()
pl$Float32$is_bool()
}
