% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_categorical}
\alias{DataType_is_categorical}
\title{Check whether the data type is a Categorical type}
\usage{
DataType_is_categorical()
}
\value{
A logical value
}
\description{
Check whether the data type is a Categorical type
}
\examples{
pl$Categorical()$is_categorical()
pl$Enum(c("a", "b"))$is_categorical()
}
