% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_float}
\alias{DataType_is_float}
\title{Check whether the data type is a float type}
\usage{
DataType_is_float()
}
\value{
A logical value
}
\description{
Check whether the data type is a float type
}
\examples{
pl$Float32$is_float()
pl$Int32$is_float()
}
