% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_integer}
\alias{DataType_is_integer}
\title{Check whether the data type is an integer type}
\usage{
DataType_is_integer()
}
\value{
A logical value
}
\description{
Check whether the data type is an integer type
}
\examples{
pl$Int32$is_integer()
pl$Float32$is_integer()
}
