% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_known}
\alias{DataType_is_known}
\title{Check whether the data type is known}
\usage{
DataType_is_known()
}
\value{
A logical value
}
\description{
Check whether the data type is known
}
\examples{
pl$String$is_known()
pl$Unknown$is_known()
}
