% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_list}
\alias{DataType_is_list}
\title{Check whether the data type is a list type}
\usage{
DataType_is_list()
}
\value{
A logical value
}
\description{
Check whether the data type is a list type
}
\examples{
pl$List()$is_list()
pl$Float32$is_list()
}
