% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_logical}
\alias{DataType_is_logical}
\title{Check whether the data type is a logical type}
\usage{
DataType_is_logical()
}
\value{
A logical value
}
\description{
Check whether the data type is a logical type
}
