% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_nested}
\alias{DataType_is_nested}
\title{Check whether the data type is a nested type}
\usage{
DataType_is_nested()
}
\value{
A logical value
}
\description{
Check whether the data type is a nested type
}
\examples{
pl$List()$is_nested()
pl$Array(width = 2)$is_nested()
pl$Float32$is_nested()
}
