% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_null}
\alias{DataType_is_null}
\title{Check whether the data type is a null type}
\usage{
DataType_is_null()
}
\value{
A logical value
}
\description{
Check whether the data type is a null type
}
\examples{
pl$Null$is_null()
pl$Float32$is_null()
}
