% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_numeric}
\alias{DataType_is_numeric}
\title{Check whether the data type is a numeric type}
\usage{
DataType_is_numeric()
}
\value{
A logical value
}
\description{
Check whether the data type is a numeric type
}
\examples{
pl$Float32$is_numeric()
pl$Int32$is_numeric()
pl$String$is_numeric()
}
