% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_ord}
\alias{DataType_is_ord}
\title{Check whether the data type is an ordinal type}
\usage{
DataType_is_ord()
}
\value{
A logical value
}
\description{
Check whether the data type is an ordinal type
}
\examples{
pl$String$is_ord()
pl$Categorical()$is_ord()
}
