% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_primitive}
\alias{DataType_is_primitive}
\title{Check whether the data type is a primitive type}
\usage{
DataType_is_primitive()
}
\value{
A logical value
}
\description{
Check whether the data type is a primitive type
}
\examples{
pl$Float32$is_primitive()
pl$List()$is_primitive()
}
