% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_signed_integer}
\alias{DataType_is_signed_integer}
\title{Check whether the data type is a signed integer type}
\usage{
DataType_is_signed_integer()
}
\value{
A logical value
}
\description{
Check whether the data type is a signed integer type
}
\examples{
pl$Int32$is_signed_integer()
pl$UInt32$is_signed_integer()
}
