% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_string}
\alias{DataType_is_string}
\title{Check whether the data type is a String type}
\usage{
DataType_is_string()
}
\value{
A logical value
}
\description{
Check whether the data type is a String type
}
\examples{
pl$String$is_string()
pl$Float32$is_string()
}
