% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_struct}
\alias{DataType_is_struct}
\title{Check whether the data type is a temporal type}
\usage{
DataType_is_struct()
}
\value{
A logical value
}
\description{
Check whether the data type is a temporal type
}
\examples{
pl$Struct()$is_struct()
pl$Float32$is_struct()
}
