% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_temporal}
\alias{DataType_is_temporal}
\title{Check whether the data type is a temporal type}
\usage{
DataType_is_temporal()
}
\value{
A logical value
}
\description{
Check whether the data type is a temporal type
}
\examples{
pl$Date$is_temporal()
pl$Float32$is_temporal()
}
