% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_is_unsigned_integer}
\alias{DataType_is_unsigned_integer}
\title{Check whether the data type is an unsigned integer type}
\usage{
DataType_is_unsigned_integer()
}
\value{
A logical value
}
\description{
Check whether the data type is an unsigned integer type
}
\examples{
pl$UInt32$is_unsigned_integer()
pl$Int32$is_unsigned_integer()
}
