% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_List}
\alias{DataType_List}
\title{Create List DataType}
\usage{
DataType_List(datatype = "unknown")
}
\arguments{
\item{datatype}{an inner DataType, default is "Unknown" (placeholder for when inner DataType
does not matter, e.g. as used in example)}
}
\value{
a list DataType with an inner DataType
}
\description{
Create List DataType
}
\examples{
# some nested List
pl$List(pl$List(pl$Boolean))

# check if some maybe_list is a List DataType
maybe_List = pl$List(pl$UInt64)
pl$same_outer_dt(maybe_List, pl$List())
}
\keyword{pl}
