% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_new}
\alias{DataType_new}
\title{DataType_new (simple DataType's)}
\usage{
DataType_new(str)
}
\arguments{
\item{str}{name of DataType to create}
}
\value{
DataType
}
\description{
Create a new flag like DataType
}
\details{
This function is mainly used in \code{zzz.R} \code{.onLoad} to instantiate singletons of all
flag-like DataType.

Non-flag like DataType called composite DataTypes also carries extra information
e.g. Datetime a timeunit and a TimeZone, or List which recursively carries another DataType
inside. Composite DataTypes use DataType constructors.

Any DataType can be found in pl$dtypes
}
\examples{
rpolars:::DataType_new("Int64")
}
\keyword{internal}
