% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{.DollarNames.Expr}
\alias{.DollarNames.Expr}
\title{auto complete $-access into object}
\usage{
\method{.DollarNames}{Expr}(x, pattern = "")
}
\description{
called by the interactive R session internally
}
\keyword{Expr}
