% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.DollarNames.method_environment}
\alias{.DollarNames.method_environment}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{method_environment}(x, pattern = "")
}
\arguments{
\item{x}{string, name of method in method_environment (sub-namespace)}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
Doesn't return a value. This is used for autocompletion in RStudio.
}
\description{
called by the interactive R session internally
}
\keyword{internal}
