% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{.DollarNames.polars_option_list}
\alias{.DollarNames.polars_option_list}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{polars_option_list}(x, pattern = "")
}
\arguments{
\item{x}{string name of an option}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
Doesn't return a value. This is used for autocompletion in RStudio.
}
\description{
called by the interactive R session internally
}
\keyword{internal}
