% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__background.R
\name{.DollarNames.PolarsBackgroundHandle}
\alias{.DollarNames.PolarsBackgroundHandle}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{PolarsBackgroundHandle}(x, pattern = "")
}
\arguments{
\item{x}{PolarsLazyFrame}

\item{pattern}{code-stump as string to auto-complete}
}
\description{
called by the interactive R session internally
}
\keyword{internal}
