% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerr.R
\name{.DollarNames.RPolarsErr}
\alias{.DollarNames.RPolarsErr}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RPolarsErr}(x, pattern = "")
}
\arguments{
\item{x}{RPolarsErr}

\item{pattern}{code-stump as string to auto-complete}
}
\description{
called by the interactive R session internally
}
\keyword{internal}
