% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{.DollarNames.RPolarsExpr}
\alias{.DollarNames.RPolarsExpr}
\title{Auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RPolarsExpr}(x, pattern = "")
}
\arguments{
\item{x}{Name of an \code{Expr} object}

\item{pattern}{String used to auto-complete}
}
\value{
char vec

Doesn't return a value. This is used for autocompletion in RStudio.
}
\description{
Called by the interactive R session internally
}
\keyword{internal}
