% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{.DollarNames.RPolarsGroupBy}
\alias{.DollarNames.RPolarsGroupBy}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RPolarsGroupBy}(x, pattern = "")
}
\arguments{
\item{x}{GroupBy}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
char vec
}
\description{
called by the interactive R session internally
}
\keyword{internal}
