% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{.DollarNames.RPolarsLazyFrame}
\alias{.DollarNames.RPolarsLazyFrame}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RPolarsLazyFrame}(x, pattern = "")
}
\arguments{
\item{x}{LazyFrame}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
char vec
}
\description{
called by the interactive R session internally
}
\keyword{internal}
