% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{.DollarNames.RPolarsRField}
\alias{.DollarNames.RPolarsRField}
\title{Auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RPolarsRField}(x, pattern = "")
}
\arguments{
\item{x}{Name of a \code{RPolarsRField} object}

\item{pattern}{String used to auto-complete}
}
\description{
Called by the interactive R session internally
}
\keyword{internal}
