% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{.DollarNames.RPolarsRThreadHandle}
\alias{.DollarNames.RPolarsRThreadHandle}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RPolarsRThreadHandle}(x, pattern = "")
}
\arguments{
\item{x}{RPolarsRThreadHandle}

\item{pattern}{code-stump as string to auto-complete}
}
\description{
called by the interactive R session internally
}
\keyword{internal}
