% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{.DollarNames.RPolarsVecDataFrame}
\alias{.DollarNames.RPolarsVecDataFrame}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RPolarsVecDataFrame}(x, pattern = "")
}
\arguments{
\item{x}{RPolarsVecDataFrame}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
char vec
}
\description{
called by the interactive R session internally
}
\keyword{internal}
