% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{.DollarNames.Series}
\alias{.DollarNames.Series}
\title{auto complete $-access into object}
\usage{
\method{.DollarNames}{Series}(x, pattern = "")
}
\description{
called by the interactive R session internally
}
\keyword{Series}
