% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__whenthen.R
\name{.DollarNames.Then}
\alias{.DollarNames.Then}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{Then}(x, pattern = "")
}
\arguments{
\item{x}{Then}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
char vec
}
\description{
called by the interactive R session internally
}
\keyword{internal}
