% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whenthen.R
\name{.DollarNames.When}
\alias{.DollarNames.When}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{When}(x, pattern = "")
}
\arguments{
\item{x}{When}

\item{pattern}{code-stump as string to auto-complete}
}
\description{
called by the interactive R session internally
}
\keyword{internal}
