% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{dt_microsecond}
\alias{dt_microsecond}
\alias{dt.microsecond}
\alias{arr_microsecond}
\title{Micromicrosecond}
\format{
function
}
\value{
Expr of microsecond as UInt32
}
\description{
Extract microseconds from underlying DateTime representation.
Applies to Datetime columns.
}
\examples{
pl$DataFrame(date = pl$date_range(
  as.numeric(as.POSIXct("2001-1-1"))*1E6+456789, #manually convert to us
  as.numeric(as.POSIXct("2001-1-1 00:00:6"))*1E6,
  interval = "2s654321us",
  time_unit = "us" #instruct polars input is us, and store as us
))$with_columns(
  pl$col("date")$cast(pl$Int64)$alias("datetime int64"),
  pl$col("date")$dt$microsecond()$alias("microsecond")
)
}
\keyword{ExprDT}
