% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{enable_string_cache}
\alias{enable_string_cache}
\title{enable disable global string cache}
\arguments{
\item{toggle}{Boolean. TRUE enable}
}
\value{
no return
}
\description{
e.g. join to Categorical series is only allowed if using global string cache.
Local categorical encodings may differ and is not meaningful to combine as is.
}
\examples{
pl$enable_string_cache(TRUE)
}
