% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expect_grepl_error}
\alias{expect_grepl_error}
\title{expect grepl error}
\usage{
expect_grepl_error(expr, expected_err = NULL)
}
\arguments{
\item{expr}{an R expression to test}

\item{expected_err}{a string pattern passed to grepl}
}
\value{
invisble NULL
}
\description{
expect grepl error
}
\details{
expr must raise an error and expected_err pattern must match
against the error text with grepl()
}
\examples{
#passes
expect_grepl_error(stop("orange and carrot"),"carrot")

#fails
do_not_run {
expect_grepl_error(stop("orange and carrot"),"big carrot")

}

}
