% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expect_strictly_identical}
\alias{expect_strictly_identical}
\title{more strict than expect_identical}
\usage{
expect_strictly_identical(object, expected, ...)
}
\description{
more strict than expect_identical
}
