% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__agg_groups}
\alias{expr__agg_groups}
\title{Get the group indexes of the group by operation}
\usage{
expr__agg_groups()
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
\code{agg_groups()} is deprecated as of polars 1.16.0. Use
\code{df$with_row_index()$group_by(..., .maintain_order = TRUE)$agg(pl$col("index"))}
instead.

Should be used in aggregation context only.
}
\examples{
df <- pl$DataFrame(
  group = rep(c("one", "two"), each = 3),
  value = c(94, 95, 96, 97, 97, 99)
)

df$group_by("group", .maintain_order = TRUE)$agg(pl$col("value")$agg_groups())

# Recommended approach
df$with_row_index()$group_by("group", .maintain_order = TRUE)$agg(pl$col("index"))
}
