% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__any}
\alias{expr__any}
\title{Check if any boolean value in a column is true}
\usage{
expr__any(..., ignore_nulls = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{ignore_nulls}{If \code{TRUE} (default), ignore null values. If \code{FALSE},
\href{https://en.wikipedia.org/wiki/Three-valued_logic}{Kleene logic} is used to
deal with nulls: if the column contains any null values and no \code{TRUE} values,
the output is null.}
}
\value{
A polars \link{expression}
}
\description{
Check if any boolean value in a column is true
}
\examples{
df <- pl$DataFrame(
  a = c(TRUE, FALSE),
  b = c(FALSE, FALSE),
  c = c(NA, FALSE)
)

df$select(pl$col("*")$any())

# If we set ignore_nulls = FALSE, then we don't know if any values in column
# "c" is TRUE, so it returns null
df$select(pl$col("*")$any(ignore_nulls = FALSE))
}
