% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__approx_n_unique}
\alias{expr__approx_n_unique}
\title{Approximate count of unique values}
\usage{
expr__approx_n_unique()
}
\value{
A polars \link{expression}
}
\description{
This is done using the HyperLogLog++ algorithm for cardinality estimation.
}
\examples{
df <- pl$DataFrame(n = c(1, 1, 2))
df$select(pl$col("n")$approx_n_unique())

df <- pl$DataFrame(n = 0:1000)
df$select(
  exact = pl$col("n")$n_unique(),
  approx = pl$col("n")$approx_n_unique()
)
}
