% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arccosh}
\alias{expr__arccosh}
\title{Compute inverse hyperbolic cosine}
\usage{
expr__arccosh()
}
\value{
A polars \link{expression}
}
\description{
Compute inverse hyperbolic cosine
}
\examples{
pl$DataFrame(a = c(-1, cosh(0.5), 0, 1, NA))$
  with_columns(arccosh = pl$col("a")$arccosh())
}
