% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arctan}
\alias{expr__arctan}
\title{Compute inverse tangent}
\usage{
expr__arctan()
}
\value{
A polars \link{expression}
}
\description{
Compute inverse tangent
}
\examples{
pl$DataFrame(a = c(-1, tan(0.5), 0, 1, NA_real_))$
  with_columns(arctan = pl$col("a")$arctan())
}
