% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arg_sort}
\alias{expr__arg_sort}
\title{Index of a sort}
\usage{
expr__arg_sort(..., descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{descending}{Sort in descending order.}

\item{nulls_last}{Place null values last.}
}
\value{
A polars \link{expression}
}
\description{
Get the index values that would sort this column.
}
\examples{
pl$DataFrame(
  a = c(6, 1, 0, NA, Inf, NaN)
)$with_columns(arg_sorted = pl$col("a")$arg_sort())
}
\seealso{
\link[=pl__arg_sort_by]{pl$arg_sort_by()} to find the row indices that would
sort multiple columns.
}
