% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arg_true}
\alias{expr__arg_true}
\title{Return indices where expression is true}
\usage{
expr__arg_true()
}
\value{
A polars \link{expression}
}
\description{
Return indices where expression is true
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2, 1))
df$select((pl$col("a") == 1)$arg_true())
}
