% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arg_unique}
\alias{expr__arg_unique}
\title{Get the index of the first unique value}
\usage{
expr__arg_unique()
}
\value{
A polars \link{expression}
}
\description{
Get the index of the first unique value
}
\examples{
df <- pl$DataFrame(a = 1:3, b = c(NA, 4, 4))
df$select(pl$col("a")$arg_unique())
df$select(pl$col("b")$arg_unique())
}
