% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bitwise_xor}
\alias{expr__bitwise_xor}
\title{Perform an aggregation of bitwise XORs.}
\usage{
expr__bitwise_xor()
}
\value{
A polars \link{expression}
}
\description{
Perform an aggregation of bitwise XORs.
}
\examples{
df <- pl$DataFrame(n = -1:1)
df$select(pl$col("n")$bitwise_xor())

df <- pl$DataFrame(
  grouper = c("a", "a", "a", "b", "b"),
  n = c(-1L, 0L, 1L, -1L, 1L)
)
df$group_by("grouper", .maintain_order = TRUE)$agg(pl$col("n")$bitwise_xor())
}
