% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bottom_k}
\alias{expr__bottom_k}
\title{Return the \code{k} smallest elements}
\usage{
expr__bottom_k(k = 5)
}
\arguments{
\item{k}{Number of elements to return.}
}
\value{
A polars \link{expression}
}
\description{
Non-null elements are always preferred over null elements. The output is not
guaranteed to be in any particular order, call \link[=expr__sort]{$sort()} after
this function if you wish the output to be sorted. This has time complexity
\eqn{O(n)}.
}
\examples{
df <- pl$DataFrame(value = c(1, 98, 2, 3, 99, 4))
df$select(
  top_k = pl$col("value")$top_k(k = 3),
  bottom_k = pl$col("value")$bottom_k(k = 3)
)
}
