% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bottom_k_by}
\alias{expr__bottom_k_by}
\title{Return the elements corresponding to the \code{k} smallest elements of the \code{by}
column(s)}
\usage{
expr__bottom_k_by(by, k = 5, ..., reverse = FALSE)
}
\arguments{
\item{by}{Column(s) used to determine the smallest elements. Accepts
expression input. Strings are parsed as column names.}

\item{k}{Number of elements to return.}

\item{...}{These dots are for future extensions and must be empty.}

\item{reverse}{Consider the \code{k} largest elements of the \code{by} column(s)
(instead of the \code{k} smallest). This can be specified per column by passing a
sequence of booleans.}
}
\value{
A polars \link{expression}
}
\description{
Non-null elements are always preferred over null elements. The output is not
guaranteed to be in any particular order, call \link[=expr__sort]{$sort()} after
this function if you wish the output to be sorted. This has time complexity
\eqn{O(n)}.
}
\examples{
df <- pl$DataFrame(
  a = 1:6,
  b = 6:1,
  c = c("Apple", "Orange", "Apple", "Apple", "Banana", "Banana")
)

# Get the bottom 2 rows by column a or b:
df$select(
  pl$all()$bottom_k_by("a", 2)$name$suffix("_btm_by_a"),
  pl$all()$bottom_k_by("b", 2)$name$suffix("_btm_by_b")
)

# Get the bottom 2 rows by multiple columns with given order.
df$select(
  pl$all()$
    bottom_k_by(c("c", "a"), 2, reverse = c(FALSE, TRUE))$
    name$suffix("_btm_by_ca"),
  pl$all()$
    bottom_k_by(c("c", "b"), 2, reverse = c(FALSE, TRUE))$
    name$suffix("_btm_by_cb"),
)

# Get the bottom 2 rows by column a in each group
df$group_by("c", .maintain_order = TRUE)$agg(
  pl$all()$bottom_k_by("a", 2)
)$explode(pl$all()$exclude("c"))
}
