% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cot}
\alias{expr__cot}
\title{Compute cotangent}
\usage{
expr__cot()
}
\value{
A polars \link{expression}
}
\description{
Compute cotangent
}
\examples{
pl$DataFrame(a = c(0, pi / 2, -5, NA))$
  with_columns(cotangent = pl$col("a")$cot())
}
