% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__count}
\alias{expr__count}
\title{Get the number of non-null elements in the column}
\usage{
expr__count()
}
\value{
A polars \link{expression}
}
\description{
Get the number of non-null elements in the column
}
\examples{
df <- pl$DataFrame(a = 1:3, b = c(NA, 4, 4))
df$select(pl$all()$count())
}
