% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cum_count}
\alias{expr__cum_count}
\title{Return the cumulative count of the non-null values in the column}
\usage{
expr__cum_count(..., reverse = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{reverse}{If \code{TRUE}, reverse the count.}
}
\value{
A polars \link{expression}
}
\description{
Return the cumulative count of the non-null values in the column
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  cum_count = pl$col("a")$cum_count(),
  cum_count_reversed = pl$col("a")$cum_count(reverse = TRUE)
)
}
