% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cum_sum}
\alias{expr__cum_sum}
\title{Return the cumulative sum computed at every element.}
\usage{
expr__cum_sum(..., reverse = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{reverse}{If \code{TRUE}, start with the total sum of elements and subtract
each row one by one.}
}
\value{
A polars \link{expression}
}
\description{
Return the cumulative sum computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  cum_sum = pl$col("a")$cum_sum(),
  cum_sum_reversed = pl$col("a")$cum_sum(reverse = TRUE)
)
}
