% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__degrees}
\alias{expr__degrees}
\title{Convert from radians to degrees}
\usage{
expr__degrees()
}
\value{
A polars \link{expression}
}
\description{
Convert from radians to degrees
}
\examples{
pl$DataFrame(a = c(1, 2, 4) * pi)$
  with_columns(degrees = pl$col("a")$degrees())
}
