% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__dot}
\alias{expr__dot}
\title{Compute the dot/inner product between two Expressions}
\usage{
expr__dot(other)
}
\arguments{
\item{other}{Expression to compute dot product with.}
}
\value{
A polars \link{expression}
}
\description{
Compute the dot/inner product between two Expressions
}
\examples{
df <- pl$DataFrame(a = c(1, 3, 5), b = c(2, 4, 6))
df$select(pl$col("a")$dot(pl$col("b")))
}
