% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__entropy}
\alias{expr__entropy}
\title{Compute entropy}
\usage{
expr__entropy(base = exp(1), ..., normalize = TRUE)
}
\arguments{
\item{base}{Numeric value used as base, defaults to \code{exp(1)}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{normalize}{Normalize \code{pk} if it doesn’t sum to 1.}
}
\value{
A polars \link{expression}
}
\description{
Uses the formula \verb{-sum(pk * log(pk)} where \code{pk} are discrete probabilities.
}
\examples{
df <- pl$DataFrame(a = 1:3)
df$select(pl$col("a")$entropy(base = 2))
df$select(pl$col("a")$entropy(base = 2, normalize = FALSE))
}
