% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__eq}
\alias{expr__eq}
\title{Check equality}
\usage{
expr__eq(other)
}
\arguments{
\item{other}{A literal or expression value to compare with.}
}
\value{
A polars \link{expression}
}
\description{
This propagates null values, i.e. any comparison involving \code{null} will
return \code{null}. Use \code{\link[=expr__eq_missing]{$eq_missing()}} to consider null
values as equal.
}
\examples{
df <- pl$DataFrame(x = c(NA, FALSE, TRUE), y = c(TRUE, TRUE, TRUE))
df$with_columns(
  eq = pl$col("x")$eq(pl$col("y")),
  eq_missing = pl$col("x")$eq_missing(pl$col("y"))
)
}
\seealso{
\link{expr__eq_missing}
}
