% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__eq_missing}
\alias{expr__eq_missing}
\title{Check equality without \code{null} propagation}
\usage{
expr__eq_missing(other)
}
\arguments{
\item{other}{A literal or expression value to compare with.}
}
\value{
A polars \link{expression}
}
\description{
This considers that null values are equal. It differs from
\code{\link[=expr__eq]{$eq()}} where null values are propagated.
}
\examples{
df <- pl$DataFrame(x = c(NA, FALSE, TRUE), y = c(TRUE, TRUE, TRUE))
df$with_columns(
  eq = pl$col("x")$eq("y"),
  eq_missing = pl$col("x")$eq_missing("y")
)
}
\seealso{
\link{expr__eq}
}
