% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__explode}
\alias{expr__explode}
\title{Explode a list expression}
\usage{
expr__explode(..., empty_as_null = TRUE, keep_nulls = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{empty_as_null}{Indicates to explode an empty list/array into a \code{null}.
Defaults to \code{TRUE}.}

\item{keep_nulls}{Indicates to explode a \code{null} list/array into a \code{null}.}
}
\value{
A polars \link{expression}
}
\description{
This means that every item is expanded to a new row.
}
\examples{
df <- pl$DataFrame(
  groups = c("a", "b"),
  values = list(1:2, 3:4)
)

df$select(pl$col("values")$explode())
}
