% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__filter}
\alias{expr__filter}
\title{Filter the expression based on one or more predicate expressions}
\usage{
expr__filter(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Expression(s) that evaluate
to a boolean Series.}
}
\value{
A polars \link{expression}
}
\description{
Elements where the filter does not evaluate to \code{TRUE} are discarded,
including nulls. This is mostly useful in an aggregation context. If you
want to filter on a DataFrame level, use
\code{\link[=dataframe__filter]{DataFrame$filter()}} or
\code{\link[=lazyframe__filter]{LazyFrame$filter()}}.
}
\examples{
df <- pl$DataFrame(
  group_col = c("g1", "g1", "g2"),
  b = c(1, 2, 3)
)
df

df$group_by("group_col")$agg(
  lt = pl$col("b")$filter(pl$col("b") < 2),
  gte = pl$col("b")$filter(pl$col("b") >= 2)
)
}
