% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__hash}
\alias{expr__hash}
\title{Hash elements}
\usage{
expr__hash(seed = 0)
}
\arguments{
\item{seed}{Integer, random seed parameter. Defaults to 0.}
}
\value{
A polars \link{expression}
}
\description{
Hash elements
}
\details{
This implementation of hash does not guarantee stable results across
different Polars versions. Its stability is only guaranteed within a single
version.
}
\examples{
df <- pl$DataFrame(a = c(1, 2, NA), b = c("x", NA, "z"))
df$with_columns(pl$all()$hash(seed = 10))
}
