% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__head}
\alias{expr__head}
\title{Get the first n elements}
\usage{
expr__head(n = 10)
}
\arguments{
\item{n}{Number of elements to take.}
}
\value{
A polars \link{expression}
}
\description{
Get the first n elements
}
\examples{
pl$DataFrame(x = 1:11)$select(pl$col("x")$head(3))
}
