% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__implode}
\alias{expr__implode}
\title{Aggregate values into a list}
\usage{
expr__implode(..., maintain_order = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{maintain_order}{Whether to preserve the order of elements in the list.
Setting this to \code{FALSE} can improve performance, especially within
\verb{$group_by()}.}
}
\value{
A polars \link{expression}
}
\description{
Aggregate values into a list
}
\examples{
df <- pl$DataFrame(a = 1:3, b = 4:6)
df$with_columns(pl$col("a")$implode())
}
