% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__interpolate}
\alias{expr__interpolate}
\title{Fill null values using interpolation}
\usage{
expr__interpolate(method = c("linear", "nearest"))
}
\arguments{
\item{method}{Interpolation method. Must be one of \code{"linear"} or \code{"nearest"}.}
}
\value{
A polars \link{expression}
}
\description{
Fill null values using interpolation
}
\examples{
df <- pl$DataFrame(a = c(1, NA, 3), b = c(1, NaN, 3))
df$with_columns(
  a_interpolated = pl$col("a")$interpolate(),
  b_interpolated = pl$col("b")$interpolate()
)
}
