% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__interpolate_by}
\alias{expr__interpolate_by}
\title{Fill null values using interpolation based on another column}
\usage{
expr__interpolate_by(by)
}
\arguments{
\item{by}{Column to interpolate values based on.}
}
\value{
A polars \link{expression}
}
\description{
Fill null values using interpolation based on another column
}
\examples{
df <- pl$DataFrame(a = c(1, NA, NA, 3), b = c(1, 2, 7, 8))
df$with_columns(
  a_interpolated = pl$col("a")$interpolate_by("b")
)
}
