% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_duplicated}
\alias{expr__is_duplicated}
\title{Return a boolean mask indicating duplicated values}
\usage{
expr__is_duplicated()
}
\value{
A polars \link{expression}
}
\description{
Return a boolean mask indicating duplicated values
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2, 3, 2))
df$select(pl$col("a")$is_duplicated())
}
