% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_unique}
\alias{expr__is_unique}
\title{Return a boolean mask indicating unique values}
\usage{
expr__is_unique()
}
\value{
A polars \link{expression}
}
\description{
Return a boolean mask indicating unique values
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2, 3, 2))
df$select(pl$col("a")$is_unique())
}
