% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__item}
\alias{expr__item}
\title{Get the single value}
\usage{
expr__item(..., allow_empty = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{allow_empty}{Allow having no values to return \code{null}.}
}
\value{
A polars \link{expression}
}
\description{
This raises an error if there is not exactly one value.
}
\examples{
# This function may be useful to force ourselves to handle cases where we
# expect a single value but there are several. For example, we might expect
# `$mode()` to return a single value...
df <- pl$DataFrame(x = c("a", "a", "c"))
df$select(
  mode = pl$col("x")$mode()$item()
)

# ... but this is not always the case:
df <- pl$DataFrame(x = c("a", "b", "c"))
tryCatch(
  \{
    df$select(
      mode = pl$col("x")$mode()$item()
    )
  \},
  error = function(e) print(e)
)
}
\seealso{
\link[=expr__get]{$get()} to get a single value by index.
}
