% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__last}
\alias{expr__last}
\title{Get the last value}
\usage{
expr__last()
}
\value{
A polars \link{expression}
}
\description{
Get the last value
}
\examples{
pl$DataFrame(x = 3:1)$with_columns(last = pl$col("x")$last())
}
