% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__le}
\alias{expr__le}
\title{Check lower or equal inequality}
\usage{
expr__le(other)
}
\arguments{
\item{other}{A literal or expression value to compare with.}
}
\value{
A polars \link{expression}
}
\description{
Check lower or equal inequality
}
\examples{
df <- pl$DataFrame(x = 1:3)
df$with_columns(
  with_le = pl$col("x")$le(pl$lit(2)),
  with_symbol = pl$col("x") <= pl$lit(2)
)
}
