% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__len}
\alias{expr__len}
\title{Return the number of elements in the column}
\usage{
expr__len()
}
\value{
A polars \link{expression}
}
\description{
Null values are counted in the total.
}
\examples{
df <- pl$DataFrame(a = 1:3, b = c(NA, 4, 4))
df$select(pl$all()$len())
}
