% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__log}
\alias{expr__log}
\title{Compute the logarithm}
\usage{
expr__log(base = exp(1))
}
\arguments{
\item{base}{Numeric value used as base, defaults to \code{exp(1)}.}
}
\value{
A polars \link{expression}
}
\description{
Compute the logarithm
}
\examples{
pl$DataFrame(a = c(1, 2, 4))$
  with_columns(
  log = pl$col("a")$log(),
  log_base_2 = pl$col("a")$log(base = 2)
)
}
