% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__map_batches}
\alias{expr__map_batches}
\title{Apply a custom R function to a whole Series or sequence of Series.}
\usage{
expr__map_batches(
  lambda,
  return_dtype = NULL,
  ...,
  is_elementwise = FALSE,
  returns_scalar = FALSE
)
}
\arguments{
\item{lambda}{Function to apply.}

\item{return_dtype}{Dtype of the output Series. Can be a \link[=polars_dtype]{DataType}
or a DataTypeExpr, such as one returned by \code{\link[=pl__dtype_of]{pl$dtype_of()}}.
It is recommended to set this whenever possible. If this is \code{NULL}, it tries
to infer the datatype by calling the function with dummy data and looking at
the output.}

\item{...}{These dots are for future extensions and must be empty.}

\item{is_elementwise}{Whether the function can be applied element by element.
If \code{TRUE}, the function may be optimized by the query engine. Defaults to
\code{FALSE}.}

\item{returns_scalar}{Whether the function returns a single scalar value.
Defaults to \code{FALSE}.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The output of this custom function is presumed to be either a Series, or an
R vector that will be converted into a Series by \code{\link[=as_polars_series]{as_polars_series()}}.
By default, the function is applied to the complete Series, rather than
element by element. Set \code{is_elementwise} to \code{TRUE} when the function can be
safely applied element by element.
}
\examples{
df <- pl$DataFrame(
  sine = c(0.0, 1.0, 0.0, -1.0),
  cosine = c(1.0, 0.0, -1.0, 0.0)
)
df$select(pl$all()$map_batches(\\(x) \{
  x$to_r_vector() |>
    which.max()
\}))

# Call a function that takes multiple arguments by creating a struct and
# referencing its fields inside the function call.
df <- pl$DataFrame(
  a = c(5, 1, 0, 3),
  b = c(4, 2, 3, 4),
)
df$with_columns(
  a_times_b = pl$struct("a", "b")$map_batches(
    \\(x) x$struct$field("a") * x$struct$field("b")
  )
)
}
