% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__max}
\alias{expr__max}
\title{Get the maximum value}
\usage{
expr__max()
}
\value{
A polars \link{expression}
}
\description{
Get the maximum value
}
\examples{
pl$DataFrame(x = c(1, NaN, 3))$
  with_columns(max = pl$col("x")$max())
}
