% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__mean}
\alias{expr__mean}
\title{Get mean value}
\usage{
expr__mean()
}
\value{
A polars \link{expression}
}
\description{
Get mean value
}
\examples{
pl$DataFrame(x = c(1, 3, 4, NA))$
  with_columns(mean = pl$col("x")$mean())
}
