% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__median}
\alias{expr__median}
\title{Get median value}
\usage{
expr__median()
}
\value{
A polars \link{expression}
}
\description{
Get median value
}
\examples{
pl$DataFrame(x = c(1, 3, 4, NA))$
  with_columns(median = pl$col("x")$median())
}
