% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__min_by}
\alias{expr__min_by}
\title{Get minimum value, ordered by another expression}
\usage{
expr__min_by(by)
}
\arguments{
\item{by}{Column used to determine the smallest element. Accepts expression
input. Strings are parsed as column names.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(a = c(-1, NaN, 1), b = c("x", "y", "z"))
df$with_columns(min_b_by_a = pl$col("b")$min_by("a"))
}
