% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__mode}
\alias{expr__mode}
\title{Compute the most occurring value(s)}
\usage{
expr__mode(..., maintain_order = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{maintain_order}{If \code{TRUE}, maintain order of data.
This requires more work.}
}
\value{
A polars \link{expression}
}
\description{
Compute the most occurring value(s)
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2, 3), b = c(1, 1, 2, 2))
df$select(pl$col("a")$mode())
df$select(pl$col("b")$mode())
}
