% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__mul}
\alias{expr__mul}
\title{Multiply two expressions}
\usage{
expr__mul(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of multiplication operator \code{expr * other}.
}
\examples{
df <- pl$DataFrame(x = c(1, 2, 4, 8, 16))

df$with_columns(
  `x*2` = pl$col("x")$mul(2),
  `x * xlog2` = pl$col("x")$mul(pl$col("x")$log(2))
)
}
\seealso{
\itemize{
\item \link[=s3-arithmetic]{Arithmetic operators}
}
}
