% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__n_unique}
\alias{expr__n_unique}
\title{Count unique values}
\usage{
expr__n_unique()
}
\value{
A polars \link{expression}
}
\description{
\code{null} is considered to be a unique value for the purposes of this operation.
}
\examples{
df <- pl$DataFrame(
  x = c(1, 1, 2, 2, 3),
  y = c(1, 1, 1, NA, NA)
)
df$select(
  x_unique = pl$col("x")$n_unique(),
  y_unique = pl$col("y")$n_unique()
)
}
