% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__nan_min}
\alias{expr__nan_min}
\title{Get the minimum value with NaN}
\usage{
expr__nan_min()
}
\value{
A polars \link{expression}
}
\description{
This returns \code{NaN} if there are any.
}
\examples{
pl$DataFrame(x = c(1, NA, 3, NaN, Inf))$
  with_columns(nan_min = pl$col("x")$nan_min())
}
