% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__ne}
\alias{expr__ne}
\title{Check inequality}
\usage{
expr__ne(other)
}
\arguments{
\item{other}{A literal or expression value to compare with.}
}
\value{
A polars \link{expression}
}
\description{
This propagates null values, i.e. any comparison involving \code{null} will
return \code{null}. Use \code{\link[=expr__ne_missing]{$ne_missing()}} to consider null
values as equal.
}
\examples{
df <- pl$DataFrame(x = c(NA, FALSE, TRUE), y = c(TRUE, TRUE, TRUE))
df$with_columns(
  ne = pl$col("x")$ne(pl$col("y")),
  ne_missing = pl$col("x")$ne_missing(pl$col("y"))
)
}
\seealso{
\link{expr__ne_missing}
}
