% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__pct_change}
\alias{expr__pct_change}
\title{Computes percentage change between values}
\usage{
expr__pct_change(n = 1)
}
\arguments{
\item{n}{Integer or Expr indicating the number of periods to shift for
forming percent change.}
}
\value{
A polars \link{expression}
}
\description{
Computes the percentage change (as fraction) between current element and
most-recent non-null element at least \code{n} period(s) before the current
element. By default it computes the change from the previous row.
}
\examples{
df <- pl$DataFrame(a = c(10:12, NA, 12))
df$with_columns(
  pct_change = pl$col("a")$pct_change()
)
}
