% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__peak_max}
\alias{expr__peak_max}
\title{Get a boolean mask of the local maximum peaks}
\usage{
expr__peak_max()
}
\value{
A polars \link{expression}
}
\description{
Get a boolean mask of the local maximum peaks
}
\examples{
df <- pl$DataFrame(x = c(1, 2, 3, 2, 3, 4, 5, 2))
df$with_columns(peak_max = pl$col("x")$peak_max())
}
