% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__peak_min}
\alias{expr__peak_min}
\title{Get a boolean mask of the local minimum peaks}
\usage{
expr__peak_min()
}
\value{
A polars \link{expression}
}
\description{
Get a boolean mask of the local minimum peaks
}
\examples{
df <- pl$DataFrame(x = c(1, 2, 3, 2, 3, 4, 5, 2))
df$with_columns(peak_min = pl$col("x")$peak_min())
}
