% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__pow}
\alias{expr__pow}
\title{Exponentiation using two expressions}
\usage{
expr__pow(exponent)
}
\arguments{
\item{exponent}{Numeric literal or expression value.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of exponentiation operator \code{expr ^ exponent}.
}
\examples{
df <- pl$DataFrame(x = c(1, 2, 4, 8))

df$with_columns(
  cube = pl$col("x")$pow(3),
  `x^xlog2` = pl$col("x")$pow(pl$col("x")$log(2))
)
}
\seealso{
\itemize{
\item \link[=s3-arithmetic]{Arithmetic operators}
}
}
