% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__product}
\alias{expr__product}
\title{Compute the product of an expression.}
\usage{
expr__product()
}
\value{
A polars \link{expression}
}
\description{
Compute the product of an expression.
}
\examples{
pl$DataFrame(a = 1:3, b = c(NA, 4, 4))$
  select(pl$all()$product())
}
