% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__radians}
\alias{expr__radians}
\title{Convert from degrees to radians}
\usage{
expr__radians()
}
\value{
A polars \link{expression}
}
\description{
Convert from degrees to radians
}
\examples{
pl$DataFrame(a = c(-720, -540, -360, -180, 0, 180, 360, 540, 720))$
  with_columns(radians = pl$col("a")$radians())
}
