% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__repeat_by}
\alias{expr__repeat_by}
\title{Repeat the elements in this Series as specified in the given expression}
\usage{
expr__repeat_by(by)
}
\arguments{
\item{by}{Numeric column that determines how often the values will be
repeated. The column will be coerced to UInt32. Give this dtype to make the
coercion a no-op. Accepts expression input, strings are parsed as column
names.}
}
\value{
A polars \link{expression}
}
\description{
The repeated elements are expanded into a List dtype.
}
\examples{
df <- pl$DataFrame(a = c("x", "y", "z"), n = 1:3)

df$with_columns(
  repeated = pl$col("a")$repeat_by("n")
)
}
