% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rle}
\alias{expr__rle}
\title{Compress the column data using run-length encoding}
\usage{
expr__rle()
}
\value{
A polars \link{expression}
}
\description{
Run-length encoding (RLE) encodes data by storing each run of identical
values as a single value and its length.
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2, 1, NA, 1, 3, 3))

df$select(pl$col("a")$rle())$unnest("a")
}
