% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rle_id}
\alias{expr__rle_id}
\title{Get a distinct integer ID for each run of identical values}
\usage{
expr__rle_id()
}
\value{
A polars \link{expression}
}
\description{
The ID starts at 0 and increases by one each time the value of the column
changes.
}
\details{
This functionality is especially useful for defining a new group for every
time a column’s value changes, rather than for every distinct value of that
column.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, 1, 1, 1),
  b = c("x", "x", NA, "y", "y")
)

df$with_columns(
  rle_id_a = pl$col("a")$rle_id(),
  rle_id_ab = pl$struct("a", "b")$rle_id()
)
}
