% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rolling_kurtosis}
\alias{expr__rolling_kurtosis}
\title{Compute a rolling kurtosis}
\usage{
expr__rolling_kurtosis(
  window_size,
  ...,
  fisher = TRUE,
  bias = TRUE,
  min_samples = NULL,
  center = FALSE
)
}
\arguments{
\item{window_size}{The length of the window in number of elements.}

\item{...}{These dots are for future extensions and must be empty.}

\item{fisher}{If \code{TRUE} (default), Fisher’s definition is used
(normal ==> 0.0). If \code{FALSE}, Pearson’s definition is used (normal ==> 3.0).}

\item{bias}{If \code{FALSE}, the calculations are corrected for statistical bias.}

\item{min_samples}{The number of values in the window that should be
non-null before computing a result. If \code{NULL} (default), it will be set
equal to \code{window_size}.}

\item{center}{If \code{TRUE}, set the labels at the center of the window.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The window at a given row will include the row itself, and the
\code{window_size - 1} elements before it.
}
\examples{
df <- pl$DataFrame(a = c(1, 4, 2, 9))

df$with_columns(
  rolling_kurtosis = pl$col("a")$rolling_kurtosis(window_size = 3)
)

# Center the values in the window
df$with_columns(
  rolling_kurtosis = pl$col("a")$rolling_kurtosis(window_size = 3, center = TRUE)
)
}
