% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rolling_rank}
\alias{expr__rolling_rank}
\title{Apply a rolling rank over values}
\usage{
expr__rolling_rank(
  window_size,
  method = c("average", "min", "max", "dense", "random"),
  ...,
  seed = NULL,
  min_samples = NULL,
  center = FALSE
)
}
\arguments{
\item{window_size}{The length of the window in number of elements.}

\item{method}{The method used to assign ranks to tied elements. Must be one
of the following:
\itemize{
\item \code{"average"} (default): The average of the ranks that would have been
assigned to all the tied values is assigned to each value.
\item \code{"min"}: The minimum of the ranks that would have been assigned to all
the tied values is assigned to each value. (This is also referred to
as "competition" ranking.)
\item \code{"max"}: The maximum of the ranks that would have been assigned to all
the tied values is assigned to each value.
\item \code{"dense"}: Like \code{"min"}, but the rank of the next highest element is
assigned the rank immediately after those assigned to the tied
elements.
\item \code{"random"}: Choose a random rank for each value in a tie.
}}

\item{...}{These dots are for future extensions and must be empty.}

\item{seed}{Random seed used when \code{method = "random"}. If \code{NULL} (default), a
random seed is generated for each rolling rank operation.}

\item{min_samples}{The number of values in the window that should be
non-null before computing a result. If \code{NULL} (default), it will be set
equal to \code{window_size}.}

\item{center}{If \code{TRUE}, set the labels at the center of the window.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A window of length \code{window_size} will traverse the array. The values that
fill this window will be ranked according to the \code{method} parameter. The
resulting values will be the rank of the value that is at the end of the
sliding window.
}
\details{
If you want to compute multiple aggregation statistics over the same dynamic
window, consider using \code{\link[=expr__rolling]{$rolling()}} - this method can cache
the window size computation.
}
\examples{
df <- pl$DataFrame(a = c(1, 4, 4, 1, 9))

df$select(pl$col("a")$rolling_rank(3, method = "average"))
}
