% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rolling_skew}
\alias{expr__rolling_skew}
\title{Apply a rolling skew over values}
\usage{
expr__rolling_skew(
  window_size,
  ...,
  bias = TRUE,
  min_samples = NULL,
  center = FALSE
)
}
\arguments{
\item{window_size}{The length of the window in number of elements.}

\item{...}{These dots are for future extensions and must be empty.}

\item{bias}{If \code{FALSE}, the calculations are corrected for statistical bias.}

\item{min_samples}{The number of values in the window that should be non-null before computing
a result. If set to \code{NULL} (default), it will be set equal to \code{window_size}.}

\item{center}{Set the labels at the center of the window.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A window of length \code{window_size} will traverse the array. The values that
fill this window will (optionally) be multiplied with the weights given by
the \code{weights} vector. The resulting values will be aggregated.

The window at a given row will include the row itself, and the
\code{window_size - 1} elements before it.
}
\details{
If you want to compute multiple aggregation statistics over the same dynamic
window, consider using \code{\link[=expr__rolling]{$rolling()}} - this method can cache
the window size computation.
}
\examples{
df <- pl$DataFrame(a = c(1, 4, 2, 9))
df$with_columns(
  rolling_skew = pl$col("a")$rolling_skew(3)
)
}
