% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__shrink_dtype}
\alias{expr__shrink_dtype}
\title{Shrink numeric columns to the minimal required datatype}
\usage{
expr__shrink_dtype()
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated as of polars 1.3.0 and turned into a no-op.
Use \code{\link[=series__shrink_dtype]{<series>$shrink_dtype}} instead.
}
