% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sign}
\alias{expr__sign}
\title{Compute the sign}
\usage{
expr__sign()
}
\value{
A polars \link{expression}
}
\description{
This returns -1 if x is lower than 0, 0 if x == 0, and 1 if x is greater
than 0.
}
\examples{
df <- pl$DataFrame(a = c(-9, 0, 0, 4, NA))
df$with_columns(sign = pl$col("a")$sign())
}
