% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sinh}
\alias{expr__sinh}
\title{Compute hyperbolic sine}
\usage{
expr__sinh()
}
\value{
A polars \link{expression}
}
\description{
Compute hyperbolic sine
}
\examples{
pl$DataFrame(a = c(-1, asinh(0.5), 0, 1, NA))$
  with_columns(sinh = pl$col("a")$sinh())
}
