% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sort}
\alias{expr__sort}
\title{Sort this expression}
\usage{
expr__sort(..., descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{descending}{Sort in descending order.}

\item{nulls_last}{Place null values last.}
}
\value{
A polars \link{expression}
}
\description{
If used in a groupby context, values within each group are sorted.
}
\examples{
df <- pl$DataFrame(a = c(6, 1, 0, NA, Inf, NaN))

df$with_columns(
  sorted = pl$col("a")$sort(),
  sorted_desc = pl$col("a")$sort(descending = TRUE),
  sorted_nulls_last = pl$col("a")$sort(nulls_last = TRUE)
)

# When sorting in a group by context, values in each group are sorted.
df <- pl$DataFrame(
  group = c("one", "one", "one", "two", "two", "two"),
  value = c(1, 98, 2, 3, 99, 4)
)

df$group_by("group")$agg(pl$col("value")$sort())
}
