% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__std}
\alias{expr__std}
\title{Compute the standard deviation}
\usage{
expr__std(ddof = 1)
}
\arguments{
\item{ddof}{"Delta Degrees of Freedom": the divisor used in the calculation
is \code{N - ddof}, where \code{N} represents the number of elements. By default
\code{ddof} is 1.}
}
\value{
A polars \link{expression}
}
\description{
Compute the standard deviation
}
\examples{
pl$DataFrame(a = c(1, 3, 5, 6))$
  select(pl$all()$std())
}
