% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sum}
\alias{expr__sum}
\title{Get sum value}
\usage{
expr__sum()
}
\value{
A polars \link{expression}
}
\description{
Get sum value
}
\details{
The dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(x = c(1L, NA, 2L))$
  with_columns(sum = pl$col("x")$sum())
}
