% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__tanh}
\alias{expr__tanh}
\title{Compute hyperbolic tangent}
\usage{
expr__tanh()
}
\value{
A polars \link{expression}
}
\description{
Compute hyperbolic tangent
}
\examples{
pl$DataFrame(a = c(-1, atanh(0.5), 0, 1, NA))$
  with_columns(tanh = pl$col("a")$tanh())
}
