% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__unique}
\alias{expr__unique}
\title{Get unique values}
\usage{
expr__unique(..., maintain_order = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{maintain_order}{Maintain order of data. This requires more work.}
}
\value{
A polars \link{expression}
}
\description{
This method differs from \code{\link[=expr__value_counts]{$value_counts()}} in that it
does not return the values, only the counts and might be faster.
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2))
df$select(pl$col("a")$unique())
}
