% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__unique_counts}
\alias{expr__unique_counts}
\title{Count unique values in the order of appearance}
\usage{
expr__unique_counts()
}
\value{
A polars \link{expression}
}
\description{
This method differs from \code{\link[=expr__value_counts]{$value_counts()}} in that it
does not return the values, only the counts and might be faster.
}
\examples{
df <- pl$DataFrame(id = c("a", "b", "b", "c", "c", "c"))
df$select(pl$col("id")$unique_counts())
}
