% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__var}
\alias{expr__var}
\title{Compute the variance}
\usage{
expr__var(ddof = 1)
}
\arguments{
\item{ddof}{"Delta Degrees of Freedom": the divisor used in the calculation
is \code{N - ddof}, where \code{N} represents the number of elements. By default
\code{ddof} is 1.}
}
\value{
A polars \link{expression}
}
\description{
Compute the variance
}
\examples{
pl$DataFrame(a = c(1, 3, 5, 6))$
  select(pl$all()$var())
}
