% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_agg_groups}
\alias{Expr_agg_groups}
\title{aggregate groups}
\usage{
Expr_agg_groups()
}
\value{
Exprs
}
\description{
aggregate groups
}
\examples{
df = pl$DataFrame(list(
  group = c("one","one","one","two","two","two"),
  value =  c(94, 95, 96, 97, 97, 99)
))
df$groupby("group", maintain_order=TRUE)$agg(pl$col("value")$agg_groups())
}
\keyword{Expr}
\keyword{Get}
\keyword{Should}
\keyword{aggregation}
\keyword{be}
\keyword{by}
\keyword{context}
\keyword{group}
\keyword{in}
\keyword{indexes}
\keyword{of}
\keyword{only.}
\keyword{operation.}
\keyword{the}
\keyword{used}
