% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_alias}
\alias{Expr_alias}
\title{Rename Expr output}
\usage{
Expr_alias
}
\arguments{
\item{name}{New name of output}
}
\value{
Expr
}
\description{
Rename the output of an expression.
}
\examples{
pl$col("bob")$alias("alice")
}
\keyword{datasets}
