% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_and}
\alias{Expr_and}
\title{Apply logical AND on two expressions}
\usage{
Expr_and(other)
}
\arguments{
\item{other}{Literal or object that can be converted to a literal}
}
\value{
Expr
}
\description{
Combine two boolean expressions with AND.
}
\examples{
pl$lit(TRUE) & TRUE
pl$lit(TRUE)$and(pl$lit(TRUE))
}
