% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_append}
\alias{Expr_append}
\title{Append expressions}
\usage{
Expr_append(other, upcast = TRUE)
}
\arguments{
\item{other}{Expr or something coercible to an Expr.}

\item{upcast}{Cast both Expr to a common supertype if they have one.}
}
\value{
Expr
}
\description{
This is done by adding the chunks of \code{other} to this \code{output}.
}
\examples{
# append bottom to to row
df = pl$DataFrame(list(a = 1:3, b = c(NA_real_, 4, 5)))
df$select(pl$all()$head(1)$append(pl$all()$tail(1)))

# implicit upcast, when default = TRUE
pl$DataFrame(list())$select(pl$lit(42)$append(42L))
pl$DataFrame(list())$select(pl$lit(42)$append(FALSE))
pl$DataFrame(list())$select(pl$lit("Bob")$append(FALSE))
}
