% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_apply_in_background}
\alias{Expr_apply_in_background}
\title{Expr_apply_in_background}
\usage{
Expr_apply_in_background(f, output_type = NULL)
}
\arguments{
\item{f}{a function mapping series in each group}

\item{output_type}{NULL or one of pl$dtypes$..., the output datatype, NULL is the same as input.}
}
\value{
Expr
}
\description{
Expr_apply_in_background
}
\details{
user function return should be a series or any Robj convertible into a Series.
In PyPolars likely return must be Series. User functions do fully support \code{browser()}, helpful to
investigate.
}
