% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_approx_unique}
\alias{Expr_approx_unique}
\alias{approx_unique}
\title{Approx count unique values}
\usage{
Expr_approx_unique
}
\value{
Expr
}
\description{
This is done using the HyperLogLog++ algorithm for cardinality estimation.
}
\examples{
pl$DataFrame(iris)$select(pl$col("Species")$approx_unique())
}
\keyword{Expr}
