% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arccos}
\alias{Expr_arccos}
\alias{arccos}
\title{Arccos}
\usage{
Expr_arccos
}
\value{
Expr
}
\description{
Compute the element-wise value for the inverse cosine.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a = c(-1, cos(0.5), 0, 1, NA_real_))$select(pl$col("a")$arccos())
}
\keyword{datasets}
