% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arccosh}
\alias{Expr_arccosh}
\alias{arccosh}
\title{Arccosh}
\format{
Method
}
\usage{
Expr_arccosh
}
\value{
Expr
}
\description{
Compute the element-wise value for the inverse hyperbolic cosine.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a=c(-1,cosh(0.5),0,1,NA_real_))$select(pl$col("a")$arccosh())
}
\keyword{Expr}
