% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arcsin}
\alias{Expr_arcsin}
\title{Compute inverse sine}
\usage{
Expr_arcsin
}
\value{
Expr
}
\description{
Compute inverse sine
}
\examples{
pl$DataFrame(a = c(-1, sin(0.5), 0, 1, NA_real_))$
  with_columns(arcsin = pl$col("a")$arcsin())
}
\keyword{datasets}
