% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arcsinh}
\alias{Expr_arcsinh}
\title{Compute inverse hyperbolic sine}
\usage{
Expr_arcsinh()
}
\value{
Expr
}
\description{
Compute inverse hyperbolic sine
}
\examples{
pl$DataFrame(a = c(-1, sinh(0.5), 0, 1, NA_real_))$
  with_columns(arcsinh = pl$col("a")$arcsinh())
}
