% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arctan}
\alias{Expr_arctan}
\title{Compute inverse tangent}
\usage{
Expr_arctan()
}
\value{
Expr
}
\description{
Compute inverse tangent
}
\examples{
pl$DataFrame(a = c(-1, tan(0.5), 0, 1, NA_real_))$
  with_columns(arctan = pl$col("a")$arctan())
}
