% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arctanh}
\alias{Expr_arctanh}
\title{Compute inverse hyperbolic tangent}
\usage{
Expr_arctanh
}
\value{
Expr
}
\description{
Compute inverse hyperbolic tangent
}
\examples{
pl$DataFrame(a = c(-1, tanh(0.5), 0, 1, NA_real_))$
  with_columns(arctanh = pl$col("a")$arctanh())
}
\keyword{datasets}
