% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_argsort}
\alias{Expr_argsort}
\title{Index of a sort}
\usage{
Expr_argsort(descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{descending}{Sort in descending order. When sorting by multiple columns,
can be specified per column by passing a vector of booleans.}

\item{nulls_last}{If \code{TRUE}, place nulls values last.}
}
\description{
argsort is a alias for arg_sort
}
\examples{
pl$DataFrame(
  a = c(6, 1, 0, NA, Inf, NaN)
)$with_columns(arg_sorted = pl$col("a")$arg_sort())
}
