% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arr}
\alias{Expr_arr}
\alias{arr_ns}
\title{arr: list related methods DEPRECATED}
\usage{
Expr_arr()
}
\value{
Expr
}
\description{
DEPRECATED FROM 0.9.0 USE \verb{<Expr>$list$...} instead. Subnamespace is simple renamed.
}
\seealso{
\code{\link[=Expr_list]{<Expr>$list$...}}
}
\keyword{Expr}
