% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_all}
\alias{expr_arr_all}
\title{Evaluate whether all boolean values are true for every sub-array}
\usage{
expr_arr_all(..., ignore_nulls = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{ignore_nulls}{If \code{TRUE} (default), ignore null values. If \code{FALSE},
\href{https://en.wikipedia.org/wiki/Three-valued_logic}{Kleene logic} is used to
deal with nulls: if the column contains any null values and no \code{FALSE} values,
the output is null.}
}
\value{
A polars \link{expression}
}
\description{
Evaluate whether all boolean values are true for every sub-array
}
\examples{
df <- pl$DataFrame(
  values = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), c(NA, NA)),
)$cast(pl$Array(pl$Boolean, 2))
df$with_columns(all = pl$col("values")$arr$all())
}
