% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_arg_min}
\alias{expr_arr_arg_min}
\title{Retrieve the index of the minimum value in every sub-array}
\usage{
expr_arr_arg_min()
}
\value{
A polars \link{expression}
}
\description{
Retrieve the index of the minimum value in every sub-array
}
\examples{
df <- pl$DataFrame(
  values = list(1:2, 2:1)
)$cast(pl$Array(pl$Int32, 2))
df$with_columns(
  arg_min = pl$col("values")$arr$arg_min()
)
}
