% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_get}
\alias{expr_arr_get}
\title{Get the value by index in every sub-array}
\usage{
expr_arr_get(index, ..., null_on_oob = TRUE)
}
\arguments{
\item{index}{An Expr or something coercible to an Expr, that must return a
single index.}

\item{...}{These dots are for future extensions and must be empty.}

\item{null_on_oob}{If \code{TRUE}, return \code{null} if an index is out of bounds.
Otherwise, raise an error.}
}
\value{
A polars \link{expression}
}
\description{
This allows to extract one value per array only. Values are 0-indexed (so
index \code{0} would return the first item of every sub-array) and negative values
start from the end (so index \code{-1} returns the last item).
In R, out-of-bounds indices return \code{null} by default; pass
\code{null_on_oob = FALSE} to raise an error.
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(NA, 6)),
  idx = c(1, NA, 3)
)$cast(values = pl$Array(pl$Float64, 2))
df$with_columns(
  using_expr = pl$col("values")$arr$get("idx"),
  val_0 = pl$col("values")$arr$get(0),
  val_minus_1 = pl$col("values")$arr$get(-1),
  val_oob = pl$col("values")$arr$get(10)
)
}
