% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_last}
\alias{expr_arr_last}
\title{Get the last value of the sub-arrays}
\usage{
expr_arr_last()
}
\value{
A polars \link{expression}
}
\description{
Get the last value of the sub-arrays
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 2, 3), c(4, 5, 6))
)$cast(pl$Array(pl$Int64, 3))
df$with_columns(last = pl$col("a")$arr$last())
}
