% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_len}
\alias{expr_arr_len}
\title{Return the number of elements in each sub-array}
\usage{
expr_arr_len()
}
\value{
A polars \link{expression}
}
\description{
Return the number of elements in each sub-array
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 1, 2), c(2, 3, 4))
)$cast(pl$Array(pl$Int64, 3))
df$with_columns(len = pl$col("a")$arr$len())
}
