% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_median}
\alias{expr_arr_median}
\title{Compute the median value of the sub-arrays}
\usage{
expr_arr_median()
}
\value{
A polars \link{expression}
}
\description{
Compute the median value of the sub-arrays
}
\examples{
df <- pl$DataFrame(
  values = list(c(2, 1, 4), c(8.4, 3.2, 1)),
)$cast(pl$Array(pl$Float64, 3))
df$with_columns(median = pl$col("values")$arr$median())
}
