% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_min}
\alias{expr_arr_min}
\title{Compute the min value of the sub-arrays}
\usage{
expr_arr_min()
}
\value{
A polars \link{expression}
}
\description{
Compute the min value of the sub-arrays
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(NA, NA))
)$cast(pl$Array(pl$Float64, 2))
df$with_columns(min = pl$col("values")$arr$min())
}
