% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_backward_fill}
\alias{Expr_backward_fill}
\alias{backward_fill}
\title{Fill Nulls Backward}
\format{
a method
}
\usage{
Expr_backward_fill(limit = NULL)
}
\arguments{
\item{limit}{Expr or \verb{Into<Expr>}  The number of consecutive null values to backward fill.}
}
\value{
Expr
}
\description{
Fill missing values with the next to be seen values.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[polars]{docs_translations}}
}
\examples{
l = list(a=c(1L,rep(NA_integer_,3L),10))
pl$DataFrame(l)$select(
  pl$col("a")$backward_fill()$alias("bf_null"),
  pl$col("a")$backward_fill(limit = 0)$alias("bf_l0"),
  pl$col("a")$backward_fill(limit = 1)$alias("bf_l1")
)$to_list()
}
\keyword{Expr}
