% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_bin}
\alias{Expr_bin}
\alias{bin_ns}
\title{bin: binary related methods}
\usage{
Expr_bin()
}
\value{
Expr
}
\description{
Create an object namespace of all binary related methods.
See the individual method pages for full details
}
\examples{

# missing

}
\keyword{Expr}
