% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-binary.R
\name{expr_bin_ends_with}
\alias{expr_bin_ends_with}
\title{Check if string values end with a binary substring}
\usage{
expr_bin_ends_with(suffix)
}
\arguments{
\item{suffix}{Suffix substring.}
}
\value{
A polars \link{expression}
}
\description{
Check if string values end with a binary substring
}
\examples{
colors <- pl$DataFrame(
  name = c("black", "yellow", "blue"),
  code = as_polars_series(c("x00x00x00", "xffxffx00", "x00x00xff"))$cast(pl$Binary),
  suffix = as_polars_series(c("x00", "xffx00", "xffxff"))$cast(pl$Binary)
)

colors$select(
  "name",
  ends_with_lit = pl$col("code")$bin$ends_with("xff"),
  ends_with_expr = pl$col("code")$bin$ends_with(pl$col("suffix"))
)
}
