% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-binary.R
\name{expr_bin_reinterpret}
\alias{expr_bin_reinterpret}
\title{Interpret bytes as another type}
\usage{
expr_bin_reinterpret(..., dtype, endianness = c("little", "big"))
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{dtype}{A Polars DataType or DataTypeExpr indicating which type to
interpret binary column into.}

\item{endianness}{Which endianness to use when interpreting bytes. Must be
one of \code{"little"} (default) or \code{"big"}.}
}
\value{
A polars \link{expression}
}
\description{
Supported types are numerical or temporal dtypes, or an \code{\link[=pl__Array]{Array}}
of these dtypes.
}
\examples{
\dontshow{if (\{
requireNamespace("blob", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
df <- pl$DataFrame(x = blob::as_blob(c(5L, 35L)))

df$with_columns(
  bin2int = pl$col("x")$bin$reinterpret(
    dtype = pl$UInt8,
    endianness = "little"
  )
)
\dontshow{\}) # examplesIf}
}
