% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cast}
\alias{Expr_cast}
\alias{cast}
\title{Cast between DataType(s)}
\usage{
Expr_cast(dtype, strict = TRUE)
}
\arguments{
\item{dtype}{DataType to cast to.}

\item{strict}{bool if true an error will be thrown if cast failed at resolve time.}
}
\value{
Expr
}
\description{
Cast between DataType(s)
}
\examples{
df = pl$DataFrame(list(a = 1:3, b = 1:3))
df$with_columns(
  pl$col("a")$cast(pl$dtypes$Float64, TRUE)$alias("f64"),
  pl$col("a")$cast(pl$dtypes$Int32, TRUE)$alias("i32")
)
}
\keyword{Expr}
