% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cat}
\alias{Expr_cat}
\alias{cat_ns}
\title{cat: related methods}
\usage{
Expr_cat()
}
\value{
Expr
}
\description{
Create an object namespace of all cat related methods.
See the individual method pages for full details
}
\examples{

#missing

}
\keyword{Expr}
