% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-categorical.R
\name{expr_cat_get_categories}
\alias{expr_cat_get_categories}
\title{Get the categories stored in this data type}
\usage{
expr_cat_get_categories()
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\verb{$cat$get_categories()} is deprecated. To get the distinct values present
in a Categorical column, use \verb{$unique()}. For the fixed category list of
an Enum, use its \code{dtype$categories}.
}
\examples{
df <- pl$DataFrame(
  cats = factor(c("z", "z", "k", "a", "b")),
  vals = factor(c(3, 1, 2, 2, 3))
)
df

df$select(
  pl$col("cats")$cat$get_categories()
)
df$select(
  pl$col("vals")$cat$get_categories()
)
}
