% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-categorical.R
\name{expr_cat_set_ordering}
\alias{expr_cat_set_ordering}
\title{Set Ordering}
\usage{
expr_cat_set_ordering(ordering)
}
\arguments{
\item{ordering}{string either 'physical' or 'lexical'
\itemize{
\item \code{"physical"}: use the physical representation of the categories to
determine the order (default).
\item \code{"lexical"}: use the string values to determine the order.
}}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Determine how this categorical series should be sorted.
}
