% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_ceil}
\alias{Expr_ceil}
\title{Ceiling}
\usage{
Expr_ceil
}
\value{
Expr
}
\description{
Rounds up to the nearest integer value.
Only works on floating point Series.
}
\examples{
pl$DataFrame(list(
  a = c(0.33, 0.5, 1.02, 1.5, NaN , NA, Inf, -Inf)
))$select(
  pl$col("a")$ceil()
)
}
\keyword{Expr}
