% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_clip}
\alias{Expr_clip}
\title{Clip elements}
\usage{
Expr_clip(lower_bound = NULL, upper_bound = NULL)
}
\arguments{
\item{lower_bound}{Lower bound. Accepts expression input. Strings are parsed
as column names and other non-expression inputs are parsed as literals.}

\item{upper_bound}{Upper bound. Accepts expression input. Strings are parsed
as column names and other non-expression inputs are parsed as literals.}
}
\value{
Expr
}
\description{
Set values outside the given boundaries to the boundary value. This only
works for numeric and temporal values.
}
\examples{
df = pl$DataFrame(foo = c(-50L, 5L, NA_integer_, 50L), bound = c(1, 10, 1, 1))

# With the two bounds
df$with_columns(clipped = pl$col("foo")$clip(1, 10))

# Without lower bound
df$with_columns(clipped = pl$col("foo")$clip(upper_bound = 10))

# Using another column as lower bound
df$with_columns(clipped = pl$col("foo")$clip(lower_bound = "bound"))
}
