% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_clip_max}
\alias{Expr_clip_max}
\title{Clip elements above maximum value}
\usage{
<Expr>$clip_max(max)
}
\arguments{
\item{max}{Maximum value, Expr returning a numeric.}
}
\description{
Replace all values above a maximum value by this maximum value.
}
\examples{
pl$DataFrame(foo = c(-50L, 5L, NA_integer_, 50L))$
  with_columns(clipped = pl$col("foo")$clip_max(10))
}
