% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_clip_min}
\alias{Expr_clip_min}
\title{Clip elements below minimum value}
\usage{
<Expr>$clip_min(min)
}
\arguments{
\item{min}{Minimum value, Expr returning a numeric.}
}
\description{
Replace all values below a minimum value by this minimum value.
}
\examples{
pl$DataFrame(foo = c(-50L, 5L, NA_integer_, 50L))$
  with_columns(clipped = pl$col("foo")$clip_min(1))
}
