% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cos}
\alias{Expr_cos}
\title{Compute cosine}
\usage{
Expr_cos()
}
\value{
Expr
}
\description{
Compute cosine
}
\examples{
pl$DataFrame(a = c(0, pi / 2, pi, NA_real_))$
  with_columns(cosine = pl$col("a")$cos())
}
