% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cosh}
\alias{Expr_cosh}
\alias{cosh}
\title{Cosh}
\format{
Method
}
\usage{
Expr_cosh
}
\value{
Expr
}
\description{
Compute the element-wise value for the hyperbolic cosine.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a=c(-1,acosh(1.5),0,1,NA_real_))$select(pl$col("a")$cosh())
}
\keyword{Expr}
