% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_count}
\alias{Expr_count}
\alias{Expr_len}
\title{Count values (len is a alias)}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
Expr_count

Expr_len
}
\value{
Expr

Expr
}
\description{
Count the number of values in this expression.
Similar to R length()
}
\examples{
pl$DataFrame(list(all=c(T,T),any=c(T,F),none=c(F,F)))$select(pl$all()$count())
#same as
pl$DataFrame(list(all=c(T,T),any=c(T,F),none=c(F,F)))$select(pl$all()$len())
}
\keyword{Expr}
