% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cumcount}
\alias{Expr_cumcount}
\title{Cumulative count}
\usage{
Expr_cumcount(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, reverse the count.}
}
\value{
Expr
}
\description{
Get an array with the cumulative count (zero-indexed) computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.

\verb{$cumcount()} does not seem to count within lists.
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  pl$col("a")$cumcount()$alias("cumcount"),
  pl$col("a")$cumcount(reverse = TRUE)$alias("cumcount_reversed")
)
}
