% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cummax}
\alias{Expr_cummax}
\title{Cumulative maximum}
\usage{
Expr_cummax(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, start from the last value.}
}
\value{
Expr
}
\description{
Get an array with the cumulative max computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(a = c(1:4, 2L))$with_columns(
  pl$col("a")$cummax()$alias("cummux"),
  pl$col("a")$cummax(reverse = TRUE)$alias("cummax_reversed")
)
}
