% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cumprod}
\alias{Expr_cumprod}
\title{Cumulative product}
\usage{
Expr_cumprod(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, start with the total product of elements and divide
each row one by one.}
}
\value{
Expr
}
\description{
Get an array with the cumulative product computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  pl$col("a")$cumprod()$alias("cumprod"),
  pl$col("a")$cumprod(reverse = TRUE)$alias("cumprod_reversed")
)
}
