% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cumsum}
\alias{Expr_cumsum}
\title{Cumulative sum}
\usage{
Expr_cumsum(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, start with the total sum of elements and substract
each row one by one.}
}
\value{
Expr
}
\description{
Get an array with the cumulative sum computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  pl$col("a")$cumsum()$alias("cumsum"),
  pl$col("a")$cumsum(reverse = TRUE)$alias("cumsum_reversed")
)
}
