% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cut}
\alias{Expr_cut}
\title{Bin continuous values into discrete categories}
\usage{
Expr_cut(
  breaks,
  ...,
  labels = NULL,
  left_closed = FALSE,
  include_breaks = FALSE
)
}
\arguments{
\item{breaks}{Unique cut points.}

\item{...}{Ignored.}

\item{labels}{Names of the categories. The number of labels must be equal to
the number of cut points plus one.}

\item{left_closed}{Set the intervals to be left-closed instead of right-closed.}

\item{include_breaks}{Include a column with the right endpoint of the bin each
observation falls in. This will change the data type of the output from a
\code{\link[=DataType_Categorical]{Categorical}} to a \code{\link[=DataType_Struct]{Struct}}.}
}
\value{
Expr of data type \code{Categorical} is \code{include_breaks} is \code{FALSE} and
of data type \code{Struct} if \code{include_breaks} is \code{TRUE}.
}
\description{
Bin continuous values into discrete categories
}
\examples{
df = pl$DataFrame(foo = c(-2, -1, 0, 1, 2))

df$with_columns(
  cut = pl$col("foo")$cut(c(-1, 1), labels = c("a", "b", "c"))
)

# Add both the category and the breakpoint
df$with_columns(
  cut = pl$col("foo")$cut(c(-1, 1), include_breaks = TRUE)
)$unnest("cut")
}
\seealso{
\code{\link[=Expr_qcut]{$qcut()}}
}
