% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_diff}
\alias{Expr_diff}
\alias{diff}
\title{Arg Sort}
\usage{
Expr_diff(n = 1, null_behavior = "ignore")
}
\arguments{
\item{n}{Number of slots to shift.}

\item{null_behavior}{option defult 'ignore', else 'drop'}
}
\value{
Expr
}
\description{
Calculate the n-th discrete difference.
}
\keyword{Expr}
