% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_div}
\alias{Expr_div}
\alias{/.Expr}
\title{Div}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_div

\method{/}{Expr}(e1, e2)
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
Divide
}
\examples{
#three syntaxes same result
pl$lit(5) / 10
pl$lit(5) / pl$lit(10)
pl$lit(5)$div(pl$lit(10))
}
\keyword{Expr}
\keyword{Expr_operators}
