% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_drop_nans}
\alias{Expr_drop_nans}
\title{Drop NaN(s)}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_drop_nans
}
\value{
Expr
}
\description{
Drop floating point NaN values.
Similar to R syntax \code{x[!is.nan(x)]}
}
\details{
Note that NaN values are not null values! (null corrosponds to R NA, not R NULL)
To drop null values, use method \code{drop_nulls}.

See Inf,NaN,NULL,Null/NA translations here \code{\link[polars]{docs_translations}}
}
\examples{
 pl$DataFrame(list(x=c(1,2,NaN,NA)))$select(pl$col("x")$drop_nans())
}
\keyword{Expr}
