% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_drop_nulls}
\alias{Expr_drop_nulls}
\title{Drop missing values}
\usage{
Expr_drop_nulls
}
\value{
Expr
}
\description{
Drop missing values
}
\examples{
pl$DataFrame(list(x = c(1, 2, NaN, NA)))$select(pl$col("x")$drop_nulls())
}
\seealso{
\code{drop_nans()}
}
\keyword{datasets}
