% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_dt}
\alias{Expr_dt}
\alias{dt_ns}
\title{dt: datetime related methods}
\usage{
Expr_dt()
}
\value{
Expr
}
\description{
Create an object namespace of all datetime related methods.
See the individual method pages for full details
}
\examples{

#missing

}
\keyword{Expr}
