% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_cast_time_unit}
\alias{expr_dt_cast_time_unit}
\title{Change time unit}
\usage{
expr_dt_cast_time_unit(time_unit)
}
\arguments{
\item{time_unit}{One of \code{"us"} (microseconds),
\code{"ns"} (nanoseconds) or \code{"ms"}(milliseconds). Representing the unit of time.}
}
\value{
A polars \link{expression}
}
\description{
Cast the underlying data to another time unit. This may lose precision.
}
\examples{
df <- pl$select(
  date = pl$datetime_range(
    start = as.Date("2001-1-1"),
    end = as.Date("2001-1-3"),
    interval = "1d1s"
  )
)
df$with_columns(
  cast_time_unit_ms = pl$col("date")$dt$cast_time_unit("ms"),
  cast_time_unit_ns = pl$col("date")$dt$cast_time_unit("ns"),
)
}
