% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_date}
\alias{expr_dt_date}
\title{Extract date from date(time)}
\usage{
expr_dt_date()
}
\value{
A polars \link{expression}
}
\description{
Extract date from date(time)
}
\examples{
df <- pl$DataFrame(
  datetime = as.POSIXct(c("1978-1-1 1:1:1", "1897-5-7 00:00:00"), tz = "UTC")
)
df$with_columns(
  date = pl$col("datetime")$dt$date()
)
}
