% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_day}
\alias{expr_dt_day}
\title{Extract day from underlying Date representation}
\usage{
expr_dt_day()
}
\value{
A polars \link{expression}
}
\description{
Returns the day of month starting from 1. The return value ranges from 1 to
31 (the last day of month differs across months).
}
\examples{
df <- pl$select(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d"
  )
)
df$with_columns(
  pl$col("date")$dt$day()$alias("day")
)
}
