% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_days_in_month}
\alias{expr_dt_days_in_month}
\title{Extract the number of days in the month from the underlying Date representation.}
\usage{
expr_dt_days_in_month()
}
\value{
A polars \link{expression}
}
\description{
Extract the number of days in the month from the underlying Date representation.
}
\examples{
df <- pl$DataFrame(date = as.Date(c("2020-01-01", "2020-02-01", "2020-03-01")))

df$with_columns(
  days_in_month = pl$col("date")$dt$days_in_month()
)
}
