% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_iso_year}
\alias{expr_dt_iso_year}
\title{Extract ISO year from underlying Date representation}
\usage{
expr_dt_iso_year()
}
\value{
A polars \link{expression}
}
\description{
Returns the year number in the ISO standard. This may not correspond with
the calendar year.
}
\examples{
df <- pl$DataFrame(
  date = as.Date(c("1977-01-01", "1978-01-01", "1979-01-01"))
)
df$with_columns(
  year = pl$col("date")$dt$year(),
  iso_year = pl$col("date")$dt$iso_year()
)
}
