% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_millennium}
\alias{expr_dt_millennium}
\title{Extract the millennium from underlying representation}
\usage{
expr_dt_millennium()
}
\value{
A polars \link{expression}
}
\description{
Returns the millennium number in the calendar date.
}
\examples{
df <- pl$DataFrame(
  date = as.Date(
    c("999-12-31", "1897-05-07", "2000-01-01", "2001-07-05", "3002-10-20")
  )
)
df$with_columns(
  millennium = pl$col("date")$dt$millennium()
)
}
