% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_minute}
\alias{expr_dt_minute}
\title{Extract minute from underlying Datetime representation}
\usage{
expr_dt_minute()
}
\value{
A polars \link{expression}
}
\description{
Returns the minute number from 0 to 59.
}
\examples{
df <- pl$DataFrame(
  datetime = as.POSIXct(
    c(
      "1978-01-01 01:01:01",
      "2024-10-13 05:30:14.500",
      "2065-01-01 10:20:30.06"
    ),
    "UTC"
  )
)
df$with_columns(
  pl$col("datetime")$dt$minute()$alias("minute")
)
}
