% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_month}
\alias{expr_dt_month}
\title{Extract month from underlying Date representation}
\usage{
expr_dt_month()
}
\value{
A polars \link{expression}
}
\description{
Returns the month number between 1 and 12.
}
\examples{
df <- pl$DataFrame(
  date = as.Date(c("2001-01-01", "2001-06-30", "2001-12-27"))
)
df$with_columns(
  month = pl$col("date")$dt$month()
)
}
