% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_replace}
\alias{expr_dt_replace}
\title{Replace time unit}
\usage{
expr_dt_replace(
  ...,
  year = NULL,
  month = NULL,
  day = NULL,
  hour = NULL,
  minute = NULL,
  second = NULL,
  microsecond = NULL,
  ambiguous = c("raise", "earliest", "latest", "null")
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{year}{Column or literal.}

\item{month}{Column or literal, ranging from 1-12.}

\item{day}{Column or literal, ranging from 1-31}

\item{hour}{Column or literal, ranging from 0-23.}

\item{minute}{Column or literal, ranging from 0-59.}

\item{second}{Column or literal, ranging from 0-59.}

\item{microsecond}{Column or literal, ranging from 0-999999.}

\item{ambiguous}{Determine how to deal with ambiguous datetimes.
Character vector or \link{expression} containing the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"earliest"}: Use the earliest datetime
\item \code{"latest"}: Use the latest datetime
\item \code{"null"}: Return a null value
}}
}
\value{
A polars \link{expression}
}
\description{
Replace time unit
}
\examples{
df <- pl$DataFrame(
  date = as.Date(c("2024-04-01", "2025-03-16")),
  new_day = c(10, 15)
)
df$with_columns(replaced = pl$col("date")$dt$replace(day = "new_day"))
}
