% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_second}
\alias{expr_dt_second}
\title{Extract seconds from underlying Datetime representation}
\usage{
expr_dt_second(..., fractional = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{fractional}{If \code{TRUE}, include the fractional component of the second.}
}
\value{
A polars \link{expression}
}
\description{
Returns the integer second number from 0 to 59, or a floating point number
from 0 to 60 if \code{fractional = TRUE} that includes any milli/micro/nanosecond
component.
}
\examples{
df <- pl$DataFrame(
  datetime = as.POSIXct(
    c(
      "1978-01-01 01:01:01",
      "2024-10-13 05:30:14.500",
      "2065-01-01 10:20:30.06"
    ),
    "UTC"
  )
)

df$with_columns(
  second = pl$col("datetime")$dt$second(),
  second_fractional = pl$col("datetime")$dt$second(fractional = TRUE)
)
}
