% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_time}
\alias{expr_dt_time}
\title{Extract time}
\usage{
expr_dt_time()
}
\value{
A polars \link{expression}
}
\description{
This only works on Datetime columns, it will error on Date columns.
}
\examples{
df <- pl$select(dates = pl$datetime_range(
  as.Date("2000-1-1"),
  as.Date("2000-1-2"),
  "1h"
))

df$with_columns(times = pl$col("dates")$dt$time())
}
