% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_timestamp}
\alias{expr_dt_timestamp}
\title{Get timestamp in the given time unit}
\usage{
expr_dt_timestamp(time_unit = c("us", "ns", "ms"))
}
\arguments{
\item{time_unit}{Time unit, one of 'ns', 'us', or 'ms'.}
}
\value{
A polars \link{expression}
}
\description{
Get timestamp in the given time unit
}
\examples{
df <- pl$select(
  date = pl$datetime_range(
    start = as.Date("2001-1-1"),
    end = as.Date("2001-1-3"),
    interval = "1d1s"
  )
)
df$select(
  pl$col("date"),
  pl$col("date")$dt$timestamp()$alias("timestamp_ns"),
  pl$col("date")$dt$timestamp(time_unit = "ms")$alias("timestamp_ms")
)
}
