% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_total_days}
\alias{expr_dt_total_days}
\title{Extract the days from a Duration type}
\usage{
expr_dt_total_days(..., fractional = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{fractional}{A bool to indicate whether to include the fractional component of the day.}
}
\value{
A polars \link{expression}
}
\description{
Extract the days from a Duration type
}
\examples{
df <- pl$select(
  date = pl$datetime_range(
    start = as.Date("2020-3-1"),
    end = as.Date("2020-5-1"),
    interval = "1mo1s"
  )
)
df$with_columns(
  diff_days = pl$col("date")$diff()$dt$total_days()
)
}
