% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_total_nanoseconds}
\alias{expr_dt_total_nanoseconds}
\title{Extract the nanoseconds from a Duration type}
\usage{
expr_dt_total_nanoseconds(..., fractional = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{fractional}{A bool to indicate whether to include the fractional component of the day.}
}
\value{
A polars \link{expression}
}
\description{
Extract the nanoseconds from a Duration type
}
\examples{
df <- pl$select(date = pl$datetime_range(
  start = as.POSIXct("2020-1-1", tz = "GMT"),
  end = as.POSIXct("2020-1-1 00:00:01", tz = "GMT"),
  interval = "1ms"
))
df$with_columns(
  diff_nanosec = pl$col("date")$diff()$dt$total_nanoseconds()
)
}
