% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_truncate}
\alias{expr_dt_truncate}
\title{Truncate datetime}
\usage{
expr_dt_truncate(every)
}
\arguments{
\item{every}{Either an Expr or a string indicating a column name or a
duration (see Details).}
}
\value{
A polars \link{expression}
}
\description{
Divide the date/datetime range into buckets. Each date/datetime is mapped to
the start of its bucket using the corresponding local datetime. Note that
weekly buckets start on Monday. Ambiguous results are localised using the
DST offset of the original timestamp - for example, truncating
\code{'2022-11-06 01:30:00 CST'} by \code{'1h'} results in
\code{'2022-11-06 01:00:00 CST'}, whereas truncating \code{'2022-11-06 01:30:00 CDT'}
by \code{'1h'} results in \code{'2022-11-06 01:00:00 CDT'}.
}
\details{
The \code{every} and \code{offset} argument are created with the
the following string language:
\itemize{
\item 1ns # 1 nanosecond
\item 1us # 1 microsecond
\item 1ms # 1 millisecond
\item 1s  # 1 second
\item 1m  # 1 minute
\item 1h  # 1 hour
\item 1d  # 1 day
\item 1w  # 1 calendar week
\item 1mo # 1 calendar month
\item 1y  # 1 calendar year
These strings can be combined:
\itemize{
\item 3d12h4m25s # 3 days, 12 hours, 4 minutes, and 25 seconds
}
}
}
\examples{
df <- pl$select(
  datetime = pl$datetime_range(
    as.Date("2001-01-01"),
    as.Date("2001-01-02"),
    as.difftime("0:25:0")
  )
)
df$with_columns(truncated = pl$col("datetime")$dt$truncate("1h"))

df <- pl$select(
  datetime = pl$datetime_range(
    as.POSIXct("2001-01-01 00:00"),
    as.POSIXct("2001-01-01 01:00"),
    as.difftime("0:10:0")
  )
)
df$with_columns(truncated = pl$col("datetime")$dt$truncate("30m"))
}
