% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_weekday}
\alias{expr_dt_weekday}
\title{Extract weekday from underlying Date representation}
\usage{
expr_dt_weekday()
}
\value{
A polars \link{expression}
}
\description{
Returns the ISO weekday number where Monday = 1 and Sunday = 7.
}
\examples{
df <- pl$select(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d"
  )
)
df$with_columns(
  weekday = pl$col("date")$dt$weekday()
)
}
