% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_entropy}
\alias{Expr_entropy}
\alias{entropy}
\title{Entropy}
\usage{
Expr_entropy(base = base::exp(1), normalize = TRUE)
}
\arguments{
\item{base}{Given exponential base, defaults to \code{e}}

\item{normalize}{Normalize pk if it doesn't sum to 1.}
}
\value{
Expr
}
\description{
Computes the entropy.
Uses the formula \code{-sum(pk * log(pk))} where \code{pk} are discrete probabilities.
Return Null if input is not values
}
\examples{
pl$select(pl$lit(c("a", "b", "b", "c", "c", "c"))$unique_counts()$entropy(base = 2))
}
\keyword{Expr}
