% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_exp}
\alias{Expr_exp}
\alias{exp}
\title{Compute the exponential, element-wise.}
\format{
a method
}
\usage{
Expr_exp
}
\value{
Expr
}
\description{
Compute the exponential, element-wise.
}
\examples{
log10123 = suppressWarnings(log(-1:3))
all.equal(
  pl$DataFrame(list(a = log10123))$select(pl$col("a")$exp())$as_data_frame()$a,
  exp(1)^log10123
)
}
\keyword{Expr}
