% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_explode}
\alias{Expr_explode}
\alias{explode}
\title{Explode a list or utf8 Series.}
\format{
a method
}
\usage{
Expr_explode
}
\value{
Expr
}
\description{
This means that every item is expanded to a new row.
}
\details{
explode/flatten does not support categorical
}
\examples{
pl$DataFrame(list(a=letters))$select(pl$col("a")$explode()$take(0:5))

listed_group_df =  pl$DataFrame(iris[c(1:3,51:53),])$groupby("Species")$agg(pl$all())
print(listed_group_df)
vectors_df = listed_group_df$select(
  pl$col(c("Sepal.Width","Sepal.Length"))$explode()
)
print(vectors_df)
}
\keyword{Expr}
