% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_extend_constant}
\alias{Expr_extend_constant}
\alias{extend_constant}
\title{Extend_constant}
\format{
Method
}
\usage{
Expr_extend_constant(value, n)
}
\arguments{
\item{value}{The value to extend the Series with.
This value may be None to fill with nulls.}

\item{n}{The number of values to extend.}
}
\value{
Expr
}
\description{
Extend the Series with given number of values.
}
\examples{
pl$select(
  pl$lit(c("5","Bob_is_not_a_number"))
  $cast(pl$dtypes$UInt64, strict = FALSE)
  $extend_constant(10.1, 2)
)

pl$select(
  pl$lit(c("5","Bob_is_not_a_number"))
  $cast(pl$dtypes$Utf8, strict = FALSE)
  $extend_constant("chuchu", 2)
)
}
\keyword{Expr}
