% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_extend_expr}
\alias{Expr_extend_expr}
\alias{extend_constant}
\title{Extend_expr}
\format{
Method
}
\usage{
Expr_extend_expr(value, n)
}
\arguments{
\item{value}{The expr to extend the Series with.
This value may be None to fill with nulls.}

\item{n}{The number of values to extend.}
}
\value{
Expr
}
\description{
Extend the Series with a expression repeated a number of times
}
\examples{
pl$select(
  pl$lit(c("5","Bob_is_not_a_number"))
  $cast(pl$dtypes$UInt64, strict = FALSE)
  $extend_expr(10.1, 2)
)

pl$select(
  pl$lit(c("5","Bob_is_not_a_number"))
  $cast(pl$dtypes$Utf8, strict = FALSE)
  $extend_expr("chuchu", 2)
)
}
\keyword{Expr}
