% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_fill_nan}
\alias{Expr_fill_nan}
\title{Fill floating point NaN value with a fill value}
\usage{
Expr_fill_nan(value = NULL)
}
\arguments{
\item{value}{Value used to fill \code{NaN} values.}
}
\value{
Expr
}
\description{
Fill floating point NaN value with a fill value
}
\examples{
pl$DataFrame(a = c(NaN, 1, NaN, 2, NA))$
  with_columns(
  literal = pl$col("a")$fill_nan(999),
  # implicit coercion to string
  string = pl$col("a")$fill_nan("invalid")
)
}
