% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_fill_null}
\alias{Expr_fill_null}
\title{Fill null values with a value or strategy}
\usage{
Expr_fill_null(value = NULL, strategy = NULL, limit = NULL)
}
\arguments{
\item{value}{Expr or something coercible in an Expr}

\item{strategy}{Possible choice are \code{NULL} (default, requires a non-null
\code{value}), \code{"forward"}, \code{"backward"}, \code{"min"}, \code{"max"}, \code{"mean"}, \code{"zero"},
\code{"one"}.}

\item{limit}{Number of consecutive null values to fill when using the
\code{"forward"} or \code{"backward"} strategy.}
}
\value{
Expr
}
\description{
Fill null values with a value or strategy
}
\examples{
pl$DataFrame(a = c(NA, 1, NA, 2, NA))$
  with_columns(
    value = pl$col("a")$fill_null(999),
    backward = pl$col("a")$fill_null(strategy = "backward"),
    mean = pl$col("a")$fill_null(strategy = "mean")
  )
}
