% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_filter}
\alias{Expr_filter}
\title{Filter a single column.}
\usage{
Expr_filter(predicate)
}
\arguments{
\item{predicate}{An Expr or something coercible to an Expr. Must return a
boolean.}
}
\value{
Expr
}
\description{
Mostly useful in an aggregation context. If you want to filter on a
DataFrame level, use \code{DataFrame$filter()} (or \code{LazyFrame$filter()}).
}
\examples{
df = pl$DataFrame(
  group_col = c("g1", "g1", "g2"),
  b = c(1, 2, 3)
)
df

df$group_by("group_col")$agg(
  lt = pl$col("b")$filter(pl$col("b") < 2),
  gte = pl$col("b")$filter(pl$col("b") >= 2)
)
}
