% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_flatten}
\alias{Expr_flatten}
\alias{flatten}
\title{Flatten/Explode a list or utf8 Series.}
\format{
a method
}
\usage{
Expr_flatten
}
\value{
Expr
}
\description{
Alias for explode
This means that every item is expanded to a new row.
}
\examples{
pl$DataFrame(list(a=letters))$select(pl$col("a")$flatten()$take(0:5))

listed_group_df =  pl$DataFrame(iris[c(1:3,51:53),])$groupby("Species")$agg(pl$all())
print(listed_group_df)
vectors_df = listed_group_df$select(
  pl$col(c("Sepal.Width","Sepal.Length"))$flatten()
)
print(vectors_df)
}
\keyword{Expr}
