% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_floor_div}
\alias{Expr_floor_div}
\title{Floor divide two expressions}
\usage{
Expr_floor_div(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of floor division operator \code{expr \%/\% other}.
}
\examples{
df = pl$DataFrame(x = 1:5)

df$with_columns(
  `x/2` = pl$col("x")$div(2),
  `x\%/\%2` = pl$col("x")$floor_div(2)
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
\item \code{\link[=Expr_div]{<Expr>$div()}}
\item \code{\link[=Expr_mod]{<Expr>$mod()}}
}
}
