% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_forward_fill}
\alias{Expr_forward_fill}
\title{Fill null values forward}
\usage{
<Expr>$forward_fill(limit = NULL)
}
\arguments{
\item{limit}{Number of consecutive null values to fill when using the
\code{"forward"} or \code{"backward"} strategy.}
}
\value{
Expr
}
\description{
Fill missing values with the last seen values. Syntactic sugar for
\verb{$fill_null(strategy = "forward")}.
}
\examples{
pl$DataFrame(a = c(NA, 1, NA, 2, NA))$
  with_columns(
  backward = pl$col("a")$forward_fill()
)
}
