% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_gather}
\alias{Expr_gather}
\title{Gather values by index}
\usage{
Expr_gather(indices)
}
\arguments{
\item{indices}{R scalar/vector or Series, or Expr that leads to a Series of
dtype UInt32.}
}
\value{
Expr
}
\description{
Gather values by index
}
\examples{
pl$DataFrame(a = c(1, 2, 4, 5, 8))$select(pl$col("a")$gather(c(0, 2, 4)))
}
