% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_gt_eq}
\alias{Expr_gt_eq}
\alias{>=.RPolarsExpr}
\title{Check greater or equal inequality}
\usage{
Expr_gt_eq(other)

\method{>=}{RPolarsExpr}(e1, e2)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(2) >= 2
pl$lit(2) >= pl$lit(2)
pl$lit(2)$gt_eq(pl$lit(2))
}
