% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_has_nulls}
\alias{Expr_has_nulls}
\title{Check whether the expression contains one or more null values}
\usage{
Expr_has_nulls()
}
\value{
Expr
}
\description{
Check whether the expression contains one or more null values
}
\examples{
df = pl$DataFrame(
  a = c(NA, 1, NA),
  b = c(1, NA, 2),
  c = c(1, 2, 3)
)

df$select(pl$all()$has_nulls())
}
