% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_hash}
\alias{Expr_hash}
\alias{hash}
\title{Hash elements}
\usage{
<Expr>$hash(seed = 0, seed_1 = NULL, seed_2 = NULL, seed_3 = NULL)
}
\arguments{
\item{seed}{Random seed parameter. Defaults to 0. Doesn't have any effect
for now.}

\item{seed_1, seed_2, seed_3}{Random seed parameter. Defaults to arg seed.
The column will be coerced to UInt32.}
}
\value{
Expr
}
\description{
The hash value is of type \code{UInt64}.
}
\examples{
df = pl$DataFrame(iris[1:3, c(1, 2)])
df$with_columns(pl$all()$hash(1234)$name$suffix("_hash"))
}
