% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_head}
\alias{Expr_head}
\title{Get the first n elements}
\usage{
Expr_head(n = 10)
}
\arguments{
\item{n}{Number of elements to take.}
}
\value{
Expr
}
\description{
Get the first n elements
}
\examples{
pl$DataFrame(x = 1:11)$select(pl$col("x")$head(3))
}
