% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_implode}
\alias{Expr_implode}
\alias{list}
\title{Wrap column in list}
\usage{
Expr_implode
}
\value{
Expr
}
\description{
Aggregate values into a list.
}
\details{
use to_struct to wrap a DataFrame. Notice implode() is sometimes referred to
as list() .
}
\examples{
df = pl$DataFrame(
  a = 1:3,
  b = 4:6
)
df$select(pl$all()$implode())
}
\keyword{datasets}
