% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_interpolate}
\alias{Expr_interpolate}
\alias{interpolate}
\title{Interpolate \code{Nulls}}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_interpolate
}
\value{
Expr
}
\description{
Fill nulls with linear interpolation over missing values.
Can also be used to regrid data to a new grid - see examples below.
}
\examples{
pl$empty_select(pl$lit(c(1,NA,4,NA,100))$interpolate())
}
\keyword{Expr}
