% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_between}
\alias{Expr_is_between}
\title{Check whether a value is between two values}
\usage{
<Expr>$is_between(start, end, include_bounds = FALSE)
}
\arguments{
\item{start}{Lower bound, an Expr that is either numeric or datetime.}

\item{end}{Upper bound, an Expr that is either numeric or datetime.}

\item{include_bounds}{If \code{FALSE} (default), exclude start and end. This can
also be a vector of two booleans indicating whether to include the start
and/or the end.}
}
\value{
Expr
}
\description{
This is syntactic sugar for \code{x > start & x < end} (or \code{x >= start & x <= end}).
}
\examples{
df = pl$DataFrame(num = 1:5, y = c(0, 2, 3, 3, 3))
df$with_columns(
  bet_2_4_no_bounds = pl$col("num")$is_between(2, 4),
  bet_2_4_with_bounds = pl$col("num")$is_between(2, 4, TRUE),
  bet_2_4_upper_bound = pl$col("num")$is_between(2, 4, c(FALSE, TRUE)),
  between_y_4 = pl$col("num")$is_between(pl$col("y"), 6)
)
}
