% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_duplicated}
\alias{Expr_is_duplicated}
\title{Check whether each value is duplicated}
\usage{
Expr_is_duplicated()
}
\value{
Expr
}
\description{
This is syntactic sugar for \verb{$is_unique()$not()}.
}
\examples{
pl$DataFrame(head(mtcars[, 1:2]))$
  with_columns(is_duplicated = pl$col("mpg")$is_duplicated())
}
