% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_first}
\alias{Expr_is_first}
\alias{is_unique}
\title{Get a mask of the first unique value.}
\format{
a method
}
\usage{
Expr_is_first
}
\value{
Expr (boolean)
}
\description{
Get a mask of the first unique value.
}
\examples{
pl$empty_select(pl$lit(abs(-2:2))$is_first())
}
\keyword{Expr}
