% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_last_distinct}
\alias{Expr_is_last_distinct}
\title{Check whether each value is the last occurrence}
\usage{
<Expr>$is_last_distinct
}
\value{
Expr
}
\description{
Check whether each value is the last occurrence
}
\examples{
pl$DataFrame(head(mtcars[, 1:2]))$
  with_columns(is_ulast = pl$col("mpg")$is_last_distinct())
}
\keyword{datasets}
