% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_nan}
\alias{Expr_is_nan}
\alias{is_nan}
\title{Check if elements are NaN}
\usage{
<Expr>$is_nan
}
\value{
Expr
}
\description{
Returns a boolean Series indicating which values are NaN.
}
\examples{
pl$DataFrame(list(alice = c(0, NaN, NA, Inf, -Inf)))$
  with_columns(nan = pl$col("alice")$is_nan())
}
\keyword{datasets}
