% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_not_nan}
\alias{Expr_is_not_nan}
\alias{is_not_nan}
\title{Check if elements are not NaN}
\usage{
<Expr>$is_not_nan
}
\value{
Expr
}
\description{
Returns a boolean Series indicating which values are not NaN. Syntactic sugar
for \verb{$is_nan()$not()}.
}
\examples{
pl$DataFrame(list(alice = c(0, NaN, NA, Inf, -Inf)))$
  with_columns(not_nan = pl$col("alice")$is_not_nan())
}
\keyword{datasets}
