% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_unique}
\alias{Expr_is_unique}
\title{Check whether each value is unique}
\usage{
Expr_is_unique
}
\value{
Expr
}
\description{
Check whether each value is unique
}
\examples{
pl$DataFrame(head(mtcars[, 1:2]))$
  with_columns(is_unique = pl$col("mpg")$is_unique())
}
\keyword{datasets}
