% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_kurtosis}
\alias{Expr_kurtosis}
\title{Kurtosis}
\usage{
<Expr>$kurtosis(fisher = TRUE, bias = TRUE)
}
\arguments{
\item{fisher}{If \code{TRUE} (default), Fisher’s definition is used (normal,
centered at 0). Otherwise, Pearson’s definition is used (normal, centered at
3).}

\item{bias}{If \code{FALSE}, the calculations are corrected for statistical bias.}
}
\value{
Expr
}
\description{
Compute the kurtosis (Fisher or Pearson) of a dataset.
}
\details{
Kurtosis is the fourth central moment divided by the square of the variance.
If Fisher's definition is used, then 3 is subtracted from the result to
give 0 for a normal distribution.

If bias is \code{FALSE}, then the kurtosis is calculated using \code{k} statistics to
eliminate bias coming from biased moment estimators.
}
\examples{
pl$DataFrame(a = c(1:3, 2:1))$
  with_columns(kurt = pl$col("a")$kurtosis())
}
