% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_last}
\alias{Expr_last}
\title{Get the last value}
\usage{
Expr_last()
}
\value{
Expr
}
\description{
Get the last value
}
\examples{
pl$DataFrame(x = 3:1)$with_columns(last = pl$col("x")$last())
}
