% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_limit}
\alias{Expr_limit}
\title{Get the first n elements}
\usage{
<Expr>$limit(n = 10)
}
\arguments{
\item{n}{Number of elements to take.}
}
\value{
Expr
}
\description{
This is an alias for \verb{<Expr>$head()}.
}
\examples{
pl$DataFrame(x = 1:11)$select(pl$col("x")$limit(3))
}
