% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_list}
\alias{Expr_list}
\alias{list}
\title{Wrap column in list}
\format{
a method
}
\usage{
Expr_list
}
\value{
Expr
}
\description{
Aggregate to list.
}
\details{
use to_struct to wrap a DataFrame
}
\examples{
pl$select(pl$lit(1:4)$list(), pl$lit(c("a")))
}
\keyword{Expr}
