% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_count_matches}
\alias{expr_list_count_matches}
\title{Count how often a value produced occurs}
\usage{
expr_list_count_matches(element)
}
\arguments{
\item{element}{An expression that produces a single value.}
}
\value{
A polars \link{expression}
}
\description{
Count how often a value produced occurs
}
\examples{
df <- pl$DataFrame(a = list(0, 1, c(1, 2, 3, 2), c(1, 2, 1), c(4, 4)))

df$with_columns(
  number_of_twos = pl$col("a")$list$count_matches(2)
)
}
