% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_diff}
\alias{expr_list_diff}
\title{Compute difference between sub-list values}
\usage{
expr_list_diff(n = 1, null_behavior = c("ignore", "drop"))
}
\arguments{
\item{n}{Number of slots to shift. If negative, then it starts from the end.}

\item{null_behavior}{How to handle \code{null} values. Either \code{"ignore"} (default)
or \code{"drop"}.}
}
\value{
A polars \link{expression}
}
\description{
This computes the first discrete difference between shifted items of every
list. The parameter \code{n} gives the interval between items to subtract, e.g.
if \code{n = 2} the output will be the difference between the 1st and the 3rd
value, the 2nd and 4th value, etc.
}
\examples{
df <- pl$DataFrame(s = list(1:4, c(10L, 2L, 1L)))
df$with_columns(diff = pl$col("s")$list$diff(2))

# negative value starts shifting from the end
df$with_columns(diff = pl$col("s")$list$diff(-2))
}
