% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_drop_nulls}
\alias{expr_list_drop_nulls}
\title{Drop all null values in every sub-list}
\usage{
expr_list_drop_nulls()
}
\value{
A polars \link{expression}
}
\description{
Drop all null values in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(NA, 0, NA), c(1, NaN), NA))

df$with_columns(
  without_nulls = pl$col("values")$list$drop_nulls()
)
}
