% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_eval}
\alias{expr_list_eval}
\title{Run any polars expression on the sub-lists' values}
\usage{
expr_list_eval(expr)
}
\arguments{
\item{expr}{Expression to run. Note that you can select an element with
\code{pl$element()}, \code{pl$first()}, and more. See Examples.}
}
\value{
A polars \link{expression}
}
\description{
Run any polars expression on the sub-lists' values
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 8, 3), c(3, 2), c(NA, NA, 1)),
  b = list(c("R", "is", "amazing"), c("foo", "bar"), "text")
)

df

# standardize each value inside a list, using only the values in this list
df$select(
  a_stand = pl$col("a")$list$eval(
    (pl$element() - pl$element()$mean()) / pl$element()$std()
  )
)

# count characters for each element in list. Since column "b" is list[str],
# we can apply all `$str` functions on elements in the list:
df$select(
  b_len_chars = pl$col("b")$list$eval(
    pl$element()$str$len_chars()
  )
)

# concat strings in each list
df$select(
  pl$col("b")$list$eval(pl$element()$str$join(" "))$list$first()
)
}
