% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_first}
\alias{expr_list_first}
\title{Get the first value of the sub-lists}
\usage{
expr_list_first()
}
\value{
A polars \link{expression}
}
\description{
Get the first value of the sub-lists
}
\examples{
df <- pl$DataFrame(a = list(3:1, NULL, 1:2))
df$with_columns(
  first = pl$col("a")$list$first()
)
}
