% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_join}
\alias{expr_list_join}
\title{Join elements of every sub-list}
\usage{
expr_list_join(separator, ..., ignore_nulls = FALSE)
}
\arguments{
\item{separator}{String to separate the items with. Can be an Expr. Strings
are \emph{not} parsed as columns.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Columns to concatenate into a
single string column. Accepts expression input. Strings are parsed as column
names, other non-expression inputs are parsed as literals. Non-\code{String}
columns are cast to \code{String}.}

\item{ignore_nulls}{If \code{FALSE} (default), null values will be propagated,
i.e. if the row contains any null values, the output is null.}
}
\value{
A polars \link{expression}
}
\description{
Join all string items in a sub-list and place a separator between them. This
only works if the inner dtype is \code{String}.
The default is to propagate null values.
}
\examples{
df <- pl$DataFrame(
  s = list(c("a", "b", "c"), c("x", "y"), c("e", NA)),
  separator = c("-", "+", "/")
)
df$with_columns(
  join_with_expr = pl$col("s")$list$join(pl$col("separator")),
  join_with_lit = pl$col("s")$list$join(" "),
  join_ignore_null = pl$col("s")$list$join(" ", ignore_nulls = TRUE)
)
}
