% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_len}
\alias{expr_list_len}
\title{Return the number of elements in each sub-list}
\usage{
expr_list_len()
}
\value{
A polars \link{expression}
}
\description{
Null values are counted in the total.
}
\examples{
df <- pl$DataFrame(list_of_strs = list(c("a", "b", NA), "c"))
df$with_columns(len_list = pl$col("list_of_strs")$list$len())
}
