% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_mean}
\alias{expr_list_mean}
\title{Compute the mean value in every sub-list}
\usage{
expr_list_mean()
}
\value{
A polars \link{expression}
}
\description{
Compute the mean value in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(1, 2, 3, NA), c(2, 3), NA))
df$with_columns(mean = pl$col("values")$list$mean())
}
