% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_median}
\alias{expr_list_median}
\title{Compute the median in every sub-list}
\usage{
expr_list_median()
}
\value{
A polars \link{expression}
}
\description{
Compute the median in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(-1, 0, 1), c(1, 10)))

df$with_columns(
  median = pl$col("values")$list$median()
)
}
