% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_n_unique}
\alias{expr_list_n_unique}
\title{Count the number of unique values in every sub-lists}
\usage{
expr_list_n_unique()
}
\value{
A polars \link{expression}
}
\description{
Count the number of unique values in every sub-lists
}
\examples{
df <- pl$DataFrame(values = list(c(2, 2, NA), c(1, 2, 3), NA))
df$with_columns(unique = pl$col("values")$list$n_unique())
}
