% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_reverse}
\alias{expr_list_reverse}
\title{Reverse values in every sub-list}
\usage{
expr_list_reverse()
}
\value{
A polars \link{expression}
}
\description{
Reverse values in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(1, 2, 3, NA), c(2, 3), NA))
df$with_columns(reverse = pl$col("values")$list$reverse())
}
