% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_sort}
\alias{expr_list_sort}
\title{Sort values in every sub-list}
\usage{
expr_list_sort(..., descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{descending}{Sort values in descending order.}

\item{nulls_last}{Place null values last.}
}
\value{
A polars \link{expression}
}
\description{
Sort values in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(NA, 2, 1, 3), c(Inf, 2, 3, NaN), NA))
df$with_columns(sort = pl$col("values")$list$sort())
}
