% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_tail}
\alias{expr_list_tail}
\title{Slice the last \code{n} values of every sub-list}
\usage{
expr_list_tail(n = 5L)
}
\arguments{
\item{n}{Number of values to return for each sub-list. Can be an Expr. Strings
are parsed as column names.}
}
\value{
A polars \link{expression}
}
\description{
Slice the last \code{n} values of every sub-list
}
\examples{
df <- pl$DataFrame(
  s = list(1:4, c(10L, 2L, 1L)),
  n = 1:2
)
df$with_columns(
  tail_by_expr = pl$col("s")$list$tail("n"),
  tail_by_lit = pl$col("s")$list$tail(2)
)
}
