% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_to_struct}
\alias{expr_list_to_struct}
\title{Convert the Series of type List to a Series of type Struct}
\usage{
expr_list_to_struct(fields)
}
\arguments{
\item{fields}{A character vector of field names.}
}
\value{
A polars \link{expression}
}
\description{
Convert the Series of type List to a Series of type Struct
}
\details{
The field names must be supplied explicitly.
}
\examples{
df <- pl$DataFrame(n = list(c(0, 1), c(0, 1, 2)))

# Convert list to struct with explicit field names:

df$select(pl$col("n")$list$to_struct(c("one", "two", "three")))$unnest("n")

# Convert list to struct with field name assignment by
# index from a list of names:
df$select(pl$col("n")$list$to_struct(
  fields = c("one", "two", "three"))
)$unnest("n")
}
\seealso{
\itemize{
\item \code{\link[=expr_arr_to_struct]{<expr>$arr$to_struct()}}
\item \code{\link[=series_list_to_struct]{<series>$list$to_struct()}}
}
}
