% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_unique}
\alias{expr_list_unique}
\title{Get unique values in a list}
\usage{
expr_list_unique(..., maintain_order = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{maintain_order}{Maintain order of data. This requires more work.}
}
\value{
A polars \link{expression}
}
\description{
Get unique values in a list
}
\examples{
df <- pl$DataFrame(values = list(c(2, 2, NA), c(1, 2, 3), NA))
df$with_columns(unique = pl$col("values")$list$unique())
}
