% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_var}
\alias{expr_list_var}
\title{Compute the variance in every sub-list}
\usage{
expr_list_var(ddof = 1)
}
\arguments{
\item{ddof}{"Delta Degrees of Freedom": the divisor used in the calculation
is \code{N - ddof}, where \code{N} represents the number of elements. By default
\code{ddof} is 1.}
}
\value{
A polars \link{expression}
}
\description{
Compute the variance in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(-1, 0, 1), c(1, 10)))

df$with_columns(
  var = pl$col("values")$list$var()
)
}
