% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_log}
\alias{Expr_log}
\alias{log}
\title{Natural Log}
\usage{
Expr_log(base = base::exp(1))
}
\arguments{
\item{base}{numeric base value for log, default base::exp(1)}
}
\value{
Expr
}
\description{
Compute the base x logarithm of the input array, element-wise.
}
\examples{
pl$DataFrame(list(a = exp(1)^(-1:3)))$select(pl$col("a")$log())
}
\keyword{Expr}
