% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_log10}
\alias{Expr_log10}
\alias{log10}
\title{10-base log}
\format{
a method
}
\usage{
Expr_log10
}
\value{
Expr
}
\description{
Compute the base 10 logarithm of the input array, element-wise.
}
\examples{
pl$DataFrame(list(a = 10^(-1:3)))$select(pl$col("a")$log10())
}
\keyword{Expr}
