% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_lower_bound}
\alias{Expr_lower_bound}
\title{Find the lower bound of a DataType}
\usage{
<Expr>$lower_bound
}
\value{
Expr
}
\description{
Find the lower bound of a DataType
}
\examples{
pl$DataFrame(
  x = 1:3, y = 1:3,
  schema = list(x = pl$UInt32, y = pl$Int32)
)$
  select(pl$all()$lower_bound())
}
\keyword{datasets}
