% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_lt}
\alias{Expr_lt}
\alias{<.RPolarsExpr}
\title{Check strictly lower inequality}
\usage{
Expr_lt(other)

\method{<}{RPolarsExpr}(e1, e2)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(5) < 10
pl$lit(5) < pl$lit(10)
pl$lit(5)$lt(pl$lit(10))
}
