% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_map_alias}
\alias{Expr_map_alias}
\alias{map_alias}
\title{Map alias of expression with an R function}
\usage{
Expr_map_alias(fun)
}
\arguments{
\item{fun}{an R function which takes a string as input and return a string}
}
\value{
Expr
}
\description{
Rename the output of an expression by mapping a function over the root name.
}
\examples{
pl$DataFrame(list(alice=1:3))$select(
  pl$col("alice")$alias("joe_is_not_root")$map_alias(\(x) paste0(x,"_and_bob"))
)
}
\keyword{Expr}
