% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_mean}
\alias{Expr_mean}
\title{mean}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
Expr_mean

Expr_mean
}
\value{
Expr

Expr
}
\description{
Get mean value.

Get median value.
}
\examples{
pl$DataFrame(list(x=c(1,NA,3)))$select(pl$col("x")$mean()==2) #is true
pl$DataFrame(list(x=c(1,NA,2)))$select(pl$col("x")$median()==1.5) #is true
}
\keyword{Expr}
