% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_median}
\alias{Expr_median}
\title{Get median value}
\usage{
Expr_median()
}
\value{
Expr
}
\description{
Get median value
}
\examples{
pl$DataFrame(x = c(1L, NA, 2L))$
  with_columns(median = pl$col("x")$median())
}
