% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_meta}
\alias{Expr_meta}
\alias{meta_ns}
\title{meta: related methods}
\usage{
Expr_meta()
}
\value{
Expr
}
\description{
Create an object namespace of all meta related methods.
See the individual method pages for full details
}
\examples{

# missing

}
