% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_has_multiple_outputs}
\alias{expr_meta_has_multiple_outputs}
\title{Indicate if this expression expands into multiple expressions}
\usage{
expr_meta_has_multiple_outputs()
}
\value{
A polars \link{expression}
}
\description{
Indicate if this expression expands into multiple expressions
}
\examples{
e <- pl$col(c("a", "b"))$name$suffix("_foo")
e$meta$has_multiple_outputs()
}
