% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_is_column}
\alias{expr_meta_is_column}
\title{Indicate if this expression is a basic (non-regex) unaliased column}
\usage{
expr_meta_is_column()
}
\value{
A logical value.
}
\description{
Indicate if this expression is a basic (non-regex) unaliased column
}
\examples{
e <- pl$col("foo")
e$meta$is_column()

e <- pl$col("foo") * pl$col("bar")
e$meta$is_column()

e <- pl$col("^col\\\\.*\\\\d+$")
e$meta$is_column()
}
