% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_is_regex_projection}
\alias{expr_meta_is_regex_projection}
\title{Indicate if this expression expands to columns that match a regex pattern}
\usage{
expr_meta_is_regex_projection()
}
\value{
A logical value.
}
\description{
Indicate if this expression expands to columns that match a regex pattern
}
\examples{
e <- pl$col("^.*$")$name$prefix("foo_")
e$meta$is_regex_projection()
}
