% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_ne}
\alias{expr_meta_ne}
\title{Indicate if this expression is not the same as another expression}
\usage{
expr_meta_ne(other)
}
\arguments{
\item{other}{Expression to compare with.}
}
\value{
A polars \link{expression}
}
\description{
Indicate if this expression is not the same as another expression
}
\examples{
foo_bar <- pl$col("foo")$alias("bar")
foo <- pl$col("foo")
foo_bar$meta$ne(foo)

foo_bar2 <- pl$col("foo")$alias("bar")
foo_bar$meta$ne(foo_bar2)
}
