% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_output_name}
\alias{expr_meta_output_name}
\title{Get the column name that this expression would produce}
\usage{
expr_meta_output_name(..., raise_if_undetermined = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{raise_if_undetermined}{If \code{TRUE} (default), raise an error if the
output name cannot be determined. Otherwise return \code{NA}.}
}
\value{
A polars \link{expression}
}
\description{
It may not always be possible to determine the output name as that can
depend on the schema of the context; in that case this will raise an error
if \code{raise_if_undetermined = TRUE} (the default), and return \code{NA} otherwise.
}
\examples{
e <- pl$col("foo") * pl$col("bar")
e$meta$output_name()

e_filter <- pl$col("foo")$filter(pl$col("bar") == 13)
e_filter$meta$output_name()

e_sum_over <- pl$col("foo")$sum()$over("groups")
e_sum_over$meta$output_name()

e_sum_slice <- pl$col("foo")$sum()$slice(pl$len() - 10, pl$col("bar"))
e_sum_slice$meta$output_name()

pl$len()$meta$output_name()
}
