% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_serialize}
\alias{expr_meta_serialize}
\title{Serialize this expression to a string in binary or JSON format}
\usage{
expr_meta_serialize(..., format = c("binary", "json"))
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{format}{The format in which to serialize. Must be one of:
\itemize{
\item \code{"binary"} (default): serialize to binary format (bytes).
\item \code{"json"}: serialize to JSON format (string).
}}
}
\value{
A polars \link{expression}
}
\description{
Serialize this expression to a string in binary or JSON format
}
\details{
Serialization is not stable across Polars versions: a LazyFrame serialized
in one Polars version may not be deserializable in another Polars version.
}
\examples{
\dontshow{if (\{
requireNamespace("jsonlite", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
# Serialize the expression into a binary representation.
expr <- pl$col("foo")$sum()$over("bar")
bytes <- expr$meta$serialize()
rawToChar(bytes)

pl$deserialize_expr(bytes)

# Serialize into json
expr$meta$serialize(format = "json") |>
  jsonlite::prettify()
\dontshow{\}) # examplesIf}
}
