% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_undo_aliases}
\alias{expr_meta_undo_aliases}
\title{Undo any renaming operation like \code{alias} or \code{name$keep}}
\usage{
expr_meta_undo_aliases()
}
\value{
A polars \link{expression}
}
\description{
Undo any renaming operation like \code{alias} or \code{name$keep}
}
\examples{
e <- pl$col("foo")$alias("bar")
e$meta$undo_aliases()$meta$eq(pl$col("foo"))

e <- pl$col("foo")$sum()$over("bar")
e$name$keep()$meta$undo_aliases()$meta$eq(e)
}
