% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_min}
\alias{Expr_min}
\title{min}
\usage{
Expr_min
}
\value{
Expr
}
\description{
Get minimum value.
}
\examples{
pl$DataFrame(list(x = c(1, NA, 3)))$select(pl$col("x")$min() == 1) # is true
}
\keyword{datasets}
