% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_mod}
\alias{Expr_mod}
\title{Modulo two expressions}
\usage{
Expr_mod(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of modulus operator \code{expr \%\% other}.
}
\examples{
df = pl$DataFrame(x = -5L:5L)

df$with_columns(
  `x\%\%2` = pl$col("x")$mod(2)
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
\item \code{\link[=Expr_floor_div]{<Expr>$floor_div()}}
}
}
