% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_mul}
\alias{Expr_mul}
\title{Multiply two expressions}
\usage{
Expr_mul(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of multiplication operator \code{expr * other}.
}
\examples{
df = pl$DataFrame(x = c(1, 2, 4, 8, 16))

df$with_columns(
  `x*2` = pl$col("x")$mul(2),
  `x * xlog2` = pl$col("x")$mul(pl$col("x")$log(2))
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
}
}
