% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_n_unique}
\alias{Expr_n_unique}
\alias{n_unique}
\title{Count number of unique values}
\usage{
Expr_n_unique
}
\value{
Expr
}
\description{
Count number of unique values.
Similar to R length(unique(x))
}
\examples{
pl$DataFrame(iris)$select(pl$col("Species")$n_unique())
}
\keyword{datasets}
