% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_name}
\alias{Expr_name}
\alias{name_ns}
\title{name: related methods}
\usage{
Expr_name()
}
\value{
Expr
}
\description{
Create an object namespace of all name related methods.
See the individual method pages for full details
}
\examples{

# missing

}
