% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-name.R
\name{expr_name_keep}
\alias{expr_name_keep}
\title{Keep the original root name of the expression.}
\usage{
expr_name_keep()
}
\value{
A polars \link{expression}
}
\description{
Keep the original root name of the expression.
}
\examples{
df <- pl$DataFrame(alice = 1:3)
df$select(pl$col("alice")$alias("bob")$name$keep())
}
