% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-name.R
\name{expr_name_prefix_fields}
\alias{expr_name_prefix_fields}
\title{Add a prefix to all fields name of a struct}
\usage{
expr_name_prefix_fields(prefix)
}
\arguments{
\item{prefix}{Prefix to add to the field name.}
}
\value{
A polars \link{expression}
}
\description{
Add a prefix to all fields name of a struct
}
\examples{
df <- pl$DataFrame(a = 1, b = 2)$select(
  pl$struct(pl$all())$alias("my_struct")
)

df$with_columns(
  pl$col("my_struct")$name$prefix_fields("col_")
)$unnest("my_struct")
}
