% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-name.R
\name{expr_name_to_uppercase}
\alias{expr_name_to_uppercase}
\title{Make the root column name uppercase}
\usage{
expr_name_to_uppercase()
}
\value{
A polars \link{expression}
}
\description{
Due to implementation constraints, this method can only be called as the last
expression in a chain.
}
\examples{
df <- pl$DataFrame(Foo = 1:3, bar = 4:6)
df$select(pl$all()$name$to_uppercase())
}
