% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_nan_max}
\alias{Expr_nan_max}
\title{max}
\usage{
Expr_nan_max
}
\value{
Expr
}
\description{
Get maximum value, but propagate/poison encountered \code{NaN} values.
Get maximum value.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[polars]{docs_translations}}
}
\examples{
pl$DataFrame(list(x = c(1, NaN, Inf, 3)))$select(pl$col("x")$nan_max()$is_nan()) # is true
}
\keyword{Expr}
