% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_nan_min}
\alias{Expr_nan_min}
\title{Get minimum value with NaN}
\usage{
Expr_nan_min
}
\value{
Expr
}
\description{
Get minimum value, but returns \code{NaN} if there are any.
}
\examples{
pl$DataFrame(x = c(1, NA, 3, NaN, Inf))$
  with_columns(nan_min = pl$col("x")$nan_min())
}
\keyword{datasets}
