% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_neq_missing}
\alias{Expr_neq_missing}
\title{Check inequality without \code{null} propagation}
\usage{
Expr_neq_missing(other)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
df = pl$DataFrame(x = c(NA, FALSE, TRUE), y = c(TRUE, TRUE, TRUE))
df$with_columns(
  neq = pl$col("x")$neq("y"),
  neq_missing = pl$col("x")$neq_missing("y")
)
}
\seealso{
\link{Expr_neq}
}
