% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_not}
\alias{Expr_not}
\alias{!.Expr}
\title{Negate a boolean expression}
\usage{
Expr_not

\method{!}{Expr}(x)
}
\arguments{
\item{x}{Expr}
}
\value{
Expr
}
\description{
The RHS can either be an Expr or an object that can be converted to a literal
(e.g an integer).
}
\examples{
# two syntaxes same result
pl$lit(TRUE)$not()
!pl$lit(TRUE)
}
\keyword{datasets}
