% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_not_}
\alias{Expr_not_}
\alias{!.Expr}
\title{Not !}
\usage{
Expr_not_(other)

\method{!}{Expr}(x)
}
\arguments{
\item{x}{Expr}

\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
not method and operator
}
\examples{
# two syntaxes same result
pl$lit(TRUE)$not_()
!pl$lit(TRUE)
}
\keyword{Expr}
\keyword{Expr_operators}
