% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_null_count}
\alias{Expr_null_count}
\alias{null_count}
\title{Count \code{Nulls}}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_null_count
}
\value{
Expr
}
\description{
Count \code{Nulls}
}
\examples{
pl$empty_select(pl$lit(c(NA,"a",NA,"b"))$null_count())
}
\keyword{Expr}
