% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_or}
\alias{Expr_or}
\title{Or}
\usage{
Expr_or(other)
}
\arguments{
\item{other}{Expr or into Expr}
}
\value{
Expr
}
\description{
combine to boolean expresions with OR
}
\examples{
pl$lit(TRUE) | FALSE
pl$lit(TRUE)$or(pl$lit(TRUE))
}
\keyword{Expr}
\keyword{Expr_operators}
