% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_pct_change}
\alias{Expr_pct_change}
\title{Percentage change}
\usage{
<Expr>$pct_change(n = 1)
}
\arguments{
\item{n}{Periods to shift for computing percent change.}
}
\value{
Expr
}
\description{
Computes percentage change (as fraction) between current element and most-
recent non-null element at least \code{n} period(s) before the current element.
Computes the change from the previous row by default.
}
\examples{
pl$DataFrame(a = c(10L, 11L, 12L, NA_integer_, 12L))$
  with_columns(pct_change = pl$col("a")$pct_change())
}
