% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_peak_max}
\alias{Expr_peak_max}
\title{Find local maxima}
\usage{
<Expr>$peak_max()
}
\value{
Expr
}
\description{
A local maximum is the point that marks the transition between an increase
and a decrease in a Series. The first and last values of the Series can never
be a peak.
}
\examples{
df = pl$DataFrame(x = c(1, 2, 3, 2, 3, 4, 5, 2))
df

df$with_columns(peak_max = pl$col("x")$peak_max())
}
\seealso{
\verb{$peak_min()}
}
