% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_peak_min}
\alias{Expr_peak_min}
\title{Find local minima}
\usage{
<Expr>$peak_min()
}
\value{
Expr
}
\description{
A local minimum is the point that marks the transition between a decrease
and an increase in a Series. The first and last values of the Series can never
be a peak.
}
\examples{
df = pl$DataFrame(x = c(1, 2, 3, 2, 3, 4, 5, 2))
df

df$with_columns(peak_min = pl$col("x")$peak_min())
}
\seealso{
\verb{$peak_max()}
}
