% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_pow}
\alias{Expr_pow}
\alias{pow}
\title{Exponentiation \code{^} and \verb{**}.}
\usage{
Expr_pow(exponent)
}
\arguments{
\item{exponent}{exponent}
}
\value{
Expr
}
\description{
Raise expression to the power of exponent.
}
\details{
in py-polars python \code{^} is the xor operator. The R interpreter will replace the
\verb{**} with \code{^}, such that \verb{**} means \code{^}. Except e.g. in strings like "**". Use of \verb{**} will
not work in some cases, because it is not a real separate operator in R.
}
\examples{
pl$DataFrame(a = -1:3)$select(
  pl$lit(2)$pow(pl$col("a"))
)$get_column("literal")$to_r() == 2^(-1:3)

pl$DataFrame(a = -1:3)$select(
  pl$lit(2)^(pl$col("a"))
)$get_column("literal")$to_r() == 2^(-1:3)
}
\keyword{Expr}
