% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_qcut}
\alias{Expr_qcut}
\title{Bin continuous values into discrete categories based on their quantiles}
\usage{
Expr_qcut(
  quantiles,
  ...,
  labels = NULL,
  left_closed = FALSE,
  allow_duplicates = FALSE,
  include_breaks = FALSE
)
}
\arguments{
\item{quantiles}{Either a vector of quantile probabilities between 0 and 1 or
a positive integer determining the number of bins with uniform probability.}

\item{...}{Ignored.}

\item{labels}{Names of the categories. The number of labels must be equal to
the number of cut points plus one.}

\item{left_closed}{Set the intervals to be left-closed instead of right-closed.}

\item{allow_duplicates}{If set to \code{TRUE}, duplicates in the resulting
quantiles are dropped, rather than raising an error. This can happen even
with unique probabilities, depending on the data.}

\item{include_breaks}{Include a column with the right endpoint of the bin each
observation falls in. This will change the data type of the output from a
\code{\link[=DataType_Categorical]{Categorical}} to a \code{\link[=DataType_Struct]{Struct}}.}
}
\value{
Expr of data type \code{Categorical} is \code{include_breaks} is \code{FALSE} and
of data type \code{Struct} if \code{include_breaks} is \code{TRUE}.
}
\description{
Bin continuous values into discrete categories based on their quantiles
}
\examples{
df = pl$DataFrame(foo = c(-2, -1, 0, 1, 2))

# Divide a column into three categories according to pre-defined quantile
# probabilities
df$with_columns(
  qcut = pl$col("foo")$qcut(c(0.25, 0.75), labels = c("a", "b", "c"))
)

# Divide a column into two categories using uniform quantile probabilities.
df$with_columns(
  qcut = pl$col("foo")$qcut(2, labels = c("low", "high"), left_closed = TRUE)
)

# Add both the category and the breakpoint
df$with_columns(
  qcut = pl$col("foo")$qcut(c(0.25, 0.75), include_breaks = TRUE)
)$unnest("qcut")
}
\seealso{
\code{\link[=Expr_cut]{$cut()}}
}
