% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_rechunk}
\alias{Expr_rechunk}
\alias{rechunk}
\title{Rechunk memory layout}
\format{
a method
}
\usage{
Expr_rechunk
}
\value{
Expr
}
\description{
Create a single chunk of memory for this Series.
}
\examples{
pl$DataFrame(list(a=1:3,b=4:6))$select(pl$col("a")$append(pl$col("b")))$rechunk()
}
\keyword{Expr}
