% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_reinterpret}
\alias{Expr_reinterpret}
\title{Reinterpret bits}
\usage{
Expr_reinterpret(signed = TRUE)
}
\arguments{
\item{signed}{If \code{TRUE} (default), reinterpret into Int64. Otherwise, it
will be reinterpreted in UInt64.}
}
\value{
Expr
}
\description{
Reinterpret the underlying bits as a signed/unsigned integer. This
operation is only allowed for Int64. For lower bits integers, you can
safely use the cast operation.
}
\examples{
df = pl$DataFrame(x = 1:5, schema = list(x = pl$Int64))
df$select(pl$all()$reinterpret())
}
