% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_repeat_by}
\alias{Expr_repeat_by}
\title{Repeat values}
\usage{
Expr_repeat_by(by)
}
\arguments{
\item{by}{Expr that determines how often the values will be repeated. The
column will be coerced to UInt32.}
}
\value{
Expr
}
\description{
Repeat the elements in this Series as specified in the given expression.
The repeated elements are expanded into a \code{List}.
}
\examples{
df = pl$DataFrame(a = c("w", "x", "y", "z"), n = c(-1, 0, 1, 2))
df$with_columns(repeated = pl$col("a")$repeat_by("n"))
}
