% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_reverse}
\alias{Expr_reverse}
\title{Reverse a variable}
\usage{
Expr_reverse()
}
\value{
Expr
}
\description{
Reverse a variable
}
\examples{
pl$DataFrame(list(a = 1:5))$select(pl$col("a")$reverse())
}
