% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rle}
\alias{Expr_rle}
\title{Get the lengths of runs of identical values}
\usage{
Expr_rle()
}
\value{
Expr
}
\description{
Get the lengths of runs of identical values
}
\examples{
df = pl$DataFrame(s = c(1, 1, 2, 1, NA, 1, 3, 3))
df$select(pl$col("s")$rle())$unnest("s")
}
