% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rle_id}
\alias{Expr_rle_id}
\title{Map values to run IDs}
\usage{
<Expr>$rle_id()
}
\value{
Expr
}
\description{
Similar to $rle(), but it maps each value to an ID corresponding to the run
into which it falls. This is especially useful when you want to define groups
by runs of identical values rather than the values themselves. Note that
the ID is 0-indexed.
}
\examples{
df = pl$DataFrame(a = c(1, 2, 1, 1, 1, 4))
df$with_columns(a_r = pl$col("a")$rle_id())
}
