% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rolling}
\alias{Expr_rolling}
\title{Create rolling groups based on a time or numeric column}
\usage{
Expr_rolling(
  index_column,
  period,
  offset = NULL,
  closed = "right",
  check_sorted = TRUE
)
}
\arguments{
\item{index_column}{Column used to group based on the time window. Often of
type Date/Datetime. This column must be sorted in ascending order. If this
column represents an index, it has to be either Int32 or Int64. Note that
Int32 gets temporarily cast to Int64, so if performance matters use an Int64
column.}

\item{period}{Length of the window, must be non-negative.}

\item{offset}{Offset of the window. Default is \code{-period}.}

\item{closed}{Define which sides of the temporal interval are closed
(inclusive). This can be either \code{"left"}, \code{"right"}, \code{"both"} or \code{"none"}.}

\item{check_sorted}{Check whether data is actually sorted. Checking it is
expensive so if you are sure the data within the \code{index_column} is sorted, you
can set this to \code{FALSE} but note that if the data actually is unsorted, it
will lead to incorrect output.}
}
\value{
Expr
}
\description{
Create rolling groups based on a time or numeric column
}
\examples{
# create a DataFrame with a Datetime column and an f64 column
dates = c("2020-01-01 13:45:48", "2020-01-01 16:42:13", "2020-01-01 16:45:09",
          "2020-01-02 18:12:48", "2020-01-03 19:45:32", "2020-01-08 23:16:43")

df = pl$DataFrame(dt = dates, a = c(3, 7, 5, 9, 2, 1))$
  with_columns(
    pl$col("dt")$str$strptime(pl$Datetime(tu = "us"), format = "\%Y-\%m-\%d \%H:\%M:\%S")$set_sorted()
  )

df$with_columns(
  sum_a=pl$sum("a")$rolling(index_column="dt", period="2d"),
  min_a=pl$min("a")$rolling(index_column="dt", period="2d"),
  max_a=pl$max("a")$rolling(index_column="dt", period="2d")
)

# we can use "offset" to change the start of the window period. Here, with
# offset = "1d", we start the window one day after the value in "dt", and
# then we add a 2-day window relative to the window start.
df$with_columns(
  sum_a_offset1 = pl$sum("a")$rolling(index_column = "dt", period = "2d", offset = "1d"),
  sum_a_offset2 = pl$sum("a")$rolling(index_column = "dt", period = "2d", offset = "2d")
)
}
