% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rolling_skew}
\alias{Expr_rolling_skew}
\title{Rolling skew}
\usage{
Expr_rolling_skew(window_size, bias = TRUE)
}
\arguments{
\item{window_size}{Integer specifying the length of the window.}

\item{bias}{If \code{FALSE}, the calculations are corrected for statistical bias.}
}
\value{
Expr
}
\description{
Compute the rolling (= moving) skewness over the values in this array. A
window of length \code{window_size} will traverse the array.
}
\details{
For normally distributed data, the skewness should be about zero. For
uni-modal continuous distributions, a skewness value greater than zero means
that there is more weight in the right tail of the distribution.
}
\examples{
pl$DataFrame(a = c(1, 3, 2, 4, 5, 6))$
  with_columns(roll_skew = pl$col("a")$rolling_skew(window_size = 2))
}
