% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rolling_std}
\alias{Expr_rolling_std}
\title{Rolling standard deviation}
\usage{
Expr_rolling_std(
  window_size,
  weights = NULL,
  min_periods = NULL,
  ...,
  center = FALSE,
  ddof = 1
)
}
\arguments{
\item{window_size}{Integer specifying the length of the window.}

\item{weights}{An optional slice with the same length as the window that will
be multiplied elementwise with the values in the window.}

\item{min_periods}{The number of values in the window that should be non-null
before computing a result. If \code{NULL}, it will be set equal to window size.}

\item{...}{Ignored.}

\item{center}{Set the labels at the center of the window}

\item{ddof}{An integer representing "Delta Degrees of Freedom":
the divisor used in the calculation is \code{N - ddof},
where \code{N} represents the number of elements. By default ddof is \code{1}.}
}
\value{
Expr
}
\description{
Compute the rolling (= moving) standard deviation over the values in this
array. A window of length \code{window_size} will traverse the array. The values
that fill this window will (optionally) be multiplied with the weights given
by the \code{weight} vector.
}
\details{
If you want to compute multiple aggregation statistics over the same dynamic
window, consider using \verb{$rolling()} this method can cache the window size
computation.
}
\examples{
pl$DataFrame(a = c(1, 3, 2, 4, 5, 6))$
  with_columns(roll_std = pl$col("a")$rolling_std(window_size = 2))
}
