% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rolling_sum}
\alias{Expr_rolling_sum}
\title{Rolling sum}
\usage{
Expr_rolling_sum(
  window_size,
  weights = NULL,
  min_periods = NULL,
  center = FALSE
)
}
\arguments{
\item{window_size}{Integer specifying the length of the window.}

\item{weights}{An optional slice with the same length as the window that will
be multiplied elementwise with the values in the window.}

\item{min_periods}{The number of values in the window that should be non-null
before computing a result. If \code{NULL}, it will be set equal to window size.}

\item{center}{Set the labels at the center of the window}
}
\value{
Expr
}
\description{
Compute the rolling (= moving) sum over the values in this array. A window of
length \code{window_size} will traverse the array. The values that fill this window
will (optionally) be multiplied with the weights given by the \code{weight} vector.
}
\details{
If you want to compute multiple aggregation statistics over the same dynamic
window, consider using \verb{$rolling()} this method can cache the window size
computation.
}
\examples{
pl$DataFrame(a = c(1, 3, 2, 4, 5, 6))$
  with_columns(roll_sum = pl$col("a")$rolling_sum(window_size = 2))
}
