% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_round}
\alias{Expr_round}
\title{Round}
\usage{
Expr_round(decimals)
}
\arguments{
\item{decimals}{Number of decimals to round by.}
}
\value{
Expr
}
\description{
Round underlying floating point data by \code{decimals} digits.
}
\examples{
pl$DataFrame(a = c(0.33, 0.5, 1.02, 1.5, NaN, NA, Inf, -Inf))$with_columns(
  round = pl$col("a")$round(1)
)
}
