% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sample}
\alias{Expr_sample}
\title{Take a sample}
\usage{
<Expr>$sample(
  frac = NULL,
  with_replacement = TRUE,
  shuffle = FALSE,
  seed = NULL,
  n = NULL
)
}
\arguments{
\item{frac}{Fraction of items to return (can be higher than 1). Cannot be
used with \code{n}.}

\item{with_replacement}{If \code{TRUE} (default), allow values to be sampled more
than once.}

\item{shuffle}{Shuffle the order of sampled data points (implicitly \code{TRUE} if
\code{with_replacement = TRUE}).}

\item{seed}{numeric value of 0 to 2^52 Seed for the random number generator.
If \code{NULL} (default), a random seed value between 0 and 10000 is picked.}

\item{n}{Number of items to return. Cannot be used with \code{frac}.}
}
\value{
Expr
}
\description{
Take a sample
}
\examples{
df = pl$DataFrame(a = 1:4)
df$select(pl$col("a")$sample(frac = 1, with_replacement = TRUE, seed = 1L))
df$select(pl$col("a")$sample(frac = 2, with_replacement = TRUE, seed = 1L))
df$select(pl$col("a")$sample(n = 2, with_replacement = FALSE, seed = 1L))
}
