% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_search_sorted}
\alias{Expr_search_sorted}
\title{Where to inject element(s) to maintain sorting}
\usage{
Expr_search_sorted(element)
}
\arguments{
\item{element}{Expr or scalar value.}
}
\value{
Expr
}
\description{
Find the index in self where the element should be inserted so that it doesn't
break sortedness.
}
\details{
This function looks up where to insert element to keep self column sorted.
It is assumed the self column is already sorted in ascending order (otherwise
this leads to wrong results).
}
\examples{
df = pl$DataFrame(a = c(1, 3, 4, 4, 6))
df

# in which row should 5 be inserted in order to not break the sort?
# (value is 0-indexed)
df$select(pl$col("a")$search_sorted(5))
}
