% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_set_sorted}
\alias{Expr_set_sorted}
\title{Flag an Expr as "sorted"}
\usage{
Expr_set_sorted(descending = FALSE)
}
\arguments{
\item{descending}{Sort the columns in descending order.}
}
\value{
Expr
}
\description{
This enables downstream code to use fast paths for sorted arrays. WARNING:
this doesn't check whether the data is actually sorted, you have to ensure of
that yourself.
}
\examples{
# correct use flag something correctly as ascendingly sorted
s = pl$select(pl$lit(1:4)$set_sorted()$alias("a"))$get_column("a")
s$flags

# incorrect use, flag something as not sorted ascendingly
s2 = pl$select(pl$lit(c(1, 3, 2, 4))$set_sorted()$alias("a"))$get_column("a")
s2$sort()
s2$flags # returns TRUE while it's not actually sorted
}
