% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_shrink_dtype}
\alias{Expr_shrink_dtype}
\title{Shrink numeric columns to the minimal required datatype}
\usage{
Expr_shrink_dtype()
}
\value{
Expr
}
\description{
Shrink to the dtype needed to fit the extrema of this Series. This can be
used to reduce memory pressure.
}
\examples{
df = pl$DataFrame(
  a = 1:3,
  b = c(1, 2, 3)
)
df

df$with_columns(pl$all()$shrink_dtype()$name$suffix("_shrunk"))
}
