% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_shuffle}
\alias{Expr_shuffle}
\title{Shuffle values}
\usage{
Expr_shuffle(seed = NULL)
}
\arguments{
\item{seed}{numeric value of 0 to 2^52 Seed for the random number generator.
If \code{NULL} (default), a random seed value between 0 and 10000 is picked.}
}
\value{
Expr
}
\description{
Shuffle values
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(shuff = pl$col("a")$shuffle(seed = 1))
}
