% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sign}
\alias{Expr_sign}
\title{Get the sign of elements}
\usage{
Expr_sign
}
\value{
Expr
}
\description{
Get the sign of elements
}
\examples{
pl$DataFrame(a = c(.9, -3, -0, 0, 4, NA_real_))$
  with_columns(sign = pl$col("a")$sign())
}
\keyword{datasets}
