% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sin}
\alias{Expr_sin}
\title{Compute sine}
\usage{
Expr_sin()
}
\value{
Expr
}
\description{
Compute sine
}
\examples{
pl$DataFrame(a = c(0, pi / 2, pi, NA_real_))$
  with_columns(sine = pl$col("a")$sin())
}
