% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sinh}
\alias{Expr_sinh}
\alias{sinh}
\title{Sinh}
\usage{
Expr_sinh
}
\value{
Expr
}
\description{
Compute the element-wise value for the hyperbolic sine.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a = c(-1, asinh(0.5), 0, 1, NA_real_))$select(pl$col("a")$sinh())
}
\keyword{Expr}
