% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_slice}
\alias{Expr_slice}
\alias{slice}
\title{Get a slice of an Expr}
\usage{
Expr_slice(offset, length = NULL)
}
\arguments{
\item{offset}{Numeric or expression, zero-indexed. Indicates where to start
the slice. A negative value is one-indexed and starts from the end.}

\item{length}{Maximum number of elements contained in the slice. Default is
full data.}
}
\value{
Expr
}
\description{
Performing a slice of length 1 on a subset of columns will recycle this value
in those columns but will not change the number of rows in the data. See
examples.
}
\examples{

# as head
pl$DataFrame(list(a = 0:100))$select(
  pl$all()$slice(0, 6)
)

# as tail
pl$DataFrame(list(a = 0:100))$select(
  pl$all()$slice(-6, 6)
)

pl$DataFrame(list(a = 0:100))$select(
  pl$all()$slice(80)
)

# recycling
pl$DataFrame(mtcars)$with_columns(pl$col("mpg")$slice(0, 1))
}
