% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sort}
\alias{Expr_sort}
\title{Sort an Expr}
\usage{
Expr_sort(..., descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{...}{Ignored.}

\item{descending}{A logical. If \code{TRUE}, sort in descending order.}

\item{nulls_last}{A logical. If \code{TRUE}, place \code{null} values last insead of first.}
}
\value{
Expr
}
\description{
Sort this column. If used in a groupby context, the groups are sorted.
}
\examples{
pl$DataFrame(a = c(6, 1, 0, NA, Inf, NaN))$
  with_columns(sorted = pl$col("a")$sort())
}
