% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sqrt}
\alias{Expr_sqrt}
\alias{sqrt}
\title{Square root}
\usage{
Expr_sqrt()
}
\value{
Expr
}
\description{
Compute the square root of the elements.
}
\examples{
pl$DataFrame(list(a = -1:3))$select(pl$col("a")$sqrt())
}
\keyword{Expr}
