% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_std}
\alias{Expr_std}
\title{Get standard deviation}
\usage{
Expr_std(ddof = 1)
}
\arguments{
\item{ddof}{An integer representing "Delta Degrees of Freedom":
the divisor used in the calculation is \code{N - ddof},
where \code{N} represents the number of elements. By default ddof is \code{1}.}
}
\value{
Expr
}
\description{
Get standard deviation
}
\examples{
pl$select(pl$lit(1:5)$std())
}
