% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_str}
\alias{Expr_str}
\alias{str_ns}
\title{str: string related methods}
\usage{
Expr_str()
}
\value{
Expr
}
\description{
Create an object namespace of all string related methods.
See the individual method pages for full details
}
\examples{

# missing

}
