% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_decode}
\alias{expr_str_decode}
\title{Decode a value using the provided encoding}
\usage{
expr_str_decode(encoding, ..., strict = TRUE)
}
\arguments{
\item{encoding}{Either 'hex' or 'base64'.}

\item{...}{These dots are for future extensions and must be empty.}

\item{strict}{If \code{TRUE} (default), raise an error if the underlying value
cannot be decoded. Otherwise, replace it with a null value.}
}
\value{
A polars \link{expression}
}
\description{
Decode a value using the provided encoding
}
\examples{
df <- pl$DataFrame(strings = c("foo", "bar", NA))
df$select(pl$col("strings")$str$encode("hex"))
df$with_columns(
  pl$col("strings")$str$encode("base64")$alias("base64"), # notice DataType is not encoded
  pl$col("strings")$str$encode("hex")$alias("hex") # ... and must restored with cast
)$with_columns(
  pl$col("base64")$str$decode("base64")$alias("base64_decoded")$cast(pl$String),
  pl$col("hex")$str$decode("hex")$alias("hex_decoded")$cast(pl$String)
)
}
