% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_escape_regex}
\alias{expr_str_escape_regex}
\title{Returns string values with all regular expression meta characters escaped}
\usage{
expr_str_escape_regex()
}
\value{
A polars \link{expression}
}
\description{
Returns string values with all regular expression meta characters escaped
}
\examples{
df <- pl$DataFrame(text = c("abc", "def", NA, r"(abc(\w+))"))
df$with_columns(escaped = pl$col("text")$str$escape_regex())
}
