% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_extract_all}
\alias{expr_str_extract_all}
\title{Extract all matches for the given regex pattern}
\usage{
expr_str_extract_all(pattern)
}
\arguments{
\item{pattern}{A valid regex pattern}
}
\value{
A polars \link{expression}
}
\description{
Extracts all matches for the given regex pattern. Extracts each
successive non-overlapping regex match in an individual string as an array.
}
\examples{
df <- pl$DataFrame(foo = c("123 bla 45 asd", "xyz 678 910t"))
df$select(
  pl$col("foo")$str$extract_all(r"((\d+))")$alias("extracted_nrs")
)
}
