% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_extract_groups}
\alias{expr_str_extract_groups}
\title{Extract all capture groups for the given regex pattern}
\usage{
expr_str_extract_groups(pattern)
}
\arguments{
\item{pattern}{A character of a valid regular expression pattern containing
at least one capture group, compatible with the
\href{https://docs.rs/regex/latest/regex/}{regex crate}.}
}
\value{
A polars \link{expression}
}
\description{
Extract all capture groups for the given regex pattern
}
\details{
All group names are strings. If your pattern contains unnamed groups, their
numerical position is converted to a string. See examples.
}
\examples{
df <- pl$DataFrame(
  url = c(
    "http://vote.com/ballon_dor?candidate=messi&ref=python",
    "http://vote.com/ballon_dor?candidate=weghorst&ref=polars",
    "http://vote.com/ballon_dor?error=404&ref=rust"
  )
)

pattern <- r"(candidate=(?<candidate>\w+)&ref=(?<ref>\w+))"

df$with_columns(
  captures = pl$col("url")$str$extract_groups(pattern)
)$unnest("captures")

# If the groups are unnamed, their numerical position (as a string) is used:

pattern <- r"(candidate=(\w+)&ref=(\w+))"

df$with_columns(
  captures = pl$col("url")$str$extract_groups(pattern)
)$unnest("captures")
}
