% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_extract_many}
\alias{expr_str_extract_many}
\title{Use the Aho-Corasick algorithm to extract matches}
\usage{
expr_str_extract_many(
  patterns,
  ...,
  ascii_case_insensitive = FALSE,
  overlapping = FALSE,
  leftmost = FALSE
)
}
\arguments{
\item{patterns}{String patterns to search. Accepts expression input. Strings
are parsed as column names, and other non-expression inputs are parsed as
literals. To use the same character vector for all rows, use
\code{list(c(...))} instead of \code{c(...)} (see Examples).}

\item{...}{These dots are for future extensions and must be empty.}

\item{ascii_case_insensitive}{Enable ASCII-aware case insensitive matching.
When this option is enabled, searching will be performed without respect to
case for ASCII letters (a-z and A-Z) only.}

\item{overlapping}{Whether matches can overlap.}

\item{leftmost}{Whether to guarantee in case there are overlapping matches that
the leftmost match is used. In case there are multiple candidates for the
leftmost match, the pattern which comes first in \code{patterns} is used.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This method supports matching on string literals only, and does not support
regular expression matching.
}
\examples{
df <- pl$DataFrame(values = "discontent")
patterns <- list(c("winter", "disco", "onte", "discontent"))

df$with_columns(
  matches = pl$col("values")$str$extract_many(patterns),
  matches_overlap = pl$col("values")$str$extract_many(patterns, overlapping = TRUE)
)

df <- pl$DataFrame(
  values = c("discontent", "rhapsody"),
  patterns = list(c("winter", "disco", "onte", "discontent"), c("rhap", "ody", "coalesce"))
)

df$select(pl$col("values")$str$extract_many("patterns"))
}
