% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_join}
\alias{expr_str_join}
\title{Vertically concatenate the string values in the column to a single string value.}
\usage{
expr_str_join(delimiter = "", ..., ignore_nulls = TRUE)
}
\arguments{
\item{delimiter}{The delimiter to insert between consecutive string values.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ignore_nulls}{Ignore null values (default). If \code{FALSE}, null values will be
propagated: if the column contains any null values, the output is null.}
}
\value{
A polars \link{expression}
}
\description{
Vertically concatenate the string values in the column to a single string value.
}
\examples{
# concatenate a Series of strings to a single string
df <- pl$DataFrame(foo = c(1, NA, 2))

df$select(pl$col("foo")$str$join("-"))

df$select(pl$col("foo")$str$join("-", ignore_nulls = FALSE))
}
