% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_json_decode}
\alias{expr_str_json_decode}
\title{Parse string values as JSON.}
\usage{
expr_str_json_decode(dtype, ...)
}
\arguments{
\item{dtype}{The dtype to cast the extracted value to.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A polars \link{expression}
}
\description{
Parse string values as JSON.
Throw errors if encounter invalid json strings.
}
\examples{
df <- pl$DataFrame(
  json_val = c('{"a":1, "b": true}', NA, '{"a":2, "b": false}')
)

dtype <- pl$Struct(a = pl$UInt8, b = pl$Boolean)
df$select(
  pl$col("json_val")$str$json_decode(dtype)
)$unnest("json_val")
}
\seealso{
\itemize{
\item \code{\link[=series_str_json_decode]{<series>$str$json_decode()}}
}
}
