% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_json_path_match}
\alias{expr_str_json_path_match}
\title{Extract the first match of JSON string with the provided JSONPath expression}
\usage{
expr_str_json_path_match(json_path)
}
\arguments{
\item{json_path}{A valid JSON path query string.}
}
\value{
A polars \link{expression}
}
\description{
Extract the first match of JSON string with the provided JSONPath expression
}
\details{
Throw errors if encounter invalid JSON strings. All return value will be
cast to String regardless of the original value.

Documentation on JSONPath standard can be found here: \url{https://goessner.net/articles/JsonPath/}.
}
\examples{
df <- pl$DataFrame(
  json_val = c('{"a":"1"}', NA, '{"a":2}', '{"a":2.1}', '{"a":true}')
)
df$select(pl$col("json_val")$str$json_path_match("$.a"))
}
