% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_len_chars}
\alias{expr_str_len_chars}
\title{Get the number of characters in strings}
\usage{
expr_str_len_chars()
}
\value{
A polars \link{expression}
}
\description{
Get length of the strings as UInt32 (as number of characters). Use
\verb{$str$len_bytes()} to get the number of bytes.
}
\details{
If you know that you are working with ASCII text, \code{lengths} will be
equivalent, and faster (returns length in terms of the number of bytes).
}
\examples{
pl$DataFrame(
  s = c("Café", NA, "345", "æøå")
)$select(
  pl$col("s"),
  pl$col("s")$str$len_bytes()$alias("lengths"),
  pl$col("s")$str$len_chars()$alias("n_chars")
)
}
