% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_normalize}
\alias{expr_str_normalize}
\title{Returns the Unicode normal form of the string values}
\usage{
expr_str_normalize(form = c("NFC", "NFKC", "NFD", "NFKD"))
}
\arguments{
\item{form}{Unicode form to use. Must be one of: \code{"NFC"}, \code{"NFKC"}, \code{"NFD"},
\code{"NFKD"}.}
}
\value{
A polars \link{expression}
}
\description{
This uses the forms described in Unicode Standard Annex 15:
\url{https://www.unicode.org/reports/tr15/}.
}
\examples{
df <- pl$DataFrame(text = c("01²", "ＫＡＤＯＫＡＷＡ"))

new <- df$with_columns(
  nfc = pl$col("text")$str$normalize("NFC"),
  nfkc = pl$col("text")$str$normalize("NFKC"),
)
new

new$select(pl$all()$str$len_bytes())
}
