% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_pad_start}
\alias{expr_str_pad_start}
\title{Right justify strings}
\usage{
expr_str_pad_start(length, fill_char = " ")
}
\arguments{
\item{length}{Pad the string until it reaches this length.
Strings with length equal to or greater than this value are returned as-is.
Can be integer or \link[=Expr]{expression}.}

\item{fill_char}{Fill with this ASCII character.}
}
\value{
A polars \link{expression}
}
\description{
Return the string right justified in a string of length \code{length}.
}
\examples{
df <- pl$DataFrame(a = c("cow", "monkey", NA, "hippopotamus"))
df$select(pl$col("a")$str$pad_start(8, "*"))
}
