% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_starts_with}
\alias{expr_str_starts_with}
\title{Check if string starts with a regex}
\usage{
expr_str_starts_with(prefix)
}
\arguments{
\item{prefix}{Prefix substring or Expr.}
}
\value{
A polars \link{expression}
}
\description{
Check if string values starts with a substring.
}
\details{
See also \verb{$str$contains()} and \verb{$str$ends_with()}.
}
\examples{
df <- pl$DataFrame(fruits = c("apple", "mango", NA))
df$select(
  pl$col("fruits"),
  pl$col("fruits")$str$starts_with("app")$alias("has_suffix")
)
}
