% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_strip_chars}
\alias{expr_str_strip_chars}
\title{Strip leading and trailing characters}
\usage{
expr_str_strip_chars(characters = NULL)
}
\arguments{
\item{characters}{The set of characters to be removed. All combinations of this
set of characters will be stripped. If \code{NULL} (default), all whitespace is
removed instead. This can be an Expr.}
}
\value{
A polars \link{expression}
}
\description{
Remove leading and trailing characters.
}
\details{
This function will not strip any chars beyond the first char not matched.
\code{strip_chars()} removes characters at the beginning and the end of the string.
Use \code{strip_chars_start()} and \code{strip_chars_end()} to remove characters only
from left and right respectively.
}
\examples{
df <- pl$DataFrame(foo = c(" hello", "\\tworld"))
df$select(pl$col("foo")$str$strip_chars())
df$select(pl$col("foo")$str$strip_chars(" hel rld"))
}
