% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_strip_chars_end}
\alias{expr_str_strip_chars_end}
\title{Strip trailing characters}
\usage{
expr_str_strip_chars_end(characters = NULL)
}
\arguments{
\item{characters}{The set of characters to be removed. All combinations of this
set of characters will be stripped. If \code{NULL} (default), all whitespace is
removed instead. This can be an Expr.}
}
\value{
A polars \link{expression}
}
\description{
Remove trailing characters.
}
\details{
This function will not strip any chars beyond the first char not matched.
\code{strip_chars_end()} removes characters at the end of the string only.
Use \code{strip_chars()} and \code{strip_chars_start()} to remove characters from the left
and right or only from the left respectively.
}
\examples{
df <- pl$DataFrame(foo = c(" hello", "\\tworld"))
df$select(pl$col("foo")$str$strip_chars_end(" hel\\trld"))
df$select(pl$col("foo")$str$strip_chars_end("rldhel\\t "))
}
