% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_strip_chars_start}
\alias{expr_str_strip_chars_start}
\title{Strip leading characters}
\usage{
expr_str_strip_chars_start(characters = NULL)
}
\arguments{
\item{characters}{The set of characters to be removed. All combinations of this
set of characters will be stripped. If \code{NULL} (default), all whitespace is
removed instead. This can be an Expr.}
}
\value{
A polars \link{expression}
}
\description{
Remove leading characters.
}
\details{
This function will not strip any chars beyond the first char not matched.
\code{strip_chars_start()} removes characters at the beginning of the string only.
Use \code{strip_chars()} and \code{strip_chars_end()} to remove characters from the left
and right or only from the right respectively.
}
\examples{
df <- pl$DataFrame(foo = c(" hello", "\\tworld"))
df$select(pl$col("foo")$str$strip_chars_start(" hel rld"))
}
