% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_strip_prefix}
\alias{expr_str_strip_prefix}
\title{Strip prefix}
\usage{
expr_str_strip_prefix(prefix = NULL)
}
\arguments{
\item{prefix}{The prefix to be removed.}
}
\value{
A polars \link{expression}
}
\description{
The prefix will be removed from the string exactly once, if found.
}
\details{
This method strips the exact character sequence provided in \code{prefix} from
the start of the input. To strip a set of characters in any order, use
\code{\link[=expr_str_strip_chars_start]{$strip_chars_start()}} instead.
}
\examples{
df <- pl$DataFrame(a = c("foobar", "foofoobar", "foo", "bar"))
df$with_columns(
  stripped = pl$col("a")$str$strip_prefix("foo")
)
}
