% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_strip_suffix}
\alias{expr_str_strip_suffix}
\title{Strip suffix}
\usage{
expr_str_strip_suffix(suffix = NULL)
}
\arguments{
\item{suffix}{The suffix to be removed.}
}
\value{
A polars \link{expression}
}
\description{
The suffix will be removed from the string exactly once, if found.
}
\details{
This method strips the exact character sequence provided in \code{suffix} from
the end of the input. To strip a set of characters in any order, use
\code{\link[=expr_str_strip_chars_end]{$strip_chars_end()}} instead.
}
\examples{
df <- pl$DataFrame(a = c("foobar", "foobarbar", "foo", "bar"))
df$with_columns(
  stripped = pl$col("a")$str$strip_suffix("bar")
)
}
