% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_to_datetime}
\alias{expr_str_to_datetime}
\title{Convert a String column into a Datetime column}
\usage{
expr_str_to_datetime(
  format = NULL,
  ...,
  time_unit = NULL,
  time_zone = NULL,
  strict = TRUE,
  exact = TRUE,
  cache = TRUE,
  ambiguous = c("raise", "earliest", "latest", "null")
)
}
\arguments{
\item{format}{Format to use for conversion. Refer to
\href{https://docs.rs/chrono/latest/chrono/format/strftime/index.html}{the chrono crate documentation}
for the full specification. Example: \code{"\%Y-\%m-\%d \%H:\%M:\%S"}.
If \code{NULL} (default), the format is inferred from the data.
Notice that time zone \verb{\%Z} is not supported and will just ignore timezones.
Numeric time zones like \verb{\%z} or \verb{\%:z} are supported.}

\item{...}{These dots are for future extensions and must be empty.}

\item{time_unit}{Unit of time for the resulting Datetime column. If \code{NULL} (default),
the time unit is inferred from the format string if given,
e.g.: \code{"\%F \%T\%.3f"} => \code{\link[=pl__Datetime]{pl$Datetime("ms")}}.
If no fractional second component is found, the default is \code{"us"} (microsecond).}

\item{time_zone}{for the resulting \link[=pl__Datetime]{Datetime} column.}

\item{strict}{If \code{TRUE} (default), raise an error if a single string cannot
be parsed. If \code{FALSE}, produce a polars \code{null}.}

\item{exact}{If \code{TRUE} (default), require an exact format match.
If \code{FALSE}, allow the format to match anywhere in the target string.
Note that using \code{exact = FALSE} introduces a performance
penalty - cleaning your data beforehand will almost certainly be more performant.}

\item{cache}{Use a cache of unique, converted dates to apply the datetime
conversion.}

\item{ambiguous}{Determine how to deal with ambiguous datetimes.
Character vector or \link{expression} containing the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"earliest"}: Use the earliest datetime
\item \code{"latest"}: Use the latest datetime
\item \code{"null"}: Return a null value
}}
}
\value{
A polars \link{expression}
}
\description{
Convert a String column into a Datetime column
}
\examples{
df <- pl$DataFrame(x = c("2020-01-01 01:00Z", "2020-01-01 02:00Z"))

df$select(pl$col("x")$str$to_datetime("\%Y-\%m-\%d \%H:\%M\%#z"))
df$select(pl$col("x")$str$to_datetime(time_zone = "UTC"))
}
\seealso{
\itemize{
\item \code{\link[=expr_str_strptime]{<expr>$str$strptime()}}
\item \code{\link[=series_str_to_datetime]{<series>$str$to_datetime()}}
}
}
