% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_to_lowercase}
\alias{expr_str_to_lowercase}
\title{Convert a string to lowercase}
\usage{
expr_str_to_lowercase()
}
\value{
A polars \link{expression}
}
\description{
Transform to lowercase variant.
}
\examples{
pl$select(
  pl$lit(c("A", "b", "c", "1", NA))$str$to_lowercase()
)$to_series()
}
