% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_to_time}
\alias{expr_str_to_time}
\title{Convert a String column into a Time column}
\usage{
expr_str_to_time(format = NULL, ..., strict = TRUE, cache = TRUE)
}
\arguments{
\item{format}{Format to use for conversion. Refer to
\href{https://docs.rs/chrono/latest/chrono/format/strftime/index.html}{the chrono crate documentation}
for the full specification. Example: \code{"\%Y-\%m-\%d \%H:\%M:\%S"}.
If \code{NULL} (default), the format is inferred from the data.
Notice that time zone \verb{\%Z} is not supported and will just ignore timezones.
Numeric time zones like \verb{\%z} or \verb{\%:z} are supported.}

\item{...}{These dots are for future extensions and must be empty.}

\item{strict}{If \code{TRUE} (default), raise an error if a single string cannot
be parsed. If \code{FALSE}, produce a polars \code{null}.}

\item{cache}{Use a cache of unique, converted dates to apply the datetime
conversion.}
}
\value{
A polars \link{expression}
}
\description{
Convert a String column into a Time column
}
\examples{
df <- pl$DataFrame(x = c("01:00", "02:00", "03:00"))

df$select(pl$col("x")$str$to_time("\%H:\%M"))
}
\seealso{
\itemize{
\item \code{\link[=expr_str_strptime]{<Expr>$str$strptime()}}
}
}
