% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_struct}
\alias{Expr_struct}
\alias{struct_ns}
\title{struct: related methods}
\usage{
Expr_struct()
}
\value{
Expr
}
\description{
Create an object namespace of all struct related methods.
See the individual method pages for full details
}
\examples{

#missing

}
\keyword{Expr}
