% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-struct.R
\name{expr_struct_drop}
\alias{expr_struct_drop}
\title{Drop one or more fields from the struct}
\usage{
expr_struct_drop(names, ..., strict = TRUE)
}
\arguments{
\item{names}{Character vector of the names of the fields to drop.}

\item{...}{These dots are for future extensions and must be empty.}

\item{strict}{If \code{TRUE} (default), raise an error if any of the specified
fields doesn't exist in the struct.}
}
\value{
A polars \link{expression}
}
\description{
Drop one or more fields from the struct
}
\examples{
df <- pl$DataFrame(
  aaa = c(1, 2),
  bbb = c("ab", "cd"),
  ccc = c(TRUE, NA)
)$select(struct_col = pl$struct("aaa", "bbb", "ccc"))
df

df$select(pl$col("struct_col")$struct$drop("aaa"))

# Dropping an unknown field is an error unless `strict = FALSE`
df$select(pl$col("struct_col")$struct$drop("zzz", strict = FALSE))
}
