% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-struct.R
\name{expr_struct_unnest}
\alias{expr_struct_unnest}
\title{Expand the struct into its individual fields}
\usage{
expr_struct_unnest()
}
\value{
A polars \link{expression}
}
\description{
This is an alias for \code{\link[=expr_struct_field]{Expr$struct$field("*")}}.
}
\examples{
df <- pl$DataFrame(
  aaa = c(1, 2),
  bbb = c("ab", "cd"),
  ccc = c(TRUE, NA),
  ddd = list(1:2, 3)
)$select(struct_col = pl$struct("aaa", "bbb", "ccc", "ddd"))
df

df$select(pl$col("struct_col")$struct$unnest())
}
