% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sub}
\alias{Expr_sub}
\alias{-.Expr}
\title{Sub}
\usage{
Expr_sub(other)

\method{-}{Expr}(e1, e2)
}
\arguments{
\item{other}{literal or Robj which can become a literal}

\item{e1}{lhs Expr}

\item{e2}{rhs Expr or anything which can become a literal Expression}
}
\value{
Exprs
}
\description{
Substract
}
\examples{
# three syntaxes same result
pl$lit(5) - 10
pl$lit(5) - pl$lit(10)
pl$lit(5)$sub(pl$lit(10))
-pl$lit(5)
}
\keyword{Expr}
\keyword{Expr_operators}
