% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_suffix}
\alias{Expr_suffix}
\alias{suffix}
\title{Add a suffix to a column name}
\usage{
Expr_suffix(suffix)
}
\arguments{
\item{suffix}{Suffix to be added to column name(s)}
}
\value{
Expr
}
\description{
Add a suffix to a column name
}
\examples{
dat = pl$DataFrame(mtcars)

dat$select(
  pl$col("mpg"),
  pl$col("mpg")$suffix("_foo"),
  pl$col("cyl", "drat")$suffix("_bar")
)
}
\seealso{
\code{\link[=Expr_prefix]{$prefix()}} to add a prefix
}
\keyword{Expr}
