% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_sum}
\alias{Expr_sum}
\title{sum}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_sum
}
\value{
Expr
}
\description{
Get sum value
}
\examples{
pl$DataFrame(list(x=c(1L,NA,2L)))$select(pl$col("x")$sum())#is i32 3 (Int32 not casted)
}
\keyword{Expr}
