% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_tan}
\alias{Expr_tan}
\alias{Tan}
\title{Tan}
\usage{
Expr_tan
}
\value{
Expr
}
\description{
Compute the element-wise value for the tangent.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a = c(0, pi / 2, pi, NA_real_))$select(pl$col("a")$tan())
}
\keyword{Expr}
