% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_tanh}
\alias{Expr_tanh}
\title{Compute hyperbolic tangent}
\usage{
Expr_tanh
}
\value{
Expr
}
\description{
Compute hyperbolic tangent
}
\examples{
pl$DataFrame(a = c(-1, atanh(0.5), 0, 1, NA_real_))$
  with_columns(tanh = pl$col("a")$tanh())
}
\keyword{datasets}
