% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_to_physical}
\alias{Expr_to_physical}
\alias{to_physical}
\title{To physical representation}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_to_physical
}
\value{
Expr
}
\description{
expression request underlying physical base representation
}
\examples{
pl$DataFrame(
  list(vals = c("a", "x", NA, "a"))
)$with_columns(
  pl$col("vals")$cast(pl$Categorical),
  pl$col("vals")
  $cast(pl$Categorical)
  $to_physical()
  $alias("vals_physical")
)
}
\keyword{Expr}
