% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_to_series}
\alias{Expr_to_series}
\title{Convert Literal to Series}
\usage{
Expr_to_series()
}
\value{
Series
}
\description{
Collect an expression based on literals into a Series.
}
\examples{
pl$lit(1:5)$to_series()
}
