% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_to_struct}
\alias{Expr_to_struct}
\title{Convert an Expr to a Struct}
\usage{
<Expr>$to_struct()
}
\value{
Expr
}
\description{
Convert an Expr to a Struct
}
\examples{
pl$DataFrame(iris[, 3:5])$with_columns(
  my_struct = pl$all()$to_struct()
)
}
