% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_unique}
\alias{Expr_unique}
\title{get unique values}
\usage{
Expr_unique(maintain_order = FALSE)
}
\arguments{
\item{maintain_order}{bool, if TRUE guaranteed same order, if FALSE maybe}
}
\value{
Expr
}
\description{
Get unique values of this expression.
Similar to R unique()
}
\examples{
pl$DataFrame(iris)$select(pl$col("Species")$unique())
}
\keyword{Expr}
