% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_upper_lower_bound}
\alias{Expr_upper_lower_bound}
\alias{Expr_upper_bound}
\alias{upper_bound}
\alias{Expr_lower_bound}
\alias{lower_bound}
\title{Upper bound}
\usage{
Expr_upper_bound

Expr_lower_bound
}
\value{
Expr
}
\description{
Calculate the upper/lower bound.
Returns a unit Series with the highest value possible for the dtype of this
expression.
}
\details{
Notice lower bound i32 exported to R is NA_integer_ for now
}
\examples{
pl$DataFrame(i32 = 1L, f64 = 1)$select(pl$all()$upper_bound())
pl$DataFrame(i32 = 1L, f64 = 1)$select(pl$all()$lower_bound())
}
\keyword{datasets}
