% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_value_counts}
\alias{Expr_value_counts}
\title{Value counts}
\usage{
Expr_value_counts(..., sort = FALSE, parallel = FALSE, name, normalize = FALSE)
}
\arguments{
\item{...}{Ignored.}

\item{sort}{Ensure the output is sorted from most values to least.}

\item{parallel}{Better to turn this off in the aggregation context, as it can
lead to contention.}

\item{name}{Give the resulting count column a specific name. The default is
\code{"count"} if \code{normalize = FALSE} and \code{"proportion"} if  \code{normalize = TRUE}.}

\item{normalize}{If \code{TRUE}, it gives relative frequencies of the unique
values instead of their count.}
}
\value{
Expr
}
\description{
Count all unique values and create a struct mapping value to count.
}
\examples{
df = pl$DataFrame(iris)
df$select(pl$col("Species")$value_counts())$unnest()
df$select(pl$col("Species")$value_counts(normalize = TRUE))$unnest()
}
