% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_all}
\alias{ExprArr_all}
\title{Evaluate whether all boolean values in an array are true}
\usage{
ExprArr_all()
}
\value{
Expr
}
\description{
Evaluate whether all boolean values in an array are true
}
\examples{
df = pl$DataFrame(
  list(values = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), c(NA, NA))),
  schema = list(values = pl$Array(pl$Boolean, 2))
)
df$with_columns(all = pl$col("values")$arr$all())
}
