% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_any}
\alias{ExprArr_any}
\title{Evaluate whether any boolean values in an array are true}
\usage{
ExprArr_any()
}
\value{
Expr
}
\description{
Evaluate whether any boolean values in an array are true
}
\examples{
df = pl$DataFrame(
  list(values = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), c(NA, NA))),
  schema = list(values = pl$Array(pl$Boolean, 2))
)
df$with_columns(any = pl$col("values")$arr$any())
}
