% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_arg_max}
\alias{ExprArr_arg_max}
\alias{arr_arg_max}
\title{Get the index of the maximal value in an array}
\usage{
ExprArr_arg_max()
}
\value{
Expr
}
\description{
Get the index of the maximal value in an array
}
\examples{
df = pl$DataFrame(
  values = list(1:2, 2:1),
  schema = list(values = pl$Array(pl$Int32, 2))
)
df$with_columns(
  arg_max = pl$col("values")$arr$arg_max()
)
}
