% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_arg_min}
\alias{ExprArr_arg_min}
\alias{arr_arg_min}
\title{Get the index of the minimal value in an array}
\usage{
ExprArr_arg_min()
}
\value{
Expr
}
\description{
Get the index of the minimal value in an array
}
\examples{
df = pl$DataFrame(list(s = list(1:2, 2:1)))
df$with_columns(
  arg_min = pl$col("s")$arr$arg_min()
)
}
