% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_contains}
\alias{ExprArr_contains}
\alias{arr_contains}
\title{Check if array contains a given value}
\usage{
ExprArr_contains(item)
}
\arguments{
\item{item}{Expr or something coercible to an Expr. Strings are \emph{not} parsed
as columns.}
}
\value{
Expr
}
\description{
Check if array contains a given value
}
\examples{
df = pl$DataFrame(
  values = list(0:2, 4:6, c(NA_integer_, NA_integer_, NA_integer_)),
  item = 0:2,
  schema = list(values = pl$Array(pl$Float64, 3))
)
df$with_columns(
  with_expr = pl$col("values")$arr$contains(pl$col("item")),
  with_lit = pl$col("values")$arr$contains(1)
)
}
