% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_max}
\alias{ExprArr_max}
\alias{arr_max}
\title{Find the maximum value in an array}
\usage{
ExprArr_max()
}
\value{
Expr
}
\description{
Find the maximum value in an array
}
\examples{
df = pl$DataFrame(
  list(values = list(c(1, 2), c(3, 4), c(5, 6))),
  schema = list(values = pl$Array(pl$Float64, 2))
)
df$with_columns(max = pl$col("values")$arr$max())
}
