% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_median}
\alias{ExprArr_median}
\alias{arr_median}
\title{Find the median in an array}
\usage{
ExprArr_median()
}
\value{
Expr
}
\description{
Find the median in an array
}
\examples{
df = pl$DataFrame(
  values = list(c(2, 1, 4), c(8.4, 3.2, 1)),
  schema = list(values = pl$Array(pl$Float64, 3))
)
df$with_columns(median = pl$col("values")$arr$median())
}
