% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_reverse}
\alias{ExprArr_reverse}
\alias{arr_reverse}
\title{Reverse values in an array}
\usage{
ExprArr_reverse()
}
\value{
Expr
}
\description{
Reverse values in an array
}
\examples{
df = pl$DataFrame(list(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_)))
df$with_columns(reverse = pl$col("values")$arr$reverse())
}
