% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_to_list}
\alias{ExprArr_to_list}
\title{Convert an Array column into a List column with the same inner data type}
\usage{
ExprArr_to_list()
}
\value{
\link[=Expr_class]{Expr} of \link[=DataType_List]{data type List}
}
\description{
Convert an Array column into a List column with the same inner data type
}
\examples{
df = pl$DataFrame(
  a = list(c(1, 2), c(3, 4)),
  schema = list(a = pl$Array(pl$Int8, 2))
)

df$with_columns(
  list = pl$col("a")$arr$to_list()
)
}
