% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_unique}
\alias{ExprArr_unique}
\alias{arr_unique}
\title{Get unique values in a list}
\usage{
ExprArr_unique()
}
\value{
Expr
}
\description{
Get the unique/distinct values in the list.
}
\examples{
df = pl$DataFrame(list(values = list(c(2, 2, NA), c(1, 2, 3), NA_real_)))
df$with_columns(unique = pl$col("values")$arr$unique())
}
