% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_encode}
\alias{ExprBin_encode}
\title{Encode a value using the provided encoding}
\usage{
ExprBin_encode(encoding)
}
\arguments{
\item{encoding}{A character, \code{"hex"} or \code{"base64"}. The encoding to use.}
}
\value{
\link[=Expr_class]{Expr} of data type String.
}
\description{
Encode a value using the provided encoding
}
\examples{
df = pl$DataFrame(
  name = c("black", "yellow", "blue"),
  code = as_polars_series(
    c("000000", "ffff00", "0000ff")
  )$cast(pl$Binary)$bin$decode("hex")
)

df$with_columns(encoded = pl$col("code")$bin$encode("hex"))
}
