% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_size}
\alias{ExprBin_size}
\title{Get the size of binary values in the given unit}
\usage{
ExprBin_size(unit = "b")
}
\arguments{
\item{unit}{Scale the returned size to the given unit. Can be \code{"b"}, \code{"kb"},
\code{"mb"}, \code{"gb"}, \code{"tb"}, or their full names (\code{"kilobytes"}, etc.).}
}
\value{
\link[=Expr_class]{Expr} of data type UInt or Float.
}
\description{
Get the size of binary values in the given unit
}
\examples{
df = pl$DataFrame(
  name = c("black", "yellow", "blue"),
  code_hex = as_polars_series(c("000000", "ffff00", "0000ff"))$cast(pl$Binary)
)

df$with_columns(
  n_bytes = pl$col("code_hex")$bin$size(),
  n_kilobytes = pl$col("code_hex")$bin$size("kb")
)
}
