% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_starts_with}
\alias{ExprBin_starts_with}
\title{Check if values start with a binary substring}
\usage{
ExprBin_starts_with(sub)
}
\arguments{
\item{sub}{Prefix substring.}
}
\value{
Expr returing a Boolean
}
\description{
Check if values start with a binary substring
}
\examples{
colors = pl$DataFrame(
  name = c("black", "yellow", "blue"),
  code = as_polars_series(c("x00x00x00", "xffxffx00", "x00x00xff"))$cast(pl$Binary),
  lit = as_polars_series(c("x00", "xffx00", "xffxff"))$cast(pl$Binary)
)

colors$select(
  "name",
  starts_with_lit = pl$col("code")$bin$starts_with("xff"),
  starts_with_expr = pl$col("code")$bin$starts_with(pl$col("prefix"))
)
}
