% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__categorical.R
\name{ExprCat_get_categories}
\alias{ExprCat_get_categories}
\title{Get the categories stored in this data type}
\usage{
ExprCat_get_categories()
}
\value{
A polars DataFrame with the categories for each categorical Series.
}
\description{
Get the categories stored in this data type
}
\examples{
df = pl$DataFrame(
  cats = factor(c("z", "z", "k", "a", "b")),
  vals = factor(c(3, 1, 2, 2, 3))
)
df

df$select(
  pl$col("cats")$cat$get_categories()
)
df$select(
  pl$col("vals")$cat$get_categories()
)
}
\keyword{ExprCat}
