% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_combine}
\alias{ExprDT_combine}
\title{Combine Date and Time}
\usage{
ExprDT_combine(time, time_unit = "us")
}
\arguments{
\item{time}{The number of epoch since or before (if negative) the Date. Can be
an Expr or a PTime.}

\item{time_unit}{Unit of time. One of \code{"ms"}, \code{"us"} (default) or \code{"ns"}.}
}
\value{
Date/Datetime expr
}
\description{
If the underlying expression is a Datetime then its time component is
replaced, and if it is a Date then a new Datetime is created by combining
the two values.
}
\examples{
df = pl$DataFrame(
  dtm = c(
    ISOdatetime(2022, 12, 31, 10, 30, 45),
    ISOdatetime(2023, 7, 5, 23, 59, 59)
  ),
  dt = c(ISOdate(2022, 10, 10), ISOdate(2022, 7, 5)),
  tm = c(pl$time(1, 2, 3, 456000), pl$time(7, 8, 9, 101000))
)$explode("tm")

df

df$select(
  d1 = pl$col("dtm")$dt$combine(pl$col("tm")),
  s2 = pl$col("dt")$dt$combine(pl$col("tm")),
  d3 = pl$col("dt")$dt$combine(pl$time(4, 5, 6))
)
}
