% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_epoch}
\alias{ExprDT_epoch}
\title{Epoch}
\usage{
ExprDT_epoch(time_unit = "us")
}
\arguments{
\item{time_unit}{Time unit, one of \code{"ns"}, \code{"us"}, \code{"ms"}, \code{"s"} or  \code{"d"}.}
}
\value{
Expr with datatype Int64
}
\description{
Get the time passed since the Unix EPOCH in the give time unit.
}
\examples{
df = pl$DataFrame(date = pl$date_range(as.Date("2001-1-1"), as.Date("2001-1-3")))

df$with_columns(
  epoch_ns = pl$col("date")$dt$epoch(),
  epoch_s = pl$col("date")$dt$epoch(time_unit = "s")
)
}
