% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_hour}
\alias{ExprDT_hour}
\alias{(Expr)$dt$hour}
\title{Hour}
\format{
function
}
\value{
Expr of hour as UInt32
}
\description{
Extract hour from underlying Datetime representation.
Applies to Datetime columns.
Returns the hour number from 0 to 23.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d",
    time_zone = "GMT"
  )
)
df$with_columns(
  pl$col("date")$dt$hour()$alias("hour")
)
}
\keyword{ExprDT}
