% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_is_leap_year}
\alias{ExprDT_is_leap_year}
\title{Determine whether the year of the underlying date is a leap year}
\usage{
ExprDT_is_leap_year()
}
\value{
An Expr of datatype Bool
}
\description{
Determine whether the year of the underlying date is a leap year
}
\examples{
df = pl$DataFrame(date = as.Date(c("2000-01-01", "2001-01-01", "2002-01-01")))

df$with_columns(
  leap_year = pl$col("date")$dt$is_leap_year()
)
}
