% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_microsecond}
\alias{ExprDT_microsecond}
\title{Extract microseconds from underlying Datetime representation.}
\usage{
ExprDT_microsecond()
}
\value{
\link[=Expr_class]{Expr} of data type Int32
}
\description{
Applies to Datetime columns.
}
\examples{
df = pl$DataFrame(
  datetime = as.POSIXct(
    c(
      "1978-01-01 01:01:01",
      "2024-10-13 05:30:14.500",
      "2065-01-01 10:20:30.06"
    ),
    "UTC"
  )
)

df$with_columns(
  microsecond = pl$col("datetime")$dt$microsecond()
)
}
