% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_microseconds}
\alias{ExprDT_microseconds}
\alias{(Expr)$dt$microseconds}
\title{microseconds}
\format{
function
}
\value{
Expr of i64
}
\description{
Extract the microseconds from a Duration type.
}
\examples{
df = pl$DataFrame(date = pl$date_range(
  low = as.POSIXct("2020-1-1", tz = "GMT"),
  high = as.POSIXct("2020-1-1 00:00:01", tz = "GMT"),
  interval = "1ms",
  lazy = FALSE
))
df$select(
  pl$col("date"),
  pl$col("date")$diff()$dt$microseconds()$alias("seconds_diff")
)
}
\keyword{ExprDT}
