% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_minute}
\alias{ExprDT_minute}
\alias{(Expr)$dt$minute}
\title{Minute}
\format{
function
}
\value{
Expr of minute as UInt32
}
\description{
Extract minutes from underlying Datetime representation.
Applies to Datetime columns.
Returns the minute number from 0 to 59.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d",
    time_zone = "GMT",
    lazy = FALSE
  )
)
df$with_columns(
  pl$col("date")$dt$minute()$alias("minute")
)
}
\keyword{ExprDT}
