% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_second}
\alias{ExprDT_second}
\title{Extract seconds from underlying Datetime representation}
\usage{
ExprDT_second(fractional = FALSE)
}
\arguments{
\item{fractional}{A logical.
Whether to include the fractional component of the second.}
}
\value{
\link[=Expr_class]{Expr} of data type Int8 or Float64
}
\description{
Applies to Datetime columns.
Returns the integer second number from 0 to 59, or a floating
point number from 0 < 60 if \code{fractional=TRUE} that includes
any milli/micro/nanosecond component.
}
\examples{
df = pl$DataFrame(
  datetime = as.POSIXct(
    c(
      "1978-01-01 01:01:01",
      "2024-10-13 05:30:14.500",
      "2065-01-01 10:20:30.06"
    ),
    "UTC"
  )
)

df$with_columns(
  second = pl$col("datetime")$dt$second(),
  second_fractional = pl$col("datetime")$dt$second(fractional = TRUE)
)
}
