% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_seconds}
\alias{ExprDT_seconds}
\alias{(Expr)$dt$seconds}
\title{Seconds}
\format{
function
}
\value{
Expr of i64
}
\description{
Extract the seconds from a Duration type.
}
\examples{
df = pl$DataFrame(date = pl$date_range(
    low = as.POSIXct("2020-1-1", tz = "GMT"),
    high = as.POSIXct("2020-1-1 00:04:00", tz = "GMT"),
    interval = "1m"
))
df$select(
  pl$col("date"),
  pl$col("date")$diff()$dt$seconds()$alias("seconds_diff")
)
}
\keyword{ExprDT}
