% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_strftime}
\alias{ExprDT_strftime}
\alias{(Expr)$dt$strftime}
\title{strftime}
\usage{
ExprDT_strftime(format)
}
\arguments{
\item{format}{string format very much like in R passed to chrono}
}
\value{
Date/Datetime expr
}
\description{
Format Date/Datetime with a formatting rule.
See \verb{chrono strftime/strptime <https://docs.rs/chrono/latest/chrono/format/strftime/index.html>}_.
}
\examples{
pl$lit(as.POSIXct("2021-01-02 12:13:14", tz = "GMT"))$dt$strftime("this is the year: \%Y")$to_r()
}
\keyword{ExprDT}
