% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_time}
\alias{ExprDT_time}
\title{Extract time from a Datetime Series}
\usage{
<Expr>$dt$time()
}
\value{
A Time Expr
}
\description{
This only works on Datetime Series, it will error on Date Series.
}
\examples{
df = pl$DataFrame(dates = pl$date_range(
  as.Date("2000-1-1"),
  as.Date("2000-1-2"),
  "1h",
  eager = TRUE
))

df$with_columns(times = pl$col("dates")$dt$time())
}
