% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_timestamp}
\alias{ExprDT_timestamp}
\alias{(Expr)$dt$timestamp}
\title{timestamp}
\usage{
<Expr>$dt$timestamp(tu = c("ns", "us", "ms"))
}
\arguments{
\item{tu}{string option either 'ns', 'us', or 'ms'}
}
\value{
Expr of i64
}
\description{
Return a timestamp in the given time unit.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(
    start = as.Date("2001-1-1"),
    end = as.Date("2001-1-3"),
    interval = "1d1s",
    eager = TRUE
  )
)
df$select(
  pl$col("date"),
  pl$col("date")$dt$timestamp()$alias("timestamp_ns"),
  pl$col("date")$dt$timestamp(tu = "ms")$alias("timestamp_ms")
)
}
\keyword{ExprDT}
