% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_total_days}
\alias{ExprDT_total_days}
\title{Days}
\usage{
<Expr>$dt$total_days()
}
\value{
Expr of i64
}
\description{
Extract the days from a Duration type.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(
    start = as.Date("2020-3-1"),
    end = as.Date("2020-5-1"),
    interval = "1mo",
    eager = TRUE
  )
)
df$select(
  pl$col("date"),
  diff_days = pl$col("date")$diff()$dt$total_days()
)
}
