% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_total_microseconds}
\alias{ExprDT_total_microseconds}
\title{microseconds}
\usage{
ExprDT_total_microseconds()
}
\value{
Expr of i64
}
\description{
Extract the microseconds from a Duration type.
}
\examples{
df = pl$DataFrame(date = pl$date_range(
  start = as.POSIXct("2020-1-1", tz = "GMT"),
  end = as.POSIXct("2020-1-1 00:00:01", tz = "GMT"),
  interval = "1ms"
))
df$select(
  pl$col("date"),
  diff_microsec = pl$col("date")$diff()$dt$total_microseconds()
)
}
