% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_total_minutes}
\alias{ExprDT_total_minutes}
\title{Minutes}
\usage{
ExprDT_total_minutes()
}
\value{
Expr of i64
}
\description{
Extract the minutes from a Duration type.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(
    start = as.Date("2020-1-1"),
    end = as.Date("2020-1-4"),
    interval = "1d"
  )
)
df$select(
  pl$col("date"),
  diff_minutes = pl$col("date")$diff()$dt$total_minutes()
)
}
