% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_total_nanoseconds}
\alias{ExprDT_total_nanoseconds}
\title{nanoseconds}
\value{
Expr of i64
}
\description{
Extract the nanoseconds from a Duration type.
}
\examples{
df = pl$DataFrame(date = pl$date_range(
  start = as.POSIXct("2020-1-1", tz = "GMT"),
  end = as.POSIXct("2020-1-1 00:00:01", tz = "GMT"),
  interval = "1ms",
  eager = TRUE
))
df$select(
  pl$col("date"),
  diff_nanosec = pl$col("date")$diff()$dt$total_nanoseconds()
)
}
