% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_total_seconds}
\alias{ExprDT_total_seconds}
\title{Seconds}
\value{
Expr of i64
}
\description{
Extract the seconds from a Duration type.
}
\examples{
df = pl$DataFrame(date = pl$date_range(
  start = as.POSIXct("2020-1-1", tz = "GMT"),
  end = as.POSIXct("2020-1-1 00:04:00", tz = "GMT"),
  interval = "1m",
  eager = TRUE
))
df$select(
  pl$col("date"),
  diff_sec = pl$col("date")$diff()$dt$total_seconds()
)
}
