% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_with_time_zone}
\alias{ExprDT_with_time_zone}
\alias{(Expr)$dt$with_time_zone}
\title{with_time_zone}
\format{
function
}
\arguments{
\item{tz}{Null or string time zone from base::OlsonNames()}
}
\value{
Expr of i64
}
\description{
Set time zone for a Series of type Datetime.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(low = as.Date("2001-3-1"), high = as.Date("2001-5-1"), interval = "1mo")
)
df$select(
  pl$col("date"),
  pl$col("date")
    $dt$with_time_zone("Europe/London")
    $alias("London_with"),
  pl$col("date")
    $dt$tz_localize("Europe/London")
    $alias("London_localize")
)
}
\keyword{ExprDT}
