% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_any}
\alias{ExprList_any}
\title{Evaluate whether any boolean values in a list are true}
\usage{
ExprList_any()
}
\value{
Expr
}
\description{
Evaluate whether any boolean values in a list are true
}
\examples{
df = pl$DataFrame(
  list(a = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), NA, c()))
)
df$with_columns(any = pl$col("a")$list$any())
}
