% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_arg_max}
\alias{ExprList_arg_max}
\alias{list_arg_max}
\title{Get the index of the maximal value in list}
\usage{
ExprList_arg_max()
}
\value{
Expr
}
\description{
Get the index of the maximal value in list
}
\examples{
df = pl$DataFrame(list(s = list(1:2, 2:1)))
df$with_columns(
  arg_max = pl$col("s")$list$arg_max()
)
}
