% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_arg_min}
\alias{ExprList_arg_min}
\alias{list_arg_min}
\title{Get the index of the minimal value in list}
\usage{
ExprList_arg_min()
}
\value{
Expr
}
\description{
Get the index of the minimal value in list
}
\examples{
df = pl$DataFrame(list(s = list(1:2, 2:1)))
df$with_columns(
  arg_min = pl$col("s")$list$arg_min()
)
}
