% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_first}
\alias{ExprList_first}
\alias{list_first}
\title{Get the first value in a list}
\usage{
ExprList_first()
}
\value{
Expr
}
\description{
Get the first value in a list
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2)))
df$with_columns(
  first = pl$col("a")$list$first()
)
}
