% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_gather_every}
\alias{ExprList_gather_every}
\title{Gather every nth element in a list}
\usage{
ExprList_gather_every(n, offset = 0)
}
\arguments{
\item{n}{Positive integer.}

\item{offset}{Starting index.}
}
\value{
Expr
}
\description{
Gather every nth element in a list
}
\examples{
df = pl$DataFrame(
  a = list(1:5, 6:8, 9:12),
  n = c(2, 1, 3),
  offset = c(0, 1, 0)
)

df$with_columns(
  gather_every = pl$col("a")$list$gather_every(pl$col("n"), offset = pl$col("offset"))
)
}
