% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_get}
\alias{ExprList_get}
\alias{list_get}
\title{Get the value by index in a list}
\usage{
ExprList_get(index)
}
\arguments{
\item{index}{An Expr or something coercible to an Expr, that must return a
single index. Values are 0-indexed (so index 0 would return the first item
of every sublist) and negative values start from the end (index \code{-1}
returns the last item). If the index is out of bounds, it will return a
\code{null}. Strings are parsed as column names.}
}
\value{
Expr
}
\description{
This allows to extract one value per list only. To extract several values by
index, use \code{\link[=ExprList_gather]{$list$gather()}}.
}
\examples{
df = pl$DataFrame(
  values = list(c(2, 2, NA), c(1, 2, 3), NA_real_, NULL),
  idx = c(1, 2, NA, 3)
)
df$with_columns(
  using_expr = pl$col("values")$list$get("idx"),
  val_0 = pl$col("values")$list$get(0),
  val_minus_1 = pl$col("values")$list$get(-1),
  val_oob = pl$col("values")$list$get(10)
)
}
