% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_last}
\alias{ExprList_last}
\alias{list_last}
\title{Get the last value in a list}
\usage{
ExprList_last()
}
\value{
Expr
}
\description{
Get the last value in a list
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2)))
df$with_columns(
  last = pl$col("a")$list$last()
)
}
