% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_len}
\alias{ExprList_len}
\alias{list_len}
\title{Get the length of each list}
\usage{
ExprList_len()
}
\value{
Expr
}
\description{
Return the number of elements in each list. Null values are counted in the
total. \verb{$list$lengths()} is deprecated.
}
\examples{
df = pl$DataFrame(list(list_of_strs = list(c("a", "b", NA), "c")))
df$with_columns(len_list = pl$col("list_of_strs")$list$len())
}
