% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_min}
\alias{ExprList_min}
\alias{list_min}
\title{Find the minimum value in a list}
\usage{
ExprList_min()
}
\value{
Expr
}
\description{
Find the minimum value in a list
}
\examples{
df = pl$DataFrame(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_))
df$with_columns(min = pl$col("values")$list$min())
}
