% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_n_unique}
\alias{ExprList_n_unique}
\alias{list_n_unique}
\title{Get the number of unique values in a list}
\usage{
ExprList_n_unique()
}
\value{
Expr
}
\description{
Get the number of unique values in a list
}
\examples{
df = pl$DataFrame(values = list(c(2, 2, NA), c(1, 2, 3), NA_real_))
df$with_columns(unique = pl$col("values")$list$n_unique())
}
