% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_reverse}
\alias{ExprList_reverse}
\alias{list_reverse}
\title{Reverse values in a list}
\usage{
ExprList_reverse()
}
\value{
Expr
}
\description{
Reverse values in a list
}
\examples{
df = pl$DataFrame(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_))
df$with_columns(reverse = pl$col("values")$list$reverse())
}
