% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_sample}
\alias{ExprList_sample}
\title{Sample from this list}
\usage{
ExprList_sample(
  n = NULL,
  ...,
  fraction = NULL,
  with_replacement = FALSE,
  shuffle = FALSE,
  seed = NULL
)
}
\arguments{
\item{n}{Number of items to return. Cannot be used with \code{fraction}.}

\item{...}{Ignored.}

\item{fraction}{Fraction of items to return. Cannot be used with \code{n}. Can be
larger than 1 if \code{with_replacement} is \code{TRUE}.}

\item{with_replacement}{If \code{TRUE} (default), allow values to be sampled more
than once.}

\item{shuffle}{Shuffle the order of sampled data points (implicitly \code{TRUE} if
\code{with_replacement = TRUE}).}

\item{seed}{numeric value of 0 to 2^52 Seed for the random number generator.
If \code{NULL} (default), a random seed value between 0 and 10000 is picked.}
}
\value{
Expr
}
\description{
Sample from this list
}
\examples{
df = pl$DataFrame(
  values = list(1:3, NA_integer_, c(NA_integer_, 3L), 5:7),
  n = c(1, 1, 1, 2)
)

df$with_columns(
  sample = pl$col("values")$list$sample(n = pl$col("n"), seed = 1)
)
}
