% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_set_symmetric_difference}
\alias{ExprList_set_symmetric_difference}
\title{Get the symmetric difference of two list variables}
\usage{
ExprList_set_symmetric_difference(other)
}
\arguments{
\item{other}{Other list variable. Can be an Expr or something coercible to an
Expr.}
}
\value{
Expr
}
\description{
This returns all elements that are in only one of the two lists. To get only
elements that are in the first list but not in the second one, use
\code{\link[=ExprList_set_difference]{$set_difference()}}.
}
\details{
Note that the datatypes inside the list must have a common supertype. For
example, the first column can be \code{list[i32]} and the second one can be
\code{list[i8]} because it can be cast to \code{list[i32]}. However, the second column
cannot be e.g \code{list[f32]}.
}
\examples{
df = pl$DataFrame(
  a = list(1:3, NA_integer_, c(NA_integer_, 3L), 5:7),
  b = list(2:4, 3L, c(3L, 4L, NA_integer_), c(6L, 8L))
)

df$with_columns(
  symmetric_difference = pl$col("a")$list$set_symmetric_difference("b")
)
}
