% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_shift}
\alias{ExprList_shift}
\title{Shift list values by \code{n} indices}
\usage{
ExprList_shift(n = 1)
}
\arguments{
\item{n}{Number of indices to shift forward. If a negative value is
passed, values are shifted in the opposite direction instead.}
}
\value{
Expr
}
\description{
Shift list values by \code{n} indices
}
\examples{
df = pl$DataFrame(
  s = list(1:4, c(10L, 2L, 1L)),
  idx = 1:2
)
df$with_columns(
  shift_by_expr = pl$col("s")$list$shift(pl$col("idx")),
  shift_by_lit = pl$col("s")$list$shift(2)
)
}
