% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_slice}
\alias{ExprList_slice}
\alias{list_slice}
\title{Slice list}
\usage{
ExprList_slice(offset, length = NULL)
}
\arguments{
\item{offset}{Start index. Negative indexing is supported. Can be an Expr.
Strings are parsed as column names.}

\item{length}{Length of the slice. If \code{NULL} (default), the slice is taken to
the end of the list. Can be an Expr. Strings are parsed as column names.}
}
\value{
Expr
}
\description{
This extracts \code{length} values at most, starting at index \code{offset}. This can
return less than \code{length} values if \code{length} is larger than the number of
values.
}
\examples{
df = pl$DataFrame(
  s = list(1:4, c(10L, 2L, 1L)),
  idx_off = 1:2,
  len = c(4, 1)
)
df$with_columns(
  slice_by_expr = pl$col("s")$list$slice("idx_off", "len"),
  slice_by_lit = pl$col("s")$list$slice(2, 3)
)
}
