% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_sort}
\alias{ExprList_sort}
\title{Sort an Expr}
\usage{
<Expr>$list$sort(descending = FALSE)
}
\arguments{
\item{descending}{Sort values in descending order}
}
\value{
Expr
}
\description{
Sort this column. If used in a groupby context, the groups are sorted.
}
