% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_has_multiple_outputs}
\alias{ExprMeta_has_multiple_outputs}
\title{Indicate if an expression has multiple outputs}
\usage{
ExprMeta_has_multiple_outputs()
}
\value{
Boolean
}
\description{
Indicate if an expression has multiple outputs
}
\examples{
e = (pl$col("alice") + pl$col("eve"))$alias("bob")
e$meta$has_multiple_outputs()

# pl$all() select multiple cols to modify them, so it has multiple outputs
pl$all()$meta$has_multiple_outputs()
}
