% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_neq}
\alias{ExprMeta_neq}
\title{Check if two expressions are different}
\usage{
ExprMeta_neq(other)
}
\arguments{
\item{other}{Expr to compare with}
}
\value{
A boolean: \code{TRUE} if different, \code{FALSE} otherwise
}
\description{
Indicate if this expression is different from another expression. See also
the counterpart \code{\link[=ExprMeta_eq]{$meta$eq()}}.
}
\examples{
# three naive expression literals
e1 = pl$lit(40) + 2
e2 = pl$lit(42)
e3 = pl$lit(40) + 2

# e1 and e3 are identical expressions
e1$meta$neq(e3)

# when evaluated, e1 and e2 are equal
e1$neq(e2)$to_r()

# however, on the meta-level, e1 and e2 are different
e1$meta$neq(e2)
}
