% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_pop}
\alias{ExprMeta_pop}
\title{Pop}
\usage{
ExprMeta_pop()
}
\value{
A list of expressions which in most cases will have a unit length.
This is not the case when an expression has multiple inputs, for instance
in a \code{\link[=pl_fold]{$fold()}} expression.
}
\description{
Pop the latest expression and return the input(s) of the popped expression.
}
\examples{
e1 = pl$lit(40) + 2
e2 = pl$lit(42)$sum()

e1
e1$meta$pop()

e2
e2$meta$pop()
}
