% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_root_names}
\alias{ExprMeta_root_names}
\title{Get the root column names}
\usage{
ExprMeta_root_names()
}
\value{
A character vector
}
\description{
This returns the names of input columns. Use
\code{\link[=ExprMeta_output_name]{$meta$output_name()}} to get the name of output
column.
}
\examples{
e = (pl$col("alice") + pl$col("eve"))$alias("bob")
e$meta$root_names()
}
