% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__name.R
\name{ExprName_map}
\alias{ExprName_map}
\title{Map alias of expression with an R function}
\usage{
ExprName_map(fun)
}
\arguments{
\item{fun}{an R function which takes a string as input and return a string}
}
\value{
Expr
}
\description{
Rename the output of an expression by mapping a function over the root name.
}
\examples{
\dontshow{if (Sys.info()["sysname"] == "Windows") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df = pl$DataFrame(var1 = 1:3, var2 = "a")

df$select(
  pl$col("*")$name$map(\(x) paste0("new_", x))
)

# $alias() is not taken into account by $name$map()
df$select(
  pl$col("var1")$alias("foobar")$name$map(\(x) paste0("new_", x))
)
\dontshow{\}) # examplesIf}
}
