% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__name.R
\name{ExprName_prefix}
\alias{ExprName_prefix}
\alias{prefix}
\title{Add a prefix to a column name}
\arguments{
\item{prefix}{Prefix to be added to column name(s)}
}
\value{
Expr
}
\description{
Add a prefix to a column name
}
\examples{
dat = pl$DataFrame(mtcars)

dat$select(
  pl$col("mpg"),
  pl$col("mpg")$name$prefix("name_"),
  pl$col("cyl", "drat")$name$prefix("bar_")
)
}
\seealso{
\code{\link[=ExprName_suffix]{$suffix()}} to add a suffix
}
\keyword{Expr}
