% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__name.R
\name{ExprName_prefix_fields}
\alias{ExprName_prefix_fields}
\title{Add a prefix to all fields name of a struct}
\usage{
ExprName_prefix_fields(prefix)
}
\arguments{
\item{prefix}{Prefix to add to the field name.}
}
\value{
Expr
}
\description{
Add a prefix to all fields name of a struct
}
\examples{
df = pl$DataFrame(a = 1, b = 2)$select(
  pl$struct(pl$all())$alias("my_struct")
)

df$with_columns(pl$col("my_struct")$name$prefix_fields("col_"))$unnest()
}
