% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__name.R
\name{ExprName_suffix}
\alias{ExprName_suffix}
\alias{suffix}
\title{Add a suffix to a column name}
\arguments{
\item{suffix}{Suffix to be added to column name(s)}
}
\value{
Expr
}
\description{
Add a suffix to a column name
}
\examples{
dat = pl$DataFrame(mtcars)

dat$select(
  pl$col("mpg"),
  pl$col("mpg")$name$suffix("_foo"),
  pl$col("cyl", "drat")$name$suffix("_bar")
)
}
\seealso{
\code{\link[=ExprName_prefix]{$prefix()}} to add a prefix
}
\keyword{Expr}
