% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__name.R
\name{ExprName_to_lowercase}
\alias{ExprName_to_lowercase}
\title{Make the root column name lowercase}
\usage{
<Expr>$name$to_lowercase()
}
\value{
Expr
}
\description{
Due to implementation constraints, this method can only be called as the last
expression in a chain.
}
\examples{
pl$DataFrame(Alice = 1:3)$with_columns(pl$col("Alice")$name$to_lowercase())
}
