% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_contains_any}
\alias{ExprStr_contains_any}
\title{Use the aho-corasick algorithm to find matches}
\arguments{
\item{patterns}{String patterns to search. Can be an Expr.}

\item{ascii_case_insensitive}{Enable ASCII-aware case insensitive matching.
When this option is enabled, searching will be performed without respect to
case for ASCII letters (a-z and A-Z) only.}
}
\value{
Expr
}
\description{
This function determines if any of the patterns find a match.
}
\examples{
df = pl$DataFrame(
  lyrics = c(
    "Everybody wants to rule the world",
    "Tell me what you want, what you really really want",
    "Can you feel the love tonight"
  )
)

df$with_columns(
  contains_any = pl$col("lyrics")$str$contains_any(c("you", "me"))
)
}
