% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_count_match}
\alias{ExprStr_count_match}
\alias{expr_str_count_match}
\title{count_match}
\arguments{
\item{pattern}{A valid regex pattern}

\item{literal}{Boolean. Treat pattern as a literal string.}
}
\value{
UInt32 array. Contain null if original value is null or regex capture nothing.
}
\description{
Count all successive non-overlapping regex matches.
}
\examples{
df = pl$DataFrame(foo = c("123 bla 45 asd", "xyz 678 910t"))
df$select(
  pl$col("foo")$str$count_match(r"{(\d)}")$alias("count digits")
)
}
\keyword{ExprStr}
