% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_ends_with}
\alias{ExprStr_ends_with}
\alias{expr_str_ends_with}
\title{ends_with}
\arguments{
\item{sub}{Suffix substring or Expr.}
}
\value{
Expr returning a Boolean
}
\description{
Check if string values end with a substring.
}
\details{
contains : Check if string contains a substring that matches a regex.
starts_with : Check if string values start with a substring.
}
\examples{
df = pl$DataFrame(fruits = c("apple", "mango", NA))
df$select(
  pl$col("fruits"),
  pl$col("fruits")$str$ends_with("go")$alias("has_suffix")
)
}
\keyword{ExprStr}
