% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_explode}
\alias{ExprStr_explode}
\title{Returns a column with a separate row for every string character}
\value{
Expr: Series of dtype Utf8.
}
\description{
Returns a column with a separate row for every string character
}
\examples{
df = pl$DataFrame(a = c("foo", "bar"))
df$select(pl$col("a")$str$explode())
}
\keyword{ExprStr}
