% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_extract_groups}
\alias{ExprStr_extract_groups}
\title{Extract all capture groups for the given regex pattern}
\usage{
ExprStr_extract_groups(pattern)
}
\arguments{
\item{pattern}{A character or \link[=Expr_class]{Expr} of a valid regex pattern,
compatible with the \href{https://docs.rs/regex/latest/regex/}{regex crate}.}
}
\value{
Expr of data type \link[=DataType_Struct]{Struct} with fields of data type
\code{String}.
}
\description{
Extract all capture groups for the given regex pattern
}
\details{
All group names are strings. If your pattern contains unnamed groups, their
numerical position is converted to a string. See examples.
}
\examples{
df = pl$DataFrame(
  url = c(
    "http://vote.com/ballon_dor?candidate=messi&ref=python",
    "http://vote.com/ballon_dor?candidate=weghorst&ref=polars",
    "http://vote.com/ballon_dor?error=404&ref=rust"
  )
)

pattern = r"(candidate=(?<candidate>\w+)&ref=(?<ref>\w+))"

df$with_columns(
  captures = pl$col("url")$str$extract_groups(pattern)
)$unnest("captures")

# If the groups are unnamed, their numerical position (as a string) is used:

pattern = r"(candidate=(\w+)&ref=(\w+))"

df$with_columns(
  captures = pl$col("url")$str$extract_groups(pattern)
)$unnest("captures")
}
