% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_json_decode}
\alias{ExprStr_json_decode}
\title{Parse string values as JSON.}
\arguments{
\item{dtype}{The dtype to cast the extracted value to. If \code{NULL}, the dtype
will be inferred from the JSON value.}

\item{infer_schema_length}{How many rows to parse to determine the schema.
If \code{NULL}, all rows are used.}
}
\value{
Expr returning a struct
}
\description{
Parse string values as JSON.
}
\details{
Throw errors if encounter invalid json strings.
}
\examples{
df = pl$DataFrame(
  json_val = c('{"a":1, "b": true}', NA, '{"a":2, "b": false}')
)
dtype = pl$Struct(pl$Field("a", pl$Int64), pl$Field("b", pl$Boolean))
df$select(pl$col("json_val")$str$json_decode(dtype))
}
\keyword{ExprStr}
