% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_json_extract}
\alias{ExprStr_json_extract}
\alias{expr_str_json_extract}
\title{json_extract}
\arguments{
\item{dtype}{The dtype to cast the extracted value to. If None, the dtype will be
inferred from the JSON value.}
}
\value{
Expr returning a boolean
}
\description{
Parse string values as JSON.
}
\details{
Throw errors if encounter invalid json strings.
}
\examples{
df = pl$DataFrame(
  json_val =  c('{"a":1, "b": true}', NA, '{"a":2, "b": false}')
)
dtype = pl$Struct(pl$Field("a", pl$Int64), pl$Field("b", pl$Boolean))
df$select(pl$col("json_val")$str$json_extract(dtype))
}
\keyword{ExprStr}
