% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_lengths}
\alias{ExprStr_lengths}
\title{Get the number of bytes in strings}
\value{
Expr of u32
}
\description{
Get length of the strings as UInt32 (as number of bytes). Use \verb{$str$n_chars()}
to get the number of characters.
}
\details{
If you know that you are working with ASCII text, \code{lengths} will be
equivalent, and faster (returns length in terms of the number of bytes).
}
\examples{
pl$DataFrame(
  s = c("Café", NA, "345", "æøå")
)$select(
  pl$col("s"),
  pl$col("s")$str$lengths()$alias("lengths"),
  pl$col("s")$str$n_chars()$alias("n_chars")
)
}
\keyword{ExprStr}
