% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_n_chars}
\alias{ExprStr_n_chars}
\alias{expr_str_n_chars}
\title{n_chars}
\format{
function
}
\value{
Expr of u32 n_chars
}
\description{
Get length of the strings as UInt32 (as number of chars).
}
\details{
If you know that you are working with ASCII text, \code{lengths} will be
equivalent, and faster (returns length in terms of the number of bytes).
}
\examples{
pl$DataFrame(
  s = c("Café", NA, "345", "東京")
)$select(
  pl$col("s"),
  pl$col("s")$str$lengths()$alias("lengths"),
  pl$col("s")$str$n_chars()$alias("n_chars")
)
}
\keyword{ExprStr}
