% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_replace}
\alias{ExprStr_replace}
\alias{expr_str_replace}
\title{replace}
\arguments{
\item{pattern}{Into\if{html}{\out{<Expr>}}, regex pattern}

\item{value}{Into\if{html}{\out{<Expr>}} replcacement}

\item{literal}{bool, Treat pattern as a literal string.}
}
\value{
Expr of Utf8 Series
}
\description{
Replace first matching regex/literal substring with a new string value.
}
\examples{
df = pl$DataFrame(id = c(1, 2), text = c("123abc", "abc456"))
df$with_columns(
   pl$col("text")$str$replace(r"{abc\b}", "ABC")
)
}
\seealso{
replace_all : Replace all matching regex/literal substrings.
}
\keyword{ExprStr}
