% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_replace_all}
\alias{ExprStr_replace_all}
\title{Replace all matching regex/literal substrings with a new string value}
\arguments{
\item{pattern}{Regex pattern, can be an Expr.}

\item{value}{Replacement, can be an Expr.}

\item{literal}{Treat pattern as a literal string.}
}
\value{
Expr of Utf8 Series
}
\description{
Replace all matching regex/literal substrings with a new string value
}
\examples{
df = pl$DataFrame(id = c(1, 2), text = c("abcabc", "123a123"))
df$with_columns(
  pl$col("text")$str$replace_all("a", "-")
)
}
\seealso{
\verb{$str$replace()}: Replace first matching regex/literal substring.
}
\keyword{ExprStr}
