% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_replace_many}
\alias{ExprStr_replace_many}
\title{Use the aho-corasick algorithm to replace many matches}
\usage{
<Expr>$str$replace_many(patterns, replace_with, ascii_case_insensitive = FALSE)
}
\arguments{
\item{patterns}{String patterns to search. Can be an Expr.}

\item{replace_with}{A vector of strings used as replacements. If this is of
length 1, then it is applied to all matches. Otherwise, it must be of same
length as the \code{patterns} argument.}

\item{ascii_case_insensitive}{Enable ASCII-aware case insensitive matching.
When this option is enabled, searching will be performed without respect to
case for ASCII letters (a-z and A-Z) only.}
}
\value{
Expr
}
\description{
This function replaces several matches at once.
}
\examples{
df = pl$DataFrame(
  lyrics = c(
    "Everybody wants to rule the world",
    "Tell me what you want, what you really really want",
    "Can you feel the love tonight"
  )
)

# a replacement of length 1 is applied to all matches
df$with_columns(
  remove_pronouns = pl$col("lyrics")$str$replace_many(c("you", "me"), "")
)

# if there are more than one replacement, the patterns and replacements are
# matched
df$with_columns(
  fake_pronouns = pl$col("lyrics")$str$replace_many(c("you", "me"), c("foo", "bar"))
)
}
