% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_rjust}
\alias{ExprStr_rjust}
\alias{expr_str_rjust}
\title{rjust}
\arguments{
\item{width}{Justify left to this length.}

\item{fillchar}{Fill with this ASCII character.}
}
\value{
Expr of Utf8
}
\description{
Return the string left justified in a string of length \code{width}.
}
\details{
Padding is done using the specified \code{fillchar}. The original string is returned if
\code{width} is less than or equal to \code{len(s)}.
}
\examples{
df = pl$DataFrame(a = c("cow", "monkey", NA, "hippopotamus"))
df$select(pl$col("a")$str$rjust(8, "*"))
}
\keyword{ExprStr}
