% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_split_exact}
\alias{ExprStr_split_exact}
\alias{expr_str_split_exact}
\title{split_exact}
\arguments{
\item{by}{Substring to split_exact by.}

\item{inclusive}{If True, include the split_exact character/string in the results.}
}
\value{
List of Utf8 type
}
\description{
split_exact the string by a substring.
}
\examples{
df = pl$DataFrame(s = c("a_1", NA, "c", "d_4"))
df$select( pl$col("s")$str$split_exact(by="_",1))
}
\keyword{ExprStr}
