% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_starts_with}
\alias{ExprStr_starts_with}
\title{Check if string starts with a regex}
\arguments{
\item{sub}{Prefix substring or Expr.}
}
\value{
Expr returning a Boolean
}
\description{
Check if string values starts with a substring.
}
\details{
See also \verb{$str$contains()} and \verb{$str$ends_with()}.
}
\examples{
df = pl$DataFrame(fruits = c("apple", "mango", NA))
df$select(
  pl$col("fruits"),
  pl$col("fruits")$str$starts_with("app")$alias("has_suffix")
)
}
\keyword{ExprStr}
