% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_str_explode}
\alias{ExprStr_str_explode}
\alias{expr_str_explode}
\title{str_explode}
\value{
Expr: Series of dtype Utf8.
}
\description{
Returns a column with a separate row for every string character.
}
\examples{
df = pl$DataFrame(a = c("foo", "bar"))
df$select(pl$col("a")$str$str_explode())
}
\keyword{ExprStr}
