% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_strip}
\alias{ExprStr_strip}
\alias{expr_str_strip}
\title{Strip}
\arguments{
\item{matches}{The set of characters to be removed. All combinations of this set of
characters will be stripped. If set to NULL (default), all whitespace is removed instead.}
}
\value{
Expr of Utf8 lowercase chars
}
\description{
Remove leading and trailing characters.
}
\details{
will not strip anyt chars beyond the first char not matched. \code{strip()} starts from
both left and right. Whereas \code{lstrip()}and \code{rstrip()} starts from left and right respectively.
}
\examples{
df = pl$DataFrame(foo = c(" hello", "\tworld"))
df$select(pl$col("foo")$str$strip())
df$select(pl$col("foo")$str$strip(" hel rld"))
df$select(pl$col("foo")$str$lstrip(" hel rld"))
df$select(pl$col("foo")$str$rstrip(" hel\trld"))
df$select(pl$col("foo")$str$rstrip("rldhel\t "))
}
\keyword{ExprStr}
