% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_strip_chars_end}
\alias{ExprStr_strip_chars_end}
\alias{expr_str_strip_chars_end}
\title{Strip trailing characters}
\usage{
<Expr>$str$strip_chars_end(matches = NULL)
}
\arguments{
\item{matches}{The set of characters to be removed. All combinations of this
set of characters will be stripped. If \code{NULL} (default), all whitespace is
removed instead. This can be an Expr.}
}
\value{
Expr of String lowercase chars
}
\description{
Remove trailing characters.
}
\details{
This function will not strip any chars beyond the first char not matched.
\code{strip_chars_end()} removes characters at the end of the string only.
Use \code{strip_chars()} and \code{strip_chars_start()} to remove characters from the left
and right or only from the left respectively.
}
\examples{
df = pl$DataFrame(foo = c(" hello", "\tworld"))
df$select(pl$col("foo")$str$strip_chars_end(" hel\trld"))
df$select(pl$col("foo")$str$strip_chars_end("rldhel\t "))
}
