% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_strip_chars_start}
\alias{ExprStr_strip_chars_start}
\alias{expr_str_strip_chars_start}
\title{Strip leading characters}
\arguments{
\item{matches}{The set of characters to be removed. All combinations of this
set of characters will be stripped. If \code{NULL} (default), all whitespace is
removed instead. This can be an Expr.}
}
\value{
Expr of Utf8 lowercase chars
}
\description{
Remove leading characters.
}
\details{
This function will not strip any chars beyond the first char not matched.
\code{strip_chars_start()} removes characters at the beginning of the string only.
Use \code{strip_chars()} and \code{strip_chars_end()} to remove characters from the left
and right or only from the right respectively.
}
\examples{
df = pl$DataFrame(foo = c(" hello", "\tworld"))
df$select(pl$col("foo")$str$strip_chars_start(" hel rld"))
}
