% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_date}
\alias{ExprStr_to_date}
\title{Convert a String column into a Date column}
\usage{
ExprStr_to_date(format = NULL, strict = TRUE, exact = TRUE, cache = TRUE)
}
\arguments{
\item{format}{Format to use for conversion. See \code{?strptime} for possible
values. Example: "\%Y-\%m-\%d". If \code{NULL} (default), the format is
inferred from the data. Notice that time zone \verb{\%Z} is not supported and will
just ignore timezones. Numeric time zones like \verb{\%z} or \verb{\%:z}  are supported.}

\item{strict}{If \code{TRUE} (default), raise an error if a single string cannot
be parsed. If \code{FALSE}, parsing failure will produce a polars \code{null}.}

\item{exact}{If \code{TRUE} (default), require an exact format match. Otherwise,
allow the format to match anywhere in the target string.}

\item{cache}{Use a cache of unique, converted dates to apply the datetime
conversion.}
}
\value{
Expr
}
\description{
Convert a String column into a Date column
}
\examples{
pl$DataFrame(str_date = c("2009-01-02", "2009-01-03", "2009-1-4", "2009 05 01"))$
  with_columns(date = pl$col("str_date")$str$to_date(strict = FALSE))
}
