% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_integer}
\alias{ExprStr_to_integer}
\title{Convert a String column into an Int64 column with base radix}
\usage{
ExprStr_to_integer(..., base = 10L, strict = TRUE)
}
\arguments{
\item{...}{Ignored.}

\item{base}{A positive integer or expression which is the base of the string
we are parsing. Characters are parsed as column names. Default: \code{10L}.}

\item{strict}{A logical. If \code{TRUE} (default), raise any ParseError or overflow
as ComputeError. If \code{FALSE}, silently convert to \code{null}.}
}
\value{
\link[=Expr_class]{Expression} of data type \code{Int64}.
}
\description{
Convert a String column into an Int64 column with base radix
}
\examples{
df = pl$DataFrame(bin = c("110", "101", "010", "invalid"))
df$with_columns(
  parsed = pl$col("bin")$str$to_integer(base = 2, strict = FALSE)
)

df = pl$DataFrame(hex = c("fa1e", "ff00", "cafe", NA))
df$with_columns(
  parsed = pl$col("hex")$str$to_integer(base = 16, strict = TRUE)
)
}
