% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_titlecase}
\alias{ExprStr_to_titlecase}
\alias{expr_str_to_titlecase}
\title{To titlecase}
\value{
Expr of Utf8 titlecase chars
}
\description{
Transform to titlecase variant.
}
\examples{

# this method is only availble with rust compiler flag "full_features" which can
be set via envvar "RPOLARS_FULL_FEATURES" and it requires rust nightly to compile.

# wrapped tryCatch to not trigger any CI error
f = \() pl$lit(c("hello there", "HI, THERE", NA))$str$to_titlecase()$lit_to_s()
tryCatch(f(), error = function(err) {as.character(err)})
}
\keyword{ExprStr}
