% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_zfill}
\alias{ExprStr_zfill}
\alias{expr_str_zfill}
\title{zfill}
\usage{
ExprStr_zfill(alignment)
}
\arguments{
\item{width}{Justify left to this length.}

\item{fillchar}{Fill with this ASCII character.}
}
\value{
Expr of Utf8
}
\description{
Return the string left justified in a string of length \code{width}.
}
\details{
Padding is done using the specified \code{fillchar}.
The original string is returned if \code{width} is less than or equal to
\code{len(s)}.

A leading sign prefix ('+'/'-') is handled by inserting the padding after the
sign character rather than before. The original string is returned if width is
less than or equal to \code{len(s)}.
}
\examples{
some_floats_expr = pl$lit(c(0,10,-5,5))

#cast to Utf8 and ljust alignment = 5, and view as R char vector
some_floats_expr$cast(pl$Utf8)$str$zfill(5)$to_r()

#cast to int and the to utf8 and then ljust alignment = 5, and view as R char vector
some_floats_expr$cast(pl$Int64)$cast(pl$Utf8)$str$zfill(5)$to_r()
}
\keyword{ExprStr}
