% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__struct.R
\name{ExprStruct_field}
\alias{ExprStruct_field}
\title{Retrieve one of the fields of this Struct as a new Series}
\usage{
ExprStruct_field(name)
}
\arguments{
\item{name}{Name of the field.}
}
\value{
Expr of datatype Struct
}
\description{
Retrieve one of the fields of this Struct as a new Series
}
\examples{
df = pl$DataFrame(
  aaa = c(1, 2),
  bbb = c("ab", "cd"),
  ccc = c(TRUE, NA),
  ddd = list(c(1, 2), 3)
)$select(
  pl$struct(pl$all())$alias("struct_col")
)
# struct field into a new Series
df$select(
  pl$col("struct_col")$struct$field("bbb"),
  pl$col("struct_col")$struct$field("ddd")
)
}
