% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__struct.R
\name{ExprStruct_with_fields}
\alias{ExprStruct_with_fields}
\title{Add or overwrite fields of this struct}
\usage{
ExprStruct_with_fields(...)
}
\arguments{
\item{...}{Field(s) to add. Accepts expression input. Strings are parsed as
column names, other non-expression inputs are parsed as literals.}
}
\value{
An \code{\link[=RPolarsExpr]{Expr}} of data type Struct.
}
\description{
This is similar to \code{\link[=DataFrame_with_columns]{with_columns}} on
\code{\link[=RPolarsDataFrame]{DataFrame}}.
}
\examples{
df = pl$DataFrame(x = c(1, 4, 9), y = c(4, 9, 16), multiply = c(10, 2, 3))$
  with_columns(coords = pl$struct(c("x", "y")))$
  select("coords", "multiply")

df

df = df$with_columns(
  pl$col("coords")$struct$with_fields(
    pl$col("coords")$struct$field("x")$sqrt(),
    y_mul = pl$col("coords")$struct$field("y") * pl$col("multiply")
  )
)

df

df$unnest("coords")
}
