% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__having}
\alias{groupby__having}
\title{Filter groups with a list of predicates after aggregation}
\usage{
groupby__having(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Expression that evaluates to
a boolean Series.}
}
\value{
An object of class \code{polars_group_by}
}
\description{
Using this method is equivalent to adding the predicates to the aggregation
and filtering afterwards.

This method can be chained and all conditions will be combined using \code{&}.
}
\examples{
df <- pl$DataFrame(x = c("a", "b", "a", "b", "c"))

# Only keep groups that contain more than one element:
df$group_by("x")$having(
  pl$len() > 1
)$agg()
}
