% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__head}
\alias{groupby__head}
\title{Get the first \code{n} rows of each group}
\usage{
groupby__head(n = 5)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A polars \link{DataFrame}
}
\description{
Get the first \code{n} rows of each group
}
\examples{
df <- pl$DataFrame(
  letters = c("c", "c", "a", "c", "a", "b"),
  nrs = 1:6
)
df

df$group_by("letters")$head(2)$sort("letters")
}
