% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__max}
\alias{groupby__max}
\title{Reduce the groups to the maximal value}
\usage{
groupby__max()
}
\value{
A polars \link{DataFrame}
}
\description{
Reduce the groups to the maximal value
}
\examples{
df <- pl$DataFrame(
  grp = c("c", "c", "a", "c", "a", "b"),
  x = c(0.5, 0.5, 4, 10, 13, 14),
  y = 1:6,
  z = c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE)
)
df

df$group_by("grp")$max()
}
