% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__n_unique}
\alias{groupby__n_unique}
\title{Count the unique values per group}
\usage{
groupby__n_unique()
}
\value{
A polars \link{DataFrame}
}
\description{
Count the unique values per group
}
\examples{
df <- pl$DataFrame(
  grp = c("c", "c", "a", "c", "a", "b"),
  x = c(0.5, 0.5, 4, 10, 13, 14),
  y = 1:6,
  z = c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE)
)
df

df$group_by("grp")$n_unique()
}
