% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__quantile}
\alias{groupby__quantile}
\title{Compute the quantile per group}
\usage{
groupby__quantile(
  quantile,
  interpolation = c(
    "nearest",
    "higher",
    "lower",
    "midpoint",
    "linear",
    "equiprobable"
  )
)
}
\arguments{
\item{quantile}{Quantile between 0.0 and 1.0.}

\item{interpolation}{Interpolation method.}
}
\value{
A polars \link{DataFrame}
}
\description{
Compute the quantile per group
}
\examples{
df <- pl$DataFrame(
  grp = c("c", "c", "a", "c", "a", "b"),
  x = c(0.5, 0.5, 4, 10, 13, 14),
  y = 1:6,
  z = c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE)
)
df

df$group_by("grp")$quantile(0.5)
}
