% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{GroupBy_agg}
\alias{GroupBy_agg}
\alias{agg}
\title{GroupBy Aggregate}
\usage{
GroupBy_agg(...)
}
\arguments{
\item{...}{exprs to aggregate}
}
\value{
aggregated DataFrame
}
\description{
Aggregatete a DataFrame over a groupby
}
\examples{
pl$DataFrame(
  list(
    foo = c("one", "two", "two", "one", "two"),
    bar = c(5, 3, 2, 4, 1)
  )
)$groupby(
"foo"
)$agg(
 pl$col("bar")$sum()$alias("bar_sum"),
 pl$col("bar")$mean()$alias("bar_tail_sum")
)

}
