% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{GroupBy_as_data_frame}
\alias{GroupBy_as_data_frame}
\title{convert to data.frame}
\usage{
GroupBy_as_data_frame(...)
}
\arguments{
\item{...}{any opt param passed to R as.data.frame}
}
\value{
R data.frame
}
\description{
convert to data.frame
}
\examples{
pl$DataFrame(iris)$as_data_frame() #R-polars back and forth
}
