% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{GroupBy_columns}
\alias{GroupBy_columns}
\title{Get and set column names of a GroupBy}
\usage{
GroupBy_columns()
}
\value{
A character vector with the column names.
}
\description{
Get and set column names of a GroupBy
}
\examples{
gb = pl$GroupBy(iris)$group_by("Species")

# get values
gb$columns
}
\keyword{GroupBy}
