% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{GroupBy_median}
\alias{GroupBy_median}
\title{GroupBy Median}
\usage{
GroupBy_median()
}
\value{
aggregated DataFrame
}
\description{
Reduce the groups to the median value.
}
\examples{
df = pl$DataFrame(
  a = c(1, 2, 2, 3, 4, 5),
  b = c(0.5, 0.5, 4, 10, 13, 14),
  c = c(TRUE, TRUE, TRUE, FALSE, FALSE, TRUE),
  d = c("Apple", "Orange", "Apple", "Apple", "Banana", "Banana")
)
df$group_by("d", maintain_order = TRUE)$median()
}
\keyword{GroupBy}
