% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{GroupBy_min}
\alias{GroupBy_min}
\title{GroupBy Min}
\usage{
GroupBy_min()
}
\value{
aggregated DataFrame
}
\description{
Reduce the groups to the minimum value.
}
\examples{
df = pl$DataFrame(
        a = c(1, 2, 2, 3, 4, 5),
        b = c(0.5, 0.5, 4, 10, 13, 14),
        c = c(TRUE, TRUE, TRUE, FALSE, FALSE, TRUE),
        d = c("Apple", "Orange", "Apple", "Apple", "Banana", "Banana")
)
df$groupby("d", maintain_order=TRUE)$min()
}
\keyword{GroupBy}
