% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{GroupBy_shift_and_fill}
\alias{GroupBy_shift_and_fill}
\title{Shift and fill}
\usage{
GroupBy_shift_and_fill(fill_value, periods = 1)
}
\arguments{
\item{fill_value}{fill None values with the result of this expression.}

\item{periods}{integer Number of periods to shift (may be negative).}
}
\value{
GroupBy
}
\description{
Shift and fill the values by a given period.
}
\examples{
pl$DataFrame(mtcars)$group_by("cyl")$shift_and_fill(99, 1)
}
\keyword{GroupBy}
