% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{GroupBy_to_data_frame}
\alias{GroupBy_to_data_frame}
\title{convert to data.frame}
\usage{
GroupBy_to_data_frame(...)
}
\arguments{
\item{...}{not used}
}
\value{
R data.frame
}
\description{
convert to data.frame
}
\examples{
pl$DataFrame(iris)$to_data_frame() # R-polars back and forth
}
