% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{GroupBy_ungroup}
\alias{GroupBy_ungroup}
\title{GroupBy_ungroup}
\usage{
GroupBy_ungroup()
}
\value{
\link[=DataFrame_class]{DataFrame}
}
\description{
Revert the group by operation.
}
\examples{
gb = pl$DataFrame(mtcars)$group_by("cyl")
gb

gb$ungroup()
}
