% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{import_polars_as_}
\alias{import_polars_as_}
\title{Bind polars function to a namespace object pl}
\usage{
import_polars_as_(name = "pl")
}
\description{
Bind polars function to a namespace object pl
}
\details{
polars has alot of function names conflicting with base R like sum and col.
To disambiguate, minipolars can be loaded as namesapce, syntactically very similar to
python's \verb{import polars as pl}. The major syntactical difference between py-polars and minipolars
is that in R \code{$} is used as method operator instead of \code{.}.
}
\examples{
import_polars_as_("pl")
pl$df(iris)$select(pl$col("Petal.Length")$sum()$over("Species"))
}
