% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_sink_ndjson}
\alias{LazyFrame_sink_ndjson}
\title{Stream the output of a query to a JSON file}
\usage{
LazyFrame_sink_ndjson(
  path,
  maintain_order = TRUE,
  type_coercion = TRUE,
  predicate_pushdown = TRUE,
  projection_pushdown = TRUE,
  simplify_expression = TRUE,
  slice_pushdown = TRUE,
  no_optimization = FALSE,
  inherit_optimization = FALSE
)
}
\arguments{
\item{path}{File path to which the result should be written.}

\item{maintain_order}{Keep the same order as the original \code{LazyFrame}. Setting
this to \code{TRUE} makes it more expensive to compute and blocks the possibility
to run on the streaming engine. The default value can be changed with
\code{pl$set_options(maintain_order = TRUE)}.}

\item{type_coercion}{Boolean. Coerce types such that operations succeed and
run on minimal required memory.}

\item{predicate_pushdown}{Boolean. Applies filters as early as possible at
scan level.}

\item{projection_pushdown}{Boolean. Select only the columns that are needed
at the scan level.}

\item{simplify_expression}{Boolean. Various optimizations, such as constant
folding and replacing expensive operations with faster alternatives.}

\item{slice_pushdown}{Boolean. Only load the required slice from the scan
level. Don't materialize sliced outputs (e.g. \code{join$head(10)}).}

\item{no_optimization}{Boolean. Sets the following parameters to \code{FALSE}:
\code{predicate_pushdown}, \code{projection_pushdown}, \code{slice_pushdown},
\code{comm_subplan_elim}, \code{comm_subexpr_elim}.}

\item{inherit_optimization}{Boolean. Use existing optimization settings
regardless the settings specified in this function call.}
}
\description{
This writes the output of a query directly to a JSON file without collecting
it in the R session first. This is useful if the output of the query is still
larger than RAM as it would crash the R session if it was collected into R.
}
\examples{
# sink table 'mtcars' from mem to JSON
tmpf = tempfile(fileext = ".json")
pl$LazyFrame(mtcars)$sink_ndjson(tmpf)

# load parquet directly into a DataFrame / memory
pl$scan_ndjson(tmpf)$collect()
}
