% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_write_json}
\alias{DataFrame_write_json}
\title{Write to JSON file}
\usage{
DataFrame_write_json(file, pretty = FALSE, row_oriented = FALSE)
}
\arguments{
\item{file}{File path to which the result should be written.}

\item{pretty}{Pretty serialize JSON.}

\item{row_oriented}{Write to row-oriented JSON. This is slower, but more
common.}
}
\value{
This doesn't return anything.
}
\description{
Write to JSON file
}
\examples{
if (require("jsonlite", quiet = TRUE)) {
  dat = pl$DataFrame(head(mtcars))
  destination = tempfile()

  dat$select(pl$col("drat", "mpg"))$write_json(destination)
  jsonlite::fromJSON(destination)

  dat$select(pl$col("drat", "mpg"))$write_json(destination, row_oriented = TRUE)
  jsonlite::fromJSON(destination)
}
}
