% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_write_ndjson}
\alias{DataFrame_write_ndjson}
\title{Write to NDJSON file}
\usage{
DataFrame_write_ndjson(file)
}
\arguments{
\item{file}{File path to which the result should be written.}
}
\value{
Invisibly returns the input DataFrame.
}
\description{
Write to NDJSON file
}
\examples{
dat = as_polars_df(head(mtcars))

destination = tempfile()
dat$select(pl$col("drat", "mpg"))$write_ndjson(destination)

pl$read_ndjson(destination)
}
