% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{is_err}
\alias{is_err}
\title{check if x ss a result and an err}
\usage{
is_err(x)
}
\arguments{
\item{x}{R object which could be a rust-like result of a list with two elements, ok and err}
}
\value{
bool if is a result object which is an err
}
\description{
check if x ss a result and an err
}
\keyword{internal}
