% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{is_ok}
\alias{is_ok}
\title{check if x ss a result and an ok}
\usage{
is_ok(x)
}
\arguments{
\item{x}{R object which could be a rust-like result of a list with two elements, ok and err}
}
\value{
bool if is a result object which is an ok
}
\description{
check if x ss a result and an ok
}
