% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_polars.R
\name{is_polars_df}
\alias{is_polars_df}
\title{Test if the object is a polars DataFrame}
\usage{
is_polars_df(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
This function tests if the object is a polars DataFrame.
}
\examples{
is_polars_df(mtcars)

is_polars_df(as_polars_df(mtcars))
}
