% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{is_polars_dtype}
\alias{is_polars_dtype}
\title{Check if input is a valid DataType}
\usage{
is_polars_dtype(x, include_unknown = FALSE)
}
\arguments{
\item{x}{An object to be tested.}

\item{include_unknown}{If \code{FALSE} (default), \code{pl$Unknown} is considered as
an invalid datatype.}
}
\value{
A boolean scalar.
}
\description{
Check if input is a valid DataType
}
\examples{
is_polars_dtype(pl$Int64)
is_polars_dtype(mtcars)
is_polars_dtype(pl$Unknown)
is_polars_dtype(pl$Unknown, include_unknown = TRUE)
}
