% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_polars.R
\name{is_polars_lf}
\alias{is_polars_lf}
\title{Test if the object is a polars LazyFrame}
\usage{
is_polars_lf(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
These functions test if the object is a polars LazyFrame.
}
\examples{
is_polars_lf(mtcars)

is_polars_lf(as_polars_lf(mtcars))
}
