% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_polars.R
\name{is_polars_series}
\alias{is_polars_series}
\title{Test if the object is a polars Series}
\usage{
is_polars_series(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
These functions test if the object is a polars Series.
}
\examples{
is_polars_series(1:3)

is_polars_series(as_polars_series(1:3))
}
