% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{is_result}
\alias{is_result}
\title{check if z is a result}
\usage{
is_result(x)
}
\arguments{
\item{x}{R object which could be a rust-like result of a list with two elements, ok and err}
}
\value{
bool if is a result object
}
\description{
check if z is a result
}
\details{
both ok and err being NULL encodes ok-value NULL. No way to encode an err-value NULL
If both ok and err has value then this is an invalid result
}
