% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{is_schema}
\alias{is_schema}
\title{check if schema}
\format{
function
}
\usage{
is_schema(x)
}
\arguments{
\item{x}{objet to test if schema}
}
\value{
bool
}
\description{
check if schema
}
\examples{
polars:::is_schema(pl$DataFrame(iris)$schema)
pl$is_schema(pl$DataFrame(iris)$schema)
polars:::is_schema(list("alice","bob"))

}
