% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-knitr.R
\name{knit_print.DataFrame}
\alias{knit_print.DataFrame}
\title{knit print polars DataFrame}
\usage{
knit_print.DataFrame(x, ...)
}
\arguments{
\item{x}{a polars DataFrame to knit_print}

\item{...}{additional arguments, not used}
}
\description{
knit print polars DataFrame
}
\keyword{internals}
