% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_to_vdf}
\alias{l_to_vdf}
\title{DataFrame-list to rust vector of DataFrame}
\usage{
l_to_vdf(l)
}
\arguments{
\item{l}{list of DataFrame}
}
\value{
RPolarsVecDataFrame
}
\description{
lifecycle: DEPRECATE, imple on rust side as a function
}
\keyword{internal}
