% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{lazy_csv_reader}
\alias{lazy_csv_reader}
\title{read csv lazuliy}
\usage{
lazy_csv_reader(
  path = "my.csv",
  sep = ",",
  has_header = TRUE,
  ignore_errors = FALSE,
  skip_rows = 0,
  n_rows = NULL,
  cache = FALSE,
  overwrite_dtype = NULL,
  low_memory = FALSE,
  comment_char = NULL,
  quote_char = "\\"",
  null_values = NULL,
  infer_schema_length = 100,
  rechunk = TRUE,
  skip_rows_after_header = 0,
  encoding = "utf8-lossy",
  row_count_name = "myrowcounter",
  row_count_offset = 42,
  parse_dates = FALSE
)
}
\arguments{
\item{parse_dates}{}
}
\description{
read csv lazuliy
}
