% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{Lazy_describe_optimized_plan}
\alias{Lazy_describe_optimized_plan}
\title{Lazy_select}
\usage{
Lazy_describe_optimized_plan()
}
\arguments{
\item{...}{any single Expr or string naming a column}
}
\value{
A new \code{lazy_polar_frame} object with applied filter.
}
\description{
select on a lazy DataFrame
}
