% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{Lazy_groupby}
\alias{Lazy_groupby}
\title{Lazy_groupby}
\usage{
Lazy_groupby(..., maintain_order = FALSE)
}
\arguments{
\item{...}{any single Expr or string naming a column}
}
\value{
A new \code{lazy_polar_frame} object with applied filter.
}
\description{
groupby on lazy_polar_frame.
}
