% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{lazy_polar_frame}
\alias{lazy_polar_frame}
\title{R6 Class polar_lazy_frame}
\description{
R6 Class polar_lazy_frame

R6 Class polar_lazy_frame
}
\details{
high-level wrapper class for Rpolarframe
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-lazy_polar_frame-new}{\code{lazy_polar_frame$new()}}
\item \href{#method-lazy_polar_frame-collect}{\code{lazy_polar_frame$collect()}}
\item \href{#method-lazy_polar_frame-select}{\code{lazy_polar_frame$select()}}
\item \href{#method-lazy_polar_frame-filter}{\code{lazy_polar_frame$filter()}}
\item \href{#method-lazy_polar_frame-groupby}{\code{lazy_polar_frame$groupby()}}
\item \href{#method-lazy_polar_frame-describe_optimized_plan}{\code{lazy_polar_frame$describe_optimized_plan()}}
\item \href{#method-lazy_polar_frame-print}{\code{lazy_polar_frame$print()}}
\item \href{#method-lazy_polar_frame-clone}{\code{lazy_polar_frame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-new"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-new}{}}}
\subsection{Method \code{new()}}{
wrap low-level LazyFrame in new R6 lazy_polar_frame object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{LazyFrame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-collect"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-collect}{}}}
\subsection{Method \code{collect()}}{
collect lazy_polar_frame. This triggers computation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$collect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{polar_frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-select"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-select}{}}}
\subsection{Method \code{select()}}{
select on lazy_polar_frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$select(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any  Exprs or strings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object with selection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-filter"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-filter}{}}}
\subsection{Method \code{filter()}}{
filter on lazy_polar_frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$filter(rexpr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rexpr}}{any single Expr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object with applied filter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-groupby"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-groupby}{}}}
\subsection{Method \code{groupby()}}{
groupby on lazy_polar_frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$groupby(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any single Expr or string naming a column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object with applied filter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-describe_optimized_plan"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-describe_optimized_plan}{}}}
\subsection{Method \code{describe_optimized_plan()}}{
print the optimized plan of a polar_lazy_frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$describe_optimized_plan()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-print"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-print}{}}}
\subsection{Method \code{print()}}{
print non-optimized plan of a polar_lazy_frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_frame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_frame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_polar_frame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
