% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{Lazy_with_columns}
\alias{Lazy_with_columns}
\title{Lazy_with_columns}
\usage{
Lazy_with_columns(...)
}
\arguments{
\item{...}{any single Expr or string naming a column}
}
\value{
A new \code{lazy_polar_frame} object with applied filter.
}
\description{
add or replace columns of lazy DataFrame
}
