% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{LazyFrame-class}
\alias{LazyFrame-class}
\alias{LazyFrame}
\title{Describe the collect plan of a Polars LazyFrame}
\format{
An object of class \code{environment} of length 24.
}
\usage{
LazyFrame
}
\description{
Describe the collect plan of a Polars LazyFrame
}
\keyword{datasets}
