% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{LazyFrame-describe_optimized_plan}
\alias{LazyFrame-describe_optimized_plan}
\title{Describe the optimized collect plan of a Polars LazyFrame}
\usage{
<LazyFrame>$describe_optimized_plan()
}
\description{
Describe the optimized collect plan of a Polars LazyFrame
}
