% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{LazyFrame-describe_plan}
\alias{LazyFrame-describe_plan}
\title{Describe the collect plan of a Polars LazyFrame}
\usage{
<LazyFrame>$describe_plan()
}
\description{
Describe the collect plan of a Polars LazyFrame
}
