% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__collect}
\alias{lazyframe__collect}
\title{Materialize this LazyFrame into a DataFrame}
\usage{
lazyframe__collect(
  ...,
  engine = c("auto", "in-memory", "streaming"),
  optimizations = pl$QueryOptFlags()
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{engine}{The engine name to use for processing the query.
One of the followings:
\itemize{
\item \code{"auto"} (default): Select the engine automatically.
In Polars 1.16, the \code{"in-memory"} engine is selected for most cases.
Starting with Polars 2.0, the streaming engine is selected instead.
Use \code{"in-memory"} to keep the current execution engine explicitly.
\item \code{"in-memory"}: Use the in-memory engine. This is an escape hatch for
queries whose incidental row order differs under streaming.
\item \code{"streaming"}: Use the streaming engine.
}}

\item{optimizations}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A \link{QueryOptFlags} object to indicate optimization passes done during query optimization.}
}
\value{
A polars \link{DataFrame}
}
\description{
By default, all query optimizations are enabled.
Individual optimizations may be disabled by setting the corresponding parameter to \code{FALSE}.
}
\examples{
lf <- pl$LazyFrame(
  a = c("a", "b", "a", "b", "b", "c"),
  b = 1:6,
  c = 6:1,
)
lf$group_by("a")$agg(pl$all()$sum())$collect()

# Collect in streaming mode
lf$group_by("a")$agg(pl$all()$sum())$collect(
  engine = "streaming"
)
}
\seealso{
\itemize{
\item \code{\link[=lazyframe__sink_parquet]{$sink_parquet()}} streams query to a parquet file.
\item \code{\link[=lazyframe__sink_ipc]{$sink_ipc()}} streams query to a arrow file.
}
}
