% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__collect_schema}
\alias{lazyframe__collect_schema}
\title{Resolve the schema of this LazyFrame}
\usage{
lazyframe__collect_schema()
}
\value{
A named list with names indicating column names and values indicating
column data types.
}
\description{
This resolves the query plan but does not trigger computations.
}
\examples{
lf <- pl$LazyFrame(
  foo = 1:3,
  bar = 6:8,
  ham = c("a", "b", "c")
)

lf$collect_schema()

lf$with_columns(
  baz = (pl$col("foo") + pl$col("bar"))$cast(pl$String),
  pl$col("bar")$cast(pl$Int64)
)$collect_schema()
}
