% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__count}
\alias{lazyframe__count}
\title{Return the number of non-null elements for each column}
\usage{
lazyframe__count()
}
\value{
A polars \link{LazyFrame}
}
\description{
Return the number of non-null elements for each column
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, NA), c = rep(NA, 4))
lf$count()$collect()
}
