% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__drop}
\alias{lazyframe__drop}
\title{Remove columns}
\usage{
lazyframe__drop(..., strict = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names or \link[=polars_selector]{selectors}
that should be removed.}

\item{strict}{Validate that all column names exist in the current schema,
and throw an exception if any do not.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Remove columns
}
\examples{
# Drop columns by passing the name of those columns
lf <- pl$LazyFrame(
  foo = 1:3,
  bar = c(6, 7, 8),
  ham = c("a", "b", "c")
)
lf$drop("ham")$collect()
lf$drop("ham", "bar")$collect()

# Drop multiple columns by passing a selector
lf$drop(cs$all())$collect()
}
