% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__drop_nulls}
\alias{lazyframe__drop_nulls}
\title{Drop all rows that contain null values}
\usage{
lazyframe__drop_nulls(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names or \link[=polars_selector]{selectors}
for which are considered. If empty (default), use all columns
(same as specifying with the selector \code{\link[=cs__all]{cs$all()}}).}
}
\value{
A polars \link{LazyFrame}
}
\description{
The original order of the remaining rows is preserved.
}
\examples{
lf <- pl$LazyFrame(
  foo = 1:3,
  bar = c(6L, NA, 8L),
  ham = c("a", "b", NA)
)

# The default behavior of this method is to drop rows where any single value
# of the row is null.
lf$drop_nulls()$collect()

# This behaviour can be constrained to consider only a subset of columns, as
# defined by name or with a selector. For example, dropping rows if there is
# a null in any of the integer columns:
lf$drop_nulls(cs$integer())$collect()
}
