% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__last}
\alias{lazyframe__last}
\title{Get the last row of the LazyFrame}
\usage{
lazyframe__last()
}
\value{
A polars \link{LazyFrame}
}
\description{
Get the last row of the LazyFrame
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$last()$collect()
}
