% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__max}
\alias{lazyframe__max}
\title{Aggregate the columns in the LazyFrame to their maximum value}
\usage{
lazyframe__max()
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns in the LazyFrame to their maximum value
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$max()$collect()
}
