% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__mean}
\alias{lazyframe__mean}
\title{Aggregate the columns in the LazyFrame to their mean value}
\usage{
lazyframe__mean()
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns in the LazyFrame to their mean value
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$mean()$collect()
}
