% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__profile}
\alias{lazyframe__profile}
\title{Collect and profile a lazy query}
\usage{
lazyframe__profile(
  ...,
  show_plot = FALSE,
  truncate_nodes = 0,
  engine = c("auto", "in-memory", "streaming"),
  optimizations = pl$QueryOptFlags(),
  type_coercion = deprecated(),
  predicate_pushdown = deprecated(),
  projection_pushdown = deprecated(),
  simplify_expression = deprecated(),
  slice_pushdown = deprecated(),
  comm_subplan_elim = deprecated(),
  comm_subexpr_elim = deprecated(),
  cluster_with_columns = deprecated(),
  collapse_joins = deprecated(),
  no_optimization = deprecated()
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{show_plot}{Show a Gantt chart of the profiling result}

\item{truncate_nodes}{Truncate the label lengths in the Gantt chart to this
number of characters. If \code{0} (default), do not truncate.}

\item{engine}{The engine name to use for processing the query.
One of the followings:
\itemize{
\item \code{"auto"} (default): Select the engine automatically.
In Polars 1.16, the \code{"in-memory"} engine is selected for most cases.
Starting with Polars 2.0, the streaming engine is selected instead.
Use \code{"in-memory"} to keep the current execution engine explicitly.
\item \code{"in-memory"}: Use the in-memory engine. This is an escape hatch for
queries whose incidental row order differs under streaming.
\item \code{"streaming"}: Use the streaming engine.
}}

\item{optimizations}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A \link{QueryOptFlags} object to indicate optimization passes done during query optimization.}

\item{type_coercion}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{type_coercion} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{predicate_pushdown}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{predicate_pushdown} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{projection_pushdown}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{projection_pushdown} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{simplify_expression}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{simplify_expression} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{slice_pushdown}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{slice_pushdown} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{comm_subplan_elim}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{comm_subplan_elim} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{comm_subexpr_elim}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{comm_subexpr_elim} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{cluster_with_columns}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{cluster_with_columns} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{collapse_joins}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{predicate_pushdown} property of a \link{QueryOptFlags} object, then pass
that to the \code{optimizations} argument instead.}

\item{no_optimization}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use the \code{optimizations} argument with
\code{\link[=QueryOptFlags]{pl$QueryOptFlags()$no_optimizations()}} instead.}
}
\value{
List of two \code{DataFrame}s: one with the collected result, the other
with the timings of each step. If \code{show_plot = TRUE}, then the plot is
also stored in the list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\verb{$profile()} is deprecated. Starting with Polars 2.0, \code{engine = "auto"} will
use the streaming engine by default. Due to the concurrent nature of the
streaming engine, the profiling information from this method would be
misleading.

This will run the query and return a list containing the
materialized DataFrame and a DataFrame that contains profiling information
of each node that is executed.
}
\details{
The units of the timings are microseconds.
}
\examples{
lf <- pl$LazyFrame(
  a = c("a", "b", "a", "b", "b", "c"),
  b = 1:6,
  c = 6:1,
)

lf$group_by("a", .maintain_order = TRUE)$agg(
  pl$all()$sum()
)$sort("a")$profile()
}
\seealso{
\itemize{
\item \code{\link[=lazyframe__collect]{$collect()}} - regular collect.
\item \code{\link[=lazyframe__sink_parquet]{$sink_parquet()}} streams query to a parquet file.
\item \code{\link[=lazyframe__sink_ipc]{$sink_ipc()}} streams query to a arrow file.
}
}
