% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__quantile}
\alias{lazyframe__quantile}
\title{Aggregate the columns to a unique quantile value}
\usage{
lazyframe__quantile(
  quantile,
  interpolation = c(
    "nearest",
    "higher",
    "lower",
    "midpoint",
    "linear",
    "equiprobable"
  )
)
}
\arguments{
\item{quantile}{Quantile between 0.0 and 1.0.}

\item{interpolation}{Interpolation method.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns to a unique quantile value
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$quantile(0.7)$collect()
}
