% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__rename}
\alias{lazyframe__rename}
\title{Rename column names}
\usage{
lazyframe__rename(..., .strict = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Either a function that takes
a character vector as input and returns a character vector as output, or
named values where names are old column names and values are the new ones.}

\item{.strict}{Validate that all column names exist in the current schema,
and throw an error if any do not. (Note that this parameter is a no-op when
passing a function to \code{...}).}
}
\value{
A polars \link{LazyFrame}
}
\description{
Rename column names
}
\details{
If existing names are swapped (e.g. 'A' points to 'B' and 'B' points to
'A'), polars will block projection and predicate pushdowns at this node.
}
\examples{
lf <- pl$LazyFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)

lf$rename(foo = "apple")$collect()
}
