% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__reverse}
\alias{lazyframe__reverse}
\title{Reverse the LazyFrame}
\usage{
lazyframe__reverse()
}
\value{
A polars \link{LazyFrame}
}
\description{
Reverse the LazyFrame
}
\examples{
lf <- pl$LazyFrame(key = c("a", "b", "c"), val = 1:3)
lf$reverse()$collect()
}
