% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__serialize}
\alias{lazyframe__serialize}
\alias{pl__deserialize_lf}
\title{Serialize the logical plan of this LazyFrame}
\usage{
lazyframe__serialize()

pl__deserialize_lf(data)
}
\arguments{
\item{data}{A raw vector of serialized \link{LazyFrame}.}
}
\value{
\itemize{
\item \verb{<lazyframe>$serialize()} returns a raw vector.
\item \code{pl$deserialize_lf()} returns a deserialized \link{LazyFrame}.
}
}
\description{
Serialize the logical plan to a binary representation (raw vector).
}
\examples{
lf <- pl$LazyFrame(a = 1:3)$sum()

# Serialize the logical plan to a binary representation.
serialized <- lf$serialize()
serialized

# The bytes can later be deserialized back into a LazyFrame.
pl$deserialize_lf(serialized)$collect()
}
