% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__slice}
\alias{lazyframe__slice}
\title{Get a slice of the LazyFrame.}
\usage{
lazyframe__slice(offset, length = NULL)
}
\arguments{
\item{offset}{Start index. Negative indexing is supported.}

\item{length}{Length of the slice. If \code{NULL} (default), all rows starting at
the offset will be selected.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Get a slice of the LazyFrame.
}
\examples{
lf <- pl$LazyFrame(x = c("a", "b", "c"), y = 1:3, z = 4:6)
lf$slice(1, 2)$collect()
}
