% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__sum}
\alias{lazyframe__sum}
\title{Aggregate the columns of this LazyFrame to their sum values}
\usage{
lazyframe__sum()
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns of this LazyFrame to their sum values
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$sum()$collect()
}
