% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__tail}
\alias{lazyframe__tail}
\title{Get the last \code{n} rows.}
\usage{
lazyframe__tail(n = 5L)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Get the last \code{n} rows.
}
\examples{
lf <- pl$LazyFrame(a = 1:6, b = 7:12)
lf$tail()$collect()
lf$tail(2)$collect()
}
\seealso{
\code{\link[=lazyframe__head]{<LazyFrame>$head()}}
}
