% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__unnest}
\alias{lazyframe__unnest}
\title{Decompose struct columns into separate columns for each of their fields}
\usage{
lazyframe__unnest(..., separator = NULL)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name of the struct column(s)
or \link[=polars_selector]{selectors}  that should be unnested.}

\item{separator}{\code{NULL} (default) or single string.
Rename output column names as combination of the struct column name,
name separator and field name.}
}
\value{
A polars \link{LazyFrame}
}
\description{
The new columns will be inserted at the location of the struct column.
}
\examples{
lf <- pl$LazyFrame(
  a = 1:5,
  b = c("one", "two", "three", "four", "five"),
  c = 6:10
)$select(
  pl$struct("b"),
  pl$struct(c("a", "c"))$alias("a_and_c")
)
lf$collect()

lf$unnest("a_and_c")$collect()
lf$unnest("a_and_c", separator = ":")$collect()
}
