% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__with_row_index}
\alias{lazyframe__with_row_index}
\title{Add a row index as the first column in the LazyFrame}
\usage{
lazyframe__with_row_index(name = "index", offset = 0)
}
\arguments{
\item{name}{Name of the index column.}

\item{offset}{Start the index at this offset. Cannot be negative.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Using this function can have a negative effect on query performance. This
may, for instance, block predicate pushdown optimization.
}
\examples{
lf <- pl$LazyFrame(x = c(1, 3, 5), y = c(2, 4, 6))
lf$with_row_index()$collect()

lf$with_row_index("id", offset = 1000)$collect()

# An index column can also be created using the expressions int_range()
# and len()$
lf$with_columns(
  index = pl$int_range(pl$len(), dtype = pl$UInt32)
)$collect()
}
