% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_cast}
\alias{LazyFrame_cast}
\title{Cast LazyFrame column(s) to the specified dtype}
\usage{
LazyFrame_cast(dtypes, ..., strict = TRUE)
}
\arguments{
\item{dtypes}{Either a datatype or a list where the names are column names and
the values are the datatypes to convert to.}

\item{...}{Ignored.}

\item{strict}{Throw an error if a cast could not be done (for instance, due to
an overflow).}
}
\value{
A LazyFrame
}
\description{
This allows to convert all columns to a datatype or to convert only specific
columns. Contrarily to the Python implementation, it is not possible to
convert all columns of a specific datatype to another datatype.
}
\examples{
lf = pl$LazyFrame(
  foo = 1:3,
  bar = c(6, 7, 8),
  ham = as.Date(c("2020-01-02", "2020-03-04", "2020-05-06"))
)

# Cast only some columns
lf$cast(list(foo = pl$Float32, bar = pl$UInt8))$collect()

# Cast all columns to the same type
lf$cast(pl$String)$collect()
}
