% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_clone}
\alias{LazyFrame_clone}
\title{Clone a LazyFrame}
\usage{
LazyFrame_clone()
}
\value{
A LazyFrame
}
\description{
This makes a very cheap deep copy/clone of an existing \code{LazyFrame}.
}
\examples{
df1 = pl$LazyFrame(iris)
df2 = df1$clone()
df3 = df1

# the clone and the original don't have the same address...
pl$mem_address(df1) != pl$mem_address(df2)

# ... but simply assigning df1 to df3 change the address anyway
pl$mem_address(df1) == pl$mem_address(df3)
}
