% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyFrame_collect}
\alias{LazyFrame_collect}
\title{LazyFrame collect}
\usage{
LazyFrame_collect()
}
\value{
collected \code{DataFrame}
}
\description{
collect DataFrame by lazy query
}
\examples{
pl$DataFrame(iris)$lazy()$filter(pl$col("Species")=="setosa")$collect()
}
\keyword{LazyFrame}
