% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_collect_in_background}
\alias{LazyFrame_collect_in_background}
\title{Collect_in_background}
\usage{
LazyFrame_collect_in_background()
}
\value{
a thread handle for the task
}
\description{
collect DataFrame by lazy query in a background thread
}
\examples{
prexpr <- pl$col("mpg")$map(\(x) x * 0.43, in_background = TRUE)$alias("kml")
pl$LazyFrame(mtcars)$with_column(prexpr)$collect_in_background()
}
\keyword{DataFrame_new}
\keyword{LazyFrame}
