% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_columns}
\alias{LazyFrame_columns}
\title{Columns}
\usage{
LazyFrame_columns()
}
\value{
A vector of column names
}
\description{
Get the column names of the LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$columns

}
\keyword{LazyFrame}
