% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_describe_optimized_plan}
\alias{LazyFrame_describe_optimized_plan}
\title{Print the optmized plan of LazyFrame}
\usage{
LazyFrame_describe_optimized_plan()
}
\description{
select on a LazyFrame
}
\keyword{LazyFrame}
