% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_drop}
\alias{LazyFrame_drop}
\title{Drop columns of a LazyFrame}
\usage{
LazyFrame_drop(..., strict = TRUE)
}
\arguments{
\item{...}{Characters of column names to drop. Passed to \code{\link[=pl_col]{pl$col()}}.}

\item{strict}{Validate that all column names exist in the schema and throw an
exception if a column name does not exist in the schema.}
}
\value{
LazyFrame
}
\description{
Drop columns of a LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$drop(c("mpg", "hp"))$collect()

# equivalent
pl$LazyFrame(mtcars)$drop("mpg", "hp")$collect()
}
