% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_dtypes}
\alias{LazyFrame_dtypes}
\title{Dtypes}
\usage{
LazyFrame_dtypes()
}
\value{
A vector of column data types
}
\description{
Get the data types of the LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$dtypes

}
\keyword{LazyFrame}
