% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_fill_null}
\alias{LazyFrame_fill_null}
\title{Fill nulls}
\usage{
LazyFrame_fill_null(fill_value)
}
\arguments{
\item{fill_value}{Value to fill nulls with.}
}
\value{
LazyFrame
}
\description{
Fill null values (which correspond to \code{NA} in R) using the
specified value or strategy.
}
\examples{
df = pl$LazyFrame(
  a = c(1.5, 2, NA, 4),
  b = c(1.5, NA, NA, 4)
)
df$fill_null(99)$collect()
}
\keyword{LazyFrame}
