% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_filter}
\alias{LazyFrame_filter}
\title{Filter rows based on conditions}
\format{
An object of class \code{character} of length 1.
}
\usage{
LazyFrame_filter(expr)
}
\arguments{
\item{expr}{one Expr or string naming a column}
}
\value{
A new \code{LazyFrame} object with add/modified column.
}
\description{
Filter rows based on conditions
}
\examples{
pl$LazyFrame(iris)$filter(pl$col("Species") == "setosa")$collect()
}
\keyword{LazyFrame}
