% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_first}
\alias{LazyFrame_first}
\title{Get the first row}
\usage{
LazyFrame_first
}
\value{
\code{LazyFrame} with a single row
}
\description{
Get the first row
}
\examples{
pl$LazyFrame(mtcars)$first()$collect()
}
\keyword{DataFrame}
