% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_get_optimization_toggle}
\alias{LazyFrame_get_optimization_toggle}
\title{Get optimization settings}
\usage{
LazyFrame_get_optimization_toggle()
}
\value{
List of optimization toggles
}
\description{
Get the current optimization toggles for the lazy query
}
\examples{
pl$LazyFrame(mtcars)$get_optimization_toggle()
}
\keyword{LazyFrame}
