% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_group_by}
\alias{LazyFrame_group_by}
\title{Group a LazyFrame}
\usage{
LazyFrame_group_by(..., maintain_order = polars_options()$maintain_order)
}
\arguments{
\item{...}{Any Expr(s) or string(s) naming a column.}

\item{maintain_order}{Keep the same order as the original \code{LazyFrame}. Setting
this to \code{TRUE} makes it more expensive to compute and blocks the possibility
to run on the streaming engine. The default value can be changed with
\code{options(polars.maintain_order = TRUE)}.}
}
\value{
LazyGroupBy (a LazyFrame with special groupby methods like \verb{$agg()})
}
\description{
This doesn't modify the data but only stores information about
the group structure. This structure can then be used by several functions
(\verb{$agg()}, \verb{$filter()}, etc.).
}
\examples{
pl$LazyFrame(
  foo = c("one", "two", "two", "one", "two"),
  bar = c(5, 3, 2, 4, 1)
)$
  group_by("foo")$
  agg(
  pl$col("bar")$sum()$name$suffix("_sum"),
  pl$col("bar")$mean()$alias("bar_tail_sum")
)$
  collect()
}
\keyword{LazyFrame}
