% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_last}
\alias{LazyFrame_last}
\title{Get the first row}
\usage{
LazyFrame_last
}
\value{
\code{LazyFrame} with a single row
}
\description{
Get the first row
}
\examples{
pl$LazyFrame(mtcars)$last()$collect()
}
\keyword{LazyFrame}
