% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_limit}
\alias{LazyFrame_limit}
\title{Limit a LazyFrame}
\usage{
LazyFrame_limit(n)
}
\arguments{
\item{n}{Positive number not larger than 2^32.}
}
\value{
A \code{LazyFrame}
}
\description{
Take some maximum number of rows.
}
\details{
Any number will converted to u32. Negative raises error.
}
\examples{
pl$LazyFrame(mtcars)$limit(4)$collect()
}
