% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\docType{data}
\name{LazyFrame_max}
\alias{LazyFrame_max}
\title{Max}
\format{
function
}
\usage{
LazyFrame_max
}
\value{
A new \code{LazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their maximum value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$max()$collect()
}
\keyword{LazyFrame}
