% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_mean}
\alias{LazyFrame_mean}
\title{Mean}
\usage{
LazyFrame_mean
}
\value{
A LazyFrame with one row
}
\description{
Aggregate the columns in the LazyFrame to their mean value.
}
\examples{
pl$LazyFrame(mtcars)$mean()$collect()
}
\keyword{LazyFrame}
