% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_median}
\alias{LazyFrame_median}
\title{Median}
\usage{
LazyFrame_median
}
\value{
A LazyFrame with one row
}
\description{
Aggregate the columns in the LazyFrame to their median value.
}
\examples{
pl$LazyFrame(mtcars)$median()$collect()
}
\keyword{LazyFrame}
