% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_min}
\alias{LazyFrame_min}
\title{Min}
\usage{
LazyFrame_min
}
\value{
A new \code{LazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their minimum value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$min()$collect()
}
\keyword{LazyFrame}
