% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\docType{data}
\name{LazyFrame_print}
\alias{LazyFrame_print}
\title{print LazyFrame internal method}
\format{
An object of class \code{character} of length 1.
}
\usage{
LazyFrame_print
}
\arguments{
\item{x}{LazyFrame}
}
\value{
self
}
\description{
can be used i the middle of a method chain
}
\examples{
 pl$DataFrame(iris)$lazy()$print()
}
\keyword{LazyFrame}
