% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_quantile}
\alias{LazyFrame_quantile}
\title{Quantile}
\usage{
LazyFrame_quantile(quantile, interpolation = "nearest")
}
\arguments{
\item{quantile}{numeric Quantile between 0.0 and 1.0.}

\item{interpolation}{string Interpolation method: "nearest", "higher", "lower", "midpoint", or "linear".}
}
\value{
LazyFrame
}
\description{
Aggregate the columns in the DataFrame to their quantile value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$quantile(.4)$collect()
}
\keyword{LazyFrame}
