% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_rename}
\alias{LazyFrame_rename}
\title{Rename column names of a LazyFrame}
\usage{
LazyFrame_rename(...)
}
\arguments{
\item{...}{One of the following:
\itemize{
\item params like \code{new_name = "old_name"} to rename selected variables.
\item as above but with params wrapped in a list
}}
}
\value{
\link[=LazyFrame_class]{LazyFrame}
}
\description{
Rename column names of a LazyFrame
}
\details{
If existing names are swapped (e.g. \code{A} points to \code{B} and \code{B} points to \code{A}),
polars will block projection and predicate pushdowns at this node.
}
\examples{
lf = pl$LazyFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)

lf$rename(apple = "foo")$collect()
}
\seealso{
\itemize{
\item \code{\link[=LazyFrame_rename_with]{<LazyFrame>$rename_with()}}
}
}
