% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_rename_with}
\alias{LazyFrame_rename_with}
\title{Rename column names of a LazyFrame with a function}
\usage{
LazyFrame_rename_with(fun)
}
\arguments{
\item{fun}{An R function that takes the old names character vector as input and
returns the new names character vector.}
}
\value{
\link[=LazyFrame_class]{LazyFrame}
}
\description{
This method is currently experimental and may
change without it being considered a breaking change.
}
\details{
If existing names are swapped (e.g. \code{A} points to \code{B} and \code{B} points to \code{A}),
polars will block projection and predicate pushdowns at this node.
}
\examples{
lf = pl$LazyFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)

lf$rename_with(
  \(column_name) paste0("c", substr(column_name, 2, 100))
)$collect()
}
\seealso{
\itemize{
\item \code{\link[=LazyFrame_rename]{<LazyFrame>$rename()}}
}
}
