% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_schema}
\alias{LazyFrame_schema}
\title{Schema}
\usage{
LazyFrame_schema()
}
\value{
A list mapping from field name to field type
}
\description{
Get the schema of the LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$schema

}
\keyword{LazyFrame}
