% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyFrame_select}
\alias{LazyFrame_select}
\title{Lazy_select}
\usage{
LazyFrame_select(...)
}
\arguments{
\item{...}{any single Expr or string naming a column}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
select on a LazyFrame
}
\keyword{LazyFrame}
