% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_select_seq}
\alias{LazyFrame_select_seq}
\title{Select and modify columns of a LazyFrame}
\usage{
LazyFrame_select_seq(...)
}
\arguments{
\item{...}{Columns to keep. Those can be expressions (e.g \code{pl$col("a")}),
column names  (e.g \code{"a"}), or list containing expressions or column names
(e.g \code{list(pl$col("a"))}).}
}
\value{
A LazyFrame
}
\description{
Similar to \code{dplyr::mutate()}. However, it discards unmentioned columns (like
\code{.()} in \code{data.table}).

This will run all expression sequentially instead of in parallel. Use this
when the work per expression is cheap. Otherwise, \verb{$select()} should be
preferred.
}
\examples{
pl$LazyFrame(iris)$select_seq(
  pl$col("Sepal.Length")$abs()$alias("abs_SL"),
  (pl$col("Sepal.Length") + 2)$alias("add_2_SL")
)
}
