% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_serialize}
\alias{LazyFrame_serialize}
\title{Serialize the logical plan of this LazyFrame to a file or string in JSON format}
\usage{
LazyFrame_serialize()
}
\value{
A character of the JSON representation of the logical plan
}
\description{
Note that not all LazyFrames can be serialized. For example, LazyFrames that
contain UDFs such as \code{\link[=Expr_map_elements]{$map_elements()}} cannot be serialized.
}
\examples{
lf = pl$LazyFrame(a = 1:3)$sum()
json = lf$serialize()
json

# The logical plan can later be deserialized back into a LazyFrame.
pl$deserialize_lf(json)$collect()
}
\seealso{
\itemize{
\item \code{\link[=pl_deserialize_lf]{pl$deserialize_lf()}}
}
}
