% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_show_graph}
\alias{LazyFrame_show_graph}
\title{Plot the query plan}
\usage{
LazyFrame_show_graph(
  optimized = TRUE,
  raw_output = FALSE,
  type_coercion = TRUE,
  predicate_pushdown = TRUE,
  projection_pushdown = TRUE,
  simplify_expression = TRUE,
  slice_pushdown = TRUE,
  comm_subplan_elim = TRUE,
  comm_subexpr_elim = TRUE,
  streaming = FALSE
)
}
\arguments{
\item{optimized}{Optimize the query plan.}

\item{raw_output}{Return the raw GraphViz code only, does not print the plot.}

\item{type_coercion}{Boolean. Coerce types such that operations succeed and
run on minimal required memory.}

\item{predicate_pushdown}{Boolean. Applies filters as early as possible at
scan level.}

\item{projection_pushdown}{Boolean. Select only the columns that are needed
at the scan level.}

\item{simplify_expression}{Boolean. Various optimizations, such as constant
folding and replacing expensive operations with faster alternatives.}

\item{slice_pushdown}{Boolean. Only load the required slice from the scan
level. Don't materialize sliced outputs (e.g. \code{join$head(10)}).}

\item{comm_subplan_elim}{Boolean. Will try to cache branching subplans that
occur on self-joins or unions.}

\item{comm_subexpr_elim}{Boolean. Common subexpressions will be cached and
reused.}

\item{streaming}{Boolean. Run parts of the query in a streaming fashion
(this is in an alpha state).}
}
\value{
If \code{raw_output = TRUE}, a character vector. Otherwise, an object of
class \code{"grViz"}.
}
\description{
Note that this requires to have the package \code{DiagrammeR} installed. Use
\code{raw_output = TRUE} to obtain the raw GraphViz code without plotting it.
}
\examples{
\dontshow{if (requireNamespace("DiagrammeR", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
query = pl$LazyFrame(mtcars)$
  filter(pl$col("drat") > 3)$
  with_columns(foo = pl$col("mpg") + pl$col("cyl"), bar = pl$mean("mpg"))

query$show_graph(raw_output = TRUE)

query$show_graph()
\dontshow{\}) # examplesIf}
}
