% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_slice}
\alias{LazyFrame_slice}
\title{Slice}
\usage{
<LazyFrame>$slice(offset, length = NULL)
}
\arguments{
\item{offset}{Start index, can be a negative value. This is 0-indexed, so
\code{offset = 1} doesn't include the first row.}

\item{length}{Length of the slice. If \code{NULL} (default), all rows starting at
the offset will be selected.}
}
\description{
Get a slice of the LazyFrame.
}
\examples{
pl$LazyFrame(mtcars)$slice(2, 4)$collect()
pl$LazyFrame(mtcars)$slice(30)$collect()
mtcars[2:6, ]
}
