% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_std}
\alias{LazyFrame_std}
\title{Std}
\usage{
<LazyFrame>$std(ddof = 1)
}
\arguments{
\item{ddof}{Delta Degrees of Freedom: the divisor used in the calculation is
N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
A LazyFrame with one row
}
\description{
Aggregate the columns of this LazyFrame to their standard
deviation values.
}
\examples{
pl$LazyFrame(mtcars)$std()$collect()
}
\keyword{LazyFrame}
