% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_tail}
\alias{LazyFrame_tail}
\title{Tail of a DataFrame}
\usage{
LazyFrame_tail(n)
}
\arguments{
\item{n}{Positive number not larger than 2^32.}
}
\value{
A LazyFrame
}
\description{
Get the last \code{n} rows.
}
\details{
Any number will converted to u32. Negative raises error.
}
\examples{
pl$LazyFrame(mtcars)$tail(2)$collect()
}
