% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_width}
\alias{LazyFrame_width}
\title{Width}
\usage{
LazyFrame_width()
}
\value{
Integer
}
\description{
Get the width of the LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$width

}
\keyword{LazyFrame}
