% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\docType{data}
\name{LazyFrame_with_column}
\alias{LazyFrame_with_column}
\title{Lazy with column}
\format{
An object of class \code{character} of length 1.
}
\usage{
LazyFrame_with_column
}
\arguments{
\item{expr}{one Expr or string naming a column}
}
\value{
A new \code{LazyFrame} object with add/modified column.
}
\description{
add or replace columns of LazyFrame
}
\keyword{LazyFrame}
