% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_with_common_subplan_elimination}
\alias{LazyFrame_with_common_subplan_elimination}
\title{With_common_subplan_elimination}
\usage{
LazyFrame_with_common_subplan_elimination(toggle = TRUE)
}
\arguments{
\item{toggle}{whether the optimization is turned on}
}
\value{
A new LazyFrame with specified optimization scheme
}
\description{
With_common_subplan_elimination
}
\examples{
pl$LazyFrame(mtcars)$
  with_common_subplan_elimination(FALSE)

}
\keyword{LazyFrame}
