% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_with_context}
\alias{LazyFrame_with_context}
\title{Add an external context to the computation graph}
\usage{
<LazyFrame>$with_context(other)
}
\arguments{
\item{other}{Data/LazyFrame to have access to. This can be a list of DataFrames
and LazyFrames.}
}
\value{
A LazyFrame
}
\description{
This allows expressions to also access columns from DataFrames or LazyFrames
that are not part of this one.
}
\examples{
lf = pl$LazyFrame(a = c(1, 2, 3), b = c("a", "c", NA))
lf_other = pl$LazyFrame(c = c("foo", "ham"))

lf$with_context(lf_other)$select(
  pl$col("b") + pl$col("c")$first()
)$collect()

# Fill nulls with the median from another lazyframe:
train_lf = pl$LazyFrame(
  feature_0 = c(-1.0, 0, 1), feature_1 = c(-1.0, 0, 1)
)
test_lf = pl$LazyFrame(
  feature_0 = c(-1.0, NA, 1), feature_1 = c(-1.0, 0, 1)
)

test_lf$with_context(train_lf$select(pl$all()$name$suffix("_train")))$select(
  pl$col("feature_0")$fill_null(pl$col("feature_0_train")$median())
)$collect()
}
