% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_with_projection_pushdown}
\alias{LazyFrame_with_projection_pushdown}
\title{With_projection_pushdown}
\usage{
LazyFrame_with_projection_pushdown(toggle = TRUE)
}
\arguments{
\item{toggle}{whether the optimization is turned on}
}
\value{
A new LazyFrame with specified optimization scheme
}
\description{
With_projection_pushdown
}
\examples{
pl$LazyFrame(mtcars)$
  with_projection_pushdown(FALSE)

}
\keyword{LazyFrame}
