% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_with_row_count}
\alias{LazyFrame_with_row_count}
\title{Lazy with_row_count}
\usage{
LazyFrame_with_row_count(name, offset = NULL)
}
\arguments{
\item{name}{string name of the created column}

\item{offset}{positive integer offset for the start of the counter}
}
\value{
A new \code{LazyFrame} object with a counter column in front
}
\description{
Add a new column at index 0 that counts the rows
}
\keyword{LazyFrame}
