% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_with_simplify_expr}
\alias{LazyFrame_with_simplify_expr}
\title{With_simplify_expr}
\usage{
LazyFrame_with_simplify_expr(toggle = TRUE)
}
\arguments{
\item{toggle}{whether the optimization is turned on}
}
\value{
A new LazyFrame with specified optimization scheme
}
\description{
With_simplify_expr
}
\examples{
pl$LazyFrame(mtcars)$
  with_simplify_expr(FALSE)

}
\keyword{LazyFrame}
