% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_with_streaming}
\alias{LazyFrame_with_streaming}
\title{With_streaming}
\usage{
LazyFrame_with_streaming(toggle = TRUE)
}
\arguments{
\item{toggle}{whether the optimization is turned on}
}
\value{
A new LazyFrame with specified optimization scheme
}
\description{
With_streaming
}
\examples{
pl$LazyFrame(mtcars)$
  with_streaming(FALSE)

}
\keyword{LazyFrame}
