% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_with_type_coercion}
\alias{LazyFrame_with_type_coercion}
\title{With_type_coercion}
\usage{
LazyFrame_with_type_coercion(toggle = TRUE)
}
\arguments{
\item{toggle}{whether the optimization is turned on}
}
\value{
A new LazyFrame with specified optimization scheme
}
\description{
With_type_coercion
}
\examples{
pl$LazyFrame(mtcars)$
  with_type_coercion(FALSE)

}
\keyword{LazyFrame}
