% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\docType{data}
\name{LazyFrame_without_optimization}
\alias{LazyFrame_without_optimization}
\title{Without_optimization}
\format{
An object of class \code{character} of length 1.
}
\usage{
LazyFrame_without_optimization
}
\value{
A new LazyFrame with optimizations disabled
}
\description{
Without_optimization
}
\examples{
pl$LazyFrame(mtcars)$
  without_optimization()

}
\keyword{LazyFrame}
