% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-group_by.R
\name{lazygroupby__head}
\alias{lazygroupby__head}
\title{Get the first \code{n} rows of each group}
\usage{
lazygroupby__head(n = 5)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Get the first \code{n} rows of each group
}
\examples{
lf <- pl$LazyFrame(
  letters = c("c", "c", "a", "c", "a", "b"),
  nrs = 1:6
)
lf$collect()

lf$group_by("letters")$head(2)$sort("letters")$collect()
}
