% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-group_by.R
\name{lazygroupby__n_unique}
\alias{lazygroupby__n_unique}
\title{Count the unique values per group}
\usage{
lazygroupby__n_unique()
}
\value{
A polars \link{LazyFrame}
}
\description{
Count the unique values per group
}
\examples{
lf <- pl$LazyFrame(
  grp = c("c", "c", "a", "c", "a", "b"),
  x = c(0.5, 0.5, 4, 10, 13, 14),
  y = 1:6,
  z = c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE)
)
lf$collect()

lf$group_by("grp")$n_unique()$collect()
}
