% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-group_by.R
\name{lazygroupby__quantile}
\alias{lazygroupby__quantile}
\title{Compute the quantile per group}
\usage{
lazygroupby__quantile(
  quantile,
  interpolation = c(
    "nearest",
    "higher",
    "lower",
    "midpoint",
    "linear",
    "equiprobable"
  )
)
}
\arguments{
\item{quantile}{Quantile between 0.0 and 1.0.}

\item{interpolation}{Interpolation method.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Compute the quantile per group
}
\examples{
lf <- pl$LazyFrame(
  grp = c("c", "c", "a", "c", "a", "b"),
  x = c(0.5, 0.5, 4, 10, 13, 14),
  y = 1:6,
  z = c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE)
)
lf$collect()

lf$group_by("grp")$quantile(0.5)$collect()
}
