% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__groupby.R
\name{LazyGroupBy_agg}
\alias{LazyGroupBy_agg}
\title{LazyGroupBy_agg}
\usage{
LazyGroupBy_agg(...)
}
\arguments{
\item{...}{any Expr or string}
}
\value{
A new \code{LazyFrame} object.
}
\description{
aggregate a polar_lazy_groupby
}
