% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__group_by.R
\name{LazyGroupBy_class}
\alias{LazyGroupBy_class}
\alias{RPolarsLazyGroupBy}
\title{Operations on Polars grouped LazyFrame}
\description{
This class comes from \code{\link[=LazyFrame_group_by]{<LazyFrame>$group_by()}}, etc.
}
\section{Active bindings}{

\subsection{columns}{

\verb{$columns} returns a character vector with the column names.
}
}

\examples{
as_polars_lf(mtcars)$group_by("cyl")$agg(
  pl$col("mpg")$sum()
)
}
