% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__group_by.R
\name{LazyGroupBy_columns}
\alias{LazyGroupBy_columns}
\title{Get and set column names of a LazyGroupBy}
\usage{
LazyGroupBy_columns()
}
\value{
A character vector with the column names.
}
\description{
Get and set column names of a LazyGroupBy
}
\examples{
lgb = pl$LazyGroupBy(iris)$lazy()$group_by("Species")

# get values
lgb$columns
}
\keyword{LazyGroupBy}
