% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_groupby.R
\name{LazyGroupBy_head}
\alias{LazyGroupBy_head}
\title{LazyGroupBy_head}
\usage{
LazyGroupBy_head(n = 1L)
}
\arguments{
\item{n}{integer number of rows to get}
}
\value{
A new \code{LazyFrame} object.
}
\description{
get n rows of head of group
}
