% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyGroupBy_tail}
\alias{LazyGroupBy_tail}
\title{LazyGroupBy_tail}
\usage{
LazyGroupBy_tail(n = 1L)
}
\arguments{
\item{n}{integer number of rows to get}
}
\value{
A new \code{lazy_polar_frame} object.
}
\description{
get n tail rows of group
}
