% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__group_by.R
\name{LazyGroupBy_ungroup}
\alias{LazyGroupBy_ungroup}
\title{LazyGroupBy_ungroup}
\usage{
Lazy<GroupBy>$ungroup()
}
\value{
A new \code{LazyFrame} object.
}
\description{
Revert the group by operation.
}
\examples{
lf = pl$LazyFrame(mtcars)
lf

lgb = lf$group_by("cyl")
lgb

lgb$ungroup()
}
