% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{list_arg_max}
\alias{list_arg_max}
\alias{Expr_list_arg_max}
\alias{Expr_arr.arg_max}
\title{Arg max sublists}
\format{
function
}
\value{
Expr
}
\description{
Retrieve the index of the maximum value in every sublist.
}
\examples{
df = pl$DataFrame(list(s = list(1:2, 2:1)))
df$select(pl$col("s")$list$arg_max())
}
\keyword{ExprArr}
