% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{[.ExprListNameSpace}
\alias{[.ExprListNameSpace}
\title{Get list}
\usage{
\method{[}{ExprListNameSpace}(x, index)
}
\arguments{
\item{x}{ExprListNameSpace}

\item{index}{value to get}
}
\description{
Get list
}
\details{
\verb{[.ExprListNameSpace} used as e.g. \code{pl$col("a")$arr[0]} same as \code{pl$col("a")$get(0)}
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) # NULL or integer() or list()
df$select(pl$col("a")$list[0])
df$select(pl$col("a")$list[c(2, 0, -1)])
}
