% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_lengths}
\alias{ExprList_lengths}
\alias{lengths}
\alias{list_lengths}
\title{Lengths arrays in list}
\format{
function
}
\value{
Expr
}
\description{
Get the length of the arrays as UInt32
}
\examples{
df = pl$DataFrame(list_of_strs = pl$Series(list(c("a", "b"), "c")))
df$with_columns(pl$col("list_of_strs")$list$lengths()$alias("list_of_strs_lengths"))
}
\keyword{ExprList}
