% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{list_tail}
\alias{list_tail}
\title{Tails of sublists}
\format{
function
}
\arguments{
\item{n}{Numeric or Expr, number of values to return for each sublist.}
}
\value{
Expr
}
\description{
tail the first \code{n} values of every sublist.
}
\examples{
df = pl$DataFrame(list(a = list(1:4, c(10L, 2L, 1L))))
df$select(pl$col("a")$list$tail(2))
}
\keyword{ExprList}
