% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{lit}
\alias{lit}
\title{polars literal}
\usage{
lit(x)
}
\arguments{
\item{x}{any R expression yielding an integer or float}
}
\value{
Rexpr, literal of that value
}
\description{
polars literal
}
\examples{
pl::lit(42L)
}
