% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{macro_add_syntax_check_to_class}
\alias{macro_add_syntax_check_to_class}
\title{add syntax verification to class}
\usage{
macro_add_syntax_check_to_class(Class_name)
}
\arguments{
\item{Class_name}{string name of env class}
}
\value{
dollarsign method with syntax verification
}
\description{
add syntax verification to class
}
\details{
this function overrides dollarclass method of a extendr env_class
to run first verify_method_call() to check for syntax error and return
more user friendly error if issues

All R functions coined 'macro'-functions use eval(parse()) but only at package build time
to solve some tricky self-referential problem. If possible to deprecate a macro in a clean way
, go ahead.

see zzz.R for usage examples
}
\seealso{
verify_method_call
}
\keyword{internal}
