% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{macro_add_syntax_check_to_class}
\alias{macro_add_syntax_check_to_class}
\title{add syntax verification to class}
\usage{
macro_add_syntax_check_to_class(Class_name)
}
\arguments{
\item{Class_name}{string name of env class}
}
\value{
dollarsign method with syntax verification
}
\description{
add syntax verification to class
}
\details{
this function overrides dollarclass method of a extendr env_class
to run first verify_method_call() to check for syntax error and return
more user friendly error if issues
}
\examples{
macro_add_verify_to_class("DataFrame")
}
\seealso{
verify_method_call
}
