% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{make_expr_arr_namespace}
\alias{make_expr_arr_namespace}
\title{make_expr_arr_namespace}
\usage{
make_expr_arr_namespace(self)
}
\value{
environment with methods to call on self
}
\description{
Internal function to yield methods within arr namespace
See the individual method pages for full details
}
\keyword{internals}
