% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{map_err}
\alias{map_err}
\title{map an Err part of Result}
\usage{
map_err(x, f)
}
\arguments{
\item{x}{any R object}

\item{f}{a closure that takes the err part as input}
}
\value{
same R object wrapped in a Err-result
}
\description{
map an Err part of Result
}
