% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{method_as_property}
\alias{method_as_property}
\title{Give a class method property behavior}
\usage{
method_as_property(f)
}
\arguments{
\item{f}{a function}
}
\value{
function subclassed into c("property","function")
}
\description{
Internal function, see use in source
}
