% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_env_elements}
\alias{move_env_elements}
\title{Move environment element from one env to another}
\usage{
move_env_elements(from_env, to_env, element_names, remove = TRUE)
}
\arguments{
\item{from_env}{env from}

\item{to_env}{env to}

\item{element_names}{names of elements to move, if named names, then name of name is to_env name}

\item{remove}{bool, actually remove element in from_env}
}
\description{
Move environment element from one env to another
}
\details{
envs are mutable
}
