% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{ncol.DataFrame}
\alias{ncol.DataFrame}
\title{The Number of Columns of a DataFrame}
\usage{
ncol.DataFrame(x)
}
\arguments{
\item{x}{DataFrame}
}
\value{
Integer
}
\description{
The Number of Columns of a DataFrame
}
