% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{ncol.LazyFrame}
\alias{ncol.LazyFrame}
\title{The Number of Columns of a LazyFrame}
\usage{
ncol.LazyFrame(x)
}
\arguments{
\item{x}{LazyFrame}
}
\value{
Integer
}
\description{
The Number of Columns of a LazyFrame
}
