% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neopolars-package.R
\docType{package}
\name{neopolars-package}
\alias{neopolars}
\alias{neopolars-package}
\title{neopolars: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
