% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{new_df}
\alias{new_df}
\title{create new dataframe}
\usage{
new_df(data)
}
\arguments{
\item{data}{a data.frame or list of mixed vectors and Rseries of equal length.}
}
\value{
Rdataframe
}
\description{
create new dataframe
}
\examples{
minipolars:::new_df(iris)
#with namespace
pl::df(iris)
pl::df(list(some_column_name = c(1,2,3,4,5)))
}
