% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{new_pf}
\alias{new_pf}
\title{create new dataframe}
\usage{
new_pf(data)
}
\arguments{
\item{data}{a data.frame or list of mixed vectors and Rseries of equal length.}
}
\value{
Rdataframe
}
\description{
create new dataframe
}
\examples{
minipolars:::new_pf(iris)
#with namespace
pl::pf(iris)
pl::pf(list(some_column_name = c(1,2,3,4,5)))
}
