% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xptr-package.R
\name{new_xptr}
\alias{new_xptr}
\title{Create an external pointer object.}
\usage{
new_xptr(address = "", tag = NULL, protected = NULL)
}
\arguments{
\item{address}{a string of pointer address}

\item{tag}{an optional tag}

\item{protected}{an objected to be protected from gc within the lifetime of the external pointer}
}
\value{
an \code{externalptr} object
}
\description{
Create an external pointer object.
}
\details{
This function is copied from the xptr-package authored by Randy Lai.
The package xptr has been a great tool for prototyping and some function(s) we could not
do without.
}
