% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\docType{data}
\name{None}
\alias{None}
\title{python None is NA}
\format{
An object of class \code{logical} of length 1.
}
\usage{
None
}
\value{
an R NA value
}
\description{
python None is NA
}
\details{
py-polars use python None to encode polars Null, where r-polars uses NA.
When copy pasting py-polars code, forgetting to change None to NA
should not be a show-stopper.

None should not be used casually.
}
\examples{
None
}
\keyword{datasets}
