% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{nrow.DataFrame}
\alias{nrow.DataFrame}
\title{The Number of Rows of a DataFrame}
\usage{
nrow.DataFrame(x)
}
\arguments{
\item{x}{DataFrame}
}
\value{
Integer
}
\description{
The Number of Rows of a DataFrame
}
