% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{nrow.LazyFrame}
\alias{nrow.LazyFrame}
\title{The Number of Rows of a LazyFrame}
\usage{
nrow.LazyFrame(x)
}
\arguments{
\item{x}{LazyFrame}
}
\value{
Integer
}
\description{
The Number of Rows of a LazyFrame
}
