% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\docType{data}
\name{object}
\alias{object}
\title{Any rpolars class object is made of this}
\format{
An object of class \code{character} of length 1.
}
\usage{
object
}
\description{
One SEXP of Rtype: "externalptr" + a class attribute
}
\details{
\itemize{
\item \code{object$method()} calls are facilitated by a \verb{$.ClassName}- s3method see 'R/after-wrappers.R'
\item Code completion is facilitted by \code{.DollarNames.ClassName}-s3method see e.g. 'R/dataframe__frame.R'
\item Implementation of property-methods as DataFrame_columns() and syntax checking is an extension to \verb{$.ClassName}
See function macro_add_syntax_check_to_class().
}
}
\examples{
#all a rpolars object is made of:
some_rpolars_object = pl$DataFrame(iris)
str(some_rpolars_object) #External Pointer tagged with a class attribute.
}
\keyword{api_object}
