% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{pl$from_arrow}
\alias{pl$from_arrow}
\alias{from_arrow}
\title{pl$from_arrow}
\arguments{
\item{data}{arrow Table or Array or ChunkedArray}

\item{rechunk}{bool rewrite in one array per column, Implemented for ChunkedArray
Array is already contiguous. Not implemented for Table. C}

\item{schema}{not used yet}

\item{schema_overrides}{not used yet}
}
\value{
DataFrame or Series
}
\description{
import Arrow Table or Array
}
\examples{
#none so for
}
\keyword{conversion}
