% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__any}
\alias{pl__any}
\title{Evaluate a bitwise OR operation}
\usage{
pl__any(names, ..., ignore_nulls = TRUE)
}
\arguments{
\item{names}{The name(s) or \link[=DataType]{data type} of the column(s) to use in
the aggregation. A character vector or a Polars data type/list of data
types can be supplied.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ignore_nulls}{If \code{TRUE} (default), ignore null values. If \code{FALSE},
\href{https://en.wikipedia.org/wiki/Three-valued_logic}{Kleene logic} is used to
deal with nulls: if the column contains any null values and no \code{TRUE} values,
the output is null.}
}
\value{
A polars \link{expression}
}
\description{
This function is syntactic sugar for \code{col(names)$any()}.
}
\examples{
df <- pl$DataFrame(
  a = c(TRUE, FALSE, TRUE),
  b = c(FALSE, FALSE, FALSE)
)

df$select(pl$any("a"))
}
