% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-lazy.R
\name{pl__arg_where}
\alias{pl__arg_where}
\title{Return indices where \code{condition} evaluates to \code{TRUE}}
\usage{
pl__arg_where(condition)
}
\arguments{
\item{condition}{Boolean expression to evaluate.}
}
\value{
A polars \link{expression}
}
\description{
Return indices where \code{condition} evaluates to \code{TRUE}
}
\examples{
df <- pl$DataFrame(a = 1:5)
df$select(
  pl$arg_where(pl$col("a") \%\% 2 == 0)
)
}
