% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-lazy.R
\name{pl__coalesce}
\alias{pl__coalesce}
\title{Folds the columns from left to right, keeping the first non-null value}
\usage{
pl__coalesce(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Non-named objects can be referenced as columns.
Each object will be converted to \link{expression} by \code{\link[=as_polars_expr]{as_polars_expr()}}.
Strings are parsed as column names, other non-expression inputs are parsed as literals.}
}
\value{
A polars \link{expression}
}
\description{
Folds the columns from left to right, keeping the first non-null value
}
\examples{
df <- pl$DataFrame(
  a = c(1, NA, NA, NA),
  b = c(1, 2, NA, NA),
  c = c(5, NA, 3, NA)
)

df$with_columns(d = pl$coalesce("a", "b", "c", 10))

df$with_columns(d = pl$coalesce(pl$col(c("a", "b", "c")), 10))
}
