% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-as_datatype.R
\name{pl__concat_str}
\alias{pl__concat_str}
\title{Horizontally concatenate columns into a single string column}
\usage{
pl__concat_str(..., separator = "", ignore_nulls = FALSE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Columns to concatenate into a
single string column. Accepts expression input. Strings are parsed as column
names, other non-expression inputs are parsed as literals. Non-\code{String}
columns are cast to \code{String}.}

\item{separator}{String that will be used to separate the values of each
column.}

\item{ignore_nulls}{If \code{FALSE} (default), null values will be propagated,
i.e. if the row contains any null values, the output is null.}
}
\value{
A polars \link{expression}
}
\description{
Operates in linear time.
}
\examples{
df <- pl$DataFrame(
  a = 1:3,
  b = c("dogs", "cats", NA),
  c = c("play", "swim", "walk")
)
df$with_columns(
  full_sentence = pl$concat_str(
    pl$col("a") * 2L,
    pl$col("b"),
    pl$col("c"),
    separator = " ",
  )
)
}
