% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-as_datatype.R
\name{pl__date}
\alias{pl__date}
\title{Create a Polars literal expression of type Date}
\usage{
pl__date(year, month, day)
}
\arguments{
\item{year}{An \link[=Expr]{polars expression} or something can be coerced to
an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}},
which represents a column or literal number of year.}

\item{month}{An \link[=Expr]{polars expression} or something can be coerced to
an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}},
which represents a column or literal number of month. Range: 1-12.}

\item{day}{An \link[=Expr]{polars expression} or something can be coerced to
an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}},
which represents a column or literal number of day. Range: 1-31.}
}
\value{
A polars \link{expression}
}
\description{
Create a Polars literal expression of type Date
}
\examples{
df <- pl$DataFrame(month = 1:3, day = 4:6)
df$with_columns(pl$date(2024, pl$col("month"), pl$col("day")))

# We can also use `pl$date()` for filtering:
df <- pl$DataFrame(
  start = rep(as.Date("2024-01-01"), 3),
  end = as.Date(c("2024-05-01", "2024-07-01", "2024-09-01"))
)
df$filter(pl$col("end") > pl$date(2024, 6, 1))
}
