% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-range-int_range.R
\name{pl__int_ranges}
\alias{pl__int_ranges}
\title{Generate a range of integers for each row of the input columns}
\usage{
pl__int_ranges(start = 0, end = NULL, step = 1, ..., dtype = pl$Int64)
}
\arguments{
\item{start}{Start of the range (inclusive). Defaults to 0.}

\item{end}{End of the range (exclusive). If \code{NULL} (default), the value
of \code{start} is used and \code{start} is set to 0.}

\item{step}{Step size of the range.}

\item{...}{These dots are for future extensions and must be empty.}

\item{dtype}{Data type of the range.}
}
\value{
A polars \link{expression}
}
\description{
Generate a range of integers for each row of the input columns
}
\examples{
df <- pl$DataFrame(start = c(1, -1), end = c(3, 2))
df$with_columns(int_range = pl$int_ranges("start", "end"))

# end can be omitted for a shorter syntax$
df$select("end", int_range = pl$int_ranges("end"))
}
