% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{pl__LazyFrame}
\alias{pl__LazyFrame}
\alias{polars_lazy_frame}
\alias{LazyFrame}
\title{Polars LazyFrame class (\code{polars_lazy_frame})}
\usage{
pl__LazyFrame(..., .schema_overrides = NULL, .strict = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Name-value pairs of objects to be converted to polars \link{Series}
by the \code{\link[=as_polars_series]{as_polars_series()}} function.
Each \link{Series} will be used as a column of the \link{DataFrame}.
All values must be the same length or length 1.
Each name will be used as the column name. If the name is empty,
the original name of the \link{Series} will be used.}

\item{.schema_overrides}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A list of polars data types or \code{NULL} (default).
Passed to the \code{\link[=dataframe__cast]{$cast()}} method as dynamic-dots.}

\item{.strict}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A logical value. Passed to the \code{\link[=dataframe__cast]{$cast()}} method's \code{.strict} argument.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Representation of a Lazy computation graph/query against a \link{DataFrame}.
This allows for whole-query optimisation in addition to parallelism,
and is the preferred (and highest-performance) mode of operation for polars.
}
\details{
The \code{pl$LazyFrame(...)} function is a shortcut for \code{pl$DataFrame(...)$lazy()}.
}
\examples{
# Constructing a LazyFrame from vectors:
pl$LazyFrame(a = 1:2, b = 3:4)

# Constructing a LazyFrame from Series:
pl$LazyFrame(pl$Series("a", 1:2), pl$Series("b", 3:4))

# Constructing a LazyFrame from a list:
data <- list(a = 1:2, b = 3:4)

## Using dynamic dots feature
pl$LazyFrame(!!!data)
}
\seealso{
\itemize{
\item \code{\link[=lazyframe__collect]{<LazyFrame>$collect()}}: Materialize a \link{LazyFrame} into a \link{DataFrame}.
}
}
