% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__min}
\alias{pl__min}
\title{Get the minimum value}
\usage{
pl__min(names, ...)
}
\arguments{
\item{names}{The name(s) or \link[=DataType]{data type} of the column(s) to use in
the aggregation. A character vector or a Polars data type/list of data
types can be supplied.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A polars \link{expression}
}
\description{
This function is syntactic sugar for \code{col(names)$min()}.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

# Get the minimum value of a column
df$select(pl$min("a"))

# Get the minimum value of multiple columns
df$select(pl$min(c("a", "b")))
}
