% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/input-ipc-functions.R
\name{pl__read_ipc_stream}
\alias{pl__read_ipc_stream}
\title{Read into a DataFrame from Arrow IPC stream format}
\usage{
pl__read_ipc_stream(
  source,
  ...,
  columns = NULL,
  n_rows = NULL,
  row_index_name = NULL,
  row_index_offset = 0L
)
}
\arguments{
\item{source}{A character of the path to an Arrow IPC stream file.}

\item{...}{These dots are for future extensions and must be empty.}

\item{columns}{A character vector of column names to read.}

\item{n_rows}{Stop reading from the source after reading \code{n_rows}.}

\item{row_index_name}{If not \code{NULL}, this will insert a row index column with
the given name.}

\item{row_index_offset}{Offset to start the row index column (only used if
the name is set by \code{row_index_name}).}
}
\value{
A polars \link{DataFrame}
}
\description{
Read into a DataFrame from Arrow IPC stream format
}
\examples{
\dontshow{if (\{
requireNamespace("nanoarrow", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
temp_file <- tempfile(fileext = ".arrows")

mtcars |>
  nanoarrow::write_nanoarrow(temp_file)

pl$read_ipc_stream(temp_file, columns = c("cyl", "am"))
\dontshow{\}) # examplesIf}
}
