% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-repeat.R
\name{pl__repeat_}
\alias{pl__repeat_}
\title{Construct a column of length n`` filled with the given value}
\usage{
pl__repeat_(value, n, ..., dtype = NULL)
}
\arguments{
\item{value}{Value to repeat.}

\item{n}{Length of the resulting column}

\item{...}{These dots are for future extensions and must be empty.}

\item{dtype}{Data type of the resulting column. If \code{NULL} (default), data
type is inferred from the given value. Defaults to Int32 for integer values,
unless Int64 is required to fit the given value. Defaults to Float64 for
float values.}
}
\value{
A polars \link{expression}
}
\description{
Construct a column of length n`` filled with the given value
}
\details{
If you want to construct a column in lazy mode and do not need a
pre-determined length, use \code{pl$lit()} instead.
}
\examples{
# Construct a column with a repeated value in a lazy context.
pl$select(pl$repeat_("z", n = 3))

# Specify an output dtype
pl$select(pl$repeat_(3, n = 3, dtype = pl$Int8))
}
