% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-lazy.R
\name{pl__row_index}
\alias{pl__row_index}
\title{Generate a row index}
\usage{
pl__row_index(name = "index")
}
\arguments{
\item{name}{Name of the returned column.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The length of the returned sequence will match the context length. If you
would like to generate sequences with custom offsets / length / step size /
datatypes, it is recommended to use \code{\link[=pl__int_range]{pl$int_range()}} instead.
}
\examples{
df <- pl$DataFrame(x = c("A", "A", "B", "B", "B"))
df$with_columns(pl$row_index(), pl$row_index("another_index"))

df$group_by("x")$agg(pl$row_index())$sort("x")

df$select(pl$row_index())
}
