% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/meta-versions.R
\name{pl__show_versions}
\alias{pl__show_versions}
\title{Print out the version of Polars and its optional dependencies}
\usage{
pl__show_versions()
}
\value{
\code{NULL} invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Print out the version of Polars and its optional dependencies.
}
\details{
\link[cli:cli-package]{cli} enhances the terminal output, especially error messages.

These packages may be used for exporting \link{Series} to R.
See \code{\link[=series__to_r_vector]{<Series>$to_r_vector()}} for details.
\itemize{
\item \link[bit64:bit64-package]{bit64}
\item \link[blob:blob]{blob}
\item \link[clock:clock-package]{clock}
\item \link[data.table:data.table-package]{data.table}
\item \link[hms:hms-package]{hms}
\item \link[tibble:tibble-package]{tibble}
\item \link[vctrs:vctrs-package]{vctrs}
}
}
\examples{
pl$show_versions()
}
