% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_all}
\alias{pl_all}
\title{New Expr referring to all columns}
\usage{
pl_all(name = NULL)
}
\arguments{
\item{name}{Character vector indicating on which columns the AND operation
should be applied.}
}
\value{
Boolean literal
}
\description{
Not to mix up with \code{Expr_object$all()} which is a 'reduce Boolean columns by
AND' method.
}
\examples{
test = pl$DataFrame(col_1 = c(TRUE, TRUE), col_2 = c(TRUE, FALSE))
test

# here, the first `$all()` selects all columns, and the second `$all()` checks
# whether all values are true in each column
test$with_columns(pl$all()$all())
}
