% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_all_horizontal}
\alias{pl_all_horizontal}
\title{Apply the AND logical rowwise}
\arguments{
\item{...}{Columns to concatenate into a single string column. Accepts
expressions. Strings are parsed as column names, other non-expression inputs
are parsed as literals.}
}
\value{
Expr
}
\description{
Apply the AND logical rowwise
}
\examples{
df = pl$DataFrame(
  a = c(TRUE, FALSE, NA, NA),
  b = c(TRUE, FALSE, NA, NA),
  c = c(TRUE, FALSE, NA, TRUE)
)
df$with_columns(
  pl$all_horizontal("a", "b", "c")$alias("all")
)
}
