% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_functions.R
\name{pl_any}
\alias{pl_any}
\title{Any}
\arguments{
\item{...}{\itemize{
\item Single string: alias for `pl$col(name)$any()``
\item Series: alias for \code{name$any()}
\item Otherwise, this function computes the bitwise OR horizontally across multiple columns.
}}
}
\description{
Evaluate a bitwise OR operation.
}
\examples{
df = pl$DataFrame(
  "a" = c(TRUE, FALSE, TRUE),
  "b" = c(FALSE, FALSE, FALSE),
  "c" = c(FALSE, TRUE, FALSE)
)
pl$any(df$select("a")$to_series())
df$select(pl$any("*"))$to_data_frame()

}
