% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_approx_n_unique}
\alias{pl_approx_n_unique}
\title{Approximate count of unique values}
\usage{
pl_approx_n_unique(...)
}
\arguments{
\item{...}{Characters indicating the column names.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(...)$approx_n_unique()},
and uses the HyperLogLog++ algorithm for cardinality estimation.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 1),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$approx_n_unique("a"))

df$select(pl$approx_n_unique("b", "c"))
}
\seealso{
\itemize{
\item \code{\link[=Expr_approx_n_unique]{<Expr>$approx_n_unique()}}
}
}
