% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_arg_sort_by}
\alias{pl_arg_sort_by}
\title{Return the row indices that would sort the columns}
\usage{
pl_arg_sort_by(
  ...,
  descending = FALSE,
  nulls_last = FALSE,
  multithreaded = TRUE,
  maintain_order = FALSE
)
}
\arguments{
\item{...}{Column(s) to arg sort by. Can be Expr(s) or something coercible
to Expr(s). Strings are parsed as column names.}

\item{descending}{Logical. Sort in descending order (default is \code{FALSE}). This must be
either of length 1 or a logical vector of the same length as the number of
Expr(s) specified in \code{by} and \code{...}.}

\item{nulls_last}{A logical or logical vector of the same length as the number of columns.
If \code{TRUE}, place \code{null} values last insead of first.}

\item{multithreaded}{A logical. If \code{TRUE}, sort using multiple threads.}

\item{maintain_order}{Whether the order should be maintained if elements are
equal. If \code{TRUE}, streaming is not possible and performance might be worse
since this requires a stable search.}
}
\value{
Expr
}
\description{
Return the row indices that would sort the columns
}
\examples{
df = pl$DataFrame(
  a = c(0, 1, 1, 0),
  b = c(3, 2, 3, 2)
)

df$with_columns(
  arg_sort_a = pl$arg_sort_by("a"),
  arg_sort_ab = pl$arg_sort_by(c("a", "b"), descending = TRUE)
)

# we can also pass Expr
df$with_columns(
  arg_sort_a = pl$arg_sort_by(pl$col("a") * -1)
)
}
\seealso{
\link[=Expr_arg_sort]{$arg_sort()} to find the row indices that would
sort an Expr.
}
