% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_arg_where}
\alias{pl_arg_where}
\title{Return indices that match a condition}
\usage{
pl_arg_where(condition)
}
\arguments{
\item{condition}{An Expr that gives a boolean.}
}
\value{
Expr
}
\description{
Return indices that match a condition
}
\examples{
df = pl$DataFrame(a = c(1, 2, 3, 4, 5))
df$select(
  pl$arg_where(pl$col("a") \%\% 2 == 0)
)
}
