% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_col}
\alias{pl_col}
\title{Create an expression representing column(s) in a dataframe}
\usage{
pl_col(...)
}
\arguments{
\item{...}{One of the following:
\itemize{
\item character vectors
\item \link[=pl_dtypes]{RPolarsDataTypes}
\item a list of \link[=pl_dtypes]{RPolarsDataTypes}
}}
}
\value{
\link[=Expr_class]{Expr} of a column or columns
}
\description{
Create an expression representing column(s) in a dataframe
}
\examples{
# a single column by a character
pl$col("foo")

# multiple columns by characters
pl$col("foo", "bar")

# multiple columns by RPolarsDataTypes
pl$col(pl$dtypes$Float64, pl$dtypes$String)

# multiple character vectors and a list of RPolarsDataTypes are also allowed
pl$col(c("foo", "bar"), "baz")
pl$col("foo", c("bar", "baz"))
pl$col(list(pl$dtypes$Float64, pl$dtypes$String))

# there are some special notations for selecting columns
df = pl$DataFrame(foo = 1:3, bar = 4:6, baz = 7:9)

## select all columns with a wildcard `"*"`
df$select(pl$col("*"))

## select multiple columns by a regular expression
## starts with `^` and ends with `$`
df$select(pl$col(c("^ba.*$")))
}
