% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_concat_str}
\alias{pl_concat_str}
\title{Horizontally concatenate columns into a single string column}
\arguments{
\item{...}{Columns to concatenate into a single string column. Accepts
expressions. Strings are parsed as column names, other non-expression inputs
are parsed as literals. Non-Utf8 columns are cast to Utf8.}

\item{separator}{String that will be used to separate the values of each
column.}
}
\value{
Expr
}
\description{
Horizontally concatenate columns into a single string column
}
\examples{
df = pl$DataFrame(
  a = c(1, 2, 3),
  b = c("dogs", "cats", NA),
  c = c("play", "swim", "walk")
)

df$with_columns(
  pl$concat_str(
    pl$col("a") * 2,
    "b",
    "c",
    pl$lit("!"),
    separator = " "
  )$alias("full_sentence")
)

}
