% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_corr}
\alias{pl_corr}
\title{Correlation}
\usage{
pl_corr(a, b, method = "pearson", propagate_nans = FALSE)
}
\arguments{
\item{a}{One column name or Expr or anything convertible Into\if{html}{\out{<Expr>}} via \code{pl$col()}.}

\item{b}{Another column name or Expr or anything convertible Into\if{html}{\out{<Expr>}} via \code{pl$col()}.}

\item{method}{str One of 'pearson' or 'spearman'}

\item{propagate_nans}{bool Used only when calculating the spearman rank correlation.
If \code{True} any \code{NaN} encountered will lead to \code{NaN} in the output.
Defaults to \code{False} where \code{NaN} are regarded as larger than any finite number and thus lead to the highest rank.}
}
\value{
Expr for the computed correlation
}
\description{
Calculates the correlation between two columns
}
\examples{
lf = as_polars_lf(data.frame(a = c(1, 8, 3), b = c(4, 5, 2)))
lf$select(pl$corr("a", "b", method = "spearman"))$collect()
}
