% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_count}
\alias{pl_count}
\title{Return the number of non-null values in the column.}
\usage{
pl_count(...)
}
\arguments{
\item{...}{Characters indicating the column names.}
}
\value{
\link[=Expr_class]{Expression} of data type \link[=pl_dtypes]{UInt32}
}
\description{
This function is syntactic sugar for \code{pl$col(...)$count()}.
}
\details{
Calling this function without any arguments returns the number of rows in the context.
This way of using the function is deprecated.
Please use \code{\link[=pl_len]{pl$len()}} instead.
}
\examples{
df = pl$DataFrame(
  a = c(1, 2, NA),
  b = c(3, NA, NA),
  c = c("foo", "bar", "foo")
)

df$select(pl$count("a"))

df$select(pl$count(c("b", "c")))
}
\seealso{
\itemize{
\item \code{\link[=pl_len]{pl$len()}}
\item \code{\link[=Expr_count]{<Expr>$count()}}
}
}
