% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_cov}
\alias{pl_cov}
\title{Covariance}
\arguments{
\item{a}{str or Expr One column}

\item{b}{str of Expr The other column}
}
\value{
Expr for the computed covariance
}
\description{
Calculates the covariance between two columns
}
\examples{
lf <- pl$LazyFrame(data.frame(a = c(1, 8, 3), b = c(4, 5, 2)))
lf$select(pl$cov("a", "b"))$collect()
}
