% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_functions.R
\name{pl_cumsum}
\alias{pl_cumsum}
\title{Cumulatively sum all values}
\arguments{
\item{...}{is a:
If one arg:
\itemize{
\item Series or Expr, same as \code{column$cumsum()}
\item string, same as \code{pl$col(column)$cumsum()}
\item numeric, same as \code{pl$lit(column)$cumsum()}
\item list of strings(column names) or expressions to cumulatively sum horizontally across multiple columns.
}

If several args, then wrapped in a list and handled as above.}
}
\value{
Expr
}
\description{
syntactic sugar for starting a expression with cumsum
}
\examples{
df = pl$DataFrame(a = 1:2, b = 3:4, c = 5:6)
df$select(pl$cumsum("a"))
df$with_columns(pl$cumsum("a", "b", "c"))$to_data_frame()
}
