% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_datetime}
\alias{pl_datetime}
\title{Create a Datetime expression}
\usage{
pl_datetime(
  year,
  month,
  day,
  hour = NULL,
  minute = NULL,
  second = NULL,
  microsecond = NULL,
  ...,
  time_unit = "us",
  time_zone = NULL,
  ambiguous = "raise"
)
}
\arguments{
\item{year}{An Expr or something coercible to an Expr, that must return an
integer. Strings are parsed as column names. Floats are cast to integers.}

\item{month}{An Expr or something coercible to an Expr, that must return an
integer between 1 and 12. Strings are parsed as column names. Floats are
cast to integers.}

\item{day}{An Expr or something coercible to an Expr, that must return an
integer between 1 and 31. Strings are parsed as column names. Floats are
cast to integers.}

\item{hour}{An Expr or something coercible to an Expr, that must return an
integer between 0 and 23. Strings are parsed as column names. Floats are
cast to integers.}

\item{minute}{An Expr or something coercible to an Expr, that must return an
integer between 0 and 59. Strings are parsed as column names. Floats are
cast to integers.}

\item{second}{An Expr or something coercible to an Expr, that must return an
integer between 0 and 59. Strings are parsed as column names. Floats are
cast to integers.}

\item{microsecond}{An Expr or something coercible to an Expr, that must
return an integer between 0 and 999,999. Strings are parsed as column
names. Floats are cast to integers.}

\item{...}{Not used.}

\item{time_unit}{Unit of time. One of \code{"ms"}, \code{"us"} (default) or \code{"ns"}.}

\item{time_zone}{Time zone string, as defined in \code{\link[=OlsonNames]{OlsonNames()}}.
Setting \code{timezone = "*"} will match any timezone, which can be useful to
select all Datetime columns containing a timezone.}

\item{ambiguous}{Determine how to deal with ambiguous datetimes:
\itemize{
\item \code{"raise"} (default): raise
\item \code{"earliest"}: use the earliest datetime
\item \code{"latest"}: use the latest datetime
}}
}
\value{
An Expr of type Datetime
}
\description{
Create a Datetime expression
}
\examples{
df = pl$DataFrame(
  year = 2019:2021,
  month = 9:11,
  day = 10:12,
  min = 55:57
)

df$with_columns(
  dt_from_cols = pl$datetime("year", "month", "day", minute = "min"),
  dt_from_lit = pl$datetime(2020, 3, 5, hour = 20:22),
  dt_from_mix = pl$datetime("year", 3, 5, second = 1)
)

# floats are coerced to integers
df$with_columns(
  dt_floats = pl$datetime(2018.8, 5.3, 1, second = 2.1)
)

# if datetime can't be constructed, it returns null
df$with_columns(
  dt_floats = pl$datetime(pl$lit("abc"), -2, 1)
)

# can control the time_unit
df$with_columns(
  dt_from_cols = pl$datetime("year", "month", "day", minute = "min", time_unit = "ms")
)
}
\seealso{
\itemize{
\item \code{\link[=pl_date]{pl$date()}}
\item \code{\link[=pl_time]{pl$time()}}
}
}
