% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{pl_deserialize_lf}
\alias{pl_deserialize_lf}
\title{Read a logical plan from a JSON file to construct a LazyFrame}
\usage{
pl_deserialize_lf(json)
}
\arguments{
\item{json}{A character of the JSON representation of the logical plan.}
}
\value{
\link[=LazyFrame_class]{LazyFrame}
}
\description{
Read a logical plan from a JSON file to construct a LazyFrame
}
\examples{
lf = pl$LazyFrame(a = 1:3)$sum()
json = lf$serialize()
pl$deserialize_lf(json)$collect()
}
\seealso{
\itemize{
\item \code{\link[=LazyFrame_serialize]{<LazyFrame>$serialize()}}
}
}
