% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polars_options.R
\name{pl_disable_string_cache}
\alias{pl_disable_string_cache}
\title{Disable the global string cache}
\usage{
pl_disable_string_cache()
}
\value{
This doesn't return any value.
}
\description{
Some functions (e.g joins) can be applied on Categorical series only allowed
if using the global string cache is enabled. This function disables
the string_cache. In general, you should use \code{pl$with_string_cache()} instead.
}
\examples{
pl$enable_string_cache()
pl$using_string_cache()
pl$disable_string_cache()
pl$using_string_cache()
}
\seealso{
\code{\link[=pl_using_string_cache]{pl$using_string_cache}}
\code{\link[=pl_enable_string_cache]{pl$enable_string_cache}}
\code{\link[=pl_with_string_cache]{pl$with_string_cache}}
}
\keyword{options}
