% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{pl_dtypes}
\alias{pl_dtypes}
\title{DataTypes polars types}
\description{
\code{DataType} any polars type (ported so far)
}
\examples{
print(ls(pl$dtypes))
pl$dtypes$Float64
pl$dtypes$Utf8

pl$List(pl$List(pl$UInt64))

pl$Struct(pl$Field("CityNames", pl$Utf8))

# Some DataType use case, this user function fails because....
\dontrun{
pl$Series(1:4)$apply(\(x) letters[x])
}
# The function changes type from Integer(Int32)[Integers] to char(Utf8)[Strings]
# specifying the output DataType: Utf8 solves the problem
pl$Series(1:4)$apply(\(x) letters[x], datatype = pl$dtypes$Utf8)

}
