% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_element}
\alias{pl_element}
\alias{element}
\title{an element in 'eval'-expr}
\value{
Expr
}
\description{
Alias for an element in evaluated in an \code{eval} expression.
}
\examples{
pl$lit(1:5)$cumulative_eval(pl$element()$first() - pl$element()$last()**2)$to_r()
}
\keyword{Expr}
