% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pl_empty_select}
\alias{pl_empty_select}
\alias{.onLoad}
\alias{empty_select}
\title{Select from a empty DataFrame}
\usage{
pl_empty_select
}
\arguments{
\item{...}{expressions passed to select}
}
\value{
DataFrame
}
\description{
Select from a empty DataFrame
}
\details{
experimental
}
\examples{
pl$empty_select(pl$lit(1:4))
}
\keyword{DataFrame}
