% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pl_enable_string_cache}
\alias{pl_enable_string_cache}
\title{Toggle the global string cache}
\arguments{
\item{toggle}{Boolean. TRUE enable, FALSE disable.}
}
\value{
enable_string_cache: no return
}
\description{
Some functions (e.g joins) can be applied on Categorical series only allowed
if using the global string cache is enabled. This function enables or disables
the string_cache and override any contexts made by \code{pl$with_string_cache()}.
In general, you should use \code{pl$with_string_cache()} instead.
}
\examples{
pl$enable_string_cache(TRUE)
pl$using_string_cache()
}
\seealso{
\code{\link[=pl_using_string_cache]{pl$using_string_cache}}
\code{\link[=pl_with_string_cache]{pl$with_string_cache}}
}
\keyword{options}
