% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{pl_expr_to_r}
\alias{pl_expr_to_r}
\title{Convert an Expr to R output}
\usage{
pl_expr_to_r(expr, df = NULL, i = 0)
}
\arguments{
\item{df}{If \code{NULL} (default), it evaluates the Expr in an empty DataFrame.
Otherwise, provide a DataFrame that the Expr should be evaluated in.}

\item{i}{Numeric column to extract. Default is zero (which gives the first
column).}
}
\value{
R object
}
\description{
This is mostly useful to debug an expression. It evaluates the Expr in an
empty DataFrame and return the first Series to R. This is an alias for
\verb{$to_r()}.
}
\examples{
pl$expr_to_r(pl$lit(1:3))
}
