% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_field}
\alias{pl_field}
\title{Quickly select a field in a Struct}
\usage{
pl_field(name)
}
\arguments{
\item{name}{Name of the field to select.}
}
\value{
An Expr with the datatype from the selected field.
}
\description{
This is syntactic sugar that should mostly be used in
\code{\link[=ExprStruct_with_fields]{$struct$with_fields()}}. \code{pl$field("x")} is
equivalent to \code{pl$col("my_struct")$struct$field("x")}.
}
\examples{
df = pl$DataFrame(x = c(1, 4, 9), y = c(4, 9, 16), multiply = c(10, 2, 3))$
  with_columns(coords = pl$struct(c("x", "y")))$
  select("coords", "multiply")

df

df = df$with_columns(
  pl$col("coords")$struct$with_fields(
    pl$field("x")$sqrt(),
    y_mul = pl$field("y") * pl$col("multiply")
  )
)

df

df$unnest("coords")
}
