% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_first}
\alias{pl_first}
\title{pl$first}
\arguments{
\item{column}{if dtype is:
\itemize{
\item Series: Take first value in \code{Series}
\item str: syntactic sugar for \verb{pl.col(..).first()}
\item NULL: expression to take first column of a context.
}}
}
\value{
Expr or first value of input Series
}
\description{
Depending on the input type this function does different things:
}
\examples{

df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)
df$select(pl$first())

df$select(pl$first("a"))

pl$first(df$get_column("a"))

}
\keyword{Expr_new}
