% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_fold}
\alias{pl_fold}
\title{Accumulate over multiple columns horizontally with an R function}
\usage{
pl_fold(acc, lambda, exprs)
}
\arguments{
\item{acc}{an Expr or Into\if{html}{\out{<Expr>}} of the initial accumulator.}

\item{lambda}{R function which takes two polars Series as input and return one.}

\item{exprs}{Expressions to aggregate over. May also be a wildcard expression.}
}
\value{
An expression that will be applied rowwise
}
\description{
This allows one to do rowwise operations, starting with an initial value
(\code{acc}). See \code{pl$reduce()} to do rowwise operations without this initial
value.
}
\examples{
df = pl$DataFrame(mtcars)

# Make the row-wise sum of all columns
df$with_columns(
  pl$fold(
    acc = pl$lit(0),
    lambda = \(acc, x) acc + x,
    exprs = pl$col("*")
  )$alias("mpg_drat_sum_folded")
)
}
