% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_implode}
\alias{pl_implode}
\title{Aggregate all column values into a list.}
\arguments{
\item{name}{Name of the column(s) that should be imploded, passed to pl$col()}
}
\value{
Expr
}
\description{
Aggregate all column values into a list.
}
\examples{
pl$DataFrame(iris)$select(pl$implode("Species"))
}
\keyword{Expr}
