% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_int_ranges}
\alias{pl_int_ranges}
\title{Generate a range of integers for each row of the input columns}
\usage{
pl_int_ranges(start = 0, end = NULL, step = 1, ..., dtype = pl$Int64)
}
\arguments{
\item{start}{Start of the range (inclusive). Defaults to 0.}

\item{end}{End of the range (exclusive). If \code{NULL} (default), the value of
\code{start} is used and \code{start} is set to 0.}

\item{step}{Step size of the range.}

\item{...}{Not used.}

\item{dtype}{Data type of the range.}
}
\value{
An Expr with the data type List(\code{dtype}) (with \code{Int64} as default of
\code{dtype}).
}
\description{
Generate a range of integers for each row of the input columns
}
\examples{
df = pl$DataFrame(start = c(1, -1), end = c(3, 2))

df$with_columns(int_range = pl$int_ranges("start", "end"))

df$with_columns(int_range = pl$int_ranges("start", "end", dtype = pl$Int16))
}
\seealso{
\code{\link[=pl_int_range]{pl$int_range()}} to generate a single range of
integers.
}
