% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{pl_is_schema}
\alias{pl_is_schema}
\title{check if schema}
\format{
function
}
\usage{
pl$is_schema(x)
}
\arguments{
\item{x}{object to test if schema}
}
\value{
bool
}
\description{
check if schema
}
\examples{
pl$is_schema(pl$DataFrame(iris)$schema)
pl$is_schema(list("alice", "bob"))
}
\keyword{functions}
