% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_last}
\alias{pl_last}
\title{Get the last value.}
\usage{
pl_last(columns = NULL)
}
\arguments{
\item{columns}{A character vector indicating the column names or \code{NULL}.
If \code{NULL} (default), returns an expression to take the last column
of the context instead.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function has different behavior depending on the input type:
\itemize{
\item \code{NULL} -> Takes last column of a context.
\item A character vector -> Syntactic sugar for \code{pl$col(columns)$last()}.
}
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "baz")
)

df$select(pl$last())

df$select(pl$last("a"))

df$select(pl$last(c("b", "c")))
}
\seealso{
\itemize{
\item \code{\link[=Expr_last]{<Expr>$last()}}
}
}
