% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{pl_LazyFrame}
\alias{pl_LazyFrame}
\title{Create new LazyFrame}
\arguments{
\item{...}{Anything that is accepted by \code{pl$DataFrame()}}
}
\value{
LazyFrame
}
\description{
This is simply a convenience function to create \code{LazyFrame}s in a quick way.
It is a wrapper around \code{pl$DataFrame()$lazy()}. Note that this should only
be used for making examples and quick demonstrations.
}
\examples{
pl$LazyFrame(
  a = list(c(1, 2, 3, 4, 5)),
  b = 1:5,
  c = letters[1:5],
  d = list(1:1, 1:2, 1:3, 1:4, 1:5)
) # directly from vectors

# from a list of vectors or data.frame
pl$LazyFrame(list(
  a = c(1, 2, 3, 4, 5),
  b = 1:5,
  c = letters[1:5],
  d = list(1L, 1:2, 1:3, 1:4, 1:5)
))
}
\keyword{LazyFrame_new}
