% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_len}
\alias{pl_len}
\title{Return the number of rows in the context.}
\usage{
pl_len()
}
\value{
\link[=Expr_class]{Expression} of data type \link[=pl_dtypes]{UInt32}
}
\description{
This is similar to \verb{COUNT(*)} in SQL.
}
\examples{
df = pl$DataFrame(
  a = c(1, 2, NA),
  b = c(3, NA, NA),
  c = c("foo", "bar", "foo")
)

df$select(pl$len())
}
\seealso{
\itemize{
\item \code{\link[=Expr_count]{<Expr>$count()}}
}
}
