% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{pl_List}
\alias{pl_List}
\title{Create List DataType}
\format{
function
}
\arguments{
\item{datatype}{an inner DataType}
}
\value{
a list DataType with an inner DataType
}
\description{
Create List DataType
}
\examples{
pl$List(pl$List(pl$Boolean))
}
\keyword{pl}
