% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_max_horizontal}
\alias{pl_max_horizontal}
\title{Get the maximum value rowwise}
\usage{
pl_max_horizontal(...)
}
\arguments{
\item{...}{Columns to concatenate into a single string column. Accepts
expressions. Strings are parsed as column names, other non-expression inputs
are parsed as literals.}
}
\value{
Expr
}
\description{
Get the maximum value rowwise
}
\examples{
df = pl$DataFrame(
  a = NA_real_,
  b = c(2:1, NA_real_, NA_real_),
  c = c(1:2, NA_real_, Inf)
)
df$with_columns(
  pl$max_horizontal("a", "b", "c", 99.9)$alias("max")
)
}
