% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_mean}
\alias{pl_mean}
\title{pl$mean}
\usage{
pl_mean(...)
}
\arguments{
\item{column}{if dtype is:
\itemize{
\item Series: Take mean value in \code{Series}
\item DataFrame or LazyFrame: Take mean value of each column
\item str: syntactic sugar for \code{pl$col(..)$mean()}
\item NULL: expression to take mean column of a context.
}}
}
\value{
Expr or mean value of input Series
}
\description{
Depending on the input type this function does different things:
}
\examples{

df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)
df$select(pl$mean("a"))

df$select(pl$mean("a", "b"))

}
\keyword{Expr_new}
