% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_mean_horizontal}
\alias{pl_mean_horizontal}
\title{Compute the mean rowwise}
\usage{
pl_mean_horizontal(...)
}
\arguments{
\item{...}{Columns to concatenate into a single string column. Accepts
expressions. Strings are parsed as column names, other non-expression inputs
are parsed as literals.}
}
\value{
Expr
}
\description{
Compute the mean rowwise
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3, 6, 7),
  b = c(4, 5, NA_real_, Inf, NaN)
)

df$with_columns(
  pl$mean_horizontal("a", "b")$alias("mean"),
  pl$mean_horizontal("a", "b", 5)$alias("mean_with_lit")
)
}
