% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pl_mem_address}
\alias{pl_mem_address}
\alias{mem_address}
\title{Get Memory Address}
\arguments{
\item{robj}{an R object}
}
\value{
String of mem address
}
\description{
mimics pl$mem_address
}
\examples{
pl$mem_address(pl$Series(1:3))
}
