% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_min}
\alias{pl_min}
\title{Get the minimum value.}
\usage{
pl_min(columns)
}
\arguments{
\item{columns}{A character vector indicating the column names.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Syntactic sugar for \code{pl$col(columns)$min()}.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$min("a"))

# Get the minimum value of multiple columns.
df$select(pl$min("^a|b$"))

df$select(pl$min(c("a", "b")))
}
\seealso{
\itemize{
\item \code{\link[=Expr_min]{<Expr>$min()}}
\item \code{\link[=pl_min_horizontal]{pl$min_horizontal()}}
}
}
