% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_n_unique}
\alias{pl_n_unique}
\title{Count unique values.}
\usage{
pl_n_unique(...)
}
\arguments{
\item{...}{Characters indicating the column names, passed to \code{\link[=pl_col]{pl$col()}}.
See \code{\link[=pl_col]{?pl_col}} for details.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(...)$n_unique()}.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 1),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$n_unique("a"))

df$select(pl$n_unique("b", "c"))
}
\seealso{
\itemize{
\item \code{\link[=Expr_n_unique]{<Expr>$n_unique()}}
}
}
