% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pl_set_options}
\alias{pl_set_options}
\alias{pl_reset_options}
\title{Set polars options}
\usage{
pl_set_options(
  ...,
  strictly_immutable = TRUE,
  maintain_order = FALSE,
  do_not_repeat_call = FALSE,
  debug_polars = FALSE,
  no_messages = FALSE,
  rpool_cap = 4,
  int64_conversion = c("bit64", "double", "string")
)

pl_reset_options()
}
\arguments{
\item{...}{Ignored.}

\item{strictly_immutable}{Keep polars strictly immutable. Polars/arrow is in
general pro "immutable objects". Immutability is also classic in R. To mimic
the Python-polars API, set this to \code{FALSE.}}

\item{maintain_order}{Default for all \code{maintain_order} options (present in
\verb{$group_by()} or \verb{$unique()} for example).}

\item{do_not_repeat_call}{Do not print the call causing the error in error
messages. The default (\code{FALSE}) is to show them.}

\item{debug_polars}{Print additional information to debug Polars.}

\item{no_messages}{Hide messages.}

\item{rpool_cap}{The maximum number of R sessions that can be used to process
R code in the background. See Details.}

\item{int64_conversion}{How should Int64 values be handled when converting a
polars object to R?
\itemize{
\item \code{"double"} (default) converts the integer values to double.
\item \code{"bit64"} uses \code{bit64::as.integer64()} to do the conversion (requires
the package \code{bit64} to be attached).
\item \code{"string"} converts Int64 values to character.
}}
}
\value{
\code{polars_options()} returns a named list with the value (\code{TRUE} or \code{FALSE}) of
each option.

\code{pl$set_options()} silently modifies the options values.

\code{pl$reset_options()} silently resets the options to their default values.
}
\description{
[DEPRECATED] Please use \code{options()} to set options, \code{polars_options()} to
get them, and \code{polars_options_reset()} to reset them. See the documentation
of \code{\link[=polars_options]{polars_options()}} for details.
}
