% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{pl_PolarsLazyFrame}
\alias{pl_PolarsLazyFrame}
\title{Create new PolarsLazyFrame}
\arguments{
\item{...}{Anything that is accepted by \code{pl$DataFrame()}}
}
\value{
PolarsLazyFrame
}
\description{
This is simply a convenience function to create \code{PolarsLazyFrame}s in a quick way.
It is a wrapper around \code{pl$DataFrame()$lazy()}. Note that this should only
be used for making examples and quick demonstrations.
}
\examples{
pl$PolarsLazyFrame(
  a = list(c(1, 2, 3, 4, 5)),
  b = 1:5,
  c = letters[1:5],
  d = list(1:1, 1:2, 1:3, 1:4, 1:5)
) # directly from vectors

# from a list of vectors or data.frame
pl$PolarsLazyFrame(list(
  a = c(1, 2, 3, 4, 5),
  b = 1:5,
  c = letters[1:5],
  d = list(1L, 1:2, 1:3, 1:4, 1:5)
))

}
\keyword{PolarsLazyFrame_new}
