% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_functions.R
\name{pl_quantile}
\alias{pl_quantile}
\title{Quantile}
\arguments{
\item{column}{Column name.}
}
\value{
LazyFrame
}
\description{
Syntactic sugar for \code{pl$col("foo")$quantile(...)}
}
\examples{
df = pl$DataFrame(mtcars)
df$select(pl$quantile("mpg", .25))
}
