% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_reduce}
\alias{pl_reduce}
\title{Accumulate over multiple columns horizontally / rowwise with a left fold}
\arguments{
\item{lambda}{Function to apply over the accumulator and the value.}

\item{exprs}{Expressions to aggregate over. May also be a wildcard expression.}
}
\value{
An expression that will be applied rowwise
}
\description{
\code{pl$fold()} and \code{pl$reduce()} allows one to do rowwise operations. The only
difference between them is that \code{pl$fold()} has an additional argument (\code{acc})
that contains the value that will be initialized when the fold starts.
}
\examples{
df = pl$DataFrame(mtcars)

# Make the rowwise sum of all columns and add 1 to it
df$with_columns(
  pl$reduce(
     lambda = \(acc, x) acc + x, exprs = pl$col("mpg", "drat")
  )
)
}
