% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{pl_rep}
\alias{pl_rep}
\alias{extend_constant}
\title{Repeat a series}
\format{
functino
}
\arguments{
\item{value}{expr or any valid input to pl$lit (literal)
This value may be None to fill with nulls.}

\item{n}{Numeric the number of times to repeat, must be non-negative and finite}

\item{rechunk}{bool default = TRUE, if true memory layout will be rewritten}
}
\value{
Expr
}
\description{
This expression emulates R rep()
}
\examples{
pl$select(pl$rep(1:3, n = 5))
}
\keyword{Expr}
