% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_rolling_corr}
\alias{pl_rolling_corr}
\title{Rolling correlation}
\arguments{
\item{a}{str or Expr One column}

\item{b}{str of Expr The other column}

\item{window_size}{int The length of the window}

\item{min_periods}{NULL or int The number of values in the window that should be non-null before computing a result.
If NULL, it will be set equal to window size.}

\item{ddof}{integer Delta Degrees of Freedom: the divisor used in the calculation is N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
Expr for the computed rolling correlation
}
\description{
Calculates the rolling correlation between two columns
}
\examples{
lf <- pl$LazyFrame(data.frame(a = c(1, 8, 3), b = c(4, 5, 2)))
lf$select(pl$rolling_corr("a", "b", window_size = 2))$collect()
}
