% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pl_select}
\alias{pl_select}
\title{Select from an empty DataFrame}
\format{
method
}
\value{
DataFrame
}
\description{
Select from an empty DataFrame
}
\details{
param ... expressions passed to select
\code{pl$select} is a shorthand for \code{pl$DataFrame(list())$select}

NB param of this function
}
\examples{
pl$select(
  pl$lit(1:4)$alias("ints"),
  pl$lit(letters[1:4])$alias("letters")
)
}
\keyword{DataFrame}
