% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{pl_SQLContext}
\alias{pl_SQLContext}
\title{Initialise a new SQLContext}
\arguments{
\item{...}{Name-value pairs of \link{LazyFrame} like objects to register.}
}
\value{
RPolarsSQLContext
}
\description{
Create a new SQLContext and register the given LazyFrames.
}
\examples{
ctx = pl$SQLContext(mtcars = mtcars)
ctx
}
