% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_std}
\alias{pl_std}
\title{Standard deviation}
\usage{
pl$std(column, ddof = 1)
}
\arguments{
\item{column}{Column name.}

\item{ddof}{Delta Degrees of Freedom: the divisor used in the calculation is
N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
Expr or Series matching type of input column
}
\description{
syntactic sugar for starting a expression with std
}
