% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{pl_Struct_datatype}
\alias{pl_Struct_datatype}
\title{Create Struct DataType}
\format{
function
}
\arguments{
\item{datatype}{an inner DataType}
}
\value{
a list DataType with an inner DataType
}
\description{
Struct DataType Constructor
}
\examples{
# create a Struct-DataType
pl$List(pl$List(pl$Boolean))

# Find any DataType via pl$dtypes
print(pl$dtypes)
}
