% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_sum_horizontal}
\alias{pl_sum_horizontal}
\title{Compute the sum rowwise}
\usage{
pl_sum_horizontal(...)
}
\arguments{
\item{...}{Columns to concatenate into a single string column. Accepts
expressions. Strings are parsed as column names, other non-expression inputs
are parsed as literals.}
}
\value{
Expr
}
\description{
Compute the sum rowwise
}
\examples{
df = pl$DataFrame(
  a = NA_real_,
  b = c(3:4, NA_real_, NA_real_),
  c = c(1:2, NA_real_, -Inf)
)
df$with_columns(
  pl$sum_horizontal("a", "b", "c", 2)$alias("sum")
)
}
