% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_time}
\alias{pl_time}
\title{Create a Time expression}
\usage{
pl_time(hour = NULL, minute = NULL, second = NULL, microsecond = NULL)
}
\arguments{
\item{hour}{An Expr or something coercible to an Expr, that must return an
integer between 0 and 23. Strings are parsed as column names. Floats are
cast to integers.}

\item{minute}{An Expr or something coercible to an Expr, that must return an
integer between 0 and 59. Strings are parsed as column names. Floats are
cast to integers.}

\item{second}{An Expr or something coercible to an Expr, that must return an
integer between 0 and 59. Strings are parsed as column names. Floats are
cast to integers.}

\item{microsecond}{An Expr or something coercible to an Expr, that must
return an integer between 0 and 999,999. Strings are parsed as column
names. Floats are cast to integers.}
}
\value{
An Expr of type Time
}
\description{
Create a Time expression
}
\examples{
df = pl$DataFrame(hour = 19:21, min = 9:11, sec = 10:12, micro = 1)

df$with_columns(
  time_from_cols = pl$time("hour", "min", "sec", "micro"),
  time_from_lit = pl$time(12, 3, 5),
  time_from_mix = pl$time("hour", 3, 5)
)

# floats are coerced to integers
df$with_columns(
  time_floats = pl$time(12.5, 5.3, 1)
)

# if time can't be constructed, it returns null
df$with_columns(
  time_floats = pl$time(pl$lit("abc"), -2, 1)
)
}
\seealso{
\itemize{
\item \code{\link[=pl_datetime]{pl$datetime()}}
\item \code{\link[=pl_date]{pl$date()}}
}
}
