% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_unique}
\alias{pl_unique}
\title{count n unique values}
\arguments{
\item{column}{if dtype is:
\itemize{
\item String: syntactic sugar for \code{pl$col(column)$approx_unique()}, returns Expr
\item Expr: syntactic sugar for \code{column$approx_unique()}, returns Expr
}}
}
\value{
Expr or value

Expr or value
}
\description{
Depending on the input type this function does different things:

This is done using the HyperLogLog++ algorithm for cardinality estimation.
}
\examples{
#column as Series
pl$n_unique(pl$Series(1:4)) == 4

#column as String
expr = pl$n_unique("bob")
print(expr)
pl$DataFrame(bob = 1:4)$select(expr)

#colum as Expr
pl$DataFrame(bob = 1:4)$select(pl$n_unique(pl$col("bob")))
#column as Series
pl$approx_unique(pl$Series(1:4)) == 4

#column as String
expr = pl$approx_unique("bob")
print(expr)
pl$DataFrame(bob = 1:4)$select(expr)

#colum as Expr
pl$DataFrame(bob = 1:4)$select(pl$approx_unique(pl$col("bob")))
}
\keyword{Expr_new}
