% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pl_using_string_cache}
\alias{pl_using_string_cache}
\title{enable disable global string cache}
\value{
using_string_cache: Boolean
}
\description{
get if currently global string cache is active.
}
\examples{
pl$using_string_cache()
}
\seealso{
\code{\link[=pl_with_string_cache]{pl$with_string_cache}}
\code{\link[=pl_enable_string_cache]{pl$enable_enable_cache}}
}
\keyword{options}
