% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error__trait.R
\name{plain}
\alias{plain}
\title{Internal generic method to add plain text to error message}
\usage{
plain(err, msg)
}
\arguments{
\item{err}{some error type object}

\item{msg}{string to add}
}
\value{
condition
}
\description{
Internal generic method to add plain text to error message
}
\keyword{internal}
