% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/polars-package.R
\docType{package}
\name{polars-package}
\alias{polars}
\alias{polars-package}
\title{polars: R Bindings for the 'polars' Rust Library}
\description{
Lightning-fast 'DataFrame' library written in 'Rust'. Convert
R data to 'Polars' data and vice versa. Perform fast, lazy,
larger-than-memory and optimized data queries. 'Polars' is
interoperable with the package 'arrow', as both are based on the
'Apache Arrow' Columnar Format.
}
\seealso{
Useful links:
\itemize{\item\url{https://pola-rs.github.io/r-polars/}
\item\url{https://github.com/pola-rs/r-polars}
\item\url{https://rpolars.r-universe.dev/polars}
\item Report bugs at \url{https://github.com/pola-rs/r-polars/issues}
}
}
\author{
\strong{Maintainer}: Tatsuya Shima \email{ts1s1andn@gmail.com}



Authors:
\itemize{\item Tatsuya Shima \email{ts1s1andn@gmail.com}
\item Etienne Bacher \email{etienne.bacher@protonmail.com} (\href{https://orcid.org/0000-0002-9271-5075}{ORCID})
\item Authors of the dependency Rust crates
}
}
\keyword{internal}
