% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{polars_class_object}
\alias{polars_class_object}
\title{Any polars class object is made of this}
\value{
not applicable
}
\description{
One SEXP of Rtype: "externalptr" + a class attribute
}
\details{
\itemize{
\item \code{object$method()} calls are facilitated by a \verb{$.ClassName}- s3method see 'R/after-wrappers.R'
\item Code completion is facilitated by \code{.DollarNames.ClassName}-s3method see e.g. 'R/dataframe__frame.R'
\item Implementation of property-methods as DataFrame_columns() and syntax checking is an extension to \verb{$.ClassName}
See function macro_add_syntax_check_to_class().
}
}
\examples{
# all a polars object is only made of:
some_polars_object = pl$DataFrame(iris)
str(some_polars_object) # External Pointer tagged with a class attribute.

# All state is stored on rust side.

# The single exception from the rule is class "GroupBy", where objects also have
# two private attributes "groupby_input" and "maintain_order".
str(pl$DataFrame(iris)$group_by("Species"))
}
\keyword{api_object}
