% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/autocompletion.R
\name{polars_code_completion_activate}
\alias{polars_code_completion_activate}
\alias{polars_code_completion_deactivate}
\title{Polars code completion}
\usage{
polars_code_completion_activate(..., verbose = TRUE)

polars_code_completion_deactivate(..., verbose = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{verbose}{Inform whether code completion is active or not.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This only works in RStudio.
}
\details{
This intercepts Rstudio internal functions \code{.rs.getCompletionsFunction} &
\code{.rs.getCompletionsDollar} in the loaded session environment \code{tools:rstudio}.

Pressing tab will literally evaluate the left-hand-side in order to compute
possible suggestions. Therefore, it may be slow on large data or large queries.
}
\examples{
if (interactive()) \{
  # Activate completion
  polars_code_completion_activate()

  # Add a "$" and press tab to see functions for LazyFrame
  pl$LazyFrame(x = 1)

  # Add a "$" after pl$col("x") and press tab to see functions for expressions
  pl$LazyFrame(x = 1)$select(pl$col("x"))

  # Add a "$" in count_matches() and press tab to see possible arguments
  pl$LazyFrame(x = 1)$select(pl$col("x")$str$count_matches())

  # Deactivate (restarting the R session also deactivates code completion)
  polars_code_completion_deactivate()
\}
}
