% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/polars_info.R
\name{polars_info}
\alias{polars_info}
\title{Report information of the package}
\usage{
polars_info()
}
\value{
A list with information of the package
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function reports the following information:
\itemize{
\item Package versions (the Polars R package version and the dependent Rust
Polars crate version)
\item Compatibility information of Polars.
\item Number of threads used by Polars
\item Rust feature flags (See \code{vignette("install", "polars")} for details)
}
}
\examples{
polars_info()

polars_info()$versions

polars_info()$features$nightly
}
