% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/polars_options.R
\name{polars_options}
\alias{polars_options}
\alias{polars_options_reset}
\title{Get and reset polars options}
\usage{
polars_options()

polars_options_reset()
}
\value{
\code{polars_options()} returns a named list where the names are option names and
values are option values.

\code{polars_options_reset()} doesn't return anything.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{polars_options()} returns a list of options for polars. Options
can be set with \code{\link[=options]{options()}}. Note that \strong{options must be prefixed with
"polars."}, e.g to modify the option \code{to_r_vector.int64} you need to pass
\code{options(polars.to_r_vector.int64 =)}. See below for a description of all
options.

\code{polars_options_reset()} brings all polars options back to their default
value.
}
\details{
The following options are available (in alphabetical order, with the
default value in parenthesis):
\itemize{
\item \code{compat_level} will affect
\code{\link[=as_nanoarrow_array_stream.polars_series]{as_nanoarrow_array_stream(<series>)}}'s
\code{polars_compat_level} argument and \code{\link[=lazyframe__sink_ipc]{<lazyframe>$sink_ipc()}}'s
\code{compat_level} argument. See the documentation of those functions for details.
\item for all \verb{to_r_vector.*} options, see arguments of \link[=series__to_r_vector]{to_r_vector()}.
\item \code{df_knitr_print} (TODO: possible values??)
}
}
\examples{
\dontshow{if (\{
requireNamespace("withr", quietly = TRUE) && requireNamespace("hms", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
library(hms)
polars_options()
withr::with_options(
  list(polars.to_r_vector.int64 = "character"),
  polars_options()
)
\dontshow{\}) # examplesIf}
}
