% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/output-partition.R
\name{pl__PartitionBy}
\alias{pl__PartitionBy}
\alias{PartitionBy}
\alias{polars_partitioning_scheme}
\title{Partitioning scheme to write files}
\usage{
pl__PartitionBy(
  base_path,
  ...,
  key = NULL,
  include_key = NULL,
  max_rows_per_file = NULL,
  approximate_bytes_per_file = NULL
)
}
\arguments{
\item{base_path}{The base path for the output files.
Use the \code{mkdir} option of the \verb{sink_*} methods to ensure directories
in the path are created.}

\item{...}{These dots are for future extensions and must be empty.}

\item{key}{Something can be coerced to a list of \link[=polars_expr]{expressions}, or \code{NULL} (default).
Used to partition by.}

\item{include_key}{A bool indicating whether to include the key columns in the output files.
Can only be used if \code{key} is specified, otherwise should be \code{NULL}.}

\item{max_rows_per_file}{An integer-ish value indicating the maximum size in rows of each of the generated files.}

\item{approximate_bytes_per_file}{An integer-ish value indicating approximate number of bytes to write to each file, or \code{NULL}.
This is measured as the estimated size of the DataFrame in memory.
Defaults to approximately 4GB when \code{key} is specified without \code{max_rows_per_file};
otherwise unlimited.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Partitioning schemes are used to write multiple files with \verb{sink_*} and
\verb{write_*} methods.
\itemize{
\item \code{\link[=polars_partitioning_scheme]{pl$PartitionBy()}}: Configuration for writing to
multiple output files. Supports partitioning by key expressions, file size limits,
or both.
}
}
\examples{
\dontshow{if (\{
requireNamespace("withr", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
# Partitioning by columns
temp_dir_1 <- withr::local_tempdir()
as_polars_lf(mtcars)$sink_parquet(
  pl$PartitionBy(
    temp_dir_1,
    key = c("cyl", "am"),
    include_key = FALSE
  ),
  mkdir = TRUE
)
list.files(temp_dir_1, recursive = TRUE)

# Partitioning by max row size
temp_dir_2 <- withr::local_tempdir()
as_polars_lf(mtcars)$sink_csv(
  pl$PartitionBy(
    temp_dir_2,
    max_rows_per_file = 10
  ),
  mkdir = TRUE
)

files <- list.files(temp_dir_2, full.names = TRUE)
files
lapply(files, \\(x) nrow(read.csv(x)))

# Partitioning by both key and size
temp_dir_3 <- withr::local_tempdir()
as_polars_lf(mtcars)$sink_parquet(
  pl$PartitionBy(
    temp_dir_3,
    key = "cyl",
    max_rows_per_file = 5,
    approximate_bytes_per_file = 1000000
  ),
  mkdir = TRUE
)
list.files(temp_dir_3, recursive = TRUE)
\dontshow{\}) # examplesIf}
}
