% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{polars_runtime_flags}
\alias{polars_runtime_flags}
\alias{runtime_state}
\title{internal keeping of state at runtime}
\format{
An object of class \code{environment} of length 0.
}
\usage{
runtime_state
}
\description{
This environment is used internally for the package to remember
what has been going on. Currently only used to throw one-time warnings()
}
\keyword{internal}
