% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{polars_series}
\alias{polars_series}
\title{series wrapper}
\usage{
polars_series(x, name = NULL)
}
\arguments{
\item{x}{input to form new series from}

\item{name}{name of series}
}
\value{
a polar_series object
}
\description{
series wrapper
}
\examples{
new_polar_series(1:3,name="some_name")
}
