% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__background.R
\name{PolarsBackgroundHandle_is_exhausted}
\alias{PolarsBackgroundHandle_is_exhausted}
\title{PolarsBackgroundHandle}
\usage{
PolarsBackgroundHandle_is_exhausted()
}
\value{
Bool
}
\description{
PolarsBackgroundHandle
}
\examples{
lazy_df = pl$DataFrame(iris[, 1:3])$lazy()$select(pl$all()$first())
handle = lazy_df$collect_background()
handle$is_exhausted()
df = handle$join()
handle$is_exhausted()
}
\keyword{internal}
