% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__background.R
\name{PolarsBackgroundHandle_join}
\alias{PolarsBackgroundHandle_join}
\title{PolarsBackgroundHandle}
\usage{
PolarsBackgroundHandle_join()
}
\value{
DataFrame
}
\description{
PolarsBackgroundHandle
}
\examples{
lazy_df = pl$DataFrame(iris[, 1:3])$lazy()$select(pl$all()$first())
handle = lazy_df$collect_background()
df = handle$join()

}
\keyword{internal}
