% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_drop}
\alias{PolarsLazyFrame_drop}
\title{Drop}
\usage{
PolarsLazyFrame_drop(columns)
}
\arguments{
\item{columns}{character vector Name of the column(s) that should be removed from the dataframe.}
}
\value{
PolarsLazyFrame
}
\description{
Remove columns from the dataframe.
}
\examples{
pl$DataFrame(mtcars)$lazy()$drop(c("mpg", "hp"))
}
\keyword{PolarsLazyFrame}
