% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_filter}
\alias{PolarsLazyFrame_filter}
\title{Apply filter to PolarsLazyFrame}
\format{
An object of class \code{character} of length 1.
}
\usage{
PolarsLazyFrame_filter(expr)
}
\arguments{
\item{expr}{one Expr or string naming a column}
}
\value{
A new \code{PolarsLazyFrame} object with add/modified column.
}
\description{
Filter rows with an Expression definining a boolean column
}
\examples{
pl$DataFrame(iris)$lazy()$filter(pl$col("Species") == "setosa")$collect()
}
\keyword{PolarsLazyFrame}
