% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_first}
\alias{PolarsLazyFrame_first}
\title{First}
\usage{
PolarsLazyFrame_first
}
\value{
A new \code{DataFrame} object with applied filter.
}
\description{
Get the first row of the DataFrame.
}
\examples{
pl$DataFrame(mtcars)$lazy()$first()$collect()
}
\keyword{DataFrame}
