% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_head}
\alias{PolarsLazyFrame_head}
\title{Head}
\usage{
PolarsLazyFrame_head(n)
}
\arguments{
\item{n}{positive numeric or integer number not larger than 2^32}
}
\value{
A new \code{PolarsLazyFrame} object with applied filter.
}
\description{
Get the first n rows.
}
\details{
any number will converted to u32. Negative raises error
}
\examples{
pl$DataFrame(mtcars)$lazy()$head(4)$collect()
}
\keyword{PolarsLazyFrame}
