% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_join}
\alias{PolarsLazyFrame_join}
\title{PolarsLazyFrame join}
\usage{
PolarsLazyFrame_join(
  other,
  left_on = NULL,
  right_on = NULL,
  on = NULL,
  how = c("inner", "left", "outer", "semi", "anti", "cross"),
  suffix = "_right",
  allow_parallel = TRUE,
  force_parallel = FALSE
)
}
\arguments{
\item{other}{PolarsLazyFrame}

\item{left_on}{names of columns in self PolarsLazyFrame, order should match. Type, see on param.}

\item{right_on}{names of columns in other PolarsLazyFrame, order should match. Type, see on param.}

\item{on}{named columns as char vector of named columns, or list of expressions and/or strings.}

\item{how}{a string selecting one of the following methods: inner, left, outer, semi, anti, cross}

\item{suffix}{name to added right table}

\item{allow_parallel}{bool}

\item{force_parallel}{bool}
}
\value{
A new \code{PolarsLazyFrame} object with applied join.
}
\description{
join a PolarsLazyFrame
}
\keyword{PolarsLazyFrame}
