% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_last}
\alias{PolarsLazyFrame_last}
\title{Last}
\usage{
PolarsLazyFrame_last
}
\value{
A new \code{PolarsLazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their maximum value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$last()$collect()
}
\keyword{PolarsLazyFrame}
