% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_max}
\alias{PolarsLazyFrame_max}
\title{Max}
\usage{
PolarsLazyFrame_max
}
\value{
A new \code{PolarsLazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their maximum value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$max()$collect()
}
\keyword{PolarsLazyFrame}
