% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_mean}
\alias{PolarsLazyFrame_mean}
\title{Mean}
\usage{
PolarsLazyFrame_mean
}
\value{
A new \code{PolarsLazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their mean value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$mean()$collect()
}
\keyword{PolarsLazyFrame}
