% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_median}
\alias{PolarsLazyFrame_median}
\title{Median}
\usage{
PolarsLazyFrame_median
}
\value{
A new \code{PolarsLazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their median value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$median()$collect()
}
\keyword{PolarsLazyFrame}
