% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_min}
\alias{PolarsLazyFrame_min}
\title{Min}
\usage{
PolarsLazyFrame_min
}
\value{
A new \code{PolarsLazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their minimum value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$min()$collect()
}
\keyword{PolarsLazyFrame}
