% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_print}
\alias{PolarsLazyFrame_print}
\title{print PolarsLazyFrame internal method}
\format{
An object of class \code{character} of length 1.
}
\usage{
PolarsLazyFrame_print(x)
}
\arguments{
\item{x}{PolarsLazyFrame}
}
\value{
self
}
\description{
can be used i the middle of a method chain
}
\examples{
pl$DataFrame(iris)$lazy()$print()
}
\keyword{PolarsLazyFrame}
