% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_rename}
\alias{PolarsLazyFrame_rename}
\title{Rename columns of a PolarsLazyFrame}
\usage{
PolarsLazyFrame_rename(...)
}
\arguments{
\item{...}{One of the following:
\itemize{
\item params like \code{new_name = "old_name"} to rename selected variables.
\item as above but, but params wrapped in a list
}}
}
\value{
PolarsLazyFrame
}
\description{
Rename columns of a PolarsLazyFrame
}
\examples{
pl$DataFrame(mtcars)$
  lazy()$
  rename(miles_per_gallon = "mpg", horsepower = "hp")$
  collect()

}
\keyword{PolarsLazyFrame}
