% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_select}
\alias{PolarsLazyFrame_select}
\title{Lazy_select}
\usage{
PolarsLazyFrame_select(...)
}
\arguments{
\item{...}{any single Expr or string naming a column}
}
\value{
A new \code{PolarsLazyFrame} object with applied filter.
}
\description{
select on a PolarsLazyFrame
}
\keyword{PolarsLazyFrame}
