% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_shift}
\alias{PolarsLazyFrame_shift}
\title{Shift}
\usage{
PolarsLazyFrame_shift(periods = 1)
}
\arguments{
\item{periods}{integer Number of periods to shift (may be negative).}
}
\value{
PolarsLazyFrame
}
\description{
Shift the values by a given period.
}
\examples{
pl$DataFrame(mtcars)$lazy()$shift(2)$collect()
}
\keyword{PolarsLazyFrame}
