% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_shift_and_fill}
\alias{PolarsLazyFrame_shift_and_fill}
\title{Shift and fill}
\usage{
PolarsLazyFrame_shift_and_fill(fill_value, periods = 1)
}
\arguments{
\item{fill_value}{fill None values with the result of this expression.}

\item{periods}{integer Number of periods to shift (may be negative).}
}
\value{
PolarsLazyFrame
}
\description{
Shift the values by a given period and fill the resulting null values.
}
\examples{
pl$DataFrame(mtcars)$lazy()$shift_and_fill(0., 2.)$collect()$as_data_frame()
}
\keyword{PolarsLazyFrame}
