% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_std}
\alias{PolarsLazyFrame_std}
\title{Std}
\usage{
PolarsLazyFrame_std(ddof = 1)
}
\arguments{
\item{ddof}{integer Delta Degrees of Freedom: the divisor used in the calculation is N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
A new \code{PolarsLazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns of this PolarsLazyFrame to their standard deviation values.
}
\examples{
pl$DataFrame(mtcars)$lazy()$std()$collect()
}
\keyword{PolarsLazyFrame}
