% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_sum}
\alias{PolarsLazyFrame_sum}
\title{Sum}
\usage{
PolarsLazyFrame_sum
}
\value{
PolarsLazyFrame
}
\description{
Aggregate the columns of this DataFrame to their sum values.
}
\examples{
pl$DataFrame(mtcars)$lazy()$sum()$collect()
}
\keyword{PolarsLazyFrame}
