% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_tail}
\alias{PolarsLazyFrame_tail}
\title{Tail}
\usage{
PolarsLazyFrame_tail(n)
}
\arguments{
\item{n}{positive numeric or integer number not larger than 2^32}
}
\value{
A new \code{PolarsLazyFrame} object with applied filter.
}
\description{
take last n rows of query
}
\details{
any number will converted to u32. Negative raises error
}
\examples{
pl$DataFrame(mtcars)$lazy()$tail(2)$collect()
}
\keyword{PolarsLazyFrame}
