% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_var}
\alias{PolarsLazyFrame_var}
\title{Var}
\usage{
PolarsLazyFrame_var(ddof = 1)
}
\arguments{
\item{ddof}{integer Delta Degrees of Freedom: the divisor used in the calculation is N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
A new \code{PolarsLazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns of this PolarsLazyFrame to their variance values.
}
\examples{
pl$DataFrame(mtcars)$lazy()$var()$collect()
}
\keyword{PolarsLazyFrame}
