% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_with_column}
\alias{PolarsLazyFrame_with_column}
\title{Lazy with column}
\format{
An object of class \code{character} of length 1.
}
\usage{
PolarsLazyFrame_with_column(expr)
}
\arguments{
\item{expr}{one Expr or string naming a column}
}
\value{
A new \code{PolarsLazyFrame} object with add/modified column.
}
\description{
add or replace columns of PolarsLazyFrame
}
\keyword{PolarsLazyFrame}
