% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{PolarsLazyFrame_with_columns}
\alias{PolarsLazyFrame_with_columns}
\title{Lazy with columns}
\usage{
PolarsLazyFrame_with_columns(...)
}
\arguments{
\item{...}{any single Expr or string naming a column}
}
\value{
A new \code{PolarsLazyFrame} object with added/modified columns.
}
\description{
add or replace columns of PolarsLazyFrame
}
\keyword{PolarsLazyFrame}
