% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{prepare_alpha}
\alias{prepare_alpha}
\title{prepare alpha}
\usage{
prepare_alpha(com = NULL, span = NULL, half_life = NULL, alpha = NULL)
}
\arguments{
\item{com}{numeric or NULL}

\item{span}{numeric or NULL}

\item{half_life}{numeric or NULL}

\item{alpha}{numeric or NULL}
}
\value{
numeric
}
\description{
internal function for emw_x expressions
}
\keyword{internal}
