% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{print()}
\alias{print()}
\alias{print.DataFrame}
\alias{DataFrame_print}
\title{s3 method print DataFrame}
\usage{
\method{print}{DataFrame}(x)

DataFrame_print(x)
}
\arguments{
\item{x}{DataFrame}
}
\value{
self

self
}
\description{
s3 method print DataFrame

internal method print DataFrame
}
\examples{
pl$DataFrame(iris)
pl$DataFrame(iris)
}
