% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__whenthen.R
\name{print.ChainedThen}
\alias{print.ChainedThen}
\title{print ChainedThen}
\usage{
\method{print}{ChainedThen}(x, ...)
}
\arguments{
\item{x}{ChainedThen object}

\item{...}{not used}
}
\value{
self
}
\description{
print ChainedThen
}
\examples{
print(pl$when(pl$col("a") > 2)$then(pl$lit("more than two"))$when(pl$col("b") < 5))
}
\keyword{WhenThen}
\keyword{internal}
