% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__whenthen.R
\name{print.ChainedWhen}
\alias{print.ChainedWhen}
\title{print ChainedWhen}
\usage{
\method{print}{ChainedWhen}(x, ...)
}
\arguments{
\item{x}{ChainedWhen object}

\item{...}{not used}
}
\value{
self
}
\description{
print ChainedWhen
}
\examples{
#
print(pl$when(pl$col("a") > 2)$then(pl$lit("more than two"))$when(pl$col("b") < 5))
}
\keyword{WhenThen}
\keyword{internal}
