% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{print.DataFrame}
\alias{print.DataFrame}
\title{S3 method to print a DataFrame}
\usage{
\method{print}{DataFrame}(x, ...)
}
\arguments{
\item{x}{DataFrame}

\item{...}{not used}
}
\value{
self
}
\description{
S3 method to print a DataFrame
}
\examples{
pl$DataFrame(iris)
}
\keyword{internal}
