% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{print.DataType}
\alias{print.DataType}
\title{print a polars datatype}
\usage{
\method{print}{DataType}(x)
}
\arguments{
\item{x}{DataType}
}
\value{
self
}
\description{
print a polars datatype
}
\examples{
minipolars:::DataType$new("Boolean")
}
