% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{print.Expr}
\alias{print.Expr}
\title{Print expr}
\usage{
\method{print}{Expr}(x, ...)
}
\arguments{
\item{x}{Expr}

\item{...}{not used}
}
\value{
self
}
\description{
Print expr
}
\examples{
pl$col("some_column")$sum()$over("some_other_column")
}
\keyword{Expr}
\keyword{internal}
