% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{print.GroupBy}
\alias{print.GroupBy}
\title{print GroupBy}
\usage{
\method{print}{GroupBy}(x)
}
\arguments{
\item{x}{DataFrame}
}
\value{
self
}
\description{
print GroupBy
}
\examples{
pl$DataFrame(iris)$groupby("Species")
}
