% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{print.LazyFrame}
\alias{print.LazyFrame}
\title{print polars polars_lazy_frame}
\usage{
\method{print}{LazyFrame}(x)
}
\arguments{
\item{x}{Rlazyfrane}
}
\value{
self
}
\description{
print polars polars_lazy_frame
}
\examples{
#TODO give example
}
