% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{print.polars_series}
\alias{print.polars_series}
\title{Print rseries}
\usage{
\method{print}{polars_series}(x)
}
\arguments{
\item{x}{Series}
}
\value{
selfie
}
\description{
Print rseries
}
\examples{
pl::series(letters,"lowercase_letters")
}
