% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__background.R
\name{print.PolarsBackgroundHandle}
\alias{print.PolarsBackgroundHandle}
\title{print LazyFrame s3 method}
\usage{
\method{print}{PolarsBackgroundHandle}(x, ...)
}
\arguments{
\item{x}{DataFrame}

\item{...}{not used}
}
\value{
self
}
\description{
print LazyFrame s3 method
}
\examples{
lazy_df = pl$DataFrame(iris[, 1:3])$lazy()$select(pl$all()$first())
handle = lazy_df$collect_background()
handle$is_exhausted()
df = handle$join()
handle$is_exhausted()
}
\keyword{LazyFrame}
\keyword{internal}
