% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{print.PolarsLazyFrame}
\alias{print.PolarsLazyFrame}
\title{print PolarsLazyFrame s3 method}
\usage{
\method{print}{PolarsLazyFrame}(x, ...)
}
\arguments{
\item{x}{DataFrame}

\item{...}{not used}
}
\value{
self
}
\description{
print PolarsLazyFrame s3 method
}
\examples{
print(pl$DataFrame(iris)$lazy())
}
\keyword{PolarsLazyFrame}
\keyword{internal}
