% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{print.Rdataframe}
\alias{print.Rdataframe}
\title{print dataframe}
\usage{
\method{print}{Rdataframe}(x)
}
\arguments{
\item{x}{polar_frame}
}
\value{
self
}
\description{
print dataframe
}
\examples{
pl::pf(iris)
}
