% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{print.Rdatatype}
\alias{print.Rdatatype}
\title{print a polars datatype}
\usage{
\method{print}{Rdatatype}(x)
}
\arguments{
\item{x}{Rdatatype}
}
\value{
self
}
\description{
print a polars datatype
}
\examples{
minipolars:::Rdatatype$new("Boolean")
}
