% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{print.Rexpr}
\alias{print.Rexpr}
\title{Print expr}
\usage{
\method{print}{Rexpr}(x)
}
\arguments{
\item{x}{Rexpr}
}
\value{
self
}
\description{
Print expr
}
\examples{
pl::col("some_column")$sum()$over("some_other_column")
}
