% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{print.RPolarsDataType}
\alias{print.RPolarsDataType}
\title{print a polars datatype}
\usage{
\method{print}{RPolarsDataType}(x, ...)
}
\arguments{
\item{x}{DataType}

\item{...}{not used}
}
\value{
self
}
\description{
print a polars datatype
}
\examples{
pl$dtypes$Boolean # implicit print
}
\keyword{internal}
