% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__whenthen.R
\name{print.RPolarsThen}
\alias{print.RPolarsThen}
\title{print Then}
\usage{
\method{print}{RPolarsThen}(x, ...)
}
\arguments{
\item{x}{When object}

\item{...}{not used}
}
\value{
self
}
\description{
print Then
}
\examples{
print(pl$when(pl$col("a") > 2)$then(pl$lit("more than two")))
}
\keyword{WhenThen}
\keyword{internal}
