% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__whenthen.R
\name{print.RPolarsWhen}
\alias{print.RPolarsWhen}
\title{print When}
\usage{
\method{print}{RPolarsWhen}(x, ...)
}
\arguments{
\item{x}{When object}

\item{...}{not used}
}
\value{
self
}
\description{
print When
}
\examples{
print(pl$when(pl$col("a") > 2))
}
\keyword{WhenThen}
\keyword{internal}
