% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{print.RThreadHandle}
\alias{print.RThreadHandle}
\title{s3 method print RThreadHandle}
\usage{
\method{print}{RThreadHandle}(x, ...)
}
\arguments{
\item{x}{RThreadHandle}

\item{...}{not used}
}
\value{
self
}
\description{
s3 method print RThreadHandle
}
\examples{
handle = pl$LazyFrame()$select(pl$lit(2) + 2)$collect_in_background()
print(handle)
handle$join()
print(handle)
}
\keyword{internal}
