% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{print.Series}
\alias{print.Series}
\title{Print rseries}
\usage{
\method{print}{Series}(x)
}
\arguments{
\item{x}{Series}
}
\value{
self
}
\description{
Print rseries
}
\examples{
pl::series(letters,"lowercase_letters")
}
