% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whenthen.R
\name{print.When}
\alias{print.When}
\title{print When}
\usage{
\method{print}{When}(x, ...)
}
\arguments{
\item{x}{When object}

\item{...}{not used}
}
\value{
self
}
\description{
print When
}
\examples{
print(pl$when(pl$col("a")>2))
}
\keyword{WhenThen}
