% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whenthen.R
\name{print.WhenThen}
\alias{print.WhenThen}
\title{print When}
\usage{
\method{print}{WhenThen}(x, ...)
}
\arguments{
\item{x}{When object}

\item{...}{not used}
}
\value{
self
}
\description{
print When
}
\examples{
print(pl$when(pl$col("a")>2)$then(pl$lit("more than two")))
}
\keyword{WhenThen}
\keyword{internal}
