% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__whenthen.R
\name{print.WhenThenThen}
\alias{print.WhenThenThen}
\title{print When}
\usage{
\method{print}{WhenThenThen}(x, ...)
}
\arguments{
\item{x}{When object}

\item{...}{not used}
}
\value{
self
}
\description{
print When
}
\examples{
#
print(pl$when(pl$col("a")>2)$then(pl$lit("more than two"))$when(pl$col("b")<5))
}
\keyword{WhenThen}
\keyword{internal}
