% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_env}
\alias{print_env}
\title{Print recursively an environment, used in some documentation}
\usage{
print_env(api, name, max_depth = 10)
}
\arguments{
\item{api}{env}

\item{name}{name of env}

\item{max_depth}{numeric/int max levels to recursive iterate through}
}
\value{
Print recursively an environment to the console
}
\description{
Print recursively an environment, used in some documentation
}
\keyword{internal}
