% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pstop}
\alias{pstop}
\title{Internal preferred function to throw errors}
\usage{
pstop(err, call = sys.call(1L))
}
\arguments{
\item{call}{calling context}

\item{msg}{error msg string}
}
\value{
throws an error
}
\description{
Internal preferred function to throw errors
}
\examples{
f = function() pstop("this aint right!!")
f()
}
\keyword{internals}
