% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_private_with_pub_methods}
\alias{replace_private_with_pub_methods}
\title{remove private method}
\usage{
replace_private_with_pub_methods(env, class_pattern, keep = c())
}
\arguments{
\item{env}{class envrionment to modify. Envs are mutable so return needed}

\item{class_pattern}{a regex string matching declared public functions of that class}
}
\description{
extendr places the naked internal calls to rust in env-classes. This function
can be used to delete them and replaces them with the public methods. Which are any function
matching pattern typically '^CLASSNAME' e.g. '^DataFrame_' or '^Series_'. Likely only used in
zzz.R
}
\examples{
replace_private_with_pub_methods(minipolars:::DataFrame, "^DataFrame")
}
