% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restruct_list}
\alias{restruct_list}
\title{restruct list}
\usage{
restruct_list(l)
}
\arguments{
\item{l}{list}
}
\value{
restructed list
}
\description{
lifecycle:: Deprecate
Restruct an object where structs where previously unnested
}
\details{
It was much easier impl export unnested struct from polars. This function
restructs exported unnested structs.
This function should be replaced with rust code writing this output
directly before nesting.
This hack relies on rust uses the tag "is_struct" to mark what should be re-structed.
}
\keyword{internal}
