% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{RField_class}
\alias{RField_class}
\title{Create Field}
\arguments{
\item{name}{Field name}

\item{datatype}{DataType}
}
\value{
A object of with DataType \code{"RField"} containing its name and its
DataType.
}
\description{
A Field is composed of a name and a DataType. Fields are used in Structs-
datatypes and Schemas to represent everything of the Series/Column except the
raw values.
}
\examples{
pl$Field("city_names", pl$Utf8)
}
