% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{RField_datatype}
\alias{RField_datatype}
\title{Get/set Field datatype}
\usage{
<RField>$datatype()
}
\description{
Get/set Field datatype
}
\examples{
field = pl$Field("Cities", pl$String)
field$datatype

field$datatype = pl$Categorical #<- is fine too
field$datatype
}
\keyword{DataFrame}
