% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_rolling.R
\name{RollingGroupBy_agg}
\alias{RollingGroupBy_agg}
\title{Aggregate over a RollingGroupBy}
\usage{
RollingGroupBy_agg(...)
}
\arguments{
\item{...}{Exprs to aggregate over. Those can also be passed wrapped in a
list, e.g \verb{$agg(list(e1,e2,e3))}.}
}
\value{
An aggregated \link[=DataFrame_class]{DataFrame}
}
\description{
Aggregate a DataFrame over a rolling window created with \verb{$rolling()}.
}
\examples{
df = pl$DataFrame(
  dt = c("2020-01-01", "2020-01-01", "2020-01-01", "2020-01-02", "2020-01-03", "2020-01-08"),
  a = c(3, 7, 5, 9, 2, 1)
)$with_columns(
  pl$col("dt")$str$strptime(pl$Date, format = NULL)$set_sorted()
)

df$rolling(index_column = "dt", period = "2d")$agg(
  pl$col("a"),
  pl$sum("a")$alias("sum_a"),
  pl$min("a")$alias("min_a"),
  pl$max("a")$alias("max_a")
)
}
