% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{RThreadHandle_is_finished}
\alias{RThreadHandle_is_finished}
\title{Ask if RThreadHandle is finished?}
\usage{
<RThreadHandle>$is_finished()
}
\value{
trinary value: \code{TRUE} if finished, \code{FALSE} if not, and \code{NULL} if the handle was exhausted
with \code{\link[=RThreadHandle_join]{<RThreadHandle>$join()}}.
}
\description{
Ask if RThreadHandle is finished?
}
\keyword{RThreadHandle}
