% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{RThreadHandle_RThreadHandle_class}
\alias{RThreadHandle_RThreadHandle_class}
\title{The RThreadHandle class}
\value{
see methods
}
\description{
A handle to some polars query running in a background thread.
}
\details{
\verb{<LazyFrame>$collect_in_background()} will execute a polars query detached from the R session
and return a \code{RThreadHandle} immediately. The RThreadHandle has the methods \code{is_finished()} and
\verb{$join()}.

The background thread may access the pool of extra R sessions to process R code
embedded in polars query via \verb{$map(...,background = TRUE)} or \verb{$apply(background=TRUE)}. Use
\code{pl$set_global_rpool_cap()} to limit number of parallel R sessions. Extra R sessions are spawned
and used if \code{background} arg is set to TRUE.

Starting polars \verb{<LazyFrame>$collect_in_background()} with e.g. some
\verb{$map(...,background = FALSE)} will raise an Error as the main R session is not available to
process the query.
}
\examples{
prexpr <- pl$col("mpg")$map(\(x) {Sys.sleep(1.5);x * 0.43}, in_background = TRUE)$alias("kml")
handle = pl$LazyFrame(mtcars)$with_column(prexpr)$collect_in_background()
if(!handle$is_finished()) print("not done yet")
df = handle$join() #get result
df

}
\seealso{
\link{LazyFrame_collect_in_background} \link{Expr_map} \link{Expr_apply}
}
\keyword{RThreadHandle}
