% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-s3-knitr.R, R/series-s3-knitr.R
\name{knit_print.polars_data_frame}
\alias{knit_print.polars_data_frame}
\alias{knit_print.polars_series}
\title{knit print polars DataFrame}
\usage{
\method{knit_print}{polars_data_frame}(x, ...)

\method{knit_print}{polars_series}(x, ...)
}
\arguments{
\item{x}{A polars object}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\value{
\code{x} invisibly or \link[knitr:asis_output]{knit_asis} object.
}
\description{
Mimics Python Polars' NotebookFormatter
for HTML outputs.
}
\details{
Outputs HTML tables if the output format is HTML
and the document's \code{df_print} option is not \code{"default"} or \code{"tibble"}.

Or, the output format can be enforced with R's \code{options} function as follows:
\itemize{
\item \code{options(polars.df_knitr_print  = "default")} for the default print method.
\item \code{options(polars.df_knitr_print  = "html")} for the HTML table.
}
}
\examples{
\dontshow{if (\{
requireNamespace("knitr", quietly = TRUE) && requireNamespace("withr", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
# Using the default print method
withr::with_options(
  list(polars.df_knitr_print = "default"),
  knitr::knit_print(as_polars_df(mtcars))
)

# Returning HTML table
withr::with_options(
  list(polars.df_knitr_print = "html"),
  knitr::knit_print(as_polars_df(mtcars))
)
\dontshow{\}) # examplesIf}
}
