% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.character.Series}
\alias{as.character.Series}
\title{Convert to a character vector}
\usage{
\method{as.character}{Series}(x, ..., str_length = NULL)
}
\arguments{
\item{x}{A Polars Series}

\item{...}{Not used.}

\item{str_length}{An integer. If specified, utf8 or categorical type Series
will be formatted to a string of this length.}
}
\description{
Convert to a character vector
}
\examples{
s = pl$Series(c("foo", "barbaz"))
as.character(s)
as.character(s, str_length = 3)
}
