% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-s3-base.R
\name{as.data.frame.polars_data_frame}
\alias{as.data.frame.polars_data_frame}
\title{Export the polars object as an R DataFrame}
\usage{
\method{as.data.frame}{polars_data_frame}(
  x,
  ...,
  int64 = "double",
  ambiguous = "raise",
  non_existent = "raise"
)
}
\arguments{
\item{x}{A polars object}

\item{...}{Ignored}

\item{int64}{Determine how to convert Polars' Int64, UInt32, or UInt64 type values to R type.
One of the followings:
\itemize{
\item \code{"double"}: Convert to the R's \link{double} type.
\item \code{"character"}: Convert to the R's \link{character} type.
\item \code{"integer"}: Convert to the R's \link{integer} type.
If the value is out of the range of R's integer type, export as \link{NA_integer_}.
\item \code{"integer64"}: Convert to the \link[bit64:bit64-package]{bit64::integer64} class.
The \link[bit64:bit64-package]{bit64} package must be installed.
If the value is out of the range of \link[bit64:bit64-package]{bit64::integer64}, export as \link[bit64:as.integer64.character]{bit64::NA_integer64_}.
}}

\item{ambiguous}{Determine how to deal with ambiguous datetimes.
Character vector or Expression containing the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"earliest"}: Use the earliest datetime
\item \code{"latest"}: Use the latest datetime
\item \code{"null"}: Return a null value
}}

\item{non_existent}{Determine how to deal with non-existent datetimes.
One of the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"null"}: Return a null value
}}
}
\value{
An \link[=data.frame]{R data frame}
}
\description{
This S3 method calls \code{\link[=as.list.polars_data_frame]{as.list(<DataFrame>)}} to
convert the \link[=DataFrame]{polars DataFrame} to a list of \link[=vector]{vectors}, then
makes the list to an \link[=data.frame]{R data frame}.
}
\examples{
df <- as_polars_df(list(a = 1:3, b = 4:6))

as.data.frame(df)
}
