% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-s3-base.R
\name{as.list.polars_data_frame}
\alias{as.list.polars_data_frame}
\title{Export the polars object as an R list}
\usage{
\method{as.list}{polars_data_frame}(
  x,
  ...,
  as_series = FALSE,
  int64 = "double",
  ambiguous = "raise",
  non_existent = "raise"
)
}
\arguments{
\item{x}{A polars object}

\item{...}{Ignored}

\item{as_series}{Whether to convert each column to an \link[=vector]{R vector} or a \link{Series}.
If \code{TRUE}, return a list of \link{Series}, otherwise a list of \link[=vector]{vectors} (default).}

\item{int64}{Determine how to convert Polars' Int64, UInt32, or UInt64 type values to R type.
One of the followings:
\itemize{
\item \code{"double"}: Convert to the R's \link{double} type.
\item \code{"character"}: Convert to the R's \link{character} type.
\item \code{"integer"}: Convert to the R's \link{integer} type.
If the value is out of the range of R's integer type, export as \link{NA_integer_}.
\item \code{"integer64"}: Convert to the \link[bit64:bit64-package]{bit64::integer64} class.
The \link[bit64:bit64-package]{bit64} package must be installed.
If the value is out of the range of \link[bit64:bit64-package]{bit64::integer64}, export as \link[bit64:as.integer64.character]{bit64::NA_integer64_}.
}}

\item{ambiguous}{Determine how to deal with ambiguous datetimes.
Character vector or Expression containing the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"earliest"}: Use the earliest datetime
\item \code{"latest"}: Use the latest datetime
\item \code{"null"}: Return a null value
}}

\item{non_existent}{Determine how to deal with non-existent datetimes.
One of the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"null"}: Return a null value
}}
}
\value{
A \link{list}
}
\description{
This S3 method calls \code{\link[=dataframe__get_columns]{<DataFrame>$get_columns()}} or
\code{\link[=dataframe__to_r_list]{<DataFrame>$to_r_list()}} depending on the \code{as_series} argument.
}
\details{
Arguments other than \code{x} and \code{as_series} are passed to \code{\link[=series__to_r_vector]{<Series>$to_r_vector()}},
so they are ignored when \code{as_series=TRUE}.
}
\examples{
df <- as_polars_df(list(a = 1:3, b = 4:6))

as.list(df, as_series = TRUE)
as.list(df, as_series = FALSE)
}
\seealso{
\itemize{
\item \code{\link[=dataframe__get_columns]{<DataFrame>$get_columns()}}
\item \code{\link[=dataframe__to_r_list]{<DataFrame>$to_r_list()}}
}
}
