% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-nanoarrow.R
\name{as_nanoarrow_array_stream.RPolarsDataFrame}
\alias{as_nanoarrow_array_stream.RPolarsDataFrame}
\title{Create a nanoarrow_array_stream from a Polars object}
\usage{
\method{as_nanoarrow_array_stream}{RPolarsDataFrame}(x, ..., schema = NULL, future = FALSE)
}
\arguments{
\item{x}{\link[=DataFrame_class]{A Polars DataFrame}}

\item{...}{Ignored}

\item{schema}{must stay at default value NULL}

\item{future}{Setting this to \code{TRUE} will write Polars' internal data structures that
might not be available by other Arrow implementations.
This functionality is considered \strong{unstable}.
It may be changed at any point without it being considered a breaking change.}
}
\description{
Create a nanoarrow_array_stream from a Polars object
}
\examples{
library(nanoarrow)
pl_df = as_polars_df(mtcars)

nanoarrow_array_stream = as_nanoarrow_array_stream(pl_df)
as.data.frame(nanoarrow_array_stream)
}
