% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{c.RPolarsSeries}
\alias{c.RPolarsSeries}
\title{Combine to a Series}
\usage{
\method{c}{RPolarsSeries}(x, ...)
}
\arguments{
\item{x}{A Polars Series}

\item{...}{Series(s) or any object that can be converted to a Series.}
}
\value{
a combined Series
}
\description{
Combine to a Series
}
\details{
All objects must have the same datatype. Combining does not rechunk. Read more
about R vectors, Series and chunks in \code{\link[polars]{docs_translations}}:
}
\examples{
s = c(pl$Series(1:5), 3:1, NA_integer_)
s$chunk_lengths() # the series contain three unmerged chunks
}
