% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{[.DataFrame}
\alias{[.DataFrame}
\alias{[.LazyFrame}
\alias{[.Series}
\title{Extract Parts of a Polars Object}
\usage{
\method{[}{DataFrame}(x, i, j, drop = TRUE)

\method{[}{LazyFrame}(x, i, j, drop = TRUE)

\method{[}{Series}(x, i)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}

\item{i}{Rows to select}

\item{j}{Columns to select, either by index or by name.}

\item{drop}{Convert to a Polars Series if only one column is selected.}
}
\description{
Mimics the behavior of [\code{x[i, j, drop = TRUE]}][Extract] for \link{data.frame} or R vector.
}
\details{
\verb{<Series>[i]} is equivalent to \verb{pl$select(<Series>)[i, , drop = TRUE]}.
}
\examples{
df = pl$DataFrame(data.frame(a = 1:3, b = letters[1:3]))

df[1, ]
df[1]
df[, "b"]
df[pl$col("a") >= 2, ]
}
\seealso{
\code{\link[=DataFrame_select]{<DataFrame>$select()}},
\code{\link[=LazyFrame_select]{<LazyFrame>$select()}},
\code{\link[=DataFrame_filter]{<DataFrame>$filter()}},
\code{\link[=LazyFrame_filter]{<LazyFrame>$filter()}}
}
