% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{head.RPolarsDataFrame}
\alias{head.RPolarsDataFrame}
\alias{head.RPolarsLazyFrame}
\title{Take the first n rows}
\usage{
\method{head}{RPolarsDataFrame}(x, n = 6L, ...)

\method{head}{RPolarsLazyFrame}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}

\item{n}{Number of rows}

\item{...}{Not used}
}
\description{
Take the first n rows
}
