% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-knitr.R
\name{knit_print.RPolarsDataFrame}
\alias{knit_print.RPolarsDataFrame}
\title{knit print polars DataFrame}
\usage{
knit_print.RPolarsDataFrame(x, ...)
}
\arguments{
\item{x}{a polars DataFrame to knit_print}

\item{...}{additional arguments, not used}
}
\value{
invisible x or NULL
}
\description{
Mimics python-polars' NotebookFormatter
for HTML outputs.
}
\details{
Outputs HTML tables if the output format is HTML
and the document's \code{df_print} option is not \code{"default"} or \code{"tibble"}.

Or, the output format can be enforced with R's \code{options} function as follows:
\itemize{
\item \code{options(polars.df_knitr_print  = "default")} for the default print method.
\item \code{options(polars.df_knitr_print  = "html")} for the HTML table.
}
}
\keyword{DataFrame}
