% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{length.RPolarsDataFrame}
\alias{length.RPolarsDataFrame}
\alias{length.RPolarsLazyFrame}
\alias{length.Series}
\title{Get the length}
\usage{
\method{length}{RPolarsDataFrame}(x)

\method{length}{RPolarsLazyFrame}(x)

\method{length}{Series}(x)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame}, \link[=LazyFrame_class]{LazyFrame}, or
\link[=Series_class]{Series}}
}
\description{
Get the length
}
