% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{max.RPolarsDataFrame}
\alias{max.RPolarsDataFrame}
\alias{max.RPolarsLazyFrame}
\alias{max.Series}
\title{Compute the maximum value}
\usage{
\method{max}{RPolarsDataFrame}(x, ...)

\method{max}{RPolarsLazyFrame}(x, ...)

\method{max}{Series}(x, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame}, \link[=LazyFrame_class]{LazyFrame}, or
\link[=Series_class]{Series}}

\item{...}{Not used.}
}
\description{
Compute the maximum value
}
