% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{median.DataFrame}
\alias{median.DataFrame}
\alias{median.RPolarsLazyFrame}
\alias{median.Series}
\title{Compute the median}
\usage{
\method{median}{DataFrame}(x, ...)

\method{median}{RPolarsLazyFrame}(x, ...)

\method{median}{Series}(x, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame}, \link[=LazyFrame_class]{LazyFrame}, or
\link[=Series_class]{Series}}

\item{...}{Not used.}
}
\description{
Compute the median
}
