% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{na.omit.RPolarsLazyFrame}
\alias{na.omit.RPolarsLazyFrame}
\alias{na.omit.DataFrame}
\title{Drop missing values}
\usage{
\method{na.omit}{RPolarsLazyFrame}(object, subset = NULL, ...)

\method{na.omit}{DataFrame}(object, subset = NULL, ...)
}
\arguments{
\item{object}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}

\item{subset}{Character vector of column names to drop missing values from.}

\item{...}{Not used.}
}
\description{
Drop missing values
}
\examples{
df = pl$DataFrame(data.frame(a = c(NA, 2:10), b = c(1, NA, 3:10)))$lazy()
na.omit(df)
na.omit(df, subset = "a")
na.omit(df, subset = c("a", "b"))
}
