% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{row.names.RPolarsDataFrame}
\alias{row.names.RPolarsDataFrame}
\title{Get the row names}
\usage{
\method{row.names}{RPolarsDataFrame}(x)
}
\arguments{
\item{x}{A Polars DataFrame}
}
\description{
Get the row names
}
