% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{[.DataFrame}
\alias{[.DataFrame}
\alias{[.LazyFrame}
\title{Take a subset of rows and columns}
\usage{
\method{[}{DataFrame}(x, i, j, ..., drop = TRUE)

\method{[}{LazyFrame}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A Polars DataFrame or LazyFrame}

\item{i}{Rows to select}

\item{j}{Columns to select, either by index or by name.}

\item{...}{Not used.}

\item{drop}{Convert to a Polars Series if only one column is selected.}
}
\description{
Take a subset of rows and columns
}
