% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{sum.DataFrame}
\alias{sum.DataFrame}
\alias{sum.LazyFrame}
\alias{sum.Series}
\title{Compute the sum}
\usage{
\method{sum}{DataFrame}(x, ...)

\method{sum}{LazyFrame}(x, ...)

\method{sum}{Series}(x, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame}, \link[=LazyFrame_class]{LazyFrame}, or
\link[=Series_class]{Series}}

\item{...}{Not used.}
}
\description{
Compute the sum
}
