% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{unique.DataFrame}
\alias{unique.DataFrame}
\alias{unique.LazyFrame}
\title{Drop duplicated rows}
\usage{
\method{unique}{DataFrame}(x, incomparables = FALSE, subset = NULL, keep = "first", ...)

\method{unique}{LazyFrame}(x, incomparables = FALSE, subset = NULL, keep = "first", ...)
}
\arguments{
\item{x}{A Polars DataFrame or LazyFrame}

\item{incomparables}{Not used.}

\item{subset}{Character vector of column names to drop duplicated values from.}

\item{keep}{Either \code{"first"}, \code{"last"}, or \code{"none"}.}

\item{...}{Not used.}
}
\description{
Drop duplicated rows
}
\examples{
df = pl$DataFrame(
  x = as.numeric(c(1, 1:5)),
  y = as.numeric(c(1, 1:5)),
  z = as.numeric(c(1, 1, 1:4))
)
unique(df)
}
