% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{same_outer_datatype}
\alias{same_outer_datatype}
\title{check if x is a valid RPolarsDataType}
\arguments{
\item{lhs}{an RPolarsDataType}

\item{rhs}{an RPolarsDataType}
}
\value{
bool TRUE if outer datatype is the same.
}
\description{
check if x is a valid RPolarsDataType
}
\examples{
# TRUE
pl$same_outer_dt(pl$Datetime("us"),pl$Datetime("ms"))
pl$same_outer_dt(pl$list(pl$Int64),pl$list(pl$Float32))

#FALSE
pl$same_outer_dt(pl$Int64,pl$Float64)
}
\keyword{functions}
