% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{read_csv}
\alias{read_csv}
\title{Read csv to DataFrame}
\arguments{
\item{...}{any argument passed to scan_csv}
}
\value{
DataFrame
}
\description{
Read csv to DataFrame
}
\examples{
df = pl$read_csv("https://j.mp/iriscsv")
}
