% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{series}
\alias{series}
\title{onstructor to tempoariliy replace polars_series}
\usage{
series(x, name = NULL)
}
\arguments{
\item{x}{any vector}

\item{name}{string}
}
\value{
Series
}
\description{
onstructor to tempoariliy replace polars_series
}
\examples{
Series(1:4)
}
