% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__serialize}
\alias{series__serialize}
\alias{pl__deserialize_series}
\title{Serialize and deserialize a Series}
\usage{
series__serialize()

pl__deserialize_series(data)
}
\arguments{
\item{data}{A raw vector of serialized \link{Series}.}
}
\value{
\itemize{
\item \verb{<Series>$serialize()} returns a raw vector of serialized \link{Series}.
\item \code{pl$deserialize_series()} returns a deserialized \link{Series}.
}
}
\description{
Serialize and deserialize a Series
}
\details{
Similar to \code{polars.Series.__getstate__()} and \code{polars.Series.__setstate__()} in Python Polars.
}
\examples{
serialized <- as_polars_series(1:3)$serialize()
serialized

pl$deserialize_series(serialized)
}
