% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__shrink_dtype}
\alias{series__shrink_dtype}
\title{Shrink numeric values to the minimal required datatype}
\usage{
series__shrink_dtype()
}
\value{
A \link[=Series]{polars Series}
}
\description{
Shrink numeric values to the minimal required datatype
}
\examples{
s <- as_polars_series(1:6)
s

s$shrink_dtype()
}
