% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_abs}
\alias{Series_abs}
\alias{abs}
\title{Take absolute value of Series}
\usage{
<Series>$abs()
}
\value{
Series
}
\description{
Take absolute value of Series
}
\examples{
pl$Series(-2:2)$abs()
}
\keyword{Series}
