% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{Series_add}
\alias{Series_add}
\alias{add}
\alias{+.Series}
\title{add Series}
\usage{
Series_add(other)

\method{+}{Series}(s1, s2)
}
\value{
Series
}
\description{
Series arithmetics
}
\examples{
pl$Series(1:3)$add(11:13)
pl$Series(1:3)$add(pl$Series(11:13))
pl$Series(1:3)$add(1L)
1L + pl$Series(1:3)
pl$Series(1:3) + 1L
}
\keyword{Series}
