% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\docType{data}
\name{Series_alias}
\alias{Series_alias}
\alias{alias}
\title{Alias}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_alias
}
\arguments{
\item{name}{a String as the new name}
}
\value{
Series
}
\description{
Change name of Series
}
\examples{
pl$Series(1:3,name = "alice")$alias("bob")
}
\keyword{Series}
