% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_any}
\alias{Series_any}
\title{Reduce boolean Series with ANY}
\usage{
Series_any()
}
\value{
A boolean scalar
}
\description{
Reduce boolean Series with ANY
}
\examples{
pl$Series(c(TRUE, FALSE, NA))$any()
}
