% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\docType{data}
\name{Series_arg_min}
\alias{Series_arg_min}
\alias{Expr_arg_min}
\title{idx to min value}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_arg_min
}
\value{
bool
}
\description{
idx to min value
}
\examples{
pl$Series(c(5,1))$arg_min()
}
\keyword{Series}
