% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_clear}
\alias{Series_clear}
\title{Create an empty or n-row null-filled copy of the Series}
\usage{
Series_clear(n = 0)
}
\arguments{
\item{n}{Number of (null-filled) rows to return in the cleared frame.}
}
\value{
A n-value null-filled Series with an identical schema
}
\description{
Returns a n-row null-filled Series with an identical schema. \code{n} can be
greater than the current number of values in the Series.
}
\examples{
s = pl$Series(name = "a", values = 1:3)

s$clear()

s$clear(n = 5)
}
