% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_clone}
\alias{Series_clone}
\title{Clone a Series}
\usage{
Series_clone()
}
\value{
Series
}
\description{
This makes a very cheap deep copy/clone of an existing
\code{\link[=Series_class]{Series}}. Rarely useful as \code{Series} are nearly 100\%
immutable. Any modification of a \code{Series} should lead to a clone anyways, but
this can be useful when dealing with attributes (see examples).
}
\examples{
df1 = pl$Series(1:10)

# Make a function to take a Series, add an attribute, and return a Series
give_attr = function(data) {
  attr(data, "created_on") = "2024-01-29"
  data
}
df2 = give_attr(df1)

# Problem: the original Series also gets the attribute while it shouldn't!
attributes(df1)

# Use $clone() inside the function to avoid that
give_attr = function(data) {
  data = data$clone()
  attr(data, "created_on") = "2024-01-29"
  data
}
df1 = pl$Series(1:10)
df2 = give_attr(df1)

# now, the original Series doesn't get this attribute
attributes(df1)
}
