% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_compare}
\alias{Series_compare}
\alias{compare}
\alias{==.RPolarsSeries}
\alias{!=.RPolarsSeries}
\alias{<.RPolarsSeries}
\alias{>.RPolarsSeries}
\alias{<=.RPolarsSeries}
\alias{>=.RPolarsSeries}
\title{Compare Series}
\usage{
<Series>$compare(other, op)

\method{==}{RPolarsSeries}(s1, s2)

\method{!=}{RPolarsSeries}(s1, s2)

\method{<}{RPolarsSeries}(s1, s2)

\method{>}{RPolarsSeries}(s1, s2)

\method{<=}{RPolarsSeries}(s1, s2)

\method{>=}{RPolarsSeries}(s1, s2)
}
\arguments{
\item{other}{A Series or something a Series can be created from}

\item{op}{the chosen operator a String either: 'equal', 'not_equal', 'lt', 'gt', 'lt_eq' or 'gt_eq'}

\item{s1}{lhs Series}

\item{s2}{rhs Series or any into Series}
}
\value{
Series
}
\description{
compare two Series
}
\examples{
pl$Series(1:5) == pl$Series(c(1:3, NA_integer_, 10L))
}
\keyword{Series}
