% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_div}
\alias{Series_div}
\title{Divide Series}
\usage{
Series_div(other)
}
\arguments{
\item{other}{\link[=Series_class]{Series} like object of numeric.
Converted to \link[=Series_class]{Series} by \code{\link[=as_polars_series]{as_polars_series()}} in this method.}
}
\value{
\link[=Series_class]{Series}
}
\description{
Method equivalent of division operator \code{series / other}.
}
\examples{
pl$Series(1:3)$div(11:13)
pl$Series(1:3)$div(pl$Series(11:13))
pl$Series(1:3)$div(1L)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
}
}
