% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_dtype}
\alias{Series_dtype}
\title{Get data type of Series}
\usage{
<Series>$dtype()
}
\value{
DataType
}
\description{
Get data type of Series
}
\examples{
pl$Series(1:4)$dtype
pl$Series(c(1, 2))$dtype
pl$Series(letters)$dtype
}
\keyword{Series}
