% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_equals}
\alias{Series_equals}
\title{Are two Series equal?}
\usage{
Series_equals(other, null_equal = FALSE, strict = FALSE)
}
\arguments{
\item{other}{Series to compare with.}

\item{null_equal}{If \code{TRUE}, consider that null values are equal. Overridden
by \code{strict}.}

\item{strict}{If \code{TRUE}, do not allow similar DataType comparison. Overrides
\code{null_equal}.}
}
\value{
A logical value
}
\description{
This checks whether two Series are equal in values and in their name.
}
\examples{
as_polars_series(1:4)$equals(as_polars_series(1:4))

# names are different
as_polars_series(1:4, "bob")$equals(as_polars_series(1:4))

# nulls are different by default
as_polars_series(c(1:4, NA))$equals(as_polars_series(c(1:4, NA)))
as_polars_series(c(1:4, NA))$equals(as_polars_series(c(1:4, NA)), null_equal = TRUE)

# datatypes are ignored by default
as_polars_series(1:4)$cast(pl$Int16)$equals(as_polars_series(1:4))
as_polars_series(1:4)$cast(pl$Int16)$equals(as_polars_series(1:4), strict = TRUE)
}
