% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_expr}
\alias{Series_expr}
\title{Any expr method on a Series}
\usage{
Series_expr()
}
\value{
Expr
}
\description{
Call an expression on a Series
See the individual Expr method pages for full details
}
\details{
This is a shorthand of writing  something like
\code{pl$DataFrame(s)$select(pl$col("sname")$expr)$to_series(0)}

This subnamespace is experimental. Submit an issue if anything
unexpected happend.
}
\examples{
s = pl$Series(list(1:3,1:2,NULL))
s$expr$first()
s$expr$alias("alice")
}
\keyword{Series}
