% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_flags}
\alias{Series_flags}
\title{Get data type of Series}
\usage{
Series_flags()
}
\value{
DataType
}
\description{
Get data type of Series
}
\details{
property sorted flags are not settable, use set_sorted
}
\examples{
pl$Series(1:4)$sort()$flags
}
\keyword{Series}
