% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_is_sorted}
\alias{Series_is_sorted}
\title{Check if the Series is sorted}
\usage{
Series_is_sorted(descending = FALSE)
}
\arguments{
\item{descending}{Check if the Series is sorted in descending order.}
}
\value{
A boolean scalar
}
\description{
Check if the Series is sorted
}
\examples{
pl$Series(1:4)$sort()$is_sorted()
}
\seealso{
Use \code{\link[=Series_set_sorted]{$set_sorted()}} to add a "sorted" flag to the Series
that could be used for faster operations later on.
}
