% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_is_unique}
\alias{Series_is_unique}
\title{Series element(s) is unique}
\usage{
Series_is_unique()
}
\value{
Series
}
\description{
return Boolean vector for all elements that occurs only once
}
\examples{
pl$Series(c(1:2,2L))$is_unique()

}
\keyword{Series}
