% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_item}
\alias{Series_item}
\title{Return the element at the given index}
\usage{
Series_item(index = NULL)
}
\arguments{
\item{index}{Index of the item to return.}
}
\value{
A value of length 1
}
\description{
Return the element at the given index
}
\examples{
s1 = pl$Series(values = 1)

s1$item()

s2 = pl$Series(values = 9:7)

s2$cum_sum()$item(-1)
}
