% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\docType{data}
\name{Series_len}
\alias{Series_len}
\alias{len}
\title{Series_len}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_len
}
\value{
numeric
}
\description{
Length of this Series.
}
\examples{
pl$Series(1:10)$len()

}
\keyword{Series}
