% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_mul}
\alias{Series_mul}
\title{Multiply Series}
\usage{
Series_mul(other)
}
\arguments{
\item{other}{\link[=Series_class]{Series} like object of numeric.
Converted to \link[=Series_class]{Series} by \code{\link[=as_polars_series]{as_polars_series()}} in this method.}
}
\value{
\link[=Series_class]{Series}
}
\description{
Method equivalent of multiplication operator \code{series * other}.
}
\examples{
as_polars_series(1:3)$mul(11:13)
as_polars_series(1:3)$mul(as_polars_series(11:13))
as_polars_series(1:3)$mul(1L)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
}
}
