% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_n_chunks}
\alias{Series_n_chunks}
\title{Get the number of chunks that this Series contains.}
\usage{
Series_n_chunks()
}
\value{
A numeric value
}
\description{
Get the number of chunks that this Series contains.
}
\examples{
s = as_polars_series(1:3)
s$n_chunks()

# Concatenate Series with rechunk = TRUE
s2 = as_polars_series(4:6)
pl$concat(s, s2, rechunk = TRUE)$n_chunks()

# Concatenate Series with rechunk = FALSE
pl$concat(s, s2, rechunk = FALSE)$n_chunks()
}
