% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_n_unique}
\alias{Series_n_unique}
\title{Count unique values in Series}
\usage{
Series_n_unique()
}
\value{
A numeric scalar
}
\description{
Count unique values in Series
}
\examples{
pl$Series(c(1, 2, 1, 4, 4, 1, 5))$n_unique()
}
