% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_name}
\alias{Series_name}
\alias{name}
\title{Property: Name}
\usage{
<Series>$name()
}
\value{
String the name
}
\description{
Get name of Series
}
\examples{
pl$Series(1:3, name = "alice")$name
}
\keyword{Series}
