% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_rename}
\alias{Series_rename}
\title{Rename a series}
\usage{
Series_rename(name, in_place = FALSE)
}
\arguments{
\item{name}{New name.}

\item{in_place}{Rename in-place, which breaks immutability. If \code{TRUE}, you
need to run \code{options(polars.strictly_immutable = FALSE)} before, otherwise
it will throw an error.}
}
\value{
\link[=Series_class]{Series}
}
\description{
Rename a series
}
\examples{
pl$Series(1:4, "bob")$rename("alice")
}
