% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_rep}
\alias{Series_rep}
\title{Duplicate and concatenate a series}
\usage{
Series_rep(n, rechunk = TRUE)
}
\arguments{
\item{n}{Number of times to repeat}

\item{rechunk}{If \code{TRUE} (default), reallocate object in memory which can
speed up some calculations. If \code{FALSE}, the Series will take less space in
memory.}
}
\value{
Series
}
\description{
Note that this function doesn't exist in Python Polars.
}
\examples{
pl$Series(1:2, "bob")$rep(3)
}
