% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_series_equal}
\alias{Series_series_equal}
\alias{series_equal}
\title{Are Series's equal?}
\format{
method
}
\usage{
Series_series_equal(other, null_equal = FALSE, strict = FALSE)
}
\arguments{
\item{other}{Series to compare with}

\item{null_equal}{bool if TRUE, (Null==Null) is true and not Null/NA. Overridden by strict.}

\item{strict}{bool if TRUE, do not allow similar DataType comparison. Overrides null_equal.}
}
\value{
bool
}
\description{
Check if series is equal with another Series.
}
\examples{
pl$Series(1:4,"bob")$series_equal(pl$Series(1:4))
}
\keyword{Series}
