% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_set_sorted}
\alias{Series_set_sorted}
\alias{is_sorted}
\alias{Series}
\alias{set_sorted}
\title{is_sorted}
\usage{
Series_set_sorted(reverse = FALSE, in_place = FALSE)
}
\arguments{
\item{reverse}{bool if TRUE, signals series is Descendingly sorted, otherwise Ascendingly.}

\item{in_place}{if TRUE, will set flag mutably and return NULL. Remember to use
pl$set_rpolars_options(strictly_immutable = F) otherwise an error will be thrown. If FALSE
will return a cloned Series with set_flag which in the very most cases should be just fine.}

\item{nulls_last}{bool where to keep nulls, default same as reverse}
}
\value{
DataType

Series invisible
}
\description{
is_sorted
}
\details{
property sorted flags are not settable, use set_sorted
}
\examples{
pl$Series(1:4)$sort()$is_sorted()
Set sorted
s = pl$Series(1:4)$set_sorted()
s$flags
}
\keyword{Series}
