% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_shape}
\alias{Series_shape}
\title{Shape of series}
\usage{
Series_shape()
}
\value{
dimension vector of Series
}
\description{
Shape of series
}
\examples{
identical(pl$Series(1:2)$shape, 2:1)
}
\keyword{Series}
