% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_sort}
\alias{Series_sort}
\title{Sort a Series}
\usage{
Series_sort(
  ...,
  descending = FALSE,
  nulls_last = FALSE,
  multithreaded = TRUE,
  in_place = FALSE
)
}
\arguments{
\item{...}{Ignored.}

\item{descending}{A logical. If \code{TRUE}, sort in descending order.}

\item{nulls_last}{A logical. If \code{TRUE}, place \code{null} values last insead of first.}

\item{multithreaded}{A logical. If \code{TRUE}, sort using multiple threads.}

\item{in_place}{If \code{TRUE}, this will set the flag mutably and return \code{NULL}.
Remember to use \code{options(polars.strictly_immutable = FALSE)} before using
this parameter, otherwise an error will occur. If \code{FALSE} (default), it will
return a cloned Series with the flag.}
}
\value{
\link[=Series_class]{Series}
}
\description{
Sort a Series
}
\examples{
as_polars_series(c(1.5, NA, 1, NaN, Inf, -Inf))$sort()
as_polars_series(c(1.5, NA, 1, NaN, Inf, -Inf))$sort(nulls_last = TRUE)
}
