% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_std}
\alias{Series_std}
\title{Get the standard deviation of this Series.}
\format{
method
}
\usage{
Series_std(ddof = 1)
}
\arguments{
\item{ddof}{“Delta Degrees of Freedom”: the divisor used in the calculation is N - ddof,
where N represents the number of elements.
By default ddof is 1.}
}
\value{
bool
}
\description{
Get the standard deviation of this Series.
}
\examples{
pl$Series(1:4,"bob")$std()
}
\keyword{Series}
