% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/series-string.R
\name{series_str_json_decode}
\alias{series_str_json_decode}
\title{Parse string values as JSON}
\usage{
series_str_json_decode(dtype = NULL, ..., infer_schema_length = 100L)
}
\arguments{
\item{dtype}{The \link[=DataType]{dtype} to cast the extracted value to, or \code{NULL} (default).
If \code{NULL}, the dtype will be inferred from the JSON value.}

\item{...}{These dots are for future extensions and must be empty.}

\item{infer_schema_length}{The maximum number of rows to scan for schema inference.
If set to \code{NULL}, the full data may be scanned \emph{(this is slow)}.
Only used if the \code{dtype} argument is \code{NULL}.}
}
\value{
A \link[=Series]{polars Series}
}
\description{
Throws an error if invalid JSON strings are encountered.
}
\examples{
s1 <- as_polars_series(c('{"a":1, "b": true}', NA, '{"a":2, "b": false}'))

s2 <- s1$str$json_decode()
s2
s2$dtype

s3 <- s1$str$json_decode(pl$Struct(a = pl$UInt8, b = pl$Boolean))
s3
s3$dtype
}
\seealso{
\itemize{
\item \code{\link[=expr_str_json_decode]{<expr>$str$json_decode()}}
}
}
