% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/series-string.R
\name{series_str_to_decimal}
\alias{series_str_to_decimal}
\title{Convert a String Series into a Decimal Series}
\usage{
series_str_to_decimal(..., scale = NULL, inference_length = 100L)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{scale}{Number of digits after the comma to use for the decimals, or \code{NULL} (default).
If \code{NULL}, the method will infer the scale from the data.}

\item{inference_length}{Number of elements to parse to determine the
\code{precision} and \code{scale} of the \link[=pl__Decimal]{decimal data type}.}
}
\value{
A \link[=Series]{polars Series}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
s <- as_polars_series(c(
  "40.12",
  "3420.13",
  "120134.19",
  "3212.98",
  "12.90",
  "143.09",
  "143.9"
))

s$str$to_decimal()
s$str$to_decimal(scale = 4)
}
\seealso{
\itemize{
\item \code{\link[=expr_str_to_decimal]{<expr>$str$to_decimal()}}
}
}
