% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/series-struct.R
\name{series_struct_unnest}
\alias{series_struct_unnest}
\title{Convert this struct Series to a DataFrame with a separate column for each field}
\usage{
series_struct_unnest()
}
\value{
A polars \link{DataFrame}
}
\description{
Convert this struct Series to a DataFrame with a separate column for each field
}
\examples{
s <- as_polars_series(data.frame(a = c(1, 3), b = c(2, 4)))
s$struct$unnest()
}
\seealso{
\itemize{
\item \code{\link[=as_polars_df]{as_polars_df()}}
}
}
