% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_to_frame}
\alias{Series_to_frame}
\title{Convert Series to DataFrame}
\usage{
Series_to_frame()
}
\value{
DataFrame
}
\description{
Convert Series to DataFrame
}
\examples{
# default will be a DataFrame with empty name
pl$Series(1:4)$to_frame()

pl$Series(1:4, "bob")$to_frame()
}
