% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_to_frames}
\alias{Series_to_frames}
\alias{Series_to_frame}
\title{Series to DataFrame}
\format{
method
}
\usage{
Series_to_frame()
}
\value{
Series
}
\description{
Series to DataFrame
}
\examples{
pl$Series(1:4,"bob")$to_frame()
}
\keyword{Series}
