% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_value_count}
\alias{Series_value_count}
\alias{Series_value_counts}
\title{Value Counts as DataFrame}
\usage{
Series_value_counts(sorted = TRUE, multithreaded = FALSE)
}
\arguments{
\item{sorted}{bool, default TRUE: sort table by value; FALSE: random}

\item{multithreaded}{bool, default FALSE, process multithreaded. Likely faster
to have TRUE for a big Series. If called within an already multithreaded context
such calling apply on a GroupBy with many groups, then likely slightly faster to leave FALSE.}
}
\value{
DataFrame
}
\description{
Value Counts as DataFrame
}
\examples{
pl$Series(iris$Species, "flower species")$value_counts()
}
\keyword{Series}
