% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_var}
\alias{Series_var}
\title{Compute the variance of a Series}
\usage{
Series_var(ddof = 1)
}
\arguments{
\item{ddof}{Delta Degrees of Freedom: the divisor used in the calculation is
N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
A new \code{Series} object with applied aggregation.
}
\description{
Compute the variance of a Series
}
\examples{
pl$Series(1:10)$var()
}
