% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__string.R
\name{SeriesStr_to_lowercase}
\alias{SeriesStr_to_lowercase}
\title{Convert a string to lowercase}
\usage{
SeriesStr_to_lowercase()
}
\value{
Expr of String lowercase chars
}
\description{
Transform to lowercase variant.
}
\examples{
pl$Series(c("A", "B", "C"))$str$to_lowercase()
}
