% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__string.R
\name{SeriesStr_to_uppercase}
\alias{SeriesStr_to_uppercase}
\title{Convert a string to uppercase}
\usage{
SeriesStr_to_uppercase()
}
\value{
Expr of String uppercase chars
}
\description{
Transform to uppercase variant.
}
\examples{
pl$Series(c("a", "b", "c"))$str$to_uppercase()
}
