% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{show_all_public_methods}
\alias{show_all_public_methods}
\title{show all public methods}
\description{
methods are listed by their Class
}
\examples{
pl$show_all_public_methods()
}
\keyword{functions}
