% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{sink_parquet}
\alias{sink_parquet}
\alias{LazyFrame_sink_parquet}
\title{LazyFrame stream output to parquet file}
\usage{
LazyFrame_sink_parquet(
  path,
  method = "zstd",
  level = 3,
  stats = FALSE,
  row_group_size = NULL,
  data_pagesize_limit = NULL,
  order = TRUE
)
}
\arguments{
\item{path}{string, the path of the parquet file}

\item{method}{string, the compression method. One of \link{'uncompressed', 'snappy', 'gzip', 'lzo', 'brotli', 'zstd'}}

\item{level}{null or int, the level of compression. Only used if method is one of \link{'gzip', 'brotli', 'zstd'}}

\item{stats}{bool, whether compute and write column statistics.}

\item{row_group_size}{NULL or positive integer. If set NULL a single row group will be created.}

\item{data_pagesize_limit}{NULL or positive integer. If set NULL the limit will be 2^20 bytes.}

\item{order}{bool, whether maintain the order the data was processed.}
}
\description{
Stream the content of LazyFrame into a parquet file.
}
