% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/sql-context.R
\name{sql_context__register}
\alias{sql_context__register}
\title{Register a single frame as a table, using the given name}
\usage{
sql_context__register(name, frame = NULL)
}
\arguments{
\item{name}{Name of the table.}

\item{frame}{Object to associate with this table name.}
}
\value{
An object of class \code{"polars_sql_context"}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(x = 1)
ctx <- pl$SQLContext()
ctx$register("frame_data", df)$execute("SELECT * FROM frame_data")$collect()
}
