% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/sql-context.R
\name{sql_context__register_many}
\alias{sql_context__register_many}
\title{Register multiple eager/lazy frames as tables, using the associated names}
\usage{
sql_context__register_many(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Elements that are known in the
current \code{SQLContext}. It accepts any R object that can be converted to a
LazyFrame via \code{as_polars_lf()}. All elements must be named.}
}
\value{
An object of class \code{"polars_sql_context"}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(x = 1)
df2 <- pl$DataFrame(x = 2)
df3 <- pl$DataFrame(x = 3)

ctx <- pl$SQLContext()
ctx$register_many(tab1 = df, tab2 = df2, tab3 = df3)
}
