% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/sql-context.R
\name{sql_context__unregister}
\alias{sql_context__unregister}
\title{Unregister one or more frames by name}
\usage{
sql_context__unregister(names)
}
\arguments{
\item{names}{Names of the tables to unregister.}
}
\value{
An object of class \code{"polars_sql_context"}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(ints = 9:5)
lf1 <- pl$LazyFrame(text = letters[1:3])
lf2 <- pl$LazyFrame(misc = "testing1234")

# Register with a SQLContext object:
ctx <- pl$SQLContext(test1 = df, test2 = lf1, test3 = lf2)
ctx$tables()

# Unregister one or more of the tables:
ctx$unregister(c("test1", "test3"))$tables()
ctx$unregister("test2")$tables()
}
