% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_class}
\alias{SQLContext_class}
\title{Run SQL queries against DataFrame/LazyFrame data.}
\description{
Run SQL queries against DataFrame/LazyFrame data.
}
\examples{
lf = pl$LazyFrame(a = 1:3, b = c("x", NA, "z"))
res = pl$SQLContext(frame = lf)$execute(
  "SELECT b, a*2 AS two_a FROM frame WHERE b IS NOT NULL"
)
res$collect()
}
\keyword{SQLContext}
