% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_execute}
\alias{SQLContext_execute}
\title{Execute SQL query against the registered data}
\usage{
SQLContext_execute(query, eager = FALSE)
}
\arguments{
\item{query}{A valid string SQL query.}

\item{eager}{A logical flag indicating whether to collect the result immediately.
If FALSE (default), a \link{LazyFrame} is returned. If TRUE, a \link{DataFrame} is returned.}
}
\value{
A \link{LazyFrame} or \link{DataFrame} depending on the value of \code{eager}.
}
\description{
Parse the given SQL query and execute it against the registered frame data.
}
\examples{
query = "SELECT * FROM mtcars WHERE cyl = 4"
pl$SQLContext(mtcars = mtcars)$execute(query)
pl$SQLContext(mtcars = mtcars)$execute(query, eager = TRUE)
}
