% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_tables}
\alias{SQLContext_tables}
\title{List registered tables}
\usage{
SQLContext_tables()
}
\value{
A character vector of the registered table names.
}
\description{
Return a character vector of the registered table names.
}
\examples{
ctx = pl$SQLContext()
ctx$tables()
ctx$register("df1", mtcars)
ctx$tables()
ctx$register("df2", mtcars)
ctx$tables()
}
