% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_unregister}
\alias{SQLContext_unregister}
\title{Unregister tables by name}
\usage{
SQLContext_unregister(names)
}
\arguments{
\item{names}{A character vector of table names to unregister.}
}
\value{
Returns the \link{RPolarsSQLContext} object invisibly.
}
\description{
Unregister tables by name.
}
\examples{
# Initialise a new SQLContext and register the given tables.
ctx = pl$SQLContext(x = mtcars, y = mtcars, z = mtcars)
ctx$tables()

# Unregister some tables.
ctx$unregister(c("x", "y"))
ctx$tables()
}
