% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_string}
\alias{str_string}
\title{Simple viewer of an R object based on str()}
\usage{
str_string(x, collapse = " ")
}
\arguments{
\item{x}{object to view.}

\item{collapse}{word to glue possible multilines with}
}
\value{
string
}
\description{
Simple viewer of an R object based on str()
}
\examples{
polars:::str_string(list(a = 42, c(1, 2, 3, NA)))
}
\keyword{internal}
