% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_functions.R
\name{sum}
\alias{sum}
\title{new Expr with sum}
\arguments{
\item{column}{is a:
\itemize{
\item Series or Expr, same as \code{column$sum()}
\item string, same as \code{pl$col(column)$sum()}
\item numeric, same as \code{pl$lit(column)$sum()}
\item list of strings(column names) or exprressions to add up as expr1 + expr2 + expr3 + ...
wild cards allowed.
}}
}
\value{
Expr
}
\description{
syntactic sugar for starting a expression with sum
}
\examples{

#column as string
pl$DataFrame(iris)$select(pl$sum("Petal.Width"))

#column as Expr (prefer pl$col("Petal.Width")$sum())
pl$DataFrame(iris)$select(pl$sum(pl$col("Petal.Width")))

#column as numeric
pl$DataFrame()$select(pl$sum(1:5))

#column as list
pl$DataFrame(a=1:2,b=3:4,c=5:6)$with_column(pl$sum(list("a","c")))
pl$DataFrame(a=1:2,b=3:4,c=5:6)$with_column(pl$sum(list("a","c", 42L)))

#three eqivalent lines
pl$DataFrame(a=1:2,b=3:4,c=5:6)$with_column(pl$sum(list("a","c", pl$sum(list("a","b","c")))))
pl$DataFrame(a=1:2,b=3:4,c=5:6)$with_column(pl$sum(list(pl$col("a")+pl$col("b"),"c")))
pl$DataFrame(a=1:2,b=3:4,c=5:6)$with_column(pl$sum(list("*")))
}
