% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{threadpool_size}
\alias{threadpool_size}
\alias{pl_threadpool_size}
\title{Get the number of threads in the Polars thread pool.}
\usage{
pl_threadpool_size()
}
\value{
The number of threads
}
\description{
The threadpool size can be overridden by setting the
\code{POLARS_MAX_THREADS} environment variable before process start.
(The thread pool is not behind a lock, so it cannot be modified once set).
It is strongly recommended not to override this value as it will be
set automatically by the engine.
}
\examples{
pl$threadpool_size()
}
