% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error__trait.R
\name{to_condition}
\alias{to_condition}
\title{Internal generic method to convert an error_type to condition.}
\usage{
to_condition(err)
}
\arguments{
\item{err}{any type which impl as.character}

\item{call}{calling context}
}
\value{
condition
}
\description{
Internal generic method to convert an error_type to condition.
}
\details{
this method is needed to preserve state of err without upcasting to a string message
an implementation will describe how to store the error in the condition
}
\keyword{internal}
