% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-knitr.R
\name{to_html_table}
\alias{to_html_table}
\title{Generate HTML table from a DataFrame}
\usage{
to_html_table(x, max_cols = 75, max_rows = 40)
}
\arguments{
\item{x}{DataFrame}

\item{max_cols}{an integer of maximum number of columns to display}

\item{max_rows}{an integer of maximum number of rows to display}
}
\value{
A string of HTML code
}
\description{
Generate HTML table from a DataFrame
}
\examples{
to_html_table(mtcars, 3, 3)
}
