% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{to_list}
\alias{to_list}
\alias{DataFrame_to_list}
\title{return polars DataFrame as R lit of vectors}
\usage{
DataFrame_to_list(unnest_structs = TRUE)
}
\arguments{
\item{unnest_structs}{bool default true, as calling $unnest() on any struct column}
}
\value{
R list of vectors
}
\description{
return polars DataFrame as R lit of vectors
}
\details{
This implementation for simplicity reasons relies on unnesting all structs before
exporting to R. unnest_structs = FALSE, the previous struct columns will be re-
nested. A struct in a R is a lists of lists, where each row is a list of values.
Such a structure is not very typical or efficient in R.
}
\examples{
pl$DataFrame(iris)$to_list()
}
\keyword{DataFrame}
