% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_conversion.R
\name{unwrap_err}
\alias{unwrap_err}
\title{rust-like unwrap_err, internal use only}
\usage{
unwrap_err(result)
}
\arguments{
\item{result}{a Result, see rust_result.R#'}
}
\value{
some error type
}
\description{
rust-like unwrap_err, internal use only
}
\details{
throwed error info is sparse because only for internal errors
}
\keyword{internal}
