% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{unwrap_or}
\alias{unwrap_or}
\title{unwrap return or if err}
\usage{
unwrap_or(x, or)
}
\arguments{
\item{x}{any R object}

\item{or}{any R value}
}
\value{
pl
}
\description{
unwrap return or if err
}
\keyword{internal}
