% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error__trait.R
\name{upgrade_err}
\alias{upgrade_err}
\title{Internal generic method to add plain text to error message}
\usage{
upgrade_err(err)
}
\arguments{
\item{err}{some error type object}

\item{msg}{string to add}
}
\value{
condition
}
\description{
Internal generic method to add plain text to error message
}
\details{
polars converts any other error types to RPolarsErr.
An error type can choose to implement this to improve the translation.
As fall back the error will be deparsed into a string with rust Debug, see rdbg()
}
\keyword{internal}
