% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error__trait.R
\name{when_calling}
\alias{when_calling}
\title{Internal generic method to add call to error}
\usage{
when_calling(err, call)
}
\arguments{
\item{err}{any type which impl as.character}

\item{call}{calling context}
}
\value{
err as string
}
\description{
Internal generic method to add call to error
}
\details{
Additional details...
}
\examples{
#
}
\keyword{internal}
