% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whenthen.R
\name{WhenThenThen_when}
\alias{WhenThenThen_when}
\title{WhenThenThen_when}
\usage{
WhenThenThen_when(predicate)
}
\arguments{
\item{predicate}{Expr}
}
\value{
Expr: Series of dtype Utf8.
}
\description{
WhenThenThen_when
}
\examples{
#TODO
}
\keyword{WhenThen}
