% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{wrap_proto_schema}
\alias{wrap_proto_schema}
\title{wrap proto schema}
\format{
function
}
\usage{
wrap_proto_schema(x)
}
\arguments{
\item{x}{either schema, or incomplete schema where dataType can be NULL
or schema is just char vec, implicitly the same as if all DataType are NULL,
mean undefined.}
}
\value{
bool
}
\description{
wrap proto schema
}
\examples{
polars:::wrap_proto_schema(c("alice", "bob"))
polars:::wrap_proto_schema(list("alice" = pl$Int64, "bob" = NULL))
}
\keyword{internal}
