% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{wrap_s}
\alias{wrap_s}
\title{Wrap as Series}
\usage{
wrap_s(x)
}
\arguments{
\item{x}{a Series or something-turned-into-Series}
}
\value{
Series
}
\description{
input is either already a Series of will be passed to the Series constructor
}
