% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_completion.R
\name{activate_polars_rstudio_completion}
\alias{activate_polars_rstudio_completion}
\title{activate_polars_rstudio_completion}
\description{
activate_polars_rstudio_completion
}
\details{
Modifies rstudio auto-completion functions by wrapping them.

Any other package can also wrap these or other completion functions.
Multiple wrappers can also co-exists, just keep the signature the same.
}
\examples{
.dev$activate_polars_rstudio_completion()
}
