% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr$lengths}
\alias{arr$lengths}
\alias{make_expr_arr_namespace}
\alias{lengths}
\title{make_expr_arr_namespace}
\usage{
make_expr_arr_namespace(self)
}
\value{
environment with methods to call on self
Lenghts arrays in list

Expr
}
\description{
Internal function to yield methods within arr namespace
See the individual method pages for full details

Get the length of the arrays as UInt32.
}
\examples{
df = pl$DataFrame(list_of_strs = pl$Series(list(c("a","b"),"c")))
df$with_column(pl$col("list_of_strs")$arr$lengths()$alias("list_of_strs_lengths"))
}
\keyword{ExprArr}
\keyword{internals}
