% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_get}
\alias{arr_get}
\alias{Expr_arr_get}
\alias{Expr_arr.get}
\alias{[.ExprArrNameSpace}
\title{Get list}
\format{
function
}
\usage{
\method{[}{ExprArrNameSpace}(x, index)
}
\arguments{
\item{x}{ExprArrNameSpace}

\item{index}{value to get}
}
\value{
Expr
}
\description{
Get the value by index in the sublists.
}
\details{
\verb{[.ExprArrNameSpace} used as e.g. \code{pl$col("a")$arr[0]} same as \code{pl$col("a")$get(0)}
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) # NULL or integer() or list()
df$select(pl$col("a")$arr$get(0))
df$select(pl$col("a")$arr$get(c(2, 0, -1)))
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) # NULL or integer() or list()
df$select(pl$col("a")$arr[0])
df$select(pl$col("a")$arr[c(2, 0, -1)])
}
\keyword{ExprArr}
