% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_get}
\alias{arr_get}
\alias{arr.get}
\alias{[.ExprListNameSpace}
\title{Get list}
\format{
function

function
}
\usage{
\method{[}{ExprListNameSpace}(x, idx)
}
\arguments{
\item{index}{numeric vector or Expr of length 1 or same length of Series.
if length 1 pick same value from each sublist, if length as Series/column,
pick by individual index across sublists.

So index \code{0} would return the first item of every sublist
and index \code{-1} would return the last item of every sublist
if an index is out of bounds, it will return a \code{None}.}
}
\value{
Expr

Expr
}
\description{
Get the value by index in the sublists.

Get the value by index in the sublists.
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) #NULL or integer() or list()
df$select(pl$col("a")$arr$get(0))
df$select(pl$col("a")$arr$get(c(2,0,-1)))
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) #NULL or integer() or list()
df$select(pl$col("a")$arr$get(0))
df$select(pl$col("a")$arr$get(c(2,0,-1)))
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) #NULL or integer() or list()
df$select(pl$col("a")$arr[0])
df$select(pl$col("a")$arr[c(2,0,-1)])
}
\keyword{ExprArr}
