% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_sort}
\alias{arr_sort}
\alias{arr.sort}
\title{Sort lists}
\format{
function
}
\value{
Expr
}
\description{
Sort the arrays in the list.
}
\examples{
df = pl$DataFrame(list(
  values = list(3:1,c(9L,1:2))
))
df$select(pl$col("values")$arr$sort())
}
\keyword{ExprArr}
