% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{arr_take}
\alias{arr_take}
\alias{arr.take}
\title{take in sublists}
\format{
function
}
\arguments{
\item{index}{R list of integers for each sub-element or Expr or Series of type \code{List[usize]}}

\item{null_on_oob}{boolean}
}
\value{
Expr
}
\description{
Get the take value of the sublists.
}
\examples{
df = pl$DataFrame(list(a = list(c(3, 2, 1), 1, c(1, 2)))) #
idx = pl$Series(list(0:1, integer(), c(1L, 999L)))
df$select(pl$col("a")$arr$take(pl$lit(idx), null_on_oob = TRUE))

# with implicit conversion to Expr
df$select(pl$col("a")$arr$take(list(0:1, integer(), c(1L, 999L)), null_on_oob = TRUE))

# by some column name, must cast to an Int/Uint type to work
df$select(pl$col("a")$arr$take(pl$col("a")$cast(pl$List(pl$UInt64)), null_on_oob = TRUE))
}
\keyword{ExprArr}
