% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.character.Series}
\alias{as.character.Series}
\title{as.character.Series}
\usage{
\method{as.character}{Series}(x, ..., format = FALSE, str_length = 15)
}
\arguments{
\item{x}{Series}

\item{...}{not used}

\item{format}{a logical. If \code{TRUE}, the Series will be formatted.}

\item{str_length}{an integer. If \code{format = TRUE},
utf8 or categorical type Series will be formatted to a string of this length.}
}
\description{
as.character.Series
}
\examples{
s = pl$Series(c("foo", "barbaz"))
as.character(s)
as.character(s, format = TRUE)
as.character(s, format = TRUE, str_length = 3)
}
