% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{as.list.Expr}
\alias{as.list.Expr}
\title{S3 method to convert an Expr to a list}
\usage{
\method{as.list}{Expr}(x, ...)
}
\arguments{
\item{x}{Expr}

\item{...}{Not used.}
}
\value{
One Expr wrapped in a list
}
\description{
S3 method to convert an Expr to a list
}
\keyword{internal}
