% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_polars_df.R
\name{as_polars_df}
\alias{as_polars_df}
\alias{as_polars_df.default}
\alias{as_polars_df.polars_data_frame}
\alias{as_polars_df.polars_group_by}
\alias{as_polars_df.polars_lazy_frame}
\alias{as_polars_df.list}
\alias{as_polars_df.data.frame}
\title{Create a Polars DataFrame from an R object}
\usage{
as_polars_df(x, ...)

\method{as_polars_df}{default}(x, ...)

\method{as_polars_df}{polars_data_frame}(x, ...)

\method{as_polars_df}{polars_group_by}(x, ...)

\method{as_polars_df}{polars_lazy_frame}(x, ...)

\method{as_polars_df}{list}(x, ...)

\method{as_polars_df}{data.frame}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Additional arguments passed to the methods.}
}
\value{
A Polars DataFrame
}
\description{
The \code{\link[=as_polars_df]{as_polars_df()}} function creates a \link[=DataFrame]{polars DataFrame} from various R objects.
\link[=DataFrame]{Polars DataFrame} is based on a sequence of \link[=Series]{Polars Series},
so basically, the input object is converted to a \link{list} of
\link[=Series]{Polars Series} by \code{\link[=as_polars_series]{as_polars_series()}},
then a \link[=DataFrame]{Polars DataFrame} is created from the list.
}
\details{
The default method of \code{\link[=as_polars_df]{as_polars_df()}} throws an error,
so we need to define methods for the classes we want to support.
}
\examples{
# list
as_polars_df(list(a = 1:2, b = c("foo", "bar")))

# data.frame
as_polars_df(data.frame(a = 1:2, b = c("foo", "bar")))
}
\seealso{
\itemize{
\item \code{\link[=dataframe__to_r_list]{<DataFrame>$to_r_list()}}: Export the DataFrame as an R list of R vectors.
}
}
