% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_polars.R
\name{as_polars_df}
\alias{as_polars_df}
\alias{as_polars_df.default}
\alias{as_polars_df.data.frame}
\alias{as_polars_df.DataFrame}
\alias{as_polars_df.GroupBy}
\alias{as_polars_df.Series}
\alias{as_polars_df.LazyFrame}
\alias{as_polars_df.ArrowTabular}
\alias{as_polars_df.RecordBatchReader}
\title{To polars DataFrame}
\usage{
as_polars_df(x, ...)

\method{as_polars_df}{default}(x, ...)

\method{as_polars_df}{data.frame}(x, ...)

\method{as_polars_df}{DataFrame}(x, ...)

\method{as_polars_df}{GroupBy}(x, ...)

\method{as_polars_df}{Series}(x, ...)

\method{as_polars_df}{LazyFrame}(
  x,
  n_rows = Inf,
  ...,
  type_coercion = TRUE,
  predicate_pushdown = TRUE,
  projection_pushdown = TRUE,
  simplify_expression = TRUE,
  slice_pushdown = TRUE,
  comm_subplan_elim = TRUE,
  comm_subexpr_elim = TRUE,
  streaming = FALSE,
  no_optimization = FALSE,
  inherit_optimization = FALSE,
  collect_in_background = FALSE
)

\method{as_polars_df}{ArrowTabular}(x, ..., rechunk = TRUE, schema = NULL, schema_overrides = NULL)

\method{as_polars_df}{RecordBatchReader}(x, ..., rechunk = TRUE, schema = NULL, schema_overrides = NULL)
}
\arguments{
\item{x}{Object to convert to a polars DataFrame.}

\item{...}{Additional arguments passed to methods.}

\item{n_rows}{Number of rows to fetch. Defaults to \code{Inf}, meaning all rows.}

\item{type_coercion}{Boolean. Coerce types such that operations succeed and
run on minimal required memory.}

\item{predicate_pushdown}{Boolean. Applies filters as early as possible at
scan level.}

\item{projection_pushdown}{Boolean. Select only the columns that are needed
at the scan level.}

\item{simplify_expression}{Boolean. Various optimizations, such as constant
folding and replacing expensive operations with faster alternatives.}

\item{slice_pushdown}{Boolean. Only load the required slice from the scan
level. Don't materialize sliced outputs (e.g. \code{join$head(10)}).}

\item{comm_subplan_elim}{Boolean. Will try to cache branching subplans that
occur on self-joins or unions.}

\item{comm_subexpr_elim}{Boolean. Common subexpressions will be cached and
reused.}

\item{streaming}{Boolean. Run parts of the query in a streaming fashion
(this is in an alpha state).}

\item{no_optimization}{Boolean. Sets the following parameters to \code{FALSE}:
\code{predicate_pushdown}, \code{projection_pushdown}, \code{slice_pushdown},
\code{comm_subplan_elim}, \code{comm_subexpr_elim}.}

\item{inherit_optimization}{Boolean. Use existing optimization settings
regardless the settings specified in this function call.}

\item{collect_in_background}{Boolean. Detach this query from R session.
Computation will start in background. Get a handle which later can be converted
into the resulting DataFrame. Useful in interactive mode to not lock R session.}

\item{rechunk}{bool rewrite in one array per column, Implemented for ChunkedArray
Array is already contiguous. Not implemented for Table. C}

\item{schema}{named list of DataTypes or char vec of names. Same length as arrow table.
If schema names or types do not match arrow table, the columns will be renamed/recast.
NULL default is to import columns as is. Takes no effect for Array or ChunkedArray}

\item{schema_overrides}{named list of DataTypes. Name some columns to recast by the DataType.
Takes not effect for Array or ChunkedArray}
}
\description{
\code{\link[=as_polars_df]{as_polars_df()}} is a generic function that converts an R object to a
polars DataFrame. It is basically a wrapper for \link[=pl_DataFrame]{pl$DataFrame()},
but has special implementations for Apache Arrow-based objects such as
polars \link[=LazyFrame_class]{LazyFrame} and \link[arrow:Table]{arrow::Table}.
}
\details{
For \link[=LazyFrame_class]{LazyFrame} objects, this function is a shortcut for
\link[=LazyFrame_collect]{$collect()} or \link[=LazyFrame_fetch]{$fetch()}, depending on
whether the number of rows to fetch is infinite or not.
}
\examples{
\dontshow{if (requireNamespace("arrow", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
at = arrow::as_arrow_table(mtcars)

# Convert an arrow Table to a polars LazyFrame
lf = as_polars_df(at)$lazy()

# Collect all rows
as_polars_df(lf)

# Fetch 5 rows
as_polars_df(lf, 5)
\dontshow{\}) # examplesIf}
}
