% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_polars_expr.R
\name{as_polars_expr}
\alias{as_polars_expr}
\alias{as_polars_expr.default}
\alias{as_polars_expr.polars_expr}
\alias{as_polars_expr.polars_series}
\alias{as_polars_expr.character}
\alias{as_polars_expr.logical}
\alias{as_polars_expr.integer}
\alias{as_polars_expr.double}
\alias{as_polars_expr.raw}
\alias{as_polars_expr.NULL}
\title{Create a Polars Expression from an R object}
\usage{
as_polars_expr(x, ...)

\method{as_polars_expr}{default}(x, ...)

\method{as_polars_expr}{polars_expr}(x, ...)

\method{as_polars_expr}{polars_series}(x, ...)

\method{as_polars_expr}{character}(x, ..., str_as_lit = FALSE)

\method{as_polars_expr}{logical}(x, ...)

\method{as_polars_expr}{integer}(x, ...)

\method{as_polars_expr}{double}(x, ...)

\method{as_polars_expr}{raw}(x, ...)

\method{as_polars_expr}{`NULL`}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Additional arguments passed to the methods.}

\item{str_as_lit}{A logical value indicating whether to treat the character vector as literal values
or column names (default).}
}
\value{
A polars Expression
}
\description{
The \code{\link[=as_polars_expr]{as_polars_expr()}} function creates a polars Expression from various R objects.
This function is used internally by various polars functions that accept Expressions.
}
\details{
Because R objects are typically mapped to \link{Series}, this function often calls \code{\link[=as_polars_series]{as_polars_series()}} internally.
However, unlike R, Polars has scalars of length 1, so if an R object is converted to a \link{Series} of length 1,
this function will use \verb{<Expr>$first()} at the end to convert it to a scalar value.
If you want to implement your own conversion from an R class to a Polars object,
define an S3 method for \code{\link[=as_polars_series]{as_polars_series()}} instead of this function.
\subsection{Default S3 method}{

Create a \link{Series} by calling \code{\link[=as_polars_series]{as_polars_series()}} and then convert that \link{Series} to an Expression.
If the length of the \link{Series} is \code{1}, it will be converted to a scalar value using \verb{<Expr>$first()} at the end.
}

\subsection{S3 method for \link{character}}{

If the \code{str_as_lit} argument is \code{FALSE}, this function will call \code{pl$col()} and
the character vector is treated as column names.
}
}
\examples{
# character
as_polars_expr("a")
as_polars_expr(c("a", "b"))
as_polars_expr("a", str_as_lit = TRUE)
as_polars_expr(c("a", "b"), str_as_lit = TRUE)

# logical
as_polars_expr(TRUE)
as_polars_expr(c(TRUE, FALSE))

# integer
as_polars_expr(1L)
as_polars_expr(c(1L, 2L))

# double
as_polars_expr(1)
as_polars_expr(c(1, 2))

# raw
as_polars_expr(charToRaw("foo"))

# default method (for list)
as_polars_expr(list(1))
as_polars_expr(list(1, 2))

# polars_series
## Unlike the default method, this method does not call `$first()`
as_polars_series(1) |>
  as_polars_expr()
}
\seealso{
\itemize{
\item \code{\link[=as_polars_series]{as_polars_series()}}: R -> Polars type mapping is defined by this function.
}
}
