% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_polars_lf.R
\name{as_polars_lf}
\alias{as_polars_lf}
\alias{as_polars_lf.default}
\alias{as_polars_lf.polars_lazy_frame}
\title{Create a Polars LazyFrame from an R object}
\usage{
as_polars_lf(x, ...)

\method{as_polars_lf}{default}(x, ...)

\method{as_polars_lf}{polars_lazy_frame}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Additional arguments passed to the methods.}
}
\value{
A polars \link{LazyFrame}
}
\description{
The \code{\link[=as_polars_lf]{as_polars_lf()}} function creates a \link{LazyFrame} from various R objects.
It is basically a shortcut for \link[=as_polars_df]{as_polars_df(x, ...)} with the
\verb{$lazy()}method.
}
\details{
\subsection{Default S3 method}{

Create a \link{DataFrame} by calling \code{\link[=as_polars_df]{as_polars_df()}} and then create
a \link{LazyFrame} from the \link{DataFrame}.
Additional arguments \code{...} are passed to \code{\link[=as_polars_df]{as_polars_df()}}.
}
}
