% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_polars.R
\name{as_polars_lf}
\alias{as_polars_lf}
\alias{as_polars_lf.default}
\alias{as_polars_lf.RPolarsLazyFrame}
\alias{as_polars_lf.RPolarsLazyGroupBy}
\title{To polars LazyFrame}
\usage{
as_polars_lf(x, ...)

\method{as_polars_lf}{default}(x, ...)

\method{as_polars_lf}{RPolarsLazyFrame}(x, ...)

\method{as_polars_lf}{RPolarsLazyGroupBy}(x, ...)
}
\arguments{
\item{x}{Object to convert to a polars DataFrame.}

\item{...}{Additional arguments passed to methods.}
}
\value{
a \link[=LazyFrame_class]{LazyFrame}
}
\description{
\code{\link[=as_polars_lf]{as_polars_lf()}} is a generic function that converts an R object to a
polars LazyFrame. It is basically a shortcut for \link[=as_polars_df]{as_polars_df(x, ...)} with the
\link[=DataFrame_lazy]{$lazy()} method.
}
\examples{
as_polars_lf(mtcars)
}
