% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_polars.R
\name{as_polars_series}
\alias{as_polars_series}
\alias{as_polars_series.default}
\alias{as_polars_series.POSIXlt}
\alias{as_polars_series.data.frame}
\alias{as_polars_series.vctrs_rcrd}
\title{To polars Series}
\usage{
as_polars_series(x, name = NULL, ...)

\method{as_polars_series}{default}(x, name = NULL, ...)

\method{as_polars_series}{POSIXlt}(x, name = NULL, ...)

\method{as_polars_series}{data.frame}(x, name = NULL, ...)

\method{as_polars_series}{vctrs_rcrd}(x, name = NULL, ...)
}
\arguments{
\item{x}{Object to convert into a polars Series}

\item{name}{A string to use as the name of the Series.
If \code{NULL} (default), unnamed Series will be created.}

\item{...}{Additional arguments passed to methods.}
}
\value{
a \link[=Series_class]{RPolarsSeries}
}
\description{
\code{\link[=as_polars_series]{as_polars_series()}} is a generic function that converts an R object to a
polars Series. It is basically a wrapper for \link[=pl_Series]{pl$Series()}.
}
\examples{
as_polars_series(1:4)

as_polars_series(list(1:4))

as_polars_series(data.frame(a = 1:4))
}
