% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{c.Series}
\alias{c.Series}
\title{Immutable combine series}
\usage{
\method{c}{Series}(x, ...)
}
\arguments{
\item{x}{a Series}

\item{...}{Series(s) or any object into Series meaning \code{pl$Series(object)} returns a series}
}
\value{
a combined Series
}
\description{
Immutable combine series
}
\details{
append datatypes has to match. Combine does not rechunk.
Read more about R vectors, Series and chunks in \code{\link[rpolars]{docs_translations}}:
}
\examples{
s = c(pl$Series(1:5),3:1,NA_integer_)
s$chunk_lengths() #the series contain three unmerged chunks
}
