% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{$<-.DataFrame}
\alias{$<-.DataFrame}
\title{generic setter method}
\usage{
\method{$}{DataFrame}(self, name) <- value
}
\value{
value
}
\description{
set value of properties of DataFrames
}
\examples{
pl$DataFrame(iris)$columns()
}
\keyword{DataFrame}
