% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{$<-.DataFrame}
\alias{$<-.DataFrame}
\title{generic setter method}
\usage{
\method{$}{DataFrame}(self, name) <- value
}
\arguments{
\item{self}{DataFrame}

\item{name}{name method/property to set}

\item{value}{value to insert}
}
\value{
value
}
\description{
set value of properties of DataFrames
}
\details{
settable polars object properties may appear to be R objects, but they are not.
See \verb{[[method_name]]} example
}
\examples{
# For internal use
# is only activated for following methods of DataFrame
ls(DataFrame.property_setters)

# specific use case for one object property 'columns' (names)
df = pl$DataFrame(iris)

# get values
df$columns

# set + get values
df$columns = letters[1:5] #<- is fine too
df$columns

# Rstudio is not using the standard R code completion tool
# and it will backtick any special characters. It is possible
# to completely customize the R / Rstudio code completion except
# it will trigger Rstudio to backtick any completion! Also R does
# not support package isolated customization.


# Concrete example if tabbing on 'df$' the raw R suggestion is df$columns<-
# however Rstudio backticks it into df$`columns<-`
# to make life simple, this is valid polars syntax also, and can be used in fast scripting
df$`columns<-` = letters[5:1]

# for stable code prefer e.g.  df$columns = letters[5:1]

# to see inside code of a property use the [[]] syntax instead
df[["columns"]] # to see property code, .pr is the internal polars api into rust polars
DataFrame.property_setters$columns # and even more obscure to see setter code
}
\keyword{DataFrame}
