% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_functions.R
\name{coalesce}
\alias{coalesce}
\title{Coalesce}
\arguments{
\item{...}{is a:
If one arg:
\itemize{
\item Series or Expr, same as \code{column$sum()}
\item string, same as \code{pl$col(column)$sum()}
\item numeric, same as \code{pl$lit(column)$sum()}
\item list of strings(column names) or exprressions to add up as expr1 + expr2 + expr3 + ...
}

If several args, then wrapped in a list and handled as above.}
}
\value{
Expr
}
\description{
Folds the expressions from left to right, keeping the first non-null value.
}
\examples{
df = pl$DataFrame(
  a = NA_real_,
  b = c(1:2,NA_real_,NA_real_),
  c = c(1:3,NA_real_)
)
#use coalesce to get first non Null value for each row, otherwise insert 99.9
df$with_column(
  pl$coalesce("a", "b", "c", 99.9)$alias("d")
)

}
