% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_columns}
\alias{DataFrame_columns}
\title{columns, names columns}
\usage{
DataFrame_columns
}
\value{
char vec of column names
}
\description{
get column names as DataFrames
}
\examples{
df = pl$DataFrame(iris)$columns
df$columns
df$columns = letters[1:5]
df$columns
}
\keyword{DataFrame}
