% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_DataTypeVector}
\alias{construct_DataTypeVector}
\title{construct data type vector}
\usage{
construct_DataTypeVector(l)
}
\arguments{
\item{l}{list of Expr or string}
}
\value{
extptr to rust vector of RPolarsDataType's
}
\description{
lifecycle: Deprecate, move to rust side
}
\keyword{internal}
