% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs}
\alias{cs}
\alias{polars_selector}
\alias{Selector}
\alias{selector__as_expr}
\title{Polars column selector function namespace}
\usage{
cs

selector__as_expr()
}
\description{
\code{cs} is an \link[=environment-class]{environment class} object that stores all
selector functions of the R Polars API which mimics the Python Polars API.
It is intended to work the same way in Python as if you had imported
Python Polars Selectors with \verb{import polars.selectors as cs}.
}
\section{Supported operators}{

There are 4 supported operators for selectors:
\itemize{
\item \code{&} to combine conditions with AND, e.g. select columns that contain
\code{"oo"} \emph{and} end with \code{"t"} with \code{cs$contains("oo") & cs$ends_with("t")};
\item \code{|} to combine conditions with OR, e.g. select columns that contain
\code{"oo"} \emph{or} end with \code{"t"} with \code{cs$contains("oo") | cs$ends_with("t")};
\item \code{-} to subtract conditions, e.g. select all columns that have alphanumeric
names except those that contain \code{"a"} with
\code{cs$alphanumeric() - cs$contains("a")};
\item \code{!} to invert the selection, e.g. select all columns that \emph{are not} of data
type \code{String} with \code{!cs$string()}.
}

Note that Python Polars uses \code{~} instead of \code{!} to invert selectors.

If we want to apply operators on the data instead of the selector sets,
\verb{<selector>$as_expr()} can be used to materialize the selector as a normal
expression.
}

\examples{
cs

df <- pl$DataFrame(
  colx = c("aa", "bb", "cc"),
  coly = c(TRUE, FALSE, TRUE),
  colz = c(1, 2, 3),
)

# Inverting the boolean selector will choose the non-boolean columns:
df$select(!cs$boolean())

# To invert the values in the selected boolean columns,
# we need to materialize the selector as a standard expression instead:
df$select(!cs$boolean()$as_expr())
}
