% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\docType{data}
\name{cs}
\alias{cs}
\title{Polars column selector function namespace}
\format{
An object of class \code{polars_object} of length 2.
}
\usage{
cs
}
\description{
\code{cs} is an \link[=environment-class]{environment class} object
that stores all selector functions of the R Polars API
which mimics the Python Polars API.
It is intended to work the same way in Python as if you had imported
Python Polars Selectors with \verb{import polars.selectors as cs}.
}
\examples{
cs

# How many members are in the `cs` environment?
length(cs)
}
\keyword{datasets}
