% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__all}
\alias{cs__all}
\title{Select all columns}
\usage{
cs__all()
}
\value{
A Polars selector
}
\description{
Select all columns
}
\examples{
df <- pl$DataFrame(dt = as.Date(c("2000-1-1")), value = 10)

# Select all columns, casting them to string:
df$select(cs$all()$cast(pl$String))

# Select all columns except for those matching the given dtypes:
df$select(cs$all() - cs$numeric())
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
