% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__alpha}
\alias{cs__alpha}
\title{Select all columns with alphabetic names (e.g. only letters)}
\usage{
cs__alpha(ascii_only = FALSE, ..., ignore_spaces = FALSE)
}
\arguments{
\item{ascii_only}{Indicate whether to consider only ASCII alphabetic
characters, or the full Unicode range of valid letters (accented,
idiographic, etc).}

\item{...}{These dots are for future extensions and must be empty.}

\item{ignore_spaces}{Indicate whether to ignore the presence of spaces in
column names; if so, only the other (non-space) characters are considered.}
}
\value{
A Polars selector
}
\description{
Select all columns with alphabetic names (e.g. only letters)
}
\details{
Matching column names cannot contain any non-alphabetic characters. Note
that the definition of “alphabetic” consists of all valid Unicode alphabetic
characters (\verb{p\{Alphabetic\}}) by default; this can be changed by setting
\code{ascii_only = TRUE}.
}
\examples{
df <- pl$DataFrame(
  no1 = c(100, 200, 300),
  café = c("espresso", "latte", "mocha"),
  `t or f` = c(TRUE, FALSE, NA),
  hmm = c("aaa", "bbb", "ccc"),
  都市 = c("東京", "大阪", "京都")
)

# Select columns with alphabetic names; note that accented characters and
# kanji are recognised as alphabetic here:
df$select(cs$alpha())

# Constrain the definition of “alphabetic” to ASCII characters only:
df$select(cs$alpha(ascii_only = TRUE))
df$select(cs$alpha(ascii_only = TRUE, ignore_spaces = TRUE))

# Select all columns except for those with alphabetic names:
df$select(!cs$alpha())
df$select(!cs$alpha(ignore_spaces = TRUE))
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
