% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__alphanumeric}
\alias{cs__alphanumeric}
\title{Select all columns with alphanumeric names (e.g. only letters and the digits 0-9)}
\usage{
cs__alphanumeric(ascii_only = FALSE, ..., ignore_spaces = FALSE)
}
\arguments{
\item{ascii_only}{Indicate whether to consider only ASCII alphabetic
characters, or the full Unicode range of valid letters (accented,
idiographic, etc).}

\item{...}{These dots are for future extensions and must be empty.}

\item{ignore_spaces}{Indicate whether to ignore the presence of spaces in
column names; if so, only the other (non-space) characters are considered.}
}
\value{
A Polars selector
}
\description{
Select all columns with alphanumeric names (e.g. only letters and the digits 0-9)
}
\details{
Matching column names cannot contain any non-alphabetic characters. Note
that the definition of “alphabetic” consists of all valid Unicode alphabetic
characters (\verb{p\{Alphabetic\}}) and digit characters (\code{d}) by default; this can
be changed by setting \code{ascii_only = TRUE}.
}
\examples{
df <- pl$DataFrame(
  `1st_col` = c(100, 200, 300),
  flagged = c(TRUE, FALSE, TRUE),
  `00prefix` = c("01:aa", "02:bb", "03:cc"),
  `last col` = c("x", "y", "z")
)

# Select columns with alphanumeric names:
df$select(cs$alphanumeric())
df$select(cs$alphanumeric(ignore_spaces = TRUE))

# Select all columns except for those with alphanumeric names:
df$select(!cs$alphanumeric())
df$select(!cs$alphanumeric(ignore_spaces = TRUE))
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
