% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__array}
\alias{cs__array}
\title{Select all array columns}
\usage{
cs__array(inner = NULL, ..., width = NULL)
}
\value{
A Polars \link[=cs]{selector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(
  foo = list(c("xx", "yy"), c("x", "y")),
  bar = list(c(123, 456), c(789, 101)),
  baz = c(2.0, 5.5),
  .schema_overrides = list(
    foo = pl$Array(pl$String, 2),
    bar = pl$Array(pl$Int64, 2)
  )
)

# Select all array columns:
df$select(cs$array())

# Select all columns except for those that are array:
df$select(!cs$array())

# If you want to select specific array columns,
# you can use the `by_dtype()` selector:
df$select(cs$by_dtype(pl$Array(pl$String, 2)))
}
\seealso{
\itemize{
\item \link{cs} for the documentation on operators supported by selectors.
\item \code{\link[=cs__by_dtype]{cs$by_dtype()}}: Select all columns matching the given dtype(s).
\item \code{\link[=cs__list]{cs$list()}}: Select all list columns.
\item \code{\link[=cs__nested]{cs$nested()}}: Select all nested columns.
}
}
