% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__binary}
\alias{cs__binary}
\title{Select all binary columns}
\usage{
cs__binary()
}
\value{
A Polars selector
}
\description{
Select all binary columns
}
\examples{
df <- pl$DataFrame(
  a = charToRaw("hello"),
  b = "world",
  c = charToRaw("!"),
  d = ":"
)

# Select binary columns:
df$select(cs$binary())

# Select all columns except for those that are binary:
df$select(!cs$binary())
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
