% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__by_dtype}
\alias{cs__by_dtype}
\title{Select all columns matching the given dtypes}
\usage{
cs__by_dtype(dtypes, ...)
}
\arguments{
\item{dtypes}{A Polars data type or list of Polars data types to select.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A Polars \link[=cs]{selector}
}
\description{
Select all columns matching the given dtypes
}
\examples{
df <- pl$DataFrame(
  dt = as.Date(c("1999-12-31", "2024-1-1", "2010-7-5")),
  value = c(1234500, 5000555, -4500000),
  other = c("foo", "bar", "foo")
)

# Select all columns with date or string dtypes:
df$select(cs$by_dtype(c(pl$Date, pl$String)))

# Select all columns that are not of date or string dtype:
df$select(!cs$by_dtype(c(pl$Date, pl$String)))

# Group by string columns and sum the numeric columns:
df$group_by(cs$string())$agg(cs$numeric()$sum())$sort("other")
}
\seealso{
\link{cs} for the documentation on operators supported by selectors.
}
