% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__contains}
\alias{cs__contains}
\title{Select columns whose names contain the given literal substring(s)}
\usage{
cs__contains(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Substring(s) that matching
column names should contain.}
}
\value{
A Polars selector
}
\description{
Select columns whose names contain the given literal substring(s)
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123, 456),
  baz = c(2.0, 5.5),
  zap = c(FALSE, TRUE)
)

# Select columns that contain the substring "ba":
df$select(cs$contains("ba"))

# Select columns that contain the substring "ba" or the letter "z":
df$select(cs$contains("ba", "z"))

# Select all columns except for those that contain the substring "ba":
df$select(!cs$contains("ba"))
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
