% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__datetime}
\alias{cs__datetime}
\title{Select all datetime columns}
\usage{
cs__datetime(time_unit = c("ms", "us", "ns"), time_zone = list("*", NULL))
}
\arguments{
\item{time_unit}{One (or more) of the allowed time unit precision strings,
\code{"ms"}, \code{"us"}, and \code{"ns"}. Default is to select columns with any valid
timeunit.}

\item{time_zone}{One of the followings. The value or each element of the vector
will be passed to the \code{time_zone} argument of the \code{\link[=DataType]{pl$Datetime()}} function:
\itemize{
\item A character vector of one or more timezone strings, as defined in \code{\link[=OlsonNames]{OlsonNames()}}.
\item \code{NULL} to select Datetime columns that do not have a timezone.
\item \code{"*"} to select Datetime columns that have any timezone.
\item A list of single timezone strings , \code{"*"}, and \code{NULL} to select Datetime columns
that do not have a timezone or have the (specific) timezone.
For example, the default value \code{list("*", NULL)} selects all Datetime columns.
}}
}
\value{
A Polars selector
}
\description{
Select all datetime columns
}
\examples{
df <- pl$DataFrame(
  tstamp_tokyo = as.POSIXct(
    c(
      "1999-7-21  5:20:16:987654",
      "2000-5-16  6:21:21:123465"
    ),
    tz = "Asia/Tokyo"
  ),
  tstamp_utc = as.POSIXct(
    c(
      "1999-7-21  5:20:16:987654",
      "2000-5-16  6:21:21:123465"
    ),
    tz = "UTC"
  ),
  tstamp = as.POSIXct(c(
    "1999-7-21  5:20:16:987654",
    "2000-5-16  6:21:21:123465"
  )),
  dt = as.Date(c(
    "1999-12-31", "2010-7-5"
  )),
  .schema_overrides = list(
    tstamp_tokyo = pl$Datetime("ns", "Asia/Tokyo"),
    tstamp_utc = pl$Datetime("us", "UTC"),
    tstamp = pl$Datetime("us")
  )
)

# Select all datetime columns:
df$select(cs$datetime())

# Select all datetime columns that have "us" precision:
df$select(cs$datetime("us"))

# Select all datetime columns that have any timezone:
df$select(cs$datetime(time_zone = "*"))

# Select all datetime columns that have a specific timezone:
df$select(cs$datetime(time_zone = "UTC"))

# Select all datetime columns that have NO timezone:
df$select(cs$datetime(time_zone = NULL))

# Select all columns except for datetime columns:
df$select(!cs$datetime())
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
