% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__decimal}
\alias{cs__decimal}
\title{Select all decimal columns}
\usage{
cs__decimal()
}
\value{
A Polars selector
}
\description{
Select all decimal columns
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123, 456),
  baz = c("2.0005", "-50.5555"),
  .schema_overrides = list(
    bar = pl$Decimal(),
    baz = pl$Decimal(scale = 5, precision = 10)
  )
)

# Select decimal columns:
df$select(cs$decimal())

# Select all columns except for those that are decimal:
df$select(!cs$decimal())
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
