% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__digit}
\alias{cs__digit}
\title{Select all columns having names consisting only of digits}
\usage{
cs__digit(ascii_only = FALSE)
}
\arguments{
\item{ascii_only}{Indicate whether to consider only ASCII alphabetic
characters, or the full Unicode range of valid letters (accented,
idiographic, etc).}
}
\value{
A Polars selector
}
\description{
Select all columns having names consisting only of digits
}
\details{
Matching column names cannot contain any non-digit characters. Note that the
definition of "digit" consists of all valid Unicode digit characters (\code{d})
by default; this can be changed by setting \code{ascii_only = TRUE}.
}
\examples{
df <- pl$DataFrame(
  key = c("aaa", "bbb"),
  `2001` = 1:2,
  `2025` = 3:4
)

# Select columns with digit names:
df$select(cs$digit())

# Select all columns except for those with digit names:
df$select(!cs$digit())

# Demonstrate use of ascii_only flag (by default all valid unicode digits
# are considered, but this can be constrained to ascii 0-9):
df <- pl$DataFrame(`१९९९` = 1999, `२०७७` = 2077, `3000` = 3000)
df$select(cs$digit())
df$select(cs$digit(ascii_only = TRUE))
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
