% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__exclude}
\alias{cs__exclude}
\title{Select all columns except those matching the given columns, datatypes, or
selectors}
\usage{
cs__exclude(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names to exclude.}
}
\value{
A Polars selector
}
\description{
Select all columns except those matching the given columns, datatypes, or
selectors
}
\details{
If excluding a single selector it is simpler to write as \code{!selector} instead.
}
\examples{
df <- pl$DataFrame(
  aa = 1:3,
  ba = c("a", "b", NA),
  cc = c(NA, 2.5, 1.5)
)

# Exclude by column name(s):
df$select(cs$exclude("ba", "xx"))

# Exclude using a column name, a selector, and a dtype:
df$select(cs$exclude("aa", cs$string(), pl$Int32))
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
