% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__first}
\alias{cs__first}
\title{Select the first column in the current scope}
\usage{
cs__first()
}
\value{
A Polars selector
}
\description{
Select the first column in the current scope
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123L, 456L),
  baz = c(2.0, 5.5),
  zap = c(FALSE, TRUE)
)

# Select the first column:
df$select(cs$first())

# Select everything except for the first column:
df$select(!cs$first())
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
