% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__float}
\alias{cs__float}
\title{Select all float columns.}
\usage{
cs__float()
}
\value{
A Polars selector
}
\description{
Select all float columns.
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123L, 456L),
  baz = c(2.0, 5.5),
  zap = c(FALSE, TRUE),
  .schema_overrides = list(baz = pl$Float32, zap = pl$Float64),
)

# Select all float columns:
df$select(cs$float())

# Select all columns except for those that are float:
df$select(!cs$float())
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
