% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__last}
\alias{cs__last}
\title{Select the last column in the current scope}
\usage{
cs__last()
}
\value{
A Polars selector
}
\description{
Select the last column in the current scope
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123L, 456L),
  baz = c(2.0, 5.5),
  zap = c(FALSE, TRUE)
)

# Select the last column:
df$select(cs$last())

# Select everything except for the last column:
df$select(!cs$last())
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
