% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__matches}
\alias{cs__matches}
\title{Select all columns that match the given regex pattern}
\usage{
cs__matches(pattern)
}
\arguments{
\item{pattern}{A valid regular expression pattern, compatible with the
\verb{regex crate <https://docs.rs/regex/latest/regex/>}_.}
}
\value{
A Polars selector
}
\description{
Select all columns that match the given regex pattern
}
\examples{
df <- pl$DataFrame(
  foo = c("x", "y"),
  bar = c(123, 456),
  baz = c(2.0, 5.5),
  zap = c(0, 1)
)

# Match column names containing an "a", preceded by a character that is not
# "z":
df$select(cs$matches("[^z]a"))

# Do not match column names ending in "R" or "z" (case-insensitively):
df$select(!cs$matches(r"((?i)R|z$)"))
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
