% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__nested}
\alias{cs__nested}
\title{Select all nested columns}
\usage{
cs__nested()
}
\value{
A Polars \link[=cs]{selector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A nested column is a \link[=pl__List]{list}, \link[=pl__Array]{array} or \link[=pl__Struct]{struct}.
}
\examples{
df <- pl$DataFrame(
  foo = data.frame(a = c("xx", "x"), b = c("yy", "y")),
  bar = c(123, 456),
  baz = c(2, 5.5),
  wow = list(c(1, 2), c(3)),
)

# Select all nested columns:
df$select(cs$nested())

# Select all columns except for those that are nested:
df$select(!cs$nested())
}
\seealso{
\itemize{
\item \link{cs} for the documentation on operators supported by selectors.
\item \code{\link[=cs__by_dtype]{cs$by_dtype()}}: Select all columns matching the given dtype(s).
\item \code{\link[=cs__list]{cs$list()}}: Select all list columns.
\item \code{\link[=cs__array]{cs$array()}}: Select all array columns.
\item \code{\link[=cs__struct]{cs$struct()}}: Select all struct columns.
}
}
