% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__string}
\alias{cs__string}
\title{Select all String (and, optionally, Categorical) string columns.}
\usage{
cs__string(..., include_categorical = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{include_categorical}{If \code{TRUE}, also select categorical columns.}
}
\value{
A Polars selector
}
\description{
Select all String (and, optionally, Categorical) string columns.
}
\examples{
df <- pl$DataFrame(
  w = c("xx", "yy", "xx", "yy", "xx"),
  x = c(1, 2, 1, 4, -2),
  y = c(3.0, 4.5, 1.0, 2.5, -2.0),
  z = c("a", "b", "a", "b", "b")
)$with_columns(
  z = pl$col("z")$cast(pl$Categorical())
)

# Group by all string columns, sum the numeric columns, then sort by the
# string cols:
df$group_by(cs$string())$agg(cs$numeric()$sum())$sort(cs$string())

# Group by all string and categorical columns:
df$
  group_by(cs$string(include_categorical = TRUE))$
  agg(cs$numeric()$sum())$
  sort(cs$string(include_categorical = TRUE))
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
