% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{cs__time}
\alias{cs__time}
\title{Select all time columns}
\usage{
cs__time()
}
\value{
A Polars selector
}
\description{
Select all time columns
}
\examples{
\dontshow{if (requireNamespace("hms", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- pl$DataFrame(
  dtm = as.POSIXct(c("2001-5-7 10:25", "2031-12-31 00:30")),
  dt = as.Date(c("1999-12-31", "2024-8-9")),
  tm = hms::parse_hms(c("0:0:0", "23:59:59"))
)

# Select time columns:
df$select(cs$time())

# Select all columns except for those that are time:
df$select(!cs$time())
\dontshow{\}) # examplesIf}
}
\seealso{
\link{cs} for the documentation on operators supported by Polars
selectors.
}
