% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{DataFrame}
\alias{DataFrame}
\alias{print()}
\alias{print.DataFrame}
\alias{DataFrame_lazy}
\alias{DataFrame_select}
\alias{select}
\alias{DataFrame_with_columns}
\alias{with_columns}
\alias{DataFrame_limit}
\alias{limit}
\title{Polars DataFrame}
\usage{
\method{print}{DataFrame}(x)

DataFrame_select(...)

DataFrame_with_columns(...)

select(...)
}
\arguments{
\item{x}{DataFrame}

\item{...}{any expressions or strings}

\item{n}{positive numeric or integer number not larger than 2^32}

\item{data}{object inheriting data.frame or list of equal length vectors and/or Series.}
}
\value{
self

DataFrame

DataFrame
}
\description{
\code{DataFrame}-object is an \code{externalptr} to rust polars DataFrame with $methods() exposed.
Most methods return another \code{DataFrame}-class instance or similar which allows for method chaining.

take limit of n rows of query
}
\details{
any number will converted to u32. Negative raises error
}
\examples{
pl$DataFrame(iris)$lazy()$filter(pl$col("Sepal.Length") > 5)$collect()
pl$DataFrame(iris)
#use of lazy method
pl$DataFrame(iris)$lazy()$filter(pl$col("Sepal.Length") >= 7.7)$collect()
}
\keyword{DataFrame}
