% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame}
\alias{DataFrame}
\alias{DataFrame_constructor}
\title{create new DataFrame}
\usage{
DataFrame(data)
}
\arguments{
\item{data}{a data.frame or list of mixed vectors and Series of equal length.}

\item{make_names_unique}{default TRUE, any duplicated names will be prefixed a running number}
}
\value{
DataFrame
}
\description{
create new DataFrame
}
\examples{
pl$DataFrame(iris)
pl$DataFrame(list(a= c(1,2,3,4,5), b=1:5, c = letters[1:5]))
}
\keyword{DataFrame}
