% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{DataFrame}
\alias{DataFrame}
\alias{lazy}
\alias{select}
\alias{DataFrame_select}
\alias{filter}
\alias{DataFrame_filter}
\title{Polars DataFrame}
\usage{
DataFrame_select(...)

DataFrame_filter(bool_expr)
}
\arguments{
\item{...}{any expressions or strings}

\item{bool_expr}{Polars expression which will evaluate to a bool pl$Series}

\item{data}{object inheriting data.frame or list of equal length vectors and/or Series.}
}
\description{
\code{DataFrame}-object is an \code{externalptr} to rust polars DataFrame with $methods() exposed.
Most methods return another \code{DataFrame}-class instance or similar which allows for method chaining.
\href{http://commonmark.org/help}{Commonmark web site}
}
\examples{
pl$DataFrame(iris)$lazy()$filter(pl$col("Sepal.Length") > 5)$collect()
#use of lazy method
pl$DataFrame(iris)$lazy()$filter(pl$col("Sepal.Length") >= 7.7)$collect()
}
