% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__cast}
\alias{dataframe__cast}
\title{Cast DataFrame column(s) to the specified dtype}
\usage{
dataframe__cast(..., strict = TRUE)
}
\value{
A polars \link{DataFrame}
}
\description{
Cast DataFrame column(s) to the specified dtype
}
\examples{
df <- pl$DataFrame(
  foo = 1:3,
  bar = c(6, 7, 8),
  ham = as.Date(c("2020-01-02", "2020-03-04", "2020-05-06"))
)

# Cast only some columns
df$cast(list(foo = pl$Float32, bar = pl$UInt8))

# Cast all columns to the same type
df$cast(pl$String)
}
