% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__explode}
\alias{dataframe__explode}
\title{Explode the frame to long format by exploding the given columns}
\usage{
dataframe__explode(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names, expressions, or
a selector defining them. The underlying columns being exploded must be of
the \code{List} or \code{Array} data type.}
}
\value{
A polars \link{DataFrame}
}
\description{
Explode the frame to long format by exploding the given columns
}
\examples{
df <- pl$DataFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)

df$explode("numbers")
}
