% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__fill_nan}
\alias{dataframe__fill_nan}
\title{Fill floating point \code{NaN} value with a fill value}
\usage{
dataframe__fill_nan(value)
}
\arguments{
\item{value}{Value used to fill \code{NaN} values.}
}
\value{
A polars \link{DataFrame}
}
\description{
Fill floating point \code{NaN} value with a fill value
}
\examples{
df <- pl$DataFrame(
  a = c(1.5, 2, NaN, 4),
  b = c(1.5, NaN, NaN, 4)
)
df$fill_nan(99)$collect()
}
