% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__fill_null}
\alias{dataframe__fill_null}
\title{Fill null values using the specified value or strategy}
\usage{
dataframe__fill_null(
  value = NULL,
  strategy = NULL,
  limit = NULL,
  ...,
  matches_supertype = TRUE
)
}
\arguments{
\item{value}{Value used to fill null values.}

\item{strategy}{Strategy used to fill null values. Must be one of:
\code{"forward"}, \code{"backward"}, \code{"min"}, \code{"max"}, \code{"mean"}, \code{"zero"}, \code{"one"},
or \code{NULL} (default).}

\item{limit}{Number of consecutive null values to fill when using the
\code{"forward"} or \code{"backward"} strategy.}

\item{...}{These dots are for future extensions and must be empty.}

\item{matches_supertype}{Fill all matching supertypes of the fill \code{value}
literal.}
}
\value{
A polars \link{DataFrame}
}
\description{
Fill null values using the specified value or strategy
}
\examples{
df <- pl$DataFrame(
  a = c(1.5, 2, NA, 4),
  b = c(1.5, NA, NA, 4)
)
df$fill_null(99)

df$fill_null(strategy = "forward")

df$fill_null(strategy = "max")

df$fill_null(strategy = "zero")
}
