% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__get_column}
\alias{dataframe__get_column}
\title{Get a single column by name}
\usage{
dataframe__get_column(name)
}
\arguments{
\item{name}{Name of the column to retrieve.}
}
\value{
A \link[=Series]{polars Series}
}
\description{
Get a single column by name
}
\examples{
df <- pl$DataFrame(foo = 1:3, bar = 4:6)
df$get_column("foo")

tryCatch(
  df$get_column("baz"),
  error = function(e) print(e)
)
}
