% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__get_column_index}
\alias{dataframe__get_column_index}
\title{Find the index of a column by name}
\usage{
dataframe__get_column_index(name)
}
\arguments{
\item{name}{Name of the column to find.}
}
\value{
Numeric value (0-indexed) indicating the index of the column
}
\description{
Find the index of a column by name
}
\examples{
df <- pl$DataFrame(foo = 1:3, bar = 4:6, ham = c("a", "b", "c"))
df$get_column_index("ham")

tryCatch(
  df$get_column_index("sandwich"),
  error = function(e) print(e)
)
}
