% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__get_columns}
\alias{dataframe__get_columns}
\title{Get the DataFrame as a list of Series}
\usage{
dataframe__get_columns()
}
\value{
A list of \link{Series}
}
\description{
Get the DataFrame as a list of Series
}
\examples{
df <- pl$DataFrame(foo = c(1, 2, 3), bar = c(4, 5, 6))
df$get_columns()

df <- pl$DataFrame(
  a = 1:4,
  b = c(0.5, 4, 10, 13),
  c = c(TRUE, TRUE, FALSE, TRUE)
)
df$get_columns()
}
\seealso{
\itemize{
\item \code{\link[=as.list.polars_data_frame]{as.list(<polars_data_frame>)}}
}
}
