% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__group_by}
\alias{dataframe__group_by}
\title{Group a DataFrame}
\usage{
dataframe__group_by(..., maintain_order = FALSE)
}
\value{
\link[=GroupBy_class]{GroupBy} (a DataFrame with special groupby methods like \verb{$agg()})
}
\description{
Group a DataFrame
}
\details{
Within each group, the order of the rows is always preserved,
regardless of the \code{maintain_order} argument.
}
\examples{
df <- pl$DataFrame(
  a = c("a", "b", "a", "b", "c"),
  b = c(1, 2, 1, 3, 3),
  c = c(5, 4, 3, 2, 1)
)

df$group_by("a")$agg(pl$col("b")$sum())

# Set `maintain_order = TRUE` to ensure the order of the groups is
# consistent with the input.
df$group_by("a", maintain_order = TRUE)$agg(pl$col("c"))

# Group by multiple columns by passing a list of column names.
df$group_by(c("a", "b"))$agg(pl$max("c"))

# Or pass some arguments to group by multiple columns in the same way.
# Expressions are also accepted.
df$group_by("a", pl$col("b") \%/\% 2)$agg(
  pl$col("c")$mean()
)

# The columns will be renamed to the argument names.
df$group_by(d = "a", e = pl$col("b") \%/\% 2)$agg(
  pl$col("c")$mean()
)
}
\seealso{
\itemize{
\item \code{\link[=DataFrame_partition_by]{<DataFrame>$partition_by()}}
}
}
