% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__is_duplicated}
\alias{dataframe__is_duplicated}
\title{Get a mask of all duplicated rows in this DataFrame.}
\usage{
dataframe__is_duplicated()
}
\value{
A \link[=Series]{polars Series}
}
\description{
Get a mask of all duplicated rows in this DataFrame.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, 3, 1),
  b = c("x", "y", "z", "x")
)
df$is_duplicated()

# This mask can be used to visualize the duplicated lines like this:
df$filter(df$is_duplicated())
}
