% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__lazy}
\alias{dataframe__lazy}
\title{Convert an existing DataFrame to a LazyFrame}
\usage{
dataframe__lazy()
}
\value{
A polars \link{LazyFrame}
}
\description{
Start a new lazy query from a DataFrame.
}
\examples{
pl$DataFrame(a = 1:2, b = c(NA, "a"))$lazy()
}
