% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__mean_horizontal}
\alias{dataframe__mean_horizontal}
\title{Take the mean of all values horizontally across columns.}
\usage{
dataframe__mean_horizontal()
}
\value{
A \link[=Series]{polars Series}
}
\description{
Take the mean of all values horizontally across columns.
}
\examples{
df <- pl$DataFrame(
  foo = c(1, 2, 3),
  bar = c(4.0, 5.0, 6.0),
)
df$mean_horizontal()
}
