% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__pivot}
\alias{dataframe__pivot}
\title{Pivot a frame from long to wide format}
\usage{
dataframe__pivot(
  on,
  on_columns = NULL,
  ...,
  index = NULL,
  values = NULL,
  aggregate_function = NULL,
  maintain_order = TRUE,
  sort_columns = FALSE,
  separator = "_",
  column_naming = c("auto", "combine")
)
}
\arguments{
\item{on}{The column(s) whose values will be used as the new columns of the output.}

\item{on_columns}{What value combinations will be considered for the output table.
If \code{NULL} (default), all unique values found in the \code{on} column(s) will be used.}

\item{...}{These dots are for future extensions and must be empty.}

\item{index}{The column(s) that remain from the input to the output. The
output will have one row for each unique combination of the \code{index}'s values.
If \code{NULL}, all remaining columns not specified in \code{on} and \code{values} will be used.
At least one of \code{index} and \code{values} must be specified.}

\item{values}{The existing column(s) of values which will be moved under the
new columns from \code{index}. If an aggregation is specified, these are the
values on which the aggregation will be computed. If \code{NULL}, all remaining
columns not specified in \code{on} and \code{index} will be used. At least one of
\code{index} and \code{values} must be specified.}

\item{aggregate_function}{Choose from:
\itemize{
\item \code{NULL} (default): no aggregation takes place, will raise error if multiple values
are in group. Same as \code{pl$element()$item(allow_empty = TRUE)}.
\item A predefined aggregate function string, one of \code{"min"}, \code{"max"},
\code{"first"}, \code{"last"}, \code{"sum"}, \code{"mean"}, \code{"median"}, \code{"len"}, \code{"item"}.
Same as \verb{pl$element()$<function>()}.
\item An \link[=Expr]{expression} to do the aggregation.
}}

\item{maintain_order}{Ensure the values of \code{index} are sorted by discovery
order.}

\item{sort_columns}{Sort the transposed columns by name. Default is by order
of discovery.}

\item{separator}{Used as separator/delimiter in generated column names in
case of multiple values columns.}

\item{column_naming}{How resulting column names will be constructed.
\code{"auto"} (default) combines with separator if there are multiple \code{values}
columns, otherwise just uses the \code{on_columns} names. \code{"combine"} always
combines the \code{values} columns' names with the \code{on_columns} names.}
}
\value{
A polars \link{DataFrame}
}
\description{
Reshape data from long to wide format, known as "pivot wider".
}
\examples{
# Suppose we have a dataframe of test scores achieved by some students,
# where each row represents a distinct test.
df <- pl$DataFrame(
  name = c("Cady", "Cady", "Karen", "Karen"),
  subject = c("maths", "physics", "maths", "physics"),
  test_1 = c(98, 99, 61, 58),
  test_2 = c(100, 100, 60, 60)
)
df

# Using pivot(), we can reshape so we have one row per student, with
# different subjects as columns, and their `test_1` scores as values:
df$pivot("subject", index = "name", values = "test_1")

# You can use selectors too - here we include all test scores
# in the pivoted table:
df$pivot("subject", values = cs$starts_with("test"))

# If you end up with multiple values per cell, you can specify how to
# aggregate them with `aggregate_function`:
df <- pl$DataFrame(
  ix = c(1, 1, 2, 2, 1, 2),
  col = c("a", "a", "a", "a", "b", "b"),
  foo = c(0, 1, 2, 2, 7, 1),
  bar = c(0, 2, 0, 0, 9, 4)
)
df

df$pivot("col", index = "ix", aggregate_function = "sum")

# You can also pass a custom aggregation function using `pl$element()`:
df <- pl$DataFrame(
  col1 = c("a", "a", "a", "b", "b", "b"),
  col2 = c("x", "x", "x", "x", "y", "y"),
  col3 = c(6, 7, 3, 2, 5, 7),
)
df$pivot(
  "col2",
  index = "col1",
  values = "col3",
  aggregate_function = pl$element()$tanh()$mean(),
)
}
