% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__rechunk}
\alias{dataframe__rechunk}
\title{Rechunk the data in this DataFrame to a contiguous allocation}
\usage{
dataframe__rechunk()
}
\value{
A polars \link{DataFrame}
}
\description{
This will make sure all subsequent operations have optimal and predictable
performance.
}
