% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__remove}
\alias{dataframe__remove}
\title{Remove rows, dropping those that match the given predicate expression(s)}
\usage{
dataframe__remove(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Expression that evaluates to
a boolean Series.}
}
\value{
A polars \link{LazyFrame}
}
\description{
The original order of the remaining rows is preserved. Rows where the filter
does not evaluate to \code{TRUE} are retained (this includes rows where the
predicate evaluates as \code{null}).
}
\examples{
df <- pl$DataFrame(
  foo = c(2, 3, NA, 4, 0),
  bar = c(5, 6, NA, NA, 0),
  ham = c("a", "b", NA, "c", "d")
)

# Remove rows matching a condition:
df$remove(pl$col("bar") >= 5)

# Discard rows based on multiple conditions, combined with and/or operators:
df$remove((pl$col("foo") >= 0) & (pl$col("bar") >= 0))

df$remove((pl$col("foo") >= 0) | (pl$col("bar") >= 0))

# Remove rows by comparing two columns against each other:
df$remove(pl$col("foo")$ne_missing(pl$col("bar")))
}
