% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__sample}
\alias{dataframe__sample}
\title{Sample from this DataFrame}
\usage{
dataframe__sample(
  n = NULL,
  ...,
  fraction = NULL,
  with_replacement = FALSE,
  shuffle = FALSE,
  seed = NULL
)
}
\arguments{
\item{n}{Number of items to return. Can be an Expr. Strings are parsed as
column names. Cannot be used with \code{fraction}. Defaults to 1 if \code{fraction}
is \code{NULL}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{fraction}{Fraction of items to return. Can be an Expr. Strings are
parsed as column names. Cannot be used with \code{n}.}

\item{with_replacement}{Allow values to be sampled more than once.}

\item{shuffle}{Shuffle the order of sampled data points.}

\item{seed}{Seed for the random number generator. If \code{NULL} (default), a
random seed is generated for each sample operation.}
}
\value{
A polars \link{DataFrame}
}
\description{
Sample from this DataFrame
}
\examples{
df <- pl$DataFrame(
  foo = 1:3,
  bar = 6:8,
  ham = c("a", "b", "c")
)
df$sample(n = 2, seed = 0)
}
