% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__select}
\alias{dataframe__select}
\title{Select and modify columns of a DataFrame}
\usage{
dataframe__select(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Name-value pairs of objects to be converted to polars \link[=Expr]{expressions}
by the \code{\link[=as_polars_expr]{as_polars_expr()}} function.
Characters are parsed as column names, other non-expression inputs are parsed as \link[=pl__lit]{literals}.
Each name will be used as the expression name.}
}
\value{
A polars \link{DataFrame}
}
\description{
Select and perform operations on a subset of columns only. This discards
unmentioned columns (like \code{.()} in \code{data.table} and contrarily to
\code{dplyr::mutate()}).

One cannot use new variables in subsequent expressions in the same
\verb{$select()} call. For instance, if you create a variable \code{x}, you will only
be able to use it in another \verb{$select()} or \verb{$with_columns()} call.
}
\examples{
as_polars_df(iris)$select(
  abs_SL = pl$col("Sepal.Length")$abs(),
  add_2_SL = pl$col("Sepal.Length") + 2
)
}
