% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__serialize}
\alias{dataframe__serialize}
\alias{pl__deserialize_df}
\title{Serialize the DataFrame to a binary format}
\usage{
dataframe__serialize()

pl__deserialize_df(data)
}
\arguments{
\item{data}{A raw vector of serialized \link{DataFrame}.}
}
\value{
\itemize{
\item \verb{<dataframe>$serialize()} returns raw vector of serialized \link{DataFrame}.
\item \code{pl$deserialize_df()} returns a deserialized \link{DataFrame}.
}
}
\description{
Serialize the DataFrame to a binary format.
Currently, this format is uncompressed Arrow IPC stream format,
so other Apache Arrow implementations may be able to read it.
}
\examples{
df <- pl$DataFrame(
  foo = 1:3,
  bar = 6:8,
)$cast(bar = pl$UInt8)

# Serialize the DataFrame to a binary format
serialized <- df$serialize()
serialized

# The bytes can later be deserialized back to a DataFrame
pl$deserialize_df(serialized)

# Other Apache Arrow implementations may be able to read it.
if (requireNamespace("arrow", quietly = TRUE)) {
  arrow::read_ipc_stream(serialized, as_data_frame = FALSE)
}
}
