% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__sort}
\alias{dataframe__sort}
\title{Sort a DataFrame}
\usage{
dataframe__sort(
  ...,
  descending = FALSE,
  nulls_last = FALSE,
  multithreaded = TRUE,
  maintain_order = FALSE
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{maintain_order}{Keep the same order as the original data. This is
more expensive to compute. Setting this to \code{TRUE} blocks the possibility to
run on the streaming engine.}
}
\value{
A polars \link{DataFrame}
}
\description{
Sort a DataFrame
}
\examples{
df <- mtcars
df$mpg[1] <- NA
df <- as_polars_df(df)
df$sort("mpg")
df$sort("mpg", nulls_last = TRUE)
df$sort("cyl", "mpg")
df$sort(c("cyl", "mpg"))
df$sort(c("cyl", "mpg"), descending = TRUE)
df$sort(c("cyl", "mpg"), descending = c(TRUE, FALSE))
df$sort(pl$col("cyl"), pl$col("mpg"))
}
