% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__sum}
\alias{dataframe__sum}
\title{Aggregate the columns of this DataFrame to their sum values}
\usage{
dataframe__sum()
}
\value{
A polars \link{DataFrame}
}
\description{
Aggregate the columns of this DataFrame to their sum values
}
\examples{
df <- pl$DataFrame(a = 1:4, b = c(1, 2, 1, 1))
df$sum()
}
