% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-to_r_list.R
\name{dataframe__to_r_list}
\alias{dataframe__to_r_list}
\title{Export the polars DataFrame as an R list of R vectors}
\usage{
dataframe__to_r_list(
  ...,
  int64 = "double",
  as_clock_class = FALSE,
  struct = "dataframe",
  decimal = "double",
  ambiguous = "raise",
  non_existent = "raise"
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{int64}{Determine how to convert Polars' Int64, UInt32, or UInt64 type values to R type.
One of the followings:
\itemize{
\item \code{"double"} (default): Convert to the R's \link{double} type.
Accuracy may be degraded.
\item \code{"character"}: Convert to the R's \link{character} type.
\item \code{"integer"}: Convert to the R's \link{integer} type.
If the value is out of the range of R's integer type, export as \link{NA_integer_}.
\item \code{"integer64"}: Convert to the \link[bit64:bit64-package]{bit64::integer64} class.
The \link[bit64:bit64-package]{bit64} package must be installed.
If the value is out of the range of \link[bit64:bit64-package]{bit64::integer64}, export as \link[bit64:as.integer64.character]{bit64::NA_integer64_}.
}}

\item{as_clock_class}{A logical value indicating whether to export datetimes and duration as
the \link[clock:clock-package]{clock} package's classes.
\itemize{
\item \code{FALSE} (default): Duration values are exported as \link{difftime}
and datetime values are exported as \link{POSIXct}.
Accuracy may be degraded.
\item \code{TRUE}: Duration values are exported as \link[clock:duration-helper]{clock_duration},
datetime without timezone values are exported as \link[clock:as_naive_time]{clock_naive_time},
and datetime with timezone values are exported as \link[clock:as_zoned_time]{clock_zoned_time}.
For this case, the \link[clock:clock-package]{clock} package must be installed.
Accuracy will be maintained.
}}

\item{struct}{Determine how to convert Polars' Struct type values to R class.
One of the followings:
\itemize{
\item \code{"dataframe"} (default): Convert to the R's \link{data.frame} class.
\item \code{"tibble"}: Convert to the \link[tibble:tbl_df-class]{tibble} class.
If the \link[tibble:tibble-package]{tibble} package is not installed, a warning will be shown.
}}

\item{decimal}{Determine how to convert Polars' Decimal type values to R type.
One of the followings:
\itemize{
\item \code{"double"} (default): Convert to the R's \link{double} type.
\item \code{"character"}: Convert to the R's \link{character} type.
}}

\item{ambiguous}{Determine how to deal with ambiguous datetimes.
Only applicable when \code{as_clock_class} is set to \code{FALSE} and
datetime without timezone values are exported as \link{POSIXct}.
Character vector or Expression containing the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"earliest"}: Use the earliest datetime
\item \code{"latest"}: Use the latest datetime
\item \code{"null"}: Return a \code{NA} value
}}

\item{non_existent}{Determine how to deal with non-existent datetimes.
Only applicable when \code{as_clock_class} is set to \code{FALSE} and
datetime without timezone values are exported as \link{POSIXct}.
One of the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"null"}: Return a \code{NA} value
}}
}
\value{
A \link{list} containing \link{vector}s and \link[=data.frame]{R dataframe}s
}
\description{
This method is a shortcut of \verb{<DataFrame>$to_struct()$to_r_vector(ensure_vector = TRUE)}.
Different from \code{\link[=dataframe__get_columns]{<DataFrame>$get_columns()}}, all columns are exported as R objects.
}
\examples{
df <- pl$DataFrame(foo = c(1, 2, 3), bar = c(4, 5, 6))
df$to_r_list()

df <- pl$DataFrame(
  a = 1:4,
  b = c(0.5, 4, 10, 13),
  c = c(TRUE, TRUE, FALSE, TRUE)
)
df$to_r_list()
}
\seealso{
\itemize{
\item \code{\link[=dataframe__get_columns]{<DataFrame>$get_columns()}}: Export the polars \link{DataFrame} as a \link{list} of \link{Series}.
\item \code{\link[=series__to_r_vector]{<Series>$to_r_vector()}}: Export the polars \link{Series} as an R vector.
}
}
