% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__to_series}
\alias{dataframe__to_series}
\title{Select column as Series at index location}
\usage{
dataframe__to_series(index = 0)
}
\arguments{
\item{index}{Index of the column to return as Series. Defaults to 0, which is
the first column.}
}
\value{
Series or NULL
}
\description{
Select column as Series at index location
}
\examples{
df <- as_polars_df(iris[1:10, ])

# default is to extract the first column
df$to_series()

# Polars is 0-indexed, so we use index = 1 to extract the *2nd* column
df$to_series(index = 1)

# doesn't error if the column isn't there
df$to_series(index = 8)
}
