% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__to_struct}
\alias{dataframe__to_struct}
\title{Convert a DataFrame to a Series of type Struct}
\usage{
dataframe__to_struct(name = "")
}
\arguments{
\item{name}{A character. Name for the struct \link{Series}.}
}
\value{
A \link{Series} of the struct type
}
\description{
Convert a DataFrame to a Series of type Struct
}
\examples{
df <- pl$DataFrame(
  a = 1:5,
  b = c("one", "two", "three", "four", "five"),
)
df$to_struct("nums")
}
\seealso{
\itemize{
\item \code{\link[=as_polars_series]{as_polars_series()}}
}
}
