% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__unique}
\alias{dataframe__unique}
\title{Drop duplicate rows}
\usage{
dataframe__unique(
  subset = NULL,
  ...,
  keep = c("any", "none", "first", "last"),
  maintain_order = FALSE
)
}
\arguments{
\item{subset}{Column name(s) or selector(s), to consider when identifying
duplicate rows. If \code{NULL} (default), use all columns.}

\item{...}{These dots are for future extensions and must be empty.}

\item{keep}{Which of the duplicate rows to keep. Must be one of:
\itemize{
\item \code{"any"}: does not give any guarantee of which row is kept. This allows
more optimizations.
\item \code{"none"}: don’t keep duplicate rows.
\item \code{"first"}: keep first unique row.
\item \code{"last"}: keep last unique row.
}}

\item{maintain_order}{Keep the same order as the original data. This is
more expensive to compute. Setting this to \code{TRUE} blocks the possibility to
run on the streaming engine.}
}
\value{
A polars \link{DataFrame}
}
\description{
Drop duplicate rows
}
\examples{
df <- pl$DataFrame(
  foo = c(1, 2, 3, 1),
  bar = c("a", "a", "a", "a"),
  ham = c("b", "b", "b", "b"),
)
df$unique(maintain_order = TRUE)

df$unique(subset = c("bar", "ham"), maintain_order = TRUE)

df$unique(keep = "last", maintain_order = TRUE)
}
