% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__unnest}
\alias{dataframe__unnest}
\title{Decompose struct columns into separate columns for each of their fields}
\usage{
dataframe__unnest(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Name of the struct column(s)
that should be unnested.}
}
\value{
A polars \link{DataFrame}
}
\description{
The new columns will be inserted at the location of the struct column.
}
\examples{
df <- pl$DataFrame(
  a = 1:5,
  b = c("one", "two", "three", "four", "five"),
  c = 6:10
)$
  select(
  pl$struct("b"),
  pl$struct(c("a", "c"))$alias("a_and_c")
)
df

df$unnest("a_and_c")
df$unnest(pl$col("a_and_c"))
}
