% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__var}
\alias{dataframe__var}
\title{Aggregate the columns in the DataFrame to their variance value}
\usage{
dataframe__var(ddof = 1)
}
\arguments{
\item{ddof}{"Delta Degrees of Freedom": the divisor used in the calculation
is \code{N - ddof}, where \code{N} represents the number of elements. By default
\code{ddof} is 1.}
}
\value{
A polars \link{DataFrame}
}
\description{
Aggregate the columns in the DataFrame to their variance value
}
\examples{
df <- pl$DataFrame(a = 1:4, b = c(1, 2, 1, 1))
df$var()
df$var(ddof = 0)
}
