% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__with_row_index}
\alias{dataframe__with_row_index}
\title{Add a row index as the first column in the DataFrame}
\usage{
dataframe__with_row_index(name = "index", offset = 0)
}
\arguments{
\item{name}{Name of the index column.}

\item{offset}{Start the index at this offset. Cannot be negative.}
}
\value{
A polars \link{DataFrame}
}
\description{
Add a row index as the first column in the DataFrame
}
\examples{
df <- pl$DataFrame(x = c(1, 3, 5), y = c(2, 4, 6))
df$with_row_index()

df$with_row_index("id", offset = 1000)

# An index column can also be created using the expressions int_range()
# and len()$
df$with_columns(
  index = pl$int_range(pl$len(), dtype = pl$UInt32)
)
}
