% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-ipc.R
\name{dataframe__write_ipc_stream}
\alias{dataframe__write_ipc_stream}
\title{Write to Arrow IPC stream format.}
\usage{
dataframe__write_ipc_stream(
  path,
  ...,
  compression = c("zstd", "lz4", "uncompressed"),
  compat_level = c("newest", "oldest")
)
}
\arguments{
\item{path}{A character. File path to which the file should be written.}

\item{...}{These dots are for future extensions and must be empty.}

\item{compression}{Determines the compression algorithm.
Must be one of:
\itemize{
\item \code{"uncompressed"} or \code{NULL}: Write an uncompressed Arrow file.
\item \code{"lz4"}: Fast compression/decompression.
\item \code{"zstd"} (default): Good compression performance.
}}

\item{compat_level}{Determines the compatibility level when exporting
Polars' internal data structures. When specifying a new compatibility level,
Polars exports its internal data structures that might not be interpretable by
other Arrow implementations. The level can be specified as the name
(e.g., \code{"newest"}) or as a scalar integer (Currently, \code{0} or \code{1} is supported).
\itemize{
\item \code{"newest"} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} (default):
Use the highest level, currently same as \code{1} (Low compatibility).
\item \code{"oldest"}: Same as \code{0} (High compatibility).
}}
}
\value{
\code{NULL} invisibly.
}
\description{
Write to Arrow IPC stream format.
}
\examples{
\dontshow{if (requireNamespace("nanoarrow", quiet = TRUE)) withAutoprint(\{ # examplesIf}
tmpf <- tempfile()
as_polars_df(mtcars)$write_ipc_stream(tmpf)
pl$read_ipc_stream(tmpf)
\dontshow{\}) # examplesIf}
}
