% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-frame.R
\name{dataframe__write_json}
\alias{dataframe__write_json}
\title{Serialize to JSON representation}
\usage{
dataframe__write_json(file)
}
\arguments{
\item{file}{File path to which the result will be written.}
}
\value{
The input DataFrame is returned.
}
\description{
Serialize to JSON representation
}
\examples{
\dontshow{if (requireNamespace("jsonlite", quiet = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dat <- as_polars_df(head(mtcars))
destination <- tempfile()

dat$select(pl$col("drat", "mpg"))$write_json(destination)
jsonlite::fromJSON(destination)
\dontshow{\}) # examplesIf}
}
