% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-json.R
\name{dataframe__write_ndjson}
\alias{dataframe__write_ndjson}
\title{Serialize to newline delimited JSON representation}
\usage{
dataframe__write_ndjson(
  file,
  ...,
  compression = c("uncompressed", "gzip", "zstd"),
  compression_level = NULL,
  check_extension = TRUE
)
}
\arguments{
\item{file}{File path to which the result will be written.}

\item{...}{These dots are for future extensions and must be empty.}

\item{compression}{What compression format to use. Must be one of
\code{uncompressed} (default), \code{gzip}, or \code{zstd}.}

\item{compression_level}{The compression level to use, typically 0-9 or \code{NULL}
to let the engine choose.}

\item{check_extension}{Whether to check if the filename matches the
compression settings. Will raise an error if compression is set to
\code{"uncompressed"} and the filename ends in one of \code{".gz"}, \code{".zst"}, \code{".zstd"},
or if \code{compression != "uncompressed"} and the file uses a mismatched extension.
Only applies if file is a path.}
}
\value{
\code{NULL} invisibly.
}
\description{
Serialize to newline delimited JSON representation
}
\examples{
\dontshow{if (requireNamespace("jsonlite", quiet = TRUE)) withAutoprint(\{ # examplesIf}
dat <- as_polars_df(head(mtcars))
destination <- tempfile()

dat$select(pl$col("drat", "mpg"))$write_ndjson(destination)
jsonlite::stream_in(file(destination))
\dontshow{\}) # examplesIf}
}
