% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-json-functions.R
\name{dataframe__write_ndjson}
\alias{dataframe__write_ndjson}
\title{Serialize to newline delimited JSON representation}
\usage{
dataframe__write_ndjson(file)
}
\arguments{
\item{file}{File path to which the result will be written.}
}
\value{
The input DataFrame is returned.
}
\description{
Serialize to newline delimited JSON representation
}
\examples{
\dontshow{if (requireNamespace("jsonlite", quiet = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dat <- as_polars_df(head(mtcars))
destination <- tempfile()

dat$select(pl$col("drat", "mpg"))$write_ndjson(destination)
jsonlite::stream_in(file(destination))
\dontshow{\}) # examplesIf}
}
