% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_class}
\alias{DataFrame_class}
\title{Inner workings of the DataFrame-class}
\value{
Not applicable
}
\description{
The \code{DataFrame}-class is simply two environments of respectively the public
and private methods/function calls to the polars Rust side. The instantiated
\code{DataFrame}-object is an \code{externalptr} to a low-level Rust polars DataFrame
object.

The S3 method \code{.DollarNames.RPolarsDataFrame} exposes all public \verb{$foobar()}-methods
which are callable onto the object. Most methods return another \code{DataFrame}-
class instance or similar which allows for method chaining. This class system
could be called "environment classes" (in lack of a better name) and is the
same class system \code{extendr} provides, except here there are both a public and
private set of methods. For implementation reasons, the private methods are
external and must be called from \code{.pr$DataFrame$methodname()}. Also, all
private methods must take any \code{self} as an argument, thus they are pure
functions. Having the private methods as pure functions solved/simplified
self-referential complications.
}
\details{
Check out the source code in \href{https://github.com/pola-rs/r-polars/blob/main/R/dataframe__frame.R}{R/dataframe_frame.R}
to see how public methods are derived from private methods. Check out
\href{https://github.com/pola-rs/r-polars/blob/main/R/extendr-wrappers.R}{extendr-wrappers.R}
to see the \code{extendr}-auto-generated methods. These are moved to \code{.pr} and
converted into pure external functions in \href{https://github.com/pola-rs/r-polars/blob/main/R/after-wrappers.R}{after-wrappers.R}. In \href{https://github.com/pola-rs/r-polars/blob/main/R/zzz.R}{zzz.R}
(named \code{zzz} to be last file sourced) the \code{extendr}-methods are removed and
replaced by any function prefixed \code{DataFrame_}.
}
\examples{
# see all public exported method names (normally accessed via a class
# instance with $)
ls(.pr$env$DataFrame)

# see all private methods (not intended for regular use)
ls(.pr$DataFrame)


# make an object
df = pl$DataFrame(iris)


# use a public method/property
df$shape
df2 = df

# use a private method, which has mutability
result = .pr$DataFrame$set_column_from_robj(df, 150:1, "some_ints")

# Column exists in both dataframes-objects now, as they are just pointers to
# the same object
# There are no public methods with mutability.
df$columns
df2$columns

# set_column_from_robj-method is fallible and returned a result which could
# be "ok" or an error.
# No public method or function will ever return a result.
# The `result` is very close to the same as output from functions decorated
# with purrr::safely.
# To use results on the R side, these must be unwrapped first such that
# potentially errors can be thrown. `unwrap(result)` is a way to communicate
# errors happening on the Rust side to the R side. `Extendr` default behavior
# is to use `panic!`(s) which would cause some unnecessarily confusing and
# some very verbose error messages on the inner workings of rust.
# `unwrap(result)` in this case no error, just a NULL because this mutable
# method does not return any ok-value.

# Try unwrapping an error from polars due to unmatching column lengths
err_result = .pr$DataFrame$set_column_from_robj(df, 1:10000, "wrong_length")
tryCatch(unwrap(err_result, call = NULL), error = \(e) cat(as.character(e)))
}
\keyword{DataFrame}
