% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_clone}
\alias{DataFrame_clone}
\title{Clone a DataFrame}
\usage{
<DataFrame>$clone()
}
\value{
DataFrame
}
\description{
Rarely useful as DataFrame is nearly 100\% immutable
Any modification of a DataFrame would lead to a clone anyways.
}
\examples{
df1 = pl$DataFrame(iris);
df2 =  df1$clone();
df3 = df1
pl$mem_address(df1) != pl$mem_address(df2)
pl$mem_address(df1) == pl$mem_address(df3)

}
\keyword{DataFrame}
