% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_clone}
\alias{DataFrame_clone}
\title{Clone a DataFrame}
\usage{
<DataFrame>$clone()
}
\value{
A DataFrame
}
\description{
This is rarely useful as a DataFrame is nearly 100\% immutable.
Any modification of a DataFrame would lead to a clone anyway.
}
\examples{
df1 = pl$DataFrame(iris)
df2 = df1$clone()
df3 = df1

# the clone and the original don't have the same address...
pl$mem_address(df1) != pl$mem_address(df2)

# ... but simply assigning df1 to df3 change the address anyway
pl$mem_address(df1) == pl$mem_address(df3)
}
\keyword{DataFrame}
