% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_columns}
\alias{DataFrame_columns}
\title{column names}
\usage{
DataFrame_columns()
}
\value{
char vec of column names
}
\description{
get column names as DataFrames
}
\examples{
pl$DataFrame(iris)$columns()
}
\keyword{DataFrame}
