% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_columns}
\alias{DataFrame_columns}
\title{get/set columns (the names columns)}
\usage{
DataFrame_columns()
}
\value{
char vec of column names
}
\description{
get/set column names of DataFrame object
}
\examples{
df = pl$DataFrame(iris)

#get values
df$columns

#set + get values
df$columns = letters[1:5] #<- is fine too
df$columns
}
\keyword{DataFrame}
