% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_describe}
\alias{DataFrame_describe}
\title{Summary statistics for a DataFrame}
\usage{
DataFrame_describe(percentiles = c(0.25, 0.75))
}
\arguments{
\item{percentiles}{One or more percentiles to include in the summary statistics.
All values must be in the range \verb{[0; 1]}.}
}
\value{
DataFrame
}
\description{
This returns the total number of rows, the number of missing
values, the mean, standard deviation, min, max, median and the percentiles
specified in the argument \code{percentiles}.
}
\examples{
pl$DataFrame(iris)$describe()
}
\keyword{DataFrame}
