% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_drop}
\alias{DataFrame_drop}
\title{Drop}
\usage{
DataFrame_drop(columns)
}
\arguments{
\item{character}{vector Name of the column(s) that should be removed from the dataframe.}
}
\value{
DataFrame
}
\description{
Remove columns from the dataframe.
}
\examples{
pl$DataFrame(mtcars)$drop(c("mpg", "hp"))
}
\keyword{DataFrame}
