% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_drop}
\alias{DataFrame_drop}
\title{Drop columns of a DataFrame}
\usage{
DataFrame_drop(...)
}
\arguments{
\item{...}{Characters of column names to drop. Passed to \code{\link[=pl_col]{pl$col()}}.}
}
\value{
DataFrame
}
\description{
Drop columns of a DataFrame
}
\examples{
pl$DataFrame(mtcars)$drop(c("mpg", "hp"))

# equivalent
pl$DataFrame(mtcars)$drop("mpg", "hp")
}
