% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_drop_nulls}
\alias{DataFrame_drop_nulls}
\title{Drop nulls}
\usage{
DataFrame_drop_nulls(subset = NULL)
}
\arguments{
\item{subset}{string or vector of strings. Column name(s) for which null values are considered. If set to NULL (default), use all columns.}
}
\value{
DataFrame
}
\description{
Drop all rows that contain null values.
}
\examples{
tmp = mtcars
tmp[1:3, "mpg"] = NA
tmp[4, "hp"] = NA
pl$DataFrame(tmp)$drop_nulls()$height
pl$DataFrame(tmp)$drop_nulls("mpg")$height
pl$DataFrame(tmp)$drop_nulls(c("mpg", "hp"))$height
}
\keyword{DataFrame}
