% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_dtype_strings}
\alias{DataFrame_dtype_strings}
\title{DataFrame dtype strings}
\format{
An object of class \code{character} of length 1.
}
\usage{
DataFrame_dtype_strings
}
\value{
string vector
}
\description{
Get column types as strings.
}
\examples{
pl$DataFrame(iris)$dtype_strings()
}
\keyword{DataFrame}
