% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_dtypes}
\alias{DataFrame_dtypes}
\alias{DataFrame_schema}
\title{Data types information}
\usage{
DataFrame_dtypes()

DataFrame_schema()
}
\value{
\verb{$dtypes} returns an unnamed list with the data type of each column.
\verb{$schema} returns a named list with the column names and the data type of
each column.
}
\description{
Get the data type of all columns. You can see all available
types with \code{names(pl$dtypes)}. The data type of each column is also shown
when printing the DataFrame.
}
\examples{
pl$DataFrame(iris)$dtypes

pl$DataFrame(iris)$schema
}
\keyword{DataFrame}
