% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_dtypes}
\alias{DataFrame_dtypes}
\alias{width,}
\alias{nrow}
\alias{DataFrame_schema}
\title{DataFrame dtypes}
\usage{
DataFrame_dtypes()

DataFrame_schema()
}
\value{
width as numeric scalar

width as numeric scalar
}
\description{
Get dtypes of columns in DataFrame.
Dtypes can also be found in column headers when printing the DataFrame.

Get dtypes of columns in DataFrame.
Dtypes can also be found in column headers when printing the DataFrame.
}
\examples{
pl$DataFrame(iris)$dtypes

pl$DataFrame(iris)$schema

}
\keyword{DataFrame}
