% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_explode}
\alias{DataFrame_explode}
\title{Explode the DataFrame to long format by exploding the given columns}
\usage{
DataFrame_explode(columns, ...)
}
\arguments{
\item{columns}{Name of the column(s) to explode. Columns must be of datatype
List or Utf8. Accepts \code{col} expressions as input as well.}

\item{...}{More columns to explode as above but provided one Expr per argument.}
}
\value{
DataFrame
}
\description{
Explode the DataFrame to long format by exploding the given columns
}
\examples{
df = pl$DataFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)
df

df$explode("numbers")
}
\keyword{DataFrame}
