% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_explode}
\alias{DataFrame_explode}
\title{Explode the DataFrame to long format by exploding the given columns}
\usage{
DataFrame_explode(...)
}
\arguments{
\item{...}{Column(s) to be exploded as individual \verb{Into<Expr>} or list/vector of \verb{Into<Expr>}.
A handful of places in rust-polars, only the plain variant \code{Expr::Column} is accepted. This is
currenly on of such places. Therefore \code{pl$col("name")} and \code{pl$all()} is allowed, not
\code{pl$col("name")$alias("newname")}. \code{"name"} is implcitly converted to \code{pl$col("name")}.}
}
\value{
DataFrame
}
\description{
Explode the DataFrame to long format by exploding the given columns
}
\examples{
df = pl$DataFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)
df

df$explode("numbers")
}
\keyword{DataFrame}
