% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_fill_nan}
\alias{DataFrame_fill_nan}
\title{Fill NaN}
\usage{
DataFrame_fill_nan(fill_value)
}
\arguments{
\item{fill_value}{Value to fill NaN with.}
}
\value{
DataFrame
}
\description{
Fill floating point NaN values by an Expression evaluation.
}
\examples{
df = pl$DataFrame(
  a = c(1.5, 2, NaN, 4),
  b = c(1.5, NaN, NaN, 4)
)
df$fill_nan(99)
}
\keyword{DataFrame}
