% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_filter}
\alias{DataFrame_filter}
\title{Filter rows of a DataFrame}
\usage{
DataFrame_filter(...)
}
\arguments{
\item{...}{Polars expressions which will evaluate to a boolean.}
}
\value{
A DataFrame with only the rows where the conditions are \code{TRUE}.
}
\description{
Filter rows with an Expression defining a boolean column.
Multiple expressions are combined with \code{&} (AND).
This is equivalent to \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\details{
Rows where the condition returns \code{NA} are dropped.
}
\examples{
df = pl$DataFrame(iris)

df$filter(pl$col("Sepal.Length") > 5)

# This is equivalent to
# df$filter(pl$col("Sepal.Length") > 5 & pl$col("Petal.Width") < 1)
df$filter(pl$col("Sepal.Length") > 5, pl$col("Petal.Width") < 1)

# rows where condition is NA are dropped
iris2 = iris
iris2[c(1, 3, 5), "Species"] = NA
df = pl$DataFrame(iris2)

df$filter(pl$col("Species") == "setosa")
}
\keyword{DataFrame}
