% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_first}
\alias{DataFrame_first}
\title{First}
\usage{
DataFrame_first()
}
\value{
A new \code{DataFrame} object with applied filter.
}
\description{
Get the first row of the DataFrame.
}
\examples{
pl$DataFrame(mtcars)$first()
}
\keyword{DataFrame}
