% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_first}
\alias{DataFrame_first}
\title{Get the first row}
\usage{
DataFrame_first()
}
\value{
\code{DataFrame} with a single row
}
\description{
Get the first row
}
\examples{
pl$DataFrame(mtcars)$first()
}
\keyword{DataFrame}
