% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{DataFrame_get_column}
\alias{DataFrame_get_column}
\title{get column as Series from DataFrame}
\usage{
DataFrame_get_column(name)
}
\arguments{
\item{name}{name of column to get as series}
}
\value{
a Series
}
\description{
get column as Series from DataFrame
}
\examples{
pl$DataFrame(iris)$get_column("Species")
}
\keyword{DataFrame}
