% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_get_column}
\alias{DataFrame_get_column}
\alias{get_column}
\title{Get Column}
\usage{
DataFrame_get_column(name)
}
\value{
Series
}
\description{
get one column by name as series
}
\examples{
df = pl$DataFrame(iris[1,])
df$get_column("Species")
}
\keyword{DataFrame}
