% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_get_column}
\alias{DataFrame_get_column}
\title{Get Column (as one Series)}
\usage{
DataFrame_get_column(name)
}
\arguments{
\item{name}{name of column to extract as Series}
}
\value{
Series
}
\description{
get one column by name as series
}
\examples{
df = pl$DataFrame(iris[1,])
df$get_column("Species")
}
