% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_get_columns}
\alias{DataFrame_get_columns}
\title{Get columns (as Series)}
\format{
An object of class \code{character} of length 1.
}
\usage{
DataFrame_get_columns
}
\value{
list of series
}
\description{
get columns as list of series
}
\examples{
df = pl$DataFrame(iris[1,])
df$get_columns()
}
\keyword{DataFrame}
