% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_get_columns}
\alias{DataFrame_get_columns}
\title{Get columns (as Series)}
\usage{
<DataFrame>$get_columns
}
\value{
A list of series
}
\description{
Extract all DataFrame columns as a list of Polars series.
}
\examples{
df = pl$DataFrame(iris[1:2, ])
df$get_columns()
}
\keyword{DataFrame}
