% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_groupby}
\alias{DataFrame_groupby}
\alias{groupby}
\title{groupby DataFrame}
\usage{
DataFrame_groupby(..., maintain_order = FALSE)
}
\arguments{
\item{...}{any expression}

\item{maintain_order}{bool}
}
\value{
GroupBy (subclass of DataFrame)
}
\description{
DataFrame$groupby(..., maintain_order = FALSE)
}
\keyword{DataFrame}
