% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_head}
\alias{DataFrame_head}
\title{Get the first \emph{n} rows}
\usage{
DataFrame_head(n)
}
\arguments{
\item{n}{Positive numeric or integer number not larger than 2^32. Any number
will be converted to \code{u32}. Negative raises error.}
}
\value{
DataFrame
}
\description{
Get the first \emph{n} rows of the query.
}
\examples{
pl$DataFrame(mtcars)$head(5)
}
\keyword{DataFrame}
