% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_head}
\alias{DataFrame_head}
\title{Head of a DataFrame}
\usage{
<DataFrame>$head(n)
}
\arguments{
\item{n}{Positive number not larger than 2^32.}
}
\value{
DataFrame
}
\description{
Get the first \code{n} rows of the query.
}
\details{
Any number will converted to u32. Negative raises error.
}
\keyword{DataFrame}
