% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_height}
\alias{DataFrame_height}
\alias{height}
\alias{nrow}
\title{Height of DataFrame}
\usage{
DataFrame_height()
}
\value{
height as numeric
}
\description{
Get height(nrow) of DataFrame
}
\examples{
pl$DataFrame(iris)$height

}
\keyword{DataFrame}
