% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_last}
\alias{DataFrame_last}
\title{Get the last row}
\usage{
DataFrame_last()
}
\value{
\code{DataFrame} with a single row
}
\description{
Get the last row
}
\examples{
pl$DataFrame(mtcars)$last()
}
\keyword{DataFrame}
