% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_lazy}
\alias{DataFrame_lazy}
\alias{lazy}
\title{DataFrame to LazyFrame}
\usage{
DataFrame_lazy()
}
\value{
a LazyFrame
}
\description{
Start a new lazy query from a DataFrame
}
\examples{
pl$DataFrame(iris)$lazy()

#use of lazy method
pl$DataFrame(iris)$lazy()$filter(pl$col("Sepal.Length") >= 7.7)$collect()
}
\keyword{DataFrame}
