% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_lazy}
\alias{DataFrame_lazy}
\alias{lazy}
\title{Convert an existing DataFrame to a LazyFrame}
\format{
An object of class \code{character} of length 1.
}
\usage{
DataFrame_lazy
}
\value{
A LazyFrame
}
\description{
Start a new lazy query from a DataFrame.
}
\examples{
pl$DataFrame(iris)$lazy()

}
\keyword{DataFrame}
\keyword{LazyFrame_new}
