% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_limit}
\alias{DataFrame_limit}
\title{Limit a DataFrame}
\usage{
DataFrame_limit(n)
}
\arguments{
\item{n}{Positive number not larger than 2^32.}
}
\value{
DataFrame
}
\description{
Take some maximum number of rows.
}
\details{
Any number will converted to u32. Negative raises error.
}
\examples{
pl$DataFrame(iris)$limit(6)
}
\keyword{DataFrame}
