% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_mean}
\alias{DataFrame_mean}
\title{Mean}
\usage{
DataFrame_mean()
}
\value{
A new \code{DataFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their mean value.
}
\examples{
pl$DataFrame(mtcars)$mean()
}
\keyword{DataFrame}
