% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_n_chunks}
\alias{DataFrame_n_chunks}
\title{Get the number of chunks of the Series' in a DataFrame}
\usage{
DataFrame_n_chunks(strategy = "all")
}
\arguments{
\item{strategy}{string either 'all' or 'first'. 'first' only returns chunks for first Series.}
}
\value{
real vector of chunk counts per Series.
}
\description{
Get the number of chunks of the Series' in a DataFrame
}
\examples{
df = pl$concat(
  1:10,
  pl$concat(1:5,1:5, rechunk = FALSE, how = "vertical")$rename("b"),
  how = "horizontal"
)
df
df$n_chunks()
}
\keyword{DataFrame}
