% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_null_count}
\alias{DataFrame_null_count}
\title{Count null values}
\format{
function
}
\usage{
<DataFrame>$null_count
}
\value{
DataFrame
}
\description{
Create a new DataFrame that shows the null (which correspond
to \code{NA} in R) counts per column.
}
\examples{
x = mtcars
x[1, 2:3] = NA
pl$DataFrame(x)$null_count()
}
\keyword{DataFrame}
