% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_quantile}
\alias{DataFrame_quantile}
\title{Quantile}
\usage{
<DataFrame>$quantile(quantile, interpolation = "nearest")
}
\arguments{
\item{quantile}{Numeric of length 1 between 0 and 1.}

\item{interpolation}{Interpolation method: "nearest", "higher", "lower",
"midpoint", or "linear".}
}
\value{
DataFrame
}
\description{
Aggregate the columns in the DataFrame to a unique quantile
value. Use \verb{$describe()} to specify several quantiles.
}
\examples{
pl$DataFrame(mtcars)$quantile(.4)
}
\keyword{DataFrame}
