% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_rename}
\alias{DataFrame_rename}
\title{Rename column names of a DataFrame}
\usage{
DataFrame_rename(...)
}
\arguments{
\item{...}{One of the following:
\itemize{
\item Params like \code{new_name = "old_name"} to rename selected variables.
\item As above but with params wrapped in a list
\item An R function that takes the old names character vector as input and
returns the new names character vector.
}}
}
\value{
\link[=DataFrame_class]{DataFrame}
}
\description{
Rename column names of a DataFrame
}
\details{
If existing names are swapped (e.g. \code{A} points to \code{B} and \code{B} points to \code{A}),
polars will block projection and predicate pushdowns at this node.
}
\examples{
df = pl$DataFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)

df$rename(apple = "foo")

df$rename(
  \(column_name) paste0("c", substr(column_name, 2, 100))
)
}
