% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_rename}
\alias{DataFrame_rename}
\title{Rename columns of a DataFrame}
\usage{
DataFrame_rename(existing, new)
}
\arguments{
\item{existing}{A character vector of names to replace.}

\item{new}{A character vector of names used as replacement.}
}
\value{
DataFrame
}
\description{
Rename columns of a DataFrame
}
\examples{
pl$DataFrame(mtcars)$rename(c("mpg", "hp"), c("miles_per_gallon", "horsepower"))
}
\keyword{DataFrame}
