% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_rename}
\alias{DataFrame_rename}
\title{Rename columns of a DataFrame}
\usage{
DataFrame_rename(mapping)
}
\arguments{
\item{mapping}{A named list like \code{list(old = "new")}.}
}
\value{
DataFrame
}
\description{
Rename columns of a DataFrame
}
\examples{
pl$DataFrame(mtcars)$lazy()$rename(list(mpg = "miles_per_gallon", hp = "horsepower"))
}
\keyword{DataFrame}
