% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_rename}
\alias{DataFrame_rename}
\title{Rename columns of a DataFrame}
\usage{
DataFrame_rename(...)
}
\arguments{
\item{...}{One of the following:
\itemize{
\item params like \code{new_name = "old_name"} to rename selected variables.
\item as above but with params wrapped in a list
}}
}
\value{
DataFrame
}
\description{
Rename columns of a DataFrame
}
\examples{
df = pl$DataFrame(mtcars)

df$rename(miles_per_gallon = "mpg", horsepower = "hp")

replacements <- list(miles_per_gallon = "mpg", horsepower = "hp")
df$rename(replacements)
}
\keyword{DataFrame}
