% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_reverse}
\alias{DataFrame_reverse}
\title{Reverse}
\usage{
DataFrame_reverse()
}
\value{
DataFrame
}
\description{
Reverse the DataFrame.
}
\examples{
pl$DataFrame(mtcars)$reverse()
}
\keyword{LazyFrame}
