% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_select}
\alias{DataFrame_select}
\alias{select}
\title{perform select on DataFrame}
\usage{
DataFrame_select(...)
}
\arguments{
\item{...}{expresssions or strings defining columns to select(keep) in context the DataFrame}
}
\description{
related to dplyr \verb{mutate()`` and data.table }.()`.
}
\keyword{DataFrame}
