% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_shape}
\alias{DataFrame_shape}
\title{Shape of  DataFrame}
\usage{
DataFrame_shape()
}
\value{
two length numeric vector of c(nrows,ncols)
}
\description{
Get shape/dimensions of DataFrame
}
\examples{
df = pl$DataFrame(iris)$shape

}
\keyword{DataFrame}
