% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_shift}
\alias{DataFrame_shift}
\title{Shift}
\usage{
DataFrame_shift(periods = 1)
}
\arguments{
\item{periods}{integer Number of periods to shift (may be negative).}
}
\value{
DataFrame
}
\description{
Shift the values by a given period.
}
\examples{
pl$DataFrame(mtcars)$shift(2)
}
\keyword{DataFrame}
