% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_shift_and_fill}
\alias{DataFrame_shift_and_fill}
\title{Shift and fill}
\usage{
DataFrame_shift_and_fill(fill_value, periods = 1)
}
\arguments{
\item{fill_value}{Fill values with the result of this expression.}

\item{periods}{Integer indicating the number of periods to shift (may be
negative).}
}
\value{
DataFrame
}
\description{
Shift the values by a given period and fill the resulting null values.
}
\examples{
pl$DataFrame(mtcars)$shift_and_fill(0, 2)
}
\keyword{DataFrame}
