% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_slice}
\alias{DataFrame_slice}
\title{Slice}
\usage{
DataFrame_slice(offset, length = NULL)
}
\arguments{
\item{offset}{integer}

\item{length}{integer or NULL}
}
\value{
LazyFrame
}
\description{
Get a slice of this DataFrame.
}
\examples{
pl$DataFrame(mtcars)$slice(2, 4)
mtcars[2:6,]
}
\keyword{LazyFrame}
