% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_sort}
\alias{DataFrame_sort}
\title{DataFrame Sort}
\usage{
DataFrame_sort(by, ..., descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{by}{Iterable Into\if{html}{\out{<Expr>}}, e.g. one Expr, or list mixed Expr and column name strings, or a
charachter vector of column names. The column(s) to sort by.}

\item{...}{more columns to sort by as above but provided one Expr per argument.}

\item{descending}{Sort descending? Defulat = FALSE logical vector of length 1 or same length
as number of Expr's from above by + ....}

\item{nulls_last}{Bool default FALSE, place all nulls_last?}
}
\value{
LazyFrame
}
\description{
sort a DataFrame by on or more Expr.
}
\details{
by and ... args allow to either provide e.g. a list of Expr or something which can
be converted into an Expr e.g. \verb{$sort(list(e1,e2,e3))},
or provide each Expr as an individual argument \verb{$sort(e1,e2,e3)}´ ... or both.
}
\examples{
df = pl$DataFrame(
 a = c(1L, 2L, NA),
 b = c(6, 5, 4),
 c = c("a", "b", "c")
)
df$sort("a")
}
\keyword{DataFrame}
