% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_std}
\alias{DataFrame_std}
\title{Std}
\usage{
<DataFrame>$std(ddof = 1)
}
\arguments{
\item{ddof}{Delta Degrees of Freedom: the divisor used in the calculation is
N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
A DataFrame with one row.
}
\description{
Aggregate the columns of this DataFrame to their standard
deviation values.
}
\examples{
pl$DataFrame(mtcars)$std()
}
\keyword{DataFrame}
