% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_sum}
\alias{DataFrame_sum}
\title{Sum}
\usage{
DataFrame_sum()
}
\value{
A new \code{DataFrame} object with applied aggregation.
}
\description{
Aggregate the columns of this DataFrame to their sum values.
}
\examples{
pl$DataFrame(mtcars)$sum()
}
\keyword{DataFrame}
