% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_tail}
\alias{DataFrame_tail}
\title{Get the last \emph{n} rows}
\usage{
DataFrame_tail(n)
}
\arguments{
\item{n}{Positive numeric or integer number not larger than 2^32. Any number
will be converted to \code{u32}. Negative raises error.}
}
\value{
DataFrame
}
\description{
Get the last \emph{n} rows.
}
\examples{
pl$DataFrame(mtcars)$tail(5)
}
\keyword{DataFrame}
