% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_data_frame}
\alias{DataFrame_to_data_frame}
\title{Return Polars DataFrame as R data.frame}
\usage{
DataFrame_to_data_frame(..., int64_conversion = pl$options$int64_conversion)
}
\arguments{
\item{...}{Any args pased to \code{as.data.frame()}.}

\item{int64_conversion}{How should Int64 values be handled when converting a
polars object to R?
\itemize{
\item \code{"double"} (default) converts the integer values to double.
\item \code{"bit64"} uses \code{bit64::as.integer64()} to do the conversion (requires
the package \code{bit64} to be attached).
\item \code{"string"} converts Int64 values to character.
}}
}
\value{
An R data.frame
}
\description{
Return Polars DataFrame as R data.frame
}
\examples{
df = pl$DataFrame(iris[1:3, ])
df$to_data_frame()
}
\keyword{DataFrame}
