% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_data_frame}
\alias{DataFrame_to_data_frame}
\alias{as.data.frame.DataFrame}
\title{Return Polars DataFrame as R data.frame}
\usage{
DataFrame_to_data_frame(...)

\method{as.data.frame}{DataFrame}(x, ...)
}
\arguments{
\item{...}{any params passed to as.data.frame}

\item{x}{DataFrame}
}
\value{
An R data.frame

data.frame
}
\description{
Return Polars DataFrame as R data.frame
}
\examples{
df = pl$DataFrame(iris[1:3, ])
df$to_data_frame()
}
\keyword{DataFrame}
