% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_data_frame}
\alias{DataFrame_to_data_frame}
\title{Return Polars DataFrame as R data.frame}
\usage{
<DataFrame>$to_data_frame(...)
}
\arguments{
\item{...}{Any args pased to \code{as.data.frame()}.}
}
\value{
An R data.frame
}
\description{
Return Polars DataFrame as R data.frame
}
\examples{
df = pl$DataFrame(iris[1:3, ])
df$to_data_frame()
}
\keyword{DataFrame}
