% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_ipc.R
\name{DataFrame_to_raw_ipc}
\alias{DataFrame_to_raw_ipc}
\title{Write Arrow IPC data to a raw vector}
\usage{
DataFrame_to_raw_ipc(
  compression = c("uncompressed", "zstd", "lz4"),
  ...,
  compat_level = FALSE
)
}
\arguments{
\item{compression}{\code{NULL} or a character of the compression method,
\code{"uncompressed"} or "lz4" or "zstd". \code{NULL} is equivalent to \code{"uncompressed"}.
Choose "zstd" for good compression performance. Choose "lz4"
for fast compression/decompression.}

\item{...}{Ignored.}

\item{compat_level}{Use a specific compatibility level when exporting Polars’
internal data structures. This can be:
\itemize{
\item an integer indicating the compatibility version (currently only 0 for oldest
and 1 for newest);
\item a logical value with \code{TRUE} for the newest version and \code{FALSE} for the oldest
version.
}}
}
\value{
A raw vector
}
\description{
Write Arrow IPC data to a raw vector
}
\examples{
df = pl$DataFrame(
  foo = 1:5,
  bar = 6:10,
  ham = letters[1:5]
)

raw_ipc = df$to_raw_ipc()

pl$read_ipc(raw_ipc)

if (require("arrow", quietly = TRUE)) {
  arrow::read_ipc_file(raw_ipc, as_data_frame = FALSE)
}
}
\seealso{
\itemize{
\item \code{\link[=DataFrame_write_ipc]{<DataFrame>$write_ipc()}}
}
}
