% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_series}
\alias{DataFrame_to_series}
\title{Get Series by idx, if there}
\usage{
DataFrame_to_series(idx = 0)
}
\arguments{
\item{idx}{numeric default 0, zero-index of what column to return as Series}
}
\value{
Series or NULL
}
\description{
get one column by idx as series from DataFrame.
Unlike get_column this method will not fail if no series found at idx but
return a NULL, idx is zero idx.
}
\examples{
pl$DataFrame(a=1:4)$to_series()
}
\keyword{DataFrame}
