% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_to_struct}
\alias{DataFrame_to_struct}
\alias{to_struct}
\title{to_struct}
\usage{
DataFrame_to_struct(name = "")
}
\arguments{
\item{name}{name of new Series}
}
\value{
to_struct() returns a Series
}
\description{
to_struct
}
\examples{
#round-trip conversion from DataFrame with two columns
df = pl$DataFrame(a=1:5,b=c("one","two","three","four","five"))
s = df$to_struct()
s
s$to_r() # to r list
df_s = s$to_frame() #place series in a new DataFrame
df_s$unnest() # back to starting df
}
\keyword{DataFrame}
