% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_unique}
\alias{DataFrame_unique}
\title{Drop duplicated rows}
\usage{
<DataFrame>$unique(subset = NULL, keep = "first", maintain_order = FALSE)
}
\arguments{
\item{subset}{A character vector with the names of the column(s) to use to
identify duplicates. If \code{NULL} (default), use all columns.}

\item{keep}{Which of the duplicate rows to keep:
\itemize{
\item "first": Keep first unique row.
\item "last": Keep last unique row.
\item "none": Don’t keep duplicate rows.
}}

\item{maintain_order}{Keep the same order as the original \code{DataFrame}. Setting
this to \code{TRUE} makes it more expensive to compute and blocks the possibility
to run on the streaming engine. The default value can be changed with
\code{pl$set_options(maintain_order = TRUE)}.}
}
\value{
DataFrame
}
\description{
Drop duplicated rows
}
\examples{
df = pl$DataFrame(
  x = sample(10, 100, rep = TRUE),
  y = sample(10, 100, rep = TRUE)
)
df$height

df$unique()$height
df$unique(subset = "x")$height

df$unique(keep = "last")

# only keep unique rows
df$unique(keep = "none")
}
\keyword{DataFrame}
