% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_unnest}
\alias{DataFrame_unnest}
\title{Unnest a DataFrame struct columns.}
\usage{
DataFrame_unnest(names = NULL)
}
\arguments{
\item{names}{Names of the struct columns to unnest. If \code{NULL} (default), then
all "struct" columns are unnested.}
}
\value{
A DataFrame where all "struct" columns are unnested. Non-struct
columns are not modified.
}
\description{
Unnest a DataFrame struct columns.
}
\examples{
df = pl$DataFrame(a = 1:5, b = c("one", "two", "three", "four", "five"))
df = df$to_struct()$to_frame()
df

df$unnest()
}
\keyword{DataFrame}
