% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_width}
\alias{DataFrame_width}
\alias{width,}
\alias{nrow}
\title{Width of DataFrame}
\usage{
DataFrame_width()
}
\value{
width as numeric scalar
}
\description{
Get width(ncol) of DataFrame
}
\examples{
pl$DataFrame(iris)$width()

}
\keyword{DataFrame}
