% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_frame.R
\name{DataFrame_with_columns}
\alias{DataFrame_with_columns}
\alias{with_columns}
\title{with columns}
\usage{
DataFrame_with_columns(...)
}
\arguments{
\item{...}{any expressions or strings}
}
\value{
DataFrame
}
\description{
with columns
}
\keyword{DataFrame}
