% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes-classes.R
\name{datatype__to_dtype_expr}
\alias{datatype__to_dtype_expr}
\title{Return a DataTypeExpr with a static DataType}
\usage{
datatype__to_dtype_expr()
}
\value{
A polars datatype expression.
This is not the same as a \link[=polars_expr]{polars expression}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Return a DataTypeExpr with a static DataType.
}
