% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_Array}
\alias{DataType_Array}
\title{Create Array DataType}
\usage{
DataType_Array(datatype = "unknown", width)
}
\arguments{
\item{datatype}{An inner DataType. The default is \code{"Unknown"} and is only a
placeholder for when inner DataType does not matter, e.g. as used in example.}

\item{width}{The length of the arrays.}
}
\value{
An array DataType with an inner DataType
}
\description{
The Array and List datatypes are very similar. The only difference is that
sub-arrays all have the same length while sublists can have different lengths.
Array methods can be accessed via the \verb{$arr} subnamespace.
}
\examples{
# basic Array
pl$Array(pl$Int32, 4)
# some nested Array
pl$Array(pl$Array(pl$Boolean, 3), 2)
}
