% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_Categorical}
\alias{DataType_Categorical}
\title{Create Categorical DataType}
\format{
function
}
\usage{
DataType_Categorical(ordering = "physical")
}
\arguments{
\item{ordering}{Either \code{"physical"} (default) or \code{"lexical"}.}
}
\value{
A Categorical DataType
}
\description{
Create Categorical DataType
}
\details{
When a
categorical variable is created, its string values (or "lexical" values) are
stored and encoded as integers ("physical" values) by order of appearance.

For example, suppose we have a categorical variable whose first values are
\code{"Peach"}, \code{"Apple"}, and "\verb{Pear"}. When this categorical variable is read,
it each unique value will be mapped to an integer by order of appearance:
\code{"Peach"} will be mapped to 0, \code{"Apple"} to 1, and \code{"Pear"} to 2.

If we order by \code{"lexical"} values, we look at the string... TODO
}
\examples{
pl$Categorical()

pl$Categorical(ordering = "lexical")
}
