% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\docType{data}
\name{DataType_constructors}
\alias{DataType_constructors}
\title{internal collection of datatype constructors}
\format{
An object of class \code{list} of length 0.
}
\usage{
DataType_constructors
}
\description{
internal collection of datatype constructors
}
\keyword{datasets}
