% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{DataType_Datetime}
\alias{DataType_Datetime}
\title{Data type representing a calendar date and time of day.}
\usage{
DataType_Datetime(time_unit = "us", time_zone = NULL)
}
\arguments{
\item{time_unit}{Unit of time. One of \code{"ms"}, \code{"us"} (default) or \code{"ns"}.}

\item{time_zone}{Time zone string, as defined in \code{\link[=OlsonNames]{OlsonNames()}}.
When using to match dtypes, can use \code{"*"} to check for Datetime columns that
have any timezone.}
}
\value{
Datetime DataType
}
\description{
The underlying representation of this type is a 64-bit signed integer.
The integer indicates the number of time units since the Unix epoch (1970-01-01 00:00:00).
The number can be negative to indicate datetimes before the epoch.
}
\examples{
pl$Datetime("ns", "Pacific/Samoa")

df = pl$DataFrame(
  naive_time = as.POSIXct("1900-01-01"),
  zoned_time = as.POSIXct("1900-01-01", "UTC")
)
df

df$select(pl$col(pl$Datetime("us", "*")))
}
