% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype-expr-expr.R
\name{datatype_expr__inner_dtype}
\alias{datatype_expr__inner_dtype}
\title{Get the inner DataType of a List or Array}
\usage{
datatype_expr__inner_dtype()
}
\value{
A polars datatype expression.
This is not the same as a \link[=polars_expr]{polars expression}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get the inner DataType of a List or Array.
}
\examples{
df <- pl$DataFrame(
  a = list(1L),
  b = list(list("a")),
  c = list(data.frame(x = 1, y = 2))
)

df$select(
  a_inner_dtype = pl$dtype_of("a")$inner_dtype()$display(),
  b_inner_dtype = pl$dtype_of("b")$inner_dtype()$display(),
  c_inner_dtype = pl$dtype_of("c")$inner_dtype()$display()
)
}
