% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{.DollarNames.DataFrame}
\alias{.DollarNames.DataFrame}
\title{auto complete $-access into object}
\usage{
\method{.DollarNames}{DataFrame}(x, pattern = "")
}
\description{
called by the interactive R session internally
}
\keyword{DataFrame}
