% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{.DollarNames.DataFrame}
\alias{.DollarNames.DataFrame}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{DataFrame}(x, pattern = "")
}
\arguments{
\item{x}{DataFrame}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
char vec

Doesn't return a value. This is used for autocompletion in RStudio.
}
\description{
called by the interactive R session internally
}
\keyword{internal}
