% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{.DollarNames.GroupBy}
\alias{.DollarNames.GroupBy}
\title{auto complete $-access into object}
\usage{
\method{.DollarNames}{GroupBy}(x, pattern = "")
}
\description{
called by the interactive R session internally
}
\keyword{GroupBy}
