% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{.DollarNames.LazyFrame}
\alias{.DollarNames.LazyFrame}
\title{auto complete $-access into object}
\usage{
\method{.DollarNames}{LazyFrame}(x, pattern = "")
}
\description{
called by the interactive R session internally
}
\examples{
e = pl$lit("any polars object")$
e$ #place cursor after $ this line and press tab

#manually call like this (should never be needed)
.DollarNames.Expr
}
\keyword{LazyFrame}
