% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{.DollarNames.RField}
\alias{.DollarNames.RField}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{RField}(x, pattern = "")
}
\arguments{
\item{x}{string name of an RField}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
Doesn't return a value. This is used for autocompletion in RStudio.
}
\description{
called by the interactive R session internally
}
\keyword{internal}
