% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{.DollarNames.Series}
\alias{.DollarNames.Series}
\title{auto complete $-access into a polars object}
\usage{
\method{.DollarNames}{Series}(x, pattern = "")
}
\arguments{
\item{x}{Series}

\item{pattern}{code-stump as string to auto-complete}
}
\value{
char vec
}
\description{
called by the interactive R session internally
}
\keyword{internal}
