% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\docType{data}
\name{.pr}
\alias{.pr}
\title{polars-API: private calls to rust-polars}
\format{
An object of class \code{environment} of length 13.
}
\usage{
.pr
}
\description{
\code{.pr}
Original extendr bindings converted into pure functions
}
\examples{
#.pr$DataFrame$print() is an external function where self is passed as arg
rpolars:::.pr$DataFrame$print(self = pl$DataFrame(iris))
rpolars:::print_env(.pr,".pr the collection of private method calls to rust-polars")
}
\keyword{api_private}
