% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{.pr}
\alias{.pr}
\title{polars-API: internal extendr bindings to polars}
\format{
An object of class \code{environment} of length 3.
}
\usage{
not for public use
}
\description{
\code{.pr}
Contains original rextendr bindings to polars
}
\keyword{datasets}
