% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{.pr}
\alias{.pr}
\title{polars-API: internal extendr bindings to polars}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.pr
}
\description{
\code{.pr}
Contains original rextendr bindings to polars
}
\examples{
#.pr$DataFrame$print() is an external function where self is passed as arg
minipolars:::.pr$DataFrame$print(self = DataFrame(iris))
}
\keyword{datasets}
