% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{dt_epoch}
\alias{dt_epoch}
\alias{dt.epoch}
\title{Epoch}
\format{
function
}
\arguments{
\item{tu}{string option either 'ns', 'us', 'ms', 's' or  'd'}
}
\value{
Expr of epoch as UInt32
}
\description{
Get the time passed since the Unix EPOCH in the give time unit.
}
\details{
ns and perhaps us will exceed integerish limit if returning to
R as flaot64/double.
}
\examples{
 pl$date_range(as.Date("2022-1-1"),lazy = TRUE)$dt$epoch("ns")$lit_to_s()
 pl$date_range(as.Date("2022-1-1"),lazy = TRUE)$dt$epoch("ms")$lit_to_s()
 pl$date_range(as.Date("2022-1-1"),lazy = TRUE)$dt$epoch("s")$lit_to_s()
 pl$date_range(as.Date("2022-1-1"),lazy = TRUE)$dt$epoch("d")$lit_to_s()
}
\keyword{ExprDT}
