% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_dynamic.R
\name{DynamicGroupBy_class}
\alias{DynamicGroupBy_class}
\title{Operations on Polars DataFrame grouped on time or integer values}
\value{
not applicable
}
\description{
Operations on Polars DataFrame grouped on time or integer values
}
