% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_dynamic.R
\name{DynamicGroupBy_ungroup}
\alias{DynamicGroupBy_ungroup}
\title{Ungroup a DynamicGroupBy object}
\usage{
<DynamicGroupBy>$ungroup()
}
\value{
\link[=DataFrame_class]{DataFrame}
}
\description{
Revert the \verb{$group_by_dynamic()} operation. Doing
\verb{<DataFrame>$group_by_dynamic(...)$ungroup()} returns the original \code{DataFrame}.
}
\examples{
df = pl$DataFrame(
  time = pl$date_range(
    start = strptime("2021-12-16 00:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC"),
    end = strptime("2021-12-16 03:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC"),
    interval = "30m",
    eager = TRUE,
  ),
  n = 0:6
)
df

df$group_by_dynamic("time", every = "1h")$ungroup()
}
