% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expect_grepl_error}
\alias{expect_grepl_error}
\title{expect grepl error}
\usage{
expect_grepl_error(expr, expected_err = NULL, do_not_repeat_call = TRUE, ...)
}
\arguments{
\item{expr}{an R expression to test}

\item{expected_err}{one or several string patterns passed to grepl}

\item{do_not_repeat_call}{bool, prevent error-handler to add call to err msg
useful for grepping the same error message, without grep-patterns becomes
included in the error message. Latter leads to false positive outcomes.}

\item{...}{args passed to expect_identical which will run if grepl fails}
}
\value{
invisble NULL
}
\description{
expect grepl error
}
\details{
expr must raise an error and expected_err pattern must match
against the error text with grepl()
}
\examples{
# passes as "carrot" is in "orange and carrot"
rpolars:::expect_grepl_error(stop("orange and carrot"),"carrot")
rpolars:::expect_grepl_error(stop("orange and carrot"),c("carrot","orange"))
}
\keyword{internal}
