% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expect_strictly_identical}
\alias{expect_strictly_identical}
\title{expect strictly identical}
\usage{
expect_strictly_identical(object, expected, ...)
}
\arguments{
\item{object}{R object to test}

\item{expected}{R object of how it exactly should be}

\item{...}{other args passed to expect_identical}
}
\value{
testthat::expect_identical return
}
\description{
stricter than testthat::expect_identical in regards of NaN an NA's
}
\details{
objects are evaluated on prints via utils::str()
}
\examples{
if (FALSE){

  #to testthat NA_real and NaN are the same !?
  #https://github.com/r-lib/waldo/issues/150
  testthat::expect_identical(NA_real_,NaN)

  rpolars:::expect_strictly_identical(NA_real_,NaN)
}
}
\keyword{internal}
