% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{lit}
\alias{lit}
\alias{Expr.map}
\title{polars literal}
\usage{
Expr.map(lambda, output_type = NULL, `_agg_list` = NULL)
}
\arguments{
\item{x}{any R expression yielding an integer, float or bool}
}
\value{
Expr, literal of that value
}
\description{
polars literal
}
\examples{
pl$col("some_column") / pl$lit(42)
}
