% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{Expr}
\alias{Expr}
\alias{print.Expr}
\alias{lit}
\alias{Expr_map}
\title{Polars Expr}
\usage{
\method{print}{Expr}(x)

Expr_map(lambda, output_type = NULL, `_agg_list` = NULL)
}
\arguments{
\item{x}{any R expression yielding an integer, float or bool}
}
\value{
self

Expr, literal of that value
}
\description{
Polars pl$Expr
}
\examples{
#Expr has the following methods/constructors
ls(minipolars:::Expr)

pl$col("this_column")$sum()$over("that_column")
pl$col("some_column")$sum()$over("some_other_column")
pl$col("some_column") / pl$lit(42)
}
