% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__abs}
\alias{expr__abs}
\title{Compute absolute values}
\usage{
expr__abs()
}
\value{
A polars \link{expression}
}
\description{
Compute absolute values
}
\examples{
df <- pl$DataFrame(a = -1:2)
df$with_columns(abs = pl$col("a")$abs())
}
