% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__agg_groups}
\alias{expr__agg_groups}
\title{Get the group indexes of the group by operation}
\usage{
expr__agg_groups()
}
\value{
A polars \link{expression}
}
\description{
Should be used in aggregation context only.
}
\examples{
df <- pl$DataFrame(
  group = rep(c("one", "two"), each = 3),
  value = c(94, 95, 96, 97, 97, 99)
)

df$group_by("group", maintain_order = TRUE)$agg(pl$col("value")$agg_groups())
}
