% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__and}
\alias{expr__and}
\title{Apply logical AND on two expressions}
\usage{
expr__and(other)
}
\arguments{
\item{other}{Element to add. Can be a string (only if \code{expr} is a string), a
numeric value or an other expression.}
}
\value{
A polars \link{expression}
}
\description{
Combine two boolean expressions with AND.
}
\examples{
pl$lit(TRUE) & TRUE
pl$lit(TRUE)$and(pl$lit(TRUE))
}
