% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arccos}
\alias{expr__arccos}
\title{Compute inverse cosine}
\usage{
expr__arccos()
}
\value{
A polars \link{expression}
}
\description{
Compute inverse cosine
}
\examples{
pl$DataFrame(a = c(-1, cos(0.5), 0, 1, NA))$
  with_columns(arccos = pl$col("a")$arccos())
}
