% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arcsin}
\alias{expr__arcsin}
\title{Compute inverse sine}
\usage{
expr__arcsin()
}
\value{
A polars \link{expression}
}
\description{
Compute inverse sine
}
\examples{
pl$DataFrame(a = c(-1, sin(0.5), 0, 1, NA))$
  with_columns(arcsin = pl$col("a")$arcsin())
}
