% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__arcsinh}
\alias{expr__arcsinh}
\title{Compute inverse hyperbolic sine}
\usage{
expr__arcsinh()
}
\value{
A polars \link{expression}
}
\description{
Compute inverse hyperbolic sine
}
\examples{
pl$DataFrame(a = c(-1, sinh(0.5), 0, 1, NA))$
  with_columns(arcsinh = pl$col("a")$arcsinh())
}
