% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__backward_fill}
\alias{expr__backward_fill}
\title{Fill missing values with the next non-null value}
\usage{
expr__backward_fill(limit = NULL)
}
\arguments{
\item{fill}{The number of consecutive null values to backward fill.}
}
\value{
A polars \link{expression}
}
\description{
Fill missing values with the next non-null value
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, NA),
  b = c(4, NA, 6),
  c = c(NA, NA, 2)
)
df$select(pl$all()$backward_fill())
df$select(pl$all()$backward_fill(limit = 1))
}
