% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bin_intervals}
\alias{expr__bin_intervals}
\title{Bin values into discrete intervals delimited by breakpoints}
\usage{
expr__bin_intervals(
  intervals,
  ...,
  labels = NULL,
  include_intervals = FALSE,
  right_closed = FALSE
)
}
\arguments{
\item{intervals}{Numeric breakpoints, or a positive integer giving the number
of equal-width bins.}

\item{...}{These dots are for future extensions and must be empty.}

\item{labels}{Names of the bins. The number of labels must equal the number
of bins. Use \code{NULL} to return \code{UInt32} bin indices.}

\item{include_intervals}{Return a struct with the bin and its left and right
boundaries.}

\item{right_closed}{Use right-closed \verb{(left, right]} bins instead of
left-closed \verb{[left, right)} bins.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(foo = -2:2)
df$select(bin = pl$col("foo")$bin_intervals(
  c(-1, 1), labels = c("low", "mid", "high"), right_closed = TRUE
))
}
