% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bin_ranks}
\alias{expr__bin_ranks}
\title{Bin values by their position in sorted order}
\usage{
expr__bin_ranks(ranks, ..., labels = NULL, include_intervals = FALSE)
}
\arguments{
\item{ranks}{Non-decreasing cumulative fractions in \verb{[0, 1]}, or a positive
integer giving the number of near-equal-sized bins.}

\item{...}{These dots are for future extensions and must be empty.}

\item{labels}{Names of the bins. The number of labels must equal the number
of bins. Use \code{NULL} to return \code{UInt32} bin indices.}

\item{include_intervals}{Return a struct with the bin and its left and right
boundaries.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(foo = -2:2)
df$select(bin = pl$col("foo")$bin_ranks(
  c(0.25, 0.75), labels = c("low", "mid", "high")
))
}
