% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bitwise_and}
\alias{expr__bitwise_and}
\title{Perform an aggregation of bitwise ANDs.}
\usage{
expr__bitwise_and()
}
\value{
A polars \link{expression}
}
\description{
Perform an aggregation of bitwise ANDs.
}
\examples{
df <- pl$DataFrame(n = -1:1)
df$select(pl$col("n")$bitwise_and())

df <- pl$DataFrame(
  grouper = c("a", "a", "a", "b", "b"),
  n = c(-1L, 0L, 1L, -1L, 1L)
)
df$group_by("grouper", .maintain_order = TRUE)$agg(pl$col("n")$bitwise_and())
}
