% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bitwise_count_ones}
\alias{expr__bitwise_count_ones}
\title{Evaluate the number of set bits.}
\usage{
expr__bitwise_count_ones()
}
\value{
A polars \link{expression}
}
\description{
Evaluate the number of set bits.
}
\examples{
df <- pl$DataFrame(n = c(-1L, 0L, 2L, 1L))
df$with_columns(set_bits = pl$col("n")$bitwise_count_ones())
}
