% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bitwise_trailing_ones}
\alias{expr__bitwise_trailing_ones}
\title{Evaluate the number least-significant set bits before seeing an unset bit.}
\usage{
expr__bitwise_trailing_ones()
}
\value{
A polars \link{expression}
}
\description{
Evaluate the number least-significant set bits before seeing an unset bit.
}
\examples{
df <- pl$DataFrame(n = c(-1L, 0L, 2L, 1L))
df$with_columns(trailing_ones = pl$col("n")$bitwise_trailing_ones())
}
