% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__bitwise_trailing_zeros}
\alias{expr__bitwise_trailing_zeros}
\title{Evaluate the number least-significant unset bits before seeing a set bit.}
\usage{
expr__bitwise_trailing_zeros()
}
\value{
A polars \link{expression}
}
\description{
Evaluate the number least-significant unset bits before seeing a set bit.
}
\examples{
df <- pl$DataFrame(n = c(-1L, 0L, 2L, 1L))
df$with_columns(trailing_zeros = pl$col("n")$bitwise_trailing_zeros())
}
