% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cbrt}
\alias{expr__cbrt}
\title{Compute cube root}
\usage{
expr__cbrt()
}
\value{
A polars \link{expression}
}
\description{
Compute cube root
}
\examples{
pl$DataFrame(a = c(1, 2, 4))$
  with_columns(cbrt = pl$col("a")$cbrt())
}
