% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__clip}
\alias{expr__clip}
\title{Set values outside the given boundaries to the boundary value}
\usage{
expr__clip(lower_bound = NULL, upper_bound = NULL)
}
\arguments{
\item{lower_bound}{Lower bound. Accepts expression input. Non-expression
inputs are parsed as literals.}

\item{upper_bound}{Upper bound. Accepts expression input. Non-expression
inputs are parsed as literals.}
}
\value{
A polars \link{expression}
}
\description{
This method only works for numeric and temporal columns. To clip other data
types, consider writing a when-then-otherwise expression.
}
\details{
This method only works for numeric and temporal columns. To clip other data
types, consider writing a when-then-otherwise expression.
}
\examples{
df <- pl$DataFrame(a = c(-50, 5, 50, NA))

# Specifying both a lower and upper bound:
df$with_columns(
  clip = pl$col("a")$clip(1, 10)
)

# Specifying only a single bound:
df$with_columns(
  clip = pl$col("a")$clip(upper_bound = 10)
)
}
