% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cos}
\alias{expr__cos}
\title{Compute cosine}
\usage{
expr__cos()
}
\value{
A polars \link{expression}
}
\description{
Compute cosine
}
\examples{
pl$DataFrame(a = c(0, pi / 2, pi, NA))$
  with_columns(cosine = pl$col("a")$cos())
}
