% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cum_min}
\alias{expr__cum_min}
\title{Return the cumulative min computed at every element.}
\usage{
expr__cum_min(..., reverse = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{reverse}{If \code{TRUE}, start from the last value.}
}
\value{
A polars \link{expression}
}
\description{
Return the cumulative min computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(a = c(1:4, 2L))$with_columns(
  cum_min = pl$col("a")$cum_min(),
  cum_min_reversed = pl$col("a")$cum_min(reverse = TRUE)
)
}
