% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__cut}
\alias{expr__cut}
\title{Bin continuous values into discrete categories}
\usage{
expr__cut(
  breaks,
  ...,
  labels = NULL,
  left_closed = FALSE,
  include_breaks = FALSE
)
}
\arguments{
\item{breaks}{List of unique cut points.}

\item{...}{These dots are for future extensions and must be empty.}

\item{labels}{Names of the categories. The number of labels must be equal to
the number of cut points plus one.}

\item{left_closed}{Set the intervals to be left-closed instead of
right-closed.}

\item{include_breaks}{Include a column with the right endpoint of the bin
each observation falls in. This will change the data type of the output from
a Categorical to a Struct.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Divide a column into three categories.
df <- pl$DataFrame(foo = -2:2)
df$with_columns(
  cut = pl$col("foo")$cut(c(-1, 1), labels = c("a", "b", "c"))
)

# Add both the category and the breakpoint.
df$with_columns(
  cut = pl$col("foo")$cut(c(-1, 1), include_breaks = TRUE)
)$unnest("cut")
}
