% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__drop_nans}
\alias{expr__drop_nans}
\title{Drop all floating point NaN values}
\usage{
expr__drop_nans()
}
\value{
A polars \link{expression}
}
\description{
The original order of the remaining elements is preserved. A \code{NaN} value is
not the same as a \code{null} value. To drop \code{null} values, use
\link[=expr__drop_nulls]{$drop_nulls()}.
}
\examples{
df <- pl$DataFrame(a = c(1, NA, 3, NaN))
df$select(pl$col("a")$drop_nans())
}
