% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__drop_nulls}
\alias{expr__drop_nulls}
\title{Drop all floating point null values}
\usage{
expr__drop_nulls()
}
\value{
A polars \link{expression}
}
\description{
The original order of the remaining elements is preserved. A \code{null} value is
not the same as a \code{NaN} value. To drop \code{NaN} values, use
\link[=expr__drop_nans]{$drop_nans()}.
}
\examples{
df <- pl$DataFrame(a = c(1, NA, 3, NaN))
df$select(pl$col("a")$drop_nulls())
}
