% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__exclude}
\alias{expr__exclude}
\title{Exclude columns from a multi-column expression.}
\usage{
expr__exclude(...)
}
\arguments{
\item{...}{The name or datatype of the column(s) to exclude. Accepts regular
expression input. Regular expressions should start with \code{^} and end with \code{$}.}
}
\value{
A polars \link{expression}
}
\description{
Exclude columns from a multi-column expression.
}
\examples{
df <- pl$DataFrame(aa = 1:2, ba = c("a", NA), cc = c(NA, 2.5))
df

# Exclude by column name(s):
df$select(pl$all()$exclude("ba"))

# Exclude by regex, e.g. removing all columns whose names end with the
# letter "a":
df$select(pl$all()$exclude("^.*a$"))

# Exclude by dtype(s), e.g. removing all columns of type Int64 or Float64:
df$select(pl$all()$exclude(pl$Int64, pl$Float64))
}
