% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__exp}
\alias{expr__exp}
\title{Compute the exponential}
\usage{
expr__exp()
}
\value{
A polars \link{expression}
}
\description{
Compute the exponential
}
\examples{
pl$DataFrame(a = c(1, 2, 4))$
  with_columns(exp = pl$col("a")$exp())
}
