% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__explode}
\alias{expr__explode}
\title{Explode a list expression}
\usage{
expr__explode()
}
\value{
A polars \link{expression}
}
\description{
This means that every item is expanded to a new row.
}
\examples{
df <- pl$DataFrame(
  groups = c("a", "b"),
  values = list(1:2, 3:4)
)

df$select(pl$col("values")$explode())
}
