% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__extend_constant}
\alias{expr__extend_constant}
\title{Extend the Series with \code{n} copies of a value}
\usage{
expr__extend_constant(value, n)
}
\arguments{
\item{value}{A constant literal value or a unit expression with which to
extend the expression result Series. This can be \code{NA} to extend with nulls.}

\item{n}{The number of additional values that will be added.}
}
\value{
A polars \link{expression}
}
\description{
Extend the Series with \code{n} copies of a value
}
\examples{
df <- pl$DataFrame(values = 1:3)
df$select(pl$col("values")$extend_constant(99, n = 2))
}
