% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__fill_nan}
\alias{expr__fill_nan}
\title{Fill floating point \code{NaN} value with a fill value}
\usage{
expr__fill_nan(value)
}
\arguments{
\item{value}{Value used to fill \code{NaN} values.}
}
\value{
A polars \link{expression}
}
\description{
Fill floating point \code{NaN} value with a fill value
}
\examples{
df <- pl$DataFrame(a = c(1, NA, 2, NaN))
df$with_columns(
  filled_nan = pl$col("a")$fill_nan(99)
)
}
