% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__flatten}
\alias{expr__flatten}
\title{Flatten a list or string column}
\usage{
expr__flatten()
}
\value{
A polars \link{expression}
}
\description{
This is an alias for \link[=expr__explode]{$explode()}.
}
\examples{
df <- pl$DataFrame(
  group = c("a", "b", "b"),
  values = list(1:2, 2:3, 4)
)

df$group_by("group")$agg(pl$col("values")$flatten())
}
