% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__floor_div}
\alias{expr__floor_div}
\alias{expr__floordiv}
\title{Floor divide using two expressions}
\usage{
expr__floor_div(other)

expr__floordiv(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of floor division operator \code{expr \%/\% other}.
\verb{$floordiv()} is an alias for \verb{$floor_div()}, which exists for compatibility
with Python Polars.
}
\examples{
df <- pl$DataFrame(x = 1:5)

df$with_columns(
  `x/2` = pl$col("x")$true_div(2),
  `x\%/\%2` = pl$col("x")$floor_div(2)
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
\item \code{\link[=expr__true_div]{<Expr>$true_div()}}
\item \code{\link[=expr__mod]{<Expr>$mod()}}
}
}
