% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__forward_fill}
\alias{expr__forward_fill}
\title{Fill missing values with the last non-null value}
\usage{
expr__forward_fill(limit = NULL)
}
\arguments{
\item{fill}{The number of consecutive null values to forward fill.}
}
\value{
A polars \link{expression}
}
\description{
Fill missing values with the last non-null value
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, NA),
  b = c(4, NA, 6),
  c = c(2, NA, NA)
)
df$select(pl$all()$forward_fill())
df$select(pl$all()$forward_fill(limit = 1))
}
