% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__gather_every}
\alias{expr__gather_every}
\title{Take every \code{n}-th value in the Series and return as a new Series}
\usage{
expr__gather_every(n, offset = 0)
}
\arguments{
\item{n}{Gather every n-th row.}

\item{offset}{Starting index.}
}
\value{
A polars \link{expression}
}
\description{
Take every \code{n}-th value in the Series and return as a new Series
}
\examples{
df <- pl$DataFrame(foo = 1:9)
df$select(pl$col("foo")$gather_every(3))
df$select(pl$col("foo")$gather_every(3, offset = 1))
}
