% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__get}
\alias{expr__get}
\title{Return a single value by index}
\usage{
expr__get(index, ..., null_on_oob = FALSE)
}
\arguments{
\item{index}{An expression that leads to a UInt32 dtyped Series.}

\item{...}{These dots are for future extensions and must be empty.}

\item{null_on_oob}{If \code{TRUE}, return \code{null} if an index is out of bounds.
Otherwise, raise an error.}
}
\value{
A polars \link{expression}
}
\description{
Return a single value by index
}
\examples{
df <- pl$DataFrame(
  group = c("one", "one", "one", "two", "two", "two"),
  value = c(1, 98, 2, 3, 99, 4)
)
df$group_by("group", maintain_order = TRUE)$agg(
  pl$col("value")$get(1)
)
}
