% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__get}
\alias{expr__get}
\title{Return a single value by index}
\usage{
expr__get(index)
}
\arguments{
\item{index}{An expression that leads to a UInt32 dtyped Series.}
}
\value{
A polars \link{expression}
}
\description{
Return a single value by index
}
\examples{
df <- pl$DataFrame(
  group = c("one", "one", "one", "two", "two", "two"),
  value = c(1, 98, 2, 3, 99, 4)
)
df$group_by("group", maintain_order = TRUE)$agg(
  pl$col("value")$get(1)
)
}
