% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__has_nulls}
\alias{expr__has_nulls}
\title{Check whether the expression contains one or more null values}
\usage{
expr__has_nulls()
}
\value{
A polars \link{expression}
}
\description{
Check whether the expression contains one or more null values
}
\examples{
df <- pl$DataFrame(
  a = c(NA, 1, NA),
  b = c(10, NA, 300),
  c = c(350, 650, 850)
)
df$select(pl$all()$has_nulls())
}
