% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__implode}
\alias{expr__implode}
\title{Aggregate values into a list}
\usage{
expr__implode()
}
\value{
A polars \link{expression}
}
\description{
Aggregate values into a list
}
\examples{
df <- pl$DataFrame(a = 1:3, b = 4:6)
df$with_columns(pl$col("a")$implode())
}
