% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__index_of}
\alias{expr__index_of}
\title{Get the index of the first occurrence of a value, or \code{NA} if it's not found}
\usage{
expr__index_of(element)
}
\arguments{
\item{element}{Value to find.}
}
\value{
A polars \link{expression}
}
\description{
Get the index of the first occurrence of a value, or \code{NA} if it's not found
}
\examples{
df <- pl$DataFrame(a = c(1, NA, 17))
df$select(
  seventeen = pl$col("a")$index_of(17),
  null = pl$col("a")$index_of(NA),
  fiftyfive = pl$col("a")$index_of(55),
)
}
