% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_close}
\alias{expr__is_close}
\title{Check if this expression is close, i.e. almost equal, to the other expression}
\usage{
expr__is_close(other, ..., abs_tol = 0, rel_tol = 1e-09, nans_equal = FALSE)
}
\arguments{
\item{other}{A literal or expression value to compare with.}

\item{...}{These dots are for future extensions and must be empty.}

\item{abs_tol}{Absolute tolerance. This is the maximum allowed absolute
difference between two values. Must be non-negative.}

\item{rel_tol}{Relative tolerance. This is the maximum allowed difference
between two values, relative to the larger absolute value. Must be
non-negative.}

\item{nans_equal}{Whether \code{NaN} values should be considered equal.}
}
\value{
A polars \link{expression}
}
\description{
Two values \code{a} and \code{b} are considered close if the following condition holds:
\eqn{|a - b| \le \max \{ \text{rel\_tol} \cdot \max \{ |a|, |b| \}, \text{abs\_tol} \}}
}
\examples{
df <- pl$DataFrame(a = c(1.5, 2.0, 2.5), b = c(1.55, 2.2, 3.0))
df$with_columns(
  is_close = pl$col("a")$is_close("b", abs_tol = 0.1)
)
}
