% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_finite}
\alias{expr__is_finite}
\title{Check if elements are finite}
\usage{
expr__is_finite()
}
\value{
A polars \link{expression}
}
\description{
Check if elements are finite
}
\examples{
df <- pl$DataFrame(a = c(1, 2), b = c(3, Inf))
df$with_columns(
  a_finite = pl$col("a")$is_finite(),
  b_finite = pl$col("b")$is_finite()
)
}
