% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_first_distinct}
\alias{expr__is_first_distinct}
\title{Return a boolean mask indicating the first occurrence of each distinct value}
\usage{
expr__is_first_distinct()
}
\value{
A polars \link{expression}
}
\description{
Return a boolean mask indicating the first occurrence of each distinct value
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2, 3, 2))
df$with_columns(
  is_first_distinct = pl$col("a")$is_first_distinct()
)
}
