% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_in}
\alias{expr__is_in}
\title{Check if elements of an expression are present in another expression}
\usage{
expr__is_in(other)
}
\arguments{
\item{other}{Accepts expression input. Strings are parsed as column names.}
}
\value{
A polars \link{expression}
}
\description{
Check if elements of an expression are present in another expression
}
\examples{
df <- pl$DataFrame(
  sets = list(1:3, 1:2, 9:10),
  optional_members = 1:3
)
df$with_columns(
  contains = pl$col("optional_members")$is_in("sets")
)
}
