% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__is_infinite}
\alias{expr__is_infinite}
\title{Check if elements are infinite}
\usage{
expr__is_infinite()
}
\value{
A polars \link{expression}
}
\description{
Check if elements are infinite
}
\examples{
df <- pl$DataFrame(a = c(1, 2), b = c(3, Inf))
df$with_columns(
  a_infinite = pl$col("a")$is_infinite(),
  b_infinite = pl$col("b")$is_infinite()
)
}
