% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__limit}
\alias{expr__limit}
\title{Get the first n rows}
\usage{
expr__limit(n = 10)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A polars \link{expression}
}
\description{
This is an alias for \link[=expr__head]{$head()}.
}
\examples{
df <- pl$DataFrame(a = 1:9)
df$select(pl$col("a")$limit(3))
}
