% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__log10}
\alias{expr__log10}
\title{Compute the base-10 logarithm}
\usage{
expr__log10()
}
\value{
A polars \link{expression}
}
\description{
Compute the base-10 logarithm
}
\examples{
pl$DataFrame(a = c(1, 2, 4))$
  with_columns(log10 = pl$col("a")$log10())
}
