% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__log1p}
\alias{expr__log1p}
\title{Compute the natural logarithm plus one}
\usage{
expr__log1p()
}
\value{
A polars \link{expression}
}
\description{
This computes \code{log(1 + x)} but is more numerically stable for \code{x} close to
zero.
}
\examples{
pl$DataFrame(a = c(1, 2, 4))$
  with_columns(log1p = pl$col("a")$log1p())
}
