% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__lt}
\alias{expr__lt}
\title{Check strictly lower inequality}
\usage{
expr__lt(other)
}
\arguments{
\item{other}{A literal or expression value to compare with.}
}
\value{
A polars \link{expression}
}
\description{
Check strictly lower inequality
}
\examples{
df <- pl$DataFrame(x = 1:3)
df$with_columns(
  with_lt = pl$col("x")$lt(pl$lit(2)),
  with_symbol = pl$col("x") < pl$lit(2)
)
}
