% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__map_batches}
\alias{expr__map_batches}
\title{Apply a custom R function to a whole Series or sequence of Series.}
\usage{
expr__map_batches(lambda, return_dtype = NULL, ...)
}
\arguments{
\item{lambda}{Function to apply.}

\item{return_dtype}{Dtype of the output Series.
It is recommended to set this whenever possible. If this is \code{NULL}, it tries
to infer the datatype by calling the function with dummy data and looking at
the output.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The output of this custom function is presumed to be either a Series, or an
R vector that will be converted into a Series by \code{\link[=as_polars_series]{as_polars_series()}}.
}
\examples{
df <- pl$DataFrame(
  sine = c(0.0, 1.0, 0.0, -1.0),
  cosine = c(1.0, 0.0, -1.0, 0.0)
)
df$select(pl$all()$map_batches(\\(x) \{
  x$to_r_vector() |>
    which.max()
\}))

# Call a function that takes multiple arguments by creating a struct and
# referencing its fields inside the function call.
df <- pl$DataFrame(
  a = c(5, 1, 0, 3),
  b = c(4, 2, 3, 4),
)
df$with_columns(
  a_times_b = pl$struct("a", "b")$map_batches(
    \\(x) x$struct$field("a") * x$struct$field("b")
  )
)
}
