% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__max_by}
\alias{expr__max_by}
\title{Get maximum value, ordered by another expression}
\usage{
expr__max_by(by)
}
\arguments{
\item{by}{Column used to determine the largest element. Accepts expression
input. Strings are parsed as column names.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
df <- pl$DataFrame(a = c(-1, NaN, 1), b = c("x", "y", "z"))
df$with_columns(max_b_by_a = pl$col("b")$max_by("a"))
}
