% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__mod}
\alias{expr__mod}
\title{Modulo using two expressions}
\usage{
expr__mod(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of modulus operator \code{expr \%\% other}.
}
\examples{
df <- pl$DataFrame(x = -5L:5L)

df$with_columns(
  `x\%\%2` = pl$col("x")$mod(2)
)
}
\seealso{
\itemize{
\item \link[=s3-arithmetic]{Arithmetic operators}
\item \code{\link[=expr__floor_div]{<Expr>$floor_div()}}
}
}
