% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__mode}
\alias{expr__mode}
\title{Compute the most occurring value(s)}
\usage{
expr__mode()
}
\value{
A polars \link{expression}
}
\description{
Compute the most occurring value(s)
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2, 3), b = c(1, 1, 2, 2))
df$select(pl$col("a")$mode())
df$select(pl$col("b")$mode())
}
