% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__null_count}
\alias{expr__null_count}
\title{Count null values}
\usage{
expr__null_count()
}
\value{
A polars \link{expression}
}
\description{
Count null values
}
\examples{
df <- pl$DataFrame(
  a = c(NA, 1, NA),
  b = c(10, NA, 300),
  c = c(1, 2, 2)
)
df$select(pl$all()$null_count())
}
