% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__or}
\alias{expr__or}
\title{Apply logical OR on two expressions}
\usage{
expr__or(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> One or more integer or
boolean expressions to evaluate/combine.}
}
\value{
A polars \link{expression}
}
\description{
Combine two boolean expressions with OR.
}
\examples{
df <- pl$DataFrame(
  x = c(5, 6, 7, 4, 8),
  y = c(1.5, 2.5, 1.0, 4.0, -5.75),
  z = c(-9, 2, -1, 4, 8),
)
df$with_columns(
  (pl$col("x") == pl$col("y"))$or(
    pl$col("y") == pl$col("z"),
    pl$col("y")$cast(pl$Int32) == pl$col("z"),
  )$alias("any")
)
}
