% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__qcut}
\alias{expr__qcut}
\title{Bin continuous values into discrete categories based on their quantiles}
\usage{
expr__qcut(
  quantiles,
  ...,
  labels = NULL,
  left_closed = FALSE,
  allow_duplicates = FALSE,
  include_breaks = FALSE
)
}
\arguments{
\item{quantiles}{Either a vector of quantile probabilities between 0 and 1
or a positive integer determining the number of bins with uniform
probability.}

\item{...}{These dots are for future extensions and must be empty.}

\item{labels}{Names of the categories. The number of labels must be equal to
the number of categories.}

\item{left_closed}{Set the intervals to be left-closed instead of
right-closed.}

\item{allow_duplicates}{If \code{TRUE}, duplicates in the resulting quantiles are
dropped, rather than raising an error. This can happen even with unique
probabilities, depending on the data.}

\item{include_breaks}{Include a column with the right endpoint of the bin
each observation falls in. This will change the data type of the output from
a Categorical to a Struct.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Divide a column into three categories according to pre-defined quantile
# probabilities.
df <- pl$DataFrame(foo = -2:2)
df$with_columns(
  qcut = pl$col("foo")$qcut(c(0.25, 0.75), labels = c("a", "b", "c"))
)

# Divide a column into two categories using uniform quantile probabilities.
df$with_columns(
  qcut = pl$col("foo")$qcut(2, labels = c("low", "high"), left_closed = TRUE)
)

# Add both the category and the breakpoint.
df$with_columns(
  qcut = pl$col("foo")$qcut(c(0.25, 0.75), include_breaks = TRUE)
)$unnest("qcut")
}
