% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rank}
\alias{expr__rank}
\title{Assign ranks to data, dealing with ties appropriately}
\usage{
expr__rank(
  method = c("average", "min", "max", "dense", "ordinal", "random"),
  ...,
  descending = FALSE,
  seed = NULL
)
}
\arguments{
\item{method}{The method used to assign ranks to tied elements. Must be one
of the following:
\itemize{
\item \code{"average"} (default): The average of the ranks that would have been
assigned to all the tied values is assigned to each value.
\item \code{"min"}: The minimum of the ranks that would have been assigned to all
the tied values is assigned to each value. (This is also referred to
as "competition" ranking.)
\item \code{"max"} : The maximum of the ranks that would have been assigned to all
the tied values is assigned to each value.
\item \code{"dense"}: Like 'min', but the rank of the next highest element is assigned
the rank immediately after those assigned to the tied elements.
\item \code{"ordinal"} : All values are given a distinct rank, corresponding to the
order that the values occur in the Series.
\item \code{"random"} : Like 'ordinal', but the rank for ties is not dependent on the
order that the values occur in the Series.
}}

\item{...}{These dots are for future extensions and must be empty.}

\item{descending}{Rank in descending order.}

\item{seed}{Integer. Only used if \code{method = "random"}.}
}
\value{
A polars \link{expression}
}
\description{
Assign ranks to data, dealing with ties appropriately
}
\examples{
# Default is to use the "average" method to break ties
df <- pl$DataFrame(a = c(3, 6, 1, 1, 6))
df$with_columns(rank = pl$col("a")$rank())

# Ordinal method
df$with_columns(rank = pl$col("a")$rank("ordinal"))

# Use "rank" with "over" to rank within groups:
df <- pl$DataFrame(
  a = c(1, 1, 2, 2, 2),
  b = c(6, 7, 5, 14, 11)
)
df$with_columns(
  rank = pl$col("b")$rank()$over("a")
)
}
