% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rechunk}
\alias{expr__rechunk}
\title{Create a single chunk of memory for this Series}
\usage{
expr__rechunk()
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\verb{<expr>$rechunk()} is deprecated. Rechunking within a query is not
well-defined. Use \verb{$rechunk()} on the DataFrame after collecting the results
instead.
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2))

# Create a Series with 3 nulls, append column a then rechunk
df$select(pl$repeat_(NA, 3)$append(pl$col("a"))$rechunk())
}
