% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__reinterpret}
\alias{expr__reinterpret}
\title{Reinterpret the underlying bits of a same-size numeric type.}
\usage{
expr__reinterpret(..., signed = NULL, dtype = NULL)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{signed}{Whether to reinterpret as a signed or unsigned integer of the
same width as the input. Must specify exactly one of \code{signed} and \code{dtype}.}

\item{dtype}{Target data type for the reinterpretation. Must specify exactly
one of \code{signed} and \code{dtype}.}
}
\value{
A polars \link{expression}
}
\description{
Reinterpret the underlying bits of a same-size numeric type.
}
\examples{
df <- pl$DataFrame(a = c(1, 1, 2))$cast(pl$UInt64)

# Reinterpret column a as Int64
df$with_columns(
  reinterpreted = pl$col("a")$reinterpret(dtype = pl$Int64)
)
}
