% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__reshape}
\alias{expr__reshape}
\title{Reshape this Expr to a flat column or an Array column}
\usage{
expr__reshape(dimensions)
}
\arguments{
\item{dimensions}{A integer vector of length of the dimension size.
If \code{-1} is used as the value for the first dimension, that dimension is inferred.
Because the size of the Column may not be known in advance,
it is only possible to use -1 for the first dimension.}
}
\value{
A polars \link{expression}
}
\description{
Reshape this Expr to a flat column or an Array column
}
\details{
If a single dimension is given, results in an expression of the original data
type. If a multiple dimensions are given, results in an expression of data
type Array with shape dimensions.
}
\examples{
df <- pl$DataFrame(foo = 1:8)

df$select(pl$col("foo")$reshape(8))
df$select(pl$col("foo")$reshape(c(2, 4)))

# Using `-1` for the first dimension to infer the other dimension
df$select(pl$col("foo")$reshape(c(-1, 4)))
df$select(pl$col("foo")$reshape(c(-1, 2)))

# We can have more than 2 dimensions
df$select(pl$col("foo")$reshape(c(2, 2, 2)))
df$select(pl$col("foo")$reshape(c(-1, 2, 2)))
}
