% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__reshape}
\alias{expr__reshape}
\title{Reshape this Expr to a flat Series or a Series of Lists}
\usage{
expr__reshape(dimensions)
}
\arguments{
\item{dimensions}{A integer vector of length of the dimension size.
If \code{-1} is used in any of the dimensions, that dimension is inferred.
Currently, more than two dimensions not supported.}
}
\value{
A polars \link{expression}
}
\description{
Reshape this Expr to a flat Series or a Series of Lists
}
\details{
If a single dimension is given, results in an expression of the original data
type. If a multiple dimensions are given, results in an expression of data
type List with shape equal to the dimensions.
}
\examples{
df <- pl$DataFrame(foo = 1:9)

df$select(pl$col("foo")$reshape(9))
df$select(pl$col("foo")$reshape(c(3, 3)))

# Use `-1` to infer the other dimension
df$select(pl$col("foo")$reshape(c(-1, 3)))
df$select(pl$col("foo")$reshape(c(3, -1)))

# One can specify more than 2 dimensions by using the Array type
df <- pl$DataFrame(foo = 1:12)
df$select(
  pl$col("foo")$reshape(c(3, 2, 2), nested_type = pl$Array(pl$Float32, 2))
)
}
