% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__reverse}
\alias{expr__reverse}
\title{Reverse an expression}
\usage{
expr__reverse()
}
\value{
A polars \link{expression}
}
\description{
Reverse an expression
}
\examples{
df <- pl$DataFrame(
  a = 1:5,
  fruits = c("banana", "banana", "apple", "apple", "banana"),
  b = 5:1
)

df$with_columns(
  pl$all()$reverse()$name$suffix("_reverse")
)
}
