% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__rolling_quantile_by}
\alias{expr__rolling_quantile_by}
\title{Apply a rolling quantile based on another column}
\usage{
expr__rolling_quantile_by(
  by,
  window_size,
  ...,
  quantile,
  interpolation = c("nearest", "higher", "lower", "midpoint", "linear"),
  min_periods = 1,
  closed = c("right", "both", "left", "none")
)
}
\arguments{
\item{by}{This column must be of dtype Datetime or Date. Accepts expression
input. Strings are parsed as column names.}

\item{window_size}{The length of the window. Can be a dynamic temporal size
indicated by a timedelta or the following string language:
\itemize{
\item 1ns (1 nanosecond)
\item 1us (1 microsecond)
\item 1ms (1 millisecond)
\item 1s (1 second)
\item 1m (1 minute)
\item 1h (1 hour)
\item 1d (1 calendar day)
\item 1w (1 calendar week)
\item 1mo (1 calendar month)
\item 1q (1 calendar quarter)
\item 1y (1 calendar year)
}

Or combine them: \code{"3d12h4m25s"} # 3 days, 12 hours, 4 minutes, and 25 seconds

By "calendar day", we mean the corresponding time on the next day
(which may not be 24 hours, due to daylight savings). Similarly for
"calendar week", "calendar month", "calendar quarter", and "calendar year".}

\item{quantile}{Quantile between 0.0 and 1.0.}

\item{interpolation}{Interpolation method. Must be one of \code{"nearest"},
\code{"higher"}, \code{"lower"}, \code{"midpoint"}, \code{"linear"}.}

\item{min_periods}{The number of values in the window that should be
non-null before computing a result. If \code{NULL} (default), it will be set
equal to \code{window_size}.}

\item{closed}{Define which sides of the interval are closed (inclusive).
Default is \code{"right"}.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Given a \code{by} column \verb{<t_0, t_1, ..., t_n>}, then \code{closed = "right"} (the
default) means the windows will be:
\itemize{
\item \verb{(t_0 - window_size, t_0]}
\item \verb{(t_1 - window_size, t_1]}
\item …
\item \verb{(t_n - window_size, t_n]}
}
}
\details{
If you want to compute multiple aggregation statistics over the same dynamic
window, consider using \code{\link[=expr__rolling]{$rolling()}} - this method can cache
the window size computation.
}
\examples{
df_temporal <- pl$select(
  index = 0:24,
  date = pl$datetime_range(
    as.POSIXct("2001-01-01"),
    as.POSIXct("2001-01-02"),
    "1h"
  )
)

# Compute the rolling quantile with the temporal windows closed on the right
# (default)
df_temporal$with_columns(
  rolling_row_quantile = pl$col("index")$rolling_quantile_by(
    "date",
    window_size = "2h"
  )
)

# Compute the rolling quantile with the closure of windows on both sides
df_temporal$with_columns(
  rolling_row_quantile = pl$col("index")$rolling_quantile_by(
    "date",
    window_size = "2h",
    closed = "both"
  )
)
}
