% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__round}
\alias{expr__round}
\title{Round underlying floating point data by decimals digits}
\usage{
expr__round(decimals = 0L, mode = c("half_to_even", "half_away_from_zero"))
}
\arguments{
\item{decimals}{Number of decimals to round by.}

\item{mode}{Rounding mode. One of the following:
\itemize{
\item \code{"half_to_even"} (default): round to the nearest even number;
\item \code{"half_away_from_zero"}: round to the nearest number away from zero.
}}
}
\value{
A polars \link{expression}
}
\description{
Round underlying floating point data by decimals digits
}
\examples{
df <- pl$DataFrame(a = c(0.33, 0.52, 1.02, 1.17))
df$select(pl$col("a")$round(1))

df <- pl$DataFrame(
  f64 = c(-3.5, -2.5, -1.5, -0.5, 0.5, 1.5, 2.5, 3.5),
  d = c("-3.5", "-2.5", "-1.5", "-0.5", "0.5", "1.5", "2.5", "3.5")
)$cast(d = pl$Decimal(scale = 1))

df$with_columns(
  pl$all()$round(mode = "half_away_from_zero")$name$suffix("_away"),
  pl$all()$round(mode = "half_to_even")$name$suffix("_to_even"),
)
}
