% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__round}
\alias{expr__round}
\title{Round underlying floating point data by decimals digits}
\usage{
expr__round(decimals)
}
\arguments{
\item{decimals}{Number of decimals to round by.}
}
\value{
A polars \link{expression}
}
\description{
Round underlying floating point data by decimals digits
}
\examples{
df <- pl$DataFrame(a = c(0.33, 0.52, 1.02, 1.17))

df$with_columns(
  rounded = pl$col("a")$round(1)
)
}
