% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__round_sig_figs}
\alias{expr__round_sig_figs}
\title{Round to a number of significant figures}
\usage{
expr__round_sig_figs(digits)
}
\arguments{
\item{digits}{Number of significant figures to round to.}
}
\value{
A polars \link{expression}
}
\description{
Round to a number of significant figures
}
\examples{
df <- pl$DataFrame(a = c(0.01234, 3.333, 1234))

df$with_columns(
  rounded = pl$col("a")$round_sig_figs(2)
)
}
