% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__search_sorted}
\alias{expr__search_sorted}
\title{Find indices where elements should be inserted to maintain order}
\usage{
expr__search_sorted(element, side = c("any", "left", "right"))
}
\arguments{
\item{element}{Expression or scalar value.}

\item{side}{Must be one of the following:
\itemize{
\item \code{"any"}: the index of the first suitable location found is given;
\item \code{"left"}: the index of the leftmost suitable location found is given;
\item \code{"right"}: the index the rightmost suitable location found is given.
}}
}
\value{
A polars \link{expression}
}
\description{
This returns -1 if x is lower than 0, 0 if x == 0, and 1 if x is greater
than 0.
}
\examples{
df <- pl$DataFrame(values = c(1, 2, 3, 5))
df$select(
  zero = pl$col("values")$search_sorted(0),
  three = pl$col("values")$search_sorted(3),
  six = pl$col("values")$search_sorted(6),
)
}
