% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__set_sorted}
\alias{expr__set_sorted}
\title{Flags the expression as "sorted"}
\usage{
expr__set_sorted(..., descending = FALSE, nulls_last = !descending)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{descending}{Whether the Series order is descending.}

\item{nulls_last}{Whether the nulls are at the end.}
}
\value{
A polars \link{expression}
}
\description{
Enables downstream code to user fast paths for sorted arrays.

\strong{Warning:} This can lead to incorrect results if the data is NOT sorted!!
Use with care!
}
\examples{
df <- pl$DataFrame(a = 1:3)
df$select(pl$col("a")$set_sorted()$max())
}
