% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__shift}
\alias{expr__shift}
\title{Shift values by the given number of indices}
\usage{
expr__shift(n = 1, ..., fill_value = NULL)
}
\arguments{
\item{n}{Number of indices to shift forward. If a negative value is
passed, values are shifted in the opposite direction instead.}

\item{...}{These dots are for future extensions and must be empty.}

\item{fill_value}{Fill the resulting null values with this value. In Polars
1.16, a bare character string is interpreted as a column and emits a
warning. In Polars 2.0, bare strings are literals; use \code{pl$col()} to
preserve the old column behavior or \code{pl$lit()} for the new behavior.}
}
\value{
A polars \link{expression}
}
\description{
Shift values by the given number of indices
}
\examples{
# By default, values are shifted forward by one index.
df <- pl$DataFrame(a = 1:4)
df$with_columns(shift = pl$col("a")$shift())

# Pass a negative value to shift in the opposite direction instead.
df$with_columns(shift = pl$col("a")$shift(-2))

# Specify fill_value to fill the resulting null values.
df$with_columns(shift = pl$col("a")$shift(-2, fill_value = 100))
}
