% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__shrink_dtype}
\alias{expr__shrink_dtype}
\title{Shrink numeric columns to the minimal required datatype}
\usage{
expr__shrink_dtype()
}
\value{
A polars \link{expression}
}
\description{
Shrink to the dtype needed to fit the extrema of this Series. This can be
used to reduce memory pressure.
}
\examples{
df <- pl$DataFrame(a = c(-112, 2, 112))$cast(pl$Int64)
df$with_columns(
  shrunk = pl$col("a")$shrink_dtype()
)
}
