% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__shuffle}
\alias{expr__shuffle}
\title{Shuffle the contents of this expression}
\usage{
expr__shuffle(seed = NULL)
}
\arguments{
\item{seed}{Integer indicating the seed for the random number generator. If
\code{NULL} (default), a random seed is generated each time the shuffle is called.}
}
\value{
A polars \link{expression}
}
\description{
Note this is shuffled independently of any other column or Expression.
If you want each row to stay the same use
\code{\link[=dataframe_sample]{df$sample(shuffle = TRUE)}}.
}
\examples{
df <- pl$DataFrame(a = 1:3)
df$with_columns(
  shuffled = pl$col("a")$shuffle(seed = 1)
)
}
