% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sort_by}
\alias{expr__sort_by}
\title{Sort this column by the ordering of another column, or multiple other
columns.}
\usage{
expr__sort_by(
  ...,
  descending = FALSE,
  nulls_last = FALSE,
  multithreaded = TRUE,
  maintain_order = FALSE
)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column(s) to sort by. Accepts
expression input. Strings are parsed as column names.}

\item{descending}{Sort in descending order. When sorting by multiple
columns, can be specified per column by passing a sequence of booleans.}

\item{nulls_last}{Place null values last; can specify a single boolean
applying to all columns or a sequence of booleans for per-column control.}

\item{multithreaded}{Sort using multiple threads.}

\item{maintain_order}{Whether the order should be maintained if elements are
equal.}
}
\value{
A polars \link{expression}
}
\description{
If used in a groupby context, values within each group are sorted.
}
\examples{
df <- pl$DataFrame(
  group = c("a", "a", "b", "b"),
  value1 = c(1, 3, 4, 2),
  value2 = c(8, 7, 6, 5)
)

# by one column/expression
df$with_columns(
  sorted = pl$col("group")$sort_by("value1")
)

# by two columns/expressions
df$with_columns(
  sorted = pl$col("group")$sort_by(
    "value2", pl$col("value1"),
    descending = c(TRUE, FALSE)
  )
)

# by some expression
df$with_columns(
  sorted = pl$col("group")$sort_by(pl$col("value1") + pl$col("value2"))
)

# in an aggregation context, values are sorted within groups
df$group_by("group")$agg(
  pl$col("value1")$sort_by("value2")
)
}
