% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sqrt}
\alias{expr__sqrt}
\title{Compute square root}
\usage{
expr__sqrt()
}
\value{
A polars \link{expression}
}
\description{
Compute square root
}
\examples{
pl$DataFrame(a = c(1, 2, 4))$
  with_columns(sqrt = pl$col("a")$sqrt())
}
