% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__std}
\alias{expr__std}
\title{Compute the standard deviation}
\usage{
expr__std(ddof = 1)
}
\value{
A polars \link{expression}
}
\description{
Compute the standard deviation
}
\examples{
pl$DataFrame(a = c(1, 3, 5, 6))$
  select(pl$all()$std())
}
