% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__sub}
\alias{expr__sub}
\title{Substract two expressions}
\usage{
expr__sub(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
A polars \link{expression}
}
\description{
Method equivalent of subtraction operator \code{expr - other}.
}
\examples{
df <- pl$DataFrame(x = 0:4)

df$with_columns(
  `x-2` = pl$col("x")$sub(2),
  `x-expr` = pl$col("x")$sub(pl$col("x")$cum_sum())
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
}
}
