% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__tail}
\alias{expr__tail}
\title{Get the last n elements}
\usage{
expr__tail(n = 10)
}
\arguments{
\item{n}{Number of elements to take.}
}
\value{
A polars \link{expression}
}
\description{
Get the last n elements
}
\examples{
pl$DataFrame(x = 1:11)$select(pl$col("x")$tail(3))
}
