% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__tan}
\alias{expr__tan}
\title{Compute tangent}
\usage{
expr__tan()
}
\value{
A polars \link{expression}
}
\description{
Compute tangent
}
\examples{
pl$DataFrame(a = c(0, pi / 2, pi, NA))$
  with_columns(tangent = pl$col("a")$tan())
}
