% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__to_physical}
\alias{expr__to_physical}
\title{Cast to physical representation of the logical dtype}
\usage{
expr__to_physical()
}
\value{
A polars \link{expression}
}
\description{
The following data types will be changed:
\itemize{
\item Date -> Int32
\item Datetime -> Int64
\item Time -> Int64
\item Duration -> Int64
\item Categorical -> UInt32
\item List(inner) -> List(physical of inner)
}

Other data types will be left unchanged.
}
\examples{
df <- pl$DataFrame(a = factor(c("a", "x", NA, "a")))
df$with_columns(
  phys = pl$col("a")$to_physical()
)
}
