% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__truncate}
\alias{expr__truncate}
\title{Truncate numeric data toward zero to \code{decimals} number of decimal places}
\usage{
expr__truncate(decimals = 0L)
}
\arguments{
\item{decimals}{Number of decimal places to truncate to.}
}
\value{
A polars \link{expression}
}
\description{
Truncate numeric data toward zero to \code{decimals} number of decimal places
}
\details{
Truncation discards the fractional part beyond the given number of decimals.
For example, when truncating to 0 decimals, 0.25, -0.25, 0.99, and -0.99 will
all round to 0. When truncating to 1 decimal 1.9999 rounds to 1.9 and -1.9999
rounds to -1.9. There is no tiebreak behaviour at midpoint values as there is
with \code{\link[=expr__round]{$round()}} so 0.5 and -0.5 will also round to 0 when
\code{decimals = 1}.
}
\examples{
df <- pl$DataFrame(n = c(-9.9999, 0.12345, 1.0251, 8.8765))
df$with_columns(
  t0 = pl$col("n")$truncate(0),
  t1 = pl$col("n")$truncate(1),
  t2 = pl$col("n")$truncate(2),
  t3 = pl$col("n")$truncate(3),
  t4 = pl$col("n")$truncate(4),
)
}
