% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__upper_bound}
\alias{expr__upper_bound}
\title{Calculate the upper bound}
\usage{
expr__upper_bound()
}
\value{
A polars \link{expression}
}
\description{
Returns a unit Series with the highest value possible for the dtype of this
expression.
}
\examples{
df <- pl$DataFrame(a = 1:3)
df$select(pl$col("a")$upper_bound())
}
