% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__value_counts}
\alias{expr__value_counts}
\title{Count the occurrences of unique values}
\usage{
expr__value_counts(
  ...,
  sort = FALSE,
  parallel = FALSE,
  name = NULL,
  normalize = FALSE
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{sort}{Sort the output by count in descending order. If \code{FALSE}
(default), the order of the output is random.}

\item{parallel}{Execute the computation in parallel. This option should
likely not be enabled in a group by context, as the computation is already
parallelized per group.}

\item{name}{Give the resulting count field a specific name. If \code{normalize}
is \code{TRUE} it defaults to \code{"proportion"}, otherwise it defaults to \code{"count"}.}

\item{normalize}{If \code{TRUE}, gives relative frequencies of the unique values.}
}
\value{
A polars \link{expression}
}
\description{
Count the occurrences of unique values
}
\examples{
df <- pl$DataFrame(color = c("red", "blue", "red", "green", "blue", "blue"))
df$select(pl$col("color")$value_counts())

# Sort the output by (descending) count and customize the count field name.
df <- df$select(pl$col("color")$value_counts(sort = TRUE, name = "n"))
df

df$unnest("color")
}
