% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__var}
\alias{expr__var}
\title{Compute the variance}
\usage{
expr__var(ddof = 1)
}
\value{
A polars \link{expression}
}
\description{
Compute the variance
}
\examples{
pl$DataFrame(a = c(1, 3, 5, 6))$
  select(pl$all()$var())
}
