% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{expr__xor}
\alias{expr__xor}
\title{Apply logical XOR on two expressions}
\usage{
expr__xor(other)
}
\arguments{
\item{other}{Element to add. Can be a string (only if \code{expr} is a string), a
numeric value or an other expression.}
}
\value{
A polars \link{expression}
}
\description{
Combine two boolean expressions with XOR.
}
\examples{
pl$lit(TRUE)$xor(pl$lit(FALSE))
}
