% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_abs}
\alias{Expr_abs}
\title{Compute the absolute values}
\usage{
Expr_abs()
}
\value{
Expr
}
\description{
Compute the absolute values
}
\examples{
pl$DataFrame(a = -1:1)$
  with_columns(abs = pl$col("a")$abs())
}
