% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_add}
\alias{Expr_add}
\title{Add two expressions}
\usage{
Expr_add(other)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
df = pl$DataFrame(x = 1:5)

df$with_columns(
  `x+int` = pl$col("x")$add(2L),
  `x+expr` = pl$col("x")$add(pl$col("x")$cum_prod())
)

df = pl$DataFrame(
  x = c("a", "d", "g"),
  y = c("b", "e", "h"),
  z = c("c", "f", "i")
)

df$with_columns(
  pl$col("x")$add(pl$col("y"))$add(pl$col("z"))$alias("xyz")
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
}
}
