% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_add}
\alias{Expr_add}
\alias{+.Expr}
\title{Add}
\usage{
Expr_add(other)

\method{+}{Expr}(e1, e2)
}
\arguments{
\item{other}{literal or Robj which can become a literal}

\item{e1}{lhs Expr}

\item{e2}{rhs Expr or anything which can become a literal Expression}
}
\value{
Exprs
}
\description{
Addition
}
\examples{
#three syntaxes same result
pl$lit(5) + 10
pl$lit(5) + pl$lit(10)
pl$lit(5)$add(pl$lit(10))
+pl$lit(5) #unary use resolves to same as pl$lit(5)
}
\keyword{Expr}
\keyword{Expr_operators}
