% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_agg_groups}
\alias{Expr_agg_groups}
\title{Aggregate groups}
\usage{
<Expr>$agg_groups
}
\value{
Expr
}
\description{
Get the group indexes of the group by operation. Should be used in aggregation
context only.
}
\examples{
df = pl$DataFrame(list(
  group = c("one", "one", "one", "two", "two", "two"),
  value = c(94, 95, 96, 97, 97, 99)
))
df$group_by("group", maintain_order = TRUE)$agg(pl$col("value")$agg_groups())
}
\keyword{datasets}
