% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_alias}
\alias{Expr_alias}
\title{Rename Expr output}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_alias(name)
}
\arguments{
\item{name}{string new name of output}
}
\value{
Expr
}
\description{
Rename the output of an expression.
}
\examples{
pl$col("bob")$alias("alice")
}
\keyword{Expr}
