% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_alias}
\alias{Expr_alias}
\title{Rename Expr output}
\usage{
Expr_alias(name)
}
\arguments{
\item{string}{new name of output}
}
\value{
Expr
}
\description{
Rename Expr output
}
\examples{
pl$col("bob")$alias("alice")
}
\keyword{Expr}
\keyword{Rename}
\keyword{an}
\keyword{expression.}
\keyword{of}
\keyword{output}
\keyword{the}
