% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_all}
\alias{Expr_all}
\title{All (is true)}
\usage{
Expr_all()
}
\value{
Boolean literal
}
\description{
All (is true)
}
\details{
last \code{all()} in example is this Expr method, the first \code{pl$all()} refers
to "all-columns" and is an expression constructor
}
\examples{
pl$DataFrame(list(all=c(T,T),any=c(T,F),none=c(F,F)))$select(pl$all()$all())
}
\keyword{Boolean}
\keyword{Check}
\keyword{Expr}
\keyword{`TRUE`.}
\keyword{a}
\keyword{all}
\keyword{are}
\keyword{boolean}
\keyword{column}
\keyword{if}
\keyword{in}
\keyword{values}
