% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_all}
\alias{Expr_all}
\title{All, is true}
\usage{
Expr_all
}
\value{
Boolean literal
}
\description{
Check if all boolean values in a Boolean column are \code{TRUE}.
This method is an expression - not to be confused with
\code{pl$all} which is a function to select all columns.
}
\details{
last \code{all()} in example is this Expr method, the first \code{pl$all()} refers
to "all-columns" and is an expression constructor
}
\examples{
pl$DataFrame(
  all=c(TRUE,TRUE),
  any=c(TRUE,FALSE),
  none=c(FALSE,FALSE)
)$select(
  pl$all()$all()
)
}
\keyword{Expr}
