% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_all}
\alias{Expr_all}
\title{Apply logical AND on a column}
\usage{
Expr_all(drop_nulls = TRUE)
}
\arguments{
\item{drop_nulls}{Logical. Default TRUE, as name says.}
}
\value{
Boolean literal
}
\description{
Check if all values in a Boolean column are \code{TRUE}. This method is an
expression - not to be confused with \code{pl$all()} which is a function to select
all columns.
}
\examples{
pl$DataFrame(
  all = c(TRUE, TRUE),
  any = c(TRUE, FALSE),
  none = c(FALSE, FALSE)
)$select(
  # the first $all() selects all columns, the second one applies the AND
  # logical on the values
  pl$all()$all()
)
}
