% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_and}
\alias{Expr_and}
\title{And}
\usage{
Expr_and(other)
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Expr
}
\description{
combine to boolean exprresions with AND
}
\examples{
pl$lit(TRUE) & TRUE
pl$lit(TRUE)$and(pl$lit(TRUE))
}
\keyword{Expr}
\keyword{Expr_operators}
