% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_any}
\alias{Expr_any}
\title{Any (is true)}
\usage{
Expr_any()
}
\value{
Boolean literal
}
\description{
Any (is true)
}
\examples{
pl$DataFrame(list(all=c(T,T),any=c(T,F),none=c(F,F)))$select(pl$all()$any())
}
\keyword{Boolean}
\keyword{Check}
\keyword{Expr}
\keyword{`TRUE`.}
\keyword{a}
\keyword{any}
\keyword{boolean}
\keyword{column}
\keyword{if}
\keyword{in}
\keyword{is}
\keyword{value}
