% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_approx_n_unique}
\alias{Expr_approx_n_unique}
\title{Approx count unique values}
\usage{
<Expr>$approx_n_unique
}
\value{
Expr
}
\description{
This is done using the HyperLogLog++ algorithm for cardinality estimation.
}
\examples{
pl$DataFrame(iris[, 4:5])$
  with_columns(count = pl$col("Species")$approx_n_unique())
}
\keyword{datasets}
