% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arccos}
\alias{Expr_arccos}
\title{Compute inverse cosine}
\usage{
Expr_arccos
}
\value{
Expr
}
\description{
Compute inverse cosine
}
\examples{
pl$DataFrame(a = c(-1, cos(0.5), 0, 1, NA_real_))$
  with_columns(arccos = pl$col("a")$arccos())
}
\keyword{datasets}
