% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arccosh}
\alias{Expr_arccosh}
\title{Compute inverse hyperbolic cosine}
\usage{
Expr_arccosh
}
\value{
Expr
}
\description{
Compute inverse hyperbolic cosine
}
\examples{
pl$DataFrame(a = c(-1, cosh(0.5), 0, 1, NA_real_))$
  with_columns(arccosh = pl$col("a")$arccosh())
}
\keyword{datasets}
