% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_arcsin}
\alias{Expr_arcsin}
\alias{arcsin}
\title{Arcsin}
\format{
Method
}
\usage{
Expr_arcsin
}
\value{
Expr
}
\description{
Compute the element-wise value for the inverse sine.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a=c(-1,sin(0.5),0,1,NA_real_))$select(pl$col("a")$arcsin())
}
\keyword{Expr}
