% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arcsinh}
\alias{Expr_arcsinh}
\alias{arcsinh}
\title{Arcsinh}
\usage{
Expr_arcsinh
}
\value{
Expr
}
\description{
Compute the element-wise value for the inverse hyperbolic sine.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a=c(-1,sinh(0.5),0,1,NA_real_))$select(pl$col("a")$arcsinh())
}
\keyword{Expr}
