% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_arctan}
\alias{Expr_arctan}
\alias{arctan}
\title{Arctan}
\format{
Method
}
\usage{
Expr_arctan
}
\value{
Expr
}
\description{
Compute the element-wise value for the inverse tangent.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a=c(-1,tan(0.5),0,1,NA_real_))$select(pl$col("a")$arctan())
}
\keyword{Expr}
