% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arctanh}
\alias{Expr_arctanh}
\alias{arctanh}
\title{Arctanh}
\usage{
Expr_arctanh
}
\value{
Expr
}
\description{
Compute the element-wise value for the inverse hyperbolic tangent.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a = c(-1, tanh(0.5), 0, 1, NA_real_))$select(pl$col("a")$arctanh())
}
\keyword{Expr}
