% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_arg_max}
\alias{Expr_arg_max}
\alias{arg_max}
\title{Index of min value}
\format{
a method
}
\usage{
Expr_arg_max
}
\value{
Expr
}
\description{
Get the index of the minimal value.
}
\examples{
df = pl$DataFrame(list(
  a = c(6, 1, 0, NA, Inf, NaN),
))$select(pl$arg_max())
}
\keyword{Expr}
