% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arg_min}
\alias{Expr_arg_min}
\title{Index of min value}
\usage{
Expr_arg_min
}
\value{
Expr
}
\description{
Get the index of the minimal value.
}
\examples{
pl$DataFrame(list(
  a = c(6, 1, 0, NA, Inf, NaN)
))$select(pl$col("a")$arg_min())
}
\keyword{datasets}
