% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_arg_min}
\alias{Expr_arg_min}
\alias{arg_min}
\title{Index of min value}
\format{
a method
}
\usage{
Expr_arg_min
}
\value{
Expr
}
\description{
Get the index of the minimal value.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[minipolars]{docs_translations}}
}
\examples{
pl$DataFrame(list(
  a = c(6, 1, 0, NA, Inf, NaN)
))$select(pl$col("a")$arg_min())
}
\keyword{Expr}
