% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arg_sort}
\alias{Expr_arg_sort}
\alias{Expr_argsort}
\alias{argsort}
\title{Index of a sort}
\usage{
Expr_arg_sort(descending = FALSE, nulls_last = FALSE)

Expr_argsort(descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{descending}{Sort in descending order. When sorting by multiple columns,
can be specified per column by passing a vector of booleans.}

\item{nulls_last}{If \code{TRUE}, place nulls values last.}
}
\value{
Expr
}
\description{
Get the index values that would sort this column.

argsort is a alias for arg_sort
}
\examples{
pl$DataFrame(
  a = c(6, 1, 0, NA, Inf, NaN)
)$with_columns(arg_sorted = pl$col("a")$arg_sort())
}
