% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arg_sort}
\alias{Expr_arg_sort}
\alias{arg_sort}
\alias{Expr_argsort}
\alias{argsort}
\title{Index of a sort}
\usage{
Expr_arg_sort(reverse = FALSE, nulls_last = FALSE)

Expr_argsort(reverse = FALSE, nulls_last = FALSE)
}
\arguments{
\item{reverse}{bool default FALSE, reverses sort}

\item{nulls_last}{bool, default FALSE, place Nulls last}
}
\value{
Expr
}
\description{
Get the index values that would sort this column.
If 'reverse=True` the smallest elements will be given.

argsort is a alias for arg_sort
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[polars]{docs_translations}}
}
\examples{
pl$DataFrame(list(
  a = c(6, 1, 0, NA, Inf, NaN)
))$select(pl$col("a")$arg_sort())
}
\keyword{Expr}
