% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arg_unique}
\alias{Expr_arg_unique}
\alias{arg_unique}
\title{Index of First Unique Value.}
\usage{
Expr_arg_unique
}
\value{
Expr
}
\description{
Index of First Unique Value.
}
\examples{
pl$select(pl$lit(c(1:2, 1:3))$arg_unique())
}
\keyword{Expr}
