% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_arr}
\alias{Expr_arr}
\alias{arr_ns}
\title{arr: list related methods}
\usage{
Expr_arr()
}
\value{
Expr
}
\description{
Create an object namespace of all list related methods.
See the individual method pages for full details
}
\examples{
df_with_list = pl$DataFrame(
  group = c(1, 1, 2, 2, 3),
  value = c(1:5)
)$groupby(
  "group",
  maintain_order = TRUE
)$agg(
  pl$col("value") * 3L
)
df_with_list$with_columns(
  pl$col("value")$arr$lengths()$alias("group_size")
)
}
\keyword{Expr}
