% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_all}
\alias{expr_arr_all}
\title{Evaluate whether all boolean values are true for every sub-array}
\usage{
expr_arr_all()
}
\value{
A polars \link{expression}
}
\description{
Evaluate whether all boolean values are true for every sub-array
}
\examples{
df <- pl$DataFrame(
  values = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), c(NA, NA)),
)$cast(pl$Array(pl$Boolean, 2))
df$with_columns(all = pl$col("values")$arr$all())
}
