% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_contains}
\alias{expr_arr_contains}
\title{Check if sub-arrays contain the given item}
\usage{
expr_arr_contains(item, ..., nulls_equal = TRUE)
}
\arguments{
\item{item}{Item that will be checked for membership. Can be an Expr or
something coercible to an Expr. Strings are not parsed as columns.}

\item{...}{These dots are for future extensions and must be empty.}

\item{nulls_equal}{If \code{TRUE}, treat null as a distinct value. Null values will not propagate.}
}
\value{
A polars \link{expression}
}
\description{
Check if sub-arrays contain the given item
}
\examples{
df <- pl$DataFrame(
  values = list(0:2, 4:6, c(NA, NA, NA)),
  item = c(0L, 4L, 2L),
)$cast(values = pl$Array(pl$Float64, 3))
df$with_columns(
  with_expr = pl$col("values")$arr$contains(
    pl$col("item")$cast(pl$Float64)
  ),
  with_lit = pl$col("values")$arr$contains(
    pl$lit(1L)$cast(pl$Float64)
  )
)
}
