% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_contains}
\alias{expr_arr_contains}
\title{Check if sub-arrays contain the given item}
\usage{
expr_arr_contains(item)
}
\arguments{
\item{item}{Expr or something coercible to an Expr. Strings are \emph{not} parsed
as columns.}
}
\value{
A polars \link{expression}
}
\description{
Check if sub-arrays contain the given item
}
\examples{
df <- pl$DataFrame(
  values = list(0:2, 4:6, c(NA, NA, NA)),
  item = c(0L, 4L, 2L),
)$cast(values = pl$Array(pl$Float64, 3))
df$with_columns(
  with_expr = pl$col("values")$arr$contains(pl$col("item")),
  with_lit = pl$col("values")$arr$contains(1)
)
}
