% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_count_matches}
\alias{expr_arr_count_matches}
\title{Count how often a value occurs in every sub-array}
\usage{
expr_arr_count_matches(element)
}
\arguments{
\item{element}{An Expr or something coercible to an Expr that produces a
single value.}
}
\value{
A polars \link{expression}
}
\description{
Count how often a value occurs in every sub-array
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 2), c(1, 1), c(2, 2))
)$cast(pl$Array(pl$Int64, 2))
df$with_columns(number_of_twos = pl$col("values")$arr$count_matches(2))
}
