% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_eval}
\alias{expr_arr_eval}
\title{Run any polars expression on the sub-array's values}
\usage{
expr_arr_eval(expr, ..., as_list = FALSE)
}
\arguments{
\item{expr}{Expression to run. Note that you can select an element with
\code{pl$element()}, \code{pl$first()}, and more. See Examples.}

\item{...}{These dots are for future extensions and must be empty.}

\item{as_list}{Collect the resulting data as a list. This allows for
expressions which output a variable amount of data.}
}
\value{
A polars \link{expression}
}
\description{
Run any polars expression on the sub-array's values
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 1), c(8, 5), c(3, 2))
)$cast(pl$Array(pl$Float64, 2))

df$with_columns(
  cum_sum = pl$col("a")$arr$eval(pl$element()$cum_sum())
)

# This would error without `as_list = TRUE` since `$unique()` doesn't return
# the same number of values in each row.
df$with_columns(
  cum_sum = pl$col("a")$arr$eval(pl$element()$unique(), as_list = TRUE)
)
}
