% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R, R/expr-expr.R
\name{expr_arr_explode}
\alias{expr_arr_explode}
\title{Explode array in separate rows}
\usage{
expr_arr_explode(..., empty_as_null = TRUE, keep_nulls = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{empty_as_null}{Indicates to explode an empty list/array into a \code{null}.}

\item{keep_nulls}{Indicates to explode a \code{null} list/array into a \code{null}.}
}
\value{
A polars \link{expression}
}
\description{
Returns a column with a separate row for every array element.
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 2, 3), c(4, 5, 6))
)$cast(pl$Array(pl$Int64, 3))
df$select(pl$col("a")$arr$explode())
}
