% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_explode}
\alias{expr_arr_explode}
\title{Explode array in separate rows}
\usage{
expr_arr_explode()
}
\value{
A polars \link{expression}
}
\description{
Returns a column with a separate row for every array element.
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 2, 3), c(4, 5, 6))
)$cast(pl$Array(pl$Int64, 3))
df$select(pl$col("a")$arr$explode())
}
