% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_join}
\alias{expr_arr_join}
\title{Join elements in every sub-array}
\usage{
expr_arr_join(separator, ..., ignore_nulls = FALSE)
}
\arguments{
\item{separator}{String to separate the items with. Can be an Expr. Strings
are not parsed as columns.}

\item{...}{Dots which should be empty.}
}
\value{
A polars \link{expression}
}
\description{
Join all string items in a sub-array and place a separator between them. This
only works if the inner type of the array is \code{String}.
}
\examples{
df <- pl$DataFrame(
  values = list(c("a", "b", "c"), c("x", "y", "z"), c("e", NA, NA)),
  separator = c("-", "+", "/"),
)$cast(values = pl$Array(pl$String, 3))
df$with_columns(
  join_with_expr = pl$col("values")$arr$join(pl$col("separator")),
  join_with_lit = pl$col("values")$arr$join(" "),
  join_ignore_null = pl$col("values")$arr$join(" ", ignore_nulls = TRUE)
)
}
