% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_n_unique}
\alias{expr_arr_n_unique}
\title{Count the number of unique values in every sub-array}
\usage{
expr_arr_n_unique()
}
\value{
A polars \link{expression}
}
\description{
Count the number of unique values in every sub-array
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 1, 2), c(2, 3, 4))
)$cast(pl$Array(pl$Int64, 3))
df$with_columns(n_unique = pl$col("a")$arr$n_unique())
}
