% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_reverse}
\alias{expr_arr_reverse}
\title{Reverse values in every sub-array}
\usage{
expr_arr_reverse()
}
\value{
A polars \link{expression}
}
\description{
Reverse values in every sub-array
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(NA, 6))
)$cast(pl$Array(pl$Float64, 2))
df$with_columns(reverse = pl$col("values")$arr$reverse())
}
