% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_shift}
\alias{expr_arr_shift}
\title{Shift values in every sub-array by the given number of indices}
\usage{
expr_arr_shift(n = 1)
}
\arguments{
\item{n}{Number of indices to shift forward. If a negative value is passed,
values are shifted in the opposite direction instead.}
}
\value{
A polars \link{expression}
}
\description{
Shift values in every sub-array by the given number of indices
}
\examples{
df <- pl$DataFrame(
  values = list(1:3, c(2L, NA, 5L)),
  idx = 1:2,
)$cast(values = pl$Array(pl$Int32, 3))
df$with_columns(
  shift_by_expr = pl$col("values")$arr$shift(pl$col("idx")),
  shift_by_lit = pl$col("values")$arr$shift(2)
)
}
