% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_sort}
\alias{expr_arr_sort}
\title{Sort values in every sub-array}
\usage{
expr_arr_sort(..., descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{descending}{Sort in descending order.}

\item{nulls_last}{Place null values last.}
}
\description{
Sort values in every sub-array
}
\examples{
df <- pl$DataFrame(
  values = list(c(2, 1), c(3, 4), c(NA, 6))
)$cast(pl$Array(pl$Float64, 2))
df$with_columns(sort = pl$col("values")$arr$sort(nulls_last = TRUE))
}
