% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_sum}
\alias{expr_arr_sum}
\title{Compute the sum of the sub-arrays}
\usage{
expr_arr_sum()
}
\value{
A polars \link{expression}
}
\description{
Compute the sum of the sub-arrays
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(NA, 6))
)$cast(pl$Array(pl$Float64, 2))
df$with_columns(sum = pl$col("values")$arr$sum())
}
