% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_to_list}
\alias{expr_arr_to_list}
\title{Convert an Array column into a List column with the same inner data type}
\usage{
expr_arr_to_list()
}
\value{
A polars \link{expression}
}
\description{
Convert an Array column into a List column with the same inner data type
}
\examples{
df <- pl$DataFrame(
  a = list(c(1, 2), c(3, 4))
)$cast(pl$Array(pl$Int8, 2))

df$with_columns(
  list = pl$col("a")$arr$to_list()
)
}
