% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_unique}
\alias{expr_arr_unique}
\title{Get the unique values in every sub-array}
\usage{
expr_arr_unique(..., maintain_order = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A polars \link{expression}
}
\description{
Get the unique values in every sub-array
}
\examples{
df <- pl$DataFrame(
  values = list(c(1, 1, 2), c(4, 4, 4), c(NA, 6, 7)),
)$cast(pl$Array(pl$Float64, 3))
df$with_columns(unique = pl$col("values")$arr$unique())
}
