% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-array.R
\name{expr_arr_var}
\alias{expr_arr_var}
\title{Compute the variance of the sub-arrays}
\usage{
expr_arr_var(ddof = 1)
}
\arguments{
\item{ddof}{"Delta Degrees of Freedom": the divisor used in the calculation
is \code{N - ddof}, where \code{N} represents the number of elements. By default
\code{ddof} is 1.}
}
\value{
A polars \link{expression}
}
\description{
Compute the variance of the sub-arrays
}
\examples{
df <- pl$DataFrame(
  values = list(c(2, 1, 4), c(8.4, 3.2, 1)),
)$cast(pl$Array(pl$Float64, 3))
df$with_columns(var = pl$col("values")$arr$var())
}
