% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-binary.R
\name{expr_bin_encode}
\alias{expr_bin_encode}
\title{Encode a value using the provided encoding}
\usage{
expr_bin_encode(encoding)
}
\arguments{
\item{encoding}{A character, \code{"hex"} or \code{"base64"}. The encoding to use.}
}
\value{
A polars \link{expression}
}
\description{
Encode a value using the provided encoding
}
\examples{
df <- pl$DataFrame(
  name = c("black", "yellow", "blue"),
  code = as_polars_series(
    c("000000", "ffff00", "0000ff")
  )$cast(pl$Binary)$bin$decode("hex")
)

df$with_columns(encoded = pl$col("code")$bin$encode("hex"))
}
