% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-binary.R
\name{expr_bin_get}
\alias{expr_bin_get}
\title{Get the byte value at the given index}
\usage{
expr_bin_get(index, ..., null_on_oob = FALSE)
}
\arguments{
\item{index}{Index to return per binary value.}

\item{...}{These dots are for future extensions and must be empty.}

\item{null_on_oob}{If \code{TRUE}, return \code{null} if an index is out of bounds.
Otherwise, raise an error.}
}
\value{
A polars \link{expression}
}
\description{
For example, index \code{0} would return the first byte of every binary value
and index \code{-1} would return the last byte of every binary value.
}
\examples{
df <- pl$DataFrame(x = c("\x01\x02\x03", "", "\x04\x05"))$cast(pl$Binary)

df$with_columns(
  get = pl$col("x")$bin$get(0, null_on_oob = TRUE)
)
}
