% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-binary.R
\name{expr_bin_starts_with}
\alias{expr_bin_starts_with}
\title{Check if values start with a binary substring}
\usage{
expr_bin_starts_with(prefix)
}
\arguments{
\item{sub}{Prefix substring.}
}
\value{
A polars \link{expression}
}
\description{
Check if values start with a binary substring
}
\examples{
colors <- pl$DataFrame(
  name = c("black", "yellow", "blue"),
  code = as_polars_series(c("x00x00x00", "xffxffx00", "x00x00xff"))$cast(pl$Binary),
  prefix = as_polars_series(c("x00", "xffx00", "xffxff"))$cast(pl$Binary)
)

colors$select(
  "name",
  starts_with_lit = pl$col("code")$bin$starts_with("xff"),
  starts_with_expr = pl$col("code")$bin$starts_with(pl$col("prefix"))
)
}
