% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-categorical.R
\name{expr_cat_physical}
\alias{expr_cat_physical}
\title{Get the physical values of a Categorical or Enum data type}
\usage{
expr_cat_physical()
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get the physical representation of a \link[=polars_dtype]{Categorical} or an
\link[=polars_dtype]{Enum} column, i.e. the integer codes used to store the
categories.
}
\examples{
df <- pl$DataFrame(x = c("foo", "bar", "foo", "x", NA))$cast(
  pl$Enum(c("bar", "foo", "x"))
)

df$with_columns(physical = pl$col("x")$cat$physical())
}
\seealso{
\itemize{
\item \code{\link[=expr_cat_to]{<expr>$cat$to()}}
}
}
