% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-categorical.R
\name{expr_cat_to}
\alias{expr_cat_to}
\title{Convert to a Categorical or Enum data type}
\usage{
expr_cat_to(dtype, ..., strict = TRUE)
}
\arguments{
\item{dtype}{A Polars DataType or DataTypeExpr. Must be a
\link[=polars_dtype]{Categorical} or an \link[=polars_dtype]{Enum}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{strict}{If \code{TRUE} (default), raise an error if a value is not a valid
category. If \code{FALSE}, such values become \code{null}.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Convert physical values to a Categorical or Enum data type. The input must
be of the physical type of the target data type, i.e. \code{UInt32} for
\link[=polars_dtype]{Categorical} and \code{UInt8}, \code{UInt16} or \code{UInt32} for
\link[=polars_dtype]{Enum} (depending on the number of categories).
}
\examples{
dtype <- pl$Enum(c("bar", "foo", "x"))

df <- pl$DataFrame(x = c(1, 0, 1, 2, NA))$cast(pl$UInt8)
df$with_columns(cat = pl$col("x")$cat$to(dtype))

# Values that are not valid categories are converted to `null` when
# `strict = FALSE`
df2 <- pl$DataFrame(x = c(1, 0, 4))$cast(pl$UInt8)
df2$with_columns(cat = pl$col("x")$cat$to(dtype, strict = FALSE))
}
\seealso{
\itemize{
\item \code{\link[=expr_cat_physical]{<expr>$cat$physical()}}
}
}
