% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_class}
\alias{Expr_class}
\title{Polars Expressions}
\description{
\code{Expr}essions are all the functions and methods that are applicable
to a Polars DataFrame. They can be split into the following categories (following
the \href{https://pola-rs.github.io/polars/py-polars/html/reference/expressions}{Py-Polars classification}:
\itemize{
\item Aggregate
\item Binary
\item Categorical
\item Computation
\item Functions
\item List
\item Meta
\item String
\item Struct
\item Temporal
}
}
