% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_clip}
\alias{Expr_clip}
\title{Clip elements}
\usage{
Expr_clip(min, max)
}
\arguments{
\item{min}{Minimum value, Expr returning a numeric.}

\item{max}{Maximum value, Expr returning a numeric.}
}
\value{
Expr
}
\description{
Clip (limit) the values in an array to a \code{min} and \code{max} boundary. This only
works for numerical types.
}
\examples{
pl$DataFrame(foo = c(-50L, 5L, NA_integer_, 50L))$
  with_columns(clipped = pl$col("foo")$clip(1, 10))
}
