% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cosh}
\alias{Expr_cosh}
\title{Compute hyperbolic cosine}
\usage{
Expr_cosh()
}
\value{
Expr
}
\description{
Compute hyperbolic cosine
}
\examples{
pl$DataFrame(a = c(-1, acosh(2), 0, 1, NA_real_))$
  with_columns(cosh = pl$col("a")$cosh())
}
