% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_count}
\alias{Expr_count}
\alias{Expr_len}
\title{Count elements}
\usage{
Expr_count()

Expr_len()
}
\value{
Expr
}
\description{
Count the number of elements in this expression. Note that \code{NULL} values are
also counted. \verb{$len()} is an alias.
}
\examples{
pl$DataFrame(
  all = c(TRUE, TRUE),
  any = c(TRUE, FALSE),
  none = c(FALSE, FALSE)
)$select(
  pl$all()$count()
)
}
