% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_count}
\alias{Expr_count}
\title{Count values}
\usage{
Expr_count()
}
\value{
Expr
}
\description{
Count the number of values in this expression.
Similar to R length()
}
\examples{
pl$DataFrame(list(all=c(T,T),any=c(T,F),none=c(F,F)))$select(pl$all()$count())
}
