% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_count}
\alias{Expr_count}
\title{Count values}
\usage{
Expr_count()
}
\value{
Expr
}
\description{
Count values
}
\examples{
pl$DataFrame(list(all=c(T,T),any=c(T,F),none=c(F,F)))$select(pl$all()$count())
}
\keyword{Count}
\keyword{Expr}
\keyword{R}
\keyword{Similar}
\keyword{expression.}
\keyword{in}
\keyword{length()}
\keyword{number}
\keyword{of}
\keyword{the}
\keyword{this}
\keyword{to}
\keyword{values}
