% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_count}
\alias{Expr_count}
\alias{Expr_len}
\title{Count values (len is a alias)}
\usage{
Expr_count

Expr_len
}
\value{
Expr

Expr
}
\description{
Count the number of values in this expression.
Similar to R length()
}
\examples{
pl$DataFrame(
  all = c(TRUE, TRUE),
  any = c(TRUE, FALSE),
  none = c(FALSE, FALSE)
)$select(
  pl$all()$count()
)
pl$DataFrame(
  all = c(TRUE, TRUE),
  any = c(TRUE, FALSE),
  none = c(FALSE, FALSE)
)$select(
  pl$all()$len(),
  pl$col("all")$first()$len()$alias("all_first")
)
}
\keyword{Expr}
