% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cum_count}
\alias{Expr_cum_count}
\title{Cumulative count}
\usage{
Expr_cum_count(reverse = FALSE)
}
\arguments{
\item{reverse}{If \code{TRUE}, reverse the count.}
}
\value{
Expr
}
\description{
Get an array with the cumulative count (zero-indexed) computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.

\verb{$cum_count()} does not seem to count within lists.
}
\examples{
pl$DataFrame(a = 1:4)$with_columns(
  pl$col("a")$cum_count()$alias("cum_count"),
  pl$col("a")$cum_count(reverse = TRUE)$alias("cum_count_reversed")
)
}
