% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cum_sum}
\alias{Expr_cum_sum}
\alias{Expr_cumsum}
\title{Cumulative sum}
\usage{
Expr_cum_sum(reverse = FALSE)
}
\arguments{
\item{reverse}{bool, default FALSE, if true roll over vector from back to forth}
}
\value{
Expr
}
\description{
Get an array with the cumulative sum computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to
Int64 before summing to prevent overflow issues.
}
\examples{
pl$DataFrame(list(a = 1:4))$select(
  pl$col("a")$cum_sum()$alias("cum_sum"),
  pl$col("a")$cum_sum(reverse = TRUE)$alias("cum_sum_reversed")
)
}
\keyword{Expr}
