% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cumprod}
\alias{Expr_cumprod}
\alias{cumprod}
\title{Cumulative product}
\usage{
Expr_cumprod(reverse = FALSE)
}
\arguments{
\item{reverse}{bool, default FALSE, if true roll over vector from back to forth}
}
\value{
Expr
}
\description{
Get an array with the cumulative product computed at every element.
}
\details{
The Dtypes Int8, UInt8, Int16 and UInt16 are cast to
Int64 before summing to prevent overflow issues.
}
\examples{
pl$DataFrame(list(a = 1:4))$select(
  pl$col("a")$cumprod()$alias("cumprod"),
  pl$col("a")$cumprod(reverse = TRUE)$alias("cumprod_reversed")
)
}
\keyword{Expr}
