% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_cumsum}
\alias{Expr_cumsum}
\title{Cumulative sum}
\usage{
Expr_cumsum(reverse = FALSE)
}
\arguments{
\item{reverse}{bool, default FALSE, if true roll over vector from back to forth}
}
\value{
Expr
}
\description{
Get an array with the cumulative sum computed at every element.
}
\details{
Dtypes in {Int8, UInt8, Int16, UInt16} are cast to
Int64 before summing to prevent overflow issues.
}
\examples{
pl$DataFrame(list(a = 1:4))$select(
  pl$col("a")$cumsum()$alias("cumsum"),
  pl$col("a")$cumsum(reverse = TRUE)$alias("cumsum_reversed")
)
}
\keyword{Expr}
