% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_diff}
\alias{Expr_diff}
\title{Difference}
\usage{
Expr_diff(n = 1, null_behavior = "ignore")
}
\arguments{
\item{n}{Number of slots to shift.}

\item{null_behavior}{String, either \code{"ignore"} (default), else \code{"drop"}.}
}
\value{
Expr
}
\description{
Calculate the n-th discrete difference.
}
\examples{
pl$DataFrame(a = c(20L, 10L, 30L, 40L))$with_columns(
  diff_default = pl$col("a")$diff(),
  diff_2_ignore = pl$col("a")$diff(2, "ignore")
)
}
