% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_diff}
\alias{Expr_diff}
\alias{diff}
\title{Diff}
\usage{
Expr_diff(n = 1, null_behavior = c("ignore", "drop"))
}
\arguments{
\item{n}{Number of slots to shift.}

\item{null_behavior}{choice default 'ignore', else 'drop'.}
}
\value{
Expr
}
\description{
Calculate the n-th discrete difference.
}
\examples{
pl$DataFrame(list(a = c(20L, 10L, 30L, 40L)))$select(
  pl$col("a")$diff()$alias("diff_default"),
  pl$col("a")$diff(2, "ignore")$alias("diff_2_ignore")
)
}
\keyword{Expr}
