% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_div}
\alias{Expr_div}
\title{Divide two expressions}
\usage{
Expr_div(other)
}
\arguments{
\item{other}{Numeric literal or expression value.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of float division operator \code{expr / other}.
}
\details{
Zero-division behaviour follows IEEE-754:
\itemize{
\item \code{0/0}: Invalid operation - mathematically undefined, returns \code{NaN}.
\item \code{n/0}: On finite operands gives an exact infinite result, e.g.: ±infinity.
}
}
\examples{
df = pl$DataFrame(
  x = -2:2,
  y = c(0.5, 0, 0, -4, -0.5)
)

df$with_columns(
  `x/2` = pl$col("x")$div(2),
  `x/y` = pl$col("x")$div(pl$col("y"))
)
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
\item \code{\link[=Expr_floor_div]{<Expr>$floor_div()}}
}
}
