% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_div}
\alias{Expr_div}
\alias{/.RPolarsExpr}
\title{Divide two expressions}
\usage{
Expr_div(other)

\method{/}{RPolarsExpr}(e1, e2)
}
\arguments{
\item{other}{Literal or object that can be converted to a literal}

\item{e1}{Expr only}

\item{e2}{Expr or anything that can be converted to a literal}
}
\value{
Expr
}
\description{
The RHS can either be an Expr or an object that can be converted to a literal
(e.g an integer).
}
\examples{
pl$lit(5) / 10
pl$lit(5) / pl$lit(10)
pl$lit(5)$div(pl$lit(10))
}
