% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_dot}
\alias{Expr_dot}
\title{Dot product}
\usage{
Expr_dot(other)
}
\arguments{
\item{other}{Literal or object that can be converted to a literal}
}
\value{
Expr
}
\description{
Compute the dot/inner product between two Expressions.
}
\examples{
pl$DataFrame(
  a = 1:4, b = c(1, 2, 3, 4)
)$with_columns(
  pl$col("a")$dot(pl$col("b"))$alias("a dot b"),
  pl$col("a")$dot(pl$col("a"))$alias("a dot a")
)
}
