% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_dot}
\alias{Expr_dot}
\alias{dot}
\title{Dot product}
\format{
a method
}
\usage{
Expr_dot(other)
}
\arguments{
\item{other}{Exprto compute dot product with.}
}
\value{
Expr
}
\description{
Compute the dot/inner product between two Expressions.
}
\examples{
l = list(a=1:4,b=c(1,2,3,5),c=c(NA_real_,1:3),d=c(6:8,NaN))
pl$DataFrame(l)$select(
  pl$col("a")$dot(pl$col("b"))$alias("a dot b"),
  pl$col("a")$dot(pl$col("a"))$alias("a dot a"),
  pl$col("a")$dot(pl$col("c"))$alias("a dot c"),
  pl$col("a")$dot(pl$col("d"))$alias("a dot d")
)
}
\keyword{Expr}
