% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_drop_nans}
\alias{Expr_drop_nans}
\title{Drop NaN}
\usage{
Expr_drop_nans()
}
\value{
Expr
}
\description{
Drop NaN
}
\details{
Note that \code{NaN} values are not \code{null} values. Null values correspond to NA
in R.
}
\examples{
pl$DataFrame(list(x = c(1, 2, NaN, NA)))$select(pl$col("x")$drop_nans())
}
\seealso{
\code{drop_nulls()}
}
