% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_drop_nulls}
\alias{Expr_drop_nulls}
\title{Drop null(s)}
\usage{
Expr_drop_nulls
}
\value{
Expr
}
\description{
Drop null values.
Similar to R syntax \code{x[!(is.na(x) & !is.nan(x))]}
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[polars]{docs_translations}}
}
\examples{
pl$DataFrame(list(x = c(1, 2, NaN, NA)))$select(pl$col("x")$drop_nulls())
}
\keyword{Expr}
