% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_base_utc_offset}
\alias{expr_dt_base_utc_offset}
\title{Base offset from UTC}
\usage{
expr_dt_base_utc_offset()
}
\value{
A polars \link{expression}
}
\description{
This computes the offset between a time zone and UTC. This is usually
constant for all datetimes in a given time zone, but may vary in the rare
case that a country switches time zone, like Samoa (Apia) did at the end of
2011. Use \link[=expr_dt_dst_offset]{$dt$dst_offset()} to take daylight
saving time into account.
}
\examples{
df <- pl$DataFrame(
  x = as.POSIXct(c("2011-12-29", "2012-01-01"), tz = "Pacific/Apia")
)
df$with_columns(base_utc_offset = pl$col("x")$dt$base_utc_offset())
}
