% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_century}
\alias{expr_dt_century}
\title{Extract the century from underlying representation}
\usage{
expr_dt_century()
}
\value{
A polars \link{expression}
}
\description{
Returns the century number in the calendar date.
}
\examples{
df <- pl$DataFrame(
  date = as.Date(
    c("999-12-31", "1897-05-07", "2000-01-01", "2001-07-05", "3002-10-20")
  )
)
df$with_columns(
  century = pl$col("date")$dt$century()
)
}
