% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_combine}
\alias{expr_dt_combine}
\title{Combine Date and Time}
\usage{
expr_dt_combine(time, time_unit = c("us", "ns", "ms"))
}
\arguments{
\item{time}{The number of epoch since or before (if negative) the Date. Can
be an Expr or a PTime.}

\item{time_unit}{One of \code{"us"} (default, microseconds),
\code{"ns"} (nanoseconds) or \code{"ms"}(milliseconds). Representing the unit of time.}
}
\value{
A polars \link{expression}
}
\description{
If the underlying expression is a Datetime then its time component is
replaced, and if it is a Date then a new Datetime is created by combining
the two values.
}
\examples{
\dontshow{if (requireNamespace("hms", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- pl$DataFrame(
  dtm = c(
    ISOdatetime(2022, 12, 31, 10, 30, 45),
    ISOdatetime(2023, 7, 5, 23, 59, 59)
  ),
  dt = c(ISOdate(2022, 10, 10), ISOdate(2022, 7, 5)),
  tm = hms::parse_hms(c("1:2:3.456000", "7:8:9.101000"))
)

df

df$select(
  d1 = pl$col("dtm")$dt$combine(pl$col("tm")),
  s2 = pl$col("dt")$dt$combine(pl$col("tm")),
  d3 = pl$col("dt")$dt$combine(hms::parse_hms("4:5:6"))
)
\dontshow{\}) # examplesIf}
}
