% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_dst_offset}
\alias{expr_dt_dst_offset}
\title{Daylight savings offset from UTC}
\usage{
expr_dt_dst_offset()
}
\value{
A polars \link{expression}
}
\description{
This computes the offset between a time zone and UTC, taking into account
daylight saving time. Use \link[=expr_dt_base_utc_offset]{$dt$base_utc_offset()}
to avoid counting DST.
}
\examples{
df <- pl$DataFrame(
  x = as.POSIXct(c("2020-10-25", "2020-10-26"), tz = "Europe/London")
)
df$with_columns(dst_offset = pl$col("x")$dt$dst_offset())
}
