% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_epoch}
\alias{expr_dt_epoch}
\title{Get epoch of given Datetime}
\usage{
expr_dt_epoch(time_unit = c("us", "ns", "ms", "s", "d"))
}
\arguments{
\item{time_unit}{Time unit, one of \code{"ns"}, \code{"us"}, \code{"ms"}, \code{"s"} or  \code{"d"}.}
}
\value{
A polars \link{expression}
}
\description{
Get the time passed since the Unix EPOCH in the give time unit.
}
\examples{
df <- pl$select(date = pl$date_range(as.Date("2001-1-1"), as.Date("2001-1-3")))

df$with_columns(
  epoch_ns = pl$col("date")$dt$epoch(),
  epoch_s = pl$col("date")$dt$epoch(time_unit = "s")
)
}
