% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_hour}
\alias{expr_dt_hour}
\title{Extract hour from underlying Datetime representation}
\usage{
expr_dt_hour()
}
\value{
A polars \link{expression}
}
\description{
Returns the hour number from 0 to 23.
}
\examples{
df <- pl$select(
  date = pl$datetime_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d2h",
    time_zone = "GMT"
  )
)
df$with_columns(
  pl$col("date")$dt$hour()$alias("hour")
)
}
