% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_is_leap_year}
\alias{expr_dt_is_leap_year}
\title{Determine whether the year of the underlying date is a leap year}
\usage{
expr_dt_is_leap_year()
}
\value{
A polars \link{expression}
}
\description{
Determine whether the year of the underlying date is a leap year
}
\examples{
df <- pl$DataFrame(date = as.Date(c("2000-01-01", "2001-01-01", "2002-01-01")))

df$with_columns(
  leap_year = pl$col("date")$dt$is_leap_year()
)
}
