% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_microsecond}
\alias{expr_dt_microsecond}
\title{Extract microseconds from underlying Datetime representation}
\usage{
expr_dt_microsecond()
}
\value{
A polars \link{expression}
}
\description{
Extract microseconds from underlying Datetime representation
}
\examples{
df <- pl$DataFrame(
  datetime = as.POSIXct(
    c(
      "1978-01-01 01:01:01",
      "2024-10-13 05:30:14.500",
      "2065-01-01 10:20:30.06"
    ),
    "UTC"
  )
)

df$with_columns(
  microsecond = pl$col("datetime")$dt$microsecond()
)
}
