% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_month_start}
\alias{expr_dt_month_start}
\title{Roll backward to the first day of the month}
\usage{
expr_dt_month_start()
}
\value{
A polars \link{expression}
}
\description{
For datetimes, the time of day is preserved.
}
\examples{
df <- pl$DataFrame(date = as.Date(c("2000-01-23", "2001-01-12", "2002-01-01")))

df$with_columns(
  month_start = pl$col("date")$dt$month_start()
)
}
