% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_ordinal_day}
\alias{expr_dt_ordinal_day}
\title{Extract ordinal day from underlying Date representation}
\usage{
expr_dt_ordinal_day()
}
\value{
A polars \link{expression}
}
\description{
Returns the day of year starting from 1. The return value ranges from 1 to
366 (the last day of year differs across years).
}
\examples{
df <- pl$select(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d"
  )
)
df$with_columns(
  ordinal_day = pl$col("date")$dt$ordinal_day()
)
}
