% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_quarter}
\alias{expr_dt_quarter}
\title{Extract quarter from underlying Date representation}
\usage{
expr_dt_quarter()
}
\value{
A polars \link{expression}
}
\description{
Returns the quarter ranging from 1 to 4.
}
\examples{
df <- pl$select(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d"
  )
)
df$with_columns(
  quarter = pl$col("date")$dt$quarter()
)
}
