% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_replace_time_zone}
\alias{expr_dt_replace_time_zone}
\title{Replace time zone for an expression of type Datetime}
\usage{
expr_dt_replace_time_zone(
  time_zone,
  ...,
  ambiguous = "raise",
  non_existent = "raise"
)
}
\arguments{
\item{time_zone}{\code{NULL} or a character time zone from \code{\link[base:timezones]{base::OlsonNames()}}.
Pass \code{NULL} to unset time zone.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ambiguous}{Determine how to deal with ambiguous datetimes.
Character vector or \link{Expression} containing the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"earliest"}: Use the earliest datetime
\item \code{"latest"}: Use the latest datetime
\item \code{"null"}: Return a null value
}}

\item{non_existent}{Determine how to deal with non-existent datetimes.
One of the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"null"}: Return a null value
}}
}
\description{
Different from \code{convert_time_zone}, this will also modify the underlying timestamp
and will ignore the original time zone.
}
