% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_strftime}
\alias{expr_dt_strftime}
\title{Convert date/time/datetime to string}
\usage{
expr_dt_strftime(format)
}
\arguments{
\item{format}{Format to use. See \code{chrono} docs for specifying the format:
\url{https://docs.rs/chrono/latest/chrono/format/strftime/index.html}`.}
}
\value{
A polars \link{expression}
}
\description{
Similar to \verb{$cast(pl$String)}, but this method allows you to customize the
formatting of the resulting string. This is an alias for \verb{$dt$to_string()}.
}
\examples{
pl$DataFrame(
  datetime = c(as.POSIXct(c("2021-01-02 00:00:00", "2021-01-03 00:00:00")))
)$
  with_columns(
  datetime_string = pl$col("datetime")$dt$strftime("\%Y/\%m/\%d \%H:\%M:\%S")
)
}
