% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_total_hours}
\alias{expr_dt_total_hours}
\title{Extract the hours from a Duration type}
\usage{
expr_dt_total_hours(..., fractional = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{fractional}{A bool to indicate whether to include the fractional component of the day.}
}
\value{
A polars \link{expression}
}
\description{
Extract the hours from a Duration type
}
\examples{
df <- pl$select(
  date = pl$date_range(
    start = as.Date("2020-1-1"),
    end = as.Date("2020-1-4"),
    interval = "1d"
  )
)
df$with_columns(
  diff_hours = pl$col("date")$diff()$dt$total_hours()
)
}
