% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_week}
\alias{expr_dt_week}
\title{Extract week from underlying Date representation}
\usage{
expr_dt_week()
}
\value{
A polars \link{expression}
}
\description{
Returns the ISO week number starting from 1. The return value ranges from 1
to 53 (the last week of year differs across years).
}
\examples{
df <- pl$select(
  date = pl$date_range(
    as.Date("2020-12-25"),
    as.Date("2021-1-05"),
    interval = "1d"
  )
)
df$with_columns(
  week = pl$col("date")$dt$week()
)
}
