% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-datetime.R
\name{expr_dt_with_time_unit}
\alias{expr_dt_with_time_unit}
\title{Set time unit of a Series of dtype Datetime or Duration}
\usage{
expr_dt_with_time_unit(time_unit = c("ns", "us", "ms"))
}
\arguments{
\item{time_unit}{Time unit, one of 'ns', 'us', or 'ms'.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Cast to Int64 and then to Datetime instead.
}
\keyword{internal}
