% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_eq}
\alias{Expr_eq}
\alias{==.Expr}
\title{Equal ==}
\usage{
Expr_eq(other)

\method{==}{Expr}(e1, e2)
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
eq method and operator
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[rpolars]{docs_translations}}
}
\examples{
#' #three syntaxes same result
pl$lit(2) == 2
pl$lit(2) ==  pl$lit(2)
pl$lit(2)$eq(pl$lit(2))
}
\keyword{Expr}
\keyword{Expr_operators}
