% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_eq}
\alias{Expr_eq}
\alias{==.Expr}
\title{Check equality}
\usage{
Expr_eq(other)

\method{==}{Expr}(e1, e2)
}
\arguments{
\item{other}{Literal or object that can be converted to a literal}

\item{e1}{Expr only}

\item{e2}{Expr or anything that can be converted to a literal}
}
\value{
Expr
}
\description{
The RHS can either be an Expr or an object that can be converted to a literal
(e.g an integer).
}
\examples{
pl$lit(2) == 2
pl$lit(2) == pl$lit(2)
pl$lit(2)$eq(pl$lit(2))
}
\seealso{
\link{Expr_eq_missing}
}
