% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_eq_missing}
\alias{Expr_eq_missing}
\title{Check equality without \code{null} propagation}
\usage{
<Expr>$eq_missing(other)
}
\arguments{
\item{other}{Literal or object that can be converted to a literal}
}
\value{
Expr
}
\description{
The RHS can either be an Expr or an object that can be converted to a literal
(e.g an integer).
}
\examples{
df = pl$DataFrame(x = c(NA, FALSE, TRUE), y = c(TRUE, TRUE, TRUE))
df$with_columns(
  eq = pl$col("x")$eq("y"),
  eq_missing = pl$col("x")$eq_missing("y")
)
}
\seealso{
\link{Expr_eq}
}
