% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_explode}
\alias{Expr_explode}
\title{Explode a list or String Series}
\usage{
<Expr>$explode
}
\value{
Expr
}
\description{
This means that every item is expanded to a new row.
}
\details{
Categorical values are not supported.
}
\examples{
df = pl$DataFrame(x = c("abc", "ab"), y = c(list(1:3), list(3:5)))
df

df$select(pl$col("y")$explode())
}
\keyword{datasets}
