% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_extend_constant}
\alias{Expr_extend_constant}
\title{Extend Series with a constant}
\usage{
<Expr>$extend_constant(value, n)
}
\arguments{
\item{value}{The value to extend the Series with. This value may be \code{NULL} to
fill with nulls.}

\item{n}{The number of values to extend.}
}
\value{
Expr
}
\description{
Extend the Series with given number of values.
}
\examples{
pl$select(pl$lit(1:4)$extend_constant(10.1, 2))
pl$select(pl$lit(1:4)$extend_constant(NULL, 2))
}
