% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_first}
\alias{Expr_first}
\title{First}
\usage{
Expr_first
}
\value{
Expr
}
\description{
Get the first value.
Similar to R head(x,1)
}
\examples{
pl$DataFrame(list(x = c(1, 2, 3)))$select(pl$col("x")$first())
}
\keyword{datasets}
