% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_flatten}
\alias{Expr_flatten}
\title{Explode a list or String Series}
\usage{
Expr_flatten()
}
\value{
Expr
}
\description{
This is an alias for \verb{<Expr>$explode()}.
}
\examples{
df = pl$DataFrame(x = c("abc", "ab"), y = c(list(1:3), list(3:5)))
df

df$select(pl$col("y")$flatten())
}
