% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_forward_fill}
\alias{Expr_forward_fill}
\alias{forward_fill}
\title{Fill Nulls Forward}
\format{
a method
}
\usage{
Expr_forward_fill(limit = NULL)
}
\arguments{
\item{limit}{Expr or \verb{Into<Expr>}  The number of consecutive null values to forward fill.}
}
\value{
Expr
}
\description{
Fill missing values with last seen values.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[rpolars]{docs_translations}}
}
\examples{
l = list(a=c(1L,rep(NA_integer_,3L),10))
pl$DataFrame(l)$select(
  pl$col("a")$forward_fill()$alias("ff_null"),
  pl$col("a")$forward_fill(limit = 0)$alias("ff_l0"),
  pl$col("a")$forward_fill(limit = 1)$alias("ff_l1")
)$to_list()
}
\keyword{Expr}
