% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_gt}
\alias{Expr_gt}
\alias{>.RPolarsExpr}
\title{Check strictly greater inequality}
\usage{
Expr_gt(other)

\method{>}{RPolarsExpr}(e1, e2)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(2) > 1
pl$lit(2) > pl$lit(1)
pl$lit(2)$gt(pl$lit(1))
}
