% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_gt_eq}
\alias{Expr_gt_eq}
\alias{>=.Expr}
\title{Greater Than Or Equal <=}
\usage{
Expr_gt_eq(other)

\method{>=}{Expr}(e1, e2)
}
\arguments{
\item{other}{literal or Robj which can become a literal}

\item{e1}{lhs Expr}

\item{e2}{rhs Expr or anything which can become a literal Expression}
}
\value{
Exprs
}
\description{
gt_eq method and operator
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[polars]{docs_translations}}
}
\examples{
#' #three syntaxes same result
pl$lit(2) >= 2
pl$lit(2) >=  pl$lit(2)
pl$lit(2)$gt_eq(pl$lit(2))
}
\keyword{Expr}
\keyword{Expr_operators}
