% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_head}
\alias{Expr_head}
\title{Head}
\usage{
Expr_head(n = 10)
}
\arguments{
\item{n}{numeric number of elements to select from head}
}
\value{
Expr
}
\description{
Get the head n elements.
Similar to R head(x)
}
\examples{
#get 3 first elements
pl$DataFrame(list(x=1:11))$select(pl$col("x")$head(3))
}
\keyword{Expr}
