% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_inspect}
\alias{Expr_inspect}
\title{Inspect evaluated Series}
\usage{
Expr_inspect(fmt = "{}")
}
\arguments{
\item{fmt}{format string, should contain one set of \code{{}} where object will be
printed. This formatting mimics python "string".format() use in py-polars.}
}
\value{
Expr
}
\description{
Print the value that this expression evaluates to and pass on the value.
The printing will happen when the expression evaluates, not when it is formed.
}
\examples{
pl$select(pl$lit(1:5)$inspect(
  "Here's what the Series looked like before keeping the first two values: {}"
)$head(2))
}
