% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_between}
\alias{Expr_is_between}
\title{Check if an expression is between the given lower and upper bounds}
\usage{
Expr_is_between(lower_bound, upper_bound, closed = "both")
}
\arguments{
\item{lower_bound}{Lower bound, can be an Expr. Strings are parsed as column
names.}

\item{upper_bound}{Upper bound, can be an Expr. Strings are parsed as column
names.}

\item{closed}{Define which sides of the interval are closed (inclusive). This
can be either \code{"left"}, \code{"right"}, \code{"both"} or \code{"none"}.}
}
\value{
Expr
}
\description{
Check if an expression is between the given lower and upper bounds
}
\details{
Note that in polars, \code{NaN} are equal to other \code{NaN}s, and greater than any
non-\code{NaN} value.
}
\examples{
df = pl$DataFrame(num = 1:5)
df$with_columns(
  is_between = pl$col("num")$is_between(2, 4),
  is_between_excl_upper = pl$col("num")$is_between(2, 4, closed = "left"),
  is_between_excl_both = pl$col("num")$is_between(2, 4, closed = "none")
)

# lower and upper bounds can also be column names or expr
df = pl$DataFrame(
  num = 1:5,
  lower = c(0, 2, 3, 3, 3),
  upper = c(6, 4, 4, 8, 3.5)
)
df$with_columns(
  is_between_cols = pl$col("num")$is_between("lower", "upper"),
  is_between_expr = pl$col("num")$is_between(pl$col("lower") / 2, "upper")
)
}
