% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_duplicated}
\alias{Expr_is_duplicated}
\alias{is_duplicated}
\title{Get mask of duplicated values.}
\format{
a method
}
\usage{
Expr_is_duplicated
}
\value{
Expr (boolean)
}
\description{
Get mask of duplicated values.
}
\details{
is_duplicated is the opposite of \code{is_unique()}
Looking for R like \code{duplicated()}?, use  \code{some_expr$is_first()$is_not()}
}
\examples{
v = c(1,1,2,2,3,NA,NaN,Inf)
all.equal(
  pl$select(
    pl$lit(v)$is_unique()$alias("is_unique"),
    pl$lit(v)$is_first()$alias("is_first"),
    pl$lit(v)$is_duplicated()$alias("is_duplicated"),
    pl$lit(v)$is_first()$is_not()$alias("R_duplicated"),
  )$to_list(),
  list(
    is_unique = !v \%in\% v[duplicated(v)],
    is_first  = !duplicated(v),
    is_duplicated = v \%in\% v[duplicated(v)],
    R_duplicated = duplicated(v)
  )
)
}
\keyword{Expr}
