% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_first}
\alias{Expr_is_first}
\title{Get a mask of the first unique value.}
\format{
a method
}
\usage{
Expr_is_first
}
\value{
Expr (boolean)
}
\description{
Get a mask of the first unique value.
}
\examples{
v = c(1,1,2,2,3,NA,NaN,Inf)
all.equal(
  pl$select(
    pl$lit(v)$is_unique()$alias("is_unique"),
    pl$lit(v)$is_first()$alias("is_first"),
    pl$lit(v)$is_duplicated()$alias("is_duplicated"),
    pl$lit(v)$is_first()$is_not()$alias("R_duplicated"),
  )$to_list(),
  list(
    is_unique = !v \%in\% v[duplicated(v)],
    is_first  = !duplicated(v),
    is_duplicated = v \%in\% v[duplicated(v)],
    R_duplicated = duplicated(v)
  )
)
}
\keyword{Expr}
