% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_first}
\alias{Expr_is_first}
\title{Check whether each value is the first occurrence}
\usage{
Expr_is_first
}
\value{
Expr
}
\description{
Check whether each value is the first occurrence
}
\examples{
pl$DataFrame(head(mtcars[, 1:2]))$
  with_columns(is_ufirst = pl$col("mpg")$is_first())
}
\keyword{datasets}
