% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_in}
\alias{Expr_is_in}
\title{is_in}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_is_in
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
Divide
}
\examples{

#R Na_integer -> polars Null(Int32) is in polars Null(Int32)
pl$DataFrame(list(a=c(1:4,NA_integer_)))$select(pl$col("a")$is_in(pl$lit(NA_real_)))$as_data_frame()[[1]]



}
\keyword{Expr}
\keyword{Expr_operators}
