% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_infinite}
\alias{Expr_is_infinite}
\alias{is_infinite}
\title{Are elements infinite}
\format{
a method
}
\usage{
Expr_is_infinite
}
\value{
Expr
}
\description{
Returns a boolean output indicating which values are infinite.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[minipolars]{docs_translations}}
}
\examples{
pl$DataFrame(list(alice=c(0,NaN,NA,Inf,-Inf)))$select(pl$col("alice")$is_infinite())
}
\keyword{Expr}
