% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_infinite}
\alias{Expr_is_infinite}
\alias{is_infinite}
\title{Check if elements are infinite}
\usage{
Expr_is_infinite
}
\value{
Expr
}
\description{
Returns a boolean Series indicating which values are infinite.
}
\examples{
pl$DataFrame(list(alice = c(0, NaN, NA, Inf, -Inf)))$
  with_columns(infinite = pl$col("alice")$is_infinite())
}
\keyword{datasets}
