% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_nan}
\alias{Expr_is_nan}
\alias{is_nan}
\title{Are elements NaN's}
\format{
a method
}
\usage{
Expr_is_nan
}
\value{
Expr
}
\description{
Returns a boolean Series indicating which values are NaN.
}
\details{
Floating point NaN's are a different flag from Null(polars) which is the same as
NA_real_(R).
}
\examples{
pl$DataFrame(list(alice=c(0,NaN,NA,Inf,-Inf)))$select(pl$col("alice")$is_nan())
}
\keyword{Expr}
