% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_not}
\alias{Expr_is_not}
\alias{!.Expr}
\title{Not !}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_is_not(other)

\method{!}{Expr}(x)
}
\arguments{
\item{x}{Expr}

\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
not method and operator
}
\examples{
#two syntaxes same result
pl$lit(TRUE)$is_not()
!pl$lit(TRUE)
}
\keyword{Expr}
\keyword{Expr_operators}
