% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_not_null}
\alias{Expr_is_not_null}
\title{Find which elements are not NULL}
\usage{
Expr_is_not_null
}
\value{
Expr
}
\description{
Syntactic sugar for \verb{$is_null()$not_()}.
}
\examples{
pl$DataFrame(list(x = c(1, NA, 3)))$select(pl$col("x")$is_not_null())
}
\keyword{datasets}
