% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_not_null}
\alias{Expr_is_not_null}
\title{is_not_null}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_is_not_null
}
\value{
Expr
}
\description{
Returns a boolean Series indicating which values are not null.
Similar to R syntax !is.na(x)
null polars about the same as R NA
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[minipolars]{docs_translations}}
}
\examples{
pl$DataFrame(list(x=c(1,NA,3)))$select(pl$col("x")$is_not_null())
}
\keyword{Expr}
