% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_is_null}
\alias{Expr_is_null}
\title{Find which elements are NULL}
\usage{
Expr_is_null
}
\value{
Expr
}
\description{
Returns a boolean Series indicating which values are null.
}
\examples{
pl$DataFrame(list(x = c(1, NA, 3)))$select(pl$col("x")$is_null())
}
\keyword{datasets}
