% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_is_unique}
\alias{Expr_is_unique}
\alias{is_unique}
\title{Get mask of unique values}
\format{
a method
}
\usage{
Expr_is_unique
}
\value{
Expr (boolean)
}
\description{
Get mask of unique values
}
\examples{
pl$empty_select(pl$lit(abs(-2:2))$is_unique())
}
\keyword{Expr}
