% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_keep_name}
\alias{Expr_keep_name}
\alias{keep_name}
\title{Keep the original root name of the expression.}
\format{
a method
}
\usage{
Expr_keep_name
}
\value{
Expr
}
\description{
Keep the original root name of the expression.
}
\examples{
pl$DataFrame(list(alice=1:3))$select(pl$col("alice")$alias("bob")$keep_name())
}
\keyword{Expr}
