% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_last}
\alias{Expr_last}
\title{Last}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_last
}
\value{
Expr
}
\description{
Get the lastvalue.
Similar to R syntax tail(x,1)
}
\examples{
pl$DataFrame(list(x=c(1,2,3)))$select(pl$col("x")$last())
}
\keyword{Expr}
