% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_last}
\alias{Expr_last}
\title{Last}
\usage{
Expr_last
}
\value{
Expr
}
\description{
Get the lastvalue.
Similar to R syntax tail(x,1)
}
\examples{
pl$DataFrame(list(x = c(1, 2, 3)))$select(pl$col("x")$last())
}
\keyword{Expr}
