% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_limit}
\alias{Expr_limit}
\title{Limit}
\usage{
Expr_limit(n = 10)
}
\arguments{
\item{n}{numeric number of elements to select from head}
}
\value{
Expr
}
\description{
Alias for Head
Get the head n elements.
Similar to R head(x)
}
\examples{
# get 3 first elements
pl$DataFrame(list(x = 1:11))$select(pl$col("x")$limit(3))
}
\keyword{Expr}
