% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_agg}
\alias{expr_list_agg}
\title{Run any polars aggregation expression against the lists' elements}
\usage{
expr_list_agg(expr)
}
\arguments{
\item{expr}{Expression to run. Note that you can select an element with
\code{pl$element()}, \code{pl$first()}, and more. See Examples.}
}
\value{
A polars \link{expression}
}
\description{
This looks similar to \code{\link[=expr_list_eval]{$list$eval()}}, but the key
difference is that \verb{$list$agg()} automatically explodes the list if the
expression inside returns a scalar (while \verb{$list$eval()} always returns a
list).
}
\examples{
df <- pl$DataFrame(a = list(c(1, NA), c(42, 13), c(NA, NA)))

# The column "null_count" has dtype u32 because `$null_count()` returns a
# scalar for each sub-list. Using `$list$eval()` instead would return a
# column with dtype list(u32).
df$with_columns(
  null_count = pl$col("a")$list$agg(pl$element()$null_count())
)

# The column "no_nulls" has dtype list(u32) because the expression doesn't
# guarantee to return a scalar.
df$with_columns(
  no_nulls = pl$col("a")$list$agg(pl$element()$drop_nulls())
)
}
