% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_all}
\alias{expr_list_all}
\title{Evaluate whether all boolean values in a sub-list are true}
\usage{
expr_list_all()
}
\value{
A polars \link{expression}
}
\description{
Evaluate whether all boolean values in a sub-list are true
}
\examples{
df <- pl$DataFrame(
  a = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), NA, c())
)
df$with_columns(all = pl$col("a")$list$all())
}
