% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_any}
\alias{expr_list_any}
\title{Evaluate whether any boolean value in a sub-list is true}
\usage{
expr_list_any()
}
\value{
A polars \link{expression}
}
\description{
Evaluate whether any boolean value in a sub-list is true
}
\examples{
df <- pl$DataFrame(
  a = list(c(TRUE, TRUE), c(FALSE, TRUE), c(FALSE, FALSE), NA, c())
)
df$with_columns(any = pl$col("a")$list$any())
}
