% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_arg_max}
\alias{expr_list_arg_max}
\title{Get the index of the maximal value in list}
\usage{
expr_list_arg_max()
}
\value{
A polars \link{expression}
}
\description{
Get the index of the maximal value in list
}
\examples{
df <- pl$DataFrame(s = list(1:2, 2:1))
df$with_columns(
  arg_max = pl$col("s")$list$arg_max()
)
}
