% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_arg_min}
\alias{expr_list_arg_min}
\title{Get the index of the minimal value in list}
\usage{
expr_list_arg_min()
}
\value{
A polars \link{expression}
}
\description{
Get the index of the minimal value in list
}
\examples{
df <- pl$DataFrame(s = list(1:2, 2:1))
df$with_columns(
  arg_min = pl$col("s")$list$arg_min()
)
}
