% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_contains}
\alias{expr_list_contains}
\title{Check if sub-lists contains a given value}
\usage{
expr_list_contains(item)
}
\arguments{
\item{item}{Item that will be checked for membership. Can be an Expr or
something coercible to an Expr. Strings are not parsed as columns.}
}
\value{
A polars \link{expression}
}
\description{
Check if sub-lists contains a given value
}
\examples{
df <- pl$DataFrame(
  a = list(3:1, NULL, 1:2),
  item = 0:2
)
df$with_columns(
  with_expr = pl$col("a")$list$contains(pl$col("item")),
  with_lit = pl$col("a")$list$contains(1)
)
}
