% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_explode}
\alias{expr_list_explode}
\title{Returns a column with a separate row for every list element}
\usage{
expr_list_explode(..., empty_as_null = TRUE, keep_nulls = TRUE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{empty_as_null}{Indicates to explode an empty list/array into a \code{null}.
Defaults to \code{TRUE}.}

\item{keep_nulls}{Indicates to explode a \code{null} list/array into a \code{null}.}
}
\value{
A polars \link{expression}
}
\description{
Returns a column with a separate row for every list element
}
\examples{
df <- pl$DataFrame(a = list(c(1, 2, 3), c(4, 5, 6)))
df$select(pl$col("a")$list$explode())
}
