% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_explode}
\alias{expr_list_explode}
\title{Returns a column with a separate row for every list element}
\usage{
expr_list_explode()
}
\value{
A polars \link{expression}
}
\description{
Returns a column with a separate row for every list element
}
\examples{
df <- pl$DataFrame(a = list(c(1, 2, 3), c(4, 5, 6)))
df$select(pl$col("a")$list$explode())
}
