% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_gather_every}
\alias{expr_list_gather_every}
\title{Take every \code{n}-th value starting from offset in sub-lists}
\usage{
expr_list_gather_every(n, offset = 0)
}
\value{
A polars \link{expression}
}
\description{
Take every \code{n}-th value starting from offset in sub-lists
}
\examples{
df <- pl$DataFrame(
  a = list(1:5, 6:8, 9:12),
  n = c(2, 1, 3),
  offset = c(0, 1, 0)
)

df$with_columns(
  gather_every = pl$col("a")$list$gather_every(pl$col("n"), offset = pl$col("offset"))
)
}
