% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_join}
\alias{expr_list_join}
\title{Join elements of a list}
\usage{
expr_list_join(separator, ignore_nulls = FALSE)
}
\arguments{
\item{separator}{String to separate the items with. Can be an Expr. Strings
are \emph{not} parsed as columns.}
}
\value{
A polars \link{expression}
}
\description{
Join all string items in a sublist and place a separator between them. This
only works on columns of type \code{list[str]}.
}
\examples{
df <- pl$DataFrame(
  s = list(c("a", "b", "c"), c("x", "y"), c("e", NA)),
  separator = c("-", "+", "/")
)
df$with_columns(
  join_with_expr = pl$col("s")$list$join(pl$col("separator")),
  join_with_lit = pl$col("s")$list$join(" "),
  join_ignore_null = pl$col("s")$list$join(" ", ignore_nulls = TRUE)
)
}
