% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_last}
\alias{expr_list_last}
\title{Get the last value of the sub-lists}
\usage{
expr_list_last()
}
\value{
A polars \link{expression}
}
\description{
Get the last value of the sub-lists
}
\examples{
df <- pl$DataFrame(a = list(3:1, NULL, 1:2))
df$with_columns(
  last = pl$col("a")$list$last()
)
}
