% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_len}
\alias{expr_list_len}
\title{Get the length of each list}
\usage{
expr_list_len()
}
\value{
A polars \link{expression}
}
\description{
Return the number of elements in each list. Null values are counted in the
total.
}
\examples{
df <- pl$DataFrame(list_of_strs = list(c("a", "b", NA), "c"))
df$with_columns(len_list = pl$col("list_of_strs")$list$len())
}
