% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_min}
\alias{expr_list_min}
\title{Find the minimum value in a list}
\usage{
expr_list_min()
}
\value{
A polars \link{expression}
}
\description{
Find the minimum value in a list
}
\examples{
df <- pl$DataFrame(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_))
df$with_columns(min = pl$col("values")$list$min())
}
