% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_sample}
\alias{expr_list_sample}
\title{Sample values from every sub-list}
\usage{
expr_list_sample(
  n = NULL,
  ...,
  fraction = NULL,
  with_replacement = FALSE,
  shuffle = NULL,
  seed = NULL
)
}
\arguments{
\item{n}{Number of items to return. Can be an Expr. Strings are parsed as
column names. Cannot be used with \code{fraction}. Defaults to 1 if \code{fraction}
is \code{NULL}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{fraction}{Fraction of items to return. Can be an Expr. Strings are
parsed as column names. Cannot be used with \code{n}.}

\item{with_replacement}{Allow values to be sampled more than once.}

\item{shuffle}{If \code{TRUE}, explicitly shuffle the sampled data points. If
\code{FALSE}, maintain their relative order. If \code{NULL} (default), use a
high-performance algorithm without guaranteeing an order.}

\item{seed}{Seed for the random number generator. If \code{NULL} (default), a
random seed is generated for each sample operation.}
}
\value{
A polars \link{expression}
}
\description{
Sample values from every sub-list
}
\examples{
df <- pl$DataFrame(
  values = list(1:3, NA, c(NA, 3L), 5:7),
  n = c(1, 1, 1, 2)
)

df$with_columns(
  sample = pl$col("values")$list$sample(
    n = pl$col("n"), seed = 1
  )
)
}
