% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_set_intersection}
\alias{expr_list_set_intersection}
\title{Compute the intersection between elements of a list and other elements}
\usage{
expr_list_set_intersection(other)
}
\arguments{
\item{other}{Other list variable. Can be an Expr or something coercible to an
Expr.}
}
\value{
A polars \link{expression}
}
\description{
Compute the intersection between elements of a list and other elements
}
\details{
Note that the datatypes inside the list must have a common supertype. For
example, the first column can be \code{list[i32]} and the second one can be
\code{list[i8]} because it can be cast to \code{list[i32]}. However, the second column
cannot be e.g \code{list[f32]}.
}
\examples{
df <- pl$DataFrame(
  a = list(1:3, NA, c(NA, 3L), 5:7),
  b = list(2:4, 3L, c(3L, 4L, NA), c(6L, 8L))
)

df$with_columns(intersection = pl$col("a")$list$set_intersection("b"))
}
