% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_shift}
\alias{expr_list_shift}
\title{Shift list values by the given number of indices}
\usage{
expr_list_shift(n = 1)
}
\value{
A polars \link{expression}
}
\description{
Shift list values by the given number of indices
}
\examples{
df <- pl$DataFrame(
  s = list(1:4, c(10L, 2L, 1L)),
  idx = 1:2
)
df$with_columns(
  shift_by_expr = pl$col("s")$list$shift(pl$col("idx")),
  shift_by_lit = pl$col("s")$list$shift(2),
  shift_by_negative_lit = pl$col("s")$list$shift(-2)
)
}
