% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_std}
\alias{expr_list_std}
\title{Compute the standard deviation in every sub-list}
\usage{
expr_list_std(ddof = 1)
}
\arguments{
\item{"Delta}{Degrees of Freedom": the divisor used in the calculation is
\code{N - ddof}, where \code{N} represents the number of elements. By default ddof is
1.}
}
\value{
A polars \link{expression}
}
\description{
Compute the standard deviation in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(-1, 0, 1), c(1, 10)))

df$with_columns(
  std = pl$col("values")$list$std()
)
}
