% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_sum}
\alias{expr_list_sum}
\title{Sum all elements in every sub-list}
\usage{
expr_list_sum()
}
\value{
A polars \link{expression}
}
\description{
Sum all elements in every sub-list
}
\examples{
df <- pl$DataFrame(values = list(c(1, 2, 3, NA), c(2, 3), NA))
df$with_columns(sum = pl$col("values")$list$sum())
}
