% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-list.R
\name{expr_list_to_array}
\alias{expr_list_to_array}
\title{Convert a List column into an Array column with the same inner data type}
\usage{
expr_list_to_array(width)
}
\arguments{
\item{width}{Width of the resulting Array column.}
}
\value{
A polars \link{expression}
}
\description{
Convert a List column into an Array column with the same inner data type
}
\examples{
df <- pl$DataFrame(values = list(c(-1, 0), c(1, 10)))

df$with_columns(
  array = pl$col("values")$list$to_array(2)
)
}
