% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_lit_to_df}
\alias{Expr_lit_to_df}
\title{Convert Literal to DataFrame}
\usage{
Expr_lit_to_df()
}
\value{
Series
}
\description{
Collect an expression based on literals into a DataFrame.
}
\examples{
pl$lit(1:5)$lit_to_df()
}
