% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_log}
\alias{Expr_log}
\title{Compute the logarithm of elements}
\usage{
Expr_log(base = base::exp(1))
}
\arguments{
\item{base}{Numeric base value for logarithm, default is \code{exp(1)}.}
}
\value{
Expr
}
\description{
Compute the logarithm of elements
}
\examples{
pl$DataFrame(a = c(1, 2, 3, exp(1)))$
  with_columns(log = pl$col("a")$log())
}
