% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_log10}
\alias{Expr_log10}
\title{Compute the base-10 logarithm of elements}
\usage{
Expr_log10
}
\value{
Expr
}
\description{
Compute the base-10 logarithm of elements
}
\examples{
pl$DataFrame(a = c(1, 2, 3, exp(1)))$
  with_columns(log10 = pl$col("a")$log10())
}
\keyword{datasets}
