% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_lt}
\alias{Expr_lt}
\alias{<.Expr}
\title{Less Than <}
\usage{
Expr_lt(other)

\method{<}{Expr}(e1, e2)
}
\arguments{
\item{other}{literal or Robj which can become a literal}

\item{e1}{lhs Expr}

\item{e2}{rhs Expr or anything which can become a literal Expression}
}
\value{
Exprs
}
\description{
lt method and operator
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[rpolars]{docs_translations}}
}
\examples{
#' #three syntaxes same result
pl$lit(5) < 10
pl$lit(5) < pl$lit(10)
pl$lit(5)$lt(pl$lit(10))
}
\keyword{Expr}
\keyword{Expr_operators}
