% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_lt}
\alias{Expr_lt}
\title{Less Than <}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_lt
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
lt method and operator
}
\examples{
#' #three syntaxes same result
pl$lit(5) < 10
pl$lit(5) < pl$lit(10)
pl$lit(5)$lt(pl$lit(10))
}
\keyword{Expr}
\keyword{Expr_operators}
