% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{Expr_map}
\alias{Expr_map}
\title{polars map}
\usage{
Expr_map(lambda, output_type = NULL, `_agg_list` = NULL)
}
\arguments{
\item{lambda}{r function mapping a series}

\item{output_type}{NULL or one of pl$dtypes, the output datatype, NULL is the same as input.}

\item{`_agg_list`}{#not implemented yet}
}
\value{
Expr
}
\description{
polars map
}
\details{
in minipolars lambda return should be a series or any R vector convertable into a Series. In PyPolars likely return must be Series.
}
\examples{
pl$DataFrame(iris)$select(pl$col("Sepal.Length")$map(\(x) paste("cheese",as.character(x$to_r_vector())),pl$dtypes$Utf8))
}
