% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_map}
\alias{Expr_map}
\title{Expr_map}
\usage{
Expr_map(f, output_type = NULL, agg_list = FALSE)
}
\arguments{
\item{f}{a function mapping a series}

\item{output_type}{NULL or one of pl$dtypes$..., the output datatype, NULL is the same as input.}

\item{agg_list}{Aggregate list. Map from vector to group in groupby context.
Likely not so useful.}
}
\value{
Expr
}
\description{
Expr_map
}
\details{
user function return should be a series or any Robj convertable into a Series.
In PyPolars likely return must be Series. User functions do fully support \code{browser()}, helpful to
investigate.
}
\examples{
pl$DataFrame(iris)$select(pl$col("Sepal.Length")$map(\(x) {
  paste("cheese",as.character(x$to_vector()))
}, pl$dtypes$Utf8))
}
\keyword{Expr}
