% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_max}
\alias{Expr_max}
\title{max}
\usage{
Expr_max
}
\value{
Expr
}
\description{
Get maximum value.
}
\examples{
pl$DataFrame(list(x = c(1, NA, 3)))$select(pl$col("x")$max() == 3) # is true
}
\keyword{datasets}
