% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_max}
\alias{Expr_max}
\title{Get maximum value}
\usage{
Expr_max()
}
\value{
Expr
}
\description{
Get maximum value
}
\examples{
pl$DataFrame(x = c(1, NA, 3))$
  with_columns(max = pl$col("x")$max())
}
