% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_mean}
\alias{Expr_mean}
\title{mean}
\usage{
Expr_mean
}
\value{
Expr
}
\description{
Get mean value.
}
\examples{
pl$DataFrame(list(x=c(1,NA,3)))$select(pl$col("x")$mean()==2) #is true
}
\keyword{Expr}
