% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_mean}
\alias{Expr_mean}
\title{Get mean value}
\usage{
Expr_mean()
}
\value{
Expr
}
\description{
Get mean value
}
\examples{
pl$DataFrame(x = c(1L, NA, 2L))$
  with_columns(mean = pl$col("x")$mean())
}
