% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_eq}
\alias{expr_meta_eq}
\title{Indicate if this expression is the same as another expression}
\usage{
expr_meta_eq(other)
}
\value{
A polars \link{expression}
}
\description{
Indicate if this expression is the same as another expression
}
\examples{
foo_bar <- pl$col("foo")$alias("bar")
foo <- pl$col("foo")
foo_bar$meta$eq(foo)

foo_bar2 <- pl$col("foo")$alias("bar")
foo_bar$meta$eq(foo_bar2)
}
