% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_is_column_selection}
\alias{expr_meta_is_column_selection}
\title{Indicate if this expression only selects columns (optionally with aliasing)}
\usage{
expr_meta_is_column_selection(..., allow_aliasing = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{allow_aliasing}{If \code{FALSE} (default), any aliasing is not considered
pure column selection. Set \code{TRUE} to allow for column selection that also
includes aliasing.}
}
\value{
A polars \link{expression}
}
\description{
This can include bare columns, column matches by regex or dtype, selectors
and exclude ops, and (optionally) column/expression aliasing.
}
\examples{
e <- pl$col("foo")
e$meta$is_column_selection()

e <- pl$col("foo")$alias("bar")
e$meta$is_column_selection()

e$meta$is_column_selection(allow_aliasing = TRUE)

e <- pl$col("foo") * pl$col("bar")
e$meta$is_column_selection()
}
