% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-meta.R
\name{expr_meta_is_literal}
\alias{expr_meta_is_literal}
\title{Indicate if this expression is a literal value (optionally aliased)}
\usage{
expr_meta_is_literal(..., allow_aliasing = NULL)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{allow_aliasing}{If \code{FALSE} (default), only a bare literal will match.
Set to \code{TRUE} to also allow for aliased literals.}
}
\value{
A polars \link{expression}
}
\description{
Indicate if this expression is a literal value (optionally aliased)
}
\examples{
e <- pl$lit(123)
e$meta$is_literal()

e <- pl$lit(123)$alias("foo")
e$meta$is_literal()
e$meta$is_literal(allow_aliasing = TRUE)
}
