% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_min}
\alias{Expr_min}
\title{min}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_min
}
\value{
Expr
}
\description{
Get minimum value.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[rpolars]{docs_translations}}
}
\examples{
pl$DataFrame(list(x=c(1,NA,3)))$select(pl$col("x")$min()== 1 ) #is true
}
\keyword{Expr}
