% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_min}
\alias{Expr_min}
\title{Get minimum value}
\usage{
<Expr>$min
}
\value{
Expr
}
\description{
Get minimum value
}
\examples{
pl$DataFrame(x = c(1, NA, 3))$
  with_columns(min = pl$col("x")$min())
}
\keyword{datasets}
