% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-name.R
\name{expr_name_prefix}
\alias{expr_name_prefix}
\title{Add a prefix to a column name}
\usage{
expr_name_prefix(prefix)
}
\arguments{
\item{prefix}{Prefix to be added to column name(s)}
}
\value{
A polars \link{expression}
}
\description{
Add a prefix to a column name
}
\examples{
dat <- as_polars_df(mtcars)

dat$select(
  pl$col("mpg"),
  pl$col("mpg")$name$prefix("name_"),
  pl$col("cyl", "drat")$name$prefix("bar_")
)
}
\seealso{
\code{\link[=expr_name_suffix]{$suffix()}} to add a suffix
}
