% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-name.R
\name{expr_name_replace}
\alias{expr_name_replace}
\title{Replace matching regex/literal substring in the name with a new value}
\usage{
expr_name_replace(pattern, value, ..., literal = FALSE)
}
\arguments{
\item{pattern}{A valid regular expression pattern, compatible with the
\verb{regex crate <https://docs.rs/regex/latest/regex/>}_.}

\item{value}{String that will replace the matched substring.}

\item{...}{These dots are for future extensions and must be empty.}

\item{literal}{Treat \code{pattern} as a literal string, not a regex.}
}
\value{
A polars \link{expression}
}
\description{
This will undo any previous renaming operations on the expression.

Due to implementation constraints, this method can only be called as the last
expression in a chain. Only one name operation per expression will work.
}
\examples{
df <- pl$DataFrame(n_foo = 1, n_bar = 2)
df$select(pl$all()$name$replace("^n_", "col_"))

df$select(pl$all()$name$replace("(a|e|i|o|u)", "@"))$schema
}
