% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-name.R
\name{expr_name_suffix_fields}
\alias{expr_name_suffix_fields}
\title{Add a suffix to all fields name of a struct}
\usage{
expr_name_suffix_fields(suffix)
}
\arguments{
\item{suffix}{Suffix to add to the field name.}
}
\value{
A polars \link{expression}
}
\description{
Add a suffix to all fields name of a struct
}
\examples{
df <- pl$DataFrame(a = 1, b = 2)$select(
  pl$struct(pl$all())$alias("my_struct")
)

df$with_columns(
  pl$col("my_struct")$name$suffix_fields("_post")
)$unnest("my_struct")
}
