% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-name.R
\name{expr_name_to_lowercase}
\alias{expr_name_to_lowercase}
\title{Make the root column name lowercase}
\usage{
expr_name_to_lowercase()
}
\value{
A polars \link{expression}
}
\description{
Due to implementation constraints, this method can only be called as the last
expression in a chain.
}
\examples{
df <- pl$DataFrame(Foo = 1:3, BAR = 4:6)
df$select(pl$all()$name$to_lowercase())
}
