% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_nan_max}
\alias{Expr_nan_max}
\title{Get maximum value with NaN}
\usage{
Expr_nan_max
}
\value{
Expr
}
\description{
Get maximum value, but returns \code{NaN} if there are any.
}
\examples{
pl$DataFrame(x = c(1, NA, 3, NaN, Inf))$
  with_columns(nan_max = pl$col("x")$nan_max())
}
\keyword{datasets}
