% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_nan_max}
\alias{Expr_nan_max}
\alias{nan_min}
\title{max}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_nan_max
}
\value{
Expr
}
\description{
Get maximum value, but propagate/poison encountered \code{NaN} values.
Get maximum value.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[rpolars]{docs_translations}}s
}
\examples{
pl$DataFrame(list(x=c(1,NaN,Inf,3)))$select(pl$col("x")$nan_max()$is_nan()) #is true
}
\keyword{Expr}
