% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_neq}
\alias{Expr_neq}
\alias{!=.RPolarsExpr}
\title{Check inequality}
\usage{
Expr_neq(other)

\method{!=}{RPolarsExpr}(e1, e2)
}
\arguments{
\item{other}{numeric or string value; accepts expression input.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of addition operator \code{expr + other}.
}
\examples{
pl$lit(1) != 2
pl$lit(1) != pl$lit(2)
pl$lit(1)$neq(pl$lit(2))
}
\seealso{
\link{Expr_neq_missing}
}
