% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_neq}
\alias{Expr_neq}
\alias{!=.Expr}
\title{Not Equal !=}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_neq

\method{!=}{Expr}(e1, e2)
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
neq method and operator
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[rpolars]{docs_translations}}
}
\examples{
#' #three syntaxes same result
pl$lit(1) != 2
pl$lit(1) !=  pl$lit(2)
pl$lit(1)$neq(pl$lit(2))
}
\keyword{Expr}
\keyword{Expr_operators}
