% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_neq}
\alias{Expr_neq}
\alias{!=.RPolarsExpr}
\title{Check inequality}
\usage{
Expr_neq(other)

\method{!=}{RPolarsExpr}(e1, e2)
}
\arguments{
\item{other}{Literal or object that can be converted to a literal}

\item{e1}{Expr only}

\item{e2}{Expr or anything that can be converted to a literal}
}
\value{
Expr
}
\description{
The RHS can either be an Expr or an object that can be converted to a literal
(e.g an integer).
}
\examples{
pl$lit(1) != 2
pl$lit(1) != pl$lit(2)
pl$lit(1)$neq(pl$lit(2))
}
\seealso{
\link{Expr_neq_missing}
}
