% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_not}
\alias{Expr_not}
\title{Negate a boolean expression}
\usage{
Expr_not()
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Method equivalent of negation operator \code{!expr}.
}
\examples{
# two syntaxes same result
pl$lit(TRUE)$not()
!pl$lit(TRUE)
}
