% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_or}
\alias{Expr_or}
\title{Apply logical OR on two expressions}
\usage{
<Expr>$or(other)
}
\arguments{
\item{other}{Literal or object that can be converted to a literal}
}
\value{
Expr
}
\description{
Combine two boolean expressions with OR.
}
\examples{
pl$lit(TRUE) | FALSE
pl$lit(TRUE)$or(pl$lit(TRUE))
}
