% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_or}
\alias{Expr_or}
\title{Or}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_or
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Expr
}
\description{
combine to boolean expresions with OR
}
\examples{
pl$lit(TRUE) | FALSE
pl$lit(TRUE)$or(pl$lit(TRUE))
}
\keyword{Expr}
\keyword{Expr_operators}
