% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_over}
\alias{Expr_over}
\title{Apply window function over a subgroup}
\usage{
<Expr>$over(...)
}
\arguments{
\item{...}{Character vector indicating the columns to group by.}
}
\value{
Expr
}
\description{
This applies an expression on groups and returns the same number of rows as
the input (contrarily to \verb{$group_by()} + \verb{$agg()}).
}
\examples{
pl$DataFrame(
  val = 1:5,
  a = c("+", "+", "-", "-", "+"),
  b = c("+", "-", "+", "-", "+")
)$with_columns(
  count = pl$col("val")$count()$over("a", "b")
)

over_vars = c("a", "b")
pl$DataFrame(
  val = 1:5,
  a = c("+", "+", "-", "-", "+"),
  b = c("+", "-", "+", "-", "+")
)$with_columns(
  count = pl$col("val")$count()$over(over_vars)
)
}
