% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_pct_change}
\alias{Expr_pct_change}
\alias{pct_change}
\title{Pct change}
\usage{
Expr_pct_change(n = 1)
}
\arguments{
\item{n}{periods to shift for forming percent change.}
}
\value{
Expr
}
\description{
Computes percentage change between values.
Percentage change (as fraction) between current element and most-recent
non-null element at least \code{n} period(s) before the current element.
Computes the change from the previous row by default.
}
\examples{
df = pl$DataFrame(list(a = c(10L, 11L, 12L, NA_integer_, 12L)))
df$with_columns(pl$col("a")$pct_change()$alias("pct_change"))
}
