% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_pow}
\alias{Expr_pow}
\alias{pow}
\title{Exponentiation \code{^} or \verb{**}}
\usage{
Expr_pow(exponent)
}
\arguments{
\item{base}{real value of base}
}
\value{
Expr
}
\description{
Raise expression to the power of exponent.
}
\examples{
pl$DataFrame(list(a = -1:3))$select(pl$lit(2)$pow(pl$col("a")))$get_column("literal")$to_r()== 2^(-1:3)
pl$DataFrame(list(a = -1:3))$select(pl$lit(2) ^ (pl$col("a")))$get_column("literal")$to_r()== 2^(-1:3)
}
\keyword{Expr}
