% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_pow}
\alias{Expr_pow}
\title{Exponentiation}
\usage{
Expr_pow(exponent)
}
\arguments{
\item{exponent}{Exponent value.}
}
\value{
Expr
}
\description{
Raise expression to the power of exponent.
}
\examples{
# use via `pow`-method and the `^`-operator
pl$DataFrame(a = -1:3, b = 2:6)$with_columns(
  x = pl$col("a")$pow(2),
  y = pl$col("a")^3
)
}
