% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_prefix}
\alias{Expr_prefix}
\alias{prefix}
\title{Add a prefix to a column name}
\usage{
Expr_prefix(prefix)
}
\arguments{
\item{prefix}{Prefix to be added to column name(s)}
}
\value{
Expr
}
\description{
Add a prefix to a column name
}
\examples{
dat = pl$DataFrame(mtcars)

dat$select(
  pl$col("mpg"),
  pl$col("mpg")$prefix("foo_"),
  pl$col("cyl", "drat")$prefix("bar_")
)
}
\seealso{
\code{\link[=Expr_suffix]{$suffix()}} to add a suffix
}
\keyword{Expr}
