% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_print}
\alias{Expr_print}
\title{internal method print Expr}
\usage{
Expr_print()
}
\value{
Prints the Polars expression to the console
}
\description{
internal method print Expr
}
\examples{
pl$col("some_column")$sum()$over("some_other_column")$print()
}
\keyword{Expr}
