% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{print.Expr}
\alias{print.Expr}
\alias{Expr_print}
\title{S3 method to print an Expr}
\usage{
\method{print}{Expr}(x, ...)

Expr_print()
}
\arguments{
\item{x}{Expr}

\item{...}{Not used.}
}
\value{
No value returned, it prints in the console.
}
\description{
S3 method to print an Expr
}
\examples{
print(pl$col("some_column")$sum())
}
