% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_product}
\alias{Expr_product}
\title{Product}
\usage{
Expr_product
}
\value{
Expr
}
\description{
Compute the product of an expression.
}
\examples{
pl$DataFrame(x = c(2L, NA, 2L))$
  with_columns(product = pl$col("x")$product())
}
\keyword{datasets}
