% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_product}
\alias{Expr_product}
\alias{Product}
\title{Product}
\usage{
Expr_product
}
\value{
Expr
}
\description{
Compute the product of an expression.
}
\details{
does not support integer32 currently, .cast() to f64 or i64 first.
}
\examples{
pl$DataFrame(list(x=c(1,2,3)))$select(pl$col("x")$product()==6) #is true
}
\keyword{Expr}
