% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rechunk}
\alias{Expr_rechunk}
\alias{rechunk}
\title{Rechunk memory layout}
\usage{
Expr_rechunk
}
\value{
Expr
}
\description{
Create a single chunk of memory for this Series.
}
\details{
See rechunk() explained here \code{\link[polars]{docs_translations}}
}
\examples{
#get chunked lengths with/without rechunk
series_list = pl$DataFrame(list(a=1:3,b=4:6))$select(
  pl$col("a")$append(pl$col("b"))$alias("a_chunked"),
  pl$col("a")$append(pl$col("b"))$rechunk()$alias("a_rechunked")
)$get_columns()
lapply(series_list, \(x) x$chunk_lengths())
}
\keyword{Expr}
