% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rep}
\alias{Expr_rep}
\title{Repeat a Series}
\usage{
<Expr>$rep(n, rechunk = TRUE)
}
\arguments{
\item{n}{The number of times to repeat, must be non-negative and finite.}

\item{rechunk}{If \code{TRUE} (default), memory layout will be rewritten.}
}
\value{
Expr
}
\description{
This expression takes input and repeats it n times and append chunk.
}
\details{
If the input has length 1, this uses a special faster implementation that
doesn't require rechunking (so \code{rechunk = TRUE} has no effect).
}
\examples{
pl$select(pl$lit("alice")$rep(n = 3))
pl$select(pl$lit(1:3)$rep(n = 2))
}
