% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_repeat_by}
\alias{Expr_repeat_by}
\title{Repeat by}
\usage{
Expr_repeat_by(by)
}
\arguments{
\item{by}{Expr Numeric column that determines how often the values will be repeated.
The column will be coerced to UInt32. Give this dtype to make the coercion a
no-op.}
}
\value{
Expr
}
\description{
Repeat the elements in this Series as specified in the given expression.
The repeated elements are expanded into a \code{List}.
}
\examples{
df = pl$DataFrame(list(a = c("x", "y", "z"), n = c(0:2)))
df$select(pl$col("a")$repeat_by("n"))
}
