% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_reshape}
\alias{Expr_reshape}
\alias{arctanh}
\title{Reshape}
\format{
Method
}
\usage{
Expr_reshape(dims)
}
\value{
Expr
}
\description{
Reshape this Expr to a flat Series or a Series of Lists.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a=c(-1,tanh(0.5),0,1,NA_real_))$select(pl$col("a")$arctanh())
}
\keyword{Expr}
