% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_reshape}
\alias{Expr_reshape}
\title{Reshape}
\usage{
Expr_reshape(dims)
}
\arguments{
\item{dims}{Numeric vec of the dimension sizes. If a -1 is used in any of the
dimensions, that dimension is inferred.}
}
\value{
Expr
}
\description{
Reshape an Expr to a flat Series or a Series of Lists.
}
\examples{
pl$select(pl$lit(1:12)$reshape(c(3, 4)))
pl$select(pl$lit(1:12)$reshape(c(3, -1)))
}
