% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_reshape}
\alias{Expr_reshape}
\alias{reshape}
\title{Reshape}
\format{
Method
}
\usage{
Expr_reshape(dims)
}
\arguments{
\item{dims}{numeric vec of the dimension sizes. If a -1 is used in any of the dimensions, that
dimension is inferred.}
}
\value{
Expr
}
\description{
Reshape this Expr to a flat Series or a Series of Lists.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$empty_select(pl$lit(1:12)$reshape(c(3,4)))
pl$empty_select(pl$lit(1:12)$reshape(c(3,-1)))
}
\keyword{Expr}
