% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_reverse}
\alias{Expr_reverse}
\alias{reverse}
\title{polars reverse}
\usage{
Expr_reverse()
}
\value{
Expr
}
\description{
polars reverse
}
\examples{
pl$DataFrame(list(a = 1:5))$select(pl$col("a")$reverse())
}
\keyword{Expr}
