% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_rolling_skew}
\alias{Expr_rolling_skew}
\alias{rolling_skew}
\title{Rolling skew}
\usage{
Expr_rolling_skew(window_size, bias = TRUE)
}
\arguments{
\item{window_size}{integerish, Size of the rolling window}

\item{bias}{bool default = TRUE,  If False, then the calculations are corrected for statistical bias.}
}
\value{
Expr
}
\description{
Compute a rolling skew.
}
\examples{
#hej
}
\keyword{Expr}
