% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_round}
\alias{Expr_round}
\alias{round}
\title{round}
\format{
a method
}
\usage{
Expr_round
}
\arguments{
\item{decimals}{integer Number of decimals to round by.}
}
\value{
Expr
}
\description{
Round underlying floating point data by \code{decimals} digits.
}
\examples{
pl$DataFrame(list(
  a = c(0.33, 0.5, 1.02, 1.5, NaN , NA, Inf, -Inf)
))$select(
  pl$col("a")$round(0)
)
}
\keyword{Expr}
