% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_search_sorted}
\alias{Expr_search_sorted}
\alias{search_sorted}
\title{Where to inject element(s) to maintain sorting}
\usage{
Expr_search_sorted(element)
}
\arguments{
\item{element}{a R value into literal or an expression of an element}
}
\value{
Expr
}
\description{
Find indices in self where elements should be inserted into to maintain order.
}
\details{
This function look up where to insert element if to keep self column sorted.
It is assumed the self column is already sorted ascending, otherwise wrongs answers.
This function is a bit under documented in py-polars.
}
\examples{
pl$DataFrame(list(a = 0:100))$select(pl$col("a")$search_sorted(pl$lit(42L)))
}
