% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_set_sorted}
\alias{Expr_set_sorted}
\alias{set_sorted}
\title{Set_sorted}
\usage{
Expr_set_sorted(reverse = FALSE)
}
\arguments{
\item{reverse}{bool if TRUE Descending else Ascending}
}
\value{
Expr
}
\description{
Flags the expression as 'sorted'.
}
\examples{
# correct use flag something correctly as ascendingly sorted
s = pl$select(pl$lit(1:4)$set_sorted()$alias("a"))$get_column("a")
s$flags # see flags

# incorrect use, flag somthing as not sorted ascendingly
s2 = pl$select(pl$lit(c(1, 3, 2, 4))$set_sorted()$alias("a"))$get_column("a")
s2$sort() # sorting skipped, although not actually sorted
}
\keyword{Expr}
