% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_shift}
\alias{Expr_shift}
\alias{shift}
\title{Shift values}
\usage{
Expr_shift(periods)
}
\arguments{
\item{periods}{numeric number of periods to shift, may be negative.}
}
\value{
Expr
}
\description{
Shift values
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[polars]{docs_translations}}
}
\examples{
pl$select(
  pl$lit(0:3)$shift(-2)$alias("shift-2"),
  pl$lit(0:3)$shift(2)$alias("shift+2")
)
}
\keyword{Expr}
