% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_shift}
\alias{Expr_shift}
\title{Shift values}
\usage{
<Expr>$shift(periods = 1)
}
\arguments{
\item{periods}{Number of periods to shift, may be negative.}
}
\value{
Expr
}
\description{
Shift values
}
\examples{
pl$DataFrame(a = c(1, 2, 4, 5, 8))$
  with_columns(
  pl$col("a")$shift(-2)$alias("shift-2"),
  pl$col("a")$shift(2)$alias("shift+2")
)
}
