% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sign}
\alias{Expr_sign}
\alias{sign}
\title{Sign}
\usage{
Expr_sign
}
\value{
Expr
}
\description{
Compute the element-wise indication of the sign.
}
\examples{
pl$DataFrame(a = c(.9, -0, 0, 4, NA_real_))$select(pl$col("a")$sign())
}
\keyword{Expr}
