% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sin}
\alias{Expr_sin}
\alias{sin}
\title{Sin}
\usage{
Expr_sin
}
\value{
Expr
}
\description{
Compute the element-wise value for the sine.
}
\details{
Evaluated Series has dtype Float64
}
\examples{
pl$DataFrame(a = c(0, pi / 2, pi, NA_real_))$select(pl$col("a")$sin())
}
\keyword{datasets}
