% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sinh}
\alias{Expr_sinh}
\title{Compute hyperbolic sine}
\usage{
Expr_sinh
}
\value{
Expr
}
\description{
Compute hyperbolic sine
}
\examples{
pl$DataFrame(a = c(-1, asinh(0.5), 0, 1, NA_real_))$
  with_columns(sinh = pl$col("a")$sinh())
}
\keyword{datasets}
