% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_skew}
\alias{Expr_skew}
\title{Skewness}
\usage{
Expr_skew(bias = TRUE)
}
\arguments{
\item{bias}{If \code{FALSE}, then the calculations are corrected for statistical
bias.}
}
\value{
Expr
}
\description{
Compute the sample skewness of a data set.
}
\details{
For normally distributed data, the skewness should be about zero. For
uni-modal continuous distributions, a skewness value greater than zero means
that there is more weight in the right tail of the distribution.
}
\examples{
df = pl$DataFrame(list(a = c(1:3, 2:1)))
df$select(pl$col("a")$skew())
}
