% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_slice}
\alias{Expr_slice}
\alias{slice}
\title{Get a slice of this expression.}
\usage{
Expr_slice(offset, length = NULL)
}
\arguments{
\item{offset}{numeric or expression, zero-indexed where to start slice
negative value indicate starting (one-indexed) from back}

\item{length}{how many elements should slice contain, default NULL is max length}
}
\value{
Expr
}
\description{
Get a slice of this expression.
}
\examples{

# as head
pl$DataFrame(list(a = 0:100))$select(
  pl$all()$slice(0, 6)
)

# as tail
pl$DataFrame(list(a = 0:100))$select(
  pl$all()$slice(-6, 6)
)

pl$DataFrame(list(a = 0:100))$select(
  pl$all()$slice(80)
)
}
\keyword{Expr}
