% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sort}
\alias{Expr_sort}
\title{Sort an Expr}
\usage{
<Expr>$sort(descending = FALSE, nulls_last = FALSE)
}
\arguments{
\item{descending}{Sort in descending order. When sorting by multiple columns,
can be specified per column by passing a vector of booleans.}

\item{nulls_last}{If \code{TRUE}, place nulls values last.}
}
\value{
Expr
}
\description{
Sort this column. If used in a groupby context, the groups are sorted.
}
\examples{
pl$DataFrame(a = c(6, 1, 0, NA, Inf, NaN))$
  with_columns(sorted = pl$col("a")$sort())
}
