% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sort}
\alias{Expr_sort}
\alias{sort}
\title{Expr_sort}
\format{
a method
}
\usage{
Expr_sort(reverse = FALSE, nulls_last = FALSE)
}
\arguments{
\item{reverse}{bool default FALSE, reverses sort}

\item{nulls_last}{bool, default FALSE, place Nulls last}
}
\value{
Expr
}
\description{
Sort this column. In projection/ selection context the whole column is sorted.
If used in a groupby context, the groups are sorted.
}
\details{
See Inf,NaN,NULL,Null/NA translations here \code{\link[rpolars]{docs_translations}}
}
\examples{
pl$DataFrame(list(
  a = c(6, 1, 0, NA, Inf, NaN)
))$select(pl$col("a")$sort())
}
\keyword{Expr}
