% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sort_by}
\alias{Expr_sort_by}
\title{Sort Expr by order of others}
\usage{
Expr_sort_by(by, descending = FALSE)
}
\arguments{
\item{by}{One expression or a list of expressions and/or strings (interpreted
as column names).}

\item{descending}{Sort in descending order. When sorting by multiple columns,
can be specified per column by passing a vector of booleans.}
}
\value{
Expr
}
\description{
Sort this column by the ordering of another column, or multiple other columns.
If used in a groupby context, the groups are sorted.
}
\examples{
df = pl$DataFrame(
  group = c("a", "a", "a", "b", "b", "b"),
  value1 = c(98, 1, 3, 2, 99, 100),
  value2 = c("d", "f", "b", "e", "c", "a")
)

# by one column/expression
df$with_columns(
  sorted = pl$col("group")$sort_by("value1")
)

# by two columns/expressions
df$with_columns(
  sorted = pl$col("group")$sort_by(
    list("value2", pl$col("value1")),
    descending = c(TRUE, FALSE)
  )
)

# by some expression
df$with_columns(
  sorted = pl$col("group")$sort_by(pl$col("value1")$sort(descending = TRUE))
)
}
