% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_std}
\alias{Expr_std}
\title{Get standard deviation}
\usage{
<Expr>$std(ddof = 1)
}
\arguments{
\item{ddof}{Degrees of freedom, must be an integer between 0 and 255}
}
\value{
Expr
}
\description{
Get standard deviation
}
\examples{
pl$select(pl$lit(1:5)$std())
}
