% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_std}
\alias{Expr_std}
\title{Get Standard Deviation}
\format{
a method
}
\usage{
Expr_std(ddof = 1)
}
\arguments{
\item{ddof}{integer in range \link{0;255} degrees of freedom}
}
\value{
Expr (f64 scalar)
}
\description{
Get Standard Deviation
}
\examples{
pl$select(pl$lit(1:5)$std())
}
\keyword{Expr}
