% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_contains_any}
\alias{expr_str_contains_any}
\title{Use the Aho-Corasick algorithm to find matches}
\usage{
expr_str_contains_any(patterns, ..., ascii_case_insensitive = FALSE)
}
\arguments{
\item{patterns}{String patterns to search. Accepts expression input. To use
the same character vector for all rows, use
\code{list(c(...))} instead of \code{c(...)} (see Examples).}

\item{...}{These dots are for future extensions and must be empty.}

\item{ascii_case_insensitive}{Enable ASCII-aware case insensitive matching.
When this option is enabled, searching will be performed without respect to
case for ASCII letters (a-z and A-Z) only.}
}
\value{
A polars \link{expression}
}
\description{
This function determines if any of the patterns find a match.
}
\examples{
df <- pl$DataFrame(
  lyrics = c(
    "Everybody wants to rule the world",
    "Tell me what you want, what you really really want",
    "Can you feel the love tonight"
  )
)

df$with_columns(
  contains_any = pl$col("lyrics")$str$contains_any(list(c("you", "me")))
)
}
\seealso{
\itemize{
\item \code{\link[=expr_str_contains]{<Expr>$str$contains()}}
}
}
