% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_contains_any}
\alias{expr_str_contains_any}
\title{Use the Aho-Corasick algorithm to find matches}
\usage{
expr_str_contains_any(patterns, ..., ascii_case_insensitive = FALSE)
}
\arguments{
\item{patterns}{Character vector or something can be coerced to strings \link[=Expr_class]{Expr}
of a valid regex pattern, compatible with the \href{https://docs.rs/regex/latest/regex/}{regex crate}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ascii_case_insensitive}{Enable ASCII-aware case insensitive matching.
When this option is enabled, searching will be performed without respect to
case for ASCII letters (a-z and A-Z) only.}
}
\value{
A polars \link{expression}
}
\description{
This function determines if any of the patterns find a match.
}
\examples{
df <- pl$DataFrame(
  lyrics = c(
    "Everybody wants to rule the world",
    "Tell me what you want, what you really really want",
    "Can you feel the love tonight"
  )
)

df$with_columns(
  contains_any = pl$col("lyrics")$str$contains_any(c("you", "me"))
)
}
\seealso{
\itemize{
\item \code{\link[=expr_str_contains]{<Expr>$str$contains()}}
}
}
