% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_ends_with}
\alias{expr_str_ends_with}
\title{Check if string ends with a regex}
\usage{
expr_str_ends_with(suffix)
}
\arguments{
\item{suffix}{Suffix substring or Expr.}
}
\value{
A polars \link{expression}
}
\description{
Check if string values end with a substring.
}
\details{
See also \verb{$str$starts_with()} and \verb{$str$contains()}.
}
\examples{
df <- pl$DataFrame(fruits = c("apple", "mango", NA))
df$select(
  pl$col("fruits"),
  pl$col("fruits")$str$ends_with("go")$alias("has_suffix")
)
}
