% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_pad_start}
\alias{expr_str_pad_start}
\title{Right justify strings}
\usage{
expr_str_pad_start(length, fill_char = " ")
}
\arguments{
\item{length}{Justify right to this length.}

\item{fill_char}{Fill with this ASCII character.}
}
\value{
A polars \link{expression}
}
\description{
Return the string right justified in a string of length \code{length}.
}
\details{
Padding is done using the specified \code{fill_char}. The original string
is returned if \code{length} is less than or equal to \code{len(s)}.
}
\examples{
df <- pl$DataFrame(a = c("cow", "monkey", NA, "hippopotamus"))
df$select(pl$col("a")$str$pad_start(8, "*"))
}
