% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_replace_many}
\alias{expr_str_replace_many}
\title{Use the Aho-Corasick algorithm to replace many matches}
\usage{
expr_str_replace_many(
  patterns,
  replace_with,
  ...,
  ascii_case_insensitive = FALSE,
  leftmost = FALSE
)
}
\arguments{
\item{patterns}{String patterns to search. Accepts expression input. Bare
character vectors are interpreted as column names. For literal patterns,
use a list or an expression such as \code{pl$lit(...)$implode()}.}

\item{replace_with}{A vector of strings used as replacements. If this is of
length 1, then it is applied to all matches. Otherwise, it must be of same
length as the \code{patterns} argument.}

\item{...}{These dots are for future extensions and must be empty.}

\item{ascii_case_insensitive}{Enable ASCII-aware case insensitive matching.
When this option is enabled, searching will be performed without respect to
case for ASCII letters (a-z and A-Z) only.}

\item{leftmost}{Whether to guarantee in case there are overlapping matches that
the leftmost match is used. In case there are multiple candidates for the
leftmost match, the pattern which comes first in \code{patterns} is used.}
}
\value{
A polars \link{expression}
}
\description{
This function replaces several matches at once.
}
\examples{
df <- pl$DataFrame(
  lyrics = c(
    "Everybody wants to rule the world",
    "Tell me what you want, what you really really want",
    "Can you feel the love tonight"
  )
)

# a replacement of length 1 is applied to all matches
df$with_columns(
  remove_pronouns = pl$col("lyrics")$str$replace_many(list(c("you", "me")), "")
)

# if there are more than one replacement, the patterns and replacements are
# matched
df$with_columns(
  fake_pronouns = pl$col("lyrics")$str$replace_many(list(c("you", "me")), list(c("foo", "bar")))
)
}
