% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_reverse}
\alias{expr_str_reverse}
\title{Returns string values in reversed order}
\usage{
expr_str_reverse()
}
\value{
A polars \link{expression}
}
\description{
Returns string values in reversed order
}
\examples{
df <- pl$DataFrame(text = c("foo", "bar", NA))
df$with_columns(reversed = pl$col("text")$str$reverse())
}
