% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_slice}
\alias{expr_str_slice}
\title{Create subslices of the string values of a String Series}
\usage{
expr_str_slice(offset, length = NULL)
}
\arguments{
\item{offset}{Start index. Negative indexing is supported.}

\item{length}{Length of the slice. If \code{NULL} (default), the slice is taken to
the end of the string.}
}
\value{
A polars \link{expression}
}
\description{
Create subslices of the string values of a String Series
}
\examples{
df <- pl$DataFrame(s = c("pear", NA, "papaya", "dragonfruit"))
df$with_columns(
  pl$col("s")$str$slice(-3)$alias("s_sliced")
)
}
