% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_split}
\alias{expr_str_split}
\title{Split the string by a substring}
\usage{
expr_str_split(by, ..., inclusive = FALSE)
}
\arguments{
\item{by}{Substring to split by. Can be an Expr.}

\item{...}{Dots which should be empty.}

\item{inclusive}{If \code{TRUE}, include the split character/string in the results.}
}
\value{
A polars \link{expression}
}
\description{
Split the string by a substring
}
\examples{
df <- pl$DataFrame(s = c("foo bar", "foo-bar", "foo bar baz"))
df$select(pl$col("s")$str$split(by = " "))

df <- pl$DataFrame(
  s = c("foo^bar", "foo_bar", "foo*bar*baz"),
  by = c("_", "_", "*")
)
df
df$select(split = pl$col("s")$str$split(by = pl$col("by")))
}
