% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_splitn}
\alias{expr_str_splitn}
\title{Split the string by a substring, restricted to returning at most \code{n} items}
\usage{
expr_str_splitn(by, n)
}
\arguments{
\item{by}{Substring to split by.}

\item{n}{Number of splits to make.}
}
\value{
A polars \link{expression}
}
\description{
If the number of possible splits is less than \code{n-1}, the remaining field
elements will be null. If the number of possible splits is \code{n-1} or greater,
the last (nth) substring will contain the remainder of the string.
}
\examples{
df <- pl$DataFrame(s = c("a_1", NA, "c", "d_4_e"))
df$with_columns(
  s1 = pl$col("s")$str$splitn(by = "_", 1),
  s2 = pl$col("s")$str$splitn(by = "_", 2),
  s3 = pl$col("s")$str$splitn(by = "_", 3)
)
}
