% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_to_decimal}
\alias{expr_str_to_decimal}
\title{Convert a String column into a Decimal column}
\usage{
expr_str_to_decimal(inference_length = 100)
}
\arguments{
\item{inference_length}{Number of elements to parse to determine the
\code{precision} and \code{scale}.}
}
\value{
A polars \link{expression}
}
\description{
This method infers the needed parameters \code{precision} and \code{scale}.
}
\examples{
df <- pl$DataFrame(
  numbers = c(
    "40.12", "3420.13", "120134.19", "3212.98",
    "12.90", "143.09", "143.9"
  )
)
df$with_columns(numbers_decimal = pl$col("numbers")$str$to_decimal())
}
