% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_to_titlecase}
\alias{expr_str_to_titlecase}
\title{Convert a string to titlecase}
\usage{
expr_str_to_titlecase()
}
\value{
A polars \link{expression}
}
\description{
Transform to titlecase variant.
}
\details{
This method is only available with the "nightly" feature.
See \code{\link[=polars_info]{polars_info()}} for more details.
}
\examples{
\dontshow{if (polars_info()$features$nightly) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl$select(
  pl$lit(c("hello there", "HI, THERE", NA))$str$to_titlecase()
)$to_series()
\dontshow{\}) # examplesIf}
}
