% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-string.R
\name{expr_str_to_uppercase}
\alias{expr_str_to_uppercase}
\title{Convert a string to uppercase}
\usage{
expr_str_to_uppercase()
}
\value{
A polars \link{expression}
}
\description{
Transform to uppercase variant.
}
\examples{
pl$lit(c("A", "b", "c", "1", NA))$str$to_uppercase()$to_series()
}
