% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-struct.R
\name{expr_struct_rename_fields}
\alias{expr_struct_rename_fields}
\title{Rename the fields of the struct}
\usage{
expr_struct_rename_fields(names)
}
\arguments{
\item{names}{New names, given in the same order as the struct's fields.}
}
\value{
A polars \link{expression}
}
\description{
Rename the fields of the struct
}
\examples{
df <- pl$DataFrame(
  aaa = c(1, 2),
  bbb = c("ab", "cd"),
  ccc = c(TRUE, NA),
  ddd = list(1:2, 3)
)$select(struct_col = pl$struct("aaa", "bbb", "ccc", "ddd"))
df

df <- df$select(
  pl$col("struct_col")$struct$rename_fields(c("www", "xxx", "yyy", "zzz"))
)
df$select(pl$col("struct_col")$struct$field("*"))

# Following a rename, the previous field names cannot be referenced:
tryCatch(
  {
    df$select(pl$col("struct_col")$struct$field("aaa"))
  },
  error = function(e) print(e)
)
}
