% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-struct.R
\name{expr_struct_with_fields}
\alias{expr_struct_with_fields}
\title{Add or overwrite fields of this struct}
\usage{
expr_struct_with_fields(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Field(s) to add. Accepts
expression input. Strings are parsed as column names, other non-expression
inputs are parsed as literals.}
}
\value{
A polars \link{expression}
}
\description{
This is similar to \code{with_columns()} on DataFrame and LazyFrame.
}
\examples{
df <- pl$DataFrame(
  x = c(1, 4, 9),
  y = c(4, 9, 16),
  multiply = c(10, 2, 3)
)$select(coords = pl$struct("x", "y"), "multiply")
df

df <- df$with_columns(
  pl$col("coords")$struct$with_fields(
    pl$field("x")$sqrt(),
    y_mul = pl$field("y") * pl$col("multiply")
  )
)

df
df$select(pl$col("coords")$struct$field("*"))
}
