% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_sum}
\alias{Expr_sum}
\title{Get sum value}
\usage{
<Expr>$sum
}
\value{
Expr
}
\description{
Get sum value
}
\details{
The dtypes Int8, UInt8, Int16 and UInt16 are cast to Int64 before summing to
prevent overflow issues.
}
\examples{
pl$DataFrame(x = c(1L, NA, 2L))$
  with_columns(sum = pl$col("x")$sum())
}
\keyword{datasets}
