% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_tail}
\alias{Expr_tail}
\alias{tail}
\title{Tail}
\usage{
Expr_tail(n = 10)
}
\arguments{
\item{n}{numeric number of elements to select from tail}
}
\value{
Expr
}
\description{
Get the tail n elements.
Similar to R tail(x)
}
\examples{
# get 3 last elements
pl$DataFrame(list(x = 1:11))$select(pl$col("x")$tail(3))
}
