% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_take}
\alias{Expr_take}
\alias{take}
\title{Take values by index.}
\format{
a method
}
\usage{
Expr_take(indices)
}
\arguments{
\item{indeces}{R scalar/vector or Series, or Expr that leads to a UInt32 dtyped Series.}
}
\value{
Expr
}
\description{
Take values by index.
}
\details{
similar to R indexing syntax e.g. \code{letters[c(1,3,5)]}, however as an expression, not as eager computation
exceeding
}
\examples{
pl$empty_select( pl$lit(0:10)$take(c(1,8,0,7)))
}
\keyword{Expr}
