% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_take_every}
\alias{Expr_take_every}
\title{Take every nth element}
\usage{
Expr_take_every(n)
}
\arguments{
\item{n}{Positive integer.}
}
\value{
Expr
}
\description{
Take every nth value in the Series and return as a new Series.
}
\examples{
pl$DataFrame(a = 0:24)$select(pl$col("a")$take_every(6))
}
