% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_tan}
\alias{Expr_tan}
\title{Compute tangent}
\usage{
Expr_tan()
}
\value{
Expr
}
\description{
Compute tangent
}
\examples{
pl$DataFrame(a = c(0, pi / 2, pi, NA_real_))$
  with_columns(tangent = pl$col("a")$tan())
}
