% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_to_physical}
\alias{Expr_to_physical}
\alias{to_physical}
\title{Cast an Expr to its physical representation}
\usage{
<Expr>$to_physical
}
\value{
Expr
}
\description{
The following DataTypes will be converted:
\itemize{
\item Date -> Int32
\item Datetime -> Int64
\item Time -> Int64
\item Duration -> Int64
\item Categorical -> UInt32
\item List(inner) -> List(physical of inner)
Other data types will be left unchanged.
}
}
\examples{
pl$DataFrame(
  list(vals = c("a", "x", NA, "a", "b"))
)$with_columns(
  pl$col("vals")$cast(pl$Categorical),
  pl$col("vals")
  $cast(pl$Categorical)
  $to_physical()
  $alias("vals_physical")
)
}
\keyword{datasets}
