% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_to_r}
\alias{Expr_to_r}
\title{Convert an Expr to R output}
\usage{
Expr_to_r(
  df = NULL,
  i = 0,
  ...,
  int64_conversion = pl$options$int64_conversion
)
}
\arguments{
\item{df}{If \code{NULL} (default), it evaluates the Expr in an empty DataFrame.
Otherwise, provide a DataFrame that the Expr should be evaluated in.}

\item{i}{Numeric column to extract. Default is zero (which gives the first
column).}

\item{...}{Ignored.}

\item{int64_conversion}{How should Int64 values be handled when converting a
polars object to R?
\itemize{
\item \code{"double"} (default) converts the integer values to double.
\item \code{"bit64"} uses \code{bit64::as.integer64()} to do the conversion (requires
the package \code{bit64} to be attached).
\item \code{"string"} converts Int64 values to character.
}}
}
\value{
R object
}
\description{
This is mostly useful to debug an expression. It evaluates the Expr in an
empty DataFrame and return the first Series to R.
}
\examples{
pl$lit(1:3)$to_r()
}
