% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_to_r}
\alias{Expr_to_r}
\alias{to_r}
\alias{expr_to_r}
\alias{pl_expr_to_r}
\title{to_r: for debuging an expression}
\format{
Method
}
\usage{
Expr_to_r(df = NULL, i = 0)
}
\arguments{
\item{df}{DataFrame to evaluate in, default NULL is an empty DataFrame}

\item{i}{numeric column to extract zero index default first, expression could generate multiple
columns}
}
\value{
R object
}
\description{
debug an expression by evaluating in empty DataFrame and return first series to R
}
\examples{
pl$lit(1:3)$to_r()
pl$expr_to_r(pl$lit(1:3)
pl$expr_to_r(1:3)
}
\keyword{Expr}
