% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_to_struct}
\alias{Expr_to_struct}
\alias{expr_to_struct}
\title{to_struct}
\usage{
Expr_to_struct()
}
\value{
Expr
}
\description{
pass expr to pl$struct
}
\examples{
e = pl$all()$to_struct()$alias("my_struct")
print(e)
pl$DataFrame(iris)$select(e)
}
\keyword{Expr}
