% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_top_k}
\alias{Expr_top_k}
\alias{top_k}
\title{Top k values}
\format{
a method
}
\usage{
Expr_top_k(k, reverse = FALSE)
}
\arguments{
\item{k}{numeric k top values to get}

\item{reverse}{bool if true then k smallest values}
}
\value{
Expr
}
\description{
Return the \code{k} largest elements.
If 'reverse=True` the smallest elements will be given.
}
\details{
This has time complexity: \eqn{ O(n + k \\log{}n - \frac{k}{2}) }

See Inf,NaN,NULL,Null/NA translations here \code{\link[minipolars]{docs_translations}}
}
\examples{
df = pl$DataFrame(list(
  a = c(6, 1, 0, NA, Inf, NaN),
))$select(pl$top_k(5))
}
\keyword{Expr}
