% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_unique}
\alias{Expr_unique}
\title{Get unique values}
\usage{
Expr_unique(maintain_order = FALSE)
}
\arguments{
\item{maintain_order}{If \code{TRUE}, the unique values are returned in order of
appearance.}
}
\value{
Expr
}
\description{
Get unique values
}
\examples{
pl$DataFrame(iris)$select(pl$col("Species")$unique())
}
