% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_unique_counts}
\alias{Expr_unique_counts}
\title{Count unique values}
\usage{
Expr_unique_counts
}
\value{
Expr
}
\description{
Return a count of the unique values in the order of appearance. This method
differs from \verb{$value_counts()} in that it does not return the values, only
the counts and it might be faster.
}
\examples{
pl$DataFrame(iris)$select(pl$col("Species")$unique_counts())
}
\keyword{datasets}
