% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_value_counts}
\alias{Expr_value_counts}
\title{Value counts}
\format{
Method
}
\usage{
Expr_value_counts(multithreaded = FALSE, sort = FALSE)
}
\arguments{
\item{multithreaded}{Better to turn this off in the aggregation context, as it can lead to contention.}

\item{sort}{Ensure the output is sorted from most values to least.}
}
\value{
Expr
}
\description{
Count all unique values and create a struct mapping value to count.
}
\examples{
df = pl$DataFrame(iris)$select(pl$col("Species")$value_counts())
df
df$unnest()$as_data_frame() #recommended to unnest structs before converting to R
}
\keyword{Expr}
