% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_var}
\alias{Expr_var}
\title{Get variance}
\usage{
Expr_var(ddof = 1)
}
\arguments{
\item{ddof}{Degrees of freedom, must be an integer between 0 and 255}
}
\value{
Expr
}
\description{
Get variance
}
\examples{
pl$select(pl$lit(1:5)$var())
}
