% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_where}
\alias{Expr_where}
\title{Filter a single column.}
\usage{
Expr_where(predicate)
}
\arguments{
\item{predicate}{An Expr or something coercible to an Expr. Must return a
boolean.}
}
\value{
Expr
}
\description{
This is an alias for \verb{<Expr>$filter()}.
}
\examples{
df = pl$DataFrame(
  group_col = c("g1", "g1", "g2"),
  b = c(1, 2, 3)
)
df

df$group_by("group_col")$agg(
  lt = pl$col("b")$where(pl$col("b") < 2),
  gte = pl$col("b")$where(pl$col("b") >= 2)
)
}
