% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\docType{data}
\name{Expr_xor}
\alias{Expr_xor}
\title{Xor}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_xor
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Exprs
}
\description{
Divide
}
\examples{
pl$lit(TRUE).xor(pl$lit(FALES))
}
\keyword{Expr}
\keyword{Expr_operators}
