% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{Expr_xor}
\alias{Expr_xor}
\title{Xor}
\format{
An object of class \code{character} of length 1.
}
\usage{
Expr_xor(other)
}
\arguments{
\item{other}{literal or Robj which can become a literal}
}
\value{
Expr
}
\description{
combine to boolean expresions with XOR
}
\examples{
pl$lit(TRUE)$xor(pl$lit(FALSE))
}
\keyword{Expr}
\keyword{Expr_operators}
