% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_arg_max}
\alias{ExprArr_arg_max}
\alias{arr_arg_max}
\title{Get the index of the maximal value in list}
\usage{
ExprArr_arg_max()
}
\value{
Expr
}
\description{
Get the index of the maximal value in list
}
\examples{
df = pl$DataFrame(list(s = list(1:2, 2:1)))
df$with_columns(
  arg_max = pl$col("s")$arr$arg_max()
)
}
