% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_contains}
\alias{ExprArr_contains}
\alias{arr_contains}
\title{Check if list contains a given value}
\usage{
ExprArr_contains(item)
}
\arguments{
\item{item}{Expr or something coercible to an Expr. Strings are \emph{not} parsed
as columns.}
}
\value{
Expr
}
\description{
Check if list contains a given value
}
\examples{
df = pl$DataFrame(
  a = list(3:1, NULL, 1:2),
  item = 0:2
)
df$with_columns(
  with_expr = pl$col("a")$arr$contains(pl$col("item")),
  with_lit = pl$col("a")$arr$contains(1)
)
}
