% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_get}
\alias{ExprArr_get}
\alias{arr_get}
\title{Get the value by index in an array}
\usage{
ExprArr_get(index)
}
\arguments{
\item{index}{An Expr or something coercible to an Expr, that must return a
single index. Values are 0-indexed (so index 0 would return the first item
of every sub-array) and negative values start from the end (index \code{-1}
returns the last item). If the index is out of bounds, it will return a
\code{null}. Strings are parsed as column names.}
}
\value{
Expr
}
\description{
This allows to extract one value per array only.
}
\examples{
df = pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(NA_real_, 6)),
  idx = c(1, NA, 3),
  schema = list(values = pl$Array(pl$Float64, 2))
)
df$with_columns(
  using_expr = pl$col("values")$arr$get("idx"),
  val_0 = pl$col("values")$arr$get(0),
  val_minus_1 = pl$col("values")$arr$get(-1),
  val_oob = pl$col("values")$arr$get(10)
)
}
