% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_max}
\alias{ExprArr_max}
\alias{arr_max}
\title{Find the maximum value in an array}
\usage{
ExprArr_max()
}
\value{
Expr
}
\description{
Find the maximum value in an array
}
\examples{
df = pl$DataFrame(list(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_)))
df$with_columns(max = pl$col("values")$arr$max())
}
