% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_min}
\alias{ExprArr_min}
\alias{arr_min}
\title{Find the minimum value in an array}
\usage{
ExprArr_min()
}
\value{
Expr
}
\description{
Find the minimum value in an array
}
\details{
This method is only available with the "nightly" feature.
See \code{\link[=polars_info]{polars_info()}} for more details.
}
\examples{
\dontshow{if (polars_info()$features$nightly) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df = pl$DataFrame(
  values = list(c(1, 2), c(3, 4), c(5, 6)),
  schema = list(values = pl$Array(pl$Float64, 2))
)
df$with_columns(min = pl$col("values")$arr$min())
\dontshow{\}) # examplesIf}
}
