% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_min}
\alias{ExprArr_min}
\alias{arr_min}
\title{Find the minimum value in an array}
\usage{
ExprArr_min()
}
\value{
Expr
}
\description{
Find the minimum value in an array
}
\examples{
df = pl$DataFrame(list(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_)))
df$with_columns(min = pl$col("values")$arr$min())
}
