% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_shift}
\alias{ExprArr_shift}
\alias{arr_shift}
\title{Shift array values by \code{n} indices}
\usage{
ExprArr_shift(periods = 1)
}
\arguments{
\item{periods}{Number of places to shift (may be negative). Can be an Expr.
Strings are \emph{not} parsed as columns.}
}
\value{
Expr
}
\description{
Shift array values by \code{n} indices
}
\examples{
df = pl$DataFrame(
  values = list(1:3, c(2L, NA_integer_, 5L)),
  idx = 1:2,
  schema = list(values = pl$Array(pl$Int32, 3))
)
df$with_columns(
  shift_by_expr = pl$col("values")$arr$shift(pl$col("idx")),
  shift_by_lit = pl$col("values")$arr$shift(2)
)
}
