% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__array.R
\name{ExprArr_sort}
\alias{ExprArr_sort}
\alias{arr_sort}
\title{Sort values in a list}
\usage{
ExprArr_sort(descending = FALSE)
}
\arguments{
\item{descending}{Sort values in descending order}
}
\value{
Expr
}
\description{
Sort values in a list
}
\examples{
df = pl$DataFrame(list(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_)))
df$with_columns(sort = pl$col("values")$arr$sort())
}
