% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_contains}
\alias{ExprBin_contains}
\alias{expr_bin_contains}
\title{contains}
\arguments{
\item{lit}{The binary substring to look for}
}
\value{
Expr returning a Boolean
}
\description{
R Check if binaries in Series contain a binary substring.
}
\keyword{ExprBin}
