% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_decode}
\alias{ExprBin_decode}
\alias{expr_bin_decode}
\title{decode}
\usage{
ExprBin_decode(encoding, strict = TRUE)
}
\arguments{
\item{encoding}{binary choice either 'hex' or 'base64'}

\item{strict}{Raise an error if the underlying value cannot be decoded,
otherwise mask out with a null value.}
}
\value{
binary array with values decoded using provided encoding
}
\description{
Decode a value using the provided encoding.
}
\keyword{ExprBin}
