% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_encode}
\alias{ExprBin_encode}
\alias{expr_bin_encode}
\title{encode}
\arguments{
\item{encoding}{binary choice either 'hex' or 'base64'}
}
\value{
binary array with values encoded using provided encoding
}
\description{
Encode a value using the provided encoding.
}
\keyword{ExprBin}
