% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_ends_with}
\alias{ExprBin_ends_with}
\alias{expr_bin_ends_with}
\title{ends_with}
\value{
Expr returning a Boolean
}
\description{
Check if string values end with a binary substring.
}
\keyword{ExprBin}
