% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__binary.R
\name{ExprBin_starts_with}
\alias{ExprBin_starts_with}
\alias{expr_bin_starts_with}
\title{starts_with}
\arguments{
\item{sub}{Prefix substring.}
}
\description{
Check if values starts with a binary substring.
}
\keyword{ExprBin}
