% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__categorical.R
\name{ExprCat_set_ordering}
\alias{ExprCat_set_ordering}
\alias{expr_cat_set_ordering}
\title{Set Ordering}
\usage{
<Expr>$cat$set_ordering(ordering)
}
\arguments{
\item{ordering}{string either 'physical' or 'lexical'
\itemize{
\item 'physical' -> Use the physical representation of the categories to
determine the order (default).
\item 'lexical' -> Use the string values to determine the ordering.
}}
}
\value{
bool: TRUE if equal
}
\description{
Determine how this categorical series should be sorted.
}
\examples{
df = pl$DataFrame(
  cats = factor(c("z", "z", "k", "a", "b")),
  vals = c(3, 1, 2, 2, 3)
)

# sort by the string value of categories
df$with_columns(
  pl$col("cats")$cat$set_ordering("lexical")
)$sort("cats", "vals")

# sort by the underlying value of categories
df$with_columns(
  pl$col("cats")$cat$set_ordering("physical")
)$sort("cats", "vals")
}
\keyword{ExprCat}
