% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_convert_time_zone}
\alias{ExprDT_convert_time_zone}
\title{Convert to given time zone for an expression of type Datetime.}
\usage{
ExprDT_convert_time_zone(time_zone)
}
\arguments{
\item{time_zone}{String time zone from \code{\link[base:timezones]{base::OlsonNames()}}}
}
\value{
Expr of i64
}
\description{
If converting from a time-zone-naive datetime,
then conversion will happen as if converting from UTC,
regardless of your system’s time zone.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(
    as.POSIXct("2020-03-01", tz = "UTC"),
    as.POSIXct("2020-05-01", tz = "UTC"),
    "1mo"
  )
)

df$select(
  "date",
  London = pl$col("date")$dt$convert_time_zone("Europe/London")
)
}
