% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__datetime.R
\name{ExprDT_hours}
\alias{ExprDT_hours}
\alias{(Expr)$dt$hours}
\title{Hours}
\format{
function
}
\value{
Expr of i64
}
\description{
Extract the hours from a Duration type.
}
\examples{
df = pl$DataFrame(
  date = pl$date_range(
    low = as.Date("2020-1-1"),
    high = as.Date("2020-1-4"),
    interval = "1d",
    lazy = FALSE
  )
)
df$select(
  pl$col("date"),
  pl$col("date")$diff()$dt$hours()$alias("hours_diff")
)
}
\keyword{ExprDT}
