% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_concat}
\alias{ExprList_concat}
\alias{list_concat}
\title{concat another list}
\format{
function
}
\usage{
<Expr>$list$concat(other)
}
\arguments{
\item{other}{Rlist, Expr or column of same type as self.}
}
\value{
Expr
}
\description{
Concat the arrays in a Series dtype List in linear time.
}
\examples{
df = pl$DataFrame(
  a = list("a", "x"),
  b = list(c("b", "c"), c("y", "z"))
)
df$select(pl$col("a")$list$concat(pl$col("b")))

df$select(pl$col("a")$list$concat(pl$lit("hello from R")))

df$select(pl$col("a")$list$concat(pl$lit(list("hello", c("hello", "world")))))
}
\keyword{ExprList}
