% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_get}
\alias{ExprList_get}
\alias{Expr_list_get}
\alias{[.ExprListNameSpace}
\title{Get list}
\format{
function
}
\usage{
\method{[}{ExprListNameSpace}(x, index)
}
\arguments{
\item{x}{ExprListNameSpace}

\item{index}{value to get}
}
\value{
Expr
}
\description{
Get the value by index in the sublists.
}
\details{
\verb{[.ExprListNameSpace} used as e.g. \code{pl$col("a")$arr[0]} same as \code{pl$col("a")$get(0)}
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) # NULL or integer() or list()
df$select(pl$col("a")$list$get(0))
df$select(pl$col("a")$list$get(c(2, 0, -1)))
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) # NULL or integer() or list()
df$select(pl$col("a")$list[0])
df$select(pl$col("a")$list[c(2, 0, -1)])
}
\keyword{ExprList}
