% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_head}
\alias{ExprList_head}
\alias{list_head}
\title{Heads of sublists}
\format{
function
}
\usage{
<Expr>$list$head(n = 5L)
}
\arguments{
\item{n}{Numeric or Expr, number of values to return for each sublist.}
}
\value{
Expr
}
\description{
head the first \code{n} values of every sublist.
}
\examples{
df = pl$DataFrame(list(a = list(1:4, c(10L, 2L, 1L))))
df$select(pl$col("a")$list$head(2))
}
\keyword{ExprList}
