% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_join}
\alias{ExprList_join}
\alias{list_join}
\title{Join elements of a list}
\usage{
ExprList_join(separator, ignore_nulls = FALSE)
}
\arguments{
\item{separator}{String to separate the items with. Can be an Expr. Strings
are \emph{not} parsed as columns.}

\item{ignore_nulls}{If \code{FALSE} (default), null values are propagated: if the
row contains any null values, the output is null.}
}
\value{
Expr
}
\description{
Join all string items in a sublist and place a separator between them. This
only works on columns of type \code{list[str]}.
}
\examples{
df = pl$DataFrame(
  s = list(c("a", "b", "c"), c("x", "y"), c("e", NA)),
  separator = c("-", "+", "/")
)
df$with_columns(
  join_with_expr = pl$col("s")$list$join(pl$col("separator")),
  join_with_lit = pl$col("s")$list$join(" ")
)
}
