% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_last}
\alias{ExprList_last}
\alias{list_last}
\title{Last in sublists}
\format{
function
}
\value{
Expr
}
\description{
Get the last value of the sublists.
}
\examples{
df = pl$DataFrame(list(a = list(3:1, NULL, 1:2))) # NULL or integer() or list()
df$select(pl$col("a")$list$last())
}
\keyword{ExprList}
