% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_max}
\alias{ExprList_max}
\alias{Expr_list_max}
\title{Max lists}
\format{
function
}
\usage{
<Expr>$list$max()
}
\value{
Expr
}
\description{
Compute the max value of the lists in the array.
}
\examples{
df = pl$DataFrame(values = pl$Series(list(1L, 2:3)))
df$select(pl$col("values")$list$max())
}
\keyword{ExprList}
