% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_mean}
\alias{ExprList_mean}
\alias{list_mean}
\title{Mean of lists}
\format{
function
}
\usage{
ExprList_mean()
}
\value{
Expr
}
\description{
Compute the mean value of the lists in the array.
}
\examples{
df = pl$DataFrame(values = pl$Series(list(1L, 2:3)))
df$select(pl$col("values")$list$mean())
}
\keyword{ExprList}
