% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_min}
\alias{ExprList_min}
\alias{Expr_list_min}
\title{#' Min lists}
\format{
function
}
\value{
Expr
}
\description{
Compute the min value of the lists in the array.
}
\examples{
df = pl$DataFrame(values = pl$Series(list(1L, 2:3)))
df$select(pl$col("values")$list$min())
}
\keyword{ExprList}
