% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_reverse}
\alias{ExprList_reverse}
\alias{list_reverse}
\title{Reverse list}
\format{
function
}
\value{
Expr
}
\description{
Reverse the arrays in the list.
}
\examples{
df = pl$DataFrame(list(
  values = list(3:1, c(9L, 1:2))
))
df$select(pl$col("values")$list$reverse())
}
\keyword{ExprList}
