% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_slice}
\alias{ExprList_slice}
\alias{list_slice}
\title{Slice sublists}
\format{
function
}
\arguments{
\item{offset}{value or Expr.  Start index. Negative indexing is supported.}

\item{length}{value or Expr.
Length of the slice. If set to \code{None} (default), the slice is taken to the
end of the list.}
}
\value{
Expr
}
\description{
Slice every sublist.
}
\examples{
df = pl$DataFrame(list(s = list(1:4, c(10L, 2L, 1L))))
df$select(pl$col("s")$list$slice(2))
}
\keyword{ExprList}
