% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_sum}
\alias{ExprList_sum}
\alias{list_sum}
\title{Sum all elements in a list}
\usage{
ExprList_sum()
}
\value{
Expr
}
\description{
Sum all elements in a list
}
\examples{
df = pl$DataFrame(list(values = list(c(1, 2, 3, NA), c(2, 3), NA_real_)))
df$with_columns(sum = pl$col("values")$list$sum())
}
