% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__list.R
\name{ExprList_unique}
\alias{ExprList_unique}
\alias{list_unique}
\title{Get unique values in a list}
\usage{
ExprList_unique()
}
\value{
Expr
}
\description{
Get unique values in a list
}
\examples{
df = pl$DataFrame(values = list(c(2, 2, NA), c(1, 2, 3), NA_real_))
df$with_columns(unique = pl$col("values")$list$unique())
}
