% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_eq}
\alias{ExprMeta_eq}
\title{Check if two expressions are equivalent}
\usage{
ExprMeta_eq(other)
}
\arguments{
\item{other}{Expr to compare with}
}
\value{
A logical value
}
\description{
Indicate if this expression is the same as another expression. See also the
counterpart \code{\link[=ExprMeta_neq]{$meta$neq()}}.
}
\examples{
# three naive expression literals
e1 = pl$lit(40) + 2
e2 = pl$lit(42)
e3 = pl$lit(40) + 2

# e1 and e3 are identical expressions
e1$meta$eq(e3)

# when evaluated, e1 and e2 are equal
e1$eq(e2)$to_r()

# however, on the meta-level, e1 and e2 are NOT identical expressions
e1$meta$eq(e2)
}
