% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_has_multiple_outputs}
\alias{ExprMeta_has_multiple_outputs}
\alias{expr_has_multiple_outputs}
\title{Has multiple outputs}
\value{
Bool
}
\description{
Whether this expression expands into multiple expressions.
}
\examples{
pl$all()$meta$has_multiple_outputs()
pl$col("some_colname")$meta$has_multiple_outputs()
}
\keyword{ExprMeta}
