% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_pop}
\alias{ExprMeta_pop}
\alias{expr_meta_pop}
\title{Pop}
\value{
R list of Expr(s) usually one, only multiple if top Expr took more Expr as input.
}
\description{
Pop the latest expression and return the input(s) of the popped expression.
}
\examples{
e1 = pl$lit(40) + 2
e2 = pl$lit(42)$sum()

e1
e1$meta$pop()

e2
e2$meta$pop()
}
\keyword{ExprMeta}
