% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_root_names}
\alias{ExprMeta_root_names}
\alias{expr_meta_root_names}
\title{Root Name}
\usage{
<Expr>$meta$root_names()
}
\value{
R charvec of root names.
}
\description{
Get a vector with the root column name
}
\examples{
e = pl$col("alice")$alias("bob")
e$meta$root_names() == "alice"
e$meta$output_name() == "bob"
e$meta$undo_aliases()$meta$output_name() == "alice"
}
\keyword{ExprMeta}
