% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__meta.R
\name{ExprMeta_undo_aliases}
\alias{ExprMeta_undo_aliases}
\alias{expr_meta_undo_aliases}
\title{Undo aliases}
\usage{
ExprMeta_undo_aliases()
}
\value{
Expr with aliases undone
}
\description{
Undo any renaming operation like \code{alias} or \code{keep_name}.
}
\examples{
e = pl$col("alice")$alias("bob")
e$meta$root_names() == "alice"
e$meta$output_name() == "bob"
e$meta$undo_aliases()$meta$output_name() == "alice"
}
\keyword{ExprMeta}
