% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__name.R
\name{ExprName_map}
\alias{ExprName_map}
\title{Map alias of expression with an R function}
\usage{
<Expr>$name$map(fun)
}
\arguments{
\item{fun}{an R function which takes a string as input and return a string}
}
\value{
Expr
}
\description{
Rename the output of an expression by mapping a function over the root name.
}
\examples{
pl$DataFrame(list(alice = 1:3))$select(
  pl$col("alice")$alias("joe_is_not_root")$name$map(\\(x) paste0(x, "_and_bob"))
)
}
