% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_concat}
\alias{ExprStr_concat}
\alias{expr_str_concat}
\title{Concat}
\arguments{
\item{delimiter}{string The delimiter to insert between consecutive string values.}
}
\value{
Expr of Utf8 concatenated
}
\description{
Vertically concat the values in the Series to a single string value.
}
\examples{
#concatenate a Series of strings to a single string
df = pl$DataFrame(foo = c("1", NA, 2))
df$select(pl$col("foo")$str$concat("-"))

#Series list of strings to Series of concatenated strings
df = pl$DataFrame(list(bar = list(c("a","b", "c"), c("1","2",NA))))
df$select(pl$col("bar")$arr$eval(pl$col()$str$concat())$arr$first())
}
\keyword{ExprStr}
