% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_contains}
\alias{ExprStr_contains}
\title{Check if string contains a regex}
\arguments{
\item{pattern}{String or Expr of a string, a valid regex pattern.}

\item{literal}{Treat pattern as a literal string.}

\item{strict}{Raise an error if the underlying pattern is not a valid regex
expression, otherwise replace the invalid regex with a null value.}
}
\value{
Expr returning a Boolean
}
\description{
Check if string contains a substring that matches a regex.
}
\details{
See also \verb{$str$starts_with()} and \verb{$str$ends_with()}.
}
\examples{
df = pl$DataFrame(a = c("Crab", "cat and dog", "rab$bit", NA))
df$select(
  pl$col("a"),
  pl$col("a")$str$contains("cat|bit")$alias("regex"),
  pl$col("a")$str$contains("rab$", literal = TRUE)$alias("literal")
)
}
\keyword{ExprStr}
