% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_ends_with}
\alias{ExprStr_ends_with}
\title{Check if string ends with a regex}
\usage{
<Expr>$str$ends_with(sub)
}
\arguments{
\item{sub}{Suffix substring or Expr.}
}
\value{
Expr returning a Boolean
}
\description{
Check if string values end with a substring.
}
\details{
See also \verb{$str$starts_with()} and \verb{$str$contains()}.
}
\examples{
df = pl$DataFrame(fruits = c("apple", "mango", NA))
df$select(
  pl$col("fruits"),
  pl$col("fruits")$str$ends_with("go")$alias("has_suffix")
)
}
\keyword{ExprStr}
