% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_extract}
\alias{ExprStr_extract}
\title{Extract the target capture group from provided patterns}
\usage{
<Expr>$str$extract(pattern, group_index)
}
\arguments{
\item{pattern}{A valid regex pattern}

\item{group_index}{Index of the targeted capture group. Group 0 means the whole
pattern, first group begin at index 1 (default).}
}
\value{
String array. Contains null if original value is null or regex capture nothing.
}
\description{
Extract the target capture group from provided patterns
}
\examples{
df = pl$DataFrame(
  a = c(
    "http://vote.com/ballon_dor?candidate=messi&ref=polars",
    "http://vote.com/ballon_dor?candidat=jorginho&ref=polars",
    "http://vote.com/ballon_dor?candidate=ronaldo&ref=polars"
  )
)
df$select(
  pl$col("a")$str$extract(r"(candidate=(\w+))", 1)
)
}
\keyword{ExprStr}
