% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_lengths}
\alias{ExprStr_lengths}
\alias{expr_str_lengths}
\title{str lengths}
\format{
function
}
\usage{
ExprStr_lengths()
}
\value{
Expr of u32 lengths
}
\description{
Get length of the strings as UInt32 (as number of bytes).
}
\details{
The returned lengths are equal to the number of bytes in the UTF8 string. If you
need the length in terms of the number of characters, use \code{n_chars} instead.
}
\examples{
pl$DataFrame(
  s = c("Café", NA, "345", "東京")
)$select(
  pl$col("s"),
  pl$col("s")$str$lengths()$alias("lengths"),
  pl$col("s")$str$n_chars()$alias("n_chars")
)
}
\keyword{ExprStr}
