% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_replace}
\alias{ExprStr_replace}
\title{Replace first matching regex/literal substring with a new string value}
\usage{
ExprStr_replace(pattern, value, literal = FALSE)
}
\arguments{
\item{pattern}{Regex pattern, can be an Expr.}

\item{value}{Replacement, can be an Expr.}

\item{literal}{Treat pattern as a literal string.}
}
\value{
Expr of Utf8 Series
}
\description{
Replace first matching regex/literal substring with a new string value
}
\examples{
df = pl$DataFrame(id = c(1, 2), text = c("123abc", "abc456"))
df$with_columns(
  pl$col("text")$str$replace(r"{abc\b}", "ABC")
)
}
\seealso{
\verb{$str$replace_all()}: Replace all matching regex/literal substrings.
}
\keyword{ExprStr}
