% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_slice}
\alias{ExprStr_slice}
\alias{expr_str_slice}
\title{slice}
\arguments{
\item{pattern}{\verb{Into<Expr>}, regex pattern}

\item{value}{\verb{Into<Expr>} replacement}

\item{literal}{bool, treat pattern as a literal string.}
}
\value{
Expr: Series of dtype Utf8.
}
\description{
Create subslices of the string values of a Utf8 Series.
}
\examples{
df = pl$DataFrame(s = c("pear", NA, "papaya", "dragonfruit"))
df$with_columns(
  pl$col("s")$str$slice(-3)$alias("s_sliced")
)
}
\keyword{ExprStr}
