% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_slice}
\alias{ExprStr_slice}
\title{Create subslices of the string values of a String Series}
\usage{
<Expr>$str$slice(offset, length = NULL)
}
\arguments{
\item{offset}{Start index. Negative indexing is supported.}

\item{length}{Length of the slice. If \code{NULL} (default), the slice is taken to
the end of the string.}
}
\value{
Expr: Series of dtype String.
}
\description{
Create subslices of the string values of a String Series
}
\examples{
df = pl$DataFrame(s = c("pear", NA, "papaya", "dragonfruit"))
df$with_columns(
  pl$col("s")$str$slice(-3)$alias("s_sliced")
)
}
\keyword{ExprStr}
