% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_split}
\alias{ExprStr_split}
\alias{expr_str_split}
\title{split}
\arguments{
\item{by}{Substring to split by.}

\item{inclusive}{If True, include the split character/string in the results.}
}
\value{
List of Utf8 type
}
\description{
Split the string by a substring.
}
\examples{
df = pl$DataFrame(s = c("foo bar", "foo-bar", "foo bar baz"))
df$select(pl$col("s")$str$split(by = " "))
}
\keyword{ExprStr}
