% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_starts_with}
\alias{ExprStr_starts_with}
\alias{expr_str_starts_with}
\title{starts_with}
\arguments{
\item{sub}{Prefix substring or Expr.}
}
\value{
Expr returning a Boolean
}
\description{
Check if string values starts with a substring.
}
\details{
contains : Check if string contains a substring that matches a regex.
ends_with : Check if string values end with a substring.
}
\examples{
df = pl$DataFrame(fruits = c("apple", "mango", NA))
df$select(
  pl$col("fruits"),
  pl$col("fruits")$str$starts_with("app")$alias("has_suffix")
)
}
\keyword{ExprStr}
