% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_lowercase}
\alias{ExprStr_to_lowercase}
\title{Convert a string to lowercase}
\value{
Expr of Utf8 lowercase chars
}
\description{
Transform to lowercase variant.
}
\examples{
pl$lit(c("A", "b", "c", "1", NA))$str$to_lowercase()$lit_to_s()
}
\keyword{ExprStr}
