% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_time}
\alias{ExprStr_to_time}
\title{Convert a String column into a Time column}
\usage{
<Expr>$str$to_time(format = NULL, strict = TRUE, cache = TRUE)
}
\arguments{
\item{format}{Format to use for conversion. See \code{?strptime} for possible
values. Example: "\%H:\%M:\%S". If \code{NULL} (default), the format is
inferred from the data. Notice that time zone \verb{\%Z} is not supported and will
just ignore timezones. Numeric time zones like \verb{\%z} or \verb{\%:z}  are supported.}

\item{strict}{If \code{TRUE} (default), raise an error if a single string cannot
be parsed. If \code{FALSE}, parsing failure will produce a polars \code{null}.}

\item{cache}{Use a cache of unique, converted dates to apply the datetime
conversion.}
}
\value{
Expr
}
\description{
Convert a String column into a Time column
}
\examples{
pl$DataFrame(str_time = c("01:20:01", "28:00:02", "03:00:02"))$
  with_columns(time = pl$col("str_time")$str$to_time(strict = FALSE))
}
