% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_titlecase}
\alias{ExprStr_to_titlecase}
\alias{expr_str_to_titlecase}
\title{To titlecase}
\value{
Expr of Utf8 titlecase chars
}
\description{
Transform to titlecase variant.
}
\examples{
pl$lit(c("hello there", "HI, THERE", NA))$str$to_titlecase()$lit_to_s()
}
\keyword{ExprStr}
