% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_titlecase}
\alias{ExprStr_to_titlecase}
\title{Convert a string to titlecase}
\value{
Expr of Utf8 titlecase chars
}
\description{
Transform to titlecase variant.
}
\details{
This method is only available with the feature flag "simd" which can
be set via envvar "RPOLARS_FULL_FEATURES" and it requires
Rust nightly toolchain to compile.
See \code{\link[=polars_info]{pl$polars_info()}} for more details.
}
\examples{
\dontshow{if (pl$polars_info()$features$simd) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pl$lit(c("hello there", "HI, THERE", NA))$str$to_titlecase() |>
  as_polars_series()
\dontshow{\}) # examplesIf}
}
\keyword{ExprStr}
