% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__string.R
\name{ExprStr_to_titlecase}
\alias{ExprStr_to_titlecase}
\title{Convert a string to titlecase}
\value{
Expr of Utf8 titlecase chars
}
\description{
Transform to titlecase variant.
}
\details{
This method is only available with rust compiler flag "full_features" which can
be set via envvar "RPOLARS_FULL_FEATURES" and it requires rust nightly to compile.
Polars GitHub binary releases are compiled with "full_features".
}
\examples{
f = \() pl$lit(c("hello there", "HI, THERE", NA))$str$to_titlecase()$lit_to_s()
if (pl$polars_info()$features$full_features) {
  f()
} else {
  tryCatch(f(), error = as.character)
}
}
\keyword{ExprStr}
