% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__struct.R
\name{ExprStruct_field}
\alias{ExprStruct_field}
\alias{expr_struct_field}
\title{field}
\usage{
ExprStruct_field(name)
}
\arguments{
\item{name}{string, the Name of the struct field to retrieve.}
}
\value{
Expr: Series of same and name selected field.
}
\description{
Retrieve a \code{Struct} field as a new Series.
By default base 2.
}
\examples{
df = pl$DataFrame(
  aaa = c(1, 2),
  bbb = c("ab", "cd"),
  ccc = c(TRUE, NA),
  ddd = list(c(1, 2), 3)
)$select(
  pl$struct(pl$all())$alias("struct_col")
)
# struct field into a new Series
df$select(
  pl$col("struct_col")$struct$field("bbb"),
  pl$col("struct_col")$struct$field("ddd")
)
}
\keyword{ExprStruct}
