% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{extendr_method_to_pure_functions}
\alias{extendr_method_to_pure_functions}
\title{extendr methods into pure functions}
\usage{
extendr_method_to_pure_functions(env)
}
\arguments{
\item{env}{environment object output from extendr-wrappers.R classes}
}
\value{
env of pure function calls to rust
}
\description{
self is a global of extendr wrapper methods
this function copies the function into a new environment and
modify formals to have a self argument
}
\keyword{internals}
