% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatypes-s3-base.R
\name{format.polars_dtype}
\alias{format.polars_dtype}
\title{Format a data type}
\usage{
\method{format}{polars_dtype}(x, ..., abbreviated = FALSE)
}
\arguments{
\item{...}{Ignored.}

\item{abbreviated}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A boolean. If \code{TRUE}, use the abbreviated form of the dtype name,
e.g. "i64" instead of "Int64".}
}
\description{
Format a data type
}
\examples{
format(pl$Int64)
format(pl$Float64, abbreviated = TRUE)
}
\keyword{internal}
