% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_method_usages}
\alias{get_method_usages}
\title{Generate autocompletion suggestions for object}
\usage{
get_method_usages(env, pattern = "")
}
\arguments{
\item{env}{environment to extract usages from}

\item{pattern}{string passed to ls(pattern) to subset methods by pattern}
}
\value{
method usages
}
\description{
Generate autocompletion suggestions for object
}
\details{
used internally for auto completion in .DollarNames methods
}
\examples{
minipolars:::get_method_usages(minipolars:::DataFrame, pattern="col")
}
