% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__agg}
\alias{groupby__agg}
\title{Compute aggregations for each group of a group by operation}
\usage{
groupby__agg(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Aggregations to compute for
each group of the group by operation. Accepts expression input. Strings are
parsed as column names.}
}
\value{
A polars \link{DataFrame}
}
\description{
Compute aggregations for each group of a group by operation
}
\examples{
# Compute the aggregation of the columns for each group.
df <- pl$DataFrame(
  a = c("a", "b", "a", "b", "c"),
  b = c(1, 2, 1, 3, 3),
  c = c(5, 4, 3, 2, 1)
)
df$group_by("a")$agg(pl$col("b"), pl$col("c"))

# Compute the sum of a column for each group.
df$group_by("a")$agg(pl$col("b")$sum())

# Compute multiple aggregates at once by passing a list of expressions.
df$group_by("a")$agg(pl$sum("b"), pl$col("c")$mean())

# Use keyword arguments to easily name your expression inputs.
df$group_by("a")$agg(
  b_sum = pl$sum("b"),
  c_mean_squared = (pl$col("c") ** 2)$mean()
)
}
