% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__len}
\alias{groupby__len}
\title{Count the unique values per group}
\usage{
groupby__len(name = NULL)
}
\value{
A polars \link{DataFrame}
}
\description{
Count the unique values per group
}
\examples{
df <- pl$DataFrame(
  a = c("Apple", "Apple", "Orange"),
  b = c(1, NA, 2)
)
df$group_by("a")$len()

df$group_by("a")$len("n")
}
