% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__len}
\alias{groupby__len}
\title{Return the number of rows in each group}
\usage{
groupby__len(name = NULL)
}
\arguments{
\item{name}{Assign a name to the resulting column. If \code{NULL}, defaults to
\code{"len"}.}
}
\value{
A polars \link{DataFrame}
}
\description{
Return the number of rows in each group
}
\examples{
df <- pl$DataFrame(
  a = c("Apple", "Apple", "Orange"),
  b = c(1, NA, 2)
)
df$group_by("a")$len()

df$group_by("a")$len("n")
}
