% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__min}
\alias{groupby__min}
\title{Reduce the groups to the minimal value}
\usage{
groupby__min()
}
\value{
A polars \link{DataFrame}
}
\description{
Reduce the groups to the minimal value
}
\examples{
df <- pl$DataFrame(
  grp = c("c", "c", "a", "c", "a", "b"),
  x = c(0.5, 0.5, 4, 10, 13, 14),
  y = 1:6,
  z = c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE)
)
df

df$group_by("grp")$min()
}
