% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-group_by-general.R
\name{groupby__tail}
\alias{groupby__tail}
\title{Get the last \code{n} rows of each group}
\usage{
groupby__tail(n = 5)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A polars \link{DataFrame}
}
\description{
Get the last \code{n} rows of each group
}
\examples{
df <- pl$DataFrame(
  letters = c("c", "c", "a", "c", "a", "b"),
  nrs = 1:6
)
df

df$group_by("letters")$tail(2)$sort("letters")
}
