% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{GroupBy_agg}
\alias{GroupBy_agg}
\alias{agg}
\title{Aggregate over a GroupBy}
\usage{
GroupBy_agg(...)
}
\arguments{
\item{...}{exprs to aggregate over.
... args can also be passed wrapped in a list \verb{$agg(list(e1,e2,e3))}}
}
\value{
aggregated DataFrame
}
\description{
Aggregate a DataFrame over a groupby
}
\examples{
pl$DataFrame(
  foo = c("one", "two", "two", "one", "two"),
  bar = c(5, 3, 2, 4, 1)
)$group_by("foo")$agg(
  pl$col("bar")$sum()$name$suffix("_sum"),
  pl$col("bar")$mean()$alias("bar_tail_sum")
)
}
