% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupby.R
\name{GroupBy_first}
\alias{GroupBy_first}
\title{GroupBy First}
\usage{
GroupBy_first()
}
\value{
aggregated DataFrame
}
\description{
Reduce the groups to the first value.
}
\examples{
df = pl$DataFrame(
  a = c(1, 2, 2, 3, 4, 5),
  b = c(0.5, 0.5, 4, 10, 13, 14),
  c = c(TRUE, TRUE, TRUE, FALSE, FALSE, TRUE),
  d = c("Apple", "Orange", "Apple", "Apple", "Banana", "Banana")
)
df$groupby("d", maintain_order = TRUE)$first()
}
\keyword{GroupBy}
