% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{GroupBy_null_count}
\alias{GroupBy_null_count}
\title{GroupBy null count}
\usage{
GroupBy_null_count()
}
\value{
DataFrame
}
\description{
Create a new DataFrame that shows the null counts per column.
}
\examples{
x = mtcars
x[1:10, 3:5] = NA
pl$DataFrame(x)$group_by("cyl")$null_count()
}
\keyword{DataFrame}
