% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{import_polars_as_}
\alias{import_polars_as_}
\title{bind polars function to a namespace pl}
\usage{
import_polars_as_(name = "pl")
}
\description{
bind polars function to a namespace pl
}
\examples{
import_polars_as_("pl")
}
