% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read a parquet file}
\arguments{
\item{file}{string filepath}

\item{n_rows}{limit rows to scan}

\item{cache}{bool use cache}

\item{parallel}{String either Auto, None, Columns or RowGroups. The way to parallelized the scan.}

\item{rechunk}{bool rechunk reorganize memory layout, potentially make future operations faster , however perform reallocation now.}

\item{row_count_name}{NULL or string, if a string add a rowcount column named by this string}

\item{row_count_offset}{integer, the rowcount column can be offset by this value}

\item{low_memory}{bool, try reduce memory footprint}
}
\value{
DataFrame
}
\description{
Read a parquet file
}
