% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_ipc.R
\name{pl_scan_ipc}
\alias{pl_scan_ipc}
\title{Import data in Apache Arrow IPC format}
\usage{
pl_scan_ipc(
  source,
  ...,
  n_rows = NULL,
  cache = TRUE,
  rechunk = FALSE,
  row_index_name = NULL,
  row_index_offset = 0L,
  memory_map = TRUE
)
}
\arguments{
\item{source}{Path to a file or URL. It is possible to provide multiple paths
provided that all CSV files have the same schema. It is not possible to
provide several URLs.}

\item{...}{Ignored.}

\item{n_rows}{Maximum number of rows to read.}

\item{cache}{Cache the result after reading.}

\item{rechunk}{Reallocate to contiguous memory when all chunks / files are
parsed.}

\item{row_index_name}{If not \code{NULL}, this will insert a row index column with
the given name into the DataFrame.}

\item{row_index_offset}{Offset to start the row index column (only used if
the name is set).}

\item{memory_map}{A logical. If \code{TRUE}, try to memory map the file.
This can greatly improve performance on repeated queries as the OS may cache pages.
Only uncompressed Arrow IPC files can be memory mapped.}
}
\value{
LazyFrame
}
\description{
Import data in Apache Arrow IPC format
}
\details{
Create new LazyFrame from Apache Arrow IPC file or stream
}
\keyword{LazyFrame_new}
