% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{scan_parquet}
\alias{scan_parquet}
\title{Scan a parquet file}
\arguments{
\item{file}{string filepath}

\item{n_rows}{limit rows to scan}

\item{cache}{bool use cache}

\item{parallel}{String either Auto, None, Columns or RowGroups. The way to parallelized the scan.}

\item{rechunk}{bool rechunk reorganize memory layout, potentially make future operations faster , however perform reallocation now.}

\item{row_count_name}{NULL or string, if a string add a rowcount column named by this string}

\item{row_count_offset}{integer, the rowcount column can be offset by this value}

\item{low_memory}{bool, try reduce memory footprint}

\item{hive_partitioning}{Infer statistics and schema from hive partitioned URL
and use them to prune reads.}
}
\value{
LazyFrame
}
\description{
Scan a parquet file
}
\examples{
\dontshow{if (requireNamespace("arrow", quietly = TRUE) && arrow::arrow_with_dataset() && arrow::arrow_with_parquet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
temp_dir = tempfile()
# Write a hive-style partitioned parquet dataset
arrow::write_dataset(
  mtcars,
  temp_dir,
  partitioning = c("cyl", "gear"),
  format = "parquet",
  hive_style = TRUE
)
list.files(temp_dir, recursive = TRUE)

# Read the dataset
pl$scan_parquet(
  file.path(temp_dir, "**/*.parquet")
)$collect()
\dontshow{\}) # examplesIf}
}
\keyword{LazyFrame_new}
