% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{scan_parquet}
\alias{scan_parquet}
\title{Scan a parquet file}
\arguments{
\item{file}{string filepath}

\item{n_rows}{limit rows to scan}

\item{cache}{bool use cache}

\item{parallel}{String either Auto, None, Columns or RowGroups. The way to parallelized the scan.}

\item{rechunk}{bool rechunk reorganize memory layout, potentially make future operations faster , however perform reallocation now.}

\item{row_count_name}{NULL or string, if a string add a rowcount column named by this string}

\item{row_count_offset}{integer, the rowcount column can be offset by this value}

\item{low_memory}{bool, try reduce memory footprint}
}
\value{
LazyFrame
}
\description{
Scan a parquet file
}
\examples{
#write example file
pl$LazyFrame(mtcars)$sink_parquet("my.parquet")

# scan and get (project) only one column "cyl" but filter (apply predicate) on "hp".
lf = pl$scan_parquet("my.parquet")$
  filter(pl$col("hp") > 250)$
  select(pl$col("cyl") * 2)

# LayFrame with a logical plan (query)
print(lf)

# see optimized plan
lf$describe_optimized_plan()

# Execute and get result DataFrame
lf$collect()
}
\keyword{LazyFrame_new}
