% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{DataFrame_write_ipc}
\alias{DataFrame_write_ipc}
\title{Write to Arrow IPC file (a.k.a Feather file)}
\usage{
DataFrame_write_ipc(
  file,
  compression = c("uncompressed", "zstd", "lz4"),
  ...,
  future = FALSE
)
}
\arguments{
\item{file}{File path to which the result should be written.}

\item{compression}{\code{NULL} or a character of the compression method,
\code{"uncompressed"} or "lz4" or "zstd". \code{NULL} is equivalent to \code{"uncompressed"}.
Choose "zstd" for good compression performance. Choose "lz4"
for fast compression/decompression.}

\item{...}{Ignored.}

\item{future}{Setting this to \code{TRUE} will write Polars' internal data structures that
might not be available by other Arrow implementations.
This functionality is considered \strong{unstable}.
It may be changed at any point without it being considered a breaking change.}
}
\value{
Invisibly returns the input DataFrame.
}
\description{
Write to Arrow IPC file (a.k.a Feather file)
}
\examples{
dat = pl$DataFrame(mtcars)

destination = tempfile(fileext = ".arrow")
dat$write_ipc(destination)

if (require("arrow", quietly = TRUE)) {
  arrow::read_ipc_file(destination, as_data_frame = FALSE)
}
}
