% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{is_DataFrame_data_input}
\alias{is_DataFrame_data_input}
\title{Validate data input for create Dataframe with pl$DataFrame}
\usage{
is_DataFrame_data_input(x)
}
\arguments{
\item{robj}{any R object to test}
}
\value{
bool
}
\description{
The Dataframe constructors accepts data.frame inheritors or list of vectors and/or Series.
}
\examples{
rpolars:::is_DataFrame_data_input(iris)
rpolars:::is_DataFrame_data_input(list(1:5,pl$Series(1:5),letters[1:5]))
}
