% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{is_DataFrame_data_input}
\alias{is_DataFrame_data_input}
\title{test if suitable to construct polar.frame}
\usage{
is_DataFrame_data_input(x)
}
\arguments{
\item{robj}{any R object to test}
}
\value{
bool
}
\description{
test if suitable to construct polar.frame
}
\examples{
is_polar_data_input(iris)
}
