% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{is_polars_dtype}
\alias{is_polars_dtype}
\title{Check if input is a valid DataType}
\usage{
is_polars_dtype(x, include_unknown = FALSE)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
A boolean scalar.
}
\description{
Check if input is a valid DataType
}
\examples{
is_polars_dtype(pl$Int64)
is_polars_dtype("numeric")
is_polars_dtype(mtcars)
}
