% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{is_polars_dtype}
\alias{is_polars_dtype}
\title{chek if x is a valid RPolarsDataType}
\usage{
is_polars_dtype(x, include_unknown = FALSE)
}
\arguments{
\item{x}{a candidate}
}
\value{
a list DataType with an inner DataType
}
\description{
chek if x is a valid RPolarsDataType
}
\examples{
polars:::is_polars_dtype(pl$Int64)
}
\keyword{internal}
