% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-knitr.R
\name{knit_print.DataFrame}
\alias{knit_print.DataFrame}
\title{knit print polars DataFrame}
\usage{
knit_print.DataFrame(x, ...)
}
\arguments{
\item{x}{a polars DataFrame to knit_print}

\item{...}{additional arguments, not used}
}
\description{
Mimics \link{python-polars' NotebookFormatter}
(https://github.com/pola-rs/polars/blob/4ffcb7461302a580255a9d910d70f1f7b2508675/py-polars/polars/dataframe/_html.py)
for HTML outputs.
}
\keyword{DataFrame}
