% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{lazy_groupby}
\alias{lazy_groupby}
\title{R6 Class polar_groupby}
\description{
R6 Class polar_groupby

R6 Class polar_groupby
}
\details{
high-level wrapper class for Rpolarsgroupby
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-lazy_polar_groupby-new}{\code{lazy_groupby$new()}}
\item \href{#method-lazy_polar_groupby-agg}{\code{lazy_groupby$agg()}}
\item \href{#method-lazy_polar_groupby-apply}{\code{lazy_groupby$apply()}}
\item \href{#method-lazy_polar_groupby-head}{\code{lazy_groupby$head()}}
\item \href{#method-lazy_polar_groupby-tail}{\code{lazy_groupby$tail()}}
\item \href{#method-lazy_polar_groupby-print}{\code{lazy_groupby$print()}}
\item \href{#method-lazy_polar_groupby-clone}{\code{lazy_groupby$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_groupby-new"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_groupby-new}{}}}
\subsection{Method \code{new()}}{
wrap low-level Rlazygroupby in new R6 lazy_goupby object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_groupby$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Rlazygroupby}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_groupby} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_groupby-agg"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_groupby-agg}{}}}
\subsection{Method \code{agg()}}{
aggregate a polar_lazy_groupby
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_groupby$agg(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{any Rexpr or string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_groupby-apply"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_groupby-apply}{}}}
\subsection{Method \code{apply()}}{
one day this will apply
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_groupby$apply(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{not used}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_groupby-head"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_groupby-head}{}}}
\subsection{Method \code{head()}}{
get n rows of head of group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_groupby$head(n = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer number of rows to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_groupby-tail"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_groupby-tail}{}}}
\subsection{Method \code{tail()}}{
get n tail rows of group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_groupby$tail(n = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer number of rows to get}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{lazy_polar_frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_groupby-print"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_groupby-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_groupby$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lazy_polar_groupby-clone"></a>}}
\if{latex}{\out{\hypertarget{method-lazy_polar_groupby-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lazy_groupby$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
