% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__drop_nans}
\alias{lazyframe__drop_nans}
\title{Drop all rows that contain NaN values}
\usage{
lazyframe__drop_nans(subset = NULL)
}
\arguments{
\item{subset}{Column name(s) for which \code{NaN} values are considered. If \code{NULL}
(default), use all columns (note that only floating-point columns can
contain \code{NaN}s).}
}
\value{
A polars \link{LazyFrame}
}
\description{
The original order of the remaining rows is preserved.
}
\examples{
lf <- pl$LazyFrame(
  foo = c(1, NaN, 2.5),
  bar = c(NaN, 110, 25.5),
  ham = c("a", "b", NA)
)

# The default behavior of this method is to drop rows where any single value
# of the row is null.
lf$drop_nans()$collect()

# This behaviour can be constrained to consider only a subset of columns, as
# defined by name or with a selector. For example, dropping rows if there is
# a null in the "bar" column:
lf$drop_nans(subset = "bar")$collect()

# Dropping a row only if *all* values are NaN requires a different
# formulation:
df <- pl$LazyFrame(
  a = c(NaN, NaN, NaN, NaN),
  b = c(10.0, 2.5, NaN, 5.25),
  c = c(65.75, NaN, NaN, 10.5)
)
df$filter(!pl$all_horizontal(pl$all()$is_nan()))$collect()
}
