% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__explode}
\alias{lazyframe__explode}
\title{Explode the frame to long format by exploding the given columns}
\usage{
lazyframe__explode(..., empty_as_null = TRUE, keep_nulls = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names or \link[=polars_selector]{selectors}
defining them. The underlying columns being exploded must be of
the \code{List} or \code{Array} data type.}

\item{empty_as_null}{Indicates to explode an empty list/array into a \code{null}.}

\item{keep_nulls}{Indicates to explode a \code{null} list/array into a \code{null}.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Explode the frame to long format by exploding the given columns
}
\examples{
lf <- pl$LazyFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)

lf$explode("numbers")$collect()
}
