% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__explode}
\alias{lazyframe__explode}
\title{Explode the DataFrame to long format by exploding the given columns}
\usage{
lazyframe__explode(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column names, expressions, or
a selector defining them. The underlying columns being exploded must be of
the \code{List} or \code{Array} data type.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Explode the DataFrame to long format by exploding the given columns
}
\examples{
lf <- pl$LazyFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)

lf$explode("numbers")$collect()
}
