% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__fill_nan}
\alias{lazyframe__fill_nan}
\title{Fill floating point \code{NaN} value with a fill value}
\usage{
lazyframe__fill_nan(value)
}
\arguments{
\item{value}{Value used to fill \code{NaN} values.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Fill floating point \code{NaN} value with a fill value
}
\examples{
lf <- pl$LazyFrame(
  a = c(1.5, 2, NaN, 4),
  b = c(1.5, NaN, NaN, 4)
)
lf$fill_nan(99)$collect()
}
