% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__interpolate}
\alias{lazyframe__interpolate}
\title{Interpolate intermediate values}
\usage{
lazyframe__interpolate()
}
\value{
A polars \link{LazyFrame}
}
\description{
The interpolation method is linear.
}
\examples{
lf <- pl$LazyFrame(
  foo = c(1, NA, 9, 10),
  bar = c(6, 7, 9, NA),
  ham = c(1, NA, NA, 9)
)

lf$interpolate()$collect()
}
