% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__median}
\alias{lazyframe__median}
\title{Aggregate the columns in the LazyFrame to their median value}
\usage{
lazyframe__median()
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns in the LazyFrame to their median value
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$median()$collect()
}
