% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__min}
\alias{lazyframe__min}
\title{Aggregate the columns in the LazyFrame to their minimum value}
\usage{
lazyframe__min()
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns in the LazyFrame to their minimum value
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$min()$collect()
}
