% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__null_count}
\alias{lazyframe__null_count}
\title{Return the number of null elements for each column}
\usage{
lazyframe__null_count()
}
\value{
A polars \link{LazyFrame}
}
\description{
Return the number of null elements for each column
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, NA), c = rep(NA, 4))
lf$null_count()$collect()
}
