% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__quantile}
\alias{lazyframe__quantile}
\title{Aggregate the columns in the DataFrame to a unique quantile value}
\usage{
lazyframe__quantile(
  quantile,
  interpolation = c("nearest", "higher", "lower", "midpoint", "linear")
)
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns in the DataFrame to a unique quantile value
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$quantile(0.7)$collect()
}
