% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__remove}
\alias{lazyframe__remove}
\title{Remove rows, dropping those that match the given predicate expression(s)}
\usage{
lazyframe__remove(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Expression that evaluates to
a boolean Series.}
}
\value{
A polars \link{LazyFrame}
}
\description{
The original order of the remaining rows is preserved. Rows where the filter
does not evaluate to \code{TRUE} are retained (this includes rows where the
predicate evaluates as \code{null}).
}
\examples{
lf <- pl$LazyFrame(
  foo = c(2, 3, NA, 4, 0),
  bar = c(5, 6, NA, NA, 0),
  ham = c("a", "b", NA, "c", "d")
)

# Remove rows matching a condition:
lf$remove(pl$col("bar") >= 5)$collect()

# Discard rows based on multiple conditions, combined with and/or operators:
lf$remove((pl$col("foo") >= 0) & (pl$col("bar") >= 0))$collect()

lf$remove((pl$col("foo") >= 0) | (pl$col("bar") >= 0))$collect()

# Remove rows by comparing two columns against each other:
lf$remove(pl$col("foo")$ne_missing(pl$col("bar")))$collect()
}
