% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__remove}
\alias{lazyframe__remove}
\title{Remove rows, dropping those that match the given predicate expression(s)}
\usage{
lazyframe__remove(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Expression that evaluates to
a boolean Series.}
}
\value{
A polars \link{LazyFrame}
}
\description{
The original order of the remaining rows is preserved. Rows where the filter
does not evaluate to \code{TRUE} are retained (this includes rows where the
predicate evaluates as \code{null}).
}
\examples{
lf <- pl$LazyFrame(
  ccy = c("USD", "EUR", "USD", "JPY"),
  year = c(2021, 2022, 2023, 2023),
  total = c(3245, NA, -6680, 25000),
)

# Remove rows matching a condition. Note that the row where `total` is null
# is kept:
lf$remove(pl$col("total") >= 0)$collect()

# Note that this is *not* the same as simply inverting the condition in
# `$filter()` because `$filter()` doesn't keep predicates that evaluate to
# null:
lf$filter(pl$col("total") < 0)$collect()

# We can use multiple conditions, combined with and/or operators:
lf$remove((pl$col("total") >= 0) & (pl$col("ccy") == "USD"))$collect()

lf$remove((pl$col("total") >= 0) | (pl$col("ccy") == "USD"))$collect()
}
