% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__select}
\alias{lazyframe__select}
\title{Select and modify columns of a LazyFrame}
\usage{
lazyframe__select(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Name-value pairs of objects to be converted to polars \link[=Expr]{expressions}
by the \code{\link[=as_polars_expr]{as_polars_expr()}} function.
Characters are parsed as column names, other non-expression inputs are parsed as \link[=pl__lit]{literals}.
Each name will be used as the expression name.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Select and perform operations on a subset of columns only. This discards
unmentioned columns (like \code{.()} in \code{data.table} and contrarily to
\code{dplyr::mutate()}).

One cannot use new variables in subsequent expressions in the same
\verb{$select()} call. For instance, if you create a variable \code{x}, you will only
be able to use it in another \verb{$select()} or \verb{$with_columns()} call.
}
\examples{
# Pass the name of a column to select that column.
lf <- pl$LazyFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)
lf$select("foo")$collect()

# Multiple columns can be selected by passing a list of column names.
lf$select("foo", "bar")$collect()

# Expressions are also accepted.
lf$select(pl$col("foo"), pl$col("bar") + 1)$collect()

# Name expression (used as the column name of the output DataFrame)
lf$select(
  threshold = pl$when(pl$col("foo") > 2)$then(10)$otherwise(0)
)$collect()

# Expressions with multiple outputs can be automatically instantiated
# as Structs by setting the `POLARS_AUTO_STRUCTIFY` environment variable.
# (Experimental)
if (requireNamespace("withr", quietly = TRUE)) {
  withr::with_envvar(c(POLARS_AUTO_STRUCTIFY = "1"), {
    lf$select(
      is_odd = ((pl$col(pl$Int32) \%\% 2) == 1)$name$suffix("_is_odd"),
    )$collect()
  })
}
}
