% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__select_seq}
\alias{lazyframe__select_seq}
\title{Select columns from this LazyFrame}
\usage{
lazyframe__select_seq(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Name-value pairs of objects to be converted to polars \link[=Expr]{expressions}
by the \code{\link[=as_polars_expr]{as_polars_expr()}} function.
Characters are parsed as column names, other non-expression inputs are parsed as
\link[=pl__lit]{literals}. Each name will be used as the expression name.}
}
\value{
A polars \link{LazyFrame}
}
\description{
This will run all expression sequentially instead of in parallel. Use this
when the work per expression is cheap.
}
\examples{
lf <- pl$LazyFrame(
  foo = 1:3,
  bar = 6:8,
  ham = letters[1:3]
)
lf$select_seq("foo", bar2 = pl$col("bar") * 2)$collect()
}
