% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__serialize}
\alias{lazyframe__serialize}
\title{Serialize the logical plan of this LazyFrame to a string in JSON format}
\usage{
lazyframe__serialize()
}
\value{
A character value
}
\description{
Serialize the logical plan of this LazyFrame to a string in JSON format
}
\examples{
lf <- pl$LazyFrame(a = 1:3)$sum()
lf$serialize()
}
