% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__shift}
\alias{lazyframe__shift}
\title{Shift values by the given number of indices}
\usage{
lazyframe__shift(n = 1, ..., fill_value = NULL)
}
\arguments{
\item{n}{Number of indices to shift forward. If a negative value is passed,
values are shifted in the opposite direction instead.}

\item{...}{These dots are for future extensions and must be empty.}

\item{fill_value}{Fill the resulting null values with this value. Accepts
expression input. Non-expression inputs are parsed as literals.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Shift values by the given number of indices
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = 5:8)

# By default, values are shifted forward by one index.
lf$shift()$collect()

# Pass a negative value to shift in the opposite direction instead.
lf$shift(-2)$collect()

# Specify fill_value to fill the resulting null values.
lf$shift(-2, fill_value = 100)$collect()
}
