% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__sort}
\alias{lazyframe__sort}
\title{Sort the LazyFrame by the given columns}
\usage{
lazyframe__sort(
  ...,
  descending = FALSE,
  nulls_last = FALSE,
  multithreaded = TRUE,
  maintain_order = FALSE
)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Column(s) to sort by. Can be
character values indicating column names or Expr(s).}

\item{descending}{Sort in descending order. When sorting by multiple
columns, this can be specified per column by passing a logical vector.}

\item{nulls_last}{Place null values last. When sorting by multiple
columns, this can be specified per column by passing a logical vector.}

\item{multithreaded}{Sort using multiple threads.}

\item{maintain_order}{Whether the order should be maintained if elements are
equal. If \code{TRUE}, streaming is not possible and performance might be worse
since this requires a stable search.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Sort the LazyFrame by the given columns
}
\examples{
lf <- pl$LazyFrame(
  a = c(1, 2, NA, 4),
  b = c(6, 5, 4, 3),
  c = c("a", "c", "b", "a")
)

# Pass a single column name to sort by that column.
lf$sort("a")$collect()

# Sorting by expressions is also supported
lf$sort(pl$col("a") + pl$col("b") * 2, nulls_last = TRUE)$collect()

# Sort by multiple columns by passing a vector of columns
lf$sort(c("c", "a"), descending = TRUE)$collect()

# Or use positional arguments to sort by multiple columns in the same way
lf$sort("c", "a", descending = c(FALSE, TRUE))$collect()
}
