% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__unique}
\alias{lazyframe__unique}
\title{Drop duplicate rows from this DataFrame}
\usage{
lazyframe__unique(
  subset = NULL,
  ...,
  keep = c("any", "none", "first", "last"),
  maintain_order = FALSE
)
}
\arguments{
\item{subset}{Column name(s) or selector(s), to consider when identifying
duplicate rows. If \code{NULL} (default), use all columns.}

\item{...}{These dots are for future extensions and must be empty.}

\item{keep}{Which of the duplicate rows to keep. Must be one of:
\itemize{
\item \code{"any"}: does not give any guarantee of which row is kept. This allows
more optimizations.
\item \code{"none"}: don’t keep duplicate rows.
\item \code{"first"}: keep first unique row.
\item \code{"last"}: keep last unique row.
}}

\item{maintain_order}{Keep the same order as the original LazyFrame. This is
more expensive to compute. Setting this to \code{TRUE} blocks the possibility to
run on the streaming engine.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Drop duplicate rows from this DataFrame
}
\examples{
lf <- pl$LazyFrame(
  foo = c(1, 2, 3, 1),
  bar = c("a", "a", "a", "a"),
  ham = c("b", "b", "b", "b"),
)
lf$unique(maintain_order = TRUE)$collect()

lf$unique(subset = c("bar", "ham"), maintain_order = TRUE)$collect()

lf$unique(keep = "last", maintain_order = TRUE)$collect()
}
