% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__unpivot}
\alias{lazyframe__unpivot}
\title{Unpivot a LazyFrame from wide to long format}
\usage{
lazyframe__unpivot(
  on = NULL,
  ...,
  index = NULL,
  variable_name = NULL,
  value_name = NULL
)
}
\arguments{
\item{on}{Values to use as identifier variables. If \code{value_vars} is
empty all columns that are not in \code{id_vars} will be used.}

\item{...}{These dots are for future extensions and must be empty.}

\item{index}{Columns to use as identifier variables.}

\item{variable_name}{Name to give to the new column containing the names of
the melted columns. Defaults to "variable".}

\item{value_name}{Name to give to the new column containing the values of
the melted columns. Defaults to \code{"value"}.}
}
\value{
A polars \link{LazyFrame}
}
\description{
This function is useful to massage a LazyFrame into a format where one or
more columns are identifier variables (\code{index}) while all other columns,
considered measured variables (\code{on}), are “unpivoted” to the row axis
leaving just two non-identifier columns, "variable" and "value".
}
\examples{
lf <- pl$LazyFrame(
  a = c("x", "y", "z"),
  b = c(1, 3, 5),
  c = c(2, 4, 6)
)
lf$unpivot(index = "a", on = c("b", "c"))$collect()
}
