% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-frame.R
\name{lazyframe__var}
\alias{lazyframe__var}
\title{Aggregate the columns in the LazyFrame to their variance value}
\usage{
lazyframe__var(ddof = 1)
}
\value{
A polars \link{LazyFrame}
}
\description{
Aggregate the columns in the LazyFrame to their variance value
}
\examples{
lf <- pl$LazyFrame(a = 1:4, b = c(1, 2, 1, 1))
lf$var()$collect()
lf$var(ddof = 0)$collect()
}
