% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_clone}
\alias{LazyFrame_clone}
\title{Clone a LazyFrame
This makes a very cheap deep copy/clone of an existing \code{LazyFrame}.}
\usage{
LazyFrame_clone()
}
\value{
A LazyFrame
}
\description{
Clone a LazyFrame
This makes a very cheap deep copy/clone of an existing \code{LazyFrame}.
}
\examples{
pl$LazyFrame(mtcars)$clone()
}
