% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_collect_background}
\alias{LazyFrame_collect_background}
\title{New DataFrame from LazyFrame_object$collect()}
\usage{
LazyFrame_collect_background()
}
\value{
collected \code{DataFrame}
}
\description{
collect DataFrame by lazy query
}
\examples{
pl$DataFrame(iris)$lazy()$filter(pl$col("Species") == "setosa")$collect()
}
\keyword{DataFrame_new}
\keyword{LazyFrame}
