% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_collect_in_background}
\alias{LazyFrame_collect_in_background}
\title{Collect_in_background}
\usage{
LazyFrame_collect_in_background()
}
\value{
a thread handle for the task
}
\description{
collect DataFrame by lazy query in a background thread.
}
\details{
Useful to not block the R session while query executes. Any use of \verb{$map()} or \verb{$apply()} is
diverted to other R sessions. If R functions rely on global variables this can change the
result.
}
\examples{
prexpr <- pl$col("mpg")$map(\(x) {Sys.sleep(1.5);x * 0.43}, in_background = TRUE)$alias("kml")
handle = pl$LazyFrame(mtcars)$with_column(prexpr)$collect_in_background()
if(!handle$is_finished()) print("not done yet")
df = handle$join() #get result
df
}
\keyword{DataFrame_new}
\keyword{LazyFrame}
