% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_columns}
\alias{LazyFrame_columns}
\title{Get the column names of a LazyFrame}
\usage{
<LazyFrame>$columns()
}
\value{
A vector of column names
}
\description{
Get the column names of a LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$columns
}
\keyword{LazyFrame}
