% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\docType{data}
\name{LazyFrame_describe_plan}
\alias{LazyFrame_describe_plan}
\title{Print the non-optimized plan plan of LazyFrame}
\format{
An object of class \code{character} of length 1.
}
\usage{
LazyFrame_describe_plan
}
\description{
select on a LazyFrame
}
\keyword{LazyFrame}
