% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_drop}
\alias{LazyFrame_drop}
\title{Drop}
\usage{
LazyFrame_drop(columns)
}
\arguments{
\item{character}{vector Name of the column(s) that should be removed from the dataframe.}
}
\value{
LazyFrame
}
\description{
Remove columns from the dataframe.
}
\examples{
pl$DataFrame(mtcars)$lazy()$drop(c("mpg", "hp"))
}
\keyword{LazyFrame}
