% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_drop}
\alias{LazyFrame_drop}
\title{Drop columns of a LazyFrame}
\usage{
<LazyFrame>$drop(columns)
}
\arguments{
\item{columns}{A character vector with the names of the column(s) to remove.}
}
\value{
LazyFrame
}
\description{
Drop columns of a LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$drop(c("mpg", "hp"))
}
\keyword{LazyFrame}
