% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_drop_nulls}
\alias{LazyFrame_drop_nulls}
\title{Lazy_drop_nulls}
\usage{
LazyFrame_drop_nulls(subset = NULL)
}
\arguments{
\item{subset}{string or vector of strings. Column name(s) for which null values are considered. If set to NULL (default), use all columns.}
}
\value{
LazyFrame
}
\description{
Drop all rows that contain null values.
}
\examples{
tmp = mtcars
tmp[1:3, "mpg"] = NA
tmp[4, "hp"] = NA
pl$DataFrame(tmp)$lazy()$drop_nulls()$collect()$height
pl$DataFrame(tmp)$lazy()$drop_nulls("mpg")$collect()$height
pl$DataFrame(tmp)$lazy()$drop_nulls(c("mpg", "hp"))$collect()$height
}
\keyword{LazyFrame}
