% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_schema}
\alias{LazyFrame_schema}
\alias{LazyFrame_dtypes}
\title{Data types information}
\usage{
<LazyFrame>$schema()

LazyFrame_dtypes()
}
\value{
\verb{$dtypes} returns an unnamed list with the data type of each column.
\verb{$schema} returns a named list with the column names and the data type of
each column.
}
\description{
Get the data type of all columns. You can see all available
types with \code{names(pl$dtypes)}. The data type of each column is also shown
when printing the DataFrame.
}
\examples{
pl$LazyFrame(iris)$dtypes

pl$LazyFrame(iris)$schema
}
\keyword{LazyFrame}
