% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_explode}
\alias{LazyFrame_explode}
\title{Explode the DataFrame to long format by exploding the given columns}
\usage{
LazyFrame_explode(columns = list(), ...)
}
\arguments{
\item{columns}{Column(s) to be exploded. \verb{Into<Expr>}, list of \verb{Into<Expr>} or a char vec.
Only columns of DataType \code{List} or \code{Utf8} can be exploded.}

\item{...}{More columns to explode as above but provided as separate arguments}
}
\value{
LazyFrame
}
\description{
Explode the DataFrame to long format by exploding the given columns
}
\examples{
df = pl$LazyFrame(
  letters = c("a", "a", "b", "c"),
  numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
)
df

df$explode("numbers")$collect()
}
\keyword{LazyFrame}
