% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_fetch}
\alias{LazyFrame_fetch}
\title{Dtypes}
\usage{
LazyFrame_fetch(n_rows = 500)
}
\arguments{
\item{n_rows}{number of rows to fetch at maximum.}
}
\value{
A DataFrame of maximum n_rows
}
\description{
Get rows
}
\examples{
pl$LazyFrame(irirs)$fetch(3)
}
\keyword{LazyFrame}
