% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_filter}
\alias{LazyFrame_filter}
\title{Apply filter to LazyFrame}
\usage{
LazyFrame_filter(...)
}
\arguments{
\item{...}{Polars expressions which will evaluate to a boolean.}
}
\value{
A new \code{LazyFrame} object with add/modified column.
}
\description{
Filter rows with an Expression defining a boolean column.
Multiple expressions are combined with \code{&} (AND).
This is equivalent to \code{\link[dplyr:filter]{dplyr::filter()}}.
}
\details{
Rows where the condition returns \code{NA} are dropped.
}
\examples{
lf = pl$LazyFrame(iris)

lf$filter(pl$col("Species") == "setosa")$collect()

# This is equivalent to
# lf$filter(pl$col("Sepal.Length") > 5 & pl$col("Petal.Width") < 1)
lf$filter(pl$col("Sepal.Length") > 5, pl$col("Petal.Width") < 1)
}
\keyword{LazyFrame}
