% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_first}
\alias{LazyFrame_first}
\title{Get the first row of a LazyFrame}
\usage{
LazyFrame_first()
}
\value{
A LazyFrame with one row
}
\description{
Get the first row of a LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$first()$collect()
}
\keyword{DataFrame}
