% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyFrame_groupby}
\alias{LazyFrame_groupby}
\title{Lazy_groupby}
\usage{
LazyFrame_groupby(..., maintain_order = FALSE)
}
\arguments{
\item{...}{any single Expr or string naming a column}

\item{maintain_order}{bool should an aggregate of groupby retain order of groups or FALSE = random, slightly faster?}
}
\value{
A new \code{LazyGroupBy} object with applied groups.
}
\description{
apply groupby on LazyFrame, return LazyGroupBy
}
\keyword{LazyFrame}
\keyword{LazyFrame.}
\keyword{groupby}
\keyword{on}
