% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_head}
\alias{LazyFrame_head}
\title{Get the first \emph{n} rows}
\usage{
LazyFrame_head(n)
}
\arguments{
\item{n}{Positive numeric or integer number not larger than 2^32. Any number
will be converted to \code{u32}. Negative raises error.}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
Get the first \emph{n} rows of the query.
}
\examples{
pl$LazyFrame(mtcars)$head(5)$collect()
}
\keyword{LazyFrame}
