% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_head}
\alias{LazyFrame_head}
\title{Get the first \code{n} rows.}
\usage{
LazyFrame_head(n = 5L)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
A shortcut for \code{\link[=LazyFrame_slice]{$slice(0, n)}}.
Conider using the \code{\link[=LazyFrame_fetch]{$fetch()}} method if you want to test your query.
The \code{\link[=LazyFrame_fetch]{$fetch()}} operation will load the first \code{n} rows at
the scan level, whereas \verb{$head()} are applied at the end.
}
\details{
\verb{$limit()} is an alias for \verb{$head()}.
}
\examples{
lf = pl$LazyFrame(a = 1:6, b = 7:12)

lf$head()$collect()

lf$head(2)$collect()
}
