% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_head}
\alias{LazyFrame_head}
\title{Head of a LazyFrame}
\usage{
<LazyFrame>$head(n)
}
\arguments{
\item{n}{Positive number not larger than 2^32.}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
Get the first \code{n} rows of the query.
}
\details{
Any number will converted to u32. Negative raises error.
}
\examples{
pl$LazyFrame(mtcars)$head(4)$collect()
}
