% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_join}
\alias{LazyFrame_join}
\title{LazyFrame join}
\usage{
LazyFrame_join(
  other,
  left_on = NULL,
  right_on = NULL,
  on = NULL,
  how = c("inner", "left", "outer", "semi", "anti", "cross"),
  suffix = "_right",
  allow_parallel = TRUE,
  force_parallel = FALSE
)
}
\arguments{
\item{other}{LazyFrame}

\item{left_on}{names of columns in self LazyFrame, order should match. Type, see on param.}

\item{right_on}{names of columns in other LazyFrame, order should match. Type, see on param.}

\item{on}{named columns as char vector of named columns, or list of expressions and/or strings.}

\item{how}{a string selecting one of the following methods: inner, left, outer, semi, anti, cross}

\item{suffix}{name to added right table}

\item{allow_parallel}{bool}

\item{force_parallel}{bool}
}
\value{
A new \code{LazyFrame} object with applied join.
}
\description{
join a LazyFrame
}
\keyword{LazyFrame}
