% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_last}
\alias{LazyFrame_last}
\title{Last}
\usage{
LazyFrame_last()
}
\value{
A new \code{LazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their maximum value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$last()$collect()
}
\keyword{LazyFrame}
