% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_last}
\alias{LazyFrame_last}
\title{Get the last row of a LazyFrame}
\usage{
LazyFrame_last()
}
\value{
A LazyFrame with one row
}
\description{
Aggregate the columns in the LazyFrame to their maximum value.
}
\examples{
pl$LazyFrame(mtcars)$last()$collect()
}
\keyword{LazyFrame}
