% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyFrame_limit}
\alias{LazyFrame_limit}
\title{Limits}
\usage{
LazyFrame_limit(n)
}
\arguments{
\item{n}{positive numeric or integer number not larger than 2^32}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
take limit of n rows of query
}
\details{
any number will converted to u32. Negative raises error
}
\keyword{LazyFrame}
