% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_max}
\alias{LazyFrame_max}
\title{Max}
\usage{
<LazyFrame>$max()
}
\value{
A LazyFrame with one row
}
\description{
Aggregate the columns in the LazyFrame to their maximum value.
}
\examples{
pl$LazyFrame(mtcars)$max()$collect()
}
\keyword{LazyFrame}
