% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_mean}
\alias{LazyFrame_mean}
\title{Mean}
\format{
function
}
\usage{
LazyFrame_mean
}
\value{
A new \code{LazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their mean value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$mean()$collect()
}
\keyword{LazyFrame}
