% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_median}
\alias{LazyFrame_median}
\title{Median}
\usage{
LazyFrame_median()
}
\value{
A new \code{LazyFrame} object with applied aggregation.
}
\description{
Aggregate the columns in the DataFrame to their median value.
}
\examples{
pl$DataFrame(mtcars)$lazy()$median()$collect()
}
\keyword{LazyFrame}
