% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_min}
\alias{LazyFrame_min}
\title{Min}
\usage{
<LazyFrame>$min()
}
\value{
A LazyFrame with one row
}
\description{
Aggregate the columns in the LazyFrame to their minimum value.
}
\examples{
pl$LazyFrame(mtcars)$min()$collect()
}
\keyword{LazyFrame}
