% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_print}
\alias{LazyFrame_print}
\title{print LazyFrame internal method}
\usage{
LazyFrame_print(x)
}
\arguments{
\item{x}{LazyFrame}
}
\value{
self
}
\description{
can be used i the middle of a method chain
}
\examples{
pl$LazyFrame(iris)$print()
}
\keyword{LazyFrame}
