% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_profile}
\alias{LazyFrame_profile}
\title{Profile}
\usage{
LazyFrame_profile()
}
\value{
A pair of DataFrames, (collected result, profile stats)
}
\description{
Profile
}
\examples{
pl$LazyFrame(mtcars)$
  select(pl$col("mpg") * 0.43)
  profile()

}
\keyword{LazyFrame}
