% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_rename}
\alias{LazyFrame_rename}
\title{Rename columns of a LazyFrame}
\usage{
LazyFrame_rename(...)
}
\arguments{
\item{...}{Use \code{new_name = "old_name"} to rename selected variables.}
}
\value{
LazyFrame
}
\description{
Rename columns of a LazyFrame
}
\examples{
pl$DataFrame(mtcars)$
  lazy()$
  rename(miles_per_gallon = "mpg", horsepower = "hp")$
  collect()
}
\keyword{LazyFrame}
