% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_rename}
\alias{LazyFrame_rename}
\title{Rename columns of a DataFrame}
\usage{
LazyFrame_rename(...)
}
\arguments{
\item{...}{One of the following:
\itemize{
\item params like \code{new_name = "old_name"} to rename selected variables.
\item as above but with params wrapped in a list
}}
}
\value{
LazyFrame
}
\description{
Rename columns of a DataFrame
}
\examples{
pl$LazyFrame(mtcars)$
  lazy()$
  rename(miles_per_gallon = "mpg", horsepower = "hp")$
  collect()
}
\keyword{LazyFrame}
