% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_rename}
\alias{LazyFrame_rename}
\title{Rename columns of a LazyFrame}
\usage{
LazyFrame_rename(mapping)
}
\arguments{
\item{mapping}{A named list like \code{list(new = "old")}.}
}
\value{
LazyFrame
}
\description{
Rename columns of a LazyFrame
}
\examples{
pl$DataFrame(mtcars)$
  lazy()$
  rename(list(miles_per_gallon = "mpg", horsepower = "hp"))
}
\keyword{LazyFrame}
