% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_rename}
\alias{LazyFrame_rename}
\title{Rename columns of a LazyFrame}
\usage{
LazyFrame_rename(existing, new)
}
\arguments{
\item{existing}{A character vector of names to replace.}

\item{new}{A character vector of names used as replacement.}
}
\value{
LazyFrame
}
\description{
Rename columns of a LazyFrame
}
\examples{
pl$DataFrame(mtcars)$lazy()$rename(c("mpg", "hp"), c("miles_per_gallon", "horsepower"))
}
\keyword{LazyFrame}
