% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\docType{data}
\name{LazyFrame_reverse}
\alias{LazyFrame_reverse}
\title{Reverse}
\format{
An object of class \code{character} of length 1.
}
\usage{
LazyFrame_reverse
}
\value{
LazyFrame
}
\description{
Reverse the DataFrame.
}
\examples{
pl$DataFrame(mtcars)$lazy()$reverse()$collect()
}
\keyword{LazyFrame}
