% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_reverse}
\alias{LazyFrame_reverse}
\title{Reverse}
\usage{
LazyFrame_reverse()
}
\value{
LazyFrame
}
\description{
Reverse the DataFrame.
}
\examples{
pl$DataFrame(mtcars)$reverse()
}
\keyword{LazyFrame}
