% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_select}
\alias{LazyFrame_select}
\title{Select and modify columns of a LazyFrame}
\usage{
LazyFrame_select(...)
}
\arguments{
\item{...}{Columns to keep. Those can be expressions (e.g \code{pl$col("a")}),
column names  (e.g \code{"a"}), or list containing expressions or column names
(e.g \code{list(pl$col("a"))}).}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
Related to \code{dplyr::mutate()}. However, it discards unmentioned
columns (like \code{.()} in \code{data.table}).
}
\keyword{LazyFrame}
