% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_shift_and_fill}
\alias{LazyFrame_shift_and_fill}
\title{Shift a LazyFrame}
\usage{
LazyFrame_shift_and_fill(fill_value, periods = 1)
}
\arguments{
\item{periods}{Number of periods to shift (can be negative).}
}
\value{
LazyFrame
}
\description{
Shift the values by a given period. If the period (\code{n}) is positive,
then \code{n} rows will be inserted at the top of the LazyFrame and the last \code{n}
rows will be discarded. Vice-versa if the period is negative. In the end,
the total number of rows of the LazyFrame doesn't change.
}
\examples{
pl$LazyFrame(mtcars)$shift_and_fill(0., 2.)$collect()$as_data_frame()
}
