% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_shift_and_fill}
\alias{LazyFrame_shift_and_fill}
\title{Shift and fill}
\usage{
LazyFrame_shift_and_fill(fill_value, periods = 1)
}
\arguments{
\item{fill_value}{fill None values with the result of this expression.}

\item{periods}{integer Number of periods to shift (may be negative).}
}
\value{
LazyFrame
}
\description{
Shift the values by a given period and fill the resulting null values.
}
\examples{
pl$LazyFrame(mtcars)$lazy()$shift_and_fill(0, 2)$collect()
}
\keyword{LazyFrame}
