% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_slice}
\alias{LazyFrame_slice}
\title{Slice}
\usage{
LazyFrame_slice(offset, length = NULL)
}
\arguments{
\item{offset}{integer}

\item{length}{integer or NULL}
}
\value{
DataFrame
}
\description{
Get a slice of this DataFrame.
}
\examples{
pl$DataFrame(mtcars)$lazy()$slice(2, 4)$collect()
pl$DataFrame(mtcars)$lazy()$slice(30)$collect()
mtcars[2:6,]
}
\keyword{DataFrame}
