% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_sort}
\alias{LazyFrame_sort}
\title{Sort a LazyFrame}
\usage{
LazyFrame_sort(
  by,
  ...,
  descending = FALSE,
  nulls_last = FALSE,
  maintain_order = FALSE
)
}
\arguments{
\item{by}{Column(s) to sort by. Can be character vector of column names,
a list of Expr(s) or a list with a mix of Expr(s) and column names.}

\item{...}{More columns to sort by as above but provided one Expr per argument.}

\item{descending}{Boolean. Sort in descending order (default is \code{FALSE}). This must be
either of length 1 or a logical vector of the same length as the number of
Expr(s) specified in \code{by} and \code{...}.}

\item{nulls_last}{Boolean. Place \code{NULL}s at the end? Default is \code{FALSE}.}

\item{maintain_order}{Keep the same group order as the original \code{LazyFrame}.
Within each group, the order of rows is always preserved, regardless of this
argument. Setting this to \code{TRUE} makes it more expensive to compute and
blocks the possibility to run on the streaming engine. The default value can
be changed with \code{options(polars.maintain_order = TRUE)}.}
}
\value{
LazyFrame
}
\description{
Sort by one or more Expressions.
}
\examples{
df = mtcars
df$mpg[1] = NA
df = pl$LazyFrame(df)
df$sort("mpg")$collect()
df$sort("mpg", nulls_last = TRUE)$collect()
df$sort("cyl", "mpg")$collect()
df$sort(c("cyl", "mpg"))$collect()
df$sort(c("cyl", "mpg"), descending = TRUE)$collect()
df$sort(c("cyl", "mpg"), descending = c(TRUE, FALSE))$collect()
df$sort(pl$col("cyl"), pl$col("mpg"))$collect()
}
\keyword{LazyFrame}
