% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_sum}
\alias{LazyFrame_sum}
\title{Sum}
\usage{
LazyFrame_sum()
}
\value{
A LazyFrame with one row
}
\description{
Aggregate the columns of this LazyFrame to their sum values.
}
\examples{
pl$LazyFrame(mtcars)$sum()$collect()
}
\keyword{LazyFrame}
