% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_tail}
\alias{LazyFrame_tail}
\title{Get the last \code{n} rows.}
\usage{
LazyFrame_tail(n = 5L)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
Get the last \code{n} rows.
}
\examples{
lf = pl$LazyFrame(a = 1:6, b = 7:12)

lf$tail()$collect()

lf$tail(2)$collect()
}
\seealso{
\code{\link[=LazyFrame_head]{<LazyFrame>$head()}}
}
