% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_tail}
\alias{LazyFrame_tail}
\title{Get the last \emph{n} rows}
\usage{
LazyFrame_tail(n)
}
\arguments{
\item{n}{Positive numeric or integer number not larger than 2^32. Any number
will be converted to \code{u32}. Negative raises error.}
}
\value{
A new \code{LazyFrame} object with applied filter.
}
\description{
Get the last \emph{n} rows.
}
\examples{
pl$LazyFrame(mtcars)$tail(2)$collect()
}
