% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_unique}
\alias{LazyFrame_unique}
\title{Drop duplicated rows}
\usage{
LazyFrame_unique(subset = NULL, ..., keep = "any", maintain_order = FALSE)
}
\arguments{
\item{subset}{A character vector with the names of the column(s) to use to
identify duplicates. If \code{NULL} (default), use all columns.}

\item{...}{Not used.}

\item{keep}{Which of the duplicate rows to keep:
\itemize{
\item \code{"any"} (default): Does not give any guarantee of which row is kept. This
allows more optimizations.
\item \code{"first"}: Keep first unique row.
\item \code{"last"}: Keep last unique row.
\item \code{"none"}: Don’t keep duplicate rows.
}}

\item{maintain_order}{Keep the same order as the original data. Setting this
to \code{TRUE} makes it more expensive to compute and blocks the possibility to
run on the streaming engine.}
}
\value{
LazyFrame
}
\description{
Drop duplicated rows
}
\examples{
df = pl$LazyFrame(
  x = sample(10, 100, rep = TRUE),
  y = sample(10, 100, rep = TRUE)
)
df$collect()$height

df$unique()$collect()$height
df$unique(subset = "x")$collect()$height

df$unique(keep = "last")

# only keep unique rows
df$unique(keep = "none")
}
