% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_unique}
\alias{LazyFrame_unique}
\title{Lazy_unique}
\usage{
LazyFrame_unique(subset = NULL, keep = "first")
}
\arguments{
\item{subset}{string or vector of strings. Column name(s) to consider when identifying duplicates. If set to NULL (default), use all columns.}

\item{keep}{string. Which of the duplicate rows to keep:
\itemize{
\item "first": Keep first unique row.
\item "last": Keep last unique row.
\item "none": Don’t keep duplicate rows.
}}
}
\value{
LazyFrame
}
\description{
Drop duplicate rows from this dataframe.
}
\examples{
df = pl$DataFrame(
  x = as.numeric(c(1, 1:5)),
  y = as.numeric(c(1, 1:5)),
  z = as.numeric(c(1, 1, 1:4)))
df$lazy()$unique()$collect()$height
df$lazy()$unique(subset = c("x", "z"), keep = "last")$collect()$height
}
\keyword{LazyFrame}
