% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_unnest}
\alias{LazyFrame_unnest}
\title{Unnest the Struct columns of a LazyFrame}
\usage{
LazyFrame_unnest(names = NULL)
}
\arguments{
\item{names}{Names of the struct columns to unnest. If \code{NULL} (default), then
all "struct" columns are unnested.}
}
\value{
A LazyFrame where all "struct" columns are unnested. Non-struct
columns are not modified.
}
\description{
Unnest the Struct columns of a LazyFrame
}
\examples{
lf = pl$LazyFrame(
  a = 1:5,
  b = c("one", "two", "three", "four", "five"),
  c = 6:10
)$
  select(
  pl$col("b")$to_struct(),
  pl$col("a", "c")$to_struct()$alias("a_and_c")
)
lf$collect()

# by default, all struct columns are unnested
lf$unnest()$collect()

# we can specify specific columns to unnest
lf$unnest("a_and_c")$collect()
}
