% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{LazyFrame_width}
\alias{LazyFrame_width}
\title{Number of columns of a LazyFrame}
\usage{
<LazyFrame>$width()
}
\value{
The number of columns of a DataFrame
}
\description{
Get the number of columns (width) of a LazyFrame
}
\examples{
pl$LazyFrame(mtcars)$width

}
\keyword{LazyFrame}
