% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyFrame_with_columns}
\alias{LazyFrame_with_columns}
\title{Lazy with columns}
\usage{
LazyFrame_with_columns(...)
}
\arguments{
\item{...}{any single Expr or string naming a column}
}
\value{
A new \code{LazyFrame} object with added/modified columns.
}
\description{
add or replace columns of LazyFrame
}
\keyword{LazyFrame}
