% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-group_by.R
\name{lazygroupby__having}
\alias{lazygroupby__having}
\title{Filter groups with a list of predicates after aggregation}
\usage{
lazygroupby__having(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Expression that evaluates to
a boolean Series.}
}
\value{
A lazy groupby
}
\description{
Using this method is equivalent to adding the predicates to the aggregation
and filtering afterwards.

This method can be chained and all conditions will be combined using \code{&}.
}
\examples{
lf <- pl$LazyFrame(x = c("a", "b", "a", "b", "c"))

# Only keep groups that contain more than one element:
lf$group_by("x")$having(
  pl$len() > 1
)$agg()$collect()
}
