% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-group_by.R
\name{lazygroupby__len}
\alias{lazygroupby__len}
\title{Return the number of rows in each group}
\usage{
lazygroupby__len(name = NULL)
}
\arguments{
\item{name}{Assign a name to the resulting column. If \code{NULL}, defaults to
\code{"len"}.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Return the number of rows in each group
}
\examples{
lf <- pl$LazyFrame(
  a = c("Apple", "Apple", "Orange"),
  b = c(1, NA, 2)
)
lf$group_by("a")$len()$collect()

lf$group_by("a")$len("n")$collect()
}
