% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-group_by.R
\name{lazygroupby__mean}
\alias{lazygroupby__mean}
\title{Return the mean per group}
\usage{
lazygroupby__mean()
}
\value{
A polars \link{LazyFrame}
}
\description{
Return the mean per group
}
\examples{
lf <- pl$LazyFrame(
  grp = c("c", "c", "a", "c", "a", "b"),
  x = c(0.5, 0.5, 4, 10, 13, 14),
  y = 1:6,
  z = c(TRUE, TRUE, FALSE, TRUE, FALSE, TRUE)
)
lf$collect()

lf$group_by("grp")$mean()$collect()
}
