% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe-group_by.R
\name{lazygroupby__tail}
\alias{lazygroupby__tail}
\title{Get the last \code{n} rows of each group}
\usage{
lazygroupby__tail(n = 5)
}
\arguments{
\item{n}{Number of rows to return.}
}
\value{
A polars \link{LazyFrame}
}
\description{
Get the last \code{n} rows of each group
}
\examples{
lf <- pl$LazyFrame(
  letters = c("c", "c", "a", "c", "a", "b"),
  nrs = 1:6
)
lf$collect()

lf$group_by("letters")$tail(2)$sort("letters")$collect()
}
