% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{LazyGroupBy_apply}
\alias{LazyGroupBy_apply}
\title{LazyGroupBy_apply}
\usage{
LazyGroupBy_apply(f)
}
\arguments{
\item{f}{lambda function to apply}
}
\value{
A new \code{lazy_polar_frame} object.
}
\description{
one day this will apply
}
