% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__group_by.R
\name{LazyGroupBy_apply}
\alias{LazyGroupBy_apply}
\title{LazyGroupBy_apply}
\usage{
LazyGroupBy_apply(f)
}
\arguments{
\item{f}{R function to apply}
}
\value{
A new \code{LazyFrame} object.
}
\description{
one day this will apply
}
