% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__group_by.R
\name{LazyGroupBy_class}
\alias{LazyGroupBy_class}
\title{Operations on Polars grouped LazyFrame}
\value{
not applicable
}
\description{
Operations on Polars grouped LazyFrame
}
\details{
The LazyGroupBy class in R, is just another interface on top of the
LazyFrame (R wrapper class) in rust polars.
}
