% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__group_by.R
\name{LazyGroupBy_tail}
\alias{LazyGroupBy_tail}
\title{LazyGroupBy_tail}
\usage{
Lazy<GroupBy>$tail(n = 1L)
}
\arguments{
\item{n}{integer number of rows to get}
}
\value{
A new \code{LazyFrame} object.
}
\description{
get n tail rows of group
}
