% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{method_as_property}
\alias{method_as_property}
\title{Give a class method property behavior}
\usage{
method_as_property(f, setter = FALSE)
}
\arguments{
\item{f}{a function}

\item{setter}{bool, if true a property method can be modified by user}
}
\value{
function subclassed into c("property","function") or c("setter","property","function")
}
\description{
Internal function, see use in source
}
