% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-nanoarrow.R
\name{nanoarrow}
\alias{nanoarrow}
\alias{as_nanoarrow_array_stream.DataFrame}
\alias{array_stream}
\alias{arrow}
\alias{record_batch_reader}
\alias{arrow_table}
\alias{read}
\alias{more}
\alias{at}
\alias{\link{https://github.com/apache/arrow-nanoarrow/r}}
\alias{infer_nanoarrow_schema.DataFrame}
\alias{as_record_batch_reader.DataFrame}
\alias{as_arrow_table.DataFrame}
\title{rpolars to nanoarrow and arrow}
\usage{
as_nanoarrow_array_stream(x, ..., schema = NULL)

infer_nanoarrow_schema(x, ...)

as_record_batch_reader(x, ..., schema = NULL)

as_arrow_table(x, ..., schema = NULL)
}
\arguments{
\item{x}{a polars DataFrame}

\item{...}{not used right now}

\item{schema}{must stay at default value NULL}
}
\value{
a nanoarrow array stream

a nanoarrow array schema

an arrow record batch reader

an arrow table
}
\description{
Conversion via native apache arrow array stream (fast), THIS REQUIRES ´nanoarrow´
}
\details{
The following functions enable conversion to \code{nanoarrow} and \code{arrow}.
Conversion kindly provided by "paleolimbot / Dewey Dunnington" Author of \code{nanoarrow}.
Currently these conversions are the fastest way to convert from polars to R.

These conversion function are not exported/public are only accessible via \code{nanoarrow}
e.g. \cr\cr
\code{df = pl$DataFrame(iris)}\cr
\code{nanoarrow::nanoarrow_allocate_array_stream(df)}\cr\cr
or\cr\cr
\code{library(nanoarrow)}\cr
\code{nanoarrow_allocate_array_stream(df)}\cr
}
\examples{
library(nanoarrow)
library(rpolars)
df = pl$DataFrame(iris)
nanoarrow_array_stream = as_nanoarrow_array_stream(df)
rdf = as.data.frame(nanoarrow_array_stream)
nanoarrow_array_schema = infer_nanoarrow_schema(df)
arrow_record_batch_reader = as_record_batch_reader(df)
arrow_table = as_arrow_table(df)
}
\keyword{nanoarrow_interface}
