% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{pl$from_arrow}
\alias{pl$from_arrow}
\alias{to_struct}
\title{to_struct and unnest again}
\arguments{
\item{name}{name of new Series}
}
\value{
@to_struct() returns a Series
}
\description{
to_struct and unnest again
}
\examples{
#none so for
}
\keyword{conversion}
