% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pl}
\alias{pl}
\title{Polars top-level function namespace}
\format{
An object of class \code{polars_object} of length 45.
}
\usage{
pl
}
\description{
\code{pl} is an \link[=environment-class]{environment class} object
that stores all the top-level functions of the R Polars API
which mimics the Python Polars API.
It is intended to work the same way in Python as if you had imported
Python Polars with \verb{import polars as pl}.
}
\examples{
pl

# How many members are in the `pl` environment?
length(pl)

# Create a polars DataFrame
# In Python:
# ```python
# >>> import polars as pl
# >>> df = pl.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
# ```
# In R:
df <- pl$DataFrame(a = c(1, 2, 3), b = c(4, 5, 6))
df
}
\keyword{datasets}
