% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pl}
\alias{pl}
\title{The complete minipolars public API.}
\format{
An object of class \code{environment} of length 9.
}
\usage{
pl$col("colname")$sum() / pl$lit()  #expression ~ chain-method / literal-expression
}
\description{
\code{pl}-object is a list of all public functions and class constructors
public functions are not exported as a normal package as it would be huge namespace
collision with base:: and other functions.

Having all functions in an namespace is similar to the rust- and python- polars api.
Speaking of namespace this pl can be converted into and actual namespace by calling
import_polars_as_("pl"), but this not recemmended.
}
\keyword{datasets}
