% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\docType{data}
\name{pl}
\alias{pl}
\title{The complete rpolars public API.}
\format{
An object of class \code{environment} of length 45.
}
\usage{
pl
}
\description{
\code{pl}-object is a environment of all public functions and class constructors.
Public functions are not exported as a normal package as it would be huge namespace
collision with base:: and other functions. All object-methods are accessed with object$method()
via the new class functions.

Having all functions in an namespace is similar to the rust- and python- polars api.
}
\details{
If someone do not particularly like the letter combination \code{pl}, they are free to
bind the environment to another variable name as \code{simon_says = pl} or even do \code{attach(pl)}
}
\examples{
#how to use polars via `pl`
pl$col("colname")$sum() / pl$lit(42L)  #expression ~ chain-method / literal-expression

#pl inventory
rpolars:::print_env(pl,"rpolars public functions")

#all accessible classes and their public methods
rpolars:::print_env(rpolars:::pl_pub_class_env,"rpolars public class methods, access via object$method()")
}
\keyword{api}
