% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__all}
\alias{pl__all}
\title{Either return an expression representing all columns, or evaluate a bitwise
AND operation}
\usage{
pl__all(..., names, ignore_nulls = TRUE)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Deprecated compatibility
interface for passing multiple names or data types.}

\item{names}{The name(s) or \link[=DataType]{data type} of the column(s) to use in
the aggregation. A character vector or a Polars data type/list of data
types can be supplied.}

\item{ignore_nulls}{If \code{TRUE} (default), ignore null values. If \code{FALSE},
\href{https://en.wikipedia.org/wiki/Three-valued_logic}{Kleene logic} is used to
deal with nulls: if the column contains any null values and no \code{FALSE} values,
the output is null.}
}
\value{
A polars \link{expression}
}
\description{
If no arguments are passed, this function is syntactic sugar for \code{col("*")}.
Otherwise, this function is syntactic sugar for \code{col(names)$all()}.
}
\examples{
df <- pl$DataFrame(
  a = c(TRUE, FALSE, TRUE),
  b = c(FALSE, FALSE, FALSE)
)

# Selecting all columns
df$select(pl$all()$sum())

# Evaluate bitwise AND for a column.
df$select(pl$all(names = "a"))
}
