% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-aggregation-horizontal.R
\name{pl__all_horizontal}
\alias{pl__all_horizontal}
\title{Apply the AND logical horizontally across columns}
\usage{
pl__all_horizontal(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Columns to aggregate
horizontally. Accepts expressions. Strings are parsed as column
names, other non-expression inputs are parsed as literals.}
}
\value{
A polars \link{expression}
}
\description{
Apply the AND logical horizontally across columns
}
\details{
\href{https://en.wikipedia.org/wiki/Three-valued_logic}{Kleene logic} is used to
deal with nulls: if the column contains any null values and no \code{FALSE}
values, the output is null.
}
\examples{
df <- pl$DataFrame(
  a = c(FALSE, FALSE, TRUE, TRUE, FALSE, NA),
  b = c(FALSE, TRUE, TRUE, NA, NA, NA),
  c = c("u", "v", "w", "x", "y", "z")
)

df$with_columns(
  all = pl$all_horizontal("a", "b")
)
}
