% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-col.R
\name{pl__col}
\alias{pl__col}
\title{Create an expression representing column(s) in a DataFrame}
\usage{
pl__col(..., names)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Deprecated compatibility
interface for passing multiple names or data types.}

\item{names}{The name(s) or \link[=DataType]{data type} of the column(s) to
represent. It accepts one of the following:
\itemize{
\item A character vector representing column names
\itemize{
\item Regular expressions starting with \code{^} and ending with \code{$} are allowed.
\item Single wildcard \code{"*"}  has a special meaning: check the examples.
}
\item A \link[=DataType]{Polars DataType} or list of Polars data types
}}
}
\value{
A polars \link{expression}
}
\description{
Create an expression representing column(s) in a DataFrame
}
\examples{
# a single column by a character
pl$col("foo")

# multiple columns by characters
pl$col(c("foo", "bar"))

# multiple columns by polars data types
pl$col(c(pl$Float64, pl$String))

# Single `"*"` is converted to a wildcard expression
pl$col("*")

# The old dynamic-dots interface is deprecated:
pl$col("foo", "bar")

# there are some special notations for selecting columns
df <- pl$DataFrame(foo = 1:3, bar = 4:6, baz = 7:9)

## select all columns with a wildcard `"*"`
df$select(pl$col("*"))

## select multiple columns by a regular expression
## starts with `^` and ends with `$`
df$select(pl$col("^ba.*$"))
}
