% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__cum_sum}
\alias{pl__cum_sum}
\title{Cumulatively sum all values}
\usage{
pl__cum_sum(..., names)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Deprecated compatibility
interface for passing multiple names or data types.}

\item{names}{The name(s) or \link[=DataType]{data type} of the column(s) to use in
the aggregation. A character vector or a Polars data type/list of data
types can be supplied.}
}
\value{
A polars \link{expression}
}
\description{
This function is syntactic sugar for \code{col(names)$cum_sum()}.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

# Get the cum_sum of a column
df$select(pl$cum_sum(names = "a"))

# Get the cum_sum of multiple columns
df$select(pl$cum_sum(names = c("a", "b")))
}
