% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__cum_sum}
\alias{pl__cum_sum}
\title{Cumulatively sum all values}
\usage{
pl__cum_sum(...)
}
\arguments{
\item{...}{Name(s) of the columns to use in the aggregation.}
}
\value{
A polars \link{expression}
}
\description{
This function is syntactic sugar for \code{col(names)$cum_sum()}.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

# Get the cum_sum of a column
df$select(pl$cum_sum("a"))

# Get the cum_sum of multiple columns
df$select(pl$cum_sum("a", "b"))
}
