% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-datatype.R
\name{pl__dtype_of}
\alias{pl__dtype_of}
\title{Get a lazily evaluated DataType of a column or expression}
\usage{
pl__dtype_of(col_or_expr)
}
\description{
Get a lazily evaluated DataType of a column or expression
}
