% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-datatype.R
\name{pl__dtype_of}
\alias{pl__dtype_of}
\title{Get a lazily evaluated DataType of a column or expression}
\usage{
pl__dtype_of(col_or_expr)
}
\arguments{
\item{col_or_expr}{Either a string for the column_name or an expression.}
}
\value{
A polars datatype expression.
This is not the same as a \link[=polars_expr]{polars expression}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get a lazily evaluated DataType of a column or expression
}
