% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-as_datatype.R
\name{pl__duration}
\alias{pl__duration}
\title{Create polars Duration from distinct time components}
\usage{
pl__duration(
  ...,
  weeks = NULL,
  days = NULL,
  hours = NULL,
  minutes = NULL,
  seconds = NULL,
  milliseconds = NULL,
  microseconds = NULL,
  nanoseconds = NULL,
  time_unit = NULL
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{weeks}{Something can be coerced to an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of weeks, or \code{NULL} (default).}

\item{days}{Something can be coerced to an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of days, or \code{NULL} (default).}

\item{hours}{Something can be coerced to an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of hours, or \code{NULL} (default).}

\item{minutes}{Something can be coerced to an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of minutes, or \code{NULL} (default).}

\item{seconds}{Something can be coerced to an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of seconds, or \code{NULL} (default).}

\item{milliseconds}{Something can be coerced to an \link[=Expr]{polars expression}
by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of milliseconds, or \code{NULL} (default).}

\item{microseconds}{Something can be coerced to an \link[=Expr]{polars expression}
by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of microseconds, or \code{NULL} (default).}

\item{nanoseconds}{Something can be coerced to an \link[=Expr]{polars expression} by \code{\link[=as_polars_expr]{as_polars_expr()}}
which represents a column or literal number of nanoseconds, or \code{NULL} (default).}

\item{time_unit}{One of \code{NULL}, \code{"us"} (microseconds),
\code{"ns"} (nanoseconds) or \code{"ms"}(milliseconds). Representing the unit of time.
If \code{NULL} (default), the time unit will be inferred from the other inputs:
\code{"ns"} if \code{nanoseconds} was specified, \code{"us"} otherwise.}
}
\value{
A polars \link{expression}
}
\description{
A \link[=pl__Duration]{Duration} represents a fixed amount of time. For example,
\code{pl$duration(days = 1)} means "exactly 24 hours". By contrast,
\code{\link[=expr_dt_offset_by]{<expr>$dt$offset_by("1d")}} means "1 calendar day", which could sometimes be
23 hours or 25 hours depending on Daylight Savings Time.
For non-fixed durations such as "calendar month" or "calendar day",
please use \code{\link[=expr_dt_offset_by]{<expr>$dt$offset_by()}} instead.
}
\examples{
df <- pl$DataFrame(
  dt = as.POSIXct(c("2022-01-01", "2022-01-02")),
  add = c(1, 2)
)
df

df$select(
  add_weeks = pl$col("dt") + pl$duration(weeks = pl$col("add")),
  add_days = pl$col("dt") + pl$duration(days = pl$col("add")),
  add_seconds = pl$col("dt") + pl$duration(seconds = pl$col("add")),
  add_millis = pl$col("dt") + pl$duration(milliseconds = pl$col("add")),
  add_hours = pl$col("dt") + pl$duration(hours = pl$col("add"))
)
}
