% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-lazy.R
\name{pl__element}
\alias{pl__element}
\title{Alias for an element being evaluated in an eval expression}
\usage{
pl__element()
}
\value{
A polars \link{expression}
}
\description{
Alias for an element being evaluated in an eval expression
}
\examples{
# A horizontal rank computation by taking the elements of a list:
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2)
)
df$with_columns(
  rank = pl$concat_list(c("a", "b"))$list$eval(pl$element()$rank())
)

# A mathematical operation on array elements:
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2)
)
df$with_columns(
  a_b_doubled = pl$concat_list(c("a", "b"))$list$eval(pl$element() * 2)
)
}
