% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-lazy.R
\name{pl__explain_all}
\alias{pl__explain_all}
\title{Explain multiple LazyFrames as if passed to \code{pl$collect_all()}}
\usage{
pl__explain_all(lazy_frames, ..., optimizations = pl$QueryOptFlags())
}
\arguments{
\item{lazy_frames}{A list of LazyFrames to collect.}

\item{...}{These dots are for future extensions and must be empty.}

\item{optimizations}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A \link{QueryOptFlags} object to indicate optimization passes done during query optimization.}
}
\value{
A character value containing the query plan.
}
\description{
Common Subplan Elimination is applied on the combined plan, meaning that
diverging queries will run only once.
}
\examples{
lf <- as_polars_lf(mtcars)$with_columns(sqrt_mpg = pl$col("mpg")$sqrt())

cyl_4 <- lf$filter(pl$col("cyl") == 4)
cyl_6 <- lf$filter(pl$col("cyl") == 6)

# Note that `sqrt_mpg` is present in the plans for `cyl_4` and `cyl_6` but
# only once in the plan produced by `pl$explain_all()`. This is because
# `pl$explain_all()` eliminates parts of the queries that are shared across
# multiple plans.
pl$explain_all(list(cyl_4, cyl_6)) |>
  writeLines()
}
