% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-range-int_range.R
\name{pl__int_range}
\alias{pl__int_range}
\title{Generate a range of integers}
\usage{
pl__int_range(start = 0, end = NULL, step = 1, ..., dtype = pl$Int64)
}
\arguments{
\item{start}{Start of the range (inclusive). Defaults to 0.}

\item{end}{End of the range (exclusive). If \code{NULL} (default), the value
of \code{start} is used and \code{start} is set to 0.}

\item{step}{Step size of the range.}

\item{...}{These dots are for future extensions and must be empty.}

\item{dtype}{Data type of the range.}
}
\value{
A polars \link{expression}
}
\description{
Generate a range of integers
}
\examples{
pl$select(int = pl$int_range(0, 3))

# end can be omitted for a shorter syntax.
pl$select(int = pl$int_range(3))

# Generate an index column by using int_range in conjunction with len().
df <- pl$DataFrame(a = c(1, 3, 5), b = c(2, 4, 6))
df$select(
  index = pl$int_range(pl$len(), dtype = pl$UInt32),
  pl$all()
)
}
