% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-col.R
\name{pl__last}
\alias{pl__last}
\title{Get the last column of the context}
\usage{
pl__last()
}
\value{
A polars \link{expression}
}
\description{
Get the last column of the context
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "baz")
)

df$select(pl$last())
}
