% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-len.R
\name{pl__len}
\alias{pl__len}
\title{Return the number of rows in the context$}
\usage{
pl__len()
}
\value{
A polars \link{expression}
}
\description{
This is similar to \verb{COUNT(*)} in SQL.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 2, NA),
  b = c(3, NA, NA),
  c = c("foo", "bar", "foo"),
)
df$select(pl$len())

# Generate an index column by using len in conjunction with $int_range()
df$with_columns(
  pl$int_range(pl$len(), dtype = pl$UInt32)$alias("index")
)
}
