% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-range-linear_space.R
\name{pl__linear_space}
\alias{pl__linear_space}
\title{Create sequence of evenly-spaced points}
\usage{
pl__linear_space(
  start,
  end,
  num_samples,
  ...,
  closed = c("both", "left", "none", "right")
)
}
\arguments{
\item{start}{Lower bound of the range.}

\item{end}{Upper bound of the range.}

\item{num_samples}{Number of samples in the output sequence.}

\item{...}{These dots are for future extensions and must be empty.}

\item{closed}{Define which sides of the range are closed (inclusive).
One of the following: \code{"both"} (default), \code{"left"}, \code{"right"}, \code{"none"}.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{linear_space} works with numeric and temporal dtypes. When the \code{start} and
\code{end} parameters are \code{Date} dtypes, the output sequence consists of
equally-spaced \code{Datetime} elements with millisecond precision.
}
\examples{
pl$select(
  pl$linear_space(start = 0, end = 1, num_samples = 3)
)
pl$select(
  pl$linear_space(start = 0, end = 1, num_samples = 3, closed = "left")
)
pl$select(
  pl$linear_space(start = 0, end = 1, num_samples = 3, closed = "right")
)
pl$select(
  pl$linear_space(start = 0, end = 1, num_samples = 3, closed = "none")
)

# Date endpoints generate a sequence of Datetime values:
pl$select(
  pl$linear_space(
    start = as.Date("2025-01-01"),
    end = as.Date("2025-02-01"),
    num_samples = 3,
    closed = "right"
  )
)

# You can generate a sequence using the length of the dataframe:
df <- pl$DataFrame(a = c(1, 2, 3, 4, 5))
df$with_columns(ls = pl$linear_space(0, 1, pl$len()))
}
