% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-range-linear_space.R
\name{pl__linear_spaces}
\alias{pl__linear_spaces}
\title{Create sequence of evenly-spaced points for each row between \code{start} and
\code{end}}
\usage{
pl__linear_spaces(
  start,
  end,
  num_samples,
  ...,
  closed = c("both", "left", "none", "right"),
  as_array = FALSE
)
}
\arguments{
\item{start}{Lower bound of the range.}

\item{end}{Upper bound of the range.}

\item{num_samples}{Number of samples in the output sequence.}

\item{...}{These dots are for future extensions and must be empty.}

\item{closed}{Define which sides of the range are closed (inclusive).
One of the following: \code{"both"} (default), \code{"left"}, \code{"right"}, \code{"none"}.}

\item{as_array}{Return result as a fixed-length \code{Array}. \code{num_samples} must
be a constant.}
}
\value{
A polars \link{expression}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The number of values in each sequence is determined by \code{num_samples}.
}
\details{
\code{linear_space} works with numeric and temporal dtypes. When the \code{start} and
\code{end} parameters are \code{Date} dtypes, the output sequence consists of
equally-spaced \code{Datetime} elements with millisecond precision.
}
\examples{
df <- pl$DataFrame(start = c(1, -1), end = c(3, 2), num_samples = c(4, 5))
df$with_columns(ls = pl$linear_spaces("start", "end", "num_samples"))

df$with_columns(ls = pl$linear_spaces("start", "end", 3, as_array = TRUE))
}
