% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-as_datatype.R
\name{pl__list}
\alias{pl__list}
\title{Collect columns into a list column}
\usage{
pl__list(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Columns to collect into a
list column. Accepts expression input. Strings are parsed as column names,
other non-expression inputs are parsed as literals.}
}
\value{
A polars \link{expression}
}
\description{
Unlike \code{\link[=pl__concat_list]{pl$concat_list()}}, list-typed inputs are not
extended: the value of each input becomes a single element of the output
list. This means that \code{List(T)} inputs produce a \code{List(List(T))} output.
}
\examples{
# Wrap non-list columns into a list (same as pl$concat_list() in this case).
df <- pl$DataFrame(a = c(1, 2, 3), b = c(4, 5, 6))
df$with_columns(a_b = pl$list("a", "b"))

# Collect list columns into a list of lists. pl$concat_list() would instead
# concatenate them into a single list.
df <- pl$DataFrame(a = list(1:2, 3L, 4:5), b = list(6L, 7:8, 9L))
df$with_columns(
  list = pl$list("a", "b"),
  concat_list = pl$concat_list("a", "b")
)
}
