% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-lit.R
\name{pl__lit}
\alias{pl__lit}
\title{Return an expression representing a literal value}
\usage{
pl__lit(value, dtype = NULL)
}
\arguments{
\item{value}{An R object. Passed as the \code{x} param of \code{\link[=as_polars_expr]{as_polars_expr()}}.}

\item{dtype}{A polars data type or \code{NULL} (default).
If not \code{NULL}, casted to the specified data type.}
}
\value{
A polars \link{expression}
}
\description{
This function is a shorthand for \code{\link[=as_polars_expr]{as_polars_expr(x, as_lit = TRUE)}} and
in most cases, the actual conversion is done by \code{\link[=as_polars_series]{as_polars_series()}}.
}
\section{Literal scalar mapping}{
Since R has no scalar class, each of the following types of length 1 cases is specially
converted to a scalar literal.
\itemize{
\item character: String
\item logical: Boolean
\item integer: Int32
\item double: Float64
}

These types' \code{NA} is converted to a \code{null} literal with casting to the corresponding Polars type.

The \link{raw} type vector is converted to a Binary scalar.
\itemize{
\item raw: Binary
}

\code{NULL} is converted to a Null type \code{null} literal.
\itemize{
\item NULL: Null
}

Other scalar values cannot be created directly, so we need to \verb{$cast()} from other scalar values
or extract them with the \verb{$first()} method from a \link{Series}.
(\code{\link[=as_polars_expr]{as_polars_expr()}} and \code{\link[=pl__lit]{pl$lit()}} automatically apply the \verb{$first()} method to extract scalar values
inside the default S3 method of \code{\link[=as_polars_expr]{as_polars_expr()}}.)
}

\examples{
# Literal scalar values
pl$lit(1L)
pl$lit(5.5)
pl$lit(NULL)
pl$lit("foo_bar")

## Generally, for a vector of length 1, it is converted to a Series and then
## the `$first()` method is called to become a scalar value.
pl$lit(as.Date("2021-01-20"))
pl$lit(as.POSIXct("2023-03-31 10:30:45"))

# Literal Series data
pl$lit(1:3)
pl$lit(pl$Series("x", 1:3))
}
\seealso{
\itemize{
\item \code{\link[=as_polars_series]{as_polars_series()}}: R -> Polars type mapping is mostly defined by this function.
\item \code{\link[=as_polars_expr]{as_polars_expr()}}: Internal implementation of \code{pl$lit()}.
}
}
