% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-lit.R
\name{pl__lit}
\alias{pl__lit}
\title{Return an expression representing a literal value}
\usage{
pl__lit(value, dtype = NULL)
}
\arguments{
\item{value}{An R object. Passed as the \code{x} param of \code{\link[=as_polars_expr]{as_polars_expr()}}.}

\item{dtype}{A polars data type or \code{NULL} (default).
If not \code{NULL}, casted to the specified data type.}
}
\value{
A polars \link{expression}
}
\description{
This function is a shorthand for \code{\link[=as_polars_expr]{as_polars_expr(x, as_lit = TRUE)}} and
in most cases, the actual conversion is done by \code{\link[=as_polars_series]{as_polars_series()}}.
}
\examples{
# Literal scalar values
pl$lit(1L)
pl$lit(5.5)
pl$lit(NULL)
pl$lit("foo_bar")

## Generally, for a vector (an R object) becomes a Series with length 1,
## it is converted to a Series and then get the first value to become a scalar literal.
pl$lit(as.Date("2021-01-20"))
pl$lit(as.POSIXct("2023-03-31 10:30:45"))
pl$lit(data.frame(a = 1, b = "foo"))

# Literal Series data
pl$lit(1:3)
pl$lit(pl$Series("x", 1:3))
}
\seealso{
\itemize{
\item \code{\link[=as_polars_series]{as_polars_series()}}: R -> Polars type mapping is mostly defined by this function.
\item \code{\link[=as_polars_expr]{as_polars_expr()}}: Internal implementation of \code{pl$lit()}.
}
}
