% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__max}
\alias{pl__max}
\title{Get the maximum value}
\usage{
pl__max(..., names)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Deprecated compatibility
interface for passing multiple names or data types.}

\item{names}{The name(s) or \link[=DataType]{data type} of the column(s) to use in
the aggregation. A character vector or a Polars data type/list of data
types can be supplied.}
}
\value{
A polars \link{expression}
}
\description{
This function is syntactic sugar for \code{col(names)$max()}.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

# Get the maximum value of a column
df$select(pl$max(names = "a"))

# Get the maximum value of multiple columns
df$select(pl$max(names = c("a", "b")))
}
