% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__max}
\alias{pl__max}
\title{Get the maximum value}
\usage{
pl__max(...)
}
\arguments{
\item{...}{Name(s) of the columns to use in the aggregation.}
}
\value{
A polars \link{expression}
}
\description{
This function is syntactic sugar for \code{col(names)$max()}.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

# Get the maximum value of a column
df$select(pl$max("a"))

# Get the maximum value of multiple columns
df$select(pl$max("a", "b"))
}
