% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-aggregation-horizontal.R
\name{pl__mean_horizontal}
\alias{pl__mean_horizontal}
\title{Compute the mean horizontally across columns}
\usage{
pl__mean_horizontal(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Columns to aggregate
horizontally. Accepts expressions. Strings are parsed as column
names, other non-expression inputs are parsed as literals.}
}
\value{
A polars \link{expression}
}
\description{
Compute the mean horizontally across columns
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3)
  b = c(4, 5, NA),
  c = c("x", "y", "z")
)

df$with_columns(
  mean = pl$mean_horizontal("a", "b")
)
}
