% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-col.R
\name{pl__nth}
\alias{pl__nth}
\title{Get the nth column(s) of the context}
\usage{
pl__nth(indices, strict = TRUE)
}
\arguments{
\item{indices}{One or more indices representing the columns to retrieve.}

\item{strict}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Passed to
\code{\link[=cs__by_index]{cs$by_index()}}'s \code{require_all} argument.}
}
\value{
A polars \link{expression}
}
\description{
Get the nth column(s) of the context
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "baz")
)

df$select(pl$nth(1))
df$select(pl$nth(c(2, 0)))
}
