% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/input-json-functions.R
\name{pl__scan_ndjson}
\alias{pl__scan_ndjson}
\title{Lazily read from a local or cloud-hosted NDJSON file(s)}
\usage{
pl__scan_ndjson(
  source,
  ...,
  schema = NULL,
  schema_overrides = NULL,
  infer_schema_length = 100,
  batch_size = 1024,
  n_rows = NULL,
  low_memory = FALSE,
  row_index_name = NULL,
  row_index_offset = 0L,
  ignore_errors = FALSE,
  storage_options = NULL,
  include_file_paths = NULL
)
}
\arguments{
\item{source}{Path(s) to a file or directory. When needing to authenticate
for scanning cloud locations, see the \code{storage_options} parameter.}

\item{...}{These dots are for future extensions and must be empty.}

\item{schema}{Provide a list of \link[=DataType]{datatypes} for the fields in the
NDJSON input. Named elements are matched by name.}

\item{schema_overrides}{Overwrite dtypes during schema inference. This must
be a list of \link[=DataType]{datatypes}. Named elements are matched by name.}

\item{infer_schema_length}{The maximum number of rows to scan for schema
inference. This applies individually to each file included according to
\code{infer_schema_files}. If \code{NULL}, the full data may be scanned (this is
slow). Set \code{infer_schema = FALSE} to read all columns as \code{pl$String}.}

\item{batch_size}{Number of rows to read in each batch.}

\item{n_rows}{Stop reading from the source after reading \code{n_rows}.}

\item{low_memory}{Reduce memory pressure at the expense of performance.}

\item{row_index_name}{If not \code{NULL}, this will insert a row index column with
the given name.}

\item{row_index_offset}{Offset to start the row index column (only used if
the name is set by \code{row_index_name}).}

\item{ignore_errors}{Keep reading the file even if some lines yield errors.
You can also use \code{infer_schema = FALSE} to read all columns as UTF8 to
check which values might cause an issue.}

\item{storage_options}{Named vector containing options that indicate how to
connect to a cloud provider. The cloud providers currently supported are
AWS, GCP, and Azure.
See supported keys here:
\itemize{
\item \href{https://docs.rs/object_store/latest/object_store/aws/enum.AmazonS3ConfigKey.html}{aws}
\item \href{https://docs.rs/object_store/latest/object_store/gcp/enum.GoogleConfigKey.html}{gcp}
\item \href{https://docs.rs/object_store/latest/object_store/azure/enum.AzureConfigKey.html}{azure}
\item Hugging Face (\verb{hf://}): Accepts an API key under the token parameter
\code{c(token = YOUR_TOKEN)} or by setting the \code{HF_TOKEN} environment
variable.
}

If \code{storage_options} is not provided, Polars will try to infer the
information from environment variables.}

\item{include_file_paths}{Include the path of the source file(s) as a column
with this name.}
}
\value{
A polars \link{LazyFrame}
}
\description{
This allows the query optimizer to push down predicates and projections to
the scan level, thereby potentially reducing memory overhead.
}
\examples{
\dontshow{if (\{
requireNamespace("jsonlite", quietly = TRUE)
\}) withAutoprint(\{ # examplesIf}
ndjson_filename <- tempfile()
jsonlite::stream_out(iris, file(ndjson_filename), verbose = FALSE)
pl$scan_ndjson(ndjson_filename)$collect()
\dontshow{\}) # examplesIf}
}
