% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-context.R
\name{pl__SQLContext}
\alias{pl__SQLContext}
\title{Initialize a new \code{SQLContext}}
\usage{
pl__SQLContext(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Elements that are known in the
current \code{SQLContext}. It accepts any R object that can be converted to a
LazyFrame via \code{as_polars_lf()}. All elements must be named.}
}
\value{
An object of class \code{"polars_sql_context"}
}
\description{
Initialize a new \code{SQLContext}
}
\examples{
pl$SQLContext(mtcars = mtcars)

pl$SQLContext(mtcars = mtcars, a = data.frame(x = 1))
}
