% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/functions-aggregation-vertical.R
\name{pl__sum}
\alias{pl__sum}
\title{Sum all values}
\usage{
pl__sum(...)
}
\arguments{
\item{...}{Name(s) of the columns to use in the aggregation.}
}
\value{
A polars \link{expression}
}
\description{
This function is syntactic sugar for \code{col(names)$sum()}.
}
\examples{
df <- pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

# Get the sum of a column
df$select(pl$sum("a"))

# Get the sum of multiple columns
df$select(pl$sum("a", "b"))
}
