% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-thread_pool.R
\name{pl__thread_pool_size}
\alias{pl__thread_pool_size}
\title{Return the number of threads in the Polars thread pool}
\usage{
pl__thread_pool_size()
}
\value{
The integer of threads used by polars engine.
}
\description{
Return the number of threads in the Polars thread pool
}
\details{
The threadpool size can be overridden by setting the
\code{POLARS_MAX_THREADS} environment variable before process start.
It cannot be modified once the package is loaded.
It is strongly recommended not to override this value as it will be
set automatically by the engine.
}
\examples{
pl$thread_pool_size()
}
\seealso{
\itemize{
\item \code{\link[=polars_info]{polars_info()}} shows the thread pool size and other information.
}
}
