% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_all}
\alias{pl_all}
\title{New Expr referring to all columns}
\value{
Boolean literal
}
\description{
Not to mix up with \code{Expr_object$all()} which is a 'reduce Boolean columns by AND' method.
}
\details{
last \code{all()} in example is this Expr method, the first \code{pl$all()} refers
to "all-columns" and is an expression constructor
}
\examples{
pl$DataFrame(list(all = c(TRUE, TRUE), some = c(TRUE, FALSE)))$select(pl$all()$all())
}
\keyword{Expr_new}
