% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pl_api_register_series_namespace}
\alias{pl_api_register_series_namespace}
\title{Registering custom functionality with a polars Series}
\usage{
pl_api_register_series_namespace(name, ns_fn)
}
\arguments{
\item{name}{Name under which the functionality will be accessed.}

\item{ns_fn}{A function returns a new \link{environment} with the custom functionality.
See examples for details.}
}
\value{
\code{NULL} invisibly.
}
\description{
Registering custom functionality with a polars Series
}
\examples{
# s: polars series
math_shortcuts <- function(s) {
  # Create a new environment to store the methods
  self <- new.env(parent = emptyenv())

  # Store the series
  self$`_s` <- s

  # Add methods
  self$square <- function() self$`_s` * self$`_s`
  self$cube <- function() self$`_s` * self$`_s` * self$`_s`

  # Set the class
  class(self) <- c("polars_namespace_series", "polars_object")

  # Return the environment
  self
}

pl$api$register_series_namespace("math", math_shortcuts)

s <- as_polars_series(c(1.5, 31, 42, 64.5))
s$math$square()$rename("s^2")

s <- as_polars_series(1:5)
s$math$cube()$rename("s^3")
}
