% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_arg_sort_by}
\alias{pl_arg_sort_by}
\title{Return the row indices that would sort the columns}
\usage{
pl_arg_sort_by(exprs, ..., descending = FALSE)
}
\arguments{
\item{exprs}{Column(s) to arg sort by. Can be Expr(s) or something coercible
to Expr(s). Strings are parsed as column names.}

\item{...}{Not used.}

\item{descending}{Sort in descending order. When sorting by multiple columns,
can be specified per column by passing a vector of booleans.}
}
\value{
Expr
}
\description{
Return the row indices that would sort the columns
}
\examples{
df = pl$DataFrame(
  a = c(0, 1, 1, 0),
  b = c(3, 2, 3, 2)
)

df$with_columns(
  arg_sort_a = pl$arg_sort_by("a"),
  arg_sort_ab = pl$arg_sort_by(c("a", "b"), descending = TRUE)
)
}
\seealso{
\link[=Expr_arg_sort]{$arg_sort()} to find the row indices that would
sort an Expr.
}
