% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_concat_str}
\alias{pl_concat_str}
\title{Horizontally concatenate columns into a single string column}
\usage{
pl_concat_str(..., separator = "", ignore_nulls = FALSE)
}
\arguments{
\item{...}{Columns to concatenate into a single string column. Accepts
expressions. Strings are parsed as column names, other non-expression inputs
are parsed as literals. Non-String columns are cast to String}

\item{separator}{String that will be used to separate the values of each
column.}

\item{ignore_nulls}{If \code{FALSE} (default), null values are propagated: if the
row contains any null values, the output is null.}
}
\value{
Expr
}
\description{
Horizontally concatenate columns into a single string column
}
\examples{
df = pl$DataFrame(
  a = 1:3,
  b = c("dogs", "cats", NA),
  c = c("play", "swim", "walk")
)

df$with_columns(
  pl$concat_str(
    pl$col("a") * 2L, "b", "c", pl$lit("!"),
    separator = " "
  )$alias("full_sentence")
)

df$with_columns(
  pl$concat_str(
    pl$col("a") * 2L, "b", "c", pl$lit("!"),
    separator = " ",
    ignore_nulls = TRUE
  )$alias("full_sentence")
)
}
