% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pl_date_range}
\alias{pl_date_range}
\title{new date_range}
\arguments{
\item{low}{string or}

\item{high}{string or}

\item{interval}{string parsed to duration like "1w" "1week" "2years" "10s"}

\item{...}{not used}

\item{lazy}{bool}

\item{closed}{option one of 'both'(default), 'left', 'none' or 'right'}

\item{name}{name of series}

\item{time_unit}{option "ns" "us" "ms"}

\item{time_zone}{}
}
\value{
a datetime
}
\description{
new date_range
}
\examples{
t_ms = unclass(as.POSIXct("2022-01-01",tz = "GMT",origin="GMT")) * 1000
pl$date_range(t_ms,t_ms+7500, interval = "2s", time_unit = "ms", time_zone = "GMT")
}
