% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pl_date_range}
\alias{pl_date_range}
\title{new date_range}
\arguments{
\item{low}{string or}

\item{high}{string or}

\item{interval}{string parsed to duration like "1w" "1week" "2years" "10s"}

\item{...}{not used}

\item{lazy}{bool}

\item{closed}{option one of 'both'(default), 'left', 'none' or 'right'}

\item{name}{name of series}

\item{time_unit}{option "ns" "us" "ms"}

\item{time_zone}{optional string describing a timezone. If no time_zone is provided the
the tz of low/high input is ignored and the output Series will no have a tz predicate.
If time_zone is provided the low/high timezone will be corrected into that tz. If low/high
does not have a tz(tzone)-attribute, then local System time is assumed.}
}
\value{
a datetime
}
\description{
new date_range
}
\examples{
t_ms = unclass() * 1000
pl$date_range(
  as.POSIXct("2022-01-01"),
  as.POSIXct("2022-01-02",tz = "GMT", origin="GMT"),
  interval = "2h", time_unit = "ms", time_zone = "GMT"
)
}
