% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__eager.R
\name{pl_date_range}
\alias{pl_date_range}
\title{new date_range}
\arguments{
\item{start}{POSIXt or Date preferably with time_zone or double or integer}

\item{end}{POSIXt or Date preferably with time_zone or double or integer. If end is and
interval are missing, then single datetime is constructed.}

\item{interval}{string pl_duration or R difftime. Can be missing if end is missing also.}

\item{eager}{bool, if FALSE (default) return \code{Expr} else evaluate \code{Expr} to \code{Series}}

\item{closed}{option one of 'both'(default), 'left', 'none' or 'right'}

\item{name}{name of series}

\item{time_unit}{option string ("ns" "us" "ms") duration of one int64 value on polars side}

\item{time_zone}{optional string describing a timezone.}

\item{explode}{if TRUE (default) all created ranges will be "unlisted" into on column, if FALSE
output will be a list of ranges.}
}
\value{
a datetime
}
\description{
new date_range
}
\details{
If param time_zone is not defined the Series will have no time zone.

NOTICE: R POSIXt without defined timezones(tzone/tz), so called naive datetimes, are counter
intuitive in R. It is recommended to always set the timezone of start and end. If not output will
vary between local machine timezone, R and polars.

In R/r-polars it is perfectly fine to mix timezones of params time_zone, start and end.
}
\examples{

# All in GMT, straight forward, no mental confusion
s_gmt = pl$date_range(
  as.POSIXct("2022-01-01", tz = "GMT"),
  as.POSIXct("2022-01-02", tz = "GMT"),
  interval = "6h", time_unit = "ms", time_zone = "GMT"
)
s_gmt
s_gmt$to_r() # printed same way in R and polars becuase tagged with a time_zone/tzone

# polars assumes any input in GMT if time_zone = NULL, set GMT on start end to see same print
s_null = pl$date_range(
  as.POSIXct("2022-01-01", tz = "GMT"),
  as.POSIXct("2022-01-02", tz = "GMT"),
  interval = "6h", time_unit = "ms", time_zone = NULL
)
s_null$to_r() # back to R POSIXct. R prints non tzone tagged POSIXct in local timezone.


# use of ISOdate
t1 = ISOdate(2022, 1, 1, 0) # preset GMT
t2 = ISOdate(2022, 1, 2, 0) # preset GMT
pl$date_range(t1, t2, interval = "4h", time_unit = "ms", time_zone = "GMT")$to_r()

}
\keyword{ExprDT}
\keyword{functions}
