% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__eager.R
\name{pl_date_ranges}
\alias{pl_date_ranges}
\title{Create a column of date ranges}
\usage{
pl_date_ranges(
  start,
  end,
  interval = "1d",
  closed = "both",
  time_unit = "us",
  time_zone = NULL,
  eager = FALSE
)
}
\arguments{
\item{start}{Lower bound of the date range. POSIXt or Date preferably with
time_zone or double or integer}

\item{end}{Upper bound of the date range. POSIXt or Date preferably with
time_zone or double or integer. If \code{end} and \code{interval} are missing, then a
single datetime is constructed.}

\item{interval}{String, a Polars \code{duration} or R \code{\link[=difftime]{difftime()}}. Can be missing
if \code{end} is missing also.}

\item{closed}{Define which sides of the range are closed (inclusive). One of
\code{"both"} (default), \code{"left"}, \code{"none"} or \code{"right"}.}

\item{time_unit}{String (\code{"ns"}, \code{"us"}, \code{"ms"}) or integer. Time unit of the
resulting \code{Datetime} data type. Only takes effect if the output column is
of type \code{Datetime}.}

\item{time_zone}{Time zone of the resulting \code{Datetime} data type. Only takes
effect if the output column is of type \code{Datetime.} If \code{NULL} (default),
\verb{"GMT} is used.}

\item{eager}{If \code{FALSE} (default), returns an \code{Expr}. Otherwise, returns a
\code{Series}.}
}
\value{
An Expr (if \code{eager = FALSE}) or Series (if \code{eager = TRUE}) of format
\code{list[date]}.
}
\description{
This generate a list of date values for each pair of lower and upper
bounds. This is particularly suited to create ranges per row in Data/LazyFrame.
Use \code{\link[=pl_date_range]{pl$date_range()}} to generate a Series with date
values for a single lower bound and upper bound.
}
\examples{
df = pl$DataFrame(
  start = c(as.Date("2022-1-1"), as.Date("2022-1-2")),
  end = as.Date("2022-1-3")
)

df$with_columns(date_range = pl$date_ranges("start", "end"))
}
