% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{pl_dtypes}
\alias{pl_dtypes}
\title{DataTypes (RPolarsDataType)}
\value{
not applicable
}
\description{
\code{DataType} any polars type (ported so far)
}
\examples{
print(ls(pl$dtypes))
pl$dtypes$Float64
pl$dtypes$String

pl$List(pl$List(pl$UInt64))

pl$Struct(pl$Field("CityNames", pl$String))

# The function changes type from Int32 to String
# Specifying the output DataType: String solves the problem
pl$Series(1:4)$map_elements(\(x) letters[x], datatype = pl$dtypes$String)

}
