% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pl_enable_string_cache}
\alias{pl_enable_string_cache}
\title{enable disable global string cache}
\arguments{
\item{toggle}{Boolean. TRUE enable, FALSE disable.}
}
\value{
enable_string_cache: no return
}
\description{
e.g. join to Categorical series is only allowed if using global string cache.
Local categorical encodings may differ and is not meaningful to combine as is.
}
\details{
This function will force enable/disable the string_cache and override any contexts made
by \code{pl$with_string_cache}. Prefer to use \code{pl$with_string_cache}.
}
\examples{
pl$enable_string_cache(TRUE)
pl$using_string_cache()
}
\seealso{
\code{\link[=pl_using_string_cache]{pl$using_string_cache}}
\code{\link[=pl_with_string_cache]{pl$with_string_cache}}
}
\keyword{options}
