% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{pl_Field}
\alias{pl_Field}
\alias{RField_name}
\title{Create Field}
\usage{
RField_name()
}
\arguments{
\item{name}{string name}

\item{datatype}{DataType}
}
\value{
a list DataType with an inner DataType

name
}
\description{
get/set Field name
}
\details{
A Field is not a DataType but a name + DataType
Fields are used in Structs-datatypes and Schemas to represent
everything of the Series/Column except the raw values.
}
\examples{
#make a struct
pl$Field("city_names",pl$Utf8)

# find any DataType bundled pl$dtypes
print(pl$dtypes)

field = pl$Field("Cities",pl$Utf8)

#get name / datatype
field$name
field$datatype

#set + get values
field$name = "CityPoPulations" #<- is fine too
field$datatype = pl$UInt32

print(field)
}
\keyword{DataType}
\keyword{pl}
