% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{pl_Field}
\alias{pl_Field}
\alias{RField_class}
\alias{RPolarsRField}
\title{Create Field}
\usage{
pl_Field(name, datatype)
}
\arguments{
\item{name}{Field name}

\item{datatype}{\link[=pl_dtypes]{DataType}}
}
\value{
A object of with the data type \code{"Field"} containing its name and its
\link[=pl_dtypes]{data type}.
}
\description{
A Field is composed of a name and a data type.
Fields are used in Structs-datatypes and Schemas to represent everything of
the Series/Column except the raw values.
}
\section{Active Bindings}{

\subsection{datatype}{

\verb{$datatype} returns the \link[=pl_dtypes]{data type} of the Field.

\verb{$datatype = <RPolarsDataType>} sets the \link[=pl_dtypes]{data type} of the Field.
}

\subsection{name}{

\verb{$name} returns the name of the Field.

\verb{$name = "new_name"} sets the name of the Field.
}
}

\examples{
field = pl$Field("city_names", pl$String)

field
field$datatype
field$name

# Set the new data type
field$datatype = pl$Categorical
field$datatype

# Set the new name
field$name = "CityPoPulations"
field
}
