% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{pl_Field}
\alias{pl_Field}
\title{Create Field}
\arguments{
\item{name}{string name}

\item{datatype}{DataType}
}
\value{
a list DataType with an inner DataType
}
\description{
Create Field
}
\details{
A Field is not a DataType but a name + DataType
Fields are used in Structs-datatypes and Schemas to represent
everything of the Series/Column except the raw values.
}
\examples{
# make a struct
pl$Field("city_names", pl$Utf8)

# find any DataType bundled pl$dtypes
print(pl$dtypes)

}
\keyword{pl}
