% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_first}
\alias{pl_first}
\title{Get the first value.}
\usage{
pl_first(columns = NULL)
}
\arguments{
\item{columns}{A character vector indicating the column names or \code{NULL}.
If \code{NULL} (default), returns an expression to take the first column
of the context instead.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function has different behavior depending on the column argument:
\itemize{
\item \code{NULL} -> Takes first column of a context.
\item A character vector -> Syntactic sugar for \code{pl$col(columns)$first()}.
}
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$first())

df$select(pl$first("b"))

df$select(pl$first(c("a", "c")))
}
\seealso{
\itemize{
\item \code{\link[=Expr_first]{<Expr>$first()}}
}
}
