% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_fold_reduce}
\alias{pl_fold_reduce}
\alias{pl_fold_reduce_part2}
\title{Accumulate over multiple columns horizontally with an R function}
\arguments{
\item{acc}{an Expr or Into\if{html}{\out{<Expr>}} of the initial accumulator.}

\item{lambda}{R Function which takes two polars Series as input and return one.}

\item{exprs}{Expressions to aggregate over. May also be a wildcard expression.}
}
\value{
An expression that will be applied rowwise
}
\description{
\code{pl$fold()} and \code{pl$reduce()} allows one to do rowwise operations. The only
difference between them is that \code{pl$fold()} has an additional argument (\code{acc})
that contains the value that will be initialized when the fold starts.
}
\examples{
df = pl$DataFrame(mtcars)

# Make the row-wise sum of all columns with fold, reduce and vectorized "+"
df$with_columns(
  pl$reduce(
    lambda = \(acc, x) acc + x,
    exprs = pl$col("mpg", "drat")
  )$alias("mpg_drat_sum_reduced"),
  pl$fold(
    acc = pl$lit(0),
    lambda = \(acc, x) acc + x,
    exprs = pl$col("mpg", "drat")
  )$alias("mpg_drat_sum_folded"),
  pl$col("mpg")$alias("mpg_drat_vector_sum") + pl$col("drat")
)
}
