% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_from_epoch}
\alias{pl_from_epoch}
\title{Convert a Unix timestamp to date(time)}
\usage{
pl_from_epoch(column, time_unit = "s")
}
\arguments{
\item{column}{An Expr from which integers will be parsed. If this is a float
column, then the decimal part of the float will be ignored. Character are
parsed as column names, but other literal values must be passed to \code{pl$lit()}.}

\item{time_unit}{One of \code{"ns"}, \code{"us"}, \code{"ms"}, \code{"s"}, \code{"d"}}
}
\value{
Expr as Date or Datetime depending on the \code{time_unit}
}
\description{
Depending on the \code{time_unit} provided, this function will return a different
dtype:
\itemize{
\item \code{time_unit = "d"} returns \code{pl$Date}
\item \code{time_unit = "s"} returns \code{pl$Datetime("us")} (\code{pl$Datetime}’s default)
\item \code{time_unit = "ms"} returns \code{pl$Datetime("ms")}
\item \code{time_unit = "us"} returns \code{pl$Datetime("us")}
\item \code{time_unit = "ns"} returns \code{pl$Datetime("ns")}
}
}
\examples{
# pass an integer column
df = pl$DataFrame(timestamp = 100:150)
df$with_columns(
   timestamp_to_datetime = pl$from_epoch(pl$col("timestamp"), time_unit = "s")
)

df$with_columns(
   timestamp_to_date = pl$from_epoch(pl$col("timestamp"), time_unit = "d")
)

# pass a literal
pl$from_epoch(pl$lit(100:105), time_unit = "s")$to_series()
}
