% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_head}
\alias{pl_head}
\title{Get the first \code{n} rows.}
\usage{
pl_head(..., n = 10)
}
\arguments{
\item{...}{Characters indicating the column names.}

\item{n}{Number of rows to return.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(...)$head(n)}.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$head("a"))

df$select(pl$head("a", "b", n = 2))
}
\seealso{
\itemize{
\item \code{\link[=Expr_head]{<Expr>$head()}}
}
}
