% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_head}
\alias{pl_head}
\title{Get the first \code{n} rows.}
\arguments{
\item{column}{if dtype is:
\itemize{
\item Series: Take head value in \code{Series}
\item str or int: syntactic sugar for \verb{pl.col(..).head()}
}}

\item{n}{Number of rows to take}
}
\value{
Expr or head value of input Series
}
\description{
Get the first \code{n} rows.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

expr_head = pl$head("a")
print(expr_head)
df$select(expr_head)

df$select(pl$head("a",2))
pl$head(df$get_column("a"),2)
}
\keyword{Expr_new}
