% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_implode}
\alias{pl_implode}
\title{Aggregate all column values into a list.}
\usage{
pl_implode(...)
}
\arguments{
\item{...}{Characters indicating the column names, passed to \code{\link[=pl_col]{pl$col()}}.
See \code{\link[=pl_col]{?pl_col}} for details.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(...)$implode()}.
}
\examples{
pl$DataFrame(iris)$select(pl$implode("Species"))
}
