% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_last}
\alias{pl_last}
\title{pl$last}
\arguments{
\item{column}{if dtype is:
\itemize{
\item Series: Take last value in \code{Series}
\item str: syntactic sugar for \verb{pl.col(..).last()}
\item NULL: expression to take last column of a context.
}}
}
\value{
Expr or last value of input Series
}
\description{
Depending on the input type this function does different things:
}
\examples{

df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)
df$select(pl$last())

df$select(pl$last("a"))

pl$last(df$get_column("a"))

}
\keyword{Expr_new}
