% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_len}
\alias{pl_len}
\title{pl$count}
\usage{
pl_len(column = NULL)
}
\arguments{
\item{column}{if dtype is:
\itemize{
\item Series: count length of Series
\item str: count values of this column
\item NULL: count the number of value in this context.
}}
}
\value{
Expr or value-count in case Series
}
\description{
Count the number of values in this column/context.
}
\examples{

df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)
df$select(pl$len())


df$group_by("c", maintain_order = TRUE)$agg(pl$len())
}
\keyword{Expr_new}
