% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{pl_list}
\alias{pl_list}
\title{create list data type}
\arguments{
\item{datatype}{an inner DataType}
}
\value{
a list DataType with an inner DataType
}
\description{
create list data type
}
\examples{
pl$list(pl$list(pl$Boolean))
}
