% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_max}
\alias{pl_max}
\title{Get the maximum value.}
\usage{
pl_max(...)
}
\arguments{
\item{...}{Characters indicating the column names, passed to \code{\link[=pl_col]{pl$col()}}.
See \code{\link[=pl_col]{?pl_col}} for details.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Syntactic sugar for \code{pl$col(...)$max()}.
}
\examples{
df = pl$DataFrame(
  num_1 = c(1, 8, 3),
  num_2 = c(4, 5, 2),
  chr_1 = c("foo", "bar", "foo")
)

df$select(pl$max("num_1"))

# Get the maximum value of multiple columns.
df$select(pl$max(r"(^num_\d+$)"))

df$select(pl$max("num_1", "num_2"))
}
\seealso{
\itemize{
\item \code{\link[=Expr_max]{<Expr>$max()}}
\item \code{\link[=pl_max_horizontal]{pl$max_horizontal()}}
}
}
