% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_max}
\alias{pl_max}
\title{Find maximum value in one or several columns}
\arguments{
\item{...}{is a:
If one arg:
\itemize{
\item Series or Expr, same as \code{column$sum()}
\item string, same as \code{pl$col(column)$sum()}
\item numeric, same as \code{pl$lit(column)$sum()}
\item list of strings(column names) or expressions to add up as expr1 + expr2 + expr3 + ...
}

If several args, then wrapped in a list and handled as above.}

\item{verbose}{Show the deprecation message when several columns or Expr are
passed in \code{...}. Will be removed in 0.12.0.}
}
\value{
Expr
}
\description{
This is syntactic sugar for \code{pl$col(...)$max()}.
}
\examples{
df = pl$DataFrame(
  a = NA_real_,
  b = c(1:2, NA_real_, NA_real_),
  c = c(1:4)
)
df

}
\keyword{Expr_new}
