% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_mean}
\alias{pl_mean}
\title{Get the mean value.}
\usage{
pl_mean(...)
}
\arguments{
\item{...}{Characters indicating the column names.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(...)$mean()}.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$mean("a"))

df$select(pl$mean("a", "b"))
}
\seealso{
\itemize{
\item \code{\link[=Expr_mean]{<Expr>$mean()}}
}
}
