% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_median}
\alias{pl_median}
\title{pl$median}
\arguments{
\item{column}{if dtype is:
\itemize{
\item Series: Take median value in \code{Series}
\item DataFrame or PolarsLazyFrame: Take median value of each column
\item str: syntactic sugar for \code{pl$col(..)$median()}
\item NULL: expression to take median column of a context.
}}
}
\value{
Expr or median value of input Series
}
\description{
Depending on the input type this function does different things:
}
\examples{

df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)
df$select(pl$median("a"))

df$select(pl$median("a", "b"))

}
\keyword{Expr_new}
