% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_median}
\alias{pl_median}
\title{Get the median value.}
\usage{
pl_median(columns)
}
\arguments{
\item{columns}{A character vector indicating the column names.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(columns)$median()}.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$median("a"))

df$select(pl$median(c("a", "b")))
}
\seealso{
\itemize{
\item \code{\link[=Expr_median]{<Expr>$median()}}
}
}
