% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pl_mem_address}
\alias{pl_mem_address}
\alias{mem_address}
\title{Get Memory Address}
\arguments{
\item{robj}{an R object}
}
\value{
String of mem address
}
\description{
Get underlying mem address a rust object (via ExtPtr). Expert use only.
}
\details{
Does not give meaningful answers for regular R objects.
}
\examples{
pl$mem_address(pl$Series(1:3))
}
