% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_min}
\alias{pl_min}
\title{Get the minimum value.}
\usage{
pl_min(...)
}
\arguments{
\item{...}{Characters indicating the column names, passed to \code{\link[=pl_col]{pl$col()}}.
See \code{\link[=pl_col]{?pl_col}} for details.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Syntactic sugar for \code{pl$col(...)$min()}.
}
\examples{
df = pl$DataFrame(
  num_1 = c(1, 8, 3),
  num_2 = c(4, 5, 2),
  chr_1 = c("foo", "bar", "foo")
)

df$select(pl$min("num_1"))

# Get the minimum value of multiple columns.
df$select(pl$min(r"(^num_\d+$)"))

df$select(pl$min("num_1", "num_2"))
}
\seealso{
\itemize{
\item \code{\link[=Expr_min]{<Expr>$min()}}
\item \code{\link[=pl_min_horizontal]{pl$min_horizontal()}}
}
}
