% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_min}
\alias{pl_min}
\title{min across expressions / literals / Series}
\arguments{
\item{...}{is a:
If one arg:
\itemize{
\item Series or Expr, same as \code{column$sum()}
\item string, same as \code{pl$col(column)$sum()}
\item numeric, same as \code{pl$lit(column)$sum()}
\item list of strings(column names) or expressions to add up as expr1 + expr2 + expr3 + ...
}

If several args, then wrapped in a list and handled as above.}
}
\value{
Expr
}
\description{
Folds the expressions from left to right, keeping the first non-null value.
}
\examples{
df = pl$DataFrame(
  a = NA_real_,
  b = c(2:1, NA_real_, NA_real_),
  c = c(1:3, NA_real_),
  d = c(1:2, NA_real_, -Inf)
)
# use min to get first non Null value for each row, otherwise insert 99.9
df$with_columns(
  pl$min("a", "b", "c", 99.9)$alias("d")
)

}
\keyword{Expr_new}
