% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_reduce}
\alias{pl_reduce}
\title{Accumulate over multiple columns horizontally with an R function}
\usage{
pl_reduce(lambda, exprs)
}
\arguments{
\item{lambda}{R function which takes two polars Series as input and return one.}

\item{exprs}{Expressions to aggregate over. May also be a wildcard expression.}
}
\value{
An expression that will be applied rowwise
}
\description{
This allows one to do rowwise operations. See \code{pl$fold()} to do rowwise
operations with an initial value.
}
\examples{
df = pl$DataFrame(mtcars)

# Make the row-wise sum of all columns
df$with_columns(
  pl$reduce(
    lambda = \(acc, x) acc + x,
    exprs = pl$col("*")
  )$alias("mpg_drat_sum_reduced")
)
}
