% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{pl_select}
\alias{pl_select}
\title{Select from an empty DataFrame}
\usage{
pl$select(...)
}
\arguments{
\item{...}{\link[=Expr_class]{Expressions}}
}
\value{
a \link[=DataFrame_class]{DataFrame}
}
\description{
\code{pl$select(...)} is a shorthand for \code{pl$DataFrame(list())$select(...)}
}
\examples{
pl$select(
  pl$lit(1:4)$alias("ints"),
  pl$lit(letters[1:4])$alias("letters")
)
}
\keyword{DataFrame}
