% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pl_select}
\alias{pl_select}
\alias{.onLoad}
\title{Select from an empty DataFrame}
\usage{
pl_select
}
\arguments{
\item{...}{expressions passed to select}
}
\value{
DataFrame
}
\description{
Select from an empty DataFrame
}
\details{
experimental
}
\examples{
pl$select(pl$lit(1:4))
}
\keyword{DataFrame}
