% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{pl_SQLContext}
\alias{pl_SQLContext}
\title{Initialise a new SQLContext}
\usage{
pl_SQLContext(...)
}
\arguments{
\item{...}{Name-value pairs of \link[=LazyFrame_class]{LazyFrame} like objects to register.}
}
\value{
RPolarsSQLContext
}
\description{
Create a new SQLContext and register the given LazyFrames.
}
\examples{
\dontshow{if (pl$polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ctx = pl$SQLContext(mtcars = mtcars)
ctx
\dontshow{\}) # examplesIf}
}
