% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_std}
\alias{pl_std}
\title{Get the standard deviation.}
\usage{
pl_std(..., ddof = 1)
}
\arguments{
\item{...}{Characters indicating the column names.}

\item{ddof}{An integer representing "Delta Degrees of Freedom":
the divisor used in the calculation is \code{N - ddof},
where \code{N} represents the number of elements. By default ddof is \code{1}.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(...)$std(ddof)}.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$std("a"))

df$select(pl$std(c("a", "b")))
}
\seealso{
\itemize{
\item \code{\link[=Expr_std]{<Expr>$std()}}
}
}
