% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{pl_Struct_datatype}
\alias{pl_Struct_datatype}
\title{Create Struct DataType}
\format{
function
}
\arguments{
\item{datatype}{an inner DataType}
}
\value{
a list DataType with an inner DataType
}
\description{
Struct DataType Constructor
}
\examples{
# create a Struct-DataType
pl$List(pl$List(pl$Boolean))

# Find any DataType via pl$dtypes
print(pl$dtypes)

#check if some maybe_Struct any kind Struct DataType
maybe_Struct = pl$Struct(pl$UInt64)
pl$same_outer_dt(maybe_Struct, pl$Struct())

#this will yield FALSE is not exactly the same
maybe_Struct == pl$Struct()
}
