% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_sum}
\alias{pl_sum}
\title{Sum all values.}
\usage{
pl_sum(...)
}
\arguments{
\item{...}{Characters indicating the column names, passed to \code{\link[=pl_col]{pl$col()}}.
See \code{\link[=pl_col]{?pl_col}} for details.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Syntactic sugar for \code{pl$col(...)$sum()}.
}
\examples{
df = pl$DataFrame(col_a = 1:2, col_b = 3:4, c = 5:6)

df$select(pl$sum("col_a"))

# Sum multiple columns
df$select(pl$sum("col_a", "col_b"))

df$select(pl$sum("^col_.*$"))
}
\seealso{
\itemize{
\item \code{\link[=Expr_sum]{<Expr>$sum()}}
\item \code{\link[=pl_sum_horizontal]{pl$sum_horizontal()}}
}
}
