% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_sum}
\alias{pl_sum}
\title{Compute sum in one or several columns}
\arguments{
\item{...}{is a:
If one arg:
\itemize{
\item Series or Expr, same as \code{column$sum()}
\item string, same as \code{pl$col(column)$sum()}
\item numeric, same as \code{pl$lit(column)$sum()}
\item list of strings(column names) or expressions to add up as expr1 + expr2 + expr3 + ...
}

If several args, then wrapped in a list and handled as above.}

\item{verbose}{Show the deprecation message when several columns or Expr are
passed in \code{...}. Will be removed in 0.12.0.}
}
\value{
Expr
}
\description{
This is syntactic sugar for \code{pl$col(...)$sum()}.
}
\examples{
# column as string
pl$DataFrame(iris)$select(pl$sum("Petal.Width"))

# column as Expr (prefer pl$col("Petal.Width")$sum())
pl$DataFrame(iris)$select(pl$sum(pl$col("Petal.Width")))

df = pl$DataFrame(a = 1:2, b = 3:4, c = 5:6)

# Compute sum in several columns
df$with_columns(pl$sum("*"))
}
\keyword{Expr_new}
