% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_sum}
\alias{pl_sum}
\title{Sum all values.}
\usage{
pl_sum(columns)
}
\arguments{
\item{columns}{A character vector indicating the column names.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
Syntactic sugar for \code{pl$col(coumns)$sum()}.
}
\examples{
df = pl$DataFrame(a = 1:2, b = 3:4, c = 5:6)

df$select(pl$sum("a"))

# Sum multiple columns
df$select(pl$sum(c("a", "c")))

df$select(pl$sum("^.*[bc]$"))
}
\seealso{
\itemize{
\item \code{\link[=Expr_sum]{<Expr>$sum()}}
\item \code{\link[=pl_sum_horizontal]{pl$sum_horizontal()}}
}
}
