% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions__lazy.R
\name{pl_tail}
\alias{pl_tail}
\title{Get the last \code{n} rows.}
\usage{
pl_tail(columns, n = 10)
}
\arguments{
\item{columns}{A character vector indicating the column names.}

\item{n}{Number of rows to return.}
}
\value{
\link[=Expr_class]{Expr}
}
\description{
This function is syntactic sugar for \code{pl$col(columns)$tail(n)}.
}
\examples{
df = pl$DataFrame(
  a = c(1, 8, 3),
  b = c(4, 5, 2),
  c = c("foo", "bar", "foo")
)

df$select(pl$tail("a"))

df$select(pl$tail(c("a", "b"), 2))
}
\seealso{
\itemize{
\item \code{\link[=Expr_tail]{<Expr>$tail()}}
}
}
