% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pl_using_string_cache}
\alias{pl_using_string_cache}
\title{Check if the global string cache is enabled}
\value{
A boolean
}
\description{
This function simply checks if the global string cache is active.
}
\examples{
pl$enable_string_cache()
pl$using_string_cache()
pl$disable_string_cache()
pl$using_string_cache()
}
\seealso{
\code{\link[=pl_with_string_cache]{pl$with_string_cache}}
\code{\link[=pl_enable_string_cache]{pl$enable_enable_cache}}
}
\keyword{options}
