% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_functions.R
\name{pl_var}
\alias{pl_var}
\title{Variance}
\arguments{
\item{ddof}{integer Delta Degrees of Freedom: the divisor used in the calculation is N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\description{
syntactic sugar for starting a expression with var
}
