% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{pl_with_string_cache}
\alias{pl_with_string_cache}
\title{Evaluate one or several expressions with global string cache}
\usage{
pl_with_string_cache(expr)
}
\value{
return value of expression
}
\description{
This function only temporarily enables the global string cache.
}
\examples{
# activate string cache temporarily when constructing two DataFrame's
pl$with_string_cache({
  df1 = pl$DataFrame(head(iris, 2))
  df2 = pl$DataFrame(tail(iris, 2))
})
pl$concat(list(df1, df2))
}
\seealso{
\code{\link[=pl_using_string_cache]{pl$using_string_cache}}
\code{\link[=pl_enable_string_cache]{pl$enable_enable_cache}}
}
\keyword{options}
