% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polars-package.R
\docType{package}
\name{polars-package}
\alias{polars}
\alias{polars-package}
\title{polars: Lightning-Fast 'DataFrame' Library}
\description{
Lightning-fast 'DataFrame' library written in 'Rust'. Convert R data to 'Polars' data and vice versa. Perform fast, lazy, larger-than-memory and optimized data queries. 'Polars' is interoperable with the package 'arrow', as both are based on the 'Apache Arrow' Columnar Format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pola-rs/r-polars https://rpolars.r-universe.dev/polars}
}

}
\author{
\strong{Maintainer}: Soren Welling \email{sorhawell@gmail.com}

Authors:
\itemize{
  \item Ritchie Vink \email{ritchie46@gmail.com}
  \item Tatsuya Shima \email{ts1s1andn@gmail.com}
  \item Etienne Bacher \email{etienne.bacher@protonmail.com}
}

}
\keyword{internal}
