% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/polars_envvars.R
\name{polars_envvars}
\alias{polars_envvars}
\title{Get polars environment variables}
\usage{
polars_envvars()
}
\value{
\code{polars_envvars()} returns a named list where the names are the names
of environment variables and values are their values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get polars environment variables
}
\details{
The following envvars are available (in alphabetical order, with the
default value in parenthesis):
\itemize{
\item \code{POLARS_FMT_MAX_COLS} (\code{5}): Set the number of columns that are visible when
displaying tables. If negative, all columns are displayed.
\item \code{POLARS_FMT_MAX_ROWS} (\code{8}): Set the number of rows that are visible when
displaying tables. If negative, all rows are displayed. This applies to both
\link{DataFrame} and \link{Series}.
\item \code{POLARS_FMT_STR_LEN} (\code{32}): Maximum number of characters to display;
\item \code{POLARS_FMT_TABLE_CELL_ALIGNMENT} (\code{"LEFT"}): set the table cell alignment.
Can be \code{"LEFT"}, \code{"CENTER"}, \code{"RIGHT"};
\item \code{POLARS_FMT_TABLE_CELL_LIST_LEN} (\code{3}): Maximum number of elements of list
variables to display;
\item \code{POLARS_FMT_TABLE_CELL_NUMERIC_ALIGNMENT} (\code{"LEFT"}): Set the table cell
alignment for numeric columns. Can be \code{"LEFT"}, \code{"CENTER"}, \code{"RIGHT"};
\item \code{POLARS_FMT_TABLE_DATAFRAME_SHAPE_BELOW} (\code{"0"}): print the DataFrame shape
information below the data when displaying tables. Can be \code{"0"} or \code{"1"}.
\item \code{POLARS_FMT_TABLE_FORMATTING} (\code{"UTF8_FULL_CONDENSED"}): Set table
formatting style. Possible values:
\itemize{
\item \code{"ASCII_FULL"}: ASCII, with all borders and lines, including row dividers.
\item \code{"ASCII_FULL_CONDENSED"}: Same as ASCII_FULL, but with dense row spacing.
\item \code{"ASCII_NO_BORDERS"}: ASCII, no borders.
\item \code{"ASCII_BORDERS_ONLY"}: ASCII, borders only.
\item \code{"ASCII_BORDERS_ONLY_CONDENSED"}: ASCII, borders only, dense row spacing.
\item \code{"ASCII_HORIZONTAL_ONLY"}: ASCII, horizontal lines only.
\item \code{"ASCII_MARKDOWN"}: ASCII, Markdown compatible.
\item \code{"UTF8_FULL"}: UTF8, with all borders and lines, including row dividers.
\item \code{"UTF8_FULL_CONDENSED"}: Same as UTF8_FULL, but with dense row spacing.
\item \code{"UTF8_NO_BORDERS"}: UTF8, no borders.
\item \code{"UTF8_BORDERS_ONLY"}: UTF8, borders only.
\item \code{"UTF8_HORIZONTAL_ONLY"}: UTF8, horizontal lines only.
\item \code{"NOTHING"}: No borders or other lines.
}
\item \code{POLARS_FMT_TABLE_HIDE_COLUMN_DATA_TYPES} (\code{"0"}): Hide table column data
types (i64, f64, str etc.). Can be \code{"0"} or \code{"1"}.
\item \code{POLARS_FMT_TABLE_HIDE_COLUMN_NAMES} (\code{"0"}): Hide table column names. Can
be \code{"0"} or \code{"1"}.
\item \code{POLARS_FMT_TABLE_HIDE_COLUMN_SEPARATOR} (\code{"0"}): Hide the \code{"---"} separator
between the column names and column types. Can be \code{"0"} or \code{"1"}.
\item \code{POLARS_FMT_TABLE_HIDE_DATAFRAME_SHAPE_INFORMATION} (\code{"0"}): Hide the
DataFrame shape information when displaying tables. Can be \code{"0"} or \code{"1"}.
\item \code{POLARS_FMT_TABLE_INLINE_COLUMN_DATA_TYPE} (\code{"0"}): Moves the data type
inline with the column name (to the right, in parentheses). Can be \code{"0"}
or \code{"1"}.
\item \code{POLARS_FMT_TABLE_ROUNDED_CORNERS} (\code{"0"}): Apply rounded corners to
UTF8-styled tables (only applies to UTF8 formats).
\item \code{POLARS_MAX_THREADS} (\verb{<variable>}): Maximum number of threads used to
initialize the thread pool. The thread pool is locked once polars is loaded,
so this envvar must be set before loading the package.
\item \code{POLARS_STREAMING_CHUNK_SIZE} (\verb{<variable>}): Chunk size used in the
streaming engine. Integer larger than 1. By default, the chunk size is
determined by the schema and size of the thread pool. For some datasets
(esp. when you have large string elements) this can be too optimistic and
lead to Out of Memory errors.
\item \code{POLARS_TABLE_WIDTH} (\verb{<variable>}): Set the maximum width of a table in
characters.
\item \code{POLARS_VERBOSE} (\code{"0"}): Enable additional verbose/debug logging.
\item \code{POLARS_WARN_UNSTABLE} (\code{"0"}): Issue a warning when unstable functionality
is used. Enabling this setting may help avoid functionality that is still
evolving, potentially reducing maintenance burden from API changes and bugs.
Can be \code{"0"} or \code{"1"}.
}

The following configuration options are present in the Python API but
currently cannot be changed in R: decimal separator, thousands separator,
float precision, float formatting, trimming decimal zeros.
}
\examples{
polars_envvars()

# By default, long string values are truncated.
df <- pl$DataFrame(x = "This is a very very very long sentence.")
df

# We can temporarily allow for wider columns with `withr::with_envvar()`:
withr::with_envvar(
  list(POLARS_FMT_STR_LEN = "50"),
  print(df)
)

# Or we could set it permanently with `Sys.setenv(POLARS_FMT_STR_LEN = "50")`.
}
