% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-expr.R
\name{polars_expr}
\alias{polars_expr}
\alias{Expr}
\alias{expression}
\title{Polars expression class (\code{polars_expr})}
\description{
An expression is a tree of operations that describe how to construct one or more \link{Series}.
As the outputs are \link{Series}, it is straightforward to apply a sequence of expressions each of
which transforms the output from the previous step.
See examples for details.
}
\examples{
# An expression:
# 1. Select column `foo`,
# 2. Then sort the column (not in reversed order)
# 3. Then take the first two values of the sorted output
pl$col("foo")$sort()$head(2)

# Expressions will be evaluated inside a context, such as `<DataFrame>$select()`
df <- pl$DataFrame(
  foo = c(1, 2, 1, 2, 3),
  bar = c(5, 4, 3, 2, 1),
)

df$select(
  pl$col("foo")$sort()$head(3), # Return 3 values
  pl$col("bar")$filter(pl$col("foo") == 1)$sum(), # Return a single value
)
}
\seealso{
\itemize{
\item \code{\link[=pl__lit]{pl$lit()}}: Create a literal expression.
\item \code{\link[=pl__col]{pl$col()}}: Create an expression representing column(s) in a \link{DataFrame}.
}
}
