% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polars_info.R
\name{polars_info}
\alias{polars_info}
\title{Report information of the package}
\usage{
polars_info()
}
\value{
A list with information of the package
}
\description{
This function reports the following information:
\itemize{
\item Package versions (the Polars R package version and the dependent Rust Polars crate version)
\item \link[=pl_thread_pool_size]{Number of threads used by Polars}
\item Rust feature flags (See \code{vignette("install", "polars")} for details)
\item Code completion mode: either \code{"deactivated"}, \code{"rstudio"}, or \code{"native"}.
See \code{\link[=polars_code_completion_activate]{polars_code_completion_activate()}}.
}
}
\examples{
polars_info()

polars_info()$versions

polars_info()$features$nightly
}
