% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_options}
\alias{set_options}
\alias{reset_options}
\title{Set polars options}
\arguments{
\item{strictly_immutable}{Keep polars strictly immutable. Polars/arrow is in
general pro "immutable objects". Immutability is also classic in R. To mimic
the Python-polars API, set this to \code{FALSE.}}

\item{maintain_order}{Default for all \code{maintain_order} options (present in
\verb{$groupby()} or \verb{$unique()} for example).}

\item{do_not_repeat_call}{Do not print the call causing the error in error
messages. The default (\code{FALSE}) is to show them.}

\item{debug_polars}{Print additional information to debug Polars.}

\item{no_messages}{Hide messages.}
}
\value{
\code{pl$options} returns a named list with the value (\code{TRUE} or \code{FALSE}) of
each option.
\code{pl$set_options()} silently modifies the options values.
\code{pl$reset_options()} silently resets the options to their default values.
}
\description{
Get and set polars options. See sections "Value" and "Examples" below for
more details.
}
\examples{
pl$set_options(maintain_order = TRUE, strictly_immutable = FALSE)
pl$options

# these options only accept booleans (TRUE or FALSE)
tryCatch(
  pl$set_options(strictly_immutable = 42),
  error = function(e) print(e)
)

# reset options to their default value
pl$reset_options()
}
