% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{polars_options}
\alias{polars_options}
\title{Report information of the package}
\usage{
polars_options()
}
\value{
A list with information of the package
}
\description{
This function reports the following information:
\itemize{
\item Package versions (the R package version and the dependent Rust Polars version)
\item \link[=pl_threadpool_size]{Number of threads used by Polars}
\item Rust feature flags (See \code{vignette("install", "polars")} for details)
}
}
\examples{
polars_options()
}
