% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{print()}
\alias{print()}
\alias{print.DataFrame}
\title{s3 method print DataFrame}
\usage{
\method{print}{DataFrame}(x, ...)
}
\arguments{
\item{x}{DataFrame}

\item{...}{not used}
}
\value{
self
}
\description{
s3 method print DataFrame
}
\examples{
pl$DataFrame(iris)
}
\keyword{internal}
