% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{print.DataFrame}
\alias{print.DataFrame}
\title{print DataFrame}
\usage{
\method{print}{DataFrame}(x)
}
\arguments{
\item{x}{polar_frame}
}
\value{
self
}
\description{
print DataFrame
}
\examples{
pl::pf(iris)
}
