% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{print.GroupBy}
\alias{print.GroupBy}
\title{print GroupBy}
\usage{
\method{print}{GroupBy}(x, ...)
}
\arguments{
\item{x}{DataFrame}

\item{...}{not used}
}
\value{
self
}
\description{
print GroupBy
}
\examples{
pl$DataFrame(iris)$group_by("Species")
}
\keyword{internal}
