% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{print.LazyFrame}
\alias{print.LazyFrame}
\title{print LazyFrame s3 method}
\usage{
\method{print}{LazyFrame}(x)
}
\arguments{
\item{x}{DataFrame}
}
\value{
self
}
\description{
print LazyFrame s3 method
}
\examples{
print(pl$DataFrame(iris)$lazy())
}
\keyword{LazyFrame}
