% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R, R/series.R
\name{print.Series}
\alias{print.Series}
\title{Print rpl$Series}
\usage{
\method{print}{Series}(x)

\method{print}{Series}(x)
}
\arguments{
\item{x}{Series}
}
\value{
self

selfie
}
\description{
Print rpl$Series

Print rpl$Series
}
\examples{
pl$Series(letters,"lowercase_letters")
pl$Series(letters,"lowercase_letters")
}
