% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{pstop}
\alias{pstop}
\title{Internal preferred function to throw errors}
\usage{
pstop(err, call = sys.call(1L))
}
\arguments{
\item{err}{error msg string}

\item{call}{calling context}
}
\value{
throws an error
}
\description{
DEPRECATED USE stopf instead
}
\examples{
f = function() rpolars:::pstop("this aint right!!")
tryCatch(f(), error = \(e) as.character(e))
}
\keyword{internals}
