% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/lazyframe-opt_flags.R
\name{QueryOptFlags}
\alias{QueryOptFlags}
\alias{pl__QueryOptFlags}
\title{The set of the optimizations considered during query optimization}
\usage{
pl__QueryOptFlags(
  ...,
  predicate_pushdown = TRUE,
  projection_pushdown = TRUE,
  simplify_expression = TRUE,
  slice_pushdown = TRUE,
  comm_subplan_elim = TRUE,
  comm_subexpr_elim = TRUE,
  cluster_with_columns = TRUE,
  check_order_observe = TRUE,
  fast_projection = TRUE
)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{predicate_pushdown}{A logical, indicates predicate pushdown optimization.}

\item{projection_pushdown}{A logical, indicates projection pushdown optimization.}

\item{simplify_expression}{A logical, indicates simplify expression optimization.}

\item{slice_pushdown}{A logical, indicates slice pushdown optimization.}

\item{comm_subplan_elim}{A logical, indicates trying to cache branching subplans that occur
on self-joins or unions.}

\item{comm_subexpr_elim}{A logical, indicates trying to cache common subexpressions.}

\item{cluster_with_columns}{A logical, indicates to combine sequential independent calls
to with_columns.}

\item{check_order_observe}{A logical, indicates not to maintain order
if the order would not be observed.}

\item{fast_projection}{A logical, indicates to replace simple projections
with a faster inlined projection that skips the expression engine.}
}
\value{
A \code{QueryOptFlags} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
opt_flags <- pl$QueryOptFlags()
opt_flags
}
