% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restruct_list}
\alias{restruct_list}
\title{restruct an object where structs where previously unnested}
\usage{
restruct_list(l)
}
\arguments{
\item{l}{list}
}
\value{
restructed list
}
\description{
restruct an object where structs where previously unnested
}
\details{
this function should be repalced with rust code writing this output
directly before nesting
}
