% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_conversion.R
\name{result}
\alias{result}
\title{catches any R error and return a rust-Result}
\usage{
result(expr, msg = NULL)
}
\arguments{
\item{expr}{code to capture any error from and wrap as Result}

\item{msg}{easy way to add a context msg}
}
\value{
Result
}
\description{
catches any R error and return a rust-Result
}
\keyword{internal}
