% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{RField_name}
\alias{RField_name}
\title{get/set Field name}
\usage{
RField_name()
}
\value{
name
}
\description{
get/set Field name
}
\examples{
field = pl$Field("Cities",pl$Utf8)

#get name / datatype
field$name
field$datatype

#set + get values
field$name = "CityPoPulations" #<- is fine too
field$datatype = pl$UInt32

print(field)
}
\keyword{DataType}
