% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{RField_name}
\alias{RField_name}
\title{Get/set Field name}
\usage{
RField_name()
}
\description{
Get/set Field name
}
\examples{
field = pl$Field("Cities", pl$Utf8)
field$name

field$name = "CityPoPulations" #<- is fine too
field
}
