% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_rolling.R
\name{RollingGroupBy_class}
\alias{RollingGroupBy_class}
\title{Operations on Polars DataFrame grouped by rolling windows}
\value{
not applicable
}
\description{
Operations on Polars DataFrame grouped by rolling windows
}
