% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_rolling.R
\name{RollingGroupBy_ungroup}
\alias{RollingGroupBy_ungroup}
\title{RollingGroupBy_ungroup}
\usage{
RollingGroupBy_ungroup()
}
\value{
\link[=DataFrame_class]{DataFrame}
}
\description{
Revert the group by operation.
}
\examples{
gb = pl$DataFrame(mtcars)$group_by("cyl")
gb

gb$ungroup()
}
