% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{RThreadHandle_is_finished}
\alias{RThreadHandle_is_finished}
\title{Ask if RThreadHandle is finished?}
\usage{
RThreadHandle_is_finished()
}
\description{
Ask if RThreadHandle is finished?
}
\details{
method \verb{<RThreadHandle>$is_finished()}: Calling \verb{<RThreadHandle>$is_finished()} returns
trinary value: \code{TRUE} if finished, \code{FALSE} if not, and \code{NULL} if the handle was exhausted
(already joined).
}
\keyword{RThreadHandle}
