% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{RThreadHandle_is_finished}
\alias{RThreadHandle_is_finished}
\title{Check if the thread completes its job}
\usage{
RThreadHandle_is_finished(...)
}
\arguments{
\item{...}{a RThreadHandle}
}
\value{
a boolean indicating the whether the job has finished
or NULL if the handle has been joined
}
\description{
Check if the thread completes its job
}
\keyword{RThreadHandle}
