% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbackground.R
\name{RThreadHandle_join}
\alias{RThreadHandle_join}
\title{Wait for the thread to complete its job}
\usage{
RThreadHandle_join(...)
}
\arguments{
\item{...}{a RThreadHandle}
}
\value{
the result of the job
}
\description{
Wait for the thread to complete its job
}
\keyword{RThreadHandle}
