% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-s3-base.R, R/lazyframe-s3-base.R
\name{as.list.polars_data_frame}
\alias{as.list.polars_data_frame}
\alias{as.list.polars_lazy_frame}
\title{Export the polars object as an R list}
\usage{
\method{as.list}{polars_data_frame}(
  x,
  ...,
  as_series = TRUE,
  uint8 = c("integer", "raw"),
  int64 = c("double", "character", "integer", "integer64"),
  date = c("Date", "IDate"),
  time = c("hms", "ITime"),
  struct = c("dataframe", "tibble"),
  decimal = c("double", "character"),
  as_clock_class = FALSE,
  ambiguous = c("raise", "earliest", "latest", "null"),
  non_existent = c("raise", "null")
)

\method{as.list}{polars_lazy_frame}(
  x,
  ...,
  as_series = TRUE,
  uint8 = c("integer", "raw"),
  int64 = c("double", "character", "integer", "integer64"),
  date = c("Date", "IDate"),
  time = c("hms", "ITime"),
  struct = c("dataframe", "tibble"),
  decimal = c("double", "character"),
  as_clock_class = FALSE,
  ambiguous = c("raise", "earliest", "latest", "null"),
  non_existent = c("raise", "null")
)
}
\arguments{
\item{x}{A polars object}

\item{...}{Passed to \code{\link[=as_polars_df]{as_polars_df()}}.}

\item{as_series}{Whether to convert each column to an \link[=vector]{R vector} or a \link{Series}.
If \code{TRUE} (default), return a list of \link{Series}, otherwise a list of \link[=vector]{vectors}.}

\item{uint8}{Determine how to convert Polars' UInt8 type values to R type.
One of the followings:
\itemize{
\item \code{"integer"} (default): Convert to the R's \link{integer} type.
\item \code{"raw"}: Convert to the R's \link{raw} type.
If the value is \code{null}, export as \code{00}.
}}

\item{int64}{Determine how to convert Polars' Int64, UInt32, or UInt64 type values to R type.
One of the followings:
\itemize{
\item \code{"double"} (default): Convert to the R's \link{double} type.
Accuracy may be degraded.
\item \code{"character"}: Convert to the R's \link{character} type.
\item \code{"integer"}: Convert to the R's \link{integer} type.
If the value is out of the range of R's integer type, export as \link{NA_integer_}.
\item \code{"integer64"}: Convert to the \link[bit64:bit64-package]{bit64::integer64} class.
The \link[bit64:bit64-package]{bit64} package must be installed.
If the value is out of the range of \link[bit64:bit64-package]{bit64::integer64}, export as \link[bit64:as.integer64.character]{bit64::NA_integer64_}.
}}

\item{date}{Determine how to convert Polars' Date type values to R class.
One of the followings:
\itemize{
\item \code{"Date"} (default): Convert to the R's \link{Date} class.
\item \code{"IDate"}: Convert to the \link[data.table:IDateTime]{data.table::IDate} class.
}}

\item{time}{Determine how to convert Polars' Time type values to R class.
One of the followings:
\itemize{
\item \code{"hms"} (default): Convert to the \link[hms:hms]{hms::hms} class.
If the \link[hms:hms-package]{hms} package is not installed, a warning will be shown.
\item \code{"ITime"}: Convert to the \link[data.table:IDateTime]{data.table::ITime} class.
The \link[data.table:data.table]{data.table} package must be installed.
}}

\item{struct}{Determine how to convert Polars' Struct type values to R class.
One of the followings:
\itemize{
\item \code{"dataframe"} (default): Convert to the R's \link{data.frame} class.
\item \code{"tibble"}: Convert to the \link[tibble:tbl_df-class]{tibble} class.
If the \link[tibble:tibble-package]{tibble} package is not installed, a warning will be shown.
}}

\item{decimal}{Determine how to convert Polars' Decimal type values to R type.
One of the followings:
\itemize{
\item \code{"double"} (default): Convert to the R's \link{double} type.
\item \code{"character"}: Convert to the R's \link{character} type.
}}

\item{as_clock_class}{A logical value indicating whether to export datetimes and duration as
the \link[clock:clock-package]{clock} package's classes.
\itemize{
\item \code{FALSE} (default): Duration values are exported as \link{difftime}
and datetime values are exported as \link{POSIXct}.
Accuracy may be degraded.
\item \code{TRUE}: Duration values are exported as \link[clock:duration-helper]{clock_duration},
datetime without timezone values are exported as \link[clock:as_naive_time]{clock_naive_time},
and datetime with timezone values are exported as \link[clock:as_zoned_time]{clock_zoned_time}.
For this case, the \link[clock:clock-package]{clock} package must be installed.
Accuracy will be maintained.
}}

\item{ambiguous}{Determine how to deal with ambiguous datetimes.
Only applicable when \code{as_clock_class} is set to \code{FALSE} and
datetime without timezone values are exported as \link{POSIXct}.
Character vector or \link{expression} containing the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"earliest"}: Use the earliest datetime
\item \code{"latest"}: Use the latest datetime
\item \code{"null"}: Return a \code{NA} value
}}

\item{non_existent}{Determine how to deal with non-existent datetimes.
Only applicable when \code{as_clock_class} is set to \code{FALSE} and
datetime without timezone values are exported as \link{POSIXct}.
One of the followings:
\itemize{
\item \code{"raise"} (default): Throw an error
\item \code{"null"}: Return a \code{NA} value
}}
}
\value{
A \link{list}
}
\description{
These S3 methods call \code{\link[=dataframe__get_columns]{as_polars_df(x, ...)$get_columns()}} with
\code{\link[rlang:set_names]{rlang::set_names()}}, or, \code{as_polars_df(x, ...)$to_struct()$to_r_vector() |> as.list()}
depending on the \code{as_series} argument.
}
\details{
Arguments other than \code{x} and \code{as_series} are passed to
\code{\link[=series__to_r_vector]{<Series>$to_r_vector()}},
so they are ignored when \code{as_series=TRUE}.
}
\examples{
df <- as_polars_df(list(a = 1:3, b = 4:6))

as.list(df, as_series = TRUE)
as.list(df, as_series = FALSE)

as.list(df$lazy(), as_series = TRUE)
as.list(df$lazy(), as_series = FALSE)
}
\seealso{
\itemize{
\item \code{\link[=dataframe__get_columns]{<DataFrame>$get_columns()}}
}
}
