% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-s3-knitr.R, R/series-s3-knitr.R
\name{knit_print.polars_data_frame}
\alias{knit_print.polars_data_frame}
\alias{knit_print.polars_series}
\title{knit print polars DataFrame}
\usage{
knit_print.polars_data_frame(x, ...)

knit_print.polars_series(x, ...)
}
\arguments{
\item{x}{A polars object}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\value{
\code{x} invisibly or [knit_asis][knitr::asis_output)] object.

[knitr::asis_output)]: R:knitr::asis_output)
}
\description{
Mimics Python Polars' NotebookFormatter
for HTML outputs.
}
\details{
Outputs HTML tables if the output format is HTML
and the document's \code{df_print} option is not \code{"default"} or \code{"tibble"}.

Or, the output format can be enforced with R's \code{options} function as follows:
\itemize{
\item \code{options(polars.df_knitr_print  = "default")} for the default print method.
\item \code{options(polars.df_knitr_print  = "html")} for the HTML table.
}
}
