% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods-operator.R
\name{S3_arithmetic}
\alias{S3_arithmetic}
\alias{+.RPolarsExpr}
\alias{/.RPolarsExpr}
\alias{\%/\%.RPolarsExpr}
\alias{\%\%.RPolarsExpr}
\alias{-.RPolarsExpr}
\alias{*.RPolarsExpr}
\alias{^.RPolarsExpr}
\title{Arithmetic operators for RPolars objects}
\usage{
\method{+}{RPolarsExpr}(e1, e2)

\method{/}{RPolarsExpr}(e1, e2)

\method{\%/\%}{RPolarsExpr}(e1, e2)

\method{\%\%}{RPolarsExpr}(e1, e2)

\method{-}{RPolarsExpr}(e1, e2)

\method{*}{RPolarsExpr}(e1, e2)

\method{^}{RPolarsExpr}(e1, e2)
}
\arguments{
\item{e1}{Expr only}

\item{e2}{Expr or anything that can be converted to a literal}
}
\description{
Arithmetic operators for RPolars objects
}
\examples{
pl$lit(5) + 10
5 + pl$lit(10)
pl$lit(5) + pl$lit(10)
+pl$lit(1)

# This will not raise an error as it is not actually evaluated.
expr = pl$lit(5) + "10"
expr

# Will raise an error as it is evaluated.
tryCatch(
  expr$to_series(),
  error = function(e) e
)
}
\seealso{
\itemize{
\item \code{\link[=Expr_add]{<Expr>$add()}}
\item \code{\link[=Expr_div]{<Expr>$div()}}
\item \code{\link[=Expr_floor_div]{<Expr>$floor_div()}}
\item \code{\link[=Expr_mod]{<Expr>$mod()}}
\item \code{\link[=Expr_sub]{<Expr>$sub()}}
\item \code{\link[=Expr_mul]{<Expr>$mul()}}
\item \code{\link[=Expr_pow]{<Expr>$pow()}}
}
}
