% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.data.frame.DataFrame}
\alias{as.data.frame.DataFrame}
\alias{as.data.frame.LazyFrame}
\title{Convert to a data.frame}
\usage{
\method{as.data.frame}{DataFrame}(x, ...)

\method{as.data.frame}{LazyFrame}(x, ...)
}
\arguments{
\item{x}{A Polars DataFrame or LazyFrame}

\item{...}{Any arguments passed to \code{data.frame()}.}
}
\description{
Convert to a data.frame
}
