% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.data.frame.RPolarsDataFrame}
\alias{as.data.frame.RPolarsDataFrame}
\alias{as.data.frame.RPolarsLazyFrame}
\title{Convert to a data.frame}
\usage{
\method{as.data.frame}{RPolarsDataFrame}(x, ...)

\method{as.data.frame}{RPolarsLazyFrame}(x, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}

\item{...}{Any arguments passed to \code{\link[=data.frame]{data.frame()}}.}
}
\description{
Convert to a data.frame
}
