% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.matrix.RPolarsDataFrame}
\alias{as.matrix.RPolarsDataFrame}
\alias{as.matrix.RPolarsLazyFrame}
\title{Convert to a matrix}
\usage{
\method{as.matrix}{RPolarsDataFrame}(x, ...)

\method{as.matrix}{RPolarsLazyFrame}(x, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}

\item{...}{Any arguments passed to \code{\link[=data.frame]{data.frame()}}.}
}
\description{
Convert to a matrix
}
