% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.matrix.RPolarsDataFrame}
\alias{as.matrix.RPolarsDataFrame}
\alias{as.matrix.RPolarsLazyFrame}
\title{Convert to a matrix}
\usage{
\method{as.matrix}{RPolarsDataFrame}(x, ...)

\method{as.matrix}{RPolarsLazyFrame}(x, ...)
}
\arguments{
\item{x}{An object to convert to a \link{data.frame}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Convert to a matrix
}
