% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{as.matrix.RPolarsDataFrame}
\alias{as.matrix.RPolarsDataFrame}
\alias{as.matrix.RPolarsLazyFrame}
\title{Convert to a matrix}
\usage{
\method{as.matrix}{RPolarsDataFrame}(x, ...)

\method{as.matrix}{RPolarsLazyFrame}(x, ...)
}
\arguments{
\item{x}{An object to convert to a \link{matrix}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Equivalent to \code{as.data.frame(x, ...) |> as.matrix()}.
}
