% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.matrix.DataFrame}
\alias{as.matrix.DataFrame}
\alias{as.matrix.LazyFrame}
\title{Convert to a matrix}
\usage{
\method{as.matrix}{DataFrame}(x, ...)

\method{as.matrix}{LazyFrame}(x, ...)
}
\arguments{
\item{x}{A Polars DataFrame or LazyFrame}

\item{...}{Not used.}
}
\description{
Convert to a matrix
}
