% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{as.vector.RPolarsSeries}
\alias{as.vector.RPolarsSeries}
\title{Convert to a vector}
\usage{
\method{as.vector}{RPolarsSeries}(x, mode)
}
\arguments{
\item{x}{A Polars Series}

\item{mode}{Not used.}
}
\description{
Convert to a vector
}
