% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{as_arrow_table.RPolarsDataFrame}
\alias{as_arrow_table.RPolarsDataFrame}
\title{Create a arrow Table from a Polars object}
\usage{
\method{as_arrow_table}{RPolarsDataFrame}(x, ...)
}
\arguments{
\item{x}{\link[=DataFrame_class]{A Polars DataFrame}}

\item{...}{Ignored}
}
\description{
Create a arrow Table from a Polars object
}
\examples{
\dontshow{if (requireNamespace("arrow", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(arrow)

pl_df = as_polars_df(mtcars)
as_arrow_table(pl_df)
\dontshow{\}) # examplesIf}
}
