% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-nanoarrow.R
\name{as_nanoarrow_array_stream.RPolarsDataFrame}
\alias{as_nanoarrow_array_stream.RPolarsDataFrame}
\title{Create a nanoarrow_array_stream from a Polars object}
\usage{
as_nanoarrow_array_stream.RPolarsDataFrame(x, ..., schema = NULL)
}
\arguments{
\item{x}{\link[=DataFrame_class]{A Polars DataFrame}}

\item{...}{Ignored}

\item{schema}{must stay at default value NULL}
}
\description{
Create a nanoarrow_array_stream from a Polars object
}
\examples{
library(nanoarrow)
pl_df = as_polars_df(mtcars)

nanoarrow_array_stream = as_nanoarrow_array_stream(pl_df)
as.data.frame(nanoarrow_array_stream)
}
