% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{as_record_batch_reader.RPolarsDataFrame}
\alias{as_record_batch_reader.RPolarsDataFrame}
\title{Create a arrow RecordBatchReader from a Polars object}
\usage{
\method{as_record_batch_reader}{RPolarsDataFrame}(x, ..., compat_level = FALSE)
}
\arguments{
\item{x}{\link[=DataFrame_class]{A Polars DataFrame}}

\item{...}{Ignored}

\item{compat_level}{Use a specific compatibility level when exporting Polars’
internal data structures. This can be:
\itemize{
\item an integer indicating the compatibility version (currently only 0 for oldest
and 1 for newest);
\item a logical value with \code{TRUE} for the newest version and \code{FALSE} for the oldest
version;
\item a character value (either \code{"oldest"} or \code{"newest"}).
}}
}
\description{
Create a arrow RecordBatchReader from a Polars object
}
\examples{
\dontshow{if (requireNamespace("arrow", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(arrow)

pl_df = as_polars_df(mtcars)
as_record_batch_reader(pl_df)
\dontshow{\}) # examplesIf}
}
