% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-arrow.R
\name{as_record_batch_reader.RPolarsDataFrame}
\alias{as_record_batch_reader.RPolarsDataFrame}
\title{Create a arrow RecordBatchReader from a Polars object}
\usage{
\method{as_record_batch_reader}{RPolarsDataFrame}(x, ...)
}
\arguments{
\item{x}{\link[=DataFrame_class]{A Polars DataFrame}}

\item{...}{Ignored}
}
\description{
Create a arrow RecordBatchReader from a Polars object
}
\examples{
library(arrow)

pl_df = as_polars_df(mtcars)
as_record_batch_reader(pl_df)
}
