% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{dim.RPolarsDataFrame}
\alias{dim.RPolarsDataFrame}
\alias{dim.RPolarsLazyFrame}
\title{Get the dimensions}
\usage{
\method{dim}{RPolarsDataFrame}(x)

\method{dim}{RPolarsLazyFrame}(x)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}
}
\description{
Get the dimensions
}
