% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{dim.DataFrame}
\alias{dim.DataFrame}
\alias{dim.LazyFrame}
\title{Get the dimensions}
\usage{
\method{dim}{DataFrame}(x)

\method{dim}{LazyFrame}(x)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}
}
\description{
Get the dimensions
}
