% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{dimnames.RPolarsDataFrame}
\alias{dimnames.RPolarsDataFrame}
\alias{dimnames.RPolarsLazyFrame}
\title{Get the row and column names}
\usage{
\method{dimnames}{RPolarsDataFrame}(x)

\method{dimnames}{RPolarsLazyFrame}(x)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}
}
\description{
Get the row and column names
}
