% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{head.DataFrame}
\alias{head.DataFrame}
\alias{head.LazyFrame}
\title{Take the first n rows of a DataFrame or LazyFrame}
\usage{
\method{head}{DataFrame}(x, n = 6L, ...)

\method{head}{LazyFrame}(x, n = 6L, ...)
}
\arguments{
\item{x}{A Polars DataFrame or LazyFrame}

\item{n}{Number of rows}

\item{...}{Not used}
}
\description{
Take the first n rows of a DataFrame or LazyFrame
}
