% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-nanoarrow.R
\name{infer_nanoarrow_schema.RPolarsDataFrame}
\alias{infer_nanoarrow_schema.RPolarsDataFrame}
\alias{infer_nanoarrow_schema.RPolarsSeries}
\title{Infer nanoarrow schema from a Polars object}
\usage{
infer_nanoarrow_schema.RPolarsDataFrame(x, ..., compat_level = FALSE)

infer_nanoarrow_schema.RPolarsSeries(x, ..., compat_level = FALSE)
}
\arguments{
\item{x}{A polars object}

\item{...}{Ignored}

\item{compat_level}{Use a specific compatibility level when exporting Polars’
internal data structures. This can be:
\itemize{
\item an integer indicating the compatibility version (currently only 0 for oldest
and 1 for newest);
\item a logical value with \code{TRUE} for the newest version and \code{FALSE} for the oldest
version.
}}
}
\description{
Infer nanoarrow schema from a Polars object
}
\examples{
\dontshow{if (requireNamespace("nanoarrow", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(nanoarrow)

pl_df = as_polars_df(mtcars)$select("mpg", "cyl")
pl_s = as_polars_series(letters)

infer_nanoarrow_schema(pl_df)
infer_nanoarrow_schema(pl_s)
\dontshow{\}) # examplesIf}
}
