% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-nanoarrow.R
\name{infer_nanoarrow_schema.RPolarsDataFrame}
\alias{infer_nanoarrow_schema.RPolarsDataFrame}
\title{Infer nanoarrow schema from a Polars object}
\usage{
infer_nanoarrow_schema.RPolarsDataFrame(x, ...)
}
\arguments{
\item{x}{\link[=DataFrame_class]{A Polars DataFrame}}

\item{...}{Ignored}
}
\description{
Infer nanoarrow schema from a Polars object
}
\examples{
library(nanoarrow)
pl_df = as_polars_df(mtcars)

infer_nanoarrow_schema(pl_df)
}
