% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-nanoarrow.R
\name{infer_nanoarrow_schema.RPolarsDataFrame}
\alias{infer_nanoarrow_schema.RPolarsDataFrame}
\title{Infer nanoarrow schema from a Polars object}
\usage{
\method{infer_nanoarrow_schema}{RPolarsDataFrame}(x, ...)
}
\description{
Infer nanoarrow schema from a Polars object
}
