% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{length.DataFrame}
\alias{length.DataFrame}
\alias{length.LazyFrame}
\alias{length.Series}
\title{Get the length}
\usage{
\method{length}{DataFrame}(x)

\method{length}{LazyFrame}(x)

\method{length}{Series}(x)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame}, \link[=LazyFrame_class]{LazyFrame}, or
\link[=Series_class]{Series}}
}
\description{
Get the length
}
