% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{max.DataFrame}
\alias{max.DataFrame}
\alias{max.LazyFrame}
\alias{max.Series}
\title{Compute the maximum value}
\usage{
\method{max}{DataFrame}(x, ...)

\method{max}{LazyFrame}(x, ...)

\method{max}{Series}(x, ...)
}
\arguments{
\item{x}{A Polars DataFrame, LazyFrame, or Series}
}
\description{
Compute the maximum value
}
