% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{mean.DataFrame}
\alias{mean.DataFrame}
\alias{mean.LazyFrame}
\alias{mean.Series}
\title{Compute the mean}
\usage{
\method{mean}{DataFrame}(x, ...)

\method{mean}{LazyFrame}(x, ...)

\method{mean}{Series}(x, ...)
}
\arguments{
\item{x}{A Polars DataFrame, LazyFrame, or Series}
}
\description{
Compute the mean
}
