% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{mean.RPolarsDataFrame}
\alias{mean.RPolarsDataFrame}
\alias{mean.RPolarsLazyFrame}
\alias{mean.RPolarsSeries}
\title{Compute the mean}
\usage{
\method{mean}{RPolarsDataFrame}(x, ...)

\method{mean}{RPolarsLazyFrame}(x, ...)

\method{mean}{RPolarsSeries}(x, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame}, \link[=LazyFrame_class]{LazyFrame}, or
\link[=Series_class]{Series}}

\item{...}{Not used.}
}
\description{
Compute the mean
}
