% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{median.DataFrame}
\alias{median.DataFrame}
\alias{median.LazyFrame}
\alias{median.Series}
\title{Compute the median}
\usage{
\method{median}{DataFrame}(x, ...)

\method{median}{LazyFrame}(x, ...)

\method{median}{Series}(x, ...)
}
\arguments{
\item{x}{A Polars DataFrame, LazyFrame, or Series}

\item{...}{Not used.}
}
\description{
Compute the median
}
