% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{min.DataFrame}
\alias{min.DataFrame}
\alias{min.LazyFrame}
\alias{min.Series}
\title{Compute the minimum value}
\usage{
\method{min}{DataFrame}(x, ...)

\method{min}{LazyFrame}(x, ...)

\method{min}{Series}(x, ...)
}
\arguments{
\item{x}{A Polars DataFrame, LazyFrame, or Series}
}
\description{
Compute the minimum value
}
