% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{names.DataFrame}
\alias{names.DataFrame}
\alias{names.LazyFrame}
\title{Get the column names}
\usage{
\method{names}{DataFrame}(x)

\method{names}{LazyFrame}(x)
}
\arguments{
\item{x}{A Polars DataFrame or LazyFrame}
}
\description{
Get the column names
}
