% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{tail.RPolarsDataFrame}
\alias{tail.RPolarsDataFrame}
\alias{tail.RPolarsLazyFrame}
\title{Take the last n rows}
\usage{
\method{tail}{RPolarsDataFrame}(x, n = 6L, ...)

\method{tail}{RPolarsLazyFrame}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \link[=DataFrame_class]{DataFrame} or \link[=LazyFrame_class]{LazyFrame}}

\item{n}{Number of rows}

\item{...}{Not used}
}
\description{
Take the last n rows
}
