% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{scan_parquet}
\alias{scan_parquet}
\title{new PolarsLazyFrame from parquet file}
\usage{
scan_parquet(
  file,
  n_rows = NULL,
  cache = TRUE,
  parallel = c("Auto", "None", "Columns", "RowGroups"),
  rechunk = TRUE,
  row_count_name = NULL,
  row_count_offset = 0L,
  low_memory = FALSE
)
}
\arguments{
\item{file}{string filepath}

\item{n_rows}{limit rows to scan}

\item{cache}{bool use cache}

\item{parallel}{String either Auto, None, Columns or RowGroups. The way to parralize the scan.}

\item{rechunk}{bool rechunk reorganize memory layout, potentially make future operations faster , however perform reallocation now.}

\item{row_count_name}{NULL or string, if a string add a rowcount column named by this string}

\item{row_count_offset}{integer, the rowcount column can be offst by this value}

\item{low_memory}{bool, try reduce memory footprint}
}
\value{
PolarsLazyFrame
}
\description{
new PolarsLazyFrame from parquet file
}
\examples{
# TODO write parquet example
}
\keyword{PolarsLazyFrame_new}
