% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{Series}
\alias{Series}
\title{Create new Series}
\arguments{
\item{x}{any vector}

\item{name}{string}
}
\value{
Series
}
\description{
found in api as pl$Series named Series_constructor internally
}
\examples{
{
pl$Series(1:4)
}
}
\keyword{Series_new}
