% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{Series}
\alias{Series}
\alias{Series_constructor}
\title{Series}
\usage{
Series_constructor(x, name = NULL)
}
\arguments{
\item{x}{any vector}

\item{name}{string}
}
\value{
Series
}
\description{
Polars pl$Series

found in api as pl$Series named Series_constructor internally
}
\examples{
Series(1:4)
}
