% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__chunk_lengths}
\alias{series__chunk_lengths}
\title{Get the length of each individual chunk}
\usage{
series__chunk_lengths()
}
\value{
A numeric vector
}
\description{
Get the length of each individual chunk
}
\examples{
s <- pl$Series("a", c(1, 2, 3))
s$chunk_lengths()

s2 <- pl$Series("a", c(4, 5, 6))

# Concatenate Series with rechunk = TRUE
pl$concat(s, s2, rechunk = TRUE)$chunk_lengths()

# Concatenate Series with rechunk = FALSE
pl$concat(s, s2, rechunk = FALSE)$chunk_lengths()
}
