% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__is_empty}
\alias{series__is_empty}
\title{Check if the Series is empty}
\usage{
series__is_empty()
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Check if the Series is empty
}
\examples{
s <- pl$Series("a", integer())
s$is_empty()
}
