% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__n_chunks}
\alias{series__n_chunks}
\title{Get the number of chunks that this Series contains}
\usage{
series__n_chunks()
}
\value{
An integer value
}
\description{
Get the number of chunks that this Series contains
}
\examples{
s <- pl$Series("a", c(1, 2, 3))
s$n_chunks()

s2 <- pl$Series("a", c(4, 5, 6))

# Concatenate Series with rechunk = TRUE
pl$concat(c(s, s2), rechunk = TRUE)$n_chunks()

# Concatenate Series with rechunk = FALSE
pl$concat(c(s, s2), rechunk = FALSE)$n_chunks()
}
