% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__n_chunks}
\alias{series__n_chunks}
\title{Get the number of chunks that this Series contains}
\usage{
series__n_chunks()
}
\value{
An integer value
}
\description{
Get the number of chunks that this Series contains
}
