% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__rechunk}
\alias{series__rechunk}
\title{Create a single chunk of memory for this Series}
\usage{
series__rechunk(..., in_place = FALSE)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{in_place}{Bool to indicate if the operation should be done in place.}
}
\value{
A \link[=Series]{polars Series}
}
\description{
Create a single chunk of memory for this Series
}
\examples{
s <- pl$Series("a", c(1, 2, 3))
s$n_chunks()

s2 <- pl$Series("a", c(4, 5, 6))
s <- pl$concat(s, s2, rechunk = FALSE)
s$n_chunks()

s$rechunk()$n_chunks()
}
