% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-series.R
\name{series__to_frame}
\alias{series__to_frame}
\title{Cast this Series to a DataFrame}
\usage{
series__to_frame(name = NULL)
}
\arguments{
\item{name}{A character or \code{NULL}. If not \code{NULL},
name/rename the \link{Series} column in the new \link{DataFrame}.
If \code{NULL}, the column name is taken from the \link{Series} name.}
}
\value{
A polars \link{DataFrame}
}
\description{
Cast this Series to a DataFrame
}
\examples{
s <- pl$Series("a", c(123, 456))
df <- s$to_frame()
df

df <- s$to_frame("xyz")
df
}
\seealso{
\itemize{
\item \code{\link[=as_polars_df]{as_polars_df()}}
}
}
