% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_add}
\alias{Series_add}
\title{Add Series}
\usage{
Series_add(other)
}
\arguments{
\item{other}{\link[=Series_class]{Series} like object of numeric or string values.
Converted to \link[=Series_class]{Series} by \code{\link[=as_polars_series]{as_polars_series()}} in this method.}
}
\value{
\link[=Series_class]{Series}
}
\description{
Method equivalent of addition operator \code{series + other}.
}
\examples{
pl$Series(1:3)$add(pl$Series(11:13))
pl$Series(1:3)$add(11:13)
pl$Series(1:3)$add(1L)

pl$Series("a")$add("-z")
}
\seealso{
\itemize{
\item \link[=S3_arithmetic]{Arithmetic operators}
}
}
