% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_alias}
\alias{Series_alias}
\title{Change name of Series}
\usage{
Series_alias(name)
}
\arguments{
\item{name}{New name.}
}
\value{
Series
}
\description{
Change name of Series
}
\examples{
pl$Series(1:3, name = "alice")$alias("bob")
}
