% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_all}
\alias{Series_all}
\title{Reduce Boolean Series with ALL}
\usage{
Series_all()
}
\value{
bool
}
\description{
Reduce Boolean Series with ALL
}
\examples{
pl$Series(c(TRUE,TRUE,NA))$all()
}
\keyword{Series}
