% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\docType{data}
\name{Series_any}
\alias{Series_any}
\alias{any}
\title{Reduce Boolean Series with ANY}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_any
}
\value{
bool
}
\description{
Reduce Boolean Series with ANY
}
\examples{
pl$Series(c(TRUE,FALSE,NA))$any()
}
\keyword{Series}
