% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{Series_append}
\alias{Series_append}
\title{append (default immutable)}
\usage{
Series_append(other, immutable = TRUE)
}
\arguments{
\item{other}{Series to append}

\item{immutable}{bool should append be immutable, default TRUE as mutable operations should
be avoided in plain R API's.}
}
\value{
Series
}
\description{
append two Series, see details for mutability
}
\details{
if immutable = FLASE, the Series object will not behave as immutable. This mean
appending to this Series will affect any variable pointing to this memory location. This will break
normal scoping rules of R. Polars-clones are cheap. Mutable operations are likely never needed in
any sense.
}
