% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\docType{data}
\name{Series_arg_max}
\alias{Series_arg_max}
\title{idx to max value}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_arg_max
}
\value{
bool
}
\description{
idx to max value
}
\examples{
pl$Series(c(5,1))$arg_max()
}
\keyword{datasets}
