% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_chunk_lengths}
\alias{Series_chunk_lengths}
\title{Lengths of Series memory chunks}
\usage{
Series_chunk_lengths()
}
\value{
Numeric vector. Output length is the number of chunks, and the sum of
the output is equal to the length of the full Series.
}
\description{
Lengths of Series memory chunks
}
\examples{
chunked_series = c(pl$Series(1:3), pl$Series(1:10))
chunked_series$chunk_lengths()
}
