% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{Series_chunk_lengths}
\alias{Series_chunk_lengths}
\title{Append two Series}
\usage{
Series_chunk_lengths()

Series_chunk_lengths()
}
\value{
numeric vector

numeric vector
}
\description{
Append Series with other Series. Imutable.

Append Series with other Series. Imutable.
}
\examples{
chunked_series = c(pl$Series(1:3),pl$Series(1:10))
chunked_series$chunk_lengths()
chunked_series = c(pl$Series(1:3),pl$Series(1:10))
chunked_series$chunk_lengths()
}
