% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_chunk_lengths}
\alias{Series_chunk_lengths}
\alias{chunk_lengths}
\title{Lengths of Series memory chunks}
\usage{
Series_chunk_lengths()
}
\value{
numeric vector. Length is number of chunks. Sum of lengths is equal to size of Series.
}
\description{
Get the Lengths of Series memory chunks as vector.
}
\examples{
chunked_series = c(pl$Series(1:3), pl$Series(1:10))
chunked_series$chunk_lengths()
}
\keyword{Series}
