% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_clone}
\alias{Series_clone}
\title{Clone a Series}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_clone
}
\value{
Series
}
\description{
Rarely useful as Series are nearly 100\% immutable
Any modification of a Series should lead to a clone anyways.
}
\examples{
s1 = pl$Series(1:3);
s2 =  s1$clone();
s3 = s1
pl$mem_address(s1) != pl$mem_address(s2)
pl$mem_address(s1) == pl$mem_address(s3)

}
\keyword{Series}
