% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_compare}
\alias{Series_compare}
\alias{==.RPolarsSeries}
\alias{!=.RPolarsSeries}
\alias{<.RPolarsSeries}
\alias{>.RPolarsSeries}
\alias{<=.RPolarsSeries}
\alias{>=.RPolarsSeries}
\title{Compare Series}
\usage{
Series_compare(other, op)

\method{==}{RPolarsSeries}(s1, s2)

\method{!=}{RPolarsSeries}(s1, s2)

\method{<}{RPolarsSeries}(s1, s2)

\method{>}{RPolarsSeries}(s1, s2)

\method{<=}{RPolarsSeries}(s1, s2)

\method{>=}{RPolarsSeries}(s1, s2)
}
\arguments{
\item{other}{A Series or something a Series can be created from}

\item{op}{The chosen operator, must be one of \code{"equal"}, \code{"not_equal"},
\code{"lt"}, \code{"gt"}, \code{"lt_eq"} or \code{"gt_eq"}}

\item{s1}{lhs Series}

\item{s2}{rhs Series or any into Series}
}
\value{
Series
}
\description{
Check the (in)equality of two Series.
}
\examples{
# We can either use `compare()`...
pl$Series(1:5)$compare(pl$Series(c(1:3, NA_integer_, 10L)), op = "equal")

# ... or the more classic way
pl$Series(1:5) == pl$Series(c(1:3, NA_integer_, 10L))
}
