% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_compare}
\alias{Series_compare}
\alias{compare}
\alias{==.Series}
\alias{!=.Series}
\alias{<.Series}
\alias{>.Series}
\alias{<=.Series}
\alias{>=.Series}
\title{Compare Series}
\usage{
Series_compare(other, op)

\method{==}{Series}(s1, s2)

\method{!=}{Series}(s1, s2)

\method{<}{Series}(s1, s2)

\method{>}{Series}(s1, s2)

\method{<=}{Series}(s1, s2)

\method{>=}{Series}(s1, s2)
}
\arguments{
\item{other}{A Series or something a Series can be created from}

\item{op}{the chosen operator a String either: 'equal', 'not_equal', 'lt', 'gt', 'lt_eq' or 'gt_eq'}
}
\value{
Series
}
\description{
compare two Series
}
\examples{
pl$Series(1:5) == pl$Series(c(1:3,NA_integer_,10L))
}
\keyword{Series}
