% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_div}
\alias{Series_div}
\alias{div}
\alias{/.Series}
\title{div Series}
\usage{
Series_div(other)

\method{/}{Series}(s1, s2)
}
\value{
Series
}
\description{
Series arithmetics
}
\examples{
pl$Series(1:3)$div(11:13)
pl$Series(1:3)$div(pl$Series(11:13))
pl$Series(1:3)$div(1L)
2L / pl$Series(1:3)
pl$Series(1:3) / 2L
}
\keyword{Series}
