% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_dtype}
\alias{Series_dtype}
\alias{Series}
\alias{Series_flags}
\title{Get data type of Series}
\usage{
Series_dtype()

Series_flags()
}
\value{
DataType

DataType
}
\description{
Get data type of Series

Get data type of Series
}
\details{
property sorted flags are not settable, use set_sorted
}
\examples{
pl$Series(1:4)$dtype
pl$Series(c(1,2))$dtype
pl$Series(letters)$dtype
pl$Series(1:4)$sort()$flags
}
\keyword{Series}
