% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_is_numeric}
\alias{Series_is_numeric}
\title{Check if the Series is numeric}
\usage{
Series_is_numeric()
}
\value{
A logical value
}
\description{
This checks whether the Series DataType is in \code{pl$numeric_dtypes}.
}
\examples{
pl$Series(1:4)$is_numeric()
pl$Series(c("a", "b", "c"))$is_numeric()
pl$numeric_dtypes
}
