% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_is_sorted}
\alias{Series_is_sorted}
\alias{is_sorted}
\title{is_sorted}
\usage{
Series_is_sorted(descending = FALSE)
}
\arguments{
\item{descending}{Check if the Series is sorted in descending order.}
}
\value{
DataType
}
\description{
is_sorted
}
\details{
property sorted flags are not settable, use set_sorted
}
\examples{
pl$Series(1:4)$sort()$is_sorted()
}
\keyword{Series}
