% Generated by mini-roxygen (roxygen2 compatible): do not edit by hand
% Please edit documentation in R/series-list.R
\name{series_list_to_struct}
\alias{series_list_to_struct}
\title{Convert the series of type List to a series of type Struct}
\usage{
series_list_to_struct(fields)
}
\arguments{
\item{fields}{A character vector of field names.}
}
\value{
A \link[=Series]{polars Series}
}
\description{
Convert the series of type List to a series of type Struct
}
\examples{
# Convert list to struct with explicit field names:
s1 <- as_polars_series(list(0:2, 0:1))
s2 <- s1$list$to_struct(c("one", "two", "three"))
s2
s2$struct$fields

# Convert list to struct with field name assignment by
# index from a list of names:
s1$list$to_struct(fields = c("one", "two", "three"))$struct$unnest()
}
\seealso{
\itemize{
\item \code{\link[=expr_list_to_struct]{<expr>$list$to_struct()}}
}
}
