% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\docType{data}
\name{Series_max}
\alias{Series_max}
\title{max}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_max
}
\value{
Series
}
\description{
Reduce Series with max
}
\details{
Dtypes in {Int8, UInt8, Int16, UInt16} are cast to
Int64 before maxming to prevent overflow issues.
}
\examples{
pl$Series(c(1:2,NA,3,5))$max() # a NA is dropped always
pl$Series(c(1:2,NA,3,NaN,4,Inf))$max() # NaN carries / poisons
pl$Series(c(1:2,3,Inf,4,-Inf,5))$max() # Inf-Inf is NaN
}
\keyword{Series}
