% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{Series_mul}
\alias{Series_mul}
\alias{mul}
\alias{*.Series}
\title{mul Series}
\usage{
Series_mul(other)

\method{*}{Series}(s1, s2)
}
\value{
Series
}
\description{
Series arithmetics
}
\examples{
pl$Series(1:3)$mul(11:13)
pl$Series(1:3)$mul(pl$Series(11:13))
pl$Series(1:3)$mul(1L)
1L + pl$Series(1:3)
pl$Series(1:3) + 1L
}
\keyword{Series}
