% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_n_unique}
\alias{Series_n_unique}
\title{Count unique values in Series}
\usage{
<Series>$n_unique()
}
\value{
Expr
}
\description{
Return count of unique values in Series
}
\examples{
pl$Series(1:4)$n_unique()
}
\keyword{Series}
