% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_print}
\alias{Series_print}
\alias{print.Series}
\title{Print Series}
\usage{
\method{print}{Series}(x)
}
\arguments{
\item{x}{Series}
}
\value{
invisible(self)
}
\description{
Print Series
}
\examples{
print(Series(1:3))
}
\keyword{Series}
