% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_rename}
\alias{Series_rename}
\alias{rename}
\title{Rename a series}
\format{
method
}
\usage{
Series_rename(name, in_place = FALSE)
}
\arguments{
\item{name}{string the new name}

\item{in_place}{bool rename in-place, breaks immutability
If true will throw an error unless this option has been set:
\code{pl$set_rpolars_options(strictly_immutable = FALSE)}}
}
\value{
bool
}
\description{
Rename a series
}
\examples{
pl$Series(1:4,"bob")$rename("alice")
}
\keyword{Series}
