% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_rep}
\alias{Series_rep}
\alias{rep}
\title{duplicate and concatenate a series}
\format{
method
}
\usage{
Series_rep(n, rechunk = TRUE)
}
\arguments{
\item{n}{number of times to repeat}

\item{rechunk}{bool default true, reallocate object in memory.
If FALSE the Series will take up less space, If TRUE calculations might be faster.}
}
\value{
bool
}
\description{
duplicate and concatenate a series
}
\details{
This function in not implemented in pypolars
}
\examples{
pl$Series(1:2, "bob")$rep(3)
}
\keyword{Series}
