% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_set_sorted}
\alias{Series_set_sorted}
\title{Set a sorted flag on a Series}
\usage{
Series_set_sorted(descending = FALSE, in_place = FALSE)
}
\arguments{
\item{descending}{Sort the columns in descending order.}

\item{in_place}{If \code{TRUE}, this will set the flag mutably and return NULL.
Remember to use \code{options(polars.strictly_immutable = FALSE)} before using
this parameter, otherwise an error will occur. If \code{FALSE} (default), it will
return a cloned Series with the flag.}
}
\value{
A \link[=Series_class]{Series} with a flag
}
\description{
Set a sorted flag on a Series
}
\details{
Use \code{\link[=Series_class]{$flags}} to see the values of the sorted flags.
}
\examples{
s = pl$Series(1:4)$set_sorted()
s$flags
}
