% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_set_sorted}
\alias{Series_set_sorted}
\title{Set sorted}
\usage{
Series_set_sorted(descending = FALSE, in_place = FALSE)
}
\arguments{
\item{descending}{Sort the columns in descending order.}

\item{in_place}{if TRUE, will set flag mutably and return NULL. Remember to use
pl$set_polars_options(strictly_immutable = FALSE) otherwise an error will be thrown. If FALSE
will return a cloned Series with set_flag which in the very most cases should be just fine.}
}
\value{
Series invisible
}
\description{
Set sorted
}
\examples{
s = pl$Series(1:4)$set_sorted()
s$flags
}
\keyword{Series}
