% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_sort}
\alias{Series_sort}
\title{Sort this Series}
\usage{
Series_sort(descending = FALSE, in_place = FALSE)
}
\arguments{
\item{descending}{Sort in descending order..}

\item{in_place}{bool sort mutable in-place, breaks immutability
If true will throw an error unless this option has been set:
\code{options(polars.strictly_immutable = FALSE)}}
}
\value{
Series
}
\description{
Sort this Series
}
\examples{
pl$Series(c(1, NA, NaN, Inf, -Inf))$sort()
}
\keyword{Series}
