% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_sort}
\alias{Series_sort}
\title{Sort this Series}
\usage{
Series_sort(reverse = FALSE, in_place = FALSE)
}
\arguments{
\item{reverse}{bool reverse(descending) sort}

\item{in_place}{bool sort mutable in-place, breaks immutability
If true will throw an error unless this option has been set:
\code{pl$set_rpolars_options(strictly_immutable = F)}}
}
\value{
Series
}
\description{
Sort this Series
}
\examples{
pl$Series(c(1,NA,NaN,Inf,-Inf))$sort()
}
\keyword{Series}
