% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_std}
\alias{Series_std}
\title{Compute the standard deviation of a Series}
\usage{
Series_std(ddof = 1)
}
\arguments{
\item{ddof}{Delta Degrees of Freedom: the divisor used in the calculation is
N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
A numeric value
}
\description{
Compute the standard deviation of a Series
}
\examples{
pl$Series(1:10)$std()
}
