% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_struct_unnest}
\alias{Series_struct_unnest}
\alias{Series_struct}
\title{Convert this struct Series to a DataFrame with a separate column for
each field}
\usage{
Series_struct_unnest()
}
\value{
A DataFrame
}
\description{
Convert this struct Series to a DataFrame with a separate column for
each field
}
\examples{
s = pl$Series(values = c(1, 2), dtype = pl$Struct(foo = pl$Float64))
s$struct$unnest()
}
