% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_sub}
\alias{Series_sub}
\alias{sub}
\alias{-.Series}
\title{sub Series}
\usage{
Series_sub(other)

\method{-}{Series}(s1, s2)
}
\arguments{
\item{other}{Series or into Series}

\item{s1}{lhs Series}

\item{s2}{rhs Series or any into Series}
}
\value{
Series
}
\description{
Series arithmetics
}
\examples{
pl$Series(1:3)$sub(11:13)
pl$Series(1:3)$sub(pl$Series(11:13))
pl$Series(1:3)$sub(1L)
1L - pl$Series(1:3)
pl$Series(1:3) - 1L
}
\keyword{Series}
