% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\docType{data}
\name{Series_sum}
\alias{Series_sum}
\alias{sum}
\title{Sum}
\format{
An object of class \code{character} of length 1.
}
\usage{
Series_sum
}
\value{
Series
}
\description{
Reduce Series with sum
}
\details{
Dtypes in {Int8, UInt8, Int16, UInt16} are cast to
Int64 before summing to prevent overflow issues.
}
\examples{
pl$Series(c(1:2,NA,3,5))$sum() # a NA is dropped always
pl$Series(c(1:2,NA,3,NaN,4,Inf))$sum() # NaN carries / poisons
pl$Series(c(1:2,3,Inf,4,-Inf,5))$sum() # Inf-Inf is NaN
}
\keyword{Series}
