% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_to_lit}
\alias{Series_to_lit}
\title{Convert a Series to literal}
\usage{
Series_to_lit()
}
\value{
Expr
}
\description{
Convert a Series to literal
}
\examples{
pl$Series(list(1:1, 1:2, 1:3, 1:4))$
  print()$
  to_lit()$
  list$len()$
  sum()$
  cast(pl$dtypes$Int8)$
  to_series()
}
