% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_to_lit}
\alias{Series_to_lit}
\alias{to_lit}
\title{Series to Literal}
\usage{
Series_to_lit()
}
\value{
Expr
}
\description{
convert Series to literal to perform modification and return
}
\examples{
s = pl$Series(list(1:1, 1:2, 1:3, 1:4))

s$to_lit()
}
\keyword{Series}
