% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_value_counts}
\alias{Series_value_counts}
\title{Count the occurrences of unique values}
\usage{
Series_value_counts(
  ...,
  sort = TRUE,
  parallel = FALSE,
  name = "count",
  normalize = FALSE
)
}
\arguments{
\item{...}{Ignored.}

\item{sort}{Ensure the output is sorted from most values to least.}

\item{parallel}{Better to turn this off in the aggregation context, as it can
lead to contention.}

\item{name}{Give the resulting count column a specific name. The default is
\code{"count"} if \code{normalize = FALSE} and \code{"proportion"} if  \code{normalize = TRUE}.}

\item{normalize}{If \code{TRUE}, it gives relative frequencies of the unique
values instead of their count.}
}
\value{
DataFrame
}
\description{
Count the occurrences of unique values
}
\examples{
as_polars_series(iris$Species, name = "flower species")$value_counts()
}
