% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_value_counts}
\alias{Series_value_counts}
\title{Count the occurrences of unique values}
\usage{
Series_value_counts(sort = TRUE, parallel = FALSE)
}
\arguments{
\item{sort}{Ensure the output is sorted from most values to least.}

\item{parallel}{Better to turn this off in the aggregation context, as it can
lead to contention.}
}
\value{
DataFrame
}
\description{
Count the occurrences of unique values
}
\examples{
pl$Series(iris$Species, name = "flower species")$value_counts()
}
