% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series__series.R
\name{Series_var}
\alias{Series_var}
\title{Var}
\usage{
<Series>$var(ddof = 1)
}
\arguments{
\item{ddof}{integer Delta Degrees of Freedom: the divisor used in the calculation is N - ddof, where N represents the number of elements. By default ddof is 1.}
}
\value{
A new \code{Series} object with applied aggregation.
}
\description{
Aggregate the columns of this Series to their variance values.
}
\examples{
pl$Series(1:10)$var()
}
\keyword{R}
\keyword{scalar}
\keyword{value}
