% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after-wrappers.R
\name{show_all_public_functions}
\alias{show_all_public_functions}
\alias{pl_show_all_public_functions}
\title{show all public functions / objects}
\usage{
pl_show_all_public_functions()
}
\description{
print any object(function, RPolarsDataType) available via \verb{pl$}.
}
\examples{
\dontrun{
pl$show_all_public_functions()
}
}
\keyword{functions}
