% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{sink_ipc}
\alias{sink_ipc}
\alias{LazyFrame_sink_ipc}
\title{LazyFrame stream output to arrow ipc file}
\usage{
LazyFrame_sink_ipc(path, compression = "zstd", maintain_order = TRUE)
}
\arguments{
\item{path}{string, the path of the arrow ipc file}

\item{compression}{NULL or string, the compression method. One of {'lz4', 'zstd'} if not NULL.}

\item{maintain_order}{bool, whether maintain the order the data was processed.}
}
\description{
Stream the content of LazyFrame into an arrow ipc file.
}
