% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyframe__lazy.R
\name{sink_ipc}
\alias{sink_ipc}
\alias{LazyFrame_sink_ipc}
\title{LazyFrame stream output to arrow ipc file}
\usage{
LazyFrame_sink_ipc(
  path,
  compression = "zstd",
  maintain_order = TRUE,
  type_coercion = TRUE,
  predicate_pushdown = TRUE,
  projection_pushdown = TRUE,
  simplify_expression = TRUE,
  slice_pushdown = TRUE,
  no_optimization = FALSE
)
}
\arguments{
\item{path}{string, the path of the arrow ipc file}

\item{compression}{NULL or string, the compression method. One of {'lz4', 'zstd'} if not NULL.}

\item{maintain_order}{bool, whether maintain the order the data was processed.}

\item{type_coercion}{Boolean. Coerce types such that operations succeed and
run on minimal required memory.}

\item{predicate_pushdown}{Boolean. Applies filters as early as possible at
scan level.}

\item{projection_pushdown}{Boolean. Select only the columns that are needed at the scan level.}

\item{simplify_expression}{Boolean. Various optimizations, such as constant folding
and replacing expensive operations with faster alternatives.}

\item{slice_pushdown}{Boolean. Only load the required slice from the scan
Don't materialize sliced outputs
level. Don't materialize sliced outputs (e.g. \code{join$head(10)}).}

\item{no_optimization}{Boolean. Turn off the following optimizations:
predicate_pushdown = FALSE
projection_pushdown = FALSE
slice_pushdown = FALSE
common_subplan_elimination = FALSE}
}
\description{
Stream the content of LazyFrame into an arrow ipc file.
}
