% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-context.R
\name{sql_context__tables}
\alias{sql_context__tables}
\title{Return a list of the registered table names}
\usage{
sql_context__tables()
}
\value{
A character vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This method will return the same values as the "SHOW TABLES" SQL statement,
but as a vector instead of a frame.
}
\examples{
# Executing as SQL:
frame_data <- pl$DataFrame(x = 1)
ctx <- pl$SQLContext(hello_world=frame_data, foo = data.frame(x = 2))
ctx$execute("SHOW TABLES")$collect()

# Calling the method:
ctx$tables()
}
