% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_class}
\alias{SQLContext_class}
\title{Run SQL queries against DataFrame/LazyFrame data.}
\description{
Run SQL queries against DataFrame/LazyFrame data.
}
\details{
Currently, only available when built with the \code{full} feature.
See \code{\link[=pl_polars_info]{pl$polars_info()}} for more information.
}
\examples{
\dontshow{if (pl$polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
lf = pl$LazyFrame(a = 1:3, b = c("x", NA, "z"))
res = pl$SQLContext(frame = lf)$execute(
  "SELECT b, a*2 AS two_a FROM frame WHERE b IS NOT NULL"
)
res$collect()
\dontshow{\}) # examplesIf}
}
\keyword{SQLContext}
