% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_execute}
\alias{SQLContext_execute}
\title{Execute SQL query against the registered data}
\usage{
SQLContext_execute(query)
}
\arguments{
\item{query}{A valid string SQL query.}
}
\value{
A \link[=LazyFrame_class]{LazyFrame}
}
\description{
Parse the given SQL query and execute it against the registered frame data.
}
\examples{
\dontshow{if (polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
query = "SELECT * FROM mtcars WHERE cyl = 4"

pl$SQLContext(mtcars = mtcars)$execute(query)
\dontshow{\}) # examplesIf}
}
