% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_register}
\alias{SQLContext_register}
\title{Register a single data as a table}
\usage{
<SQLContext>$register(name, frame)
}
\arguments{
\item{name}{A string name to register the frame as.}

\item{frame}{A \link[=LazyFrame_class]{LazyFrame} like object to register.}
}
\value{
Returns the \link{SQLContext_class} object invisibly.
}
\description{
Register a single frame as a table, using the given name.
}
\examples{
\dontshow{if (polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ctx = pl$SQLContext()
ctx$register("mtcars", mtcars)

ctx$execute("SELECT * FROM mtcars LIMIT 5")$collect()
\dontshow{\}) # examplesIf}
}
