% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{SQLContext_register_many}
\alias{SQLContext_register_many}
\title{Register multiple data as tables}
\usage{
<SQLContext>$register_many(...)
}
\arguments{
\item{...}{Name-value pairs of \link[=LazyFrame_class]{LazyFrame} like objects to register.}
}
\value{
Returns the \link{SQLContext_class} object invisibly.
}
\description{
Register multiple frames as tables.
}
\examples{
\dontshow{if (polars_info()$features$sql) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ctx = pl$SQLContext()
r_df = mtcars
pl_df = pl$DataFrame(mtcars)
pl_lf = pl$LazyFrame(mtcars)

ctx$register_many(r_df = r_df, pl_df = pl_df, pl_lf = pl_lf)

ctx$execute(
  "SELECT * FROM r_df
  UNION ALL
  SELECT * FROM pl_df
  UNION ALL
  SELECT * FROM pl_lf"
)$collect()
\dontshow{\}) # examplesIf}
}
