% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{to_list}
\alias{to_list}
\alias{DataFrame_to_list}
\title{return polars DataFrame as R lit of vectors}
\usage{
DataFrame_to_list()
}
\value{
R list of vectors
}
\description{
return polars DataFrame as R lit of vectors
}
\examples{
pl$DataFrame(iris)$to_list()
}
\keyword{DataFrame}
