% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe__frame.R
\name{to_list}
\alias{to_list}
\alias{DataFrame_to_list}
\title{Return Polars DataFrame as a list of vectors}
\usage{
DataFrame_to_list(unnest_structs = TRUE)
}
\arguments{
\item{unnest_structs}{Boolean. If \code{TRUE} (default), then \verb{$unnest()} is applied
on any struct column.}
}
\value{
R list of vectors
}
\description{
Return Polars DataFrame as a list of vectors
}
\details{
For simplicity reasons, this implementation relies on unnesting all structs
before exporting to R. If \code{unnest_structs = FALSE}, then \code{struct} columns
will be returned as nested lists, where each row is a list of values. Such a
structure is not very typical or efficient in R.
}
\examples{
pl$DataFrame(iris)$to_list()
}
\keyword{DataFrame}
