% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_result.R
\name{unwrap}
\alias{unwrap}
\title{rust-like unwrapping of result. Useful to keep error handling on the R side.}
\usage{
unwrap(result, context = NULL, call = sys.call(1L))
}
\arguments{
\item{result}{a list here either element ok or err is NULL, or both if ok is litteral NULL}

\item{context}{a msg to prefix a raised error with}

\item{call}{context of error or string}
}
\value{
the ok-element of list , or a error will be thrown
}
\description{
rust-like unwrapping of result. Useful to keep error handling on the R side.
}
\examples{

structure(list(ok = "foo", err = NULL), class = "extendr_result")

tryCatch(
  unwrap(
    structure(
      list(ok = NULL, err = "something happen on the rust side"),
      class = "extendr_result"
    )
  ),
  error = function(err) as.character(err)
)
}
\keyword{internal}
