% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unwrap}
\alias{unwrap}
\title{rust-like unwrapping of result. Useful to keep error handling on the R side.}
\usage{
unwrap(result, class = "rust result error", call = sys.call(1L), ...)
}
\arguments{
\item{result}{a list here either element ok or err is NULL, or both if ok is litteral NULL}

\item{class}{class of thrown error}

\item{call}{context of error}

\item{...}{not used}
}
\value{
the ok-element of list , or a error will be thrown
}
\description{
rust-like unwrapping of result. Useful to keep error handling on the R side.
}
\examples{
unwrap(list(ok="foo",err=NULL))
}
