% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_trait.R
\name{where_in}
\alias{where_in}
\title{Internal generic method to point to which public method the user got wrong}
\usage{
where_in(err, context)
}
\arguments{
\item{err}{any type which impl as.character}

\item{call}{calling context}
}
\value{
err as string
}
\description{
Internal generic method to point to which public method the user got wrong
}
\examples{
#
}
\keyword{internal}
