% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{wrap_e}
\alias{wrap_e}
\title{wrap as literal}
\usage{
wrap_e(e)
}
\arguments{
\item{e}{an Expr(polars) or any R expression}
}
\value{
Expr
}
\description{
wrap as literal
}
\details{
tiny wrapper to allow skipping calling lit on rhs of binary operator
}
\examples{
pl::col("foo") < 5
}
