% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{wrap_e}
\alias{wrap_e}
\title{wrap as literal}
\usage{
wrap_e(e, str_to_lit = TRUE)
}
\arguments{
\item{e}{an Expr(polars) or any R expression}
}
\value{
Expr
}
\description{
use robj_to!(Expr) on rust side or rarely wrap_e on R-side
This function is only kept for reference
}
\details{
used internally to ensure an object is an expression
}
\examples{
pl$col("foo") < 5
}
\keyword{internal}
