% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{wrap_e_result}
\alias{wrap_e_result}
\title{wrap as literal}
\usage{
wrap_e_result(e, str_to_lit = TRUE)
}
\arguments{
\item{e}{an Expr(polars) or any R expression}
}
\value{
Expr
}
\description{
wrap as literal
}
\details{
used internally to ensure an object is an expression
}
\examples{
pl$col("foo") < 5
}
\keyword{internal}
