% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr__expr.R
\name{wrap_elist_result}
\alias{wrap_elist_result}
\title{internal wrap_elist_result}
\usage{
wrap_elist_result(elist, str_to_lit = TRUE)
}
\arguments{
\item{elist}{a list Expr or any R object Into\if{html}{\out{<Expr>}} (passable to pl$lit)}
}
\value{
Expr
}
\description{
make sure all elements of a list is wrapped as Expr
DEPRECATED:  prefer robj_to!(VecPlExpr) on rust side
Capture any conversion error in the result
}
\details{
Used internally to ensure an object is a list of expression
The output is wrapped in a result, which can contain an ok or
err value.
}
\examples{
polars:::wrap_elist_result(list(pl$lit(42), 42, 1:3))
}
\keyword{internal}
